Cardano.Api.Block

Description

Blocks in the blockchain

Blocks in the context of an era

data Block era where

Deprecated: Use getBlockHeader instead

A blockchain block in a particular Cardano era.

Constructors

ByronBlock :: ByronBlock -> Block ByronEra

ShelleyBlock :: forall era. ShelleyBasedEra era -> ShelleyBlock (ConsensusProtocol era) (ShelleyLedgerEra era) -> Block era

pattern Block :: BlockHeader -> [Tx era] -> Block era

Deprecated: Use getBlockHeader instead

A block consists of a header and a body containing transactions.

data BlockHeader

Constructors

BlockHeader !SlotNo !(Hash BlockHeader) !BlockNo

getBlockHeader :: Block era -> BlockHeader

getBlockTxs :: Block era -> [Tx era]

Blocks in the context of a consensus mode

data BlockInMode where

A Block in one of the eras. TODO Rename this to BlockInEra

Constructors

BlockInMode :: forall era. CardanoEra era -> Block era -> BlockInMode

fromConsensusBlock :: CardanoBlock StandardCrypto ~ block => block -> BlockInMode

toConsensusBlock :: CardanoBlock StandardCrypto ~ block => BlockInMode -> block

Points on the chain

data ChainPoint

Constructors

ChainPointAtGenesis

ChainPoint !SlotNo !(Hash BlockHeader)

newtype SlotNo

The 0-based index for the Ourboros time slot.

Constructors

SlotNo

Fields

unSlotNo :: Word64

newtype EpochNo

An epoch, i.e. the number of the epoch.

Constructors

EpochNo

Fields

unEpochNo :: Word64

toConsensusPoint :: forall ledgerera protocol. ShelleyCompatible protocol ledgerera => ChainPoint -> Point (ShelleyBlock protocol ledgerera)

Convert a Point for single Shelley-era block type

fromConsensusPoint :: ShelleyCompatible protocol ledgerera => Point (ShelleyBlock protocol ledgerera) -> ChainPoint

Convert a Point for single Shelley-era block type

fromConsensusPointHF :: forall block (xs :: [Type]). HeaderHash block ~ OneEraHash xs => Point block -> ChainPoint

Convert a Point for multi-era block type

toConsensusPointHF :: forall block (xs :: [Type]). HeaderHash block ~ OneEraHash xs => ChainPoint -> Point block

Convert a Point for multi-era block type

Tip of the chain

data ChainTip

This is like a ChainPoint but is conventionally used for the tip of the chain: that is the most recent block at the end of the chain.

It also carries the BlockNo of the chain tip.

Constructors

ChainTipAtGenesis

ChainTip !SlotNo !(Hash BlockHeader) !BlockNo

newtype BlockNo

The 0-based index of the block in the blockchain. BlockNo is <= SlotNo and is only equal at slot N if there is a block for every slot where N <= SlotNo.

Constructors

BlockNo

Fields

unBlockNo :: Word64

chainTipToChainPoint :: ChainTip -> ChainPoint

fromConsensusTip :: CardanoBlock StandardCrypto ~ block => Tip block -> ChainTip

Data family instances

data family Hash keyrole

chainPointToHeaderHash :: ChainPoint -> Maybe (Hash BlockHeader)

chainPointToSlotNo :: ChainPoint -> Maybe SlotNo

makeChainTip :: WithOrigin BlockNo -> ChainPoint -> ChainTip