Cardano.Api.Byron

Description

This module provides a library interface that is intended to be the complete API for Byron covering everything, including exposing constructors for the lower level types.

Documentation

module Cardano.Api

data family AsType t

Cryptographic key interface

data family VerificationKey keyrole

data family SigningKey keyrole

data SomeByronSigningKey

Constructors

AByronSigningKeyLegacy (SigningKey ByronKeyLegacy)

AByronSigningKey (SigningKey ByronKey)

Hashes

data family Hash keyrole

Payment addresses

Constructing and inspecting Byron payment addresses

data Address addrtype where

Constructors

ByronAddress :: Address -> Address ByronAddr

data NetworkId

Constructors

Mainnet

Testnet !NetworkMagic

Building transactions

Constructing and inspecting transactions

newtype TxId

Constructors

TxId (Hash StandardCrypto EraIndependentTxBody)

data TxIn

Constructors

TxIn TxId TxIx

data TxOut ctx era

Constructors

TxOut (AddressInEra era) (TxOutValue era) (TxOutDatum ctx era) (ReferenceScript era)

newtype TxIx

Constructors

TxIx Word

Signing transactions

Creating transaction witnesses one by one, or all in one go.

data ATxAux a

Constructors

ATxAux

Fields

aTaTx :: !(Annotated Tx a)

aTaWitness :: !(Annotated TxWitness a)

aTaAnnotation :: !a

Incremental signing and separate witnesses

data KeyWitness era where

Constructors

ByronKeyWitness :: TxInWitness -> KeyWitness ByronEra

data WitnessNetworkIdOrByronAddress

Constructors

WitnessNetworkId !NetworkId

WitnessByronAddress !(Address ByronAddr)

Errors

class Error e where

Methods

prettyError :: e -> Doc ann

data FileError e

Constructors

FileError FilePath e

FileErrorTempFile FilePath FilePath Handle

FileDoesNotExistError FilePath

FileIOError FilePath IOException

Low level protocol interaction with a Cardano node

data LocalNodeConnectInfo

Constructors

LocalNodeConnectInfo ConsensusModeParams NetworkId SocketPath

data LocalNodeClientProtocols block point tip slot tx txid txerr (query :: Type -> Type) (m :: Type -> Type)

Constructors

LocalNodeClientProtocols (LocalChainSyncClient block point tip m) (Maybe (LocalTxSubmissionClient tx txerr m ())) (Maybe (LocalStateQueryClient block point query m ())) (Maybe (LocalTxMonitorClient txid tx slot m ()))

Chain sync protocol

newtype ChainSyncClient header point tip (m :: Type -> Type) a

A chain sync protocol client, on top of some effect m. The first choice of request is within that m.

Constructors

ChainSyncClient

Fields

runChainSyncClient :: m (ClientStIdle header point tip m a)

Local tx submission

newtype LocalTxSubmissionClient tx reject (m :: Type -> Type) a

Constructors

LocalTxSubmissionClient (m (LocalTxClientStIdle tx reject m a))

Local state query

newtype LocalStateQueryClient block point (query :: Type -> Type) (m :: Type -> Type) a

Constructors

LocalStateQueryClient

Fields

runLocalStateQueryClient :: m (ClientStIdle block point query m a)

Address

newtype NetworkMagic

NetworkMagic is used to differentiate between different networks during the initial handshake.

Constructors

NetworkMagic

Fields

unNetworkMagic :: Word32

Update Proposal

newtype ByronUpdateProposal

Constructors

ByronUpdateProposal

Fields

unByronUpdateProposal :: AProposal ByteString

data ByronProtocolParametersUpdate

Constructors

ByronProtocolParametersUpdate

Fields

bPpuScriptVersion :: !(Maybe Word16)

bPpuSlotDuration :: !(Maybe Natural)

bPpuMaxBlockSize :: !(Maybe Natural)

bPpuMaxHeaderSize :: !(Maybe Natural)

bPpuMaxTxSize :: !(Maybe Natural)

bPpuMaxProposalSize :: !(Maybe Natural)

bPpuMpcThd :: !(Maybe LovelacePortion)

bPpuHeavyDelThd :: !(Maybe LovelacePortion)

bPpuUpdateVoteThd :: !(Maybe LovelacePortion)

bPpuUpdateProposalThd :: !(Maybe LovelacePortion)

bPpuUpdateProposalTTL :: !(Maybe SlotNumber)

bPpuSoftforkRule :: !(Maybe SoftforkRule)

bPpuTxFeePolicy :: !(Maybe TxFeePolicy)

bPpuUnlockStakeEpoch :: !(Maybe EpochNumber)

makeByronUpdateProposal :: NetworkId -> ProtocolVersion -> SoftwareVersion -> SystemTag -> InstallerHash -> SomeByronSigningKey -> ByronProtocolParametersUpdate -> ByronUpdateProposal

toByronLedgerUpdateProposal :: ByronUpdateProposal -> GenTx ByronBlock

makeProtocolParametersUpdate :: ByronProtocolParametersUpdate -> ProtocolParametersUpdate

Vote

newtype ByronVote

Constructors

ByronVote

Fields

unByronVote :: AVote ByteString

makeByronVote :: NetworkId -> SomeByronSigningKey -> ByronUpdateProposal -> Bool -> ByronVote

toByronLedgertoByronVote :: ByronVote -> GenTx ByronBlock

Conversions

fromByronTxIn :: TxIn -> TxIn

toByronLovelace :: Coin -> Maybe Lovelace

toByronNetworkMagic :: NetworkId -> NetworkMagic

toByronProtocolMagicId :: NetworkId -> ProtocolMagicId

toByronRequiresNetworkMagic :: NetworkId -> RequiresNetworkMagic

Hardcoded configuration parameters

applicationName :: ApplicationName

applicationVersion :: NumSoftwareVersion

softwareVersion :: SoftwareVersion

Serialization

serializeByronTx :: ATxAux ByteString -> TextEnvelope

writeByronTxFileTextEnvelopeCddl :: File content 'Out -> ATxAux ByteString -> IO (Either (FileError ()) ())