Cardano.Api.Certificate

Description

Certificates embedded in transactions

Documentation

data Certificate era where

Constructors

ShelleyRelatedCertificate :: forall era. ShelleyToBabbageEra era -> ShelleyTxCert (ShelleyLedgerEra era) -> Certificate era

ConwayCertificate :: forall era. ConwayEraOnwards era -> ConwayTxCert (ShelleyLedgerEra era) -> Certificate era

Registering stake address and delegating

data StakeAddressRequirements era where

Constructors

StakeAddrRegistrationConway :: forall era. ConwayEraOnwards era -> Coin -> StakeCredential -> StakeAddressRequirements era

StakeAddrRegistrationPreConway :: forall era. ShelleyToBabbageEra era -> StakeCredential -> StakeAddressRequirements era

data StakeDelegationRequirements era where

Constructors

StakeDelegationRequirementsConwayOnwards :: forall era. ConwayEraOnwards era -> StakeCredential -> Delegatee (EraCrypto (ShelleyLedgerEra era)) -> StakeDelegationRequirements era

StakeDelegationRequirementsPreConway :: forall era. ShelleyToBabbageEra era -> StakeCredential -> PoolId -> StakeDelegationRequirements era

makeStakeAddressDelegationCertificate :: StakeDelegationRequirements era -> Certificate era

makeStakeAddressRegistrationCertificate :: StakeAddressRequirements era -> Certificate era

makeStakeAddressUnregistrationCertificate :: StakeAddressRequirements era -> Certificate era

type PoolId = Hash StakePoolKey

Registering stake pools

data StakePoolRegistrationRequirements era where

Constructors

StakePoolRegistrationRequirementsConwayOnwards :: forall era. ConwayEraOnwards era -> PoolParams (EraCrypto (ShelleyLedgerEra era)) -> StakePoolRegistrationRequirements era

StakePoolRegistrationRequirementsPreConway :: forall era. ShelleyToBabbageEra era -> PoolParams (EraCrypto (ShelleyLedgerEra era)) -> StakePoolRegistrationRequirements era

data StakePoolRetirementRequirements era where

Constructors

StakePoolRetirementRequirementsConwayOnwards :: forall era. ConwayEraOnwards era -> PoolId -> EpochNo -> StakePoolRetirementRequirements era

StakePoolRetirementRequirementsPreConway :: forall era. ShelleyToBabbageEra era -> PoolId -> EpochNo -> StakePoolRetirementRequirements era

makeStakePoolRegistrationCertificate :: StakePoolRegistrationRequirements era -> Certificate era

makeStakePoolRetirementCertificate :: StakePoolRetirementRequirements era -> Certificate era

data StakePoolParameters

Constructors

StakePoolParameters

Fields

stakePoolId :: PoolId

stakePoolVRF :: Hash VrfKey

stakePoolCost :: Coin

stakePoolMargin :: Rational

stakePoolRewardAccount :: StakeAddress

stakePoolPledge :: Coin

stakePoolOwners :: [Hash StakeKey]

stakePoolRelays :: [StakePoolRelay]

stakePoolMetadata :: Maybe StakePoolMetadataReference

data StakePoolRelay

Constructors

StakePoolRelayIp (Maybe IPv4) (Maybe IPv6) (Maybe PortNumber)
One or both of IPv4 & IPv6

StakePoolRelayDnsARecord ByteString (Maybe PortNumber)
An DNS name pointing to a A or AAAA record.

StakePoolRelayDnsSrvRecord ByteString
A DNS name pointing to a SRV record.

data StakePoolMetadataReference

Constructors

StakePoolMetadataReference

Fields

stakePoolMetadataURL :: Text

stakePoolMetadataHash :: Hash StakePoolMetadata

Conway specific certificates

data CommitteeColdkeyResignationRequirements era where

Constructors

CommitteeColdkeyResignationRequirements :: forall era. ConwayEraOnwards era -> Credential 'ColdCommitteeRole (EraCrypto (ShelleyLedgerEra era)) -> Maybe (Anchor (EraCrypto (ShelleyLedgerEra era))) -> CommitteeColdkeyResignationRequirements era

data CommitteeHotKeyAuthorizationRequirements era where

Constructors

CommitteeHotKeyAuthorizationRequirements :: forall era. ConwayEraOnwards era -> Credential 'ColdCommitteeRole (EraCrypto (ShelleyLedgerEra era)) -> Credential 'HotCommitteeRole (EraCrypto (ShelleyLedgerEra era)) -> CommitteeHotKeyAuthorizationRequirements era

data DRepRegistrationRequirements era where

Constructors

DRepRegistrationRequirements :: forall era. ConwayEraOnwards era -> Credential 'DRepRole (EraCrypto (ShelleyLedgerEra era)) -> Coin -> DRepRegistrationRequirements era

data DRepUnregistrationRequirements era where

Constructors

DRepUnregistrationRequirements :: forall era. ConwayEraOnwards era -> Credential 'DRepRole (EraCrypto (ShelleyLedgerEra era)) -> Coin -> DRepUnregistrationRequirements era

data DRepUpdateRequirements era where

Constructors

DRepUpdateRequirements :: forall era. ConwayEraOnwards era -> Credential 'DRepRole (EraCrypto (ShelleyLedgerEra era)) -> DRepUpdateRequirements era

makeCommitteeColdkeyResignationCertificate :: CommitteeColdkeyResignationRequirements era -> Certificate era

makeCommitteeHotKeyAuthorizationCertificate :: CommitteeHotKeyAuthorizationRequirements era -> Certificate era

makeDrepRegistrationCertificate :: DRepRegistrationRequirements era -> Maybe (Anchor (EraCrypto (ShelleyLedgerEra era))) -> Certificate era

makeDrepUnregistrationCertificate :: DRepUnregistrationRequirements era -> Certificate era

makeDrepUpdateCertificate :: DRepUpdateRequirements era -> Maybe (Anchor (EraCrypto (ShelleyLedgerEra era))) -> Certificate era

makeStakeAddressAndDRepDelegationCertificate :: ConwayEraOnwards era -> StakeCredential -> Delegatee (EraCrypto (ShelleyLedgerEra era)) -> Coin -> Certificate era

Registering DReps

data DRepMetadataReference

Constructors

DRepMetadataReference

Fields

drepMetadataURL :: Text

drepMetadataHash :: Hash DRepMetadata

Special certificates

data GenesisKeyDelegationRequirements ere where

Constructors

GenesisKeyDelegationRequirements :: forall ere. ShelleyToBabbageEra ere -> Hash GenesisKey -> Hash GenesisDelegateKey -> Hash VrfKey -> GenesisKeyDelegationRequirements ere

data MirCertificateRequirements era where

Constructors

MirCertificateRequirements :: forall era. ShelleyToBabbageEra era -> MIRPot -> MIRTarget (EraCrypto (ShelleyLedgerEra era)) -> MirCertificateRequirements era

makeMIRCertificate :: MirCertificateRequirements era -> Certificate era

makeGenesisKeyDelegationCertificate :: GenesisKeyDelegationRequirements era -> Certificate era

data MIRTarget c

MIRTarget specifies if funds from either the reserves or the treasury are to be handed out to a collection of reward accounts or instead transfered to the opposite pot.

Constructors

StakeAddressesMIR !(Map (Credential 'Staking c) DeltaCoin)

SendToOppositePotMIR !Coin

data MIRPot

Constructors

ReservesMIR

TreasuryMIR

selectStakeCredentialWitness :: Certificate era -> Maybe StakeCredential

Get the stake credential witness for a certificate that requires it. Only stake address deregistration and delegation requires witnessing (witness can be script or key).

Internal conversion functions

toShelleyCertificate :: Certificate era -> TxCert (ShelleyLedgerEra era)

fromShelleyCertificate :: ShelleyBasedEra era -> TxCert (ShelleyLedgerEra era) -> Certificate era

toShelleyPoolParams :: StakePoolParameters -> PoolParams StandardCrypto

fromShelleyPoolParams :: PoolParams StandardCrypto -> StakePoolParameters

Data family instances

data family AsType t

A family of singleton types used in this API to indicate which type to use where it would otherwise be ambiguous or merely unclear.

Values of this type are passed to deserialisation functions for example.

Internal functions

filterUnRegCreds :: Certificate era -> Maybe StakeCredential

filterUnRegDRepCreds :: Certificate era -> Maybe (Credential 'DRepRole StandardCrypto)