Cardano.Api.ChainSync.ClientPipelined

Pipelined protocol type for the client

The protocol states from the point of view of the client.

newtype ChainSyncClientPipelined header point tip (m :: Type -> Type) a

Pipelined chain sync client. It can only pipeline MsgRequestNext messages, while the MsgFindIntersect are non pipelined. This has a penalty cost of an RTT, but they are sent relatively seldom and their response might impact how many messages one would like to pipeline. It also simplifies the receiver callback.

Constructors

ChainSyncClientPipelined

Fields

runChainSyncClientPipelined :: m (ClientPipelinedStIdle 'Z header point tip m a)

data ClientPipelinedStIdle (n :: N) header point tip (m :: Type -> Type) a where

Pipelined sender which starts in StIdle state. It can either

Send MsgRequestNext (no pipelining), which might be useful when we are at the tip of the chain. It can only be send when there is no pipelined message in flight (all responses were collected);

Pipeline MsgRequestNext;

Send MsgFindIntersect (no pipelining); It can only be send when there is no pipelined message in flight (all responses were collected);

Collect responses of pipelined message;

Terminate the protocol with by sending MsgDone.

Constructors

SendMsgRequestNext

Fields

:: forall (m :: Type -> Type) header point tip a. m ()

promptly invoked when MsgAwaitReply is received

-> ClientStNext 'Z header point tip m a

-> ClientPipelinedStIdle 'Z header point tip m a

SendMsgRequestNextPipelined

Fields

:: forall (m :: Type -> Type) (n :: N) header point tip a. m ()

promptly invoked when MsgAwaitReply is received

-> ClientPipelinedStIdle ('S n) header point tip m a

-> ClientPipelinedStIdle n header point tip m a

SendMsgFindIntersect :: forall point header tip (m :: Type -> Type) a. [point] -> ClientPipelinedStIntersect header point tip m a -> ClientPipelinedStIdle 'Z header point tip m a

CollectResponse :: forall (m :: Type -> Type) (n1 :: N) header point tip a. Maybe (m (ClientPipelinedStIdle ('S n1) header point tip m a)) -> ClientStNext n1 header point tip m a -> ClientPipelinedStIdle ('S n1) header point tip m a

SendMsgDone :: forall a header point tip (m :: Type -> Type). a -> ClientPipelinedStIdle 'Z header point tip m a

data ClientStNext (n :: N) header point tip (m :: Type -> Type) a

Callback for responses received after sending MsgRequestNext.

We could receive MsgAwaitReply. In this case we will wait for the next message which must be MsgRollForward or MsgRollBackward; thus we need only the two callbacks.

Constructors

ClientStNext

Fields

recvMsgRollForward :: header -> tip -> m (ClientPipelinedStIdle n header point tip m a)

Callback for MsgRollForward message.

recvMsgRollBackward :: point -> tip -> m (ClientPipelinedStIdle n header point tip m a)

Callback for MsgRollBackward message.

data ClientPipelinedStIntersect header point tip (m :: Type -> Type) a

Callbacks for messages received after sending MsgFindIntersect.

We might receive either MsgIntersectFound or MsgIntersectNotFound.

Constructors

ClientPipelinedStIntersect

Fields

recvMsgIntersectFound :: point -> tip -> m (ClientPipelinedStIdle 'Z header point tip m a)

recvMsgIntersectNotFound :: tip -> m (ClientPipelinedStIdle 'Z header point tip m a)

data ChainSyncInstruction header point tip

Data received through pipelining: either roll forward or roll backward instruction. If the server replied with MsgAwaitReply the pipelined receiver will await for the next message which must come with an instruction how to update our chain.

Note: internal API, not exposed by this module.

Constructors

RollForward !header !tip

RollBackward !point !tip

Implementation Helpers

It's generally idiomatic to use these functions to implement your pipelined client. It aids in deciding when to make pipelined requests vs process received responses.

data PipelineDecision (n :: N) where

Pipeline decision: we can do either one of these:

non-pipelined request

pipeline a request

collect or pipeline, but only when there are pipelined requests

collect, as above, only when there are pipelined requests

There might be other useful pipelining scenarios: collect a given number of requests (which also can be used to collect all outstanding requests).

Constructors

Request :: PipelineDecision 'Z

Pipeline :: forall (n :: N). PipelineDecision n

CollectOrPipeline :: forall (n1 :: N). PipelineDecision ('S n1)

Collect :: forall (n1 :: N). PipelineDecision ('S n1)

data MkPipelineDecision where

The callback gets the following arguments:

how many requests are not yet collected (in flight or already queued)

block number of client's tip

block number of server's tip

Client's tip block number and server's tip block number can only be equal (from the client's perspective) when both the client's and the server's tip headers agree. If they would not agree (server forked), then the server sends MsgRollBackward, which rolls back one block and causes the client's tip and the server's tip to differ.

In this module we implement three pipelining strategies:

pipelineDecisionMax

pipelineDecisionMin

pipelineDecisionLowHighMark

Constructors

MkPipelineDecision :: (forall (n :: N). Nat n -> WithOrigin BlockNo -> WithOrigin BlockNo -> (PipelineDecision n, MkPipelineDecision)) -> MkPipelineDecision

runPipelineDecision :: forall (n :: N). MkPipelineDecision -> Nat n -> WithOrigin BlockNo -> WithOrigin BlockNo -> (PipelineDecision n, MkPipelineDecision)

constantPipelineDecision :: (forall (n :: N). Nat n -> WithOrigin BlockNo -> WithOrigin BlockNo -> PipelineDecision n) -> MkPipelineDecision

pipelineDecisionMax :: forall (n :: N). Word16 -> Nat n -> WithOrigin BlockNo -> WithOrigin BlockNo -> PipelineDecision n

Present maximal pipelining of at most omax requests. Collect responses either when we are at the same block number as the server or when we sent more than omax requests.

If omax = 3 this pipelining strategy will generate a sequence: Pipeline Pipeline Pipeline Collect Pipeline Collect .... Pipeline Collect Collect Collect

pipelineDecisionMin :: forall (n :: N). Word16 -> Nat n -> WithOrigin BlockNo -> WithOrigin BlockNo -> PipelineDecision n

Present minimum pipelining of at most omax requests, collect responses eagerly.

pipelineDecisionLowHighMark :: Word16 -> Word16 -> MkPipelineDecision

Pipelining strategy which pipelines up to highMark requests; if the number of pipelined messages exceeds the high mark, it collects messages until there are at most lowMark outstanding requests.

Type level natural numbers

data N

Constructors

Z

S N

data Nat (n :: N) where

Bundled Patterns

pattern Succ :: forall m n. () => m ~ 'S n => Nat n -> Nat m

pattern Zero :: () => 'Z ~ n => Nat n

natToInt :: forall (n :: N). Nat n -> Int

Utilities

mapChainSyncClientPipelined :: forall header header' point point' tip tip' (m :: Type -> Type) a. Functor m => (point -> point') -> (point' -> point) -> (header' -> header) -> (tip' -> tip) -> ChainSyncClientPipelined header point tip m a -> ChainSyncClientPipelined header' point' tip' m a

Transform a ChainSyncClientPipelined by mapping over the tx header and the chain tip values.

Note the direction of the individual mapping functions corresponds to whether the types are used as protocol inputs or outputs (or both, as is the case for points).