Cardano.Api.Compatible.Certificate

Types that vary across the hardfork boundary

type family Delegatee era where ...

Equations

Delegatee DijkstraEra = Delegatee

Delegatee ConwayEra = Delegatee

Delegatee BabbageEra = Hash StakePoolKey

Delegatee AlonzoEra = Hash StakePoolKey

Delegatee MaryEra = Hash StakePoolKey

Delegatee AllegraEra = Hash StakePoolKey

Delegatee ShelleyEra = Hash StakePoolKey

Registering stake address and delegating

makeStakeAddressDelegationCertificate :: IsShelleyBasedEra era => StakeCredential -> Delegatee era -> Certificate (ShelleyLedgerEra era)

makeStakeAddressRegistrationCertificate :: IsShelleyBasedEra era => StakeRegistrationRequirements era -> Certificate (ShelleyLedgerEra era)

makeStakeAddressUnregistrationCertificate :: IsShelleyBasedEra era => StakeCredential -> Certificate (ShelleyLedgerEra era)

data StakeCredentialAndDeposit

Constructors

StakeCredentialAndDeposit StakeCredential Coin

type family StakeRegistrationRequirements era where ...

Equations

StakeRegistrationRequirements DijkstraEra = StakeCredentialAndDeposit

StakeRegistrationRequirements ConwayEra = StakeCredentialAndDeposit

StakeRegistrationRequirements BabbageEra = StakeCredential

StakeRegistrationRequirements AlonzoEra = StakeCredential

StakeRegistrationRequirements MaryEra = StakeCredential

StakeRegistrationRequirements AllegraEra = StakeCredential

StakeRegistrationRequirements ShelleyEra = StakeCredential

Registering stake pools

makeStakePoolRegistrationCertificate :: IsShelleyBasedEra era => PoolParams -> Certificate (ShelleyLedgerEra era)

makeStakePoolRetirementCertificate :: IsShelleyBasedEra era => Hash StakePoolKey -> EpochNo -> Certificate (ShelleyLedgerEra era)

Special certificates

makeMIRCertificate :: MIRPot -> MIRTarget -> Certificate (ShelleyLedgerEra BabbageEra)

makeGenesisKeyDelegationCertificate :: Hash GenesisKey -> Hash GenesisDelegateKey -> Hash VrfKey -> Certificate (ShelleyLedgerEra BabbageEra)

data MIRTarget

MIRTarget specifies if funds from either the reserves or the treasury are to be handed out to a collection of reward accounts or instead transfered to the opposite pot.

Constructors

StakeAddressesMIR !(Map (Credential 'Staking) DeltaCoin)

SendToOppositePotMIR !Coin

data MIRPot

Constructors

ReservesMIR

TreasuryMIR

selectStakeCredentialWitness :: IsShelleyBasedEra era => Certificate (ShelleyLedgerEra era) -> Maybe StakeCredential

Get the stake credential witness for a certificate that requires it. Only stake address deregistration and delegation requires witnessing (witness can be script or key).

Internal

getTxCertWitness :: ShelleyBasedEra era -> TxCert (ShelleyLedgerEra era) -> Maybe StakeCredential