Cardano.Api.Compatible.ProtocolParametersUpdate

Description

Era-based protocol parameter updates, and the era-sliced record types that back them.

Documentation

data EraBasedProtocolParametersUpdate era where

Each constructor corresponds to the set of protocol parameters available in a given era.

Constructors

ShelleyEraBasedProtocolParametersUpdate :: CommonProtocolParametersUpdate -> DeprecatedAfterMaryPParams ShelleyEra -> DeprecatedAfterBabbagePParams ShelleyEra -> ShelleyToAlonzoPParams ShelleyEra -> EraBasedProtocolParametersUpdate ShelleyEra

AllegraEraBasedProtocolParametersUpdate :: CommonProtocolParametersUpdate -> DeprecatedAfterMaryPParams AllegraEra -> ShelleyToAlonzoPParams AllegraEra -> DeprecatedAfterBabbagePParams ShelleyEra -> EraBasedProtocolParametersUpdate AllegraEra

MaryEraBasedProtocolParametersUpdate :: CommonProtocolParametersUpdate -> DeprecatedAfterMaryPParams MaryEra -> ShelleyToAlonzoPParams MaryEra -> DeprecatedAfterBabbagePParams ShelleyEra -> EraBasedProtocolParametersUpdate MaryEra

AlonzoEraBasedProtocolParametersUpdate :: CommonProtocolParametersUpdate -> ShelleyToAlonzoPParams AlonzoEra -> AlonzoOnwardsPParams AlonzoEra -> DeprecatedAfterBabbagePParams ShelleyEra -> EraBasedProtocolParametersUpdate AlonzoEra

BabbageEraBasedProtocolParametersUpdate :: CommonProtocolParametersUpdate -> AlonzoOnwardsPParams BabbageEra -> DeprecatedAfterBabbagePParams ShelleyEra -> IntroducedInBabbagePParams BabbageEra -> EraBasedProtocolParametersUpdate BabbageEra

ConwayEraBasedProtocolParametersUpdate :: CommonProtocolParametersUpdate -> AlonzoOnwardsPParams ConwayEra -> IntroducedInBabbagePParams ConwayEra -> IntroducedInConwayPParams (ShelleyLedgerEra ConwayEra) -> EraBasedProtocolParametersUpdate ConwayEra

DijkstraEraBasedProtocolParametersUpdate :: CommonProtocolParametersUpdate -> AlonzoOnwardsPParams DijkstraEra -> IntroducedInBabbagePParams DijkstraEra -> IntroducedInConwayPParams (ShelleyLedgerEra DijkstraEra) -> IntroducedInDijkstraPParams (ShelleyLedgerEra DijkstraEra) -> EraBasedProtocolParametersUpdate DijkstraEra

data AlonzoOnwardsPParams ledgerera

Constructors

AlonzoOnwardsPParams

Fields

alCostModels :: StrictMaybe CostModels

alPrices :: StrictMaybe Prices

alMaxTxExUnits :: StrictMaybe ExUnits

alMaxBlockExUnits :: StrictMaybe ExUnits

alMaxValSize :: StrictMaybe Word32

alCollateralPercentage :: StrictMaybe Word16

alMaxCollateralInputs :: StrictMaybe Word16

data CommonProtocolParametersUpdate

Protocol parameters common to each era. This can only ever be reduced if parameters are deprecated.

Constructors

CommonProtocolParametersUpdate

Fields

cppTxFeePerByteL :: StrictMaybe CoinPerByte

cppTxFeeFixedL :: StrictMaybe Coin

cppMaxBlockBodySize :: StrictMaybe Word32

cppMaxTxSize :: StrictMaybe Word32

cppMaxBlockHeaderSize :: StrictMaybe Word16

cppKeyDeposit :: StrictMaybe Coin

cppPoolDeposit :: StrictMaybe Coin

cppPoolRetireMaxEpoch :: StrictMaybe EpochInterval

cppStakePoolTargetNum :: StrictMaybe Word16

cppPoolPledgeInfluence :: StrictMaybe NonNegativeInterval

cppTreasuryExpansion :: StrictMaybe UnitInterval

cppMonetaryExpansion :: StrictMaybe UnitInterval

cppMinPoolCost :: StrictMaybe Coin

newtype DeprecatedAfterBabbagePParams ledgerera

Constructors

DeprecatedAfterBabbagePParams (StrictMaybe ProtVer)

newtype DeprecatedAfterMaryPParams ledgerera

Constructors

DeprecatedAfterMaryPParams (StrictMaybe Coin)

data ShelleyToAlonzoPParams ledgerera

Constructors

ShelleyToAlonzoPParams

Fields

(StrictMaybe Nonce)

Extra entropy

(StrictMaybe UnitInterval)

Decentralization parameter

newtype IntroducedInBabbagePParams era

Constructors

IntroducedInBabbagePParams (StrictMaybe CoinPerByte)
Coins per UTxO byte

data IntroducedInConwayPParams era

Constructors

IntroducedInConwayPParams

Fields

icPoolVotingThresholds :: StrictMaybe PoolVotingThresholds

icDRepVotingThresholds :: StrictMaybe DRepVotingThresholds

icMinCommitteeSize :: StrictMaybe Word16

icCommitteeTermLength :: StrictMaybe EpochInterval

icGovActionLifetime :: StrictMaybe EpochInterval

icGovActionDeposit :: StrictMaybe Coin

icDRepDeposit :: StrictMaybe Coin

icDRepActivity :: StrictMaybe EpochInterval

icMinFeeRefScriptCostPerByte :: StrictMaybe NonNegativeInterval

data IntroducedInDijkstraPParams era

Constructors

IntroducedInDijkstraPParams

Fields

idMaxRefScriptSizePerBlock :: StrictMaybe Word32

idMaxRefScriptSizePerTx :: StrictMaybe Word32

idRefScriptCostStride :: StrictMaybe (NonZero Word32)

idRefScriptCostMultiplier :: StrictMaybe PositiveInterval

createEraBasedProtocolParamUpdate :: ShelleyBasedEra era -> EraBasedProtocolParametersUpdate era -> PParamsUpdate (ShelleyLedgerEra era)

createPParams :: ShelleyBasedEra era -> EraBasedProtocolParametersUpdate era -> PParams (ShelleyLedgerEra era)

Update proposals to change the protocol parameters

data UpdateProposal era

Constructors

UpdateProposal !(Map (Hash GenesisKey) (EraBasedProtocolParametersUpdate era)) !EpochNo

makeShelleyUpdateProposal :: EraBasedProtocolParametersUpdate era -> [Hash GenesisKey] -> EpochNo -> UpdateProposal era

Internal conversion functions

toLedgerUpdate :: ShelleyBasedEra era -> UpdateProposal era -> Update (ShelleyLedgerEra era)

fromLedgerUpdate :: ShelleyLedgerEra era ~ ledgerera => ShelleyBasedEra era -> Update ledgerera -> UpdateProposal era

toLedgerProposedPPUpdates :: ShelleyBasedEra era -> Map (Hash GenesisKey) (EraBasedProtocolParametersUpdate era) -> ProposedPPUpdates (ShelleyLedgerEra era)

fromLedgerProposedPPUpdates :: ShelleyLedgerEra era ~ ledgerera => ShelleyBasedEra era -> ProposedPPUpdates ledgerera -> Map (Hash GenesisKey) (EraBasedProtocolParametersUpdate era)

fromLedgerPParamsUpdate :: ShelleyBasedEra era -> PParamsUpdate (ShelleyLedgerEra era) -> EraBasedProtocolParametersUpdate era

Data family instances

data family AsType t

A family of singleton types used in this API to indicate which type to use where it would otherwise be ambiguous or merely unclear.

Values of this type are passed to deserialisation functions for example.