Cardano.Api.Convenience.Query

Description

Convenience query functions

Documentation

data QueryConvenienceError

Constructors

AcqFailure AcquiringFailure

QueryEraMismatch EraMismatch

ByronEraNotSupported

QceUnsupportedNtcVersion !UnsupportedNtcVersionError

QceUnexpectedException !SomeException

newtype TxCurrentTreasuryValue

Constructors

TxCurrentTreasuryValue

Fields

unTxCurrentTreasuryValue :: Coin

determineEra :: LocalNodeConnectInfo -> ExceptT AcquiringFailure IO AnyCardanoEra

Query the node to determine which era it is in.

Simplest query related

executeQueryCardanoMode :: SocketPath -> NetworkId -> QueryInMode (Either EraMismatch result) -> ExceptT QueryConvenienceError IO result

Execute a query against the local node. The local node must be in CardanoMode.

executeQueryAnyMode :: LocalNodeConnectInfo -> QueryInMode (Either EraMismatch result) -> ExceptT QueryConvenienceError IO result

Execute a query against the local node in any mode.

queryStateForBalancedTx :: CardanoEra era -> [TxIn] -> [Certificate era] -> LocalStateQueryExpr block point QueryInMode r IO (Either QueryConvenienceError (UTxO era, LedgerProtocolParameters era, EraHistory, SystemStart, Set PoolId, Map StakeCredential Coin, Map (Credential 'DRepRole StandardCrypto) Coin, Maybe (Featured ConwayEraOnwards era TxCurrentTreasuryValue)))

A convenience function to query the relevant information, from the local node, for Cardano.Api.Convenience.Construction.constructBalancedTx

renderQueryConvenienceError :: QueryConvenienceError -> Text