Cardano.Api.Eon.Convert

Documentation

class Convert (f :: a -> Type) (g :: a -> Type) where

The Convert class is aimed at exposing a single interface that lets us convert between eons. However this is generalizable to any injective relationship between types.

Methods

convert :: forall (era :: a). f era -> g era