Cardano.Api.Era

Description

Cardano eras, sometimes we have to distinguish them.

Eras

data ByronEra

A type used as a tag to distinguish the Byron era.

data ShelleyEra

A type used as a tag to distinguish the Shelley era.

data AllegraEra

A type used as a tag to distinguish the Allegra era.

data MaryEra

A type used as a tag to distinguish the Mary era.

data AlonzoEra

A type used as a tag to distinguish the Alonzo era.

data BabbageEra

A type used as a tag to distinguish the Babbage era.

data ConwayEra

A type used as a tag to distinguish the Conway era.

Eons

eraProtVerLow :: ShelleyBasedEra era -> Version

Lookup the lower major protocol version for the shelley based era. In other words this is the major protocol version that the era has started in.

data ShelleyBasedEra era where

While the Byron and Shelley eras are quite different, there are several eras that are based on Shelley with only minor differences. It is useful to be able to treat the Shelley-based eras in a mostly-uniform way.

Values of this type witness the fact that the era is Shelley-based. This can be used to constrain the era to being a Shelley-based on. It allows non-uniform handling making case distinctions on the constructor.

Constructors

ShelleyBasedEraShelley :: ShelleyBasedEra ShelleyEra

ShelleyBasedEraAllegra :: ShelleyBasedEra AllegraEra

ShelleyBasedEraMary :: ShelleyBasedEra MaryEra

ShelleyBasedEraAlonzo :: ShelleyBasedEra AlonzoEra

ShelleyBasedEraBabbage :: ShelleyBasedEra BabbageEra

ShelleyBasedEraConway :: ShelleyBasedEra ConwayEra

type family ShelleyLedgerEra era = (ledgerera :: Type) | ledgerera -> era where ...

A type family that connects our era type tags to equivalent type tags used in the Shelley ledger library.

This type mapping connect types from this API with types in the Shelley ledger library which allows writing conversion functions in a more generic way.

Equations

ShelleyLedgerEra ShelleyEra = ShelleyEra

ShelleyLedgerEra AllegraEra = AllegraEra

ShelleyLedgerEra MaryEra = MaryEra

ShelleyLedgerEra AlonzoEra = AlonzoEra

ShelleyLedgerEra BabbageEra = BabbageEra

ShelleyLedgerEra ConwayEra = ConwayEra

class IsCardanoEra era => IsShelleyBasedEra era where

The class of eras that are based on Shelley. This allows uniform handling of Shelley-based eras, but also non-uniform by making case distinctions on the ShelleyBasedEra constructors.

Methods

shelleyBasedEra :: ShelleyBasedEra era

data AnyShelleyBasedEra where

Constructors

AnyShelleyBasedEra :: forall era. Typeable era => ShelleyBasedEra era -> AnyShelleyBasedEra

data InAnyShelleyBasedEra (thing :: Type -> Type) where

This pairs up some era-dependent type with a ShelleyBasedEra value that tells us what era it is, but hides the era type. This is useful when the era is not statically known, for example when deserialising from a file.

Constructors

InAnyShelleyBasedEra :: forall era (thing :: Type -> Type). Typeable era => ShelleyBasedEra era -> thing era -> InAnyShelleyBasedEra thing

inAnyShelleyBasedEra :: ShelleyBasedEra era -> thing era -> InAnyShelleyBasedEra thing

inEonForShelleyBasedEra :: Eon eon => a -> (eon era -> a) -> ShelleyBasedEra era -> a

Determine the value to use for a feature in a given ShelleyBasedEra.

inEonForShelleyBasedEraMaybe :: Eon eon => (eon era -> a) -> ShelleyBasedEra era -> Maybe a

forShelleyBasedEraInEon :: Eon eon => ShelleyBasedEra era -> a -> (eon era -> a) -> a

forShelleyBasedEraInEonMaybe :: Eon eon => ShelleyBasedEra era -> (eon era -> a) -> Maybe a

forShelleyBasedEraMaybeEon :: Eon eon => ShelleyBasedEra era -> Maybe (eon era)

requireShelleyBasedEra :: Applicative m => CardanoEra era -> m (Maybe (ShelleyBasedEra era))

type ShelleyBasedEraConstraints era = (HashAlgorithm HASH, Signable (VRF StandardCrypto) Seed, PraosProtocolSupportsNode (ConsensusProtocol era), ShelleyBlock (ConsensusProtocol era) (ShelleyLedgerEra era) ~ ConsensusBlockForEra era, ShelleyCompatible (ConsensusProtocol era) (ShelleyLedgerEra era), FromCBOR (ChainDepState (ConsensusProtocol era)), IsCardanoEra era, IsShelleyBasedEra era, ADDRHASH ~ Blake2b_224, Era (ShelleyLedgerEra era), EraPParams (ShelleyLedgerEra era), EraTx (ShelleyLedgerEra era), EraTxBody (ShelleyLedgerEra era), EraTxOut (ShelleyLedgerEra era), EraUTxO (ShelleyLedgerEra era), EraTxWits (ShelleyLedgerEra era), HashAnnotated (TxBody (ShelleyLedgerEra era)) EraIndependentTxBody, ShelleyEraTxCert (ShelleyLedgerEra era), FromCBOR (ChainDepState (ConsensusProtocol era)), IsCardanoEra era, IsShelleyBasedEra era, ToJSON (ChainDepState (ConsensusProtocol era)), ToJSON (PredicateFailure (EraRule "LEDGER" (ShelleyLedgerEra era))), Typeable era, (era == ByronEra) ~ 'False)

shelleyBasedEraConstraints :: ShelleyBasedEra era -> (ShelleyBasedEraConstraints era => a) -> a

data AllegraEraOnwards era where

Constructors

AllegraEraOnwardsAllegra :: AllegraEraOnwards AllegraEra

AllegraEraOnwardsMary :: AllegraEraOnwards MaryEra

AllegraEraOnwardsAlonzo :: AllegraEraOnwards AlonzoEra

AllegraEraOnwardsBabbage :: AllegraEraOnwards BabbageEra

AllegraEraOnwardsConway :: AllegraEraOnwards ConwayEra

allegraEraOnwardsConstraints :: AllegraEraOnwards era -> (AllegraEraOnwardsConstraints era => a) -> a

allegraEraOnwardsToShelleyBasedEra :: AllegraEraOnwards era -> ShelleyBasedEra era

Deprecated: Use convert instead.

type AllegraEraOnwardsConstraints era = (HashAlgorithm HASH, Signable (VRF StandardCrypto) Seed, PraosProtocolSupportsNode (ConsensusProtocol era), ShelleyBlock (ConsensusProtocol era) (ShelleyLedgerEra era) ~ ConsensusBlockForEra era, ShelleyCompatible (ConsensusProtocol era) (ShelleyLedgerEra era), ADDRHASH ~ Blake2b_224, Era (ShelleyLedgerEra era), EraPParams (ShelleyLedgerEra era), EraTx (ShelleyLedgerEra era), EraTxBody (ShelleyLedgerEra era), EraTxOut (ShelleyLedgerEra era), HashAnnotated (TxBody (ShelleyLedgerEra era)) EraIndependentTxBody, AllegraEraTxBody (ShelleyLedgerEra era), ShelleyEraTxCert (ShelleyLedgerEra era), FromCBOR (ChainDepState (ConsensusProtocol era)), FromCBOR (DebugLedgerState era), IsCardanoEra era, IsShelleyBasedEra era, ToJSON (ChainDepState (ConsensusProtocol era)), ToJSON (DebugLedgerState era), Typeable era, (era == ByronEra) ~ 'False)

class IsShelleyBasedEra era => IsAllegraBasedEra era where

Methods

allegraBasedEra :: AllegraEraOnwards era

data BabbageEraOnwards era where

Constructors

BabbageEraOnwardsBabbage :: BabbageEraOnwards BabbageEra

BabbageEraOnwardsConway :: BabbageEraOnwards ConwayEra

babbageEraOnwardsConstraints :: BabbageEraOnwards era -> (BabbageEraOnwardsConstraints era => a) -> a

babbageEraOnwardsToShelleyBasedEra :: BabbageEraOnwards era -> ShelleyBasedEra era

Deprecated: Use convert instead.

type BabbageEraOnwardsConstraints era = (HashAlgorithm HASH, Signable (VRF StandardCrypto) Seed, PraosProtocolSupportsNode (ConsensusProtocol era), ShelleyBlock (ConsensusProtocol era) (ShelleyLedgerEra era) ~ ConsensusBlockForEra era, ShelleyCompatible (ConsensusProtocol era) (ShelleyLedgerEra era), ADDRHASH ~ Blake2b_224, AlonzoEraTxOut (ShelleyLedgerEra era), BabbageEraPParams (ShelleyLedgerEra era), BabbageEraTxBody (ShelleyLedgerEra era), BabbageEraTxOut (ShelleyLedgerEra era), Era (ShelleyLedgerEra era), EraPParams (ShelleyLedgerEra era), EraTx (ShelleyLedgerEra era), EraTxBody (ShelleyLedgerEra era), EraTxOut (ShelleyLedgerEra era), EraUTxO (ShelleyLedgerEra era), HashAnnotated (TxBody (ShelleyLedgerEra era)) EraIndependentTxBody, MaryEraTxBody (ShelleyLedgerEra era), Script (ShelleyLedgerEra era) ~ AlonzoScript (ShelleyLedgerEra era), ScriptsNeeded (ShelleyLedgerEra era) ~ AlonzoScriptsNeeded (ShelleyLedgerEra era), ShelleyEraTxCert (ShelleyLedgerEra era), TxOut (ShelleyLedgerEra era) ~ BabbageTxOut (ShelleyLedgerEra era), Value (ShelleyLedgerEra era) ~ MaryValue, FromCBOR (ChainDepState (ConsensusProtocol era)), FromCBOR (DebugLedgerState era), IsCardanoEra era, IsShelleyBasedEra era, ToJSON (ChainDepState (ConsensusProtocol era)), ToJSON (DebugLedgerState era), Typeable era, (era == ByronEra) ~ 'False)

class IsAlonzoBasedEra era => IsBabbageBasedEra era where

Methods

babbageBasedEra :: BabbageEraOnwards era

data ByronToAlonzoEra era where

Constructors

ByronToAlonzoEraByron :: ByronToAlonzoEra ByronEra

ByronToAlonzoEraShelley :: ByronToAlonzoEra ShelleyEra

ByronToAlonzoEraAllegra :: ByronToAlonzoEra AllegraEra

ByronToAlonzoEraMary :: ByronToAlonzoEra MaryEra

ByronToAlonzoEraAlonzo :: ByronToAlonzoEra AlonzoEra

byronToAlonzoEraConstraints :: ByronToAlonzoEra era -> (ByronToAlonzoEraConstraints era => a) -> a

type ByronToAlonzoEraConstraints era = (IsCardanoEra era, Typeable era)

data MaryEraOnwards era where

Constructors

MaryEraOnwardsMary :: MaryEraOnwards MaryEra

MaryEraOnwardsAlonzo :: MaryEraOnwards AlonzoEra

MaryEraOnwardsBabbage :: MaryEraOnwards BabbageEra

MaryEraOnwardsConway :: MaryEraOnwards ConwayEra

maryEraOnwardsConstraints :: MaryEraOnwards era -> (MaryEraOnwardsConstraints era => a) -> a

maryEraOnwardsToShelleyBasedEra :: MaryEraOnwards era -> ShelleyBasedEra era

Deprecated: Use convert instead.

type MaryEraOnwardsConstraints era = (HashAlgorithm HASH, Signable (VRF StandardCrypto) Seed, PraosProtocolSupportsNode (ConsensusProtocol era), ShelleyBlock (ConsensusProtocol era) (ShelleyLedgerEra era) ~ ConsensusBlockForEra era, ShelleyCompatible (ConsensusProtocol era) (ShelleyLedgerEra era), ADDRHASH ~ Blake2b_224, Era (ShelleyLedgerEra era), EraPParams (ShelleyLedgerEra era), EraTx (ShelleyLedgerEra era), EraTxBody (ShelleyLedgerEra era), EraTxOut (ShelleyLedgerEra era), EraUTxO (ShelleyLedgerEra era), HashAnnotated (TxBody (ShelleyLedgerEra era)) EraIndependentTxBody, MaryEraTxBody (ShelleyLedgerEra era), ShelleyEraTxCert (ShelleyLedgerEra era), Value (ShelleyLedgerEra era) ~ MaryValue, FromCBOR (ChainDepState (ConsensusProtocol era)), FromCBOR (DebugLedgerState era), IsCardanoEra era, IsShelleyBasedEra era, ToJSON (ChainDepState (ConsensusProtocol era)), ToJSON (DebugLedgerState era), Typeable era, (era == ByronEra) ~ 'False)

class IsAllegraBasedEra era => IsMaryBasedEra era where

Methods

maryBasedEra :: MaryEraOnwards era

data ShelleyEraOnly era where

Constructors

ShelleyEraOnlyShelley :: ShelleyEraOnly ShelleyEra

shelleyEraOnlyConstraints :: ShelleyEraOnly era -> (ShelleyEraOnlyConstraints era => a) -> a

shelleyEraOnlyToShelleyBasedEra :: ShelleyEraOnly era -> ShelleyBasedEra era

Deprecated: Use convert instead.

type ShelleyEraOnlyConstraints era = (HashAlgorithm HASH, Signable (VRF StandardCrypto) Seed, PraosProtocolSupportsNode (ConsensusProtocol era), ShelleyBlock (ConsensusProtocol era) (ShelleyLedgerEra era) ~ ConsensusBlockForEra era, ShelleyCompatible (ConsensusProtocol era) (ShelleyLedgerEra era), ADDRHASH ~ Blake2b_224, Era (ShelleyLedgerEra era), EraPParams (ShelleyLedgerEra era), EraTx (ShelleyLedgerEra era), EraTxBody (ShelleyLedgerEra era), EraTxOut (ShelleyLedgerEra era), ExactEra ShelleyEra (ShelleyLedgerEra era), ExactEra ShelleyEra (ShelleyLedgerEra era), HashAnnotated (TxBody (ShelleyLedgerEra era)) EraIndependentTxBody, ProtVerAtMost (ShelleyLedgerEra era) 2, ProtVerAtMost (ShelleyLedgerEra era) 6, ProtVerAtMost (ShelleyLedgerEra era) 8, ShelleyEraTxBody (ShelleyLedgerEra era), ShelleyEraTxCert (ShelleyLedgerEra era), TxCert (ShelleyLedgerEra era) ~ ShelleyTxCert (ShelleyLedgerEra era), Value (ShelleyLedgerEra era) ~ Coin, FromCBOR (ChainDepState (ConsensusProtocol era)), FromCBOR (DebugLedgerState era), IsCardanoEra era, IsShelleyBasedEra era, ToJSON (ChainDepState (ConsensusProtocol era)), ToJSON (DebugLedgerState era), Typeable era, (era == ByronEra) ~ 'False)

data ShelleyToAllegraEra era where

Constructors

ShelleyToAllegraEraShelley :: ShelleyToAllegraEra ShelleyEra

ShelleyToAllegraEraAllegra :: ShelleyToAllegraEra AllegraEra

shelleyToAllegraEraConstraints :: ShelleyToAllegraEra era -> (ShelleyToAllegraEraConstraints era => a) -> a

shelleyToAllegraEraToShelleyBasedEra :: ShelleyToAllegraEra era -> ShelleyBasedEra era

Deprecated: Use convert instead.

type ShelleyToAllegraEraConstraints era = (HashAlgorithm HASH, Signable (VRF StandardCrypto) Seed, PraosProtocolSupportsNode (ConsensusProtocol era), ShelleyBlock (ConsensusProtocol era) (ShelleyLedgerEra era) ~ ConsensusBlockForEra era, ShelleyCompatible (ConsensusProtocol era) (ShelleyLedgerEra era), ADDRHASH ~ Blake2b_224, Era (ShelleyLedgerEra era), EraPParams (ShelleyLedgerEra era), EraTx (ShelleyLedgerEra era), EraTxBody (ShelleyLedgerEra era), EraTxOut (ShelleyLedgerEra era), EraUTxO (ShelleyLedgerEra era), HashAnnotated (TxBody (ShelleyLedgerEra era)) EraIndependentTxBody, ProtVerAtMost (ShelleyLedgerEra era) 4, ProtVerAtMost (ShelleyLedgerEra era) 6, ProtVerAtMost (ShelleyLedgerEra era) 8, ShelleyEraTxBody (ShelleyLedgerEra era), ShelleyEraTxCert (ShelleyLedgerEra era), TxCert (ShelleyLedgerEra era) ~ ShelleyTxCert (ShelleyLedgerEra era), Value (ShelleyLedgerEra era) ~ Coin, FromCBOR (ChainDepState (ConsensusProtocol era)), FromCBOR (DebugLedgerState era), IsCardanoEra era, IsShelleyBasedEra era, ToJSON (ChainDepState (ConsensusProtocol era)), ToJSON (DebugLedgerState era), Typeable era, (era == ByronEra) ~ 'False)

data ShelleyToAlonzoEra era where

Constructors

ShelleyToAlonzoEraShelley :: ShelleyToAlonzoEra ShelleyEra

ShelleyToAlonzoEraAllegra :: ShelleyToAlonzoEra AllegraEra

ShelleyToAlonzoEraMary :: ShelleyToAlonzoEra MaryEra

ShelleyToAlonzoEraAlonzo :: ShelleyToAlonzoEra AlonzoEra

shelleyToAlonzoEraConstraints :: ShelleyToAlonzoEra era -> (ShelleyToAlonzoEraConstraints era => a) -> a

shelleyToAlonzoEraToShelleyBasedEra :: ShelleyToAlonzoEra era -> ShelleyBasedEra era

type ShelleyToAlonzoEraConstraints era = (HashAlgorithm HASH, Signable (VRF StandardCrypto) Seed, PraosProtocolSupportsNode (ConsensusProtocol era), ShelleyBlock (ConsensusProtocol era) (ShelleyLedgerEra era) ~ ConsensusBlockForEra era, ShelleyCompatible (ConsensusProtocol era) (ShelleyLedgerEra era), ADDRHASH ~ Blake2b_224, Era (ShelleyLedgerEra era), EraPParams (ShelleyLedgerEra era), EraTx (ShelleyLedgerEra era), EraTxBody (ShelleyLedgerEra era), EraTxOut (ShelleyLedgerEra era), HashAnnotated (TxBody (ShelleyLedgerEra era)) EraIndependentTxBody, ProtVerAtMost (ShelleyLedgerEra era) 6, ProtVerAtMost (ShelleyLedgerEra era) 8, ShelleyEraTxBody (ShelleyLedgerEra era), ShelleyEraTxCert (ShelleyLedgerEra era), TxCert (ShelleyLedgerEra era) ~ ShelleyTxCert (ShelleyLedgerEra era), FromCBOR (ChainDepState (ConsensusProtocol era)), FromCBOR (DebugLedgerState era), IsCardanoEra era, IsShelleyBasedEra era, ToJSON (ChainDepState (ConsensusProtocol era)), ToJSON (DebugLedgerState era), Typeable era, (era == ByronEra) ~ 'False)

data ShelleyToBabbageEra era where

Constructors

ShelleyToBabbageEraShelley :: ShelleyToBabbageEra ShelleyEra

ShelleyToBabbageEraAllegra :: ShelleyToBabbageEra AllegraEra

ShelleyToBabbageEraMary :: ShelleyToBabbageEra MaryEra

ShelleyToBabbageEraAlonzo :: ShelleyToBabbageEra AlonzoEra

ShelleyToBabbageEraBabbage :: ShelleyToBabbageEra BabbageEra

shelleyToBabbageEraConstraints :: ShelleyToBabbageEra era -> (ShelleyToBabbageEraConstraints era => a) -> a

shelleyToBabbageEraToShelleyBasedEra :: ShelleyToBabbageEra era -> ShelleyBasedEra era

Deprecated: Use convert instead.

type ShelleyToBabbageEraConstraints era = (HashAlgorithm HASH, Signable (VRF StandardCrypto) Seed, PraosProtocolSupportsNode (ConsensusProtocol era), ShelleyBlock (ConsensusProtocol era) (ShelleyLedgerEra era) ~ ConsensusBlockForEra era, ShelleyCompatible (ConsensusProtocol era) (ShelleyLedgerEra era), ADDRHASH ~ Blake2b_224, Era (ShelleyLedgerEra era), EraPParams (ShelleyLedgerEra era), EraTx (ShelleyLedgerEra era), EraTxBody (ShelleyLedgerEra era), EraTxOut (ShelleyLedgerEra era), HashAnnotated (TxBody (ShelleyLedgerEra era)) EraIndependentTxBody, ProtVerAtMost (ShelleyLedgerEra era) 8, ShelleyEraTxBody (ShelleyLedgerEra era), ShelleyEraTxCert (ShelleyLedgerEra era), TxCert (ShelleyLedgerEra era) ~ ShelleyTxCert (ShelleyLedgerEra era), FromCBOR (ChainDepState (ConsensusProtocol era)), FromCBOR (DebugLedgerState era), IsCardanoEra era, IsShelleyBasedEra era, ToJSON (ChainDepState (ConsensusProtocol era)), ToJSON (DebugLedgerState era), Typeable era, (era == ByronEra) ~ 'False)

data ShelleyToMaryEra era where

Constructors

ShelleyToMaryEraShelley :: ShelleyToMaryEra ShelleyEra

ShelleyToMaryEraAllegra :: ShelleyToMaryEra AllegraEra

ShelleyToMaryEraMary :: ShelleyToMaryEra MaryEra

shelleyToMaryEraConstraints :: ShelleyToMaryEra era -> (ShelleyToMaryEraConstraints era => a) -> a

shelleyToMaryEraToShelleyBasedEra :: ShelleyToMaryEra era -> ShelleyBasedEra era

Deprecated: Use convert instead.

type ShelleyToMaryEraConstraints era = (HashAlgorithm HASH, Signable (VRF StandardCrypto) Seed, PraosProtocolSupportsNode (ConsensusProtocol era), ShelleyBlock (ConsensusProtocol era) (ShelleyLedgerEra era) ~ ConsensusBlockForEra era, ShelleyCompatible (ConsensusProtocol era) (ShelleyLedgerEra era), ADDRHASH ~ Blake2b_224, Era (ShelleyLedgerEra era), EraPParams (ShelleyLedgerEra era), EraTx (ShelleyLedgerEra era), EraTxBody (ShelleyLedgerEra era), EraTxOut (ShelleyLedgerEra era), HashAnnotated (TxBody (ShelleyLedgerEra era)) EraIndependentTxBody, ProtVerAtMost (ShelleyLedgerEra era) 4, ProtVerAtMost (ShelleyLedgerEra era) 6, ProtVerAtMost (ShelleyLedgerEra era) 8, ShelleyEraTxBody (ShelleyLedgerEra era), ShelleyEraTxCert (ShelleyLedgerEra era), TxCert (ShelleyLedgerEra era) ~ ShelleyTxCert (ShelleyLedgerEra era), FromCBOR (ChainDepState (ConsensusProtocol era)), FromCBOR (DebugLedgerState era), IsCardanoEra era, IsShelleyBasedEra era, ToJSON (ChainDepState (ConsensusProtocol era)), ToJSON (DebugLedgerState era), Typeable era, (era == ByronEra) ~ 'False)

data ConwayEraOnwards era where

Constructors

ConwayEraOnwardsConway :: ConwayEraOnwards ConwayEra

conwayEraOnwardsConstraints :: ConwayEraOnwards era -> (ConwayEraOnwardsConstraints era => a) -> a

conwayEraOnwardsToShelleyBasedEra :: ConwayEraOnwards era -> ShelleyBasedEra era

Deprecated: Use convert instead.

conwayEraOnwardsToBabbageEraOnwards :: ConwayEraOnwards era -> BabbageEraOnwards era

Deprecated: Use convert instead.

type ConwayEraOnwardsConstraints era = (HashAlgorithm HASH, Signable (VRF StandardCrypto) Seed, PraosProtocolSupportsNode (ConsensusProtocol era), ShelleyBlock (ConsensusProtocol era) (ShelleyLedgerEra era) ~ ConsensusBlockForEra era, ShelleyCompatible (ConsensusProtocol era) (ShelleyLedgerEra era), ADDRHASH ~ Blake2b_224, AlonzoEraTxOut (ShelleyLedgerEra era), AlonzoEraTxWits (ShelleyLedgerEra era), BabbageEraTxBody (ShelleyLedgerEra era), ConwayEraGov (ShelleyLedgerEra era), ConwayEraPParams (ShelleyLedgerEra era), ConwayEraTxBody (ShelleyLedgerEra era), ConwayEraTxCert (ShelleyLedgerEra era), Era (ShelleyLedgerEra era), EraGov (ShelleyLedgerEra era), EraPParams (ShelleyLedgerEra era), EraTx (ShelleyLedgerEra era), EraTxBody (ShelleyLedgerEra era), EraTxOut (ShelleyLedgerEra era), EraUTxO (ShelleyLedgerEra era), GovState (ShelleyLedgerEra era) ~ ConwayGovState (ShelleyLedgerEra era), HashAnnotated (TxBody (ShelleyLedgerEra era)) EraIndependentTxBody, MaryEraTxBody (ShelleyLedgerEra era), Script (ShelleyLedgerEra era) ~ AlonzoScript (ShelleyLedgerEra era), ScriptsNeeded (ShelleyLedgerEra era) ~ AlonzoScriptsNeeded (ShelleyLedgerEra era), ShelleyEraTxCert (ShelleyLedgerEra era), TxCert (ShelleyLedgerEra era) ~ ConwayTxCert (ShelleyLedgerEra era), Value (ShelleyLedgerEra era) ~ MaryValue, FromCBOR (ChainDepState (ConsensusProtocol era)), FromCBOR (DebugLedgerState era), IsCardanoEra era, IsShelleyBasedEra era, ToJSON (ChainDepState (ConsensusProtocol era)), ToJSON (DebugLedgerState era), Typeable era, (era == ByronEra) ~ 'False)

class IsBabbageBasedEra era => IsConwayBasedEra era where

Methods

conwayBasedEra :: ConwayEraOnwards era

data AlonzoEraOnwards era where

Constructors

AlonzoEraOnwardsAlonzo :: AlonzoEraOnwards AlonzoEra

AlonzoEraOnwardsBabbage :: AlonzoEraOnwards BabbageEra

AlonzoEraOnwardsConway :: AlonzoEraOnwards ConwayEra

alonzoEraOnwardsConstraints :: AlonzoEraOnwards era -> (AlonzoEraOnwardsConstraints era => a) -> a

alonzoEraOnwardsToShelleyBasedEra :: AlonzoEraOnwards era -> ShelleyBasedEra era

Deprecated: Use convert instead.

type AlonzoEraOnwardsConstraints era = (HashAlgorithm HASH, Signable (VRF StandardCrypto) Seed, PraosProtocolSupportsNode (ConsensusProtocol era), ShelleyBlock (ConsensusProtocol era) (ShelleyLedgerEra era) ~ ConsensusBlockForEra era, ShelleyCompatible (ConsensusProtocol era) (ShelleyLedgerEra era), ADDRHASH ~ Blake2b_224, AlonzoEraPParams (ShelleyLedgerEra era), AlonzoEraTx (ShelleyLedgerEra era), AlonzoEraTxBody (ShelleyLedgerEra era), AlonzoEraTxOut (ShelleyLedgerEra era), AlonzoEraTxWits (ShelleyLedgerEra era), Era (ShelleyLedgerEra era), EraPParams (ShelleyLedgerEra era), EraTx (ShelleyLedgerEra era), EraTxBody (ShelleyLedgerEra era), EraTxOut (ShelleyLedgerEra era), EraUTxO (ShelleyLedgerEra era), HashAnnotated (TxBody (ShelleyLedgerEra era)) EraIndependentTxBody, MaryEraTxBody (ShelleyLedgerEra era), NativeScript (ShelleyLedgerEra era) ~ Timelock (ShelleyLedgerEra era), EraPlutusContext (ShelleyLedgerEra era), Script (ShelleyLedgerEra era) ~ AlonzoScript (ShelleyLedgerEra era), ScriptsNeeded (ShelleyLedgerEra era) ~ AlonzoScriptsNeeded (ShelleyLedgerEra era), ShelleyEraTxCert (ShelleyLedgerEra era), Value (ShelleyLedgerEra era) ~ MaryValue, FromCBOR (ChainDepState (ConsensusProtocol era)), FromCBOR (DebugLedgerState era), IsCardanoEra era, IsShelleyBasedEra era, ToJSON (ChainDepState (ConsensusProtocol era)), ToJSON (DebugLedgerState era), Typeable era, (era == ByronEra) ~ 'False)

class IsMaryBasedEra era => IsAlonzoBasedEra era where

Methods

alonzoBasedEra :: AlonzoEraOnwards era

CardanoEra

data CardanoEra era where

This GADT provides a value-level representation of all the Cardano eras. This enables pattern matching on the era to allow them to be treated in a non-uniform way.

This can be used in combination with the IsCardanoEra class to get access to this value.

In combination this can often enable code that handles all eras, and does so uniformly where possible, and non-uniformly where necessary.

Constructors

ByronEra :: CardanoEra ByronEra

ShelleyEra :: CardanoEra ShelleyEra

AllegraEra :: CardanoEra AllegraEra

MaryEra :: CardanoEra MaryEra

AlonzoEra :: CardanoEra AlonzoEra

BabbageEra :: CardanoEra BabbageEra

ConwayEra :: CardanoEra ConwayEra

class HasTypeProxy era => IsCardanoEra era where

The class of Cardano eras. This allows uniform handling of all Cardano eras, but also non-uniform by making case distinctions on the CardanoEra constructors.

Methods

cardanoEra :: CardanoEra era

data AnyCardanoEra where

Constructors

AnyCardanoEra :: forall era. Typeable era => CardanoEra era -> AnyCardanoEra

anyCardanoEra :: CardanoEra era -> AnyCardanoEra

Like the AnyCardanoEra constructor but does not demand a IsCardanoEra class constraint.

data InAnyCardanoEra (thing :: Type -> Type) where

This pairs up some era-dependent type with a CardanoEra value that tells us what era it is, but hides the era type. This is useful when the era is not statically known, for example when deserialising from a file.

Constructors

InAnyCardanoEra :: forall era (thing :: Type -> Type). Typeable era => CardanoEra era -> thing era -> InAnyCardanoEra thing

inAnyCardanoEra :: CardanoEra era -> thing era -> InAnyCardanoEra thing

cardanoEraConstraints :: CardanoEra era -> (CardanoEraConstraints era => a) -> a

type family CardanoLedgerEra era = (ledgerera :: Type) | ledgerera -> era where ...

A type family that connects our era type tags to equivalent type tags used in the ledger library.

This type mapping connect types from this API with types in the ledger library which allows writing conversion functions in a more generic way.

Equations

CardanoLedgerEra ByronEra = ByronEra

CardanoLedgerEra ShelleyEra = ShelleyEra

CardanoLedgerEra AllegraEra = AllegraEra

CardanoLedgerEra MaryEra = MaryEra

CardanoLedgerEra AlonzoEra = AlonzoEra

CardanoLedgerEra BabbageEra = BabbageEra

CardanoLedgerEra ConwayEra = ConwayEra

class ToCardanoEra (eon :: Type -> Type) where

Methods

toCardanoEra :: eon era -> CardanoEra era

IsEon

class Eon (eon :: Type -> Type) where

An Eon is a span of multiple eras. Eons are used to scope functionality to particular eras such that it isn't possible construct code that uses functionality that is outside of given eras.

Methods

inEonForEra

Arguments

:: a
Value to use if the eon does not include the era

-> (eon era -> a)
Function to get the value to use if the eon includes the era

-> CardanoEra era
Era to check

-> a
The value to use

Determine the value to use in an eon (a span of multiple eras). Note that the negative case is the first argument, and the positive case is the second as per the either function convention.

data EraInEon (eon :: Type -> Type) where

Constructors

EraInEon :: forall era (eon :: Type -> Type). (Typeable era, Typeable (eon era), Eon eon) => eon era -> EraInEon eon

inEonForEraMaybe

Arguments

:: Eon eon

=> (eon era -> a)
Function to get the value to use if the eon includes the era

-> CardanoEra era
Era to check

-> Maybe a
The value to use

forEraInEon

Arguments

:: Eon eon

=> CardanoEra era
Era to check

-> a
Value to use if the eon does not include the era

-> (eon era -> a)
Function to get the value to use if the eon includes the era

-> a
The value to use

forEraInEonMaybe

Arguments

:: Eon eon

=> CardanoEra era
Era to check

-> (eon era -> a)
Function to get the value to use if the eon includes the era

-> Maybe a
The value to use

forEraMaybeEon

Arguments

:: Eon eon

=> CardanoEra era
Era to check

-> Maybe (eon era)
The eon if supported in the era

maybeEon

Arguments

:: (Eon eon, IsCardanoEra era)

=> Maybe (eon era)
The eon if supported in the era

monoidForEraInEon :: (Eon eon, Monoid a) => CardanoEra era -> (eon era -> a) -> a

monoidForEraInEonA :: (Eon eon, Applicative f, Monoid a) => CardanoEra era -> (eon era -> f a) -> f a

class Convert (f :: a -> Type) (g :: a -> Type) where

The Convert class is aimed at exposing a single interface that lets us convert between eons. However this is generalizable to any injective relationship between types.

Methods

convert :: forall (era :: a). f era -> g era

class Inject t s where

Methods

inject :: t -> s

Era-dependend features

data Featured (eon :: Type -> Type) era a where

A value only if the eon includes era

Constructors

Featured

Fields

:: forall (eon :: Type -> Type) era a. eon era

The witness that the eon includes era

-> a

The value to use

-> Featured eon era a

mkFeatured

Arguments

:: forall (eon :: Type -> Type) era a. (IsCardanoEra era, Eon eon)

=> a
a value featured in eon

-> Maybe (Featured eon era a)
Just if era is in eon

Create a Featured with automatic witness conjuring

unFeatured :: forall (eon :: Type -> Type) era a. Featured eon era a -> a

asFeaturedInEra :: forall (eon :: Type -> Type) a era. Eon eon => a -> CardanoEra era -> Maybe (Featured eon era a)

Attempt to construct a FeatureValue from a value and era. If the eon is not supported in the era, then NoFeatureValue is returned.

asFeaturedInShelleyBasedEra :: forall (eon :: Type -> Type) a era. Eon eon => a -> ShelleyBasedEra era -> Maybe (Featured eon era a)

Attempt to construct a FeatureValue from a value and a shelley-based-era.

Data family instances

data family AsType t

A family of singleton types used in this API to indicate which type to use where it would otherwise be ambiguous or merely unclear.

Values of this type are passed to deserialisation functions for example.

Era case handling

Case on CardanoEra

caseByronOrShelleyBasedEra :: a -> (ShelleyBasedEraConstraints era => ShelleyBasedEra era -> a) -> CardanoEra era -> a

caseByronOrShelleyBasedEra f g era returns f in Byron and applies g to Shelley-based eras.

Case on ShelleyBasedEra

caseByronToAlonzoOrBabbageEraOnwards :: (ByronToAlonzoEraConstraints era => ByronToAlonzoEra era -> a) -> (BabbageEraOnwardsConstraints era => BabbageEraOnwards era -> a) -> CardanoEra era -> a

caseByronToAlonzoOrBabbageEraOnwards f g era applies f to byron, shelley, allegra, mary, and alonzo; and g to babbage and later eras.

caseShelleyToAllegraOrMaryEraOnwards :: (ShelleyToAllegraEraConstraints era => ShelleyToAllegraEra era -> a) -> (MaryEraOnwardsConstraints era => MaryEraOnwards era -> a) -> ShelleyBasedEra era -> a

caseShelleyToAllegraOrMaryEraOnwards f g era applies f to shelley and allegra; and applies g to mary and later eras.

caseShelleyToMaryOrAlonzoEraOnwards :: (ShelleyToMaryEraConstraints era => ShelleyToMaryEra era -> a) -> (AlonzoEraOnwardsConstraints era => AlonzoEraOnwards era -> a) -> ShelleyBasedEra era -> a

caseShelleyToMaryOrAlonzoEraOnwards f g era applies f to shelley, allegra, and mary; and applies g to alonzo and later eras.

caseShelleyToAlonzoOrBabbageEraOnwards :: (ShelleyToAlonzoEraConstraints era => ShelleyToAlonzoEra era -> a) -> (BabbageEraOnwardsConstraints era => BabbageEraOnwards era -> a) -> ShelleyBasedEra era -> a

caseShelleyToAlonzoOrBabbageEraOnwards f g era applies f to shelley, allegra, mary, and alonzo; and applies g to babbage and later eras.

caseShelleyToBabbageOrConwayEraOnwards :: (ShelleyToBabbageEraConstraints era => ShelleyToBabbageEra era -> a) -> (ConwayEraOnwardsConstraints era => ConwayEraOnwards era -> a) -> ShelleyBasedEra era -> a

caseShelleyToBabbageOrConwayEraOnwards f g era applies f to eras before conway; and applies g to conway and later eras.