Cardano.Api.Eras.Core

Description

Cardano eras, sometimes we have to distinguish them.

Eras

data ByronEra

A type used as a tag to distinguish the Byron era.

data ShelleyEra

A type used as a tag to distinguish the Shelley era.

data AllegraEra

A type used as a tag to distinguish the Allegra era.

data MaryEra

A type used as a tag to distinguish the Mary era.

data AlonzoEra

A type used as a tag to distinguish the Alonzo era.

data BabbageEra

A type used as a tag to distinguish the Babbage era.

data ConwayEra

A type used as a tag to distinguish the Conway era.

CardanoEra

data CardanoEra era where

This GADT provides a value-level representation of all the Cardano eras. This enables pattern matching on the era to allow them to be treated in a non-uniform way.

This can be used in combination with the IsCardanoEra class to get access to this value.

In combination this can often enable code that handles all eras, and does so uniformly where possible, and non-uniformly where necessary.

Constructors

ByronEra :: CardanoEra ByronEra

ShelleyEra :: CardanoEra ShelleyEra

AllegraEra :: CardanoEra AllegraEra

MaryEra :: CardanoEra MaryEra

AlonzoEra :: CardanoEra AlonzoEra

BabbageEra :: CardanoEra BabbageEra

ConwayEra :: CardanoEra ConwayEra

class HasTypeProxy era => IsCardanoEra era where

The class of Cardano eras. This allows uniform handling of all Cardano eras, but also non-uniform by making case distinctions on the CardanoEra constructors, or the CardanoEraStyle constructors via cardanoEraStyle.

Methods

cardanoEra :: CardanoEra era

data AnyCardanoEra where

Constructors

AnyCardanoEra :: Typeable era => CardanoEra era -> AnyCardanoEra

anyCardanoEra :: CardanoEra era -> AnyCardanoEra

Like the AnyCardanoEra constructor but does not demand a IsCardanoEra class constraint.

data InAnyCardanoEra thing where

This pairs up some era-dependent type with a CardanoEra value that tells us what era it is, but hides the era type. This is useful when the era is not statically known, for example when deserialising from a file.

Constructors

InAnyCardanoEra :: Typeable era => CardanoEra era -> thing era -> InAnyCardanoEra thing

inAnyCardanoEra :: CardanoEra era -> thing era -> InAnyCardanoEra thing

type family CardanoLedgerEra era = ledgerera | ledgerera -> era where ...

A type family that connects our era type tags to equivalent type tags used in the ledger library.

This type mapping connect types from this API with types in the ledger library which allows writing conversion functions in a more generic way.

Equations

CardanoLedgerEra ByronEra = ByronEra StandardCrypto

CardanoLedgerEra ShelleyEra = ShelleyEra StandardCrypto

CardanoLedgerEra AllegraEra = AllegraEra StandardCrypto

CardanoLedgerEra MaryEra = MaryEra StandardCrypto

CardanoLedgerEra AlonzoEra = AlonzoEra StandardCrypto

CardanoLedgerEra BabbageEra = BabbageEra StandardCrypto

CardanoLedgerEra ConwayEra = ConwayEra StandardCrypto

class ToCardanoEra (eon :: Type -> Type) where

Methods

toCardanoEra :: eon era -> CardanoEra era

IsEon

class Eon (eon :: Type -> Type) where

An Eon is a span of multiple eras. Eons are used to scope functionality to particular eras such that it isn't possible construct code that uses functionality that is outside of given eras.

Methods

inEonForEra

Arguments

:: a
Value to use if the eon does not include the era

-> (eon era -> a)
Function to get the value to use if the eon includes the era

-> CardanoEra era
Era to check

-> a
The value to use

Determine the value to use in an eon (a span of multiple eras). Note that the negative case is the first argument, and the positive case is the second as per the either function convention.

data EraInEon eon where

Constructors

EraInEon :: (Typeable era, Typeable (eon era), Eon eon) => eon era -> EraInEon eon

inEonForEraMaybe

Arguments

:: Eon eon

=> (eon era -> a)
Function to get the value to use if the eon includes the era

-> CardanoEra era
Era to check

-> Maybe a
The value to use

forEraInEon

Arguments

:: Eon eon

=> CardanoEra era
Era to check

-> a
Value to use if the eon does not include the era

-> (eon era -> a)
Function to get the value to use if the eon includes the era

-> a
The value to use

forEraInEonMaybe

Arguments

:: Eon eon

=> CardanoEra era
Era to check

-> (eon era -> a)
Function to get the value to use if the eon includes the era

-> Maybe a
The value to use

forEraMaybeEon

Arguments

:: Eon eon

=> CardanoEra era
Era to check

-> Maybe (eon era)
The eon if supported in the era

maybeEon

Arguments

:: Eon eon

=> IsCardanoEra era

=> Maybe (eon era)
The eon if supported in the era

monoidForEraInEon :: Eon eon => Monoid a => CardanoEra era -> (eon era -> a) -> a

monoidForEraInEonA :: Eon eon => Applicative f => Monoid a => CardanoEra era -> (eon era -> f a) -> f a

Data family instances

data family AsType t

A family of singleton types used in this API to indicate which type to use where it would otherwise be ambiguous or merely unclear.

Values of this type are passed to deserialisation functions for example.

type CardanoEraConstraints era = (Typeable era, IsCardanoEra era)

cardanoEraConstraints :: CardanoEra era -> (CardanoEraConstraints era => a) -> a