Cardano.Api.Error

Description

Class of errors used in the Api.

Documentation

class Error e where

Methods

prettyError :: e -> Doc ann

throwErrorM :: (HasCallStack, MonadThrow m, Typeable e, Error e) => e -> m a

Throw an 'Error e' as ErrorAsException. throwErrorM will attach a call stack to the exception.

liftEitherError :: (HasCallStack, MonadThrow m, Typeable e, Error e) => Either e a -> m a

Pretty print 'Error e' and throwM it wrapped in ErrorAsException when Left.

failEitherError :: (MonadFail m, Error e) => Either e a -> m a

Pretty print 'Error e' and fail if Left.

data ErrorAsException where

An exception wrapping any 'Error e', attaching a call stack from the construction place to it.

Constructors

ErrorAsException :: forall e. (HasCallStack, Typeable e, Error e) => e -> ErrorAsException

data FileError e

Constructors

FileError FilePath e

FileErrorTempFile

Fields

FilePath

Target path

FilePath

Temporary path

Handle

FileDoesNotExistError FilePath

FileIOError FilePath IOException

fileIOExceptT :: forall (m :: Type -> Type) s e. MonadIO m => FilePath -> (FilePath -> IO s) -> ExceptT (FileError e) m s

displayError :: Error a => a -> String

renderBuildable :: Buildable a => a -> Text