Cardano.Api.Experimental

Description

This module provides an experimental library interface that is intended to replace the existing api. It is subject to dramatic changes so use with caution.

Tx related

newtype UnsignedTx era

Constructors

UnsignedTx (Tx (LedgerEra era))

newtype UnsignedTxError

Constructors

UnsignedTxError TxBodyError

makeUnsignedTx :: Era era -> TxBodyContent BuildTx era -> Either TxBodyError (UnsignedTx era)

makeKeyWitness :: Era era -> UnsignedTx era -> ShelleyWitnessSigningKey -> WitVKey 'Witness StandardCrypto

signTx :: Era era -> [BootstrapWitness StandardCrypto] -> [WitVKey 'Witness StandardCrypto] -> UnsignedTx era -> Tx (LedgerEra era)

convertTxBodyToUnsignedTx :: HasCallStack => ShelleyBasedEra era -> TxBody era -> UnsignedTx era

type EraCommonConstraints era = (AlonzoEraTx (LedgerEra era), BabbageEraTxBody (LedgerEra era), EraTx (LedgerEra era), EraUTxO (LedgerEra era), EraCrypto (LedgerEra era) ~ StandardCrypto, ShelleyLedgerEra era ~ LedgerEra era, HashAnnotated (TxBody (LedgerEra era)) EraIndependentTxBody StandardCrypto, IsEra era)

obtainCommonConstraints :: Era era -> (EraCommonConstraints era => a) -> a

hashTxBody :: HashAnnotated (TxBody era) EraIndependentTxBody StandardCrypto => TxBody era -> Hash StandardCrypto EraIndependentTxBody

evaluateTransactionExecutionUnitsShelley :: ShelleyBasedEra era -> SystemStart -> LedgerEpochInfo -> LedgerProtocolParameters era -> UTxO era -> Tx (ShelleyLedgerEra era) -> Either (TransactionValidityError era) (Map ScriptWitnessIndex (Either ScriptExecutionError (EvalTxExecutionUnitsLog, ExecutionUnits)))

data BabbageEra

data ConwayEra

data Era era where

Constructors

BabbageEra :: Era BabbageEra

ConwayEra :: Era ConwayEra

class IsEra era where

Methods

useEra :: Era era

data Some (f :: k -> Type) where

Constructors

Some :: forall {k} (f :: k -> Type) (a :: k). (Typeable a, Typeable (f a)) => f a -> Some f

type family LedgerEra era = (r :: Type) | r -> era where ...

Equations

LedgerEra BabbageEra = Babbage

LedgerEra ConwayEra = Conway

newtype DeprecatedEra era

Constructors

DeprecatedEra (ShelleyBasedEra era)

eraToSbe :: Era era -> ShelleyBasedEra era

babbageEraOnwardsToEra :: BabbageEraOnwards era -> Era era

eraToBabbageEraOnwards :: Era era -> BabbageEraOnwards era

sbeToEra :: MonadError (DeprecatedEra era) m => ShelleyBasedEra era -> m (Era era)