Cardano.Api.Experimental.AnyScript

Documentation

data AnyScript era where

Constructors

AnySimpleScript :: forall era. SimpleScript era -> AnyScript era

AnyPlutusScript :: forall (lang :: Language) era. (PlutusLanguage lang, Typeable lang) => PlutusScriptInEra lang era -> AnyScript era

data family AsType t

A family of singleton types used in this API to indicate which type to use where it would otherwise be ambiguous or merely unclear.

Values of this type are passed to deserialisation functions for example.

deserialiseAnyPlutusScriptOfLanguage :: forall era (lang :: Language). (IsEra era, PlutusLanguage lang, HasTypeProxy (SLanguage lang)) => ByteString -> SLanguage lang -> Either DecoderError (AnyScript (LedgerEra era))

deserialiseAnySimpleScript :: IsEra era => ByteString -> Either DecoderError (AnyScript (LedgerEra era))

hashAnyScript :: IsEra era => AnyScript (LedgerEra era) -> ScriptHash