Cardano.Api.Experimental.Era

Description

This module defines the protocol versions corresponding to the eras in the Cardano blockchain.

Documentation

data BabbageEra

A type used as a tag to distinguish the Babbage era.

data ConwayEra

A type used as a tag to distinguish the Conway era.

data Era era where

Represents the latest Cardano blockchain eras, including the one currently on mainnet and the upcoming one.

After a hard fork takes place, the era on mainnet before the hard fork is deprecated and, after a deprecation period, removed from cardano-api. During the deprecation period, cardano-api users should update their codebase to the new mainnet era.

Constructors

ConwayEra :: Era ConwayEra
The currently active era on the Cardano mainnet.

class IsEra era where

Type class interface for the Era type.

Methods

useEra :: Era era

data Some (f :: k -> Type) where

An existential wrapper for types of kind k -> Type. It can hold any era, for example, Some Era. The era witness can be brought back into scope, for example, using this pattern:

anyEra = Some ConwayEra
-- then later in the code
Some era <- pure anyEra
obtainCommonConstraints era foo

Constructors

Some :: forall {k} (f :: k -> Type) (a :: k). (Typeable a, Typeable (f a)) => f a -> Some f

class Inject t s where

Methods

inject :: t -> s

type family LedgerEra era = (r :: Type) | r -> era where ...

Users typically interact with the latest features on the mainnet or experiment with features from the upcoming era. Therefore, protocol versions are limited to the current mainnet era and the next (upcoming) era.

Equations

LedgerEra ConwayEra = ConwayEra

newtype DeprecatedEra era

Constructors

DeprecatedEra (ShelleyBasedEra era)

type EraCommonConstraints era = (AllegraEraScript (LedgerEra era), AlonzoEraTx (LedgerEra era), BabbageEraPParams (LedgerEra era), BabbageEraTxBody (LedgerEra era), ConwayEraTxCert (LedgerEra era), TxCert (LedgerEra era) ~ ConwayTxCert (LedgerEra era), Era (LedgerEra era), EraScript (LedgerEra era), EraTx (LedgerEra era), EraTxCert (LedgerEra era), EraTxOut (LedgerEra era), EraUTxO (LedgerEra era), NativeScript (LedgerEra era) ~ Timelock (LedgerEra era), ShelleyEraTxCert (LedgerEra era), ShelleyLedgerEra era ~ LedgerEra era, HashAnnotated (TxBody (LedgerEra era)) EraIndependentTxBody, IsCardanoEra era, IsShelleyBasedEra era, IsEra era)

obtainCommonConstraints :: Era era -> (EraCommonConstraints era => a) -> a

eraToSbe :: Era era -> ShelleyBasedEra era

Deprecated: Use convert instead.

How to deprecate an era:

Add the DEPRECATED pragma to the era type tag and constructor at the same time:

{-# DEPRECATED BabbageEra "BabbageEra no longer supported, use ConwayEra" #-}
data BabbageEra

Update the Haddock documentation for the constructor of the deprecated era, mentioning the deprecation.

data Era era where
  {-# DEPRECATED BabbageEra "BabbageEra no longer supported, use ConwayEra" #-}
  BabbageEra :: Era BabbageEra
  -- | The era currently active on Cardano's mainnet.
  ConwayEra :: Era ConwayEra

Add a new IsEra instance and update the deprecated era instance to produce a compile-time error:

instance TypeError ('Text "IsEra BabbageEra: Deprecated. Update to ConwayEra") => IsEra BabbageEra where
  useEra = error "unreachable"

instance IsEra ConwayEra where
  useEra = ConwayEra

eraToBabbageEraOnwards :: Era era -> BabbageEraOnwards era

Deprecated: Use convert instead.

sbeToEra :: MonadError (DeprecatedEra era) m => ShelleyBasedEra era -> m (Era era)