Cardano.Api.Experimental.Eras

Description

This module defines the protocol versions corresponding to the eras in the Cardano blockchain.

Documentation

data BabbageEra

A type used as a tag to distinguish the Babbage era.

data ConwayEra

A type used as a tag to distinguish the Conway era.

data Era era where

Represents the eras in Cardano's blockchain. This type represents eras currently on mainnet and new eras which are in development.

After a hardfork, the era from which we hardfork from gets deprecated and after deprecation period, gets removed. During deprecation period, consumers of cardano-api should update their codebase to the mainnet era.

Constructors

BabbageEra :: Era BabbageEra
The era currently active on Cardano's mainnet.

ConwayEra :: Era ConwayEra
The upcoming era in development.

class IsEra era where

Type class interface for the Era type.

Methods

useEra :: Era era

data Some (f :: k -> Type) where

An existential type for singleton types. Use to hold any era e.g. Some Era. One can then bring the era witness back into scope for example using this pattern: anyEra = Some ConwayEra -- then later in the code Some era <- pure anyEra obtainCommonConstraints era foo

Constructors

Some :: forall {k} (f :: k -> Type) (a :: k). (Typeable a, Typeable (f a)) => f a -> Some f

type family LedgerEra era = (r :: Type) | r -> era where ...

Users typically interact with the latest features on the mainnet or experiment with features from the upcoming era. Hence, the protocol versions are limited to the current mainnet era and the next era (upcoming era).

Equations

LedgerEra BabbageEra = Babbage

LedgerEra ConwayEra = Conway

newtype DeprecatedEra era

Constructors

DeprecatedEra (ShelleyBasedEra era)

type EraCommonConstraints era = (AlonzoEraTx (LedgerEra era), BabbageEraTxBody (LedgerEra era), EraTx (LedgerEra era), EraUTxO (LedgerEra era), EraCrypto (LedgerEra era) ~ StandardCrypto, ShelleyLedgerEra era ~ LedgerEra era, HashAnnotated (TxBody (LedgerEra era)) EraIndependentTxBody StandardCrypto, IsEra era)

obtainCommonConstraints :: Era era -> (EraCommonConstraints era => a) -> a

eraToSbe :: Era era -> ShelleyBasedEra era

How to deprecate an era

Add DEPRECATED pragma to the era type tag and the era constructor at the same time: {-# DEPRECATED BabbageEra "BabbageEra no longer supported, use ConwayEra" #-} data BabbageEra

Update haddock for the constructor of the deprecated era, mentioning deprecation.

data Era era where
  {-# DEPRECATED BabbageEra "BabbageEra no longer supported, use ConwayEra" #-}
  BabbageEra :: Era BabbageEra
  -- | The era currently active on Cardano's mainnet.
  ConwayEra :: Era ConwayEra

Add new IsEra instance and update the deprecated era instance to produce a compile-time error: @ instance TypeError ('Text "IsEra BabbageEra: Deprecated. Update to ConwayEra") => IsEra BabbageEra where useEra = error "unreachable"

instance IsEra ConwayEra where useEra = ConwayEra @

babbageEraOnwardsToEra :: BabbageEraOnwards era -> Era era

eraToBabbageEraOnwards :: Era era -> BabbageEraOnwards era

sbeToEra :: MonadError (DeprecatedEra era) m => ShelleyBasedEra era -> m (Era era)