Cardano.Api.Governance

Actions

Proposal Procedure

data Proposal era where

Constructors

Proposal :: forall era. Typeable era => ProposalProcedure (ShelleyLedgerEra era) -> Proposal era

data GovernanceAction era

Constructors

MotionOfNoConfidence (StrictMaybe (GovPurposeId 'CommitteePurpose (ShelleyLedgerEra era)))

ProposeNewConstitution (StrictMaybe (GovPurposeId 'ConstitutionPurpose (ShelleyLedgerEra era))) Anchor (StrictMaybe ScriptHash)

ProposeNewCommittee

Fields

(StrictMaybe (GovPurposeId 'CommitteePurpose (ShelleyLedgerEra era)))

[Credential 'ColdCommitteeRole]

Old constitutional committee

(Map (Credential 'ColdCommitteeRole) EpochNo)

New committee members with epoch number when each of them expires

Rational

Quorum of the committee that is necessary for a successful vote

InfoAct

TreasuryWithdrawal [(Network, StakeCredential, Coin)] !(StrictMaybe ScriptHash)
Governance policy

InitiateHardfork (StrictMaybe (GovPurposeId 'HardForkPurpose (ShelleyLedgerEra era))) ProtVer

UpdatePParams (StrictMaybe (GovPurposeId 'PParamUpdatePurpose (ShelleyLedgerEra era))) (PParamsUpdate (ShelleyLedgerEra era)) !(StrictMaybe ScriptHash)
Governance policy

data family AsType t

A family of singleton types used in this API to indicate which type to use where it would otherwise be ambiguous or merely unclear.

Values of this type are passed to deserialisation functions for example.

createAnchor :: Url -> ByteString -> Anchor

createGovernanceActionId :: TxId -> Word16 -> GovActionId

createProposalProcedure

Arguments

:: ShelleyBasedEra era

-> Network

-> Coin
Deposit

-> StakeCredential
Credential to return the deposit to.

-> GovernanceAction era

-> Anchor

-> Proposal era

fromGovernanceAction :: GovAction (ShelleyLedgerEra era) -> GovernanceAction era

fromProposalProcedure :: ShelleyBasedEra era -> Proposal era -> (Coin, StakeCredential, GovernanceAction era)

getAnchorDataFromGovernanceAction :: GovAction (ShelleyLedgerEra era) -> Maybe Anchor

Get anchor data url and hash from a governance action. A return value of Nothing means that the governance action does not contain anchor data.

proposalTypeEquality :: (Typeable eraA, Typeable eraB) => ProposalProcedure (ShelleyLedgerEra eraA) -> ProposalProcedure (ShelleyLedgerEra eraB) -> Maybe (eraA :~: eraB)

toGovernanceAction :: ShelleyBasedEra era -> GovernanceAction era -> GovAction (ShelleyLedgerEra era)

data AnyGovernanceAction

Constructors

AnyGovernanceAction (GovAction era)

Voting Procedure

data Vote

Constructors

No

Yes

Abstain

newtype VotingProcedure era

Constructors

VotingProcedure

Fields

unVotingProcedure :: VotingProcedure (ShelleyLedgerEra era)

newtype VotingProcedures era

Constructors

VotingProcedures

Fields

unVotingProcedures :: VotingProcedures (ShelleyLedgerEra era)

data family AsType t

A family of singleton types used in this API to indicate which type to use where it would otherwise be ambiguous or merely unclear.

Values of this type are passed to deserialisation functions for example.

newtype VotesMergingConflict era

A voter, and the conflicting votes of this voter (i.e. votes with the same governance action identifier)

Constructors

VotesMergingConflict (Voter, [GovActionId])

createVotingProcedure

Arguments

:: ConwayEraOnwards era

-> Vote

-> Maybe (Url, Text)
Anchor

-> VotingProcedure era

emptyVotingProcedures :: VotingProcedures era

mergeVotingProcedures

Arguments

:: VotingProcedures era
Votes to merge

-> VotingProcedures era
Votes to merge

-> Either (VotesMergingConflict era) (VotingProcedures era)
Either the conflict found, or the merged votes

mergeVotingProcedures vote1 vote2 merges vote1 and vote2 into a single vote, or fails if the votes are incompatible.

singletonVotingProcedures :: ConwayEraOnwards era -> Voter -> GovActionId -> VotingProcedure (ShelleyLedgerEra era) -> VotingProcedures era

toVote :: Vote -> Vote

Metadata

DRep off-chain metadata

This module implements validation of metadata for DRep registration and update actions, as specified bt the CIP-119 (https:/cips.cardano.orgcip/CIP-0119).

The constraints implemented in this module can be tested against a JSON ByteString by using the function validateGovActionAnchorData in Cardano.Api.Governance.Internal.Metadata.Validation with the parameter DrepRegistrationMetadata.

data CIP119

Constructors

DrepRegistrationMetadata

Government action metadata

This module implements validation of metadata for Government Actions in general, as specified bt the CIP-108 (https:/cips.cardano.orgcip/CIP-0108), except for Government Actions covered by other CIPs.

The constraints implemented in this module can be tested against a JSON ByteString by using the function validateGovActionAnchorData in Cardano.Api.Governance.Internal.Metadata.Validation with the parameter BaseGovActionMetadata.

data CIP108

Constructors

BaseGovActionMetadata

Metadata anchor

newtype AnchorUrl

The URL to build the anchor to pass to DRep registration certificate

Constructors

AnchorUrl

Fields

unAnchorUrl :: Url

newtype AnchorDataHash

The hash to build the anchor to pass to DRep registration certificate

Constructors

AnchorDataHash

Fields

unAnchorDataHash :: SafeHash AnchorData

Metadata validation

data GovActionMetadata cip

Constructors

GovActionMetadata

Fields

hashAlgorithm :: HashAlgorithm cip

authors :: Authors cip

body :: Body cip

data family HashAlgorithm cip

data family Body cip

data family Authors cip

validateGovActionAnchorData :: FromJSON (GovActionMetadata cip) => cip -> ByteString -> Either String ()

Poll

Types

data GovernancePoll

A governance poll declaration meant to be created by one of the genesis delegates and directed towards SPOs.

A poll is made of a question and some pre-defined answers to chose from. There's an optional nonce used to make poll unique (as things down the line are based on their hashes) if the same question/answers need to be asked multiple times.

Constructors

GovernancePoll

Fields

govPollQuestion :: Text

A question as a human readable text; the text can be arbitrarily large.

govPollAnswers :: [Text]

Answers as human readable texts; their positions are used for answering.

govPollNonce :: Maybe Word

An optional nonce to make the poll unique if needs be.

data family AsType t

A family of singleton types used in this API to indicate which type to use where it would otherwise be ambiguous or merely unclear.

Values of this type are passed to deserialisation functions for example.

data GovernancePollAnswer

An (unauthenticated) answer to a poll from an SPO referring to a poll by hash digest value.

Constructors

GovernancePollAnswer

Fields

govAnsPoll :: Hash GovernancePoll

The target poll

govAnsChoice :: Word

The (0-based) index of the chosen answer from that poll

data family Hash keyrole

Errors

data GovernancePollError

Constructors

ErrGovernancePollMismatch GovernancePollMismatchError

ErrGovernancePollNoAnswer

ErrGovernancePollUnauthenticated

ErrGovernancePollMalformedAnswer DecoderError

ErrGovernancePollInvalidAnswer GovernancePollInvalidAnswerError

renderGovernancePollError :: GovernancePollError -> Text

Functions

hashGovernancePoll :: GovernancePoll -> Hash GovernancePoll

verifyPollAnswer :: GovernancePoll -> InAnyShelleyBasedEra Tx -> Either GovernancePollError [Hash PaymentKey]

Verify a poll against a given transaction and returns the signatories (verification key only) when valid.

Note: signatures aren't checked as it is assumed to have been done externally (the existence of the transaction in the ledger provides this guarantee).