Cardano.Api.Governance.Actions.ProposalProcedure

Documentation

data AnyGovernanceAction

Constructors

AnyGovernanceAction (GovAction era)

data GovernanceAction era

Constructors

MotionOfNoConfidence (StrictMaybe (GovPurposeId 'CommitteePurpose (ShelleyLedgerEra era)))

ProposeNewConstitution (StrictMaybe (GovPurposeId 'ConstitutionPurpose (ShelleyLedgerEra era))) (Anchor StandardCrypto) (StrictMaybe (ScriptHash StandardCrypto))

ProposeNewCommittee

Fields

(StrictMaybe (GovPurposeId 'CommitteePurpose (ShelleyLedgerEra era)))

[Credential 'ColdCommitteeRole StandardCrypto]

Old constitutional committee

(Map (Credential 'ColdCommitteeRole StandardCrypto) EpochNo)

New committee members with epoch number when each of them expires

Rational

Quorum of the committee that is necessary for a successful vote

InfoAct

TreasuryWithdrawal [(Network, StakeCredential, Coin)] !(StrictMaybe (ScriptHash StandardCrypto))
Governance policy

InitiateHardfork (StrictMaybe (GovPurposeId 'HardForkPurpose (ShelleyLedgerEra era))) ProtVer

UpdatePParams (StrictMaybe (GovPurposeId 'PParamUpdatePurpose (ShelleyLedgerEra era))) (PParamsUpdate (ShelleyLedgerEra era)) !(StrictMaybe (ScriptHash StandardCrypto))
Governance policy

toGovernanceAction :: ShelleyBasedEra era -> GovernanceAction era -> GovAction (ShelleyLedgerEra era)

fromGovernanceAction :: EraCrypto (ShelleyLedgerEra era) ~ StandardCrypto => GovAction (ShelleyLedgerEra era) -> GovernanceAction era

newtype Proposal era

Constructors

Proposal

Fields

unProposal :: ProposalProcedure (ShelleyLedgerEra era)

createProposalProcedure

Arguments

:: ShelleyBasedEra era

-> Network

-> Coin
Deposit

-> StakeCredential
Credential to return the deposit to.

-> GovernanceAction era

-> Anchor StandardCrypto

-> Proposal era

fromProposalProcedure :: ShelleyBasedEra era -> Proposal era -> (Coin, Hash StakeKey, GovernanceAction era)

createPreviousGovernanceActionId

Arguments

:: forall era (r :: GovActionPurpose). EraCrypto (ShelleyLedgerEra era) ~ StandardCrypto

=> TxId

-> Word16
Governance action transation index

-> GovPurposeId r (ShelleyLedgerEra era)

createGovernanceActionId :: TxId -> Word16 -> GovActionId StandardCrypto

createAnchor :: Url -> ByteString -> Anchor StandardCrypto