Cardano.Api.Governance.Actions.VotingProcedure

Documentation

newtype GovernanceActionId era

Constructors

GovernanceActionId

Fields

unGovernanceActionId :: GovActionId (EraCrypto (ShelleyLedgerEra era))

newtype Voter era

Constructors

Voter (Voter (EraCrypto (ShelleyLedgerEra era)))

data Vote

Constructors

No

Yes

Abstain

toVote :: Vote -> Vote

createVotingProcedure

Arguments

:: ConwayEraOnwards era

-> Vote

-> Maybe (Url, Text)
Anchor

-> VotingProcedure era

newtype VotingProcedure era

Constructors

VotingProcedure

Fields

unVotingProcedure :: VotingProcedure (ShelleyLedgerEra era)

newtype VotingProcedures era

Constructors

VotingProcedures

Fields

unVotingProcedures :: VotingProcedures (ShelleyLedgerEra era)

emptyVotingProcedures :: VotingProcedures era

singletonVotingProcedures :: ConwayEraOnwards era -> Voter (EraCrypto (ShelleyLedgerEra era)) -> GovActionId (EraCrypto (ShelleyLedgerEra era)) -> VotingProcedure (ShelleyLedgerEra era) -> VotingProcedures era

newtype VotesMergingConflict era

A voter, and the conflicting votes of this voter (i.e. votes with the same governance action identifier)

Constructors

VotesMergingConflict (Voter (EraCrypto (ShelleyLedgerEra era)), [GovActionId (EraCrypto (ShelleyLedgerEra era))])

mergeVotingProcedures

Arguments

:: VotingProcedures era
Votes to merge

-> VotingProcedures era
Votes to merge

-> Either (VotesMergingConflict era) (VotingProcedures era)
Either the conflict found, or the merged votes

mergeVotingProcedures vote1 vote2 merges vote1 and vote2 into a single vote, or fails if the votes are incompatible.