Cardano.Api.IO

Documentation

readByteStringFile :: MonadIO m => File content 'In -> m (Either (FileError e) ByteString)

readLazyByteStringFile :: MonadIO m => File content 'In -> m (Either (FileError e) ByteString)

readTextFile :: MonadIO m => File content 'In -> m (Either (FileError e) Text)

readFileBlocking :: FilePath -> IO ByteString

writeByteStringFileWithOwnerPermissions :: FilePath -> ByteString -> IO (Either (FileError e) ())

writeByteStringFile :: MonadIO m => File content 'Out -> ByteString -> m (Either (FileError e) ())

writeByteStringOutput :: MonadIO m => Maybe (File content 'Out) -> ByteString -> m (Either (FileError e) ())

writeLazyByteStringFileWithOwnerPermissions :: File content 'Out -> ByteString -> IO (Either (FileError e) ())

writeLazyByteStringFile :: MonadIO m => File content 'Out -> ByteString -> m (Either (FileError e) ())

writeLazyByteStringOutput :: MonadIO m => Maybe (File content 'Out) -> ByteString -> m (Either (FileError e) ())

writeTextFileWithOwnerPermissions :: File content 'Out -> Text -> IO (Either (FileError e) ())

writeTextFile :: MonadIO m => File content 'Out -> Text -> m (Either (FileError e) ())

writeTextOutput :: MonadIO m => Maybe (File content 'Out) -> Text -> m (Either (FileError e) ())

newtype File content (direction :: FileDirection)

A file path with additional type information to indicate what the file is meant to contain and whether it is to be used for reading or writing.

Constructors

File

Fields

unFile :: FilePath

data FileDirection

Constructors

In
Indicate the file is to be used for reading.

Out
Indicate the file is to be used for writing.

InOut
Indicate the file is to be used for both reading and writing.

type SocketPath = File Socket 'InOut

mapFile :: forall content (direction :: FileDirection). (FilePath -> FilePath) -> File content direction -> File content direction

onlyIn :: File content 'InOut -> File content 'In

onlyOut :: File content 'InOut -> File content 'Out

intoFile :: File content 'Out -> content -> (File content 'Out -> stream -> result) -> (content -> stream) -> result

Given a way to serialise a value and a way to write the stream to a file, serialise a value into a stream, and write it to a file.

Whilst it is possible to call the serialisation and writing functions separately, doing so means the compiler is unable to match the content type of the file with the type of the content being serialised.

Using this function ensures that the content type of the file always matches with the content value and prevents any type mismatches.

checkVrfFilePermissions :: forall content (direction :: FileDirection). File content direction -> ExceptT VRFPrivateKeyFilePermissionError IO ()

writeSecrets :: FilePath -> [Char] -> [Char] -> (a -> ByteString) -> [a] -> IO ()