Cardano.Api.IO.Base

Documentation

data FileDirection

Constructors

In
Indicate the file is to be used for reading.

Out
Indicate the file is to be used for writing.

InOut
Indicate the file is to be used for both reading and writing.

newtype File content (direction :: FileDirection)

A file path with additional type information to indicate what the file is meant to contain and whether it is to be used for reading or writing.

Constructors

File

Fields

unFile :: FilePath

data Socket

type SocketPath = File Socket 'InOut

data VRFPrivateKeyFilePermissionError

Constructors

OtherPermissionsExist FilePath

GroupPermissionsExist FilePath

GenericPermissionsExist FilePath