Cardano.Api.IPC

Description

Node IPC protocols

Node interaction

Operations that involve talking to a local Cardano node.

connectToLocalNode :: MonadIO m => LocalNodeConnectInfo -> LocalNodeClientProtocolsInMode -> m ()

Establish a connection to a local node and execute the given set of protocol handlers.

connectToLocalNodeWithVersion :: MonadIO m => LocalNodeConnectInfo -> (NodeToClientVersion -> LocalNodeClientProtocolsInMode) -> m ()

Establish a connection to a local node and execute the given set of protocol handlers parameterized on the negotiated node-to-client protocol version.

data LocalNodeConnectInfo

Constructors

LocalNodeConnectInfo

Fields

localConsensusModeParams :: ConsensusModeParams

localNodeNetworkId :: NetworkId

localNodeSocketPath :: SocketPath

data LocalNodeClientParams where

This type defines the boundary between the mode-parametrised style used in this API and the block-parametrised style used by the underlying network and consensus libraries.

This interface itself is in the block-parametrised style, with the block type itself being an hidden/existential type.

It bundles together all the necessary class instances, the consensus protocol client identifier, and the set of client side mini-protocol handlers for the node-to-client protocol.

Constructors

LocalNodeClientParamsSingleBlock :: forall block. (ProtocolClient block, LedgerSupportsProtocol (ShelleyBlock (TPraos StandardCrypto) (ShelleyEra StandardCrypto))) => ProtocolClientInfoArgs block -> (NodeToClientVersion -> LocalNodeClientProtocolsForBlock block) -> LocalNodeClientParams

LocalNodeClientParamsCardano :: forall block. (ProtocolClient block, CardanoHardForkConstraints (ConsensusCryptoForBlock block)) => ProtocolClientInfoArgs block -> (NodeToClientVersion -> LocalNodeClientProtocolsForBlock block) -> LocalNodeClientParams

mkLocalNodeClientParams :: ConsensusModeParams -> (NodeToClientVersion -> LocalNodeClientProtocolsInMode) -> LocalNodeClientParams

Convert from the mode-parametrised style to the block-parametrised style.

data LocalNodeClientProtocols block point tip slot tx txid txerr (query :: Type -> Type) (m :: Type -> Type)

The protocols we can use with a local node. Use in conjunction with connectToLocalNode.

These protocols use the types from the rest of this API. The conversion to/from the types used by the underlying wire formats is handled by connectToLocalNode.

Constructors

LocalNodeClientProtocols

Fields

localChainSyncClient :: LocalChainSyncClient block point tip m

localTxSubmissionClient :: Maybe (LocalTxSubmissionClient tx txerr m ())

localStateQueryClient :: Maybe (LocalStateQueryClient block point query m ())

localTxMonitoringClient :: Maybe (LocalTxMonitorClient txid tx slot m ())

data LocalChainSyncClient block point tip (m :: Type -> Type)

Constructors

NoLocalChainSyncClient

LocalChainSyncClientPipelined (ChainSyncClientPipelined block point tip m ())

LocalChainSyncClient (ChainSyncClient block point tip m ())

type LocalNodeClientProtocolsInMode = LocalNodeClientProtocols BlockInMode ChainPoint ChainTip SlotNo TxInMode TxIdInMode TxValidationErrorInCardanoMode QueryInMode IO

Modes

TODO move to Cardano.Api

data ConsensusModeParams where

The consensus-mode-specific parameters needed to connect to a local node that is using each consensus mode.

It is in fact only the Byron era that requires extra parameters, but this is of course inherited by the CardanoMode that uses the Byron era. The reason this parameter is needed stems from unfortunate design decisions from the legacy Byron era. The slots per epoch are needed to be able to decode epoch boundary blocks from the Byron era.

It is possible in future that we may be able to eliminate this parameter by discovering it from the node during the initial handshake.

Constructors

CardanoModeParams :: EpochSlots -> ConsensusModeParams

newtype EpochSlots

The number of slots per epoch.

Constructors

EpochSlots

Fields

unEpochSlots :: Word64

Chain sync protocol

newtype ChainSyncClient header point tip (m :: Type -> Type) a

A chain sync protocol client, on top of some effect m. The first choice of request is within that m.

Constructors

ChainSyncClient

Fields

runChainSyncClient :: m (ClientStIdle header point tip m a)

newtype ChainSyncClientPipelined header point tip (m :: Type -> Type) a

Pipelined chain sync client. It can only pipeline MsgRequestNext messages, while the MsgFindIntersect are non pipelined. This has a penalty cost of an RTT, but they are sent relatively seldom and their response might impact how many messages one would like to pipeline. It also simplifies the receiver callback.

Constructors

ChainSyncClientPipelined

Fields

runChainSyncClientPipelined :: m (ClientPipelinedStIdle 'Z header point tip m a)

data BlockInMode where

A Block in one of the eras. TODO Rename this to BlockInEra

Constructors

BlockInMode :: forall era. CardanoEra era -> Block era -> BlockInMode

Local tx submission

newtype LocalTxSubmissionClient tx reject (m :: Type -> Type) a

Constructors

LocalTxSubmissionClient

Fields

runLocalTxSubmissionClient :: m (LocalTxClientStIdle tx reject m a)

data TxInMode where

A Tx in one of the eras supported by a given protocol mode.

For multi-era modes such as the CardanoMode this type is a sum of the different transaction types for all the eras. It is used in the LocalTxSubmission protocol.

Constructors

TxInMode :: forall era. ShelleyBasedEra era -> Tx era -> TxInMode
Shelley based transactions.

TxInByronSpecial :: GenTx ByronBlock -> TxInMode
Legacy Byron transactions and things we can post to the chain which are not actually transactions. This covers: update proposals, votes and delegation certs.

data TxValidationErrorInCardanoMode

A TxValidationError in one of the eras supported by a given protocol mode.

This is used in the LocalStateQuery protocol.

data TxValidationError era

The transaction validations errors that can occur from trying to submit a transaction to a local node. The errors are specific to an era.

submitTxToNodeLocal :: MonadIO m => LocalNodeConnectInfo -> TxInMode -> m (SubmitResult TxValidationErrorInCardanoMode)

data SubmitResult reason

Isomorphic with Maybe but with a name that better describes its purpose and usage.

Constructors

SubmitSuccess

SubmitFail reason

Local state query

newtype LocalStateQueryClient block point (query :: Type -> Type) (m :: Type -> Type) a

Constructors

LocalStateQueryClient

Fields

runLocalStateQueryClient :: m (ClientStIdle block point query m a)

data AcquiringFailure

Establish a connection to a node and execute a single query using the local state query protocol.

Constructors

AFPointTooOld

AFPointNotOnChain

data QueryInMode result where

Constructors

QueryCurrentEra :: QueryInMode AnyCardanoEra

QueryInEra :: forall era result1. QueryInEra era result1 -> QueryInMode (Either EraMismatch result1)

QueryEraHistory :: QueryInMode EraHistory

QuerySystemStart :: QueryInMode SystemStart

QueryChainBlockNo :: QueryInMode (WithOrigin BlockNo)

QueryChainPoint :: QueryInMode ChainPoint

data QueryInEra era result where

Constructors

QueryByronUpdateState :: QueryInEra ByronEra ByronUpdateState

QueryInShelleyBasedEra :: forall era result. ShelleyBasedEra era -> QueryInShelleyBasedEra era result -> QueryInEra era result

data QueryInShelleyBasedEra era result where

Constructors

QueryEpoch :: forall era. QueryInShelleyBasedEra era EpochNo

QueryGenesisParameters :: forall era. QueryInShelleyBasedEra era (GenesisParameters ShelleyEra)

QueryProtocolParameters :: forall era. QueryInShelleyBasedEra era (PParams (ShelleyLedgerEra era))

QueryProtocolParametersUpdate :: forall era. QueryInShelleyBasedEra era (Map (Hash GenesisKey) ProtocolParametersUpdate)

QueryStakeDistribution :: forall era. QueryInShelleyBasedEra era (Map (Hash StakePoolKey) Rational)

QueryUTxO :: forall era. QueryUTxOFilter -> QueryInShelleyBasedEra era (UTxO era)

QueryStakeAddresses :: forall era. Set StakeCredential -> NetworkId -> QueryInShelleyBasedEra era (Map StakeAddress Coin, Map StakeAddress PoolId)

QueryStakePools :: forall era. QueryInShelleyBasedEra era (Set PoolId)

QueryStakePoolParameters :: forall era. Set PoolId -> QueryInShelleyBasedEra era (Map PoolId StakePoolParameters)

QueryDebugLedgerState :: forall era. QueryInShelleyBasedEra era (SerialisedDebugLedgerState era)

QueryProtocolState :: forall era. QueryInShelleyBasedEra era (ProtocolState era)

QueryCurrentEpochState :: forall era. QueryInShelleyBasedEra era (SerialisedCurrentEpochState era)

QueryPoolState :: forall era. Maybe (Set PoolId) -> QueryInShelleyBasedEra era (SerialisedPoolState era)

QueryPoolDistribution :: forall era. Maybe (Set PoolId) -> QueryInShelleyBasedEra era (SerialisedPoolDistribution era)

QueryStakeSnapshot :: forall era. Maybe (Set PoolId) -> QueryInShelleyBasedEra era (SerialisedStakeSnapshots era)

QueryStakeDelegDeposits :: forall era. Set StakeCredential -> QueryInShelleyBasedEra era (Map StakeCredential Coin)

QueryAccountState :: forall era. QueryInShelleyBasedEra era AccountState

QueryConstitution :: forall era. QueryInShelleyBasedEra era (Constitution (ShelleyLedgerEra era))

QueryGovState :: forall era. QueryInShelleyBasedEra era (GovState (ShelleyLedgerEra era))

QueryDRepState :: forall era. Set (Credential 'DRepRole StandardCrypto) -> QueryInShelleyBasedEra era (Map (Credential 'DRepRole StandardCrypto) (DRepState StandardCrypto))

QueryDRepStakeDistr :: forall era. Set (DRep StandardCrypto) -> QueryInShelleyBasedEra era (Map (DRep StandardCrypto) Coin)

QuerySPOStakeDistr :: forall era. Set (KeyHash 'StakePool StandardCrypto) -> QueryInShelleyBasedEra era (Map (KeyHash 'StakePool StandardCrypto) Coin)

QueryCommitteeMembersState :: forall era. Set (Credential 'ColdCommitteeRole StandardCrypto) -> Set (Credential 'HotCommitteeRole StandardCrypto) -> Set MemberStatus -> QueryInShelleyBasedEra era (CommitteeMembersState StandardCrypto)

QueryStakeVoteDelegatees :: forall era. Set StakeCredential -> QueryInShelleyBasedEra era (Map StakeCredential (DRep StandardCrypto))

queryNodeLocalState :: LocalNodeConnectInfo -> Target ChainPoint -> QueryInMode result -> ExceptT AcquiringFailure IO result

Local tx monitoring

newtype LocalTxMonitorClient txid tx slot (m :: Type -> Type) a

A tx monitor client, on top of some effect m.

Constructors

LocalTxMonitorClient

Fields

runLocalTxMonitorClient :: m (ClientStIdle txid tx slot m a)

data LocalTxMonitoringQuery

Constructors

LocalTxMonitoringQueryTx TxIdInMode
Query if a particular tx exists in the mempool. Note that, the absence of a transaction does not imply anything about how the transaction was processed: it may have been dropped, or inserted in a block.

LocalTxMonitoringSendNextTx
The mempool is modeled as an ordered list of transactions and thus, can be traversed linearly. LocalTxMonitoringSendNextTx requests the next transaction from the current list. This must be a transaction that was not previously sent to the client for this particular snapshot.

LocalTxMonitoringMempoolInformation
Ask the server about the current mempool's capacity and sizes. This is fixed in a given snapshot.

data LocalTxMonitoringResult

Constructors

LocalTxMonitoringTxExists TxId SlotNo
Slot number at which the mempool snapshot was taken

LocalTxMonitoringTxDoesNotExist TxId SlotNo
Slot number at which the mempool snapshot was taken

LocalTxMonitoringNextTx (Maybe TxInMode) SlotNo
Slot number at which the mempool snapshot was taken

LocalTxMonitoringMempoolSizeAndCapacity MempoolSizeAndCapacity SlotNo
Slot number at which the mempool snapshot was taken

data MempoolSizeAndCapacity

Describes the MemPool sizes and capacity for a given snapshot.

Constructors

MempoolSizeAndCapacity

Fields

capacityInBytes :: !Word32

The maximum capacity of the mempool. Note that this may dynamically change when the ledger state is updated.

sizeInBytes :: !Word32

The summed byte size of all the transactions in the mempool.

numberOfTxs :: !Word32

The number of transactions in the mempool

queryTxMonitoringLocal :: MonadIO m => LocalNodeConnectInfo -> LocalTxMonitoringQuery -> m LocalTxMonitoringResult

data EraHistory where

Constructors

EraHistory :: forall (xs :: [Type]). CardanoBlock StandardCrypto ~ HardForkBlock xs => Interpreter xs -> EraHistory

getProgress :: SlotNo -> EraHistory -> Either PastHorizonException (RelativeTime, SlotLength)

Common queries

getLocalChainTip :: MonadIO m => LocalNodeConnectInfo -> m ChainTip

Helpers

toAcquiringFailure :: AcquireFailure -> AcquiringFailure

data NodeToClientVersion

Enumeration of node to client protocol versions.

Constructors

NodeToClientV_9
enabled CardanoNodeToClientVersion7, i.e., Alonzo

NodeToClientV_10
added GetChainBlockNo and GetChainPoint queries

NodeToClientV_11
added GetRewardInfoPools Block query

NodeToClientV_12
added LocalTxMonitor mini-protocol

NodeToClientV_13
enabled CardanoNodeToClientVersion9, i.e., Babbage

NodeToClientV_14
added GetPoolDistr, GetPoolState, GetSnapshots

NodeToClientV_15
added query to NodeToClientVersionData

NodeToClientV_16
add ImmutableTip to LocalStateQuery, enabled CardanoNodeToClientVersion11, i.e., Conway and GetStakeDelegDeposits.

data UnsupportedNtcVersionError

Constructors

UnsupportedNtcVersionError !MinNodeToClientVersion !NodeToClientVersion