Cardano.Api.InMode

Description

Transactions in the context of a consensus mode, and other types used in the transaction submission protocol.

Transaction in a consensus mode

data TxInMode where

A Tx in one of the eras supported by a given protocol mode.

For multi-era modes such as the CardanoMode this type is a sum of the different transaction types for all the eras. It is used in the LocalTxSubmission protocol.

Constructors

TxInMode :: forall era. ShelleyBasedEra era -> Tx era -> TxInMode
Shelley based transactions.

TxInByronSpecial :: GenTx ByronBlock -> TxInMode
Legacy Byron transactions and things we can post to the chain which are not actually transactions. This covers: update proposals, votes and delegation certs.

fromConsensusGenTx :: CardanoBlock StandardCrypto ~ block => GenTx block -> TxInMode

toConsensusGenTx :: CardanoBlock StandardCrypto ~ block => TxInMode -> GenTx block

Transaction id in a consensus mode

data TxIdInMode where

A TxId in one of the eras supported by a given protocol mode.

For multi-era modes such as the CardanoMode this type is a sum of the different transaction types for all the eras. It is used in the LocalTxMonitoring protocol.

TODO Rename to TxIdInEra

Constructors

TxIdInMode :: forall era. CardanoEra era -> TxId -> TxIdInMode

toConsensusTxId :: CardanoBlock StandardCrypto ~ block => TxIdInMode -> TxId (GenTx block)

Transaction validation errors

data TxValidationError era where

The transaction validations errors that can occur from trying to submit a transaction to a local node. The errors are specific to an era.

Constructors

ByronTxValidationError :: forall era. ApplyTxErr ByronBlock -> TxValidationError era

ShelleyTxValidationError :: forall era. ShelleyBasedEra era -> ApplyTxErr (ShelleyBlock (ConsensusProtocol era) (ShelleyLedgerEra era)) -> TxValidationError era

data TxValidationErrorInCardanoMode where

A TxValidationError in one of the eras supported by a given protocol mode.

This is used in the LocalStateQuery protocol.

Constructors

TxValidationErrorInCardanoMode :: forall era. TxValidationError era -> TxValidationErrorInCardanoMode

TxValidationEraMismatch :: EraMismatch -> TxValidationErrorInCardanoMode

fromConsensusApplyTxErr :: CardanoBlock StandardCrypto ~ block => ApplyTxErr block -> TxValidationErrorInCardanoMode