Cardano.Api.Internal.DRepMetadata

Description

DRep off-chain metadata

DRep off-chain metadata

newtype DRepMetadata

A representation of the required fields for off-chain drep metadata.

Constructors

DRepMetadata

Fields

unDRepMetadata :: ByteString

hashDRepMetadata :: ByteString -> (DRepMetadata, Hash DRepMetadata)

Return the decoded metadata and the hash of the original bytes.

Data family instances

data family AsType t

A family of singleton types used in this API to indicate which type to use where it would otherwise be ambiguous or merely unclear.

Values of this type are passed to deserialisation functions for example.

data family Hash keyrole