Cardano.Api.Internal.Eras

Description

Cardano eras, sometimes we have to distinguish them.

Eras

data ByronEra

A type used as a tag to distinguish the Byron era.

data ShelleyEra

A type used as a tag to distinguish the Shelley era.

data AllegraEra

A type used as a tag to distinguish the Allegra era.

data MaryEra

A type used as a tag to distinguish the Mary era.

data AlonzoEra

A type used as a tag to distinguish the Alonzo era.

data BabbageEra

A type used as a tag to distinguish the Babbage era.

data ConwayEra

A type used as a tag to distinguish the Conway era.

CardanoEra

data CardanoEra era where

This GADT provides a value-level representation of all the Cardano eras. This enables pattern matching on the era to allow them to be treated in a non-uniform way.

This can be used in combination with the IsCardanoEra class to get access to this value.

In combination this can often enable code that handles all eras, and does so uniformly where possible, and non-uniformly where necessary.

Constructors

ByronEra :: CardanoEra ByronEra

ShelleyEra :: CardanoEra ShelleyEra

AllegraEra :: CardanoEra AllegraEra

MaryEra :: CardanoEra MaryEra

AlonzoEra :: CardanoEra AlonzoEra

BabbageEra :: CardanoEra BabbageEra

ConwayEra :: CardanoEra ConwayEra

class HasTypeProxy era => IsCardanoEra era where

The class of Cardano eras. This allows uniform handling of all Cardano eras, but also non-uniform by making case distinctions on the CardanoEra constructors.

Methods

cardanoEra :: CardanoEra era

data AnyCardanoEra where

Constructors

AnyCardanoEra :: forall era. Typeable era => CardanoEra era -> AnyCardanoEra

anyCardanoEra :: CardanoEra era -> AnyCardanoEra

Like the AnyCardanoEra constructor but does not demand a IsCardanoEra class constraint.

data InAnyCardanoEra (thing :: Type -> Type) where

This pairs up some era-dependent type with a CardanoEra value that tells us what era it is, but hides the era type. This is useful when the era is not statically known, for example when deserialising from a file.

Constructors

InAnyCardanoEra :: forall era (thing :: Type -> Type). Typeable era => CardanoEra era -> thing era -> InAnyCardanoEra thing

inAnyCardanoEra :: CardanoEra era -> thing era -> InAnyCardanoEra thing

cardanoEraConstraints :: CardanoEra era -> (CardanoEraConstraints era => a) -> a

type family CardanoLedgerEra era = (ledgerera :: Type) | ledgerera -> era where ...

A type family that connects our era type tags to equivalent type tags used in the ledger library.

This type mapping connect types from this API with types in the ledger library which allows writing conversion functions in a more generic way.

Equations

CardanoLedgerEra ByronEra = ByronEra

CardanoLedgerEra ShelleyEra = ShelleyEra

CardanoLedgerEra AllegraEra = AllegraEra

CardanoLedgerEra MaryEra = MaryEra

CardanoLedgerEra AlonzoEra = AlonzoEra

CardanoLedgerEra BabbageEra = BabbageEra

CardanoLedgerEra ConwayEra = ConwayEra

class ToCardanoEra (eon :: Type -> Type) where

Methods

toCardanoEra :: eon era -> CardanoEra era

IsEon

class Eon (eon :: Type -> Type) where

An Eon is a span of multiple eras. Eons are used to scope functionality to particular eras such that it isn't possible construct code that uses functionality that is outside of given eras.

Methods

inEonForEra

Arguments

:: a
Value to use if the eon does not include the era

-> (eon era -> a)
Function to get the value to use if the eon includes the era

-> CardanoEra era
Era to check

-> a
The value to use

Determine the value to use in an eon (a span of multiple eras). Note that the negative case is the first argument, and the positive case is the second as per the either function convention.

data EraInEon (eon :: Type -> Type) where

Constructors

EraInEon :: forall era (eon :: Type -> Type). (Typeable era, Typeable (eon era), Eon eon) => eon era -> EraInEon eon

inEonForEraMaybe

Arguments

:: Eon eon

=> (eon era -> a)
Function to get the value to use if the eon includes the era

-> CardanoEra era
Era to check

-> Maybe a
The value to use

forEraInEon

Arguments

:: Eon eon

=> CardanoEra era
Era to check

-> a
Value to use if the eon does not include the era

-> (eon era -> a)
Function to get the value to use if the eon includes the era

-> a
The value to use

forEraInEonMaybe

Arguments

:: Eon eon

=> CardanoEra era
Era to check

-> (eon era -> a)
Function to get the value to use if the eon includes the era

-> Maybe a
The value to use

forEraMaybeEon

Arguments

:: Eon eon

=> CardanoEra era
Era to check

-> Maybe (eon era)
The eon if supported in the era

maybeEon

Arguments

:: (Eon eon, IsCardanoEra era)

=> Maybe (eon era)
The eon if supported in the era

monoidForEraInEon :: (Eon eon, Monoid a) => CardanoEra era -> (eon era -> a) -> a

monoidForEraInEonA :: (Eon eon, Applicative f, Monoid a) => CardanoEra era -> (eon era -> f a) -> f a

class Convert (f :: a -> Type) (g :: a -> Type) where

The Convert class is aimed at exposing a single interface that lets us convert between eons. However this is generalizable to any injective relationship between types.

Methods

convert :: forall (era :: a). f era -> g era

class Inject t s where

Methods

inject :: t -> s

Data family instances

data family AsType t

A family of singleton types used in this API to indicate which type to use where it would otherwise be ambiguous or merely unclear.

Values of this type are passed to deserialisation functions for example.

Era case handling

Case on CardanoEra

caseByronOrShelleyBasedEra :: a -> (ShelleyBasedEraConstraints era => ShelleyBasedEra era -> a) -> CardanoEra era -> a

caseByronOrShelleyBasedEra f g era returns f in Byron and applies g to Shelley-based eras.

Case on ShelleyBasedEra

caseByronToAlonzoOrBabbageEraOnwards :: (ByronToAlonzoEraConstraints era => ByronToAlonzoEra era -> a) -> (BabbageEraOnwardsConstraints era => BabbageEraOnwards era -> a) -> CardanoEra era -> a

caseByronToAlonzoOrBabbageEraOnwards f g era applies f to byron, shelley, allegra, mary, and alonzo; and g to babbage and later eras.

caseShelleyToAllegraOrMaryEraOnwards :: (ShelleyToAllegraEraConstraints era => ShelleyToAllegraEra era -> a) -> (MaryEraOnwardsConstraints era => MaryEraOnwards era -> a) -> ShelleyBasedEra era -> a

caseShelleyToAllegraOrMaryEraOnwards f g era applies f to shelley and allegra; and applies g to mary and later eras.

caseShelleyToMaryOrAlonzoEraOnwards :: (ShelleyToMaryEraConstraints era => ShelleyToMaryEra era -> a) -> (AlonzoEraOnwardsConstraints era => AlonzoEraOnwards era -> a) -> ShelleyBasedEra era -> a

caseShelleyToMaryOrAlonzoEraOnwards f g era applies f to shelley, allegra, and mary; and applies g to alonzo and later eras.

caseShelleyToAlonzoOrBabbageEraOnwards :: (ShelleyToAlonzoEraConstraints era => ShelleyToAlonzoEra era -> a) -> (BabbageEraOnwardsConstraints era => BabbageEraOnwards era -> a) -> ShelleyBasedEra era -> a

caseShelleyToAlonzoOrBabbageEraOnwards f g era applies f to shelley, allegra, mary, and alonzo; and applies g to babbage and later eras.

caseShelleyToBabbageOrConwayEraOnwards :: (ShelleyToBabbageEraConstraints era => ShelleyToBabbageEra era -> a) -> (ConwayEraOnwardsConstraints era => ConwayEraOnwards era -> a) -> ShelleyBasedEra era -> a

caseShelleyToBabbageOrConwayEraOnwards f g era applies f to eras before conway; and applies g to conway and later eras.