Cardano.Api.Internal.Genesis

Documentation

data ShelleyGenesis c

Shelley genesis information

Note that this is needed only for a pure Shelley network, hence it being defined here rather than in its own module. In mainnet, Shelley will transition naturally from Byron, and thus will never have its own genesis information.

Constructors

ShelleyGenesis

Fields

sgSystemStart :: !UTCTime

sgNetworkMagic :: !Word32

sgNetworkId :: !Network

sgActiveSlotsCoeff :: !PositiveUnitInterval

sgSecurityParam :: !Word64

sgEpochLength :: !EpochSize

sgSlotsPerKESPeriod :: !Word64

sgMaxKESEvolutions :: !Word64

sgSlotLength :: !NominalDiffTimeMicro

sgUpdateQuorum :: !Word64

sgMaxLovelaceSupply :: !Word64

sgProtocolParams :: !(PParams (ShelleyEra c))

sgGenDelegs :: !(Map (KeyHash 'Genesis c) (GenDelegPair c))

sgInitialFunds :: ListMap (Addr c) Coin

sgInitialFunds is intentionally kept lazy, as it can otherwise cause out-of-memory problems in testing and benchmarking.

sgStaking :: ShelleyGenesisStaking c

sgStaking is intentionally kept lazy, as it can otherwise cause out-of-memory problems in testing and benchmarking.

shelleyGenesisDefaults :: ShelleyGenesis StandardCrypto

Some reasonable starting defaults for constructing a ShelleyGenesis.

You must override at least the following fields for this to be useful:

sgSystemStart the time of the first block

sgNetworkMagic to a suitable testnet or mainnet network magic number.

sgGenDelegs to have some initial nodes

sgInitialFunds to have any money in the system

sgMaxLovelaceSupply must be at least the sum of the sgInitialFunds but more if you want to allow for rewards.

alonzoGenesisDefaults :: CardanoEra era -> AlonzoGenesis

Some reasonable starting defaults for constructing a AlonzoGenesis. Based on https://github.com/IntersectMBO/cardano-node/blob/master/cardano-testnet/src/Testnet/Defaults.hs The era determines Plutus V2 cost model parameters: * Conway: 185 * <= Babbage: 175

decodeAlonzoGenesis

Arguments

:: forall era t (m :: Type -> Type). MonadTransError String t m

=> Maybe (CardanoEra era)
An optional era witness in which we're reading the genesis

-> ByteString
Genesis JSON

-> t m AlonzoGenesis

Decode Alonzo genesis in an optionally era sensitive way.

Because the Plutus V2 cost model has changed between Babbage and Conway era, we need to know the era if we want to decde Alonzo Genesis with a cost model baked in. If the V2 cost model is present in genesis, you need to provide an era witness.

When an era witness is provided, for Plutus V2 model the function additionally: 1. Does extra cost model parameters name validation: Checks for mandatory 175 parameters if provided in a map form. 2. If >= Conway: adds defaults for new 10 parameters, if they were not provided (maxBound) 3. Removes extra parameters above the max count: Babbage - 175, Conway - 185.

conwayGenesisDefaults :: ConwayGenesis StandardCrypto

Some reasonable starting defaults for constructing a ConwayGenesis. Based on https://github.com/IntersectMBO/cardano-node/blob/master/cardano-testnet/src/Testnet/Defaults.hs

Configuration

type ByronGenesisConfig = Config

type ShelleyGenesisConfig = ShelleyConfig

type AlonzoGenesisConfig = AlonzoGenesis

type ConwayGenesisConfig = ConwayGenesis StandardCrypto

data ShelleyConfig

Constructors

ShelleyConfig

Fields

scConfig :: !(ShelleyGenesis StandardCrypto)

scGenesisHash :: !GenesisHashShelley

newtype GenesisHashByron

Constructors

GenesisHashByron

Fields

unGenesisHashByron :: Text

newtype GenesisHashShelley

Constructors

GenesisHashShelley

Fields

unGenesisHashShelley :: Hash Blake2b_256 ByteString

newtype GenesisHashAlonzo

Constructors

GenesisHashAlonzo

Fields

unGenesisHashAlonzo :: Hash Blake2b_256 ByteString

newtype GenesisHashConway

Constructors

GenesisHashConway

Fields

unGenesisHashConway :: Hash Blake2b_256 ByteString

Files

type ByronGenesisFile = File ByronGenesisConfig

type ShelleyGenesisFile = File ShelleyGenesisConfig

type AlonzoGenesisFile = File AlonzoGenesisConfig

type ConwayGenesisFile = File ConwayGenesisConfig