Cardano.Api.Internal.Orphans.Misc

Orphan instances

Data CharPosition

Instance details

Methods

gfoldl :: (forall d b. Data d => c (d -> b) -> d -> c b) -> (forall g. g -> c g) -> CharPosition -> c CharPosition

gunfold :: (forall b r. Data b => c (b -> r) -> c r) -> (forall r. r -> c r) -> Constr -> c CharPosition

toConstr :: CharPosition -> Constr

dataTypeOf :: CharPosition -> DataType

dataCast1 :: Typeable t => (forall d. Data d => c (t d)) -> Maybe (c CharPosition)

dataCast2 :: Typeable t => (forall d e. (Data d, Data e) => c (t d e)) -> Maybe (c CharPosition)

gmapT :: (forall b. Data b => b -> b) -> CharPosition -> CharPosition

gmapQl :: (r -> r' -> r) -> r -> (forall d. Data d => d -> r') -> CharPosition -> r

gmapQr :: forall r r'. (r' -> r -> r) -> r -> (forall d. Data d => d -> r') -> CharPosition -> r

gmapQ :: (forall d. Data d => d -> u) -> CharPosition -> [u]

gmapQi :: Int -> (forall d. Data d => d -> u) -> CharPosition -> u

gmapM :: Monad m => (forall d. Data d => d -> m d) -> CharPosition -> m CharPosition

gmapMp :: MonadPlus m => (forall d. Data d => d -> m d) -> CharPosition -> m CharPosition

gmapMo :: MonadPlus m => (forall d. Data d => d -> m d) -> CharPosition -> m CharPosition

Data DecodingError

Instance details

Methods

gfoldl :: (forall d b. Data d => c (d -> b) -> d -> c b) -> (forall g. g -> c g) -> DecodingError -> c DecodingError

gunfold :: (forall b r. Data b => c (b -> r) -> c r) -> (forall r. r -> c r) -> Constr -> c DecodingError

toConstr :: DecodingError -> Constr

dataTypeOf :: DecodingError -> DataType

dataCast1 :: Typeable t => (forall d. Data d => c (t d)) -> Maybe (c DecodingError)

dataCast2 :: Typeable t => (forall d e. (Data d, Data e) => c (t d e)) -> Maybe (c DecodingError)

gmapT :: (forall b. Data b => b -> b) -> DecodingError -> DecodingError

gmapQl :: (r -> r' -> r) -> r -> (forall d. Data d => d -> r') -> DecodingError -> r

gmapQr :: forall r r'. (r' -> r -> r) -> r -> (forall d. Data d => d -> r') -> DecodingError -> r

gmapQ :: (forall d. Data d => d -> u) -> DecodingError -> [u]

gmapQi :: Int -> (forall d. Data d => d -> u) -> DecodingError -> u

gmapM :: Monad m => (forall d. Data d => d -> m d) -> DecodingError -> m DecodingError

gmapMp :: MonadPlus m => (forall d. Data d => d -> m d) -> DecodingError -> m DecodingError

gmapMo :: MonadPlus m => (forall d. Data d => d -> m d) -> DecodingError -> m DecodingError

Data DecoderError

Instance details

Methods

gfoldl :: (forall d b. Data d => c (d -> b) -> d -> c b) -> (forall g. g -> c g) -> DecoderError -> c DecoderError

gunfold :: (forall b r. Data b => c (b -> r) -> c r) -> (forall r. r -> c r) -> Constr -> c DecoderError

toConstr :: DecoderError -> Constr

dataTypeOf :: DecoderError -> DataType

dataCast1 :: Typeable t => (forall d. Data d => c (t d)) -> Maybe (c DecoderError)

dataCast2 :: Typeable t => (forall d e. (Data d, Data e) => c (t d e)) -> Maybe (c DecoderError)

gmapT :: (forall b. Data b => b -> b) -> DecoderError -> DecoderError

gmapQl :: (r -> r' -> r) -> r -> (forall d. Data d => d -> r') -> DecoderError -> r

gmapQr :: forall r r'. (r' -> r -> r) -> r -> (forall d. Data d => d -> r') -> DecoderError -> r

gmapQ :: (forall d. Data d => d -> u) -> DecoderError -> [u]

gmapQi :: Int -> (forall d. Data d => d -> u) -> DecoderError -> u

gmapM :: Monad m => (forall d. Data d => d -> m d) -> DecoderError -> m DecoderError

gmapMp :: MonadPlus m => (forall d. Data d => d -> m d) -> DecoderError -> m DecoderError

gmapMo :: MonadPlus m => (forall d. Data d => d -> m d) -> DecoderError -> m DecoderError

Data DeserialiseFailure

Instance details

Methods

gfoldl :: (forall d b. Data d => c (d -> b) -> d -> c b) -> (forall g. g -> c g) -> DeserialiseFailure -> c DeserialiseFailure

gunfold :: (forall b r. Data b => c (b -> r) -> c r) -> (forall r. r -> c r) -> Constr -> c DeserialiseFailure

toConstr :: DeserialiseFailure -> Constr

dataTypeOf :: DeserialiseFailure -> DataType

dataCast1 :: Typeable t => (forall d. Data d => c (t d)) -> Maybe (c DeserialiseFailure)

dataCast2 :: Typeable t => (forall d e. (Data d, Data e) => c (t d e)) -> Maybe (c DeserialiseFailure)

gmapT :: (forall b. Data b => b -> b) -> DeserialiseFailure -> DeserialiseFailure

gmapQl :: (r -> r' -> r) -> r -> (forall d. Data d => d -> r') -> DeserialiseFailure -> r

gmapQr :: forall r r'. (r' -> r -> r) -> r -> (forall d. Data d => d -> r') -> DeserialiseFailure -> r

gmapQ :: (forall d. Data d => d -> u) -> DeserialiseFailure -> [u]

gmapQi :: Int -> (forall d. Data d => d -> u) -> DeserialiseFailure -> u

gmapM :: Monad m => (forall d. Data d => d -> m d) -> DeserialiseFailure -> m DeserialiseFailure

gmapMp :: MonadPlus m => (forall d. Data d => d -> m d) -> DeserialiseFailure -> m DeserialiseFailure

gmapMo :: MonadPlus m => (forall d. Data d => d -> m d) -> DeserialiseFailure -> m DeserialiseFailure

Num Coin

Instance details

Methods

(+) :: Coin -> Coin -> Coin

(-) :: Coin -> Coin -> Coin

(*) :: Coin -> Coin -> Coin

negate :: Coin -> Coin

abs :: Coin -> Coin

signum :: Coin -> Coin

fromInteger :: Integer -> Coin

Integral Coin

Instance details

Methods

quot :: Coin -> Coin -> Coin

rem :: Coin -> Coin -> Coin

div :: Coin -> Coin -> Coin

mod :: Coin -> Coin -> Coin

quotRem :: Coin -> Coin -> (Coin, Coin)

divMod :: Coin -> Coin -> (Coin, Coin)

toInteger :: Coin -> Integer

Real Coin
These instances originally existed on the Lovelace type. As the Lovelace type is deleted and we use L.Coin instead, these instances are added to L.Coin. The instances are purely for the convenience of writing expressions involving L.Coin but be aware that not all uses of these typeclasses are valid.

Instance details

Methods

toRational :: Coin -> Rational

Pretty Error

Instance details

Methods

pretty :: Error -> Doc ann

prettyList :: [Error] -> Doc ann

Semigroup (AlonzoPParams StrictMaybe era)

Instance details

Methods

(<>) :: AlonzoPParams StrictMaybe era -> AlonzoPParams StrictMaybe era -> AlonzoPParams StrictMaybe era

sconcat :: NonEmpty (AlonzoPParams StrictMaybe era) -> AlonzoPParams StrictMaybe era

stimes :: Integral b => b -> AlonzoPParams StrictMaybe era -> AlonzoPParams StrictMaybe era

Semigroup (BabbagePParams StrictMaybe era)

Instance details

Methods

(<>) :: BabbagePParams StrictMaybe era -> BabbagePParams StrictMaybe era -> BabbagePParams StrictMaybe era

sconcat :: NonEmpty (BabbagePParams StrictMaybe era) -> BabbagePParams StrictMaybe era

stimes :: Integral b => b -> BabbagePParams StrictMaybe era -> BabbagePParams StrictMaybe era

Semigroup (ConwayPParams StrictMaybe era)

Instance details

Methods

(<>) :: ConwayPParams StrictMaybe era -> ConwayPParams StrictMaybe era -> ConwayPParams StrictMaybe era

sconcat :: NonEmpty (ConwayPParams StrictMaybe era) -> ConwayPParams StrictMaybe era

stimes :: Integral b => b -> ConwayPParams StrictMaybe era -> ConwayPParams StrictMaybe era

Semigroup (ShelleyPParams StrictMaybe era)

Instance details

Methods

(<>) :: ShelleyPParams StrictMaybe era -> ShelleyPParams StrictMaybe era -> ShelleyPParams StrictMaybe era

sconcat :: NonEmpty (ShelleyPParams StrictMaybe era) -> ShelleyPParams StrictMaybe era

stimes :: Integral b => b -> ShelleyPParams StrictMaybe era -> ShelleyPParams StrictMaybe era

IsList (ListMap k a)

Instance details

Associated Types

type Item (ListMap k a)

Instance details

Defined in Cardano.Api.Internal.Orphans.Misc

type Item (ListMap k a) = (k, a)

Methods

fromList :: [Item (ListMap k a)] -> ListMap k a

fromListN :: Int -> [Item (ListMap k a)] -> ListMap k a

toList :: ListMap k a -> [Item (ListMap k a)]