Cardano.Api.Internal.Orphans.Serialisation

Documentation

data family AsType t

A family of singleton types used in this API to indicate which type to use where it would otherwise be ambiguous or merely unclear.

Values of this type are passed to deserialisation functions for example.

Orphan instances

FromJSON ParamName

Instance details

Methods

parseJSON :: Value -> Parser ParamName

parseJSONList :: Value -> Parser [ParamName]

omittedField :: Maybe ParamName

FromJSONKey ParamName

Instance details

Methods

fromJSONKey :: FromJSONKeyFunction ParamName

fromJSONKeyList :: FromJSONKeyFunction [ParamName]

ToJSON ApplyMempoolPayloadErr

Instance details

Methods

toJSON :: ApplyMempoolPayloadErr -> Value

toEncoding :: ApplyMempoolPayloadErr -> Encoding

toJSONList :: [ApplyMempoolPayloadErr] -> Value

toEncodingList :: [ApplyMempoolPayloadErr] -> Encoding

omitField :: ApplyMempoolPayloadErr -> Bool

ToJSON KeyHash

Instance details

Methods

toJSON :: KeyHash -> Value

toEncoding :: KeyHash -> Encoding

toJSONList :: [KeyHash] -> Value

toEncodingList :: [KeyHash] -> Encoding

omitField :: KeyHash -> Bool

ToJSON LovelaceError

Instance details

Methods

toJSON :: LovelaceError -> Value

toEncoding :: LovelaceError -> Encoding

toJSONList :: [LovelaceError] -> Value

toEncodingList :: [LovelaceError] -> Encoding

omitField :: LovelaceError -> Bool

ToJSON Error

Instance details

Methods

toJSON :: Error -> Value

toEncoding :: Error -> Encoding

toJSONList :: [Error] -> Value

toEncodingList :: [Error] -> Encoding

omitField :: Error -> Bool

ToJSON UTxOError

Instance details

Methods

toJSON :: UTxOError -> Value

toEncoding :: UTxOError -> Encoding

toJSONList :: [UTxOError] -> Value

toEncodingList :: [UTxOError] -> Encoding

omitField :: UTxOError -> Bool

ToJSON TxValidationError

Instance details

Methods

toJSON :: TxValidationError -> Value

toEncoding :: TxValidationError -> Encoding

toJSONList :: [TxValidationError] -> Value

toEncodingList :: [TxValidationError] -> Encoding

omitField :: TxValidationError -> Bool

ToJSON UTxOValidationError

Instance details

Methods

toJSON :: UTxOValidationError -> Value

toEncoding :: UTxOValidationError -> Encoding

toJSONList :: [UTxOValidationError] -> Value

toEncodingList :: [UTxOValidationError] -> Encoding

omitField :: UTxOValidationError -> Bool

ToJSON ApplicationNameError

Instance details

Methods

toJSON :: ApplicationNameError -> Value

toEncoding :: ApplicationNameError -> Encoding

toJSONList :: [ApplicationNameError] -> Value

toEncodingList :: [ApplicationNameError] -> Encoding

omitField :: ApplicationNameError -> Bool

ToJSON SoftwareVersionError

Instance details

Methods

toJSON :: SoftwareVersionError -> Value

toEncoding :: SoftwareVersionError -> Encoding

toJSONList :: [SoftwareVersionError] -> Value

toEncodingList :: [SoftwareVersionError] -> Encoding

omitField :: SoftwareVersionError -> Bool

ToJSON SystemTagError

Instance details

Methods

toJSON :: SystemTagError -> Value

toEncoding :: SystemTagError -> Encoding

toJSONList :: [SystemTagError] -> Value

toEncodingList :: [SystemTagError] -> Encoding

omitField :: SystemTagError -> Bool

ToJSON Error

Instance details

Methods

toJSON :: Error -> Value

toEncoding :: Error -> Encoding

toJSONList :: [Error] -> Value

toEncodingList :: [Error] -> Encoding

omitField :: Error -> Bool

ToJSON Error

Instance details

Methods

toJSON :: Error -> Value

toEncoding :: Error -> Encoding

toJSONList :: [Error] -> Value

toEncodingList :: [Error] -> Encoding

omitField :: Error -> Bool

ToJSON Adopted

Instance details

Methods

toJSON :: Adopted -> Value

toEncoding :: Adopted -> Encoding

toJSONList :: [Adopted] -> Value

toEncodingList :: [Adopted] -> Encoding

omitField :: Adopted -> Bool

ToJSON ApplicationVersion

Instance details

Methods

toJSON :: ApplicationVersion -> Value

toEncoding :: ApplicationVersion -> Encoding

toJSONList :: [ApplicationVersion] -> Value

toEncodingList :: [ApplicationVersion] -> Encoding

omitField :: ApplicationVersion -> Bool

ToJSON Error

Instance details

Methods

toJSON :: Error -> Value

toEncoding :: Error -> Encoding

toJSONList :: [Error] -> Value

toEncodingList :: [Error] -> Encoding

omitField :: Error -> Bool

ToJSON Error

Instance details

Methods

toJSON :: Error -> Value

toEncoding :: Error -> Encoding

toJSONList :: [Error] -> Value

toEncodingList :: [Error] -> Encoding

omitField :: Error -> Bool

ToJSON VotingPeriod

Instance details

Methods

toJSON :: VotingPeriod -> Value

toEncoding :: VotingPeriod -> Encoding

toJSONList :: [VotingPeriod] -> Value

toEncodingList :: [VotingPeriod] -> Encoding

omitField :: VotingPeriod -> Bool

ToJSON ChainDepState

Instance details

Methods

toJSON :: ChainDepState -> Value

toEncoding :: ChainDepState -> Encoding

toJSONList :: [ChainDepState] -> Value

toEncodingList :: [ChainDepState] -> Encoding

omitField :: ChainDepState -> Bool

ToJSON HashHeader

Instance details

Methods

toJSON :: HashHeader -> Value

toEncoding :: HashHeader -> Encoding

toJSONList :: [HashHeader] -> Value

toEncodingList :: [HashHeader] -> Encoding

omitField :: HashHeader -> Bool

ToJSON PrtclState

Instance details

Methods

toJSON :: PrtclState -> Value

toEncoding :: PrtclState -> Encoding

toJSONList :: [PrtclState] -> Value

toEncodingList :: [PrtclState] -> Encoding

omitField :: PrtclState -> Bool

ToJSON TicknState

Instance details

Methods

toJSON :: TicknState -> Value

toEncoding :: TicknState -> Encoding

toJSONList :: [TicknState] -> Value

toEncodingList :: [TicknState] -> Encoding

omitField :: TicknState -> Bool

ToJSON ByronHash

Instance details

Methods

toJSON :: ByronHash -> Value

toEncoding :: ByronHash -> Encoding

toJSONList :: [ByronHash] -> Value

toEncodingList :: [ByronHash] -> Encoding

omitField :: ByronHash -> Bool

ToJSON StakeSnapshot

Instance details

Methods

toJSON :: StakeSnapshot -> Value

toEncoding :: StakeSnapshot -> Encoding

toJSONList :: [StakeSnapshot] -> Value

toEncodingList :: [StakeSnapshot] -> Encoding

omitField :: StakeSnapshot -> Bool

ToJSON StakeSnapshots

Instance details

Methods

toJSON :: StakeSnapshots -> Value

toEncoding :: StakeSnapshots -> Encoding

toJSONList :: [StakeSnapshots] -> Value

toEncodingList :: [StakeSnapshots] -> Encoding

omitField :: StakeSnapshots -> Bool

ToJSON ShelleyHash

Instance details

Methods

toJSON :: ShelleyHash -> Value

toEncoding :: ShelleyHash -> Encoding

toJSONList :: [ShelleyHash] -> Value

toEncodingList :: [ShelleyHash] -> Encoding

omitField :: ShelleyHash -> Bool

ToJSON PraosState

Instance details

Methods

toJSON :: PraosState -> Value

toEncoding :: PraosState -> Encoding

toJSONList :: [PraosState] -> Value

toEncodingList :: [PraosState] -> Encoding

omitField :: PraosState -> Bool

ToJSON TPraosState

Instance details

Methods

toJSON :: TPraosState -> Value

toEncoding :: TPraosState -> Encoding

toJSONList :: [TPraosState] -> Value

toEncodingList :: [TPraosState] -> Encoding

omitField :: TPraosState -> Bool

ToJSONKey ApplicationName

Instance details

Methods

toJSONKey :: ToJSONKeyFunction ApplicationName

toJSONKeyList :: ToJSONKeyFunction [ApplicationName]

Generic ApplyMempoolPayloadErr

Instance details

Associated Types

type Rep ApplyMempoolPayloadErr

Instance details

Defined in Cardano.Api.Internal.Orphans.Serialisation

type Rep ApplyMempoolPayloadErr = D1 ('MetaData "ApplyMempoolPayloadErr" "Cardano.Chain.Byron.API.Mempool" "cardano-ledger-byron-1.1.0.0-7fb551a04b7ebd202180a636c363fd4c69d431c37908920ba820fedd2c0d0ade" 'False) ((C1 ('MetaCons "MempoolTxErr" 'PrefixI 'False) (S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 UTxOValidationError)) :+: C1 ('MetaCons "MempoolDlgErr" 'PrefixI 'False) (S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 Error))) :+: (C1 ('MetaCons "MempoolUpdateProposalErr" 'PrefixI 'False) (S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 Error)) :+: C1 ('MetaCons "MempoolUpdateVoteErr" 'PrefixI 'False) (S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 Error))))

Methods

from :: ApplyMempoolPayloadErr -> Rep ApplyMempoolPayloadErr x

to :: Rep ApplyMempoolPayloadErr x -> ApplyMempoolPayloadErr

Generic LovelaceError

Instance details

Associated Types

type Rep LovelaceError

Instance details

Defined in Cardano.Api.Internal.Orphans.Serialisation

type Rep LovelaceError = D1 ('MetaData "LovelaceError" "Cardano.Chain.Common.Lovelace" "cardano-ledger-byron-1.1.0.0-7fb551a04b7ebd202180a636c363fd4c69d431c37908920ba820fedd2c0d0ade" 'False) ((C1 ('MetaCons "LovelaceOverflow" 'PrefixI 'False) (S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 Word64)) :+: C1 ('MetaCons "LovelaceTooLarge" 'PrefixI 'False) (S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 Integer))) :+: (C1 ('MetaCons "LovelaceTooSmall" 'PrefixI 'False) (S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 Integer)) :+: C1 ('MetaCons "LovelaceUnderflow" 'PrefixI 'False) (S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 Word64) :*: S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 Word64))))

Methods

from :: LovelaceError -> Rep LovelaceError x

to :: Rep LovelaceError x -> LovelaceError

Generic Error

Instance details

Methods

from :: Error -> Rep Error x

to :: Rep Error x -> Error

Generic UTxOError

Instance details

Associated Types

type Rep UTxOError

Instance details

Defined in Cardano.Api.Internal.Orphans.Serialisation

type Rep UTxOError = D1 ('MetaData "UTxOError" "Cardano.Chain.UTxO.UTxO" "cardano-ledger-byron-1.1.0.0-7fb551a04b7ebd202180a636c363fd4c69d431c37908920ba820fedd2c0d0ade" 'False) (C1 ('MetaCons "UTxOMissingInput" 'PrefixI 'False) (S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 TxIn)) :+: C1 ('MetaCons "UTxOOverlappingUnion" 'PrefixI 'False) (U1 :: Type -> Type))

Methods

from :: UTxOError -> Rep UTxOError x

to :: Rep UTxOError x -> UTxOError

Generic TxValidationError

Instance details

Associated Types

type Rep TxValidationError

Instance details

Defined in Cardano.Api.Internal.Orphans.Serialisation

type Rep TxValidationError = D1 ('MetaData "TxValidationError" "Cardano.Chain.UTxO.Validation" "cardano-ledger-byron-1.1.0.0-7fb551a04b7ebd202180a636c363fd4c69d431c37908920ba820fedd2c0d0ade" 'False) (((C1 ('MetaCons "TxValidationLovelaceError" 'PrefixI 'False) (S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 Text) :*: S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 LovelaceError)) :+: C1 ('MetaCons "TxValidationFeeTooSmall" 'PrefixI 'False) (S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 Tx) :*: (S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 Lovelace) :*: S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 Lovelace)))) :+: (C1 ('MetaCons "TxValidationWitnessWrongSignature" 'PrefixI 'False) (S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 TxInWitness) :*: (S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 ProtocolMagicId) :*: S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 TxSigData))) :+: C1 ('MetaCons "TxValidationWitnessWrongKey" 'PrefixI 'False) (S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 TxInWitness) :*: S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 Address)))) :+: ((C1 ('MetaCons "TxValidationMissingInput" 'PrefixI 'False) (S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 TxIn)) :+: C1 ('MetaCons "TxValidationNetworkMagicMismatch" 'PrefixI 'False) (S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 NetworkMagic) :*: S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 NetworkMagic))) :+: (C1 ('MetaCons "TxValidationTxTooLarge" 'PrefixI 'False) (S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 Natural) :*: S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 Natural)) :+: (C1 ('MetaCons "TxValidationUnknownAddressAttributes" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "TxValidationUnknownAttributes" 'PrefixI 'False) (U1 :: Type -> Type)))))

Methods

from :: TxValidationError -> Rep TxValidationError x

to :: Rep TxValidationError x -> TxValidationError

Generic UTxOValidationError

Instance details

Associated Types

type Rep UTxOValidationError

Instance details

Defined in Cardano.Api.Internal.Orphans.Serialisation

type Rep UTxOValidationError = D1 ('MetaData "UTxOValidationError" "Cardano.Chain.UTxO.Validation" "cardano-ledger-byron-1.1.0.0-7fb551a04b7ebd202180a636c363fd4c69d431c37908920ba820fedd2c0d0ade" 'False) (C1 ('MetaCons "UTxOValidationTxValidationError" 'PrefixI 'False) (S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 TxValidationError)) :+: C1 ('MetaCons "UTxOValidationUTxOError" 'PrefixI 'False) (S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 UTxOError)))

Methods

from :: UTxOValidationError -> Rep UTxOValidationError x

to :: Rep UTxOValidationError x -> UTxOValidationError

Generic ApplicationNameError

Instance details

Associated Types

type Rep ApplicationNameError

Instance details

Defined in Cardano.Api.Internal.Orphans.Serialisation

type Rep ApplicationNameError = D1 ('MetaData "ApplicationNameError" "Cardano.Chain.Update.ApplicationName" "cardano-ledger-byron-1.1.0.0-7fb551a04b7ebd202180a636c363fd4c69d431c37908920ba820fedd2c0d0ade" 'False) (C1 ('MetaCons "ApplicationNameTooLong" 'PrefixI 'False) (S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 Text)) :+: C1 ('MetaCons "ApplicationNameNotAscii" 'PrefixI 'False) (S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 Text)))

Methods

from :: ApplicationNameError -> Rep ApplicationNameError x

to :: Rep ApplicationNameError x -> ApplicationNameError

Generic SoftwareVersionError

Instance details

Associated Types

type Rep SoftwareVersionError

Instance details

Defined in Cardano.Api.Internal.Orphans.Serialisation

type Rep SoftwareVersionError = D1 ('MetaData "SoftwareVersionError" "Cardano.Chain.Update.SoftwareVersion" "cardano-ledger-byron-1.1.0.0-7fb551a04b7ebd202180a636c363fd4c69d431c37908920ba820fedd2c0d0ade" 'False) (C1 ('MetaCons "SoftwareVersionApplicationNameError" 'PrefixI 'False) (S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 ApplicationNameError)))

Methods

from :: SoftwareVersionError -> Rep SoftwareVersionError x

to :: Rep SoftwareVersionError x -> SoftwareVersionError

Generic SystemTagError

Instance details

Associated Types

type Rep SystemTagError

Instance details

Defined in Cardano.Api.Internal.Orphans.Serialisation

type Rep SystemTagError = D1 ('MetaData "SystemTagError" "Cardano.Chain.Update.SystemTag" "cardano-ledger-byron-1.1.0.0-7fb551a04b7ebd202180a636c363fd4c69d431c37908920ba820fedd2c0d0ade" 'False) (C1 ('MetaCons "SystemTagNotAscii" 'PrefixI 'False) (S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 Text)) :+: C1 ('MetaCons "SystemTagTooLong" 'PrefixI 'False) (S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 Text)))

Methods

from :: SystemTagError -> Rep SystemTagError x

to :: Rep SystemTagError x -> SystemTagError

Generic Error

Instance details

Methods

from :: Error -> Rep Error x

to :: Rep Error x -> Error

Generic Error

Instance details

Methods

from :: Error -> Rep Error x

to :: Rep Error x -> Error

Generic Adopted

Instance details

Associated Types

type Rep Adopted

Instance details

Defined in Cardano.Api.Internal.Orphans.Serialisation

type Rep Adopted = D1 ('MetaData "Adopted" "Cardano.Chain.Update.Validation.Registration" "cardano-ledger-byron-1.1.0.0-7fb551a04b7ebd202180a636c363fd4c69d431c37908920ba820fedd2c0d0ade" 'True) (C1 ('MetaCons "Adopted" 'PrefixI 'False) (S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 ProtocolVersion)))

Methods

from :: Adopted -> Rep Adopted x

to :: Rep Adopted x -> Adopted

Generic Error

Instance details

Methods

from :: Error -> Rep Error x

to :: Rep Error x -> Error

Generic Error

Instance details

Associated Types

type Rep Error

Instance details

Defined in Cardano.Api.Internal.Orphans.Serialisation

type Rep Error = D1 ('MetaData "Error" "Cardano.Chain.Update.Validation.Voting" "cardano-ledger-byron-1.1.0.0-7fb551a04b7ebd202180a636c363fd4c69d431c37908920ba820fedd2c0d0ade" 'False) ((C1 ('MetaCons "VotingInvalidSignature" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "VotingProposalNotRegistered" 'PrefixI 'False) (S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 UpId))) :+: (C1 ('MetaCons "VotingVoterNotDelegate" 'PrefixI 'False) (S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 KeyHash)) :+: C1 ('MetaCons "VotingVoteAlreadyCast" 'PrefixI 'False) (S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 KeyHash))))

Methods

from :: Error -> Rep Error x

to :: Rep Error x -> Error

Show ParamName

Instance details

Methods

showsPrec :: Int -> ParamName -> ShowS

show :: ParamName -> String

showList :: [ParamName] -> ShowS

HasTypeProxy GovActionId

Instance details

Associated Types

data AsType GovActionId

Instance details

Defined in Cardano.Api.Internal.Orphans.Serialisation

data AsType GovActionId = AsGovActionId

Methods

proxyToAsType :: Proxy GovActionId -> AsType GovActionId

SerialiseAsRawBytes GovActionId

Instance details

Methods

serialiseToRawBytes :: GovActionId -> ByteString

deserialiseFromRawBytes :: AsType GovActionId -> ByteString -> Either SerialiseAsRawBytesError GovActionId

Pretty Coin

Instance details

Methods

pretty :: Coin -> Doc ann

prettyList :: [Coin] -> Doc ann

Pretty AssetName

Instance details

Methods

pretty :: AssetName -> Doc ann

prettyList :: [AssetName] -> Doc ann

Pretty MultiAsset

Instance details

Methods

pretty :: MultiAsset -> Doc ann

prettyList :: [MultiAsset] -> Doc ann

Pretty PolicyID

Instance details

Methods

pretty :: PolicyID -> Doc ann

prettyList :: [PolicyID] -> Doc ann

Show (AllegraUtxoPredFailure ledgerera) => ToJSON (AllegraUtxoPredFailure ledgerera)

Instance details

Methods

toJSON :: AllegraUtxoPredFailure ledgerera -> Value

toEncoding :: AllegraUtxoPredFailure ledgerera -> Encoding

toJSONList :: [AllegraUtxoPredFailure ledgerera] -> Value

toEncodingList :: [AllegraUtxoPredFailure ledgerera] -> Encoding

omitField :: AllegraUtxoPredFailure ledgerera -> Bool

Show (AlonzoUtxoPredFailure ledgerera) => ToJSON (AlonzoUtxoPredFailure ledgerera)

Instance details

Methods

toJSON :: AlonzoUtxoPredFailure ledgerera -> Value

toEncoding :: AlonzoUtxoPredFailure ledgerera -> Encoding

toJSONList :: [AlonzoUtxoPredFailure ledgerera] -> Value

toEncodingList :: [AlonzoUtxoPredFailure ledgerera] -> Encoding

omitField :: AlonzoUtxoPredFailure ledgerera -> Bool

(ToJSON (PredicateFailure (EraRule "UTXO" ledgerera)), ToJSON (PlutusPurpose AsItem ledgerera), ToJSON (PlutusPurpose AsIx ledgerera)) => ToJSON (AlonzoUtxowPredFailure ledgerera)

Instance details

Methods

toJSON :: AlonzoUtxowPredFailure ledgerera -> Value

toEncoding :: AlonzoUtxowPredFailure ledgerera -> Encoding

toJSONList :: [AlonzoUtxowPredFailure ledgerera] -> Value

toEncodingList :: [AlonzoUtxowPredFailure ledgerera] -> Encoding

omitField :: AlonzoUtxowPredFailure ledgerera -> Bool

Show (BabbageUtxoPredFailure ledgerera) => ToJSON (BabbageUtxoPredFailure ledgerera)

Instance details

Methods

toJSON :: BabbageUtxoPredFailure ledgerera -> Value

toEncoding :: BabbageUtxoPredFailure ledgerera -> Encoding

toJSONList :: [BabbageUtxoPredFailure ledgerera] -> Value

toEncodingList :: [BabbageUtxoPredFailure ledgerera] -> Encoding

omitField :: BabbageUtxoPredFailure ledgerera -> Bool

(ToJSON (PredicateFailure (EraRule "UTXO" ledgerera)), ToJSON (TxCert ledgerera), ToJSON (PlutusPurpose AsItem ledgerera), ToJSON (PlutusPurpose AsIx ledgerera)) => ToJSON (BabbageUtxowPredFailure ledgerera)

Instance details

Methods

toJSON :: BabbageUtxowPredFailure ledgerera -> Value

toEncoding :: BabbageUtxowPredFailure ledgerera -> Encoding

toJSONList :: [BabbageUtxowPredFailure ledgerera] -> Value

toEncodingList :: [BabbageUtxowPredFailure ledgerera] -> Encoding

omitField :: BabbageUtxowPredFailure ledgerera -> Bool

ToJSON a => ToJSON (TooLarge a)

Instance details

Methods

toJSON :: TooLarge a -> Value

toEncoding :: TooLarge a -> Encoding

toJSONList :: [TooLarge a] -> Value

toEncodingList :: [TooLarge a] -> Encoding

omitField :: TooLarge a -> Bool

Show (ConwayLedgerPredFailure ledgerera) => ToJSON (ConwayLedgerPredFailure ledgerera)

Instance details

Methods

toJSON :: ConwayLedgerPredFailure ledgerera -> Value

toEncoding :: ConwayLedgerPredFailure ledgerera -> Encoding

toJSONList :: [ConwayLedgerPredFailure ledgerera] -> Value

toEncodingList :: [ConwayLedgerPredFailure ledgerera] -> Encoding

omitField :: ConwayLedgerPredFailure ledgerera -> Bool

ToJSON (VKey 'Witness)

Instance details

Methods

toJSON :: VKey 'Witness -> Value

toEncoding :: VKey 'Witness -> Encoding

toJSONList :: [VKey 'Witness] -> Value

toEncodingList :: [VKey 'Witness] -> Encoding

omitField :: VKey 'Witness -> Bool

ToJSON (PredicateFailure (EraRule "LEDGER" ledgerera)) => ToJSON (ApplyTxError ledgerera)

Instance details

Methods

toJSON :: ApplyTxError ledgerera -> Value

toEncoding :: ApplyTxError ledgerera -> Encoding

toJSONList :: [ApplyTxError ledgerera] -> Value

toEncodingList :: [ApplyTxError ledgerera] -> Encoding

omitField :: ApplyTxError ledgerera -> Bool

Show (ShelleyDelegsPredFailure ledgerera) => ToJSON (ShelleyDelegsPredFailure ledgerera)

Instance details

Methods

toJSON :: ShelleyDelegsPredFailure ledgerera -> Value

toEncoding :: ShelleyDelegsPredFailure ledgerera -> Encoding

toJSONList :: [ShelleyDelegsPredFailure ledgerera] -> Value

toEncodingList :: [ShelleyDelegsPredFailure ledgerera] -> Encoding

omitField :: ShelleyDelegsPredFailure ledgerera -> Bool

(ToJSON (PredicateFailure (EraRule "UTXOW" ledgerera)), ToJSON (PredicateFailure (EraRule "DELEGS" ledgerera))) => ToJSON (ShelleyLedgerPredFailure ledgerera)

Instance details

Methods

toJSON :: ShelleyLedgerPredFailure ledgerera -> Value

toEncoding :: ShelleyLedgerPredFailure ledgerera -> Encoding

toJSONList :: [ShelleyLedgerPredFailure ledgerera] -> Value

toEncodingList :: [ShelleyLedgerPredFailure ledgerera] -> Encoding

omitField :: ShelleyLedgerPredFailure ledgerera -> Bool

ToJSON (PredicateFailure (EraRule "UTXO" ledgerera)) => ToJSON (ShelleyPpupPredFailure ledgerera)

Instance details

Methods

toJSON :: ShelleyPpupPredFailure ledgerera -> Value

toEncoding :: ShelleyPpupPredFailure ledgerera -> Encoding

toJSONList :: [ShelleyPpupPredFailure ledgerera] -> Value

toEncodingList :: [ShelleyPpupPredFailure ledgerera] -> Encoding

omitField :: ShelleyPpupPredFailure ledgerera -> Bool

Show (ShelleyUtxoPredFailure ledgerera) => ToJSON (ShelleyUtxoPredFailure ledgerera)

Instance details

Methods

toJSON :: ShelleyUtxoPredFailure ledgerera -> Value

toEncoding :: ShelleyUtxoPredFailure ledgerera -> Encoding

toJSONList :: [ShelleyUtxoPredFailure ledgerera] -> Value

toEncodingList :: [ShelleyUtxoPredFailure ledgerera] -> Encoding

omitField :: ShelleyUtxoPredFailure ledgerera -> Bool

ToJSON (PredicateFailure (EraRule "UTXO" ledgerera)) => ToJSON (ShelleyUtxowPredFailure ledgerera)

Instance details

Methods

toJSON :: ShelleyUtxowPredFailure ledgerera -> Value

toEncoding :: ShelleyUtxowPredFailure ledgerera -> Encoding

toJSONList :: [ShelleyUtxowPredFailure ledgerera] -> Value

toEncodingList :: [ShelleyUtxowPredFailure ledgerera] -> Encoding

omitField :: ShelleyUtxowPredFailure ledgerera -> Bool

Generic (TooLarge a)

Instance details

Associated Types

type Rep (TooLarge a)

Instance details

Defined in Cardano.Api.Internal.Orphans.Serialisation

type Rep (TooLarge a) = D1 ('MetaData "TooLarge" "Cardano.Chain.Update.Validation.Registration" "cardano-ledger-byron-1.1.0.0-7fb551a04b7ebd202180a636c363fd4c69d431c37908920ba820fedd2c0d0ade" 'False) (C1 ('MetaCons "TooLarge" 'PrefixI 'True) (S1 ('MetaSel ('Just "tlActual") 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 a) :*: S1 ('MetaSel ('Just "tlMaxBound") 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 a)))

Methods

from :: TooLarge a -> Rep (TooLarge a) x

to :: Rep (TooLarge a) x -> TooLarge a

Generic (ApplyTxError era)

Instance details

Associated Types

type Rep (ApplyTxError era)

Instance details

Defined in Cardano.Api.Internal.Orphans.Serialisation

type Rep (ApplyTxError era) = D1 ('MetaData "ApplyTxError" "Cardano.Ledger.Shelley.API.Mempool" "cardano-ledger-shelley-1.16.0.0-c94e4acbae2c451b736d7fa131482fce1d6ffd0e83dcc66450421e4714554169" 'True) (C1 ('MetaCons "ApplyTxError" 'PrefixI 'False) (S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 (NonEmpty (PredicateFailure (EraRule "LEDGER" era))))))

Methods

from :: ApplyTxError era -> Rep (ApplyTxError era) x

to :: Rep (ApplyTxError era) x -> ApplyTxError era

Show a => Show (SubmitResult a)

Instance details

Methods

showsPrec :: Int -> SubmitResult a -> ShowS

show :: SubmitResult a -> String

showList :: [SubmitResult a] -> ShowS

HasTypeProxy (Credential 'ColdCommitteeRole)

Instance details

Associated Types

data AsType (Credential 'ColdCommitteeRole)

Instance details

Defined in Cardano.Api.Internal.Orphans.Serialisation

data AsType (Credential 'ColdCommitteeRole) = AsColdCommitteeCredential

Methods

proxyToAsType :: Proxy (Credential 'ColdCommitteeRole) -> AsType (Credential 'ColdCommitteeRole)

HasTypeProxy (Credential 'DRepRole)

Instance details

Associated Types

data AsType (Credential 'DRepRole)

Instance details

Defined in Cardano.Api.Internal.Orphans.Serialisation

data AsType (Credential 'DRepRole) = AsDrepCredential

Methods

proxyToAsType :: Proxy (Credential 'DRepRole) -> AsType (Credential 'DRepRole)

HasTypeProxy (Credential 'HotCommitteeRole)

Instance details

Associated Types

data AsType (Credential 'HotCommitteeRole)

Instance details

Defined in Cardano.Api.Internal.Orphans.Serialisation

data AsType (Credential 'HotCommitteeRole) = AsHotCommitteeCredential

Methods

proxyToAsType :: Proxy (Credential 'HotCommitteeRole) -> AsType (Credential 'HotCommitteeRole)

SerialiseAsRawBytes (Credential 'ColdCommitteeRole)

Instance details

Methods

serialiseToRawBytes :: Credential 'ColdCommitteeRole -> ByteString

deserialiseFromRawBytes :: AsType (Credential 'ColdCommitteeRole) -> ByteString -> Either SerialiseAsRawBytesError (Credential 'ColdCommitteeRole)

SerialiseAsRawBytes (Credential 'DRepRole)

Instance details

Methods

serialiseToRawBytes :: Credential 'DRepRole -> ByteString

deserialiseFromRawBytes :: AsType (Credential 'DRepRole) -> ByteString -> Either SerialiseAsRawBytesError (Credential 'DRepRole)

SerialiseAsRawBytes (Credential 'HotCommitteeRole)

Instance details

Methods

serialiseToRawBytes :: Credential 'HotCommitteeRole -> ByteString

deserialiseFromRawBytes :: AsType (Credential 'HotCommitteeRole) -> ByteString -> Either SerialiseAsRawBytesError (Credential 'HotCommitteeRole)

ToJSON (OneEraHash xs)

Instance details

Methods

toJSON :: OneEraHash xs -> Value

toEncoding :: OneEraHash xs -> Encoding

toJSONList :: [OneEraHash xs] -> Value

toEncodingList :: [OneEraHash xs] -> Encoding

omitField :: OneEraHash xs -> Bool

ToJSON (HeaderHash blk) => ToJSON (Tip blk)

Instance details

Methods

toJSON :: Tip blk -> Value

toEncoding :: Tip blk -> Encoding

toJSONList :: [Tip blk] -> Value

toEncodingList :: [Tip blk] -> Encoding

omitField :: Tip blk -> Bool