Cardano.Api.Internal.Query

Description

Queries from local clients to the node.

Queries

data QueryInMode result where

Constructors

QueryCurrentEra :: QueryInMode AnyCardanoEra

QueryInEra :: forall era result1. QueryInEra era result1 -> QueryInMode (Either EraMismatch result1)

QueryEraHistory :: QueryInMode EraHistory

QuerySystemStart :: QueryInMode SystemStart

QueryChainBlockNo :: QueryInMode (WithOrigin BlockNo)

QueryChainPoint :: QueryInMode ChainPoint

QueryLedgerConfig :: QueryInMode (HardForkLedgerConfig (CardanoEras StandardCrypto))

data QueryInEra era result where

Constructors

QueryByronUpdateState :: QueryInEra ByronEra ByronUpdateState

QueryInShelleyBasedEra :: forall era result. ShelleyBasedEra era -> QueryInShelleyBasedEra era result -> QueryInEra era result

data QueryInShelleyBasedEra era result where

Constructors

QueryEpoch :: forall era. QueryInShelleyBasedEra era EpochNo

QueryGenesisParameters :: forall era. QueryInShelleyBasedEra era (GenesisParameters ShelleyEra)

QueryProtocolParameters :: forall era. QueryInShelleyBasedEra era (PParams (ShelleyLedgerEra era))

QueryStakeDistribution :: forall era. QueryInShelleyBasedEra era (Map (Hash StakePoolKey) Rational)

QueryUTxO :: forall era. QueryUTxOFilter -> QueryInShelleyBasedEra era (UTxO era)

QueryStakeAddresses :: forall era. Set StakeCredential -> NetworkId -> QueryInShelleyBasedEra era (Map StakeAddress Coin, Map StakeAddress PoolId)

QueryStakePools :: forall era. QueryInShelleyBasedEra era (Set PoolId)

QueryStakePoolParameters :: forall era. Set PoolId -> QueryInShelleyBasedEra era (Map PoolId StakePoolParameters)

QueryDebugLedgerState :: forall era. QueryInShelleyBasedEra era (SerialisedDebugLedgerState era)

QueryProtocolState :: forall era. QueryInShelleyBasedEra era (ProtocolState era)

QueryCurrentEpochState :: forall era. QueryInShelleyBasedEra era (SerialisedCurrentEpochState era)

QueryPoolState :: forall era. Maybe (Set PoolId) -> QueryInShelleyBasedEra era (SerialisedPoolState era)

QueryPoolDistribution :: forall era. Maybe (Set PoolId) -> QueryInShelleyBasedEra era (SerialisedPoolDistribution era)

QueryStakeSnapshot :: forall era. Maybe (Set PoolId) -> QueryInShelleyBasedEra era (SerialisedStakeSnapshots era)

QueryStakeDelegDeposits :: forall era. Set StakeCredential -> QueryInShelleyBasedEra era (Map StakeCredential Coin)

QueryAccountState :: forall era. QueryInShelleyBasedEra era AccountState

QueryConstitution :: forall era. QueryInShelleyBasedEra era (Constitution (ShelleyLedgerEra era))

QueryGovState :: forall era. QueryInShelleyBasedEra era (GovState (ShelleyLedgerEra era))

QueryRatifyState :: forall era. QueryInShelleyBasedEra era (RatifyState (ShelleyLedgerEra era))

QueryFuturePParams :: forall era. QueryInShelleyBasedEra era (Maybe (PParams (ShelleyLedgerEra era)))

QueryDRepState :: forall era. Set (Credential 'DRepRole) -> QueryInShelleyBasedEra era (Map (Credential 'DRepRole) DRepState)

QueryDRepStakeDistr :: forall era. Set DRep -> QueryInShelleyBasedEra era (Map DRep Coin)

QuerySPOStakeDistr :: forall era. Set (KeyHash 'StakePool) -> QueryInShelleyBasedEra era (Map (KeyHash 'StakePool) Coin)

QueryCommitteeMembersState :: forall era. Set (Credential 'ColdCommitteeRole) -> Set (Credential 'HotCommitteeRole) -> Set MemberStatus -> QueryInShelleyBasedEra era CommitteeMembersState

QueryStakeVoteDelegatees :: forall era. Set StakeCredential -> QueryInShelleyBasedEra era (Map StakeCredential DRep)

QueryProposals :: forall era. Set GovActionId -> QueryInShelleyBasedEra era (Seq (GovActionState (ShelleyLedgerEra era)))

QueryLedgerPeerSnapshot :: forall era. QueryInShelleyBasedEra era (Serialised LedgerPeerSnapshot)

QueryStakePoolDefaultVote :: forall era. KeyHash 'StakePool -> QueryInShelleyBasedEra era DefaultVote

data QueryUTxOFilter

Getting the whole UTxO is obviously not efficient since the result can be huge. Filtering by address is also not efficient because it requires a linear search.

The QueryUTxOFilterByTxIn is efficient since it fits with the structure of the UTxO (which is indexed by TxIn).

Constructors

QueryUTxOWhole
O(n) time and space for utxo size n

QueryUTxOByAddress (Set AddressAny)
O(n) time, O(m) space for utxo size n, and address set size m

QueryUTxOByTxIn (Set TxIn)
O(m log n) time, O(m) space for utxo size n, and address set size m

data UTxOInAnyEra where

Constructors

UTxOInAnyEra :: forall era. CardanoEra era -> UTxO era -> UTxOInAnyEra

Internal conversion functions

toConsensusQuery :: CardanoBlock StandardCrypto ~ block => QueryInMode result -> Some (Query block)

fromConsensusQueryResult :: (HasCallStack, CardanoBlock StandardCrypto ~ block) => QueryInMode result -> Query block result' -> result' -> result

Wrapper types used in queries

newtype SerialisedDebugLedgerState era

Constructors

SerialisedDebugLedgerState (Serialised (NewEpochState (ShelleyLedgerEra era)))

newtype ProtocolState era

Constructors

ProtocolState (Serialised (ChainDepState (ConsensusProtocol era)))

decodeProtocolState :: FromCBOR (ChainDepState (ConsensusProtocol era)) => ProtocolState era -> Either (ByteString, DecoderError) (ChainDepState (ConsensusProtocol era))

newtype DebugLedgerState era

Constructors

DebugLedgerState

Fields

unDebugLedgerState :: NewEpochState (ShelleyLedgerEra era)

decodeDebugLedgerState :: FromCBOR (DebugLedgerState era) => SerialisedDebugLedgerState era -> Either (ByteString, DecoderError) (DebugLedgerState era)

newtype SerialisedCurrentEpochState era

Constructors

SerialisedCurrentEpochState (Serialised (EpochState (ShelleyLedgerEra era)))

newtype CurrentEpochState era

Constructors

CurrentEpochState (EpochState (ShelleyLedgerEra era))

decodeCurrentEpochState :: ShelleyBasedEra era -> SerialisedCurrentEpochState era -> Either DecoderError (CurrentEpochState era)

newtype SerialisedPoolState era

Constructors

SerialisedPoolState (Serialised (PState (ShelleyLedgerEra era)))

newtype PoolState era

Constructors

PoolState (PState (ShelleyLedgerEra era))

decodePoolState :: (Era (ShelleyLedgerEra era), DecCBOR (PState (ShelleyLedgerEra era))) => SerialisedPoolState era -> Either DecoderError (PoolState era)

newtype SerialisedPoolDistribution era

Constructors

SerialisedPoolDistribution (Serialised (PoolDistr StandardCrypto))

newtype PoolDistribution era

Constructors

PoolDistribution

Fields

unPoolDistr :: PoolDistr StandardCrypto

decodePoolDistribution :: ShelleyBasedEra era -> SerialisedPoolDistribution era -> Either DecoderError (PoolDistribution era)

newtype SerialisedStakeSnapshots era

Constructors

SerialisedStakeSnapshots (Serialised StakeSnapshots)

newtype StakeSnapshot era

Constructors

StakeSnapshot StakeSnapshots

decodeStakeSnapshot :: SerialisedStakeSnapshots era -> Either DecoderError (StakeSnapshot era)

data EraHistory where

Constructors

EraHistory :: forall (xs :: [Type]). CardanoBlock StandardCrypto ~ HardForkBlock xs => Interpreter xs -> EraHistory

newtype SystemStart

System start

Slots are counted from the system start.

Constructors

SystemStart

Fields

getSystemStart :: UTCTime

newtype LedgerEpochInfo

Constructors

LedgerEpochInfo

Fields

unLedgerEpochInfo :: EpochInfo (Either Text)

toLedgerEpochInfo :: EraHistory -> LedgerEpochInfo

newtype SlotsInEpoch

Constructors

SlotsInEpoch Word64

newtype SlotsToEpochEnd

Constructors

SlotsToEpochEnd Word64

slotToEpoch :: SlotNo -> EraHistory -> Either PastHorizonException (EpochNo, SlotsInEpoch, SlotsToEpochEnd)

data family LedgerState blk (mk :: MapKind)

Ledger state associated with a block

This is the Consensus notion of a Ledger ledger state. Each block type is associated with one of the Ledger types for the ledger state. Virtually every concept in this codebase revolves around this type, or the referenced blk. Whenever we use the type variable l we intend to signal that the expected instantiation is either a LedgerState or some wrapper over it (like the ExtLedgerState).

This type is parametrized over mk :: MapKind to express the LedgerTables contained in such a LedgerState. See LedgerTables for a more thorough description.

The main operations we can do with a LedgerState are ticking (defined in IsLedger), and applying a block (defined in ApplyBlock).

getProgress :: SlotNo -> EraHistory -> Either PastHorizonException (RelativeTime, SlotLength)

getSlotForRelativeTime :: RelativeTime -> EraHistory -> Either PastHorizonException SlotNo

Returns the slot number for provided relative time from SystemStart

decodeBigLedgerPeerSnapshot :: Serialised LedgerPeerSnapshot -> Either (ByteString, DecoderError) LedgerPeerSnapshot

Internal conversion functions

toLedgerUTxO :: ShelleyBasedEra era -> UTxO era -> UTxO (ShelleyLedgerEra era)

fromLedgerUTxO :: ShelleyBasedEra era -> UTxO (ShelleyLedgerEra era) -> UTxO era