Cardano.Api.Internal.Tx.Body

Creating transactions using the old API

Both the old and new APIs support transaction creation. Transactions can be converted between formats, as they share the same underlying representation. cardano-api will be moving towards using the new API and deprecating the old way to ensure simplicity, closer alignment with the ledger, and easier maintenance.

In both the new and old APIs, to construct a transaction, you need to construct a TxBodyContent, and you will need at least a witness (for example, a ShelleyWitnessSigningKey), to sign the transaction. This process remains unchanged.

To learn how to create a transaction using the new API, refer to Cardano.Api.Internal.Experimental.Tx documentation.

The next examples use the following qualified modules:

import qualified Cardano.Api as Api                -- the general `cardano-api` exports (including the old API)
import qualified Cardano.Api.Script as Script      -- types related to scripts (Plutus and native)
import qualified Cardano.Api.Ledger as Ledger      -- cardano-ledger re-exports

Creating a TxBodyContent

To create a transaction, you first need to define the contents of its body. This section will show how to use the API to create a TxBodyContent for a simple transaction.

TxBodyContent datatype provides several fields because transactions can serve multiple purposes, but the function defaultTxBodyContent (exported from Cardano.Api) already provides a base TxBodyContent with all fields set to their default values that you can use as a starting point so as not to have to set all fields manually.

The defaultTxBodyContent takes, as the only parameter, the ShelleyBasedEra witness for the era you are working with. For example, if you are working with the ConwayEra, use shelleyBasedEra available in Cardano.Api, as follows:

let sbe :: Api.ShelleyBasedEra Api.ConwayEra = Api.shelleyBasedEra

This is what creating a simple TxBodyContent would look like.

First, choose a transaction output to spend (a UTXO). Specify which UTXO to spend by providing the transaction ID and the index of the output in that transaction that you want to spend.

To specify the transaction ID, you can use the deserialiseFromRawBytesHex function on the hexadecimal representation of the transaction hash. For example:

let (Right srcTxId) = Api.deserialiseFromRawBytesHex Api.AsTxId "be6efd42a3d7b9a00d09d77a5d41e55ceaf0bd093a8aa8a893ce70d9caafd978"

In real implementations, failure cases should be handled appropriately.

To specify the transaction index, use the TxIx constructor. For example:

let srcTxIx = Api.TxIx 0

Now, combine both to create a TxIn value and pair it with a witness requirement using BuildTxWith :

let txIn = ( Api.TxIn srcTxId srcTxIx
           , Api.BuildTxWith (Api.KeyWitness Api.KeyWitnessForSpending)
           )

Next, specify the address of the recipient of the transaction. If you have the bech32 representation, you can use the deserialiseAddress function to get the AddressInEra type. For example:

let (Just destAddress) = Api.deserialiseAddress (Api.AsAddressInEra eraAsType) "addr_test1vzpfxhjyjdlgk5c0xt8xw26avqxs52rtf69993j4tajehpcue4v2v"

Now, you can create a TxOut value. For simplicity, assume the output is a simple payment output of 10 ada, with no datum or reference script attached to it:

let txOut = Api.TxOut
              destAddress
              (Api.lovelaceToTxOutValue sbe 10_000_000)
              Api.TxOutDatumNone
              Script.ReferenceScriptNone

Note to set the transaction fee. For example, set it to 2 ada:

let txFee = Api.TxFeeExplicit sbe 2_000_000

Finally, you can create the TxBodyContent by using the defaultTxBodyContent function and putting everything together:

let txBodyContent = Api.defaultTxBodyContent sbe
                     & Api.setTxIns [txIn]
                     & Api.setTxOuts [txOut]
                     & Api.setTxFee txFee

The txBodyContent can now be used to create a transaction using the old or the new API.

Balancing a transaction

If you have a UTXO with exactly 12 ada, you could just construct the transaction as in the previous section directly, and it would be a valid transaction, but:

It is probably wasting ADA

There may not be exactly one UTXO of 12 ada

The transaciton may not be this simple.

For these reasons, it is recommended that you balance the transaction before proceeding with signing and submitting.

See how to balance a transaction in the Cardano.Api.Internal.Fees documentation.

Creating a ShelleyWitnessSigningKey

Signing a transaction requires a witness, for example, a ShelleyWitnessSigningKey.

Learn how to create a ShelleyWitnessSigningKey in the Cardano.Api.Internal.Tx.Sign documentation.

Creating a transaction using the old API

Now that you have a TxBodyContent and a ShelleyWitnessSigningKey, you can easily create a transaction using the old API. First, create a transaction body using the createTransactionBody function and the ShelleyBasedEra witness defined earlier.

Create the transaction body using the TransactionBodyContent created earlier:

let (Right txBody) = Api.createTransactionBody sbe txBodyContent

Then, sign the transaction using the signShelleyTransaction function and the witness:

let oldApiSignedTx :: Api.Tx Api.ConwayEra = Api.signShelleyTransaction sbe txBody [witness]

We now have a signed transaction.

Inspecting transactions

To deconstruct an old-style TxBody into a TxBodyContent, you can also use the TxBody pattern. Note that this cannot be used for constructing. For that, use ShelleyTxBody or createTransactionBody, as in the example.

To extract the TxBody and the KeyWitnesses from an old-style Tx, use the functions getTxBody and getTxWitnesses respectively, from Cardano.Api.

Appendix: Getting Shelley-based era witness from the new API

If you are using the new API, you can also derive the ShelleyBasedEra of it from ConwayEra from Cardano.Api.Internal.Experimental using the convert function:

let era = Exp.ConwayEra
let sbe = Api.convert era

Contents

parseTxId :: Parser TxId

Transaction bodies

data TxBody era where

Constructors

ShelleyTxBody

Fields

:: forall era. ShelleyBasedEra era

-> TxBody (ShelleyLedgerEra era)

-> [Script (ShelleyLedgerEra era)]

-> TxBodyScriptData era

-> Maybe (TxAuxData (ShelleyLedgerEra era))

-> TxScriptValidity era

Mark script as expected to pass or fail validation

-> TxBody era

Bundled Patterns

pattern TxBody :: TxBodyContent ViewTx era -> TxBody era

Deprecated: Use getTxBodyContent $ getTxBody instead

createTransactionBody :: HasCallStack => ShelleyBasedEra era -> TxBodyContent BuildTx era -> Either TxBodyError (TxBody era)

createAndValidateTransactionBody :: ShelleyBasedEra era -> TxBodyContent BuildTx era -> Either TxBodyError (TxBody era)

Deprecated: Use createTransactionBody instead

data TxBodyContent build era

Constructors

TxBodyContent

Fields

txIns :: TxIns build era

txInsCollateral :: TxInsCollateral era

txInsReference :: TxInsReference era

txOuts :: [TxOut CtxTx era]

txTotalCollateral :: TxTotalCollateral era

txReturnCollateral :: TxReturnCollateral CtxTx era

txFee :: TxFee era

txValidityLowerBound :: TxValidityLowerBound era

txValidityUpperBound :: TxValidityUpperBound era

txMetadata :: TxMetadataInEra era

txAuxScripts :: TxAuxScripts era

txExtraKeyWits :: TxExtraKeyWitnesses era

txProtocolParams :: BuildTxWith build (Maybe (LedgerProtocolParameters era))

txWithdrawals :: TxWithdrawals build era

txCertificates :: TxCertificates build era

txUpdateProposal :: TxUpdateProposal era

txMintValue :: TxMintValue build era

txScriptValidity :: TxScriptValidity era

txProposalProcedures :: Maybe (Featured ConwayEraOnwards era (TxProposalProcedures build era))

txVotingProcedures :: Maybe (Featured ConwayEraOnwards era (TxVotingProcedures build era))

txCurrentTreasuryValue :: Maybe (Featured ConwayEraOnwards era (Maybe Coin))

Current treasury value

txTreasuryDonation :: Maybe (Featured ConwayEraOnwards era Coin)

Treasury donation to perform

Byron only

makeByronTransactionBody :: TxIns BuildTx ByronEra -> [TxOut CtxTx ByronEra] -> Either TxBodyError (Annotated Tx ByteString)

Transaction body builders

defaultTxBodyContent :: ShelleyBasedEra era -> TxBodyContent BuildTx era

defaultTxFee :: ShelleyBasedEra era -> TxFee era

defaultTxValidityUpperBound :: ShelleyBasedEra era -> TxValidityUpperBound era

setTxIns :: TxIns build era -> TxBodyContent build era -> TxBodyContent build era

modTxIns :: (TxIns build era -> TxIns build era) -> TxBodyContent build era -> TxBodyContent build era

addTxIn :: (TxIn, BuildTxWith build (Witness WitCtxTxIn era)) -> TxBodyContent build era -> TxBodyContent build era

addTxIns :: TxIns build era -> TxBodyContent build era -> TxBodyContent build era

setTxInsCollateral :: TxInsCollateral era -> TxBodyContent build era -> TxBodyContent build era

modTxInsCollateral :: (TxInsCollateral era -> TxInsCollateral era) -> TxBodyContent build era -> TxBodyContent build era

addTxInsCollateral :: IsAlonzoBasedEra era => [TxIn] -> TxBodyContent build era -> TxBodyContent build era

addTxInCollateral :: IsAlonzoBasedEra era => TxIn -> TxBodyContent build era -> TxBodyContent build era

setTxInsReference :: TxInsReference era -> TxBodyContent build era -> TxBodyContent build era

modTxInsReference :: (TxInsReference era -> TxInsReference era) -> TxBodyContent build era -> TxBodyContent build era

addTxInsReference :: IsBabbageBasedEra era => [TxIn] -> TxBodyContent build era -> TxBodyContent build era

addTxInReference :: IsBabbageBasedEra era => TxIn -> TxBodyContent build era -> TxBodyContent build era

setTxOuts :: [TxOut CtxTx era] -> TxBodyContent build era -> TxBodyContent build era

modTxOuts :: ([TxOut CtxTx era] -> [TxOut CtxTx era]) -> TxBodyContent build era -> TxBodyContent build era

addTxOut :: TxOut CtxTx era -> TxBodyContent build era -> TxBodyContent build era

addTxOuts :: [TxOut CtxTx era] -> TxBodyContent build era -> TxBodyContent build era

setTxTotalCollateral :: TxTotalCollateral era -> TxBodyContent build era -> TxBodyContent build era

modTxTotalCollateral :: (TxTotalCollateral era -> TxTotalCollateral era) -> TxBodyContent build era -> TxBodyContent build era

setTxReturnCollateral :: TxReturnCollateral CtxTx era -> TxBodyContent build era -> TxBodyContent build era

modTxReturnCollateral :: (TxReturnCollateral CtxTx era -> TxReturnCollateral CtxTx era) -> TxBodyContent build era -> TxBodyContent build era

setTxFee :: TxFee era -> TxBodyContent build era -> TxBodyContent build era

modTxFee :: (TxFee era -> TxFee era) -> TxBodyContent build era -> TxBodyContent build era

setTxValidityLowerBound :: TxValidityLowerBound era -> TxBodyContent build era -> TxBodyContent build era

modTxValidityLowerBound :: (TxValidityLowerBound era -> TxValidityLowerBound era) -> TxBodyContent build era -> TxBodyContent build era

setTxValidityUpperBound :: TxValidityUpperBound era -> TxBodyContent build era -> TxBodyContent build era

modTxValidityUpperBound :: (TxValidityUpperBound era -> TxValidityUpperBound era) -> TxBodyContent build era -> TxBodyContent build era

setTxMetadata :: TxMetadataInEra era -> TxBodyContent build era -> TxBodyContent build era

modTxMetadata :: (TxMetadataInEra era -> TxMetadataInEra era) -> TxBodyContent build era -> TxBodyContent build era

setTxAuxScripts :: TxAuxScripts era -> TxBodyContent build era -> TxBodyContent build era

modTxAuxScripts :: (TxAuxScripts era -> TxAuxScripts era) -> TxBodyContent build era -> TxBodyContent build era

setTxExtraKeyWits :: TxExtraKeyWitnesses era -> TxBodyContent build era -> TxBodyContent build era

modTxExtraKeyWits :: (TxExtraKeyWitnesses era -> TxExtraKeyWitnesses era) -> TxBodyContent build era -> TxBodyContent build era

addTxExtraKeyWits :: IsAlonzoBasedEra era => [Hash PaymentKey] -> TxBodyContent build era -> TxBodyContent build era

setTxProtocolParams :: BuildTxWith build (Maybe (LedgerProtocolParameters era)) -> TxBodyContent build era -> TxBodyContent build era

setTxWithdrawals :: TxWithdrawals build era -> TxBodyContent build era -> TxBodyContent build era

modTxWithdrawals :: (TxWithdrawals build era -> TxWithdrawals build era) -> TxBodyContent build era -> TxBodyContent build era

setTxCertificates :: TxCertificates build era -> TxBodyContent build era -> TxBodyContent build era

modTxCertificates :: (TxCertificates build era -> TxCertificates build era) -> TxBodyContent build era -> TxBodyContent build era

setTxUpdateProposal :: TxUpdateProposal era -> TxBodyContent build era -> TxBodyContent build era

modTxUpdateProposal :: (TxUpdateProposal era -> TxUpdateProposal era) -> TxBodyContent build era -> TxBodyContent build era

setTxProposalProcedures :: Maybe (Featured ConwayEraOnwards era (TxProposalProcedures build era)) -> TxBodyContent build era -> TxBodyContent build era

setTxVotingProcedures :: Maybe (Featured ConwayEraOnwards era (TxVotingProcedures build era)) -> TxBodyContent build era -> TxBodyContent build era

setTxMintValue :: TxMintValue build era -> TxBodyContent build era -> TxBodyContent build era

modTxMintValue :: (TxMintValue build era -> TxMintValue build era) -> TxBodyContent build era -> TxBodyContent build era

addTxMintValue :: IsMaryBasedEra era => Map PolicyId [(AssetName, Quantity, BuildTxWith build (ScriptWitness WitCtxMint era))] -> TxBodyContent build era -> TxBodyContent build era

subtractTxMintValue :: IsMaryBasedEra era => Map PolicyId [(AssetName, Quantity, BuildTxWith build (ScriptWitness WitCtxMint era))] -> TxBodyContent build era -> TxBodyContent build era

Adds the negation of the provided assets and quantities to the txMintValue field of the TxBodyContent.

setTxScriptValidity :: TxScriptValidity era -> TxBodyContent build era -> TxBodyContent build era

modTxScriptValidity :: (TxScriptValidity era -> TxScriptValidity era) -> TxBodyContent build era -> TxBodyContent build era

setTxCurrentTreasuryValue :: Maybe (Featured ConwayEraOnwards era (Maybe Coin)) -> TxBodyContent build era -> TxBodyContent build era

setTxTreasuryDonation :: Maybe (Featured ConwayEraOnwards era Coin) -> TxBodyContent build era -> TxBodyContent build era

data TxBodyError

Constructors

TxBodyEmptyTxIns

TxBodyEmptyTxInsCollateral

TxBodyEmptyTxOuts

TxBodyOutputNegative !Quantity !TxOutInAnyEra

TxBodyOutputOverflow !Quantity !TxOutInAnyEra

TxBodyMetadataError ![(Word64, TxMetadataRangeError)]

TxBodyInIxOverflow !TxIn

TxBodyMissingProtocolParams

TxBodyProtocolParamsConversionError !ProtocolParametersConversionError

data TxBodyScriptData era where

Constructors

TxBodyNoScriptData :: forall era. TxBodyScriptData era

TxBodyScriptData :: forall era. AlonzoEraOnwardsConstraints era => AlonzoEraOnwards era -> TxDats (ShelleyLedgerEra era) -> Redeemers (ShelleyLedgerEra era) -> TxBodyScriptData era

data TxScriptValidity era where

A representation of whether the era supports tx script validity.

The Alonzo and subsequent eras support script validity.

Constructors

TxScriptValidityNone :: forall era. TxScriptValidity era

TxScriptValidity :: forall era. AlonzoEraOnwards era -> ScriptValidity -> TxScriptValidity era
Tx script validity is supported in transactions in the Alonzo era onwards.

data ScriptValidity

Indicates whether a script is expected to fail or pass validation.

Constructors

ScriptInvalid
Script is expected to fail validation. Transactions marked as such can include scripts that fail validation. Such transactions may be submitted to the chain, in which case the collateral will be taken upon on chain script validation failure.

ScriptValid
Script is expected to pass validation. Transactions marked as such cannot include scripts that fail validation.

scriptValidityToIsValid :: ScriptValidity -> IsValid

isValidToScriptValidity :: IsValid -> ScriptValidity

txScriptValidityToIsValid :: TxScriptValidity era -> IsValid

txScriptValidityToScriptValidity :: TxScriptValidity era -> ScriptValidity

Transaction Ids

newtype TxId

Constructors

TxId (Hash StandardCrypto EraIndependentTxBody)

getTxId :: TxBody era -> TxId

Calculate the transaction identifier for a TxBody.

getTxIdByron :: ATxAux ByteString -> TxId

getTxIdShelley :: (EraCrypto (ShelleyLedgerEra era) ~ StandardCrypto, EraTxBody (ShelleyLedgerEra era)) => ShelleyBasedEra era -> TxBody (ShelleyLedgerEra era) -> TxId

Transaction inputs

data TxIn

Constructors

TxIn TxId TxIx

type TxIns build era = [(TxIn, BuildTxWith build (Witness WitCtxTxIn era))]

indexTxIns :: TxIns BuildTx era -> [(ScriptWitnessIndex, TxIn, Witness WitCtxTxIn era)]

Index transaction inputs ordered by TxIn Please note that the result can contain also KeyWitnesses. See section 4.1 of https://github.com/intersectmbo/cardano-ledger/releases/latest/download/alonzo-ledger.pdf

newtype TxIx

Constructors

TxIx Word

genesisUTxOPseudoTxIn :: NetworkId -> Hash GenesisUTxOKey -> TxIn

Compute the TxIn of the initial UTxO pseudo-transaction corresponding to the given address in the genesis initial funds.

The Shelley initial UTxO is constructed from the sgInitialFunds which is not a full UTxO but just a map from addresses to coin values.

This gets turned into a UTxO by making a pseudo-transaction for each address, with the 0th output being the coin value. So to spend from the initial UTxO we need this same TxIn to use as an input to the spending transaction.

getReferenceInputsSizeForTxIds :: ShelleyLedgerEra era ~ ledgerera => BabbageEraOnwards era -> UTxO ledgerera -> Set TxIn -> Int

Calculate the reference inputs size in bytes for provided set of transaction IDs and UTXOs.

Transaction outputs

data CtxTx

The context is a transaction body

data CtxUTxO

The context is the UTxO

data TxOut ctx era

Constructors

TxOut (AddressInEra era) (TxOutValue era) (TxOutDatum ctx era) (ReferenceScript era)

data TxOutValue era where

Constructors

TxOutValueByron :: forall era. Coin -> TxOutValue era

TxOutValueShelleyBased :: forall era. (Eq (Value (ShelleyLedgerEra era)), Show (Value (ShelleyLedgerEra era))) => ShelleyBasedEra era -> Value (ShelleyLedgerEra era) -> TxOutValue era

data TxOutDatum ctx era where

Constructors

TxOutDatumNone :: forall ctx era. TxOutDatum ctx era

TxOutDatumHash :: forall era ctx. AlonzoEraOnwards era -> Hash ScriptData -> TxOutDatum ctx era
A transaction output that only specifies the hash of the datum, but not the full datum value.

TxOutSupplementalDatum :: forall era. AlonzoEraOnwards era -> HashableScriptData -> TxOutDatum CtxTx era
A transaction output that specifies the whole datum value. This can only be used in the context of the transaction body (i.e this is a supplemental datum), and does not occur in the UTxO. The UTxO only contains the datum hash.

TxOutDatumInline :: forall era ctx. BabbageEraOnwards era -> HashableScriptData -> TxOutDatum ctx era
A transaction output that specifies the whole datum instead of the datum hash. Note that the datum map will not be updated with this datum, it only exists at the transaction output.

toCtxUTxOTxOut :: TxOut CtxTx era -> TxOut CtxUTxO era

lovelaceToTxOutValue :: ShelleyBasedEra era -> Coin -> TxOutValue era

prettyRenderTxOut :: TxOutInAnyEra -> Text

txOutValueToLovelace :: TxOutValue era -> Coin

txOutValueToValue :: TxOutValue era -> Value

parseHash :: SerialiseAsRawBytes (Hash a) => AsType (Hash a) -> Parser (Hash a)

data TxOutInAnyEra where

Constructors

TxOutInAnyEra :: forall era. CardanoEra era -> TxOut CtxTx era -> TxOutInAnyEra

txOutInAnyEra :: CardanoEra era -> TxOut CtxTx era -> TxOutInAnyEra

Convenience constructor for TxOutInAnyEra

Other transaction body types

data TxInsCollateral era where

Constructors

TxInsCollateralNone :: forall era. TxInsCollateral era

TxInsCollateral :: forall era. AlonzoEraOnwards era -> [TxIn] -> TxInsCollateral era

data TxInsReference era where

Constructors

TxInsReferenceNone :: forall era. TxInsReference era

TxInsReference :: forall era. BabbageEraOnwards era -> [TxIn] -> TxInsReference era

data TxReturnCollateral ctx era where

Constructors

TxReturnCollateralNone :: forall ctx era. TxReturnCollateral ctx era

TxReturnCollateral :: forall era ctx. BabbageEraOnwards era -> TxOut ctx era -> TxReturnCollateral ctx era

data TxTotalCollateral era where

Constructors

TxTotalCollateralNone :: forall era. TxTotalCollateral era

TxTotalCollateral :: forall era. BabbageEraOnwards era -> Coin -> TxTotalCollateral era

data TxFee era where

Constructors

TxFeeExplicit :: forall era. ShelleyBasedEra era -> Coin -> TxFee era

data TxValidityLowerBound era where

Constructors

TxValidityNoLowerBound :: forall era. TxValidityLowerBound era

TxValidityLowerBound :: forall era. AllegraEraOnwards era -> SlotNo -> TxValidityLowerBound era

data TxValidityUpperBound era where

This was formerly known as the TTL.

Constructors

TxValidityUpperBound :: forall era. ShelleyBasedEra era -> Maybe SlotNo -> TxValidityUpperBound era

data TxMetadataInEra era where

Constructors

TxMetadataNone :: forall era. TxMetadataInEra era

TxMetadataInEra :: forall era. ShelleyBasedEra era -> TxMetadata -> TxMetadataInEra era

data TxAuxScripts era where

Constructors

TxAuxScriptsNone :: forall era. TxAuxScripts era

TxAuxScripts :: forall era. AllegraEraOnwards era -> [ScriptInEra era] -> TxAuxScripts era

data TxExtraKeyWitnesses era where

Constructors

TxExtraKeyWitnessesNone :: forall era. TxExtraKeyWitnesses era

TxExtraKeyWitnesses :: forall era. AlonzoEraOnwards era -> [Hash PaymentKey] -> TxExtraKeyWitnesses era

data TxWithdrawals build era where

Constructors

TxWithdrawalsNone :: forall build era. TxWithdrawals build era

TxWithdrawals :: forall era build. ShelleyBasedEra era -> [(StakeAddress, Coin, BuildTxWith build (Witness WitCtxStake era))] -> TxWithdrawals build era

indexTxWithdrawals :: TxWithdrawals BuildTx era -> [(ScriptWitnessIndex, StakeAddress, Coin, Witness WitCtxStake era)]

Index the withdrawals with witnesses in the order of stake addresses. See section 4.1 of https://github.com/intersectmbo/cardano-ledger/releases/latest/download/alonzo-ledger.pdf

data TxCertificates build era where

Constructors

TxCertificatesNone :: forall build era. TxCertificates build era
No certificates

TxCertificates :: forall era build. ShelleyBasedEra era -> OMap (Certificate era) (BuildTxWith build (Maybe (StakeCredential, Witness WitCtxStake era))) -> TxCertificates build era
Represents certificates present in transaction. Prefer using mkTxCertificates to constructing this type with a constructor

mkTxCertificates :: Applicative (BuildTxWith build) => ShelleyBasedEra era -> [(Certificate era, Maybe (ScriptWitness WitCtxStake era))] -> TxCertificates build era

Create TxCertificates. Note that 'Certificate era' will be deduplicated. Only Certificates with a stake credential will be in the result.

Note that, when building a transaction in Conway era, a witness is not required for staking credential registration, but this is only the case during the transitional period of Conway era and only for staking credential registration certificates without a deposit. Future eras will require a witness for registration certificates, because the one without a deposit will be removed.

indexTxCertificates :: TxCertificates BuildTx era -> [(ScriptWitnessIndex, Certificate era, StakeCredential, Witness WitCtxStake era)]

Index certificates with witnesses by the order they appear in the list (in the transaction). See section 4.1 of https://github.com/intersectmbo/cardano-ledger/releases/latest/download/alonzo-ledger.pdf

data TxUpdateProposal era where

Constructors

TxUpdateProposalNone :: forall era. TxUpdateProposal era

TxUpdateProposal :: forall era. ShelleyToBabbageEra era -> UpdateProposal -> TxUpdateProposal era

data TxMintValue build era where

Constructors

TxMintNone :: forall build era. TxMintValue build era

TxMintValue :: forall era build. MaryEraOnwards era -> Map PolicyId [(AssetName, Quantity, BuildTxWith build (ScriptWitness WitCtxMint era))] -> TxMintValue build era

txMintValueToValue :: TxMintValue build era -> Value

Convert TxMintValue to a more handy Value.

indexTxMintValue :: TxMintValue build era -> [(ScriptWitnessIndex, PolicyId, AssetName, Quantity, BuildTxWith build (ScriptWitness WitCtxMint era))]

Index the assets with witnesses in the order of policy ids. See section 4.1 of https://github.com/intersectmbo/cardano-ledger/releases/latest/download/alonzo-ledger.pdf

data TxVotingProcedures build era where

Constructors

TxVotingProceduresNone :: forall build era. TxVotingProcedures build era

TxVotingProcedures :: forall era build. VotingProcedures (ShelleyLedgerEra era) -> BuildTxWith build (Map (Voter (EraCrypto (ShelleyLedgerEra era))) (ScriptWitness WitCtxStake era)) -> TxVotingProcedures build era

mkTxVotingProcedures :: Applicative (BuildTxWith build) => [(VotingProcedures era, Maybe (ScriptWitness WitCtxStake era))] -> Either (VotesMergingConflict era) (TxVotingProcedures build era)

Create voting procedures from map of voting procedures and optional witnesses. Validates the function argument, to make sure the list of votes is legal. See mergeVotingProcedures for validation rules.

indexTxVotingProcedures :: TxVotingProcedures BuildTx era -> [(ScriptWitnessIndex, Voter (EraCrypto (ShelleyLedgerEra era)), ScriptWitness WitCtxStake era)]

Index voting procedures by the order of the votes (Ord).

data TxProposalProcedures build era where

Constructors

TxProposalProceduresNone :: forall build era. TxProposalProcedures build era
No proposals in transaction..

TxProposalProcedures :: forall era build. EraPParams (ShelleyLedgerEra era) => OMap (ProposalProcedure (ShelleyLedgerEra era)) (BuildTxWith build (Maybe (ScriptWitness WitCtxStake era))) -> TxProposalProcedures build era
Represents proposal procedures present in transaction.

mkTxProposalProcedures :: forall era build. (Applicative (BuildTxWith build), IsShelleyBasedEra era) => [(ProposalProcedure (ShelleyLedgerEra era), Maybe (ScriptWitness WitCtxStake era))] -> TxProposalProcedures build era

A smart constructor for TxProposalProcedures. It makes sure that the value produced is consistent - the witnessed proposals are also present in the first constructor parameter.

indexTxProposalProcedures :: TxProposalProcedures BuildTx era -> [(ScriptWitnessIndex, ProposalProcedure (ShelleyLedgerEra era), ScriptWitness WitCtxStake era)]

Index proposal procedures by their order (Ord). | and filter out the ones that do not have a witness.

indexWitnessedTxProposalProcedures :: TxProposalProcedures BuildTx era -> [(ProposalProcedure (ShelleyLedgerEra era), Maybe (ScriptWitnessIndex, ScriptWitness WitCtxStake era))]

Index proposal procedures by their order (Ord).

convProposalProcedures :: TxProposalProcedures build era -> OSet (ProposalProcedure (ShelleyLedgerEra era))

Returns an OSet of proposals from TxProposalProcedures.

Building vs viewing transactions

data BuildTxWith build a where

Constructors

ViewTx :: forall a. BuildTxWith ViewTx a

BuildTxWith :: forall a. a -> BuildTxWith BuildTx a

data BuildTx

data ViewTx

buildTxWithToMaybe :: BuildTxWith build a -> Maybe a

Inspecting ScriptWitnesses

data AnyScriptWitness era where

A ScriptWitness in any WitCtx. This lets us handle heterogeneous collections of script witnesses from multiple contexts.

Constructors

AnyScriptWitness :: forall witctx era. Typeable witctx => ScriptWitness witctx era -> AnyScriptWitness era

data ScriptWitnessIndex

Identify the location of a ScriptWitness within the context of a TxBody. These are indexes of the objects within the transaction that need or can use script witnesses: inputs, minted assets, withdrawals and certificates. These are simple numeric indices, enumerated from zero. Thus the indices are not stable if the transaction body is modified.

Constructors

ScriptWitnessIndexTxIn !Word32
The n'th transaction input, in the order of the TxIds.

ScriptWitnessIndexMint !Word32
The n'th minting PolicyId, in the order of the PolicyIds.

ScriptWitnessIndexCertificate !Word32
The n'th certificate, in the list order of the certificates.

ScriptWitnessIndexWithdrawal !Word32
The n'th withdrawal, in the order of the StakeAddresss.

ScriptWitnessIndexVoting !Word32
The n'th vote, in the order of the votes.

ScriptWitnessIndexProposing !Word32
The n'th proposal, in the order of the proposals.

renderScriptWitnessIndex :: ScriptWitnessIndex -> String

collectTxBodyScriptWitnesses :: ShelleyBasedEra era -> TxBodyContent BuildTx era -> [(ScriptWitnessIndex, AnyScriptWitness era)]

toScriptIndex :: AlonzoEraOnwards era -> PlutusPurpose AsIx (ShelleyLedgerEra era) -> ScriptWitnessIndex

Conversion to inline data

scriptDataToInlineDatum :: Era ledgerera => HashableScriptData -> Datum ledgerera

Conversion of ScriptData to binary data which allows for the storage of data onchain within a transaction output.

Internal conversion functions & types

convCertificates :: ShelleyBasedEra era -> TxCertificates build era -> StrictSeq (TxCert (ShelleyLedgerEra era))

convCollateralTxIns :: TxInsCollateral era -> Set (TxIn StandardCrypto)

convExtraKeyWitnesses :: TxExtraKeyWitnesses era -> Set (KeyHash 'Witness StandardCrypto)

convLanguages :: [(ScriptWitnessIndex, AnyScriptWitness era)] -> Set Language

convMintValue :: TxMintValue build era -> MultiAsset StandardCrypto

convReferenceInputs :: TxInsReference era -> Set (TxIn StandardCrypto)

convReturnCollateral :: ShelleyBasedEra era -> TxReturnCollateral ctx era -> StrictMaybe (TxOut (ShelleyLedgerEra era))

convScripts :: ShelleyLedgerEra era ~ ledgerera => [(ScriptWitnessIndex, AnyScriptWitness era)] -> [Script ledgerera]

convScriptData :: ShelleyBasedEra era -> [TxOut CtxTx era] -> [(ScriptWitnessIndex, AnyScriptWitness era)] -> TxBodyScriptData era

convTotalCollateral :: TxTotalCollateral era -> StrictMaybe Coin

convTransactionFee :: ShelleyBasedEra era -> TxFee era -> Coin

convTxIns :: TxIns BuildTx era -> Set (TxIn StandardCrypto)

convTxOuts :: forall ctx era ledgerera. (HasCallStack, ShelleyLedgerEra era ~ ledgerera) => ShelleyBasedEra era -> [TxOut ctx era] -> StrictSeq (TxOut ledgerera)

convTxUpdateProposal

Arguments

:: ShelleyBasedEra era

-> TxUpdateProposal era

-> Either TxBodyError (StrictMaybe (Update (ShelleyLedgerEra era)))
Left when there's protocol params conversion error, Right otherwise, 'Right SNothing' means that there's no update proposal

Convert transaction update proposal into ledger update proposal

convValidityLowerBound :: TxValidityLowerBound era -> Maybe SlotNo

convValidityUpperBound :: ShelleyBasedEra era -> TxValidityUpperBound era -> Maybe SlotNo

convVotingProcedures :: TxVotingProcedures build era -> VotingProcedures (ShelleyLedgerEra era)

convWithdrawals :: TxWithdrawals build era -> Withdrawals StandardCrypto

getScriptIntegrityHash :: BuildTxWith BuildTx (Maybe (LedgerProtocolParameters era)) -> Set Language -> TxBodyScriptData era -> StrictMaybe (ScriptIntegrityHash (EraCrypto (ShelleyLedgerEra era)))

mkCommonTxBody :: HasCallStack => ShelleyBasedEra era -> TxIns BuildTx era -> [TxOut ctx era] -> TxFee era -> TxWithdrawals build era -> Maybe (TxAuxData (ShelleyLedgerEra era)) -> TxBody era

A helper function that constructs a TxBody with all of the fields that are common for all eras

toAuxiliaryData :: ShelleyBasedEra era -> TxMetadataInEra era -> TxAuxScripts era -> Maybe (TxAuxData (ShelleyLedgerEra era))

In the Allegra and Mary eras the auxiliary data consists of the tx metadata and the axiliary scripts. In the Alonzo and later eras the auxiliary data consists of the tx metadata and the axiliary scripts, and the axiliary script data.

toByronTxId :: TxId -> TxId

toShelleyTxId :: TxId -> TxId StandardCrypto

toShelleyTxIn :: TxIn -> TxIn StandardCrypto

This function may overflow on the transaction index. Call sites must ensure that all uses of this function are appropriately guarded.

toShelleyTxOut :: (HasCallStack, ShelleyLedgerEra era ~ ledgerera) => ShelleyBasedEra era -> TxOut CtxUTxO era -> TxOut ledgerera

toShelleyTxOutAny :: forall ctx era ledgerera. (HasCallStack, ShelleyLedgerEra era ~ ledgerera) => ShelleyBasedEra era -> TxOut ctx era -> TxOut ledgerera

A variant of 'toShelleyTxOutAny that is used only internally to this module that works with a TxOut in any context (including CtxTx) by ignoring embedded datums (taking only their hash).

fromShelleyTxId :: TxId StandardCrypto -> TxId

fromShelleyTxIn :: TxIn StandardCrypto -> TxIn

fromShelleyTxOut :: ShelleyBasedEra era -> TxOut (ShelleyLedgerEra era) -> TxOut ctx era

fromByronTxIn :: TxIn -> TxIn

fromLedgerTxOuts :: ShelleyBasedEra era -> TxBody (ShelleyLedgerEra era) -> TxBodyScriptData era -> [TxOut CtxTx era]

renderTxIn :: TxIn -> Text

Misc helpers

calculateExecutionUnitsLovelace :: Prices -> ExecutionUnits -> Maybe Coin

Data family instances

data family AsType t

A family of singleton types used in this API to indicate which type to use where it would otherwise be ambiguous or merely unclear.

Values of this type are passed to deserialisation functions for example.

getTxBodyContent :: TxBody era -> TxBodyContent ViewTx era

validateTxIns :: [(TxIn, BuildTxWith BuildTx (Witness WitCtxTxIn era))] -> Either TxBodyError ()

guardShelleyTxInsOverflow :: [TxIn] -> Either TxBodyError ()

validateTxOuts :: ShelleyBasedEra era -> [TxOut CtxTx era] -> Either TxBodyError ()

validateMetadata :: TxMetadataInEra era -> Either TxBodyError ()

validateTxInsCollateral :: TxInsCollateral era -> Set Language -> Either TxBodyError ()

validateProtocolParameters :: BuildTxWith BuildTx (Maybe (LedgerProtocolParameters era)) -> Set Language -> Either TxBodyError ()