Cardano.Api.Internal.Tx.Sign

Description

Creating complete, signed transactions.

Example: Creating a ShelleyWitnessSigningKey

Signing a transaction requires a witness, for example, a ShelleyWitnessSigningKey.

This example uses the following qualified module:

import qualified Cardano.Api as Api                -- the general `cardano-api` exports (including the old API)

There are several ways of signing a transaction and representing a signing key. If the bech32 representation of the signing key is available, it is possible to use the deserialiseFromBech32 function as follows:

let (Right signingKey) = Api.deserialiseFromBech32 (Api.AsSigningKey Api.AsPaymentKey) "addr_sk1648253w4tf6fv5fk28dc7crsjsaw7d9ymhztd4favg3cwkhz7x8sl5u3ms"

Then, simply wrap the signing key in a ShelleyWitnessSigningKey value:

let witness = Api.WitnessPaymentKey signingKey

This could also be done using an extended key, such as AsPaymentExtendedKey and WitnessPaymentExtendedKey.

Contents

Signing transactions

Creating transaction witnesses one by one, or all in one go.

data Tx era where

Constructors

ShelleyTx :: forall era. ShelleyBasedEra era -> Tx (ShelleyLedgerEra era) -> Tx era

Bundled Patterns

pattern Tx :: TxBody era -> [KeyWitness era] -> Tx era
This pattern will be deprecated in the future. We advise against introducing new usage of it.

data ATxAux a

Constructors

ATxAux

Fields

aTaTx :: !(Annotated Tx a)

aTaWitness :: !(Annotated TxWitness a)

aTaAnnotation :: !a

getTxBody :: Tx era -> TxBody era

getByronTxBody :: ATxAux ByteString -> Annotated Tx ByteString

getTxWitnesses :: Tx era -> [KeyWitness era]

getTxWitnessesByron :: ATxAux ByteString -> [KeyWitness ByronEra]

data ScriptValidity

Indicates whether a script is expected to fail or pass validation.

Constructors

ScriptInvalid
Script is expected to fail validation. Transactions marked as such can include scripts that fail validation. Such transactions may be submitted to the chain, in which case the collateral will be taken upon on chain script validation failure.

ScriptValid
Script is expected to pass validation. Transactions marked as such cannot include scripts that fail validation.

Signing in one go

data ShelleySigningKey

We support making key witnesses with both normal and extended signing keys.

Constructors

ShelleyNormalSigningKey (SignKeyDSIGN StandardCrypto)
A normal ed25519 signing key

ShelleyExtendedSigningKey XPrv
An extended ed25519 signing key

toShelleySigningKey :: ShelleyWitnessSigningKey -> ShelleySigningKey

signByronTransaction :: NetworkId -> Annotated Tx ByteString -> [SigningKey ByronKey] -> ATxAux ByteString

signShelleyTransaction :: ShelleyBasedEra era -> TxBody era -> [ShelleyWitnessSigningKey] -> Tx era

Incremental signing and separate witnesses

makeSignedByronTransaction :: [KeyWitness era] -> Annotated Tx ByteString -> ATxAux ByteString

makeSignedTransaction :: [KeyWitness era] -> TxBody era -> Tx era

makeSignedTransaction' :: CardanoEra era -> [KeyWitness era] -> TxBody era -> Tx era

data KeyWitness era where

Constructors

ByronKeyWitness :: TxInWitness -> KeyWitness ByronEra

ShelleyBootstrapWitness :: forall era. ShelleyBasedEra era -> BootstrapWitness StandardCrypto -> KeyWitness era

ShelleyKeyWitness :: forall era. ShelleyBasedEra era -> WitVKey 'Witness StandardCrypto -> KeyWitness era

makeByronKeyWitness :: IsByronKey key => NetworkId -> Annotated Tx ByteString -> SigningKey key -> KeyWitness ByronEra

data ShelleyWitnessSigningKey

Constructors

WitnessPaymentKey (SigningKey PaymentKey)

WitnessPaymentExtendedKey (SigningKey PaymentExtendedKey)

WitnessStakeKey (SigningKey StakeKey)

WitnessStakeExtendedKey (SigningKey StakeExtendedKey)

WitnessStakePoolKey (SigningKey StakePoolKey)

WitnessGenesisKey (SigningKey GenesisKey)

WitnessGenesisExtendedKey (SigningKey GenesisExtendedKey)

WitnessGenesisDelegateKey (SigningKey GenesisDelegateKey)

WitnessGenesisDelegateExtendedKey (SigningKey GenesisDelegateExtendedKey)

WitnessGenesisUTxOKey (SigningKey GenesisUTxOKey)

WitnessCommitteeColdKey (SigningKey CommitteeColdKey)

WitnessCommitteeColdExtendedKey (SigningKey CommitteeColdExtendedKey)

WitnessCommitteeHotKey (SigningKey CommitteeHotKey)

WitnessCommitteeHotExtendedKey (SigningKey CommitteeHotExtendedKey)

WitnessDRepKey (SigningKey DRepKey)

WitnessDRepExtendedKey (SigningKey DRepExtendedKey)

makeShelleyKeyWitness :: ShelleyBasedEra era -> TxBody era -> ShelleyWitnessSigningKey -> KeyWitness era

makeShelleyKeyWitness' :: ShelleyBasedEra era -> TxBody (ShelleyLedgerEra era) -> ShelleyWitnessSigningKey -> KeyWitness era

data WitnessNetworkIdOrByronAddress

Either a network ID or a Byron address to be used in constructing a Shelley bootstrap witness.

Constructors

WitnessNetworkId !NetworkId
Network ID.

If this value is used in the construction of a Shelley bootstrap witness, the result will not consist of a derivation path. If that is required, specify a WitnessByronAddress value instead.

WitnessByronAddress !(Address ByronAddr)
Byron address.

If this value is used in the construction of a Shelley bootstrap witness, both the network ID and derivation path will be extracted from the address and used in the construction of the witness.

makeShelleyBootstrapWitness :: ShelleyBasedEra era -> WitnessNetworkIdOrByronAddress -> TxBody era -> SigningKey ByronKey -> KeyWitness era

makeShelleyBasedBootstrapWitness :: ShelleyBasedEra era -> WitnessNetworkIdOrByronAddress -> TxBody (ShelleyLedgerEra era) -> SigningKey ByronKey -> KeyWitness era

makeShelleySignature :: SignableRepresentation tosign => tosign -> ShelleySigningKey -> SignedDSIGN StandardCrypto tosign

getShelleyKeyWitnessVerificationKey :: ShelleySigningKey -> VKey 'Witness StandardCrypto

getTxBodyAndWitnesses :: Tx era -> (TxBody era, [KeyWitness era])

Data family instances

data family AsType t

A family of singleton types used in this API to indicate which type to use where it would otherwise be ambiguous or merely unclear.

Values of this type are passed to deserialisation functions for example.

data TxBody era where

Constructors

ShelleyTxBody

Fields

:: forall era. ShelleyBasedEra era

-> TxBody (ShelleyLedgerEra era)

-> [Script (ShelleyLedgerEra era)]

-> TxBodyScriptData era

-> Maybe (TxAuxData (ShelleyLedgerEra era))

-> TxScriptValidity era

Mark script as expected to pass or fail validation

-> TxBody era

data TxScriptValidity era where

A representation of whether the era supports tx script validity.

The Alonzo and subsequent eras support script validity.

Constructors

TxScriptValidityNone :: forall era. TxScriptValidity era

TxScriptValidity :: forall era. AlonzoEraOnwards era -> ScriptValidity -> TxScriptValidity era
Tx script validity is supported in transactions in the Alonzo era onwards.

scriptValidityToIsValid :: ScriptValidity -> IsValid

isValidToScriptValidity :: IsValid -> ScriptValidity

txScriptValidityToIsValid :: TxScriptValidity era -> IsValid

txScriptValidityToScriptValidity :: TxScriptValidity era -> ScriptValidity

data TxBodyScriptData era where

Constructors

TxBodyNoScriptData :: forall era. TxBodyScriptData era

TxBodyScriptData :: forall era. AlonzoEraOnwardsConstraints era => AlonzoEraOnwards era -> TxDats (ShelleyLedgerEra era) -> Redeemers (ShelleyLedgerEra era) -> TxBodyScriptData era