Cardano.Api.Keys.Byron

Description

Byron key types and their Key class instances

Key types

data ByronKey

Byron-era payment keys. Used for Byron addresses and witnessing transactions that spend from these addresses.

These use Ed25519 but with a 32byte "chaincode" used in HD derivation. The inclusion of the chaincode is a design mistake but one that cannot be corrected for the Byron era. The Shelley era PaymentKeys do not include a chaincode. It is safe to use a zero or random chaincode for new Byron keys.

This is a type level tag, used with other interfaces like Key.

data ByronKeyLegacy

Data family instances

data family AsType t

A family of singleton types used in this API to indicate which type to use where it would otherwise be ambiguous or merely unclear.

Values of this type are passed to deserialisation functions for example.

data family VerificationKey keyrole ∷ Type

The type of cryptographic verification key, for each key role.

data family SigningKey keyrole ∷ Type

The type of cryptographic signing key, for each key role.

data family Hash keyrole ∷ Type

Legacy format

class IsByronKey key where

Methods

byronKeyFormat ∷ ByronKeyFormat key

data ByronKeyFormat key where

Constructors

ByronLegacyKeyFormat ∷ ByronKeyFormat ByronKeyLegacy

ByronModernKeyFormat ∷ ByronKeyFormat ByronKey

data SomeByronSigningKey

Constructors

AByronSigningKeyLegacy (SigningKey ByronKeyLegacy)

AByronSigningKey (SigningKey ByronKey)

toByronSigningKey ∷ SomeByronSigningKey → SigningKey