Cardano.Api.Keys.Shelley

Description

Shelley key types and their Key class instances

Key types

data CommitteeColdKey

data CommitteeColdExtendedKey

data CommitteeHotKey

data CommitteeHotExtendedKey

data DRepKey

data DRepExtendedKey

data PaymentKey

Shelley-era payment keys. Used for Shelley payment addresses and witnessing transactions that spend from these addresses.

This is a type level tag, used with other interfaces like Key.

data PaymentExtendedKey

Shelley-era payment keys using extended ed25519 cryptographic keys.

They can be used for Shelley payment addresses and witnessing transactions that spend from these addresses.

These extended keys are used by HD wallets. So this type provides interoperability with HD wallets. The ITN CLI also supported this key type.

The extended verification keys can be converted (via castVerificationKey) to ordinary keys (i.e. VerificationKey PaymentKey) but this is not the case for the signing keys. The signing keys can be used to witness transactions directly, with verification via their non-extended verification key (VerificationKey PaymentKey).

This is a type level tag, used with other interfaces like Key.

data StakeKey

data StakeExtendedKey

Shelley-era stake keys using extended ed25519 cryptographic keys.

They can be used for Shelley stake addresses and witnessing transactions that use stake addresses.

These extended keys are used by HD wallets. So this type provides interoperability with HD wallets. The ITN CLI also supported this key type.

The extended verification keys can be converted (via castVerificationKey) to ordinary keys (i.e. VerificationKey StakeKey) but this is not the case for the signing keys. The signing keys can be used to witness transactions directly, with verification via their non-extended verification key (VerificationKey StakeKey).

This is a type level tag, used with other interfaces like Key.

data StakePoolKey

data GenesisKey

data GenesisExtendedKey

Shelley-era genesis keys using extended ed25519 cryptographic keys.

These serve the same role as normal genesis keys, but are here to support legacy Byron genesis keys which used extended keys.

The extended verification keys can be converted (via castVerificationKey) to ordinary keys (i.e. VerificationKey GenesisKey) but this is not the case for the signing keys. The signing keys can be used to witness transactions directly, with verification via their non-extended verification key (VerificationKey GenesisKey).

This is a type level tag, used with other interfaces like Key.

data GenesisDelegateKey

data GenesisDelegateExtendedKey

Shelley-era genesis keys using extended ed25519 cryptographic keys.

These serve the same role as normal genesis keys, but are here to support legacy Byron genesis keys which used extended keys.

The extended verification keys can be converted (via castVerificationKey) to ordinary keys (i.e. VerificationKey GenesisKey) but this is not the case for the signing keys. The signing keys can be used to witness transactions directly, with verification via their non-extended verification key (VerificationKey GenesisKey).

This is a type level tag, used with other interfaces like Key.

data GenesisUTxOKey

Data family instances

data family AsType t

A family of singleton types used in this API to indicate which type to use where it would otherwise be ambiguous or merely unclear.

Values of this type are passed to deserialisation functions for example.

data family VerificationKey keyrole

The type of cryptographic verification key, for each key role.

data family SigningKey keyrole

The type of cryptographic signing key, for each key role.

data family Hash keyrole