Cardano.Api.KeysPraos

Description

Praos consensus key types and their Key class instances

Key types

data KesKey

data VrfKey

Data family instances

data family AsType t

A family of singleton types used in this API to indicate which type to use where it would otherwise be ambiguous or merely unclear.

Values of this type are passed to deserialisation functions for example.

data family Hash keyrole ∷ Type

data family VerificationKey keyrole ∷ Type

The type of cryptographic verification key, for each key role.

data family SigningKey keyrole ∷ Type

The type of cryptographic signing key, for each key role.