Cardano.Api.Ledger.Lens

Types

newtype TxBody era

Constructors

TxBody

Fields

unTxBody :: TxBody (ShelleyLedgerEra era)

Constructors

mkAdaOnlyTxOut :: ShelleyBasedEra era -> Addr (EraCrypto (ShelleyLedgerEra era)) -> Coin -> TxOut (ShelleyLedgerEra era)

mkAdaValue :: ShelleyBasedEra era -> Coin -> Value (ShelleyLedgerEra era)

Lenses

strictMaybeL :: forall a f. Functor f => (Maybe a -> f (Maybe a)) -> StrictMaybe a -> f (StrictMaybe a)

invalidBeforeL :: Lens' ValidityInterval (Maybe SlotNo)

Lens to access the invalidBefore field of a ValidityInterval as a 'Maybe SlotNo'.

invalidHereAfterL :: Lens' ValidityInterval (Maybe SlotNo)

Lens to access the invalidHereAfter field of a ValidityInterval as a 'Maybe SlotNo'.

invalidBeforeStrictL :: Lens' ValidityInterval (StrictMaybe SlotNo)

Lens to access the invalidBefore field of a ValidityInterval as a 'StrictMaybe SlotNo'. Ideally this should be defined in cardano-ledger

invalidHereAfterStrictL :: Lens' ValidityInterval (StrictMaybe SlotNo)

Lens to access the invalidHereAfter field of a ValidityInterval as a 'StrictMaybe SlotNo'. Ideally this should be defined in cardano-ledger

invalidBeforeTxBodyL :: AllegraEraOnwards era -> Lens' (TxBody era) (Maybe SlotNo)

invalidHereAfterTxBodyL :: ShelleyBasedEra era -> Lens' (TxBody era) (Maybe SlotNo)

Compatibility lens that provides a consistent interface over ttlTxBodyL and 'vldtTxBodyL . invalidHereAfterStrictL' across all shelley based eras.

The ledger uses ttlTxBodyL in Shelley only and from Allegra onwards uses vldtTxBodyL instead.

The former is a SlotNo with no limit represented as maxBound.

The latter is a ValidityInterval which is a pair of SlotNos that represent the lower and upper bounds.

The upper bound field is similar t ttlTxBodyL except it is a 'StrictMaybe SlotNo' type where no bounds is represented by SNothing.

invalidHereAfterTxBodyL lens over both with a 'Maybe SlotNo' type representation. Withing the Shelley era, setting Nothing will set the ttl to maxBound in the underlying ledger type.

ttlAsInvalidHereAfterTxBodyL :: ShelleyEraOnly era -> Lens' (TxBody era) (Maybe SlotNo)

Compatibility lens over ttlTxBodyL which represents maxBound as Nothing and all other values as Just.

updateTxBodyL :: ShelleyToBabbageEra era -> Lens' (TxBody era) (StrictMaybe (Update (ShelleyLedgerEra era)))

txBodyL :: forall era f. Functor f => (TxBody (ShelleyLedgerEra era) -> f (TxBody (ShelleyLedgerEra era))) -> TxBody era -> f (TxBody era)

mintTxBodyL :: MaryEraOnwards era -> Lens' (TxBody era) (MultiAsset StandardCrypto)

scriptIntegrityHashTxBodyL :: AlonzoEraOnwards era -> Lens' (TxBody era) (StrictMaybe (ScriptIntegrityHash StandardCrypto))

collateralInputsTxBodyL :: AlonzoEraOnwards era -> Lens' (TxBody era) (Set (TxIn StandardCrypto))

reqSignerHashesTxBodyL :: AlonzoEraOnwards era -> Lens' (TxBody era) (Set (KeyHash 'Witness StandardCrypto))

referenceInputsTxBodyL :: BabbageEraOnwards era -> Lens' (TxBody era) (Set (TxIn StandardCrypto))

collateralReturnTxBodyL :: BabbageEraOnwards era -> Lens' (TxBody era) (StrictMaybe (TxOut (ShelleyLedgerEra era)))

totalCollateralTxBodyL :: BabbageEraOnwards era -> Lens' (TxBody era) (StrictMaybe Coin)

certsTxBodyL :: ShelleyBasedEra era -> Lens' (TxBody era) (StrictSeq (TxCert (ShelleyLedgerEra era)))

votingProceduresTxBodyL :: ConwayEraOnwards era -> Lens' (TxBody era) (VotingProcedures (ShelleyLedgerEra era))

proposalProceduresTxBodyL :: ConwayEraOnwards era -> Lens' (TxBody era) (OSet (ProposalProcedure (ShelleyLedgerEra era)))

currentTreasuryValueTxBodyL :: ConwayEraOnwards era -> Lens' (TxBody era) (StrictMaybe Coin)

treasuryDonationTxBodyL :: ConwayEraOnwards era -> Lens' (TxBody era) Coin

adaAssetL :: ShelleyBasedEra era -> Lens' (Value (ShelleyLedgerEra era)) Coin

multiAssetL :: MaryEraOnwards era -> Lens' (MaryValue StandardCrypto) (MultiAsset StandardCrypto)

valueTxOutL :: ShelleyBasedEra era -> Lens' (TxOut (ShelleyLedgerEra era)) (Value (ShelleyLedgerEra era))

valueTxOutAdaAssetL :: ShelleyBasedEra era -> Lens' (TxOut (ShelleyLedgerEra era)) Coin