Cardano.Api.LedgerEvent

Documentation

data LedgerEvent

Constructors

PoolRegistration Certificate
The given pool is being registered for the first time on chain.

PoolReRegistration Certificate
The given pool already exists and is being re-registered.

RewardsDistribution EpochNo (Map StakeCredential Lovelace)
Rewards are being distributed.

MIRDistribution MIRDistributionDetails
MIR are being distributed.

data MIRDistributionDetails

Details of fund transfers due to MIR certificates.

Note that the transfers from reserves to treasury and treasury to reserves are inverse; a transfer of 100 ADA in either direction will result in a net movement of 0, but we include both directions for assistance in debugging.

Constructors

MIRDistributionDetails

Fields

reservePayouts ∷ Map StakeCredential Lovelace

treasuryPayouts ∷ Map StakeCredential Lovelace

reservesToTreasury ∷ Lovelace

treasuryToReserves ∷ Lovelace

toLedgerEvent ∷ ConvertLedgerEvent blk ⇒ WrapLedgerEvent blk → Maybe LedgerEvent