Cardano.Api.LedgerEvents.ConvertLedgerEvent

Documentation

data LedgerEvent

Constructors

PoolRegistration
The given pool is being registered for the first time on chain.

PoolReRegistration
The given pool already exists and is being re-registered.

IncrementalRewardsDistribution EpochNo (Map StakeCredential (Set (Reward StandardCrypto)))
Incremental rewards are being computed.

RewardsDistribution EpochNo (Map StakeCredential (Set (Reward StandardCrypto)))
Reward distribution has completed.

MIRDistribution MIRDistributionDetails
MIR are being distributed.

PoolReap PoolReapDetails
Pools have been reaped and deposits refunded.

SuccessfulPlutusScript (NonEmpty (PlutusWithContext StandardCrypto))
A number of succeeded Plutus script evaluations.

FailedPlutusScript (NonEmpty (PlutusWithContext StandardCrypto))
A number of failed Plutus script evaluations.

NewGovernanceProposals (TxId StandardCrypto) AnyProposals
Newly submittted governance proposals in a single transaction.

EpochBoundaryRatificationState AnyRatificationState
The current state of governance matters at the epoch boundary. I.E the current constitution, committee, protocol parameters, etc.

toLedgerEvent :: ConvertLedgerEvent blk => WrapLedgerEvent blk -> Maybe LedgerEvent