Cardano.Api.Monad.Error

Description

This module serves purpose as a single source of abstractions used in handling MonadError and ExceptT through 'cardano-api'.

Documentation

type MonadTransError e t m = (Monad m, MonadTrans t, MonadError e (t m))

Convenience alias

type MonadIOTransError e t m = (MonadIO m, MonadIO (t m), MonadCatch m, MonadTrans t, MonadError e (t m))

Same as MonadTransError, but with also MonadIO constraint

liftExceptT :: MonadTransError e t m => ExceptT e m a -> t m a

Lift ExceptT into MonadTransError

modifyError

Arguments

:: MonadTransError e' t m

=> (e -> e')
mapping function

-> ExceptT e m a
value

-> t m a
result with modified error

Modify an ExceptT error and lift it to MonadError transformer stack.

This implementation avoids nesting problem of modifyError from mtl. The issue with modifyError (from mtl) is that when you use it on a function, you actually end up with ExceptT e1 (ExceptT e2 m a):

modifyError (f :: e2 -> e1) (foo :: ExceptT e2 (ExceptT e1 m) a) :: ExceptT e1 m a

and if you use modifyError (mtl) again, the more nested you get e.g. ExceptT e1 (ExceptT e2 (ExceptT e3 m a)). With a deeper monad stack you pay the overhead with every use of >>=.

This function avoids that, but at the cost of limiting its application to transformers.

handleIOExceptionsWith

Arguments

:: MonadError e' m

=> MonadCatch m

=> Exception e

=> (e -> e')
mapping function

-> m a
action that can throw

-> m a
result with caught exception

Wrap an exception and lift it into MonadError.

handleIOExceptionsLiftWith

Arguments

:: MonadIOTransError e' t m

=> Exception e

=> (e -> e')
mapping function

-> m a
action that can throw

-> t m a
action with caught error lifted into MonadError stack

Wrap an exception and lift it into MonadError stack.

hoistIOEither :: MonadIOTransError e t m => IO (Either e a) -> t m a

Lift an IO action that returns Either into MonadIOTransError

class Monad m => MonadError e (m :: Type -> Type) | m -> e where

Methods

throwError :: e -> m a

catchError :: m a -> (e -> m a) -> m a

newtype ExceptT e (m :: Type -> Type) a

Constructors

ExceptT (m (Either e a))

runExceptT :: ExceptT e m a -> m (Either e a)

liftEither :: MonadError e m => Either e a -> m a

mapExcept :: (Either e a -> Either e' b) -> Except e a -> Except e' b

mapExceptT :: (m (Either e a) -> n (Either e' b)) -> ExceptT e m a -> ExceptT e' n b

runExcept :: Except e a -> Either e a

withExcept :: (e -> e') -> Except e a -> Except e' a

module Control.Monad.IO.Class