Cardano.Api.Network

Network types

data NetworkId

Constructors

Mainnet

Testnet !NetworkMagic

newtype NetworkMagic

NetworkMagic is used to differentiate between different networks during the initial handshake.

Constructors

NetworkMagic

Fields

unNetworkMagic :: Word32

fromNetworkMagic :: NetworkMagic -> NetworkId

toNetworkMagic :: NetworkId -> NetworkMagic

mainnetNetworkMagic :: NetworkMagic

network reexports

data LedgerPeerSnapshot

The type of big ledger peers that is serialised or later provided by node configuration for the networking layer to connect to when syncing.

Constructors

LedgerPeerSnapshotV1 (WithOrigin SlotNo, [(AccPoolStake, (PoolStake, NonEmpty RelayAccessPoint))])
Internal use for version 1, use pattern synonym for public API

Bundled Patterns

pattern LedgerPeerSnapshot :: (WithOrigin SlotNo, [(AccPoolStake, (PoolStake, NonEmpty RelayAccessPoint))]) -> LedgerPeerSnapshot
Public API to access snapshot data. Currently access to only most recent version is available. Nonetheless, serialisation from the node into JSON is supported for older versions via internal api so that newer CLI can still support older node formats.

data Target point

Constructors

VolatileTip
The tip of the volatile chain

Cannot fail to be acquired.

SpecificPoint point
A specified point

Fails to be acquired if the point is not between VolatileTip and ImmutableTip (inclusive).

ImmutableTip
The tip of the immutable chain

Cannot fail to be acquired.

Requires at least NodeToClientV_16.

newtype Serialised (a :: k)

An already serialised value

When streaming blocks/header from disk to the network, there is often no need to deserialise them, as we'll just end up serialising them again when putting them on the wire.

Constructors

Serialised

Fields

unSerialised :: ByteString

data SubmitResult reason

Isomorphic with Maybe but with a name that better describes its purpose and usage.

Constructors

SubmitSuccess

SubmitFail reason

Internal conversion functions

toByronProtocolMagicId :: NetworkId -> ProtocolMagicId

toByronNetworkMagic :: NetworkId -> NetworkMagic

toByronRequiresNetworkMagic :: NetworkId -> RequiresNetworkMagic

toShelleyNetwork :: NetworkId -> Network

fromShelleyNetwork :: Network -> NetworkMagic -> NetworkId