Cardano.Api.Network.IPC

Description

Node IPC protocols

This module provides the client side of the node-to-client interprocess communication (IPC) for interacting with a local Cardano node. It supports querying the node for information, submitting transactions, monitoring the local mempool, and retrieving historical chain data using the ChainSync protocol.

Examples

This section provides two examples:

Querying the node to obtain basic information

Submitting a transaction to the node.

For details on how to create a transaction, see the Cardano.Api.Tx.Internal.Body documentation.

The following qualified imports from cardano-api are used in the examples below:

import qualified Cardano.Api as Api
import qualified Cardano.Api.Consensus as Consensus
import qualified Cardano.Api.Network as Network

The following explicit import from base is also required:

import Control.Monad.Except (runExceptT)

The examples assume the use of the IO monad and unqualified access to the Prelude module.

Constructing connection information

Regardless of whether the goal is to query the node or submit transactions, the first step is to gather the connection information.

The following information is required:

The number of slots per epoch. This value depends on the network the node is connected to. It can be obtained by inspecting the epochLength key in the shelley-genesis.json file used by the node. On the preview network, the value is, at the time of writing, 86_400, but it can change. This value and other genesis parameters can also be obtained using the QueryGenesisParameters query.

Network identifier. When connecting to a testnet, the network identifier is also required. It can be obtained by looking for the networkId obtained by looking up the networkMagic key in the shelley-genesis.json file that the node uses to connect to the network. For the preview network, the current identifier is 2.

Socket path. The path to the node's socket file. It can be set using the --socket-path parameter when starting the node. By default, it is typically located in the db subfolder of the node's working directory.

Then, combine all the required information into a LocalNodeConnectInfo value.

For example, assume the node is connected to the preview network and the socket file is located at /home/user/cardano-node/db/node.socket. The LocalNodeConnectInfo value can then be constructed as follows:

let connectionInfo =
      Api.LocalNodeConnectInfo
         { Api.localConsensusModeParams = Api.CardanoModeParams (Api.EpochSlots 86_400)
         , Api.localNodeNetworkId = Api.Testnet (Api.NetworkMagic 2)
         , Api.localNodeSocketPath = Api.File "/home/user/cardano-node/db/node.socket"
         }

Querying the node for the UTXO set

To obtain the set of unspent transaction outputs (UTXO set) from the network, the current era must first be determined.

Obtaining the current era

Many queries require knowing the current era. This can be hardcoded using one of the ShelleyBasedEra constructors, but it is also possible to retreive it from the node:

eEra <- runExceptT $ Api.queryNodeLocalState connectionInfo Network.VolatileTip Api.QueryCurrentEra

VolatileTip requests information from the most recent block the node is aware of. It is important to note that this information is not guaranteed to remain valid, as it may be rolled back.

Alternatively, ImmutableTip can be used to obtain information from the most recent block considered as final by the consensus algorithm. While this data is stable and will not be rolled back, it is less recent – on mainnet, it is typically about 36 hours behind the current time.

QueryCurrentEra is the constructor of the query that retrieves the node's current era.

The query returns an ExceptT monad, which can be run using the runExceptT function. This yields an eEra value of type Either AcquiringFailure AnyCardanoEra.

Below is an example of how to unwrap this value into a ShelleyBasedEra based era, assuming the node is not running Byron:

Api.AnyShelleyBasedEra sbe :: Api.AnyShelleyBasedEra <- case eEra of
  Right (Api.AnyCardanoEra era) ->
    Api.caseByronOrShelleyBasedEra
      (error "Error, we are in Byron era")
      (return . Api.AnyShelleyBasedEra)
      era
  Left Shelley.AFPointTooOld -> error "Error, point queried in the chain is too old!"
  Left Shelley.AFPointNotOnChain -> error "Error, point queried is not on chain!"

AFPointTooOld and AFPointNotOnChain errors should not occur when querying with either VolatileTip or ImmutableTip.

Obtaining the UTXO set

After determining the current era, the node can be queried for the UTXO set using the QueryUTxO query as follows:

eUtxo <-
  runExceptT $
    Api.queryNodeLocalState
      connectionInfo
      Network.VolatileTip
      (Api.QueryInEra (Api.QueryInShelleyBasedEra sbe (Api.QueryUTxO Api.QueryUTxOWhole)))

This returns, a nested type of Either AcquiringFailure (Either EraMismatch (UTXO era)). You can unwrap it as follows:

utxo <- case eUtxo of
  Right (Right (Api.UTxO utxo)) -> do
    return utxo
  Right (Left (Consensus.EraMismatch{Consensus.ledgerEraName, Consensus.otherEraName})) ->
    error
      ( "Error, we assumed era was "
          ++ show otherEraName
          ++ " but it was "
          ++ show ledgerEraName
      )
  Left Shelley.AFPointTooOld -> error "Error, point queried in the chain is too old!"
  Left Shelley.AFPointNotOnChain -> error "Error, point queried is not on chain!"

Alternatively, to avoid nested result types, you can use convenience functions and types from Cardano.Api.Query.Internal.Convenience. It is also posible to combine several queries into a single connection by using the monadic interface that can be found in the Cardano.Api.Network.IPC.Internal.Monad documentation.

The obtained utxo variable is a standard Map of type Map TxIn (TxOut CtxUTxO era).

Submitting a transaction

Assume there is a signed transaction in the latest era that you would like to submit to the node. Assume it is stored in the variable signedTx of type Tx era.

For details on how to create such a transaction, see the Cardano.Api.Tx.Internal.Body documentation.

To submit the transaction to the node, use the submitTxToNodeLocal function as follows:

result <- Api.submitTxToNodeLocal connectionInfo (Api.TxInMode sbe signedTx)

The result is a SubmitResult value, which can be inspected as follows:

case result of
  Api.SubmitSuccess -> putStrLn "Transaction submitted successfully!"
  Api.SubmitFail reason -> error $ "Error submitting transaction: " ++ show reason

If the command succeeds, the transaction gets into the node's mempool, ready to be included in a block.

Node interaction

Operations that involve talking to a local Cardano node.

connectToLocalNode :: MonadIO m => LocalNodeConnectInfo -> LocalNodeClientProtocolsInMode -> m ()

Establish a connection to a local node and execute the given set of protocol handlers.

connectToLocalNodeWithVersion :: MonadIO m => LocalNodeConnectInfo -> (NodeToClientVersion -> LocalNodeClientProtocolsInMode) -> m ()

Establish a connection to a local node and execute the given set of protocol handlers parameterized on the negotiated node-to-client protocol version.

data LocalNodeConnectInfo

Constructors

LocalNodeConnectInfo

Fields

localConsensusModeParams :: ConsensusModeParams

localNodeNetworkId :: NetworkId

localNodeSocketPath :: SocketPath

data LocalNodeClientParams where

This type defines the boundary between the mode-parametrised style used in this API and the block-parametrised style used by the underlying network and consensus libraries.

This interface itself is in the block-parametrised style, with the block type itself being an hidden/existential type.

It bundles together all the necessary class instances, the consensus protocol client identifier, and the set of client side mini-protocol handlers for the node-to-client protocol.

Constructors

LocalNodeClientParamsSingleBlock :: forall block. (ProtocolClient block, LedgerSupportsProtocol (ShelleyBlock (TPraos StandardCrypto) ShelleyEra)) => ProtocolClientInfoArgs block -> (NodeToClientVersion -> LocalNodeClientProtocolsForBlock block) -> LocalNodeClientParams

LocalNodeClientParamsCardano :: forall block. (ProtocolClient block, CardanoHardForkConstraints (ConsensusCryptoForBlock block)) => ProtocolClientInfoArgs block -> (NodeToClientVersion -> LocalNodeClientProtocolsForBlock block) -> LocalNodeClientParams

mkLocalNodeClientParams :: ConsensusModeParams -> (NodeToClientVersion -> LocalNodeClientProtocolsInMode) -> LocalNodeClientParams

Convert from the mode-parametrised style to the block-parametrised style.

data LocalNodeClientProtocols block point tip slot tx txid txerr (query :: Type -> Type) (m :: Type -> Type)

The protocols we can use with a local node. Use in conjunction with connectToLocalNode.

These protocols use the types from the rest of this API. The conversion to/from the types used by the underlying wire formats is handled by connectToLocalNode.

Constructors

LocalNodeClientProtocols

Fields

localChainSyncClient :: LocalChainSyncClient block point tip m

localTxSubmissionClient :: Maybe (LocalTxSubmissionClient tx txerr m ())

localStateQueryClient :: Maybe (LocalStateQueryClient block point query m ())

localTxMonitoringClient :: Maybe (LocalTxMonitorClient txid tx slot m ())

data LocalChainSyncClient block point tip (m :: Type -> Type)

Constructors

NoLocalChainSyncClient

LocalChainSyncClientPipelined (ChainSyncClientPipelined block point tip m ())

LocalChainSyncClient (ChainSyncClient block point tip m ())

type LocalNodeClientProtocolsInMode = LocalNodeClientProtocols BlockInMode ChainPoint ChainTip SlotNo TxInMode TxIdInMode TxValidationErrorInCardanoMode QueryInMode IO

Chain sync protocol

newtype ChainSyncClient header point tip (m :: Type -> Type) a

A chain sync protocol client, on top of some effect m. The first choice of request is within that m.

Constructors

ChainSyncClient

Fields

runChainSyncClient :: m (ClientStIdle header point tip m a)

newtype ChainSyncClientPipelined header point tip (m :: Type -> Type) a

Pipelined chain sync client. It can only pipeline MsgRequestNext messages, while the MsgFindIntersect are non pipelined. This has a penalty cost of an RTT, but they are sent relatively seldom and their response might impact how many messages one would like to pipeline. It also simplifies the receiver callback.

Constructors

ChainSyncClientPipelined

Fields

runChainSyncClientPipelined :: m (ClientPipelinedStIdle 'Z header point tip m a)

data BlockInMode where

A Block in one of the eras. TODO Rename this to BlockInEra

Constructors

BlockInMode :: forall era. CardanoEra era -> Block era -> BlockInMode

Local tx submission

newtype LocalTxSubmissionClient tx reject (m :: Type -> Type) a

Constructors

LocalTxSubmissionClient

Fields

runLocalTxSubmissionClient :: m (LocalTxClientStIdle tx reject m a)

data TxInMode where

A Tx in one of the eras supported by a given protocol mode.

For multi-era modes such as the CardanoMode this type is a sum of the different transaction types for all the eras. It is used in the LocalTxSubmission protocol.

Constructors

TxInMode :: forall era. ShelleyBasedEra era -> Tx era -> TxInMode
Shelley based transactions.

TxInByronSpecial :: GenTx ByronBlock -> TxInMode
Legacy Byron transactions and things we can post to the chain which are not actually transactions. This covers: update proposals, votes and delegation certs.

data TxValidationErrorInCardanoMode

A TxValidationError in one of the eras supported by a given protocol mode.

This is used in the LocalStateQuery protocol.

data TxValidationError era

The transaction validations errors that can occur from trying to submit a transaction to a local node. The errors are specific to an era.

submitTxToNodeLocal :: MonadIO m => LocalNodeConnectInfo -> TxInMode -> m (SubmitResult TxValidationErrorInCardanoMode)

data SubmitResult reason

Isomorphic with Maybe but with a name that better describes its purpose and usage.

Constructors

SubmitSuccess

SubmitFail reason

Local state query

newtype LocalStateQueryClient block point (query :: Type -> Type) (m :: Type -> Type) a

Constructors

LocalStateQueryClient

Fields

runLocalStateQueryClient :: m (ClientStIdle block point query m a)

data AcquiringFailure

Establish a connection to a node and execute a single query using the local state query protocol.

Constructors

AFPointTooOld

AFPointNotOnChain

data QueryInMode result where

Constructors

QueryCurrentEra :: QueryInMode AnyCardanoEra

QueryInEra :: forall era result1. QueryInEra era result1 -> QueryInMode (Either EraMismatch result1)

QueryEraHistory :: QueryInMode EraHistory

QuerySystemStart :: QueryInMode SystemStart

QueryChainBlockNo :: QueryInMode (WithOrigin BlockNo)

QueryChainPoint :: QueryInMode ChainPoint

QueryLedgerConfig :: QueryInMode (HardForkLedgerConfig (CardanoEras StandardCrypto))

data QueryInEra era result where

Constructors

QueryByronUpdateState :: QueryInEra ByronEra ByronUpdateState

QueryInShelleyBasedEra :: forall era result. ShelleyBasedEra era -> QueryInShelleyBasedEra era result -> QueryInEra era result

data QueryInShelleyBasedEra era result where

Constructors

QueryEpoch :: forall era. QueryInShelleyBasedEra era EpochNo

QueryGenesisParameters :: forall era. QueryInShelleyBasedEra era (GenesisParameters ShelleyEra)

QueryProtocolParameters :: forall era. QueryInShelleyBasedEra era (PParams (ShelleyLedgerEra era))

QueryStakeDistribution :: forall era. QueryInShelleyBasedEra era (Map (Hash StakePoolKey) Rational)

QueryUTxO :: forall era. QueryUTxOFilter -> QueryInShelleyBasedEra era (UTxO era)

QueryStakeAddresses :: forall era. Set StakeCredential -> NetworkId -> QueryInShelleyBasedEra era (Map StakeAddress Coin, Map StakeAddress PoolId)

QueryStakePools :: forall era. QueryInShelleyBasedEra era (Set PoolId)

QueryStakePoolParameters :: forall era. Set PoolId -> QueryInShelleyBasedEra era (Map PoolId StakePoolParameters)

QueryDebugLedgerState :: forall era. QueryInShelleyBasedEra era (SerialisedDebugLedgerState era)

QueryProtocolState :: forall era. QueryInShelleyBasedEra era (ProtocolState era)

QueryCurrentEpochState :: forall era. QueryInShelleyBasedEra era (SerialisedCurrentEpochState era)

QueryPoolState :: forall era. Maybe (Set PoolId) -> QueryInShelleyBasedEra era (SerialisedPoolState era)

QueryPoolDistribution :: forall era. Maybe (Set PoolId) -> QueryInShelleyBasedEra era (SerialisedPoolDistribution era)

QueryStakeSnapshot :: forall era. Maybe (Set PoolId) -> QueryInShelleyBasedEra era (SerialisedStakeSnapshots era)

QueryStakeDelegDeposits :: forall era. Set StakeCredential -> QueryInShelleyBasedEra era (Map StakeCredential Coin)

QueryAccountState :: forall era. QueryInShelleyBasedEra era AccountState

QueryConstitution :: forall era. QueryInShelleyBasedEra era (Constitution (ShelleyLedgerEra era))

QueryGovState :: forall era. QueryInShelleyBasedEra era (GovState (ShelleyLedgerEra era))

QueryRatifyState :: forall era. QueryInShelleyBasedEra era (RatifyState (ShelleyLedgerEra era))

QueryFuturePParams :: forall era. QueryInShelleyBasedEra era (Maybe (PParams (ShelleyLedgerEra era)))

QueryDRepState :: forall era. Set (Credential 'DRepRole) -> QueryInShelleyBasedEra era (Map (Credential 'DRepRole) DRepState)

QueryDRepStakeDistr :: forall era. Set DRep -> QueryInShelleyBasedEra era (Map DRep Coin)

QuerySPOStakeDistr :: forall era. Set (KeyHash 'StakePool) -> QueryInShelleyBasedEra era (Map (KeyHash 'StakePool) Coin)

QueryCommitteeMembersState :: forall era. Set (Credential 'ColdCommitteeRole) -> Set (Credential 'HotCommitteeRole) -> Set MemberStatus -> QueryInShelleyBasedEra era CommitteeMembersState

QueryStakeVoteDelegatees :: forall era. Set StakeCredential -> QueryInShelleyBasedEra era (Map StakeCredential DRep)

QueryProposals :: forall era. Set GovActionId -> QueryInShelleyBasedEra era (Seq (GovActionState (ShelleyLedgerEra era)))

QueryLedgerPeerSnapshot :: forall era. QueryInShelleyBasedEra era (Serialised LedgerPeerSnapshot)

QueryStakePoolDefaultVote :: forall era. KeyHash 'StakePool -> QueryInShelleyBasedEra era DefaultVote

queryNodeLocalState :: LocalNodeConnectInfo -> Target ChainPoint -> QueryInMode result -> ExceptT AcquiringFailure IO result

Query monad

data LocalStateQueryExpr block point (query :: Type -> Type) r (m :: Type -> Type) a

Monadic type for constructing local state query expressions.

Use queryExpr in a do block to construct queries of this type and convert the expression to a LocalStateQueryClient with setupLocalStateQueryExpr.

Some consideration was made to use Applicative instead of Monad as the abstraction in order to support pipelining, but we actually have a fair amount of code where the next query depends on the result of the former and therefore actually need Monad.

In order to make pipelining still possible we can explore the use of Selective Functors which would allow us to straddle both worlds.

executeLocalStateQueryExpr :: LocalNodeConnectInfo -> Target ChainPoint -> LocalStateQueryExpr BlockInMode ChainPoint QueryInMode () IO a -> IO (Either AcquiringFailure a)

Execute a local state query expression.

queryExpr :: QueryInMode a -> LocalStateQueryExpr block point QueryInMode r IO (Either UnsupportedNtcVersionError a)

Use queryExpr in a do block to construct monadic local state queries.

Local tx monitoring

newtype LocalTxMonitorClient txid tx slot (m :: Type -> Type) a

A tx monitor client, on top of some effect m.

Constructors

LocalTxMonitorClient

Fields

runLocalTxMonitorClient :: m (ClientStIdle txid tx slot m a)

data LocalTxMonitoringQuery

Constructors

LocalTxMonitoringQueryTx TxIdInMode
Query if a particular tx exists in the mempool. Note that, the absence of a transaction does not imply anything about how the transaction was processed: it may have been dropped, or inserted in a block.

LocalTxMonitoringSendNextTx
The mempool is modeled as an ordered list of transactions and thus, can be traversed linearly. LocalTxMonitoringSendNextTx requests the next transaction from the current list. This must be a transaction that was not previously sent to the client for this particular snapshot.

LocalTxMonitoringMempoolInformation
Ask the server about the current mempool's capacity and sizes. This is fixed in a given snapshot.

data LocalTxMonitoringResult

Constructors

LocalTxMonitoringTxExists TxId SlotNo
Slot number at which the mempool snapshot was taken

LocalTxMonitoringTxDoesNotExist TxId SlotNo
Slot number at which the mempool snapshot was taken

LocalTxMonitoringNextTx (Maybe TxInMode) SlotNo
Slot number at which the mempool snapshot was taken

LocalTxMonitoringMempoolSizeAndCapacity MempoolSizeAndCapacity SlotNo
Slot number at which the mempool snapshot was taken

data MempoolSizeAndCapacity

Describes the MemPool sizes and capacity for a given snapshot.

Constructors

MempoolSizeAndCapacity

Fields

capacityInBytes :: !Word32

The maximum capacity of the mempool. Note that this may dynamically change when the ledger state is updated.

sizeInBytes :: !Word32

The summed byte size of all the transactions in the mempool.

numberOfTxs :: !Word32

The number of transactions in the mempool

queryTxMonitoringLocal :: MonadIO m => LocalNodeConnectInfo -> LocalTxMonitoringQuery -> m LocalTxMonitoringResult

data EraHistory where

Constructors

EraHistory :: forall (xs :: [Type]). CardanoBlock StandardCrypto ~ HardForkBlock xs => Interpreter xs -> EraHistory

getProgress :: SlotNo -> EraHistory -> Either PastHorizonException (RelativeTime, SlotLength)

Common queries

getLocalChainTip :: MonadIO m => LocalNodeConnectInfo -> m ChainTip

Query versioning

isQuerySupportedInNtcVersion :: Some (Query (CardanoBlock StandardCrypto)) -> NodeToClientVersion -> Either UnsupportedNtcVersionError ()

LocalStateQuery uses versioned queries, which means it requires the Node to support a minimum (and possibly maximum) Node-to-Client version.

Background: The node to client LocalStateQuery protocol is such that it will disconnect on any unrecognised queries. This means that for a Node-to-Client connection, if a query is sent that was introduced in a Node-to-Client version that is newer than the Node-To-Client version of the connection, the node will disconnect the client. The client will not get any information about why the disconnect happened. This is a bad user experience for tools such as the CLI.

To improve the user experience the API needs to prevent the sending of queries that are newer than the Node-To-Client version of the connection by checking the version of the query before sending it. This affords the ability to return UnsupportedNtcVersionError(UnsupportedNtcVersionError), informing the caller of a Node-To-Client versioning issue.

For maintaining typeclass instances, see the NodeToClientVersion type documentation for a list of versions and the queries that were introduced for those versions.

More information about queries versioning can be found: * https://ouroboros-network.cardano.intersectmbo.org/ouroboros-network/Ouroboros-Network-NodeToClient.html#t:NodeToClientVersion * https://ouroboros-consensus.cardano.intersectmbo.org/docs/for-developers/QueryVersioning/#implementation

data NodeToClientVersion

Enumeration of node to client protocol versions.

Constructors

NodeToClientV_16
NodeToClientV_10 -- ^ added GetChainBlockNo and GetChainPoint queries | NodeToClientV_11 -- ^ added GetRewardInfoPools Block query | NodeToClientV_12 -- ^ added LocalTxMonitor mini-protocol | NodeToClientV_13 -- ^ enabled CardanoNodeToClientVersion9, i.e., Babbage | NodeToClientV_14 -- ^ added GetPoolDistr, GetPoolState, GetSnapshots | NodeToClientV_15 -- ^ added query to NodeToClientVersionData

NodeToClientV_17
added GetProposals and GetRatifyState queries

NodeToClientV_18
added GetFuturePParams query

NodeToClientV_19
added GetLedgerPeerSnapshot

NodeToClientV_20
added QueryStakePoolDefaultVote

Error types

data UnsupportedNtcVersionError

Constructors

UnsupportedNtcVersionError

Fields

!NodeToClientVersion

The version we negotiated

![NodeToClientVersion]

The versions in which this query is supported

Helpers

toAcquiringFailure :: AcquireFailure -> AcquiringFailure