Cardano.Api.Orphans

Orphan instances

ToJSON ApplyMempoolPayloadErr

Instance details

Methods

toJSON :: ApplyMempoolPayloadErr -> Value

toEncoding :: ApplyMempoolPayloadErr -> Encoding

toJSONList :: [ApplyMempoolPayloadErr] -> Value

toEncodingList :: [ApplyMempoolPayloadErr] -> Encoding

omitField :: ApplyMempoolPayloadErr -> Bool

ToJSON KeyHash

Instance details

Methods

toJSON :: KeyHash -> Value

toEncoding :: KeyHash -> Encoding

toJSONList :: [KeyHash] -> Value

toEncodingList :: [KeyHash] -> Encoding

omitField :: KeyHash -> Bool

ToJSON LovelaceError

Instance details

Methods

toJSON :: LovelaceError -> Value

toEncoding :: LovelaceError -> Encoding

toJSONList :: [LovelaceError] -> Value

toEncodingList :: [LovelaceError] -> Encoding

omitField :: LovelaceError -> Bool

ToJSON Error

Instance details

Methods

toJSON :: Error -> Value

toEncoding :: Error -> Encoding

toJSONList :: [Error] -> Value

toEncodingList :: [Error] -> Encoding

omitField :: Error -> Bool

ToJSON UTxOError

Instance details

Methods

toJSON :: UTxOError -> Value

toEncoding :: UTxOError -> Encoding

toJSONList :: [UTxOError] -> Value

toEncodingList :: [UTxOError] -> Encoding

omitField :: UTxOError -> Bool

ToJSON TxValidationError

Instance details

Methods

toJSON :: TxValidationError -> Value

toEncoding :: TxValidationError -> Encoding

toJSONList :: [TxValidationError] -> Value

toEncodingList :: [TxValidationError] -> Encoding

omitField :: TxValidationError -> Bool

ToJSON UTxOValidationError

Instance details

Methods

toJSON :: UTxOValidationError -> Value

toEncoding :: UTxOValidationError -> Encoding

toJSONList :: [UTxOValidationError] -> Value

toEncodingList :: [UTxOValidationError] -> Encoding

omitField :: UTxOValidationError -> Bool

ToJSON ApplicationNameError

Instance details

Methods

toJSON :: ApplicationNameError -> Value

toEncoding :: ApplicationNameError -> Encoding

toJSONList :: [ApplicationNameError] -> Value

toEncodingList :: [ApplicationNameError] -> Encoding

omitField :: ApplicationNameError -> Bool

ToJSON SoftwareVersionError

Instance details

Methods

toJSON :: SoftwareVersionError -> Value

toEncoding :: SoftwareVersionError -> Encoding

toJSONList :: [SoftwareVersionError] -> Value

toEncodingList :: [SoftwareVersionError] -> Encoding

omitField :: SoftwareVersionError -> Bool

ToJSON SystemTagError

Instance details

Methods

toJSON :: SystemTagError -> Value

toEncoding :: SystemTagError -> Encoding

toJSONList :: [SystemTagError] -> Value

toEncodingList :: [SystemTagError] -> Encoding

omitField :: SystemTagError -> Bool

ToJSON Error

Instance details

Methods

toJSON :: Error -> Value

toEncoding :: Error -> Encoding

toJSONList :: [Error] -> Value

toEncodingList :: [Error] -> Encoding

omitField :: Error -> Bool

ToJSON Error

Instance details

Methods

toJSON :: Error -> Value

toEncoding :: Error -> Encoding

toJSONList :: [Error] -> Value

toEncodingList :: [Error] -> Encoding

omitField :: Error -> Bool

ToJSON Adopted

Instance details

Methods

toJSON :: Adopted -> Value

toEncoding :: Adopted -> Encoding

toJSONList :: [Adopted] -> Value

toEncodingList :: [Adopted] -> Encoding

omitField :: Adopted -> Bool

ToJSON ApplicationVersion

Instance details

Methods

toJSON :: ApplicationVersion -> Value

toEncoding :: ApplicationVersion -> Encoding

toJSONList :: [ApplicationVersion] -> Value

toEncodingList :: [ApplicationVersion] -> Encoding

omitField :: ApplicationVersion -> Bool

ToJSON Error

Instance details

Methods

toJSON :: Error -> Value

toEncoding :: Error -> Encoding

toJSONList :: [Error] -> Value

toEncodingList :: [Error] -> Encoding

omitField :: Error -> Bool

ToJSON Error

Instance details

Methods

toJSON :: Error -> Value

toEncoding :: Error -> Encoding

toJSONList :: [Error] -> Value

toEncodingList :: [Error] -> Encoding

omitField :: Error -> Bool

ToJSON VotingPeriod

Instance details

Methods

toJSON :: VotingPeriod -> Value

toEncoding :: VotingPeriod -> Encoding

toJSONList :: [VotingPeriod] -> Value

toEncodingList :: [VotingPeriod] -> Encoding

omitField :: VotingPeriod -> Bool

ToJSON TicknState

Instance details

Methods

toJSON :: TicknState -> Value

toEncoding :: TicknState -> Encoding

toJSONList :: [TicknState] -> Value

toEncodingList :: [TicknState] -> Encoding

omitField :: TicknState -> Bool

ToJSON ByronHash

Instance details

Methods

toJSON :: ByronHash -> Value

toEncoding :: ByronHash -> Encoding

toJSONList :: [ByronHash] -> Value

toEncodingList :: [ByronHash] -> Encoding

omitField :: ByronHash -> Bool

ToJSONKey ApplicationName

Instance details

Methods

toJSONKey :: ToJSONKeyFunction ApplicationName

toJSONKeyList :: ToJSONKeyFunction [ApplicationName]

Data CharPosition

Instance details

Methods

gfoldl :: (forall d b. Data d => c (d -> b) -> d -> c b) -> (forall g. g -> c g) -> CharPosition -> c CharPosition

gunfold :: (forall b r. Data b => c (b -> r) -> c r) -> (forall r. r -> c r) -> Constr -> c CharPosition

toConstr :: CharPosition -> Constr

dataTypeOf :: CharPosition -> DataType

dataCast1 :: Typeable t => (forall d. Data d => c (t d)) -> Maybe (c CharPosition)

dataCast2 :: Typeable t => (forall d e. (Data d, Data e) => c (t d e)) -> Maybe (c CharPosition)

gmapT :: (forall b. Data b => b -> b) -> CharPosition -> CharPosition

gmapQl :: (r -> r' -> r) -> r -> (forall d. Data d => d -> r') -> CharPosition -> r

gmapQr :: forall r r'. (r' -> r -> r) -> r -> (forall d. Data d => d -> r') -> CharPosition -> r

gmapQ :: (forall d. Data d => d -> u) -> CharPosition -> [u]

gmapQi :: Int -> (forall d. Data d => d -> u) -> CharPosition -> u

gmapM :: Monad m => (forall d. Data d => d -> m d) -> CharPosition -> m CharPosition

gmapMp :: MonadPlus m => (forall d. Data d => d -> m d) -> CharPosition -> m CharPosition

gmapMo :: MonadPlus m => (forall d. Data d => d -> m d) -> CharPosition -> m CharPosition

Data DecodingError

Instance details

Methods

gfoldl :: (forall d b. Data d => c (d -> b) -> d -> c b) -> (forall g. g -> c g) -> DecodingError -> c DecodingError

gunfold :: (forall b r. Data b => c (b -> r) -> c r) -> (forall r. r -> c r) -> Constr -> c DecodingError

toConstr :: DecodingError -> Constr

dataTypeOf :: DecodingError -> DataType

dataCast1 :: Typeable t => (forall d. Data d => c (t d)) -> Maybe (c DecodingError)

dataCast2 :: Typeable t => (forall d e. (Data d, Data e) => c (t d e)) -> Maybe (c DecodingError)

gmapT :: (forall b. Data b => b -> b) -> DecodingError -> DecodingError

gmapQl :: (r -> r' -> r) -> r -> (forall d. Data d => d -> r') -> DecodingError -> r

gmapQr :: forall r r'. (r' -> r -> r) -> r -> (forall d. Data d => d -> r') -> DecodingError -> r

gmapQ :: (forall d. Data d => d -> u) -> DecodingError -> [u]

gmapQi :: Int -> (forall d. Data d => d -> u) -> DecodingError -> u

gmapM :: Monad m => (forall d. Data d => d -> m d) -> DecodingError -> m DecodingError

gmapMp :: MonadPlus m => (forall d. Data d => d -> m d) -> DecodingError -> m DecodingError

gmapMo :: MonadPlus m => (forall d. Data d => d -> m d) -> DecodingError -> m DecodingError

Data DecoderError

Instance details

Methods

gfoldl :: (forall d b. Data d => c (d -> b) -> d -> c b) -> (forall g. g -> c g) -> DecoderError -> c DecoderError

gunfold :: (forall b r. Data b => c (b -> r) -> c r) -> (forall r. r -> c r) -> Constr -> c DecoderError

toConstr :: DecoderError -> Constr

dataTypeOf :: DecoderError -> DataType

dataCast1 :: Typeable t => (forall d. Data d => c (t d)) -> Maybe (c DecoderError)

dataCast2 :: Typeable t => (forall d e. (Data d, Data e) => c (t d e)) -> Maybe (c DecoderError)

gmapT :: (forall b. Data b => b -> b) -> DecoderError -> DecoderError

gmapQl :: (r -> r' -> r) -> r -> (forall d. Data d => d -> r') -> DecoderError -> r

gmapQr :: forall r r'. (r' -> r -> r) -> r -> (forall d. Data d => d -> r') -> DecoderError -> r

gmapQ :: (forall d. Data d => d -> u) -> DecoderError -> [u]

gmapQi :: Int -> (forall d. Data d => d -> u) -> DecoderError -> u

gmapM :: Monad m => (forall d. Data d => d -> m d) -> DecoderError -> m DecoderError

gmapMp :: MonadPlus m => (forall d. Data d => d -> m d) -> DecoderError -> m DecoderError

gmapMo :: MonadPlus m => (forall d. Data d => d -> m d) -> DecoderError -> m DecoderError

Data DeserialiseFailure

Instance details

Methods

gfoldl :: (forall d b. Data d => c (d -> b) -> d -> c b) -> (forall g. g -> c g) -> DeserialiseFailure -> c DeserialiseFailure

gunfold :: (forall b r. Data b => c (b -> r) -> c r) -> (forall r. r -> c r) -> Constr -> c DeserialiseFailure

toConstr :: DeserialiseFailure -> Constr

dataTypeOf :: DeserialiseFailure -> DataType

dataCast1 :: Typeable t => (forall d. Data d => c (t d)) -> Maybe (c DeserialiseFailure)

dataCast2 :: Typeable t => (forall d e. (Data d, Data e) => c (t d e)) -> Maybe (c DeserialiseFailure)

gmapT :: (forall b. Data b => b -> b) -> DeserialiseFailure -> DeserialiseFailure

gmapQl :: (r -> r' -> r) -> r -> (forall d. Data d => d -> r') -> DeserialiseFailure -> r

gmapQr :: forall r r'. (r' -> r -> r) -> r -> (forall d. Data d => d -> r') -> DeserialiseFailure -> r

gmapQ :: (forall d. Data d => d -> u) -> DeserialiseFailure -> [u]

gmapQi :: Int -> (forall d. Data d => d -> u) -> DeserialiseFailure -> u

gmapM :: Monad m => (forall d. Data d => d -> m d) -> DeserialiseFailure -> m DeserialiseFailure

gmapMp :: MonadPlus m => (forall d. Data d => d -> m d) -> DeserialiseFailure -> m DeserialiseFailure

gmapMo :: MonadPlus m => (forall d. Data d => d -> m d) -> DeserialiseFailure -> m DeserialiseFailure

Generic ApplyMempoolPayloadErr

Instance details

Associated Types

type Rep ApplyMempoolPayloadErr :: Type -> Type

Methods

from :: ApplyMempoolPayloadErr -> Rep ApplyMempoolPayloadErr x

to :: Rep ApplyMempoolPayloadErr x -> ApplyMempoolPayloadErr

Generic LovelaceError

Instance details

Associated Types

type Rep LovelaceError :: Type -> Type

Methods

from :: LovelaceError -> Rep LovelaceError x

to :: Rep LovelaceError x -> LovelaceError

Generic Error

Instance details

Associated Types

type Rep Error :: Type -> Type

Methods

from :: Error -> Rep Error x

to :: Rep Error x -> Error

Generic UTxOError

Instance details

Associated Types

type Rep UTxOError :: Type -> Type

Methods

from :: UTxOError -> Rep UTxOError x

to :: Rep UTxOError x -> UTxOError

Generic TxValidationError

Instance details

Associated Types

type Rep TxValidationError :: Type -> Type

Methods

from :: TxValidationError -> Rep TxValidationError x

to :: Rep TxValidationError x -> TxValidationError

Generic UTxOValidationError

Instance details

Associated Types

type Rep UTxOValidationError :: Type -> Type

Methods

from :: UTxOValidationError -> Rep UTxOValidationError x

to :: Rep UTxOValidationError x -> UTxOValidationError

Generic ApplicationNameError

Instance details

Associated Types

type Rep ApplicationNameError :: Type -> Type

Methods

from :: ApplicationNameError -> Rep ApplicationNameError x

to :: Rep ApplicationNameError x -> ApplicationNameError

Generic SoftwareVersionError

Instance details

Associated Types

type Rep SoftwareVersionError :: Type -> Type

Methods

from :: SoftwareVersionError -> Rep SoftwareVersionError x

to :: Rep SoftwareVersionError x -> SoftwareVersionError

Generic SystemTagError

Instance details

Associated Types

type Rep SystemTagError :: Type -> Type

Methods

from :: SystemTagError -> Rep SystemTagError x

to :: Rep SystemTagError x -> SystemTagError

Generic Error

Instance details

Associated Types

type Rep Error :: Type -> Type

Methods

from :: Error -> Rep Error x

to :: Rep Error x -> Error

Generic Error

Instance details

Associated Types

type Rep Error :: Type -> Type

Methods

from :: Error -> Rep Error x

to :: Rep Error x -> Error

Generic Adopted

Instance details

Associated Types

type Rep Adopted :: Type -> Type

Methods

from :: Adopted -> Rep Adopted x

to :: Rep Adopted x -> Adopted

Generic Error

Instance details

Associated Types

type Rep Error :: Type -> Type

Methods

from :: Error -> Rep Error x

to :: Rep Error x -> Error

Generic Error

Instance details

Associated Types

type Rep Error :: Type -> Type

Methods

from :: Error -> Rep Error x

to :: Rep Error x -> Error

Num Coin

Instance details

Methods

(+) :: Coin -> Coin -> Coin

(-) :: Coin -> Coin -> Coin

(*) :: Coin -> Coin -> Coin

negate :: Coin -> Coin

abs :: Coin -> Coin

signum :: Coin -> Coin

fromInteger :: Integer -> Coin

Integral Coin

Instance details

Methods

quot :: Coin -> Coin -> Coin

rem :: Coin -> Coin -> Coin

div :: Coin -> Coin -> Coin

mod :: Coin -> Coin -> Coin

quotRem :: Coin -> Coin -> (Coin, Coin)

divMod :: Coin -> Coin -> (Coin, Coin)

toInteger :: Coin -> Integer

Real Coin
These instances originally existed on the Lovelace type. As the Lovelace type is deleted and we use L.Coin instead, these instances are added to L.Coin. The instances are purely for the convenience of writing expressions involving L.Coin but be aware that not all uses of these typeclasses are valid.

Instance details

Methods

toRational :: Coin -> Rational

Pretty Coin

Instance details

Methods

pretty :: Coin -> Doc ann

prettyList :: [Coin] -> Doc ann

Pretty MuxError

Instance details

Methods

pretty :: MuxError -> Doc ann

prettyList :: [MuxError] -> Doc ann

Show (AllegraUtxoPredFailure ledgerera) => ToJSON (AllegraUtxoPredFailure ledgerera)

Instance details

Methods

toJSON :: AllegraUtxoPredFailure ledgerera -> Value

toEncoding :: AllegraUtxoPredFailure ledgerera -> Encoding

toJSONList :: [AllegraUtxoPredFailure ledgerera] -> Value

toEncodingList :: [AllegraUtxoPredFailure ledgerera] -> Encoding

omitField :: AllegraUtxoPredFailure ledgerera -> Bool

Show (AlonzoUtxoPredFailure ledgerera) => ToJSON (AlonzoUtxoPredFailure ledgerera)

Instance details

Methods

toJSON :: AlonzoUtxoPredFailure ledgerera -> Value

toEncoding :: AlonzoUtxoPredFailure ledgerera -> Encoding

toJSONList :: [AlonzoUtxoPredFailure ledgerera] -> Value

toEncodingList :: [AlonzoUtxoPredFailure ledgerera] -> Encoding

omitField :: AlonzoUtxoPredFailure ledgerera -> Bool

(Crypto (EraCrypto ledgerera), ToJSON (PredicateFailure (EraRule "UTXO" ledgerera)), ToJSON (PlutusPurpose AsItem ledgerera), ToJSON (PlutusPurpose AsIx ledgerera)) => ToJSON (AlonzoUtxowPredFailure ledgerera)

Instance details

Methods

toJSON :: AlonzoUtxowPredFailure ledgerera -> Value

toEncoding :: AlonzoUtxowPredFailure ledgerera -> Encoding

toJSONList :: [AlonzoUtxowPredFailure ledgerera] -> Value

toEncodingList :: [AlonzoUtxowPredFailure ledgerera] -> Encoding

omitField :: AlonzoUtxowPredFailure ledgerera -> Bool

Show (BabbageUtxoPredFailure ledgerera) => ToJSON (BabbageUtxoPredFailure ledgerera)

Instance details

Methods

toJSON :: BabbageUtxoPredFailure ledgerera -> Value

toEncoding :: BabbageUtxoPredFailure ledgerera -> Encoding

toJSONList :: [BabbageUtxoPredFailure ledgerera] -> Value

toEncodingList :: [BabbageUtxoPredFailure ledgerera] -> Encoding

omitField :: BabbageUtxoPredFailure ledgerera -> Bool

(Crypto (EraCrypto ledgerera), ToJSON (PredicateFailure (EraRule "UTXO" ledgerera)), ToJSON (TxCert ledgerera), ToJSON (PlutusPurpose AsItem ledgerera), ToJSON (PlutusPurpose AsIx ledgerera)) => ToJSON (BabbageUtxowPredFailure ledgerera)

Instance details

Methods

toJSON :: BabbageUtxowPredFailure ledgerera -> Value

toEncoding :: BabbageUtxowPredFailure ledgerera -> Encoding

toJSONList :: [BabbageUtxowPredFailure ledgerera] -> Value

toEncodingList :: [BabbageUtxowPredFailure ledgerera] -> Encoding

omitField :: BabbageUtxowPredFailure ledgerera -> Bool

ToJSON a => ToJSON (TooLarge a)

Instance details

Methods

toJSON :: TooLarge a -> Value

toEncoding :: TooLarge a -> Encoding

toJSONList :: [TooLarge a] -> Value

toEncodingList :: [TooLarge a] -> Encoding

omitField :: TooLarge a -> Bool

Show (ConwayLedgerPredFailure ledgerera) => ToJSON (ConwayLedgerPredFailure ledgerera)

Instance details

Methods

toJSON :: ConwayLedgerPredFailure ledgerera -> Value

toEncoding :: ConwayLedgerPredFailure ledgerera -> Encoding

toJSONList :: [ConwayLedgerPredFailure ledgerera] -> Value

toEncodingList :: [ConwayLedgerPredFailure ledgerera] -> Encoding

omitField :: ConwayLedgerPredFailure ledgerera -> Bool

ToJSON (PredicateFailure (EraRule "LEDGER" ledgerera)) => ToJSON (ApplyTxError ledgerera)

Instance details

Methods

toJSON :: ApplyTxError ledgerera -> Value

toEncoding :: ApplyTxError ledgerera -> Encoding

toJSONList :: [ApplyTxError ledgerera] -> Value

toEncodingList :: [ApplyTxError ledgerera] -> Encoding

omitField :: ApplyTxError ledgerera -> Bool

Show (ShelleyDelegsPredFailure ledgerera) => ToJSON (ShelleyDelegsPredFailure ledgerera)

Instance details

Methods

toJSON :: ShelleyDelegsPredFailure ledgerera -> Value

toEncoding :: ShelleyDelegsPredFailure ledgerera -> Encoding

toJSONList :: [ShelleyDelegsPredFailure ledgerera] -> Value

toEncodingList :: [ShelleyDelegsPredFailure ledgerera] -> Encoding

omitField :: ShelleyDelegsPredFailure ledgerera -> Bool

(ToJSON (PredicateFailure (EraRule "UTXOW" ledgerera)), ToJSON (PredicateFailure (EraRule "DELEGS" ledgerera))) => ToJSON (ShelleyLedgerPredFailure ledgerera)

Instance details

Methods

toJSON :: ShelleyLedgerPredFailure ledgerera -> Value

toEncoding :: ShelleyLedgerPredFailure ledgerera -> Encoding

toJSONList :: [ShelleyLedgerPredFailure ledgerera] -> Value

toEncodingList :: [ShelleyLedgerPredFailure ledgerera] -> Encoding

omitField :: ShelleyLedgerPredFailure ledgerera -> Bool

(Crypto (EraCrypto ledgerera), ToJSON (PredicateFailure (EraRule "UTXO" ledgerera))) => ToJSON (ShelleyPpupPredFailure ledgerera)

Instance details

Methods

toJSON :: ShelleyPpupPredFailure ledgerera -> Value

toEncoding :: ShelleyPpupPredFailure ledgerera -> Encoding

toJSONList :: [ShelleyPpupPredFailure ledgerera] -> Value

toEncodingList :: [ShelleyPpupPredFailure ledgerera] -> Encoding

omitField :: ShelleyPpupPredFailure ledgerera -> Bool

Show (ShelleyUtxoPredFailure ledgerera) => ToJSON (ShelleyUtxoPredFailure ledgerera)

Instance details

Methods

toJSON :: ShelleyUtxoPredFailure ledgerera -> Value

toEncoding :: ShelleyUtxoPredFailure ledgerera -> Encoding

toJSONList :: [ShelleyUtxoPredFailure ledgerera] -> Value

toEncodingList :: [ShelleyUtxoPredFailure ledgerera] -> Encoding

omitField :: ShelleyUtxoPredFailure ledgerera -> Bool

(Crypto (EraCrypto ledgerera), ToJSON (PredicateFailure (EraRule "UTXO" ledgerera))) => ToJSON (ShelleyUtxowPredFailure ledgerera)

Instance details

Methods

toJSON :: ShelleyUtxowPredFailure ledgerera -> Value

toEncoding :: ShelleyUtxowPredFailure ledgerera -> Encoding

toJSONList :: [ShelleyUtxowPredFailure ledgerera] -> Value

toEncodingList :: [ShelleyUtxowPredFailure ledgerera] -> Encoding

omitField :: ShelleyUtxowPredFailure ledgerera -> Bool

ToJSON (ChainDepState StandardCrypto)

Instance details

Methods

toJSON :: ChainDepState StandardCrypto -> Value

toEncoding :: ChainDepState StandardCrypto -> Encoding

toJSONList :: [ChainDepState StandardCrypto] -> Value

toEncodingList :: [ChainDepState StandardCrypto] -> Encoding

omitField :: ChainDepState StandardCrypto -> Bool

Crypto crypto => ToJSON (HashHeader crypto)

Instance details

Methods

toJSON :: HashHeader crypto -> Value

toEncoding :: HashHeader crypto -> Encoding

toJSONList :: [HashHeader crypto] -> Value

toEncodingList :: [HashHeader crypto] -> Encoding

omitField :: HashHeader crypto -> Bool

ToJSON (PrtclState StandardCrypto)

Instance details

Methods

toJSON :: PrtclState StandardCrypto -> Value

toEncoding :: PrtclState StandardCrypto -> Encoding

toJSONList :: [PrtclState StandardCrypto] -> Value

toEncodingList :: [PrtclState StandardCrypto] -> Encoding

omitField :: PrtclState StandardCrypto -> Bool

ToJSON (StakeSnapshot crypto)

Instance details

Methods

toJSON :: StakeSnapshot crypto -> Value

toEncoding :: StakeSnapshot crypto -> Encoding

toJSONList :: [StakeSnapshot crypto] -> Value

toEncodingList :: [StakeSnapshot crypto] -> Encoding

omitField :: StakeSnapshot crypto -> Bool

Crypto crypto => ToJSON (StakeSnapshots crypto)

Instance details

Methods

toJSON :: StakeSnapshots crypto -> Value

toEncoding :: StakeSnapshots crypto -> Encoding

toJSONList :: [StakeSnapshots crypto] -> Value

toEncodingList :: [StakeSnapshots crypto] -> Encoding

omitField :: StakeSnapshots crypto -> Bool

Crypto crypto => ToJSON (ShelleyHash crypto)

Instance details

Methods

toJSON :: ShelleyHash crypto -> Value

toEncoding :: ShelleyHash crypto -> Encoding

toJSONList :: [ShelleyHash crypto] -> Value

toEncodingList :: [ShelleyHash crypto] -> Encoding

omitField :: ShelleyHash crypto -> Bool

ToJSON (PraosState StandardCrypto)

Instance details

Methods

toJSON :: PraosState StandardCrypto -> Value

toEncoding :: PraosState StandardCrypto -> Encoding

toJSONList :: [PraosState StandardCrypto] -> Value

toEncodingList :: [PraosState StandardCrypto] -> Encoding

omitField :: PraosState StandardCrypto -> Bool

ToJSON (TPraosState StandardCrypto)

Instance details

Methods

toJSON :: TPraosState StandardCrypto -> Value

toEncoding :: TPraosState StandardCrypto -> Encoding

toJSONList :: [TPraosState StandardCrypto] -> Value

toEncodingList :: [TPraosState StandardCrypto] -> Encoding

omitField :: TPraosState StandardCrypto -> Bool

Generic (TooLarge a)

Instance details

Associated Types

type Rep (TooLarge a) :: Type -> Type

Methods

from :: TooLarge a -> Rep (TooLarge a) x

to :: Rep (TooLarge a) x -> TooLarge a

Generic (ApplyTxError era)

Instance details

Associated Types

type Rep (ApplyTxError era) :: Type -> Type

Methods

from :: ApplyTxError era -> Rep (ApplyTxError era) x

to :: Rep (ApplyTxError era) x -> ApplyTxError era

Crypto c => ToJSON (VKey 'Witness c)

Instance details

Methods

toJSON :: VKey 'Witness c -> Value

toEncoding :: VKey 'Witness c -> Encoding

toJSONList :: [VKey 'Witness c] -> Value

toEncodingList :: [VKey 'Witness c] -> Encoding

omitField :: VKey 'Witness c -> Bool

ToJSON (OneEraHash xs)

Instance details

Methods

toJSON :: OneEraHash xs -> Value

toEncoding :: OneEraHash xs -> Encoding

toJSONList :: [OneEraHash xs] -> Value

toEncodingList :: [OneEraHash xs] -> Encoding

omitField :: OneEraHash xs -> Bool

ToJSON (HeaderHash blk) => ToJSON (Tip blk)

Instance details

Methods

toJSON :: Tip blk -> Value

toEncoding :: Tip blk -> Encoding

toJSONList :: [Tip blk] -> Value

toEncodingList :: [Tip blk] -> Encoding

omitField :: Tip blk -> Bool

Semigroup (AlonzoPParams StrictMaybe era)

Instance details

Methods

(<>) :: AlonzoPParams StrictMaybe era -> AlonzoPParams StrictMaybe era -> AlonzoPParams StrictMaybe era

sconcat :: NonEmpty (AlonzoPParams StrictMaybe era) -> AlonzoPParams StrictMaybe era

stimes :: Integral b => b -> AlonzoPParams StrictMaybe era -> AlonzoPParams StrictMaybe era

Semigroup (BabbagePParams StrictMaybe era)

Instance details

Methods

(<>) :: BabbagePParams StrictMaybe era -> BabbagePParams StrictMaybe era -> BabbagePParams StrictMaybe era

sconcat :: NonEmpty (BabbagePParams StrictMaybe era) -> BabbagePParams StrictMaybe era

stimes :: Integral b => b -> BabbagePParams StrictMaybe era -> BabbagePParams StrictMaybe era

Semigroup (ConwayPParams StrictMaybe era)

Instance details

Methods

(<>) :: ConwayPParams StrictMaybe era -> ConwayPParams StrictMaybe era -> ConwayPParams StrictMaybe era

sconcat :: NonEmpty (ConwayPParams StrictMaybe era) -> ConwayPParams StrictMaybe era

stimes :: Integral b => b -> ConwayPParams StrictMaybe era -> ConwayPParams StrictMaybe era

Semigroup (ShelleyPParams StrictMaybe era)

Instance details

Methods

(<>) :: ShelleyPParams StrictMaybe era -> ShelleyPParams StrictMaybe era -> ShelleyPParams StrictMaybe era

sconcat :: NonEmpty (ShelleyPParams StrictMaybe era) -> ShelleyPParams StrictMaybe era

stimes :: Integral b => b -> ShelleyPParams StrictMaybe era -> ShelleyPParams StrictMaybe era