Cardano.Api.Parser.Text

Documentation

runParser

Arguments

:: Parser a

-> Text
input text

-> Either String a

Run parser

parseDecimal :: Parser Integer

Non-negative decimal numbers parser

parseWord64 :: Parser Integer

Word64 parser.

runParserFail :: MonadFail m => Parser a -> Text -> m a

Run parser in MonadFail

module Control.Applicative

module Data.Functor

class Monad m => Stream s (m :: Type -> Type) t | s -> t where

An instance of Stream has stream type s, underlying monad m and token type t determined by the stream

Some rough guidelines for a "correct" instance of Stream:

unfoldM uncons gives the [t] corresponding to the stream

A Stream instance is responsible for maintaining the "position within the stream" in the stream state s. This is trivial unless you are using the monad in a non-trivial way.

Methods

uncons :: s -> m (Maybe (t, s))

try :: forall s u (m :: Type -> Type) a. ParsecT s u m a -> ParsecT s u m a

The parser try p behaves like parser p, except that it pretends that it hasn't consumed any input when an error occurs.

This combinator is used whenever arbitrary look ahead is needed. Since it pretends that it hasn't consumed any input when p fails, the (<|>) combinator will try its second alternative even when the first parser failed while consuming input.

The try combinator can for example be used to distinguish identifiers and reserved words. Both reserved words and identifiers are a sequence of letters. Whenever we expect a certain reserved word where we can also expect an identifier we have to use the try combinator. Suppose we write:

 expr        = letExpr <|> identifier <?> "expression"

 letExpr     = do{ string "let"; ... }
 identifier  = many1 letter

If the user writes "lexical", the parser fails with: unexpected 'x', expecting 't' in "let". Indeed, since the (<|>) combinator only tries alternatives when the first alternative hasn't consumed input, the identifier parser is never tried (because the prefix "le" of the string "let" parser is already consumed). The right behaviour can be obtained by adding the try combinator:

 expr        = letExpr <|> identifier <?> "expression"

 letExpr     = do{ try (string "let"); ... }
 identifier  = many1 letter

satisfy :: forall s (m :: Type -> Type) u. Stream s m Char => (Char -> Bool) -> ParsecT s u m Char

The parser satisfy f succeeds for any character for which the supplied function f returns True. Returns the character that is actually parsed.

char :: forall s (m :: Type -> Type) u. Stream s m Char => Char -> ParsecT s u m Char

char c parses a single character c. Returns the parsed character (i.e. c).

 semiColon  = char ';'

eof :: forall s (m :: Type -> Type) t u. (Stream s m t, Show t) => ParsecT s u m ()

This parser only succeeds at the end of the input. This is not a primitive parser but it is defined using notFollowedBy.

 eof  = notFollowedBy anyToken <?> "end of input"

string :: forall s (m :: Type -> Type) u. Stream s m Char => String -> ParsecT s u m String

string s parses a sequence of characters given by s. Returns the parsed string (i.e. s).

 divOrMod    =   string "div"
             <|> string "mod"

Consider using string'.

choice :: forall s (m :: Type -> Type) t u a. Stream s m t => [ParsecT s u m a] -> ParsecT s u m a

choice ps tries to apply the parsers in the list ps in order, until one of them succeeds. Returns the value of the succeeding parser.

count :: forall s (m :: Type -> Type) t u a. Stream s m t => Int -> ParsecT s u m a -> ParsecT s u m [a]

count n p parses n occurrences of p. If n is smaller or equal to zero, the parser equals to return []. Returns a list of n values returned by p.

between :: forall s (m :: Type -> Type) t u open close a. Stream s m t => ParsecT s u m open -> ParsecT s u m close -> ParsecT s u m a -> ParsecT s u m a

between open close p parses open, followed by p and close. Returns the value returned by p.

 braces  = between (symbol "{") (symbol "}")

option :: forall s (m :: Type -> Type) t a u. Stream s m t => a -> ParsecT s u m a -> ParsecT s u m a

option x p tries to apply parser p. If p fails without consuming input, it returns the value x, otherwise the value returned by p.

 priority  = option 0 (do{ d <- digit
                         ; return (digitToInt d)
                         })

many1 :: forall s u (m :: Type -> Type) a. ParsecT s u m a -> ParsecT s u m [a]

many1 p applies the parser p one or more times. Returns a list of the returned values of p.

 word  = many1 letter

skipMany :: forall s u (m :: Type -> Type) a. ParsecT s u m a -> ParsecT s u m ()

skipMany p applies the parser p zero or more times, skipping its result.

 spaces  = skipMany space

skipMany1 :: forall s (m :: Type -> Type) t u a. Stream s m t => ParsecT s u m a -> ParsecT s u m ()

skipMany1 p applies the parser p one or more times, skipping its result.

sepBy :: forall s (m :: Type -> Type) t u a sep. Stream s m t => ParsecT s u m a -> ParsecT s u m sep -> ParsecT s u m [a]

sepBy p sep parses zero or more occurrences of p, separated by sep. Returns a list of values returned by p.

 commaSep p  = p `sepBy` (symbol ",")

sepBy1 :: forall s (m :: Type -> Type) t u a sep. Stream s m t => ParsecT s u m a -> ParsecT s u m sep -> ParsecT s u m [a]

sepBy1 p sep parses one or more occurrences of p, separated by sep. Returns a list of values returned by p.

endBy :: forall s (m :: Type -> Type) t u a sep. Stream s m t => ParsecT s u m a -> ParsecT s u m sep -> ParsecT s u m [a]

endBy p sep parses zero or more occurrences of p, separated and ended by sep. Returns a list of values returned by p.

  cStatements  = cStatement `endBy` semi

endBy1 :: forall s (m :: Type -> Type) t u a sep. Stream s m t => ParsecT s u m a -> ParsecT s u m sep -> ParsecT s u m [a]

endBy1 p sep parses one or more occurrences of p, separated and ended by sep. Returns a list of values returned by p.

chainr :: forall s (m :: Type -> Type) t u a. Stream s m t => ParsecT s u m a -> ParsecT s u m (a -> a -> a) -> a -> ParsecT s u m a

chainr p op x parses zero or more occurrences of p, separated by op Returns a value obtained by a right associative application of all functions returned by op to the values returned by p. If there are no occurrences of p, the value x is returned.

chainl :: forall s (m :: Type -> Type) t u a. Stream s m t => ParsecT s u m a -> ParsecT s u m (a -> a -> a) -> a -> ParsecT s u m a

chainl p op x parses zero or more occurrences of p, separated by op. Returns a value obtained by a left associative application of all functions returned by op to the values returned by p. If there are zero occurrences of p, the value x is returned.

chainr1 :: forall s (m :: Type -> Type) t u a. Stream s m t => ParsecT s u m a -> ParsecT s u m (a -> a -> a) -> ParsecT s u m a

chainr1 p op x parses one or more occurrences of |p|, separated by op Returns a value obtained by a right associative application of all functions returned by op to the values returned by p.

chainl1 :: forall s (m :: Type -> Type) t u a. Stream s m t => ParsecT s u m a -> ParsecT s u m (a -> a -> a) -> ParsecT s u m a

chainl1 p op parses one or more occurrences of p, separated by op Returns a value obtained by a left associative application of all functions returned by op to the values returned by p. This parser can for example be used to eliminate left recursion which typically occurs in expression grammars.

 expr    = term   `chainl1` addop
 term    = factor `chainl1` mulop
 factor  = parens expr <|> integer

 mulop   =   do{ symbol "*"; return (*)   }
         <|> do{ symbol "/"; return (div) }

 addop   =   do{ symbol "+"; return (+) }
         <|> do{ symbol "-"; return (-) }

manyTill :: forall s (m :: Type -> Type) t u a end. Stream s m t => ParsecT s u m a -> ParsecT s u m end -> ParsecT s u m [a]

manyTill p end applies parser p zero or more times until parser end succeeds. Returns the list of values returned by p. This parser can be used to scan comments:

 simpleComment   = do{ string "<!--"
                     ; manyTill anyChar (try (string "-->"))
                     }

Note the overlapping parsers anyChar and string "-->", and therefore the use of the try combinator.

getState :: forall (m :: Type -> Type) s u. Monad m => ParsecT s u m u

Returns the current user state.

setState :: forall (m :: Type -> Type) u s. Monad m => u -> ParsecT s u m ()

An alias for putState for backwards compatibility.

sourceColumn :: SourcePos -> Column

Extracts the column number from a source position.

sourceLine :: SourcePos -> Line

Extracts the line number from a source position.

label :: forall s u (m :: Type -> Type) a. ParsecT s u m a -> String -> ParsecT s u m a

A synonym for <?>, but as a function instead of an operator.

lookAhead :: forall s (m :: Type -> Type) t u a. Stream s m t => ParsecT s u m a -> ParsecT s u m a

lookAhead p parses p without consuming any input.

If p fails and consumes some input, so does lookAhead. Combine with try if this is undesirable.

data Consumed a

Constructors

Consumed a

Empty !a

data State s u

Constructors

State

Fields

stateInput :: s

statePos :: !SourcePos

stateUser :: !u

data Reply s u a

Constructors

Ok a !(State s u) ParseError

Error ParseError

data ParseError

The abstract data type ParseError represents parse errors. It provides the source position (SourcePos) of the error and a list of error messages (Message). A ParseError can be returned by the function parse. ParseError is an instance of the Show and Eq classes.

updateState :: forall (m :: Type -> Type) u s. Monad m => (u -> u) -> ParsecT s u m ()

An alias for modifyState for backwards compatibility.

parse :: Stream s Identity t => Parsec s () a -> SourceName -> s -> Either ParseError a

parse p filePath input runs a parser p over Identity without user state. The filePath is only used in error messages and may be the empty string. Returns either a ParseError (Left) or a value of type a (Right).

 main    = case (parse numbers "" "11, 2, 43") of
            Left err  -> print err
            Right xs  -> print (sum xs)

 numbers = commaSep integer

unexpected :: forall s (m :: Type -> Type) t u a. Stream s m t => String -> ParsecT s u m a

The parser unexpected msg always fails with an unexpected error message msg without consuming any input.

The parsers fail, (<?>) and unexpected are the three parsers used to generate error messages. Of these, only (<?>) is commonly used. For an example of the use of unexpected, see the definition of notFollowedBy.

(<?>) :: forall s u (m :: Type -> Type) a. ParsecT s u m a -> String -> ParsecT s u m a infix 0

The parser p <?> msg behaves as parser p, but whenever the parser p fails without consuming any input, it replaces expect error messages with the expect error message msg.

This is normally used at the end of a set alternatives where we want to return an error message in terms of a higher level construct rather than returning all possible characters. For example, if the expr parser from the try example would fail, the error message is: '...: expecting expression'. Without the (<?>) combinator, the message would be like '...: expecting "let" or letter', which is less friendly.

notFollowedBy :: forall s (m :: Type -> Type) t a u. (Stream s m t, Show a) => ParsecT s u m a -> ParsecT s u m ()

notFollowedBy p only succeeds when parser p fails. This parser does not consume any input. This parser can be used to implement the 'longest match' rule. For example, when recognizing keywords (for example let), we want to make sure that a keyword is not followed by a legal identifier character, in which case the keyword is actually an identifier (for example lets). We can program this behaviour as follows:

 keywordLet  = try (do{ string "let"
                      ; notFollowedBy alphaNum
                      })

NOTE: Currently, notFollowedBy exhibits surprising behaviour when applied to a parser p that doesn't consume any input; specifically

notFollowedBy . notFollowedBy is not equivalent to lookAhead, and

notFollowedBy eof never fails.

See haskell/parsec#8 for more details.

data SourcePos

The abstract data type SourcePos represents source positions. It contains the name of the source (i.e. file name), a line number and a column number. SourcePos is an instance of the Show, Eq and Ord class.

type Column = Int

type Line = Int

type SourceName = String

sourceName :: SourcePos -> SourceName

Extracts the name of the source from a source position.

incSourceLine :: SourcePos -> Line -> SourcePos

Increments the line number of a source position.

incSourceColumn :: SourcePos -> Column -> SourcePos

Increments the column number of a source position.

setSourceName :: SourcePos -> SourceName -> SourcePos

Set the name of the source.

setSourceLine :: SourcePos -> Line -> SourcePos

Set the line number of a source position.

setSourceColumn :: SourcePos -> Column -> SourcePos

Set the column number of a source position.

errorPos :: ParseError -> SourcePos

Extracts the source position from the parse error

type Parsec s u = ParsecT s u Identity

data ParsecT s u (m :: Type -> Type) a

ParserT monad transformer and Parser type

ParsecT s u m a is a parser with stream type s, user state type u, underlying monad m and return type a. Parsec is strict in the user state. If this is undesirable, simply use a data type like data Box a = Box a and the state type Box YourStateType to add a level of indirection.

unknownError :: State s u -> ParseError

sysUnExpectError :: String -> SourcePos -> Reply s u a

runParsecT :: Monad m => ParsecT s u m a -> State s u -> m (Consumed (m (Reply s u a)))

Low-level unpacking of the ParsecT type. To run your parser, please look to runPT, runP, runParserT, runParser and other such functions.

mkPT :: Monad m => (State s u -> m (Consumed (m (Reply s u a)))) -> ParsecT s u m a

Low-level creation of the ParsecT type. You really shouldn't have to do this.

parsecMap :: forall a b s u (m :: Type -> Type). (a -> b) -> ParsecT s u m a -> ParsecT s u m b

parserReturn :: forall a s u (m :: Type -> Type). a -> ParsecT s u m a

parserBind :: forall s u (m :: Type -> Type) a b. ParsecT s u m a -> (a -> ParsecT s u m b) -> ParsecT s u m b

mergeErrorReply :: ParseError -> Reply s u a -> Reply s u a

parserFail :: forall s u (m :: Type -> Type) a. String -> ParsecT s u m a

parserZero :: forall s u (m :: Type -> Type) a. ParsecT s u m a

parserZero always fails without consuming any input. parserZero is defined equal to the mzero member of the MonadPlus class and to the empty member of the Alternative class.

parserPlus :: forall s u (m :: Type -> Type) a. ParsecT s u m a -> ParsecT s u m a -> ParsecT s u m a

labels :: forall s u (m :: Type -> Type) a. ParsecT s u m a -> [String] -> ParsecT s u m a

tokens :: forall s (m :: Type -> Type) t u. (Stream s m t, Eq t) => ([t] -> String) -> (SourcePos -> [t] -> SourcePos) -> [t] -> ParsecT s u m [t]

token

Arguments

:: Stream s Identity t

=> (t -> String)
Token pretty-printing function.

-> (t -> SourcePos)
Computes the position of a token.

-> (t -> Maybe a)
Matching function for the token to parse.

-> Parsec s u a

The parser token showTok posFromTok testTok accepts a token t with result x when the function testTok t returns Just x. The source position of the t should be returned by posFromTok t and the token can be shown using showTok t.

This combinator is expressed in terms of tokenPrim. It is used to accept user defined token streams. For example, suppose that we have a stream of basic tokens tupled with source positions. We can then define a parser that accepts single tokens as:

 mytoken x
   = token showTok posFromTok testTok
   where
     showTok (pos,t)     = show t
     posFromTok (pos,t)  = pos
     testTok (pos,t)     = if x == t then Just t else Nothing

tokenPrim

Arguments

:: forall s (m :: Type -> Type) t a u. Stream s m t

=> (t -> String)
Token pretty-printing function.

-> (SourcePos -> t -> s -> SourcePos)
Next position calculating function.

-> (t -> Maybe a)
Matching function for the token to parse.

-> ParsecT s u m a

The parser tokenPrim showTok nextPos testTok accepts a token t with result x when the function testTok t returns Just x. The token can be shown using showTok t. The position of the next token should be returned when nextPos is called with the current source position pos, the current token t and the rest of the tokens toks, nextPos pos t toks.

This is the most primitive combinator for accepting tokens. For example, the char parser could be implemented as:

 char c
   = tokenPrim showChar nextPos testChar
   where
     showChar x        = "'" ++ x ++ "'"
     testChar x        = if x == c then Just x else Nothing
     nextPos pos x xs  = updatePosChar pos x

tokenPrimEx :: forall s (m :: Type -> Type) t u a. Stream s m t => (t -> String) -> (SourcePos -> t -> s -> SourcePos) -> Maybe (SourcePos -> t -> s -> u -> u) -> (t -> Maybe a) -> ParsecT s u m a

manyAccum :: forall a s u (m :: Type -> Type). (a -> [a] -> [a]) -> ParsecT s u m a -> ParsecT s u m [a]

runPT :: Stream s m t => ParsecT s u m a -> u -> SourceName -> s -> m (Either ParseError a)

runP :: Stream s Identity t => Parsec s u a -> u -> SourceName -> s -> Either ParseError a

runParserT :: Stream s m t => ParsecT s u m a -> u -> SourceName -> s -> m (Either ParseError a)

The most general way to run a parser. runParserT p state filePath input runs parser p on the input list of tokens input, obtained from source filePath with the initial user state st. The filePath is only used in error messages and may be the empty string. Returns a computation in the underlying monad m that return either a ParseError (Left) or a value of type a (Right).

parseTest :: (Stream s Identity t, Show a) => Parsec s () a -> s -> IO ()

The expression parseTest p input applies a parser p against input input and prints the result to stdout. Used for testing parsers.

getPosition :: forall (m :: Type -> Type) s u. Monad m => ParsecT s u m SourcePos

Returns the current source position. See also SourcePos.

getInput :: forall (m :: Type -> Type) s u. Monad m => ParsecT s u m s

Returns the current input

setPosition :: forall (m :: Type -> Type) s u. Monad m => SourcePos -> ParsecT s u m ()

setPosition pos sets the current source position to pos.

setInput :: forall (m :: Type -> Type) s u. Monad m => s -> ParsecT s u m ()

setInput input continues parsing with input. The getInput and setInput functions can for example be used to deal with #include files.

getParserState :: forall (m :: Type -> Type) s u. Monad m => ParsecT s u m (State s u)

Returns the full parser state as a State record.

setParserState :: forall (m :: Type -> Type) s u. Monad m => State s u -> ParsecT s u m (State s u)

setParserState st set the full parser state to st.

updateParserState :: forall s u (m :: Type -> Type). (State s u -> State s u) -> ParsecT s u m (State s u)

updateParserState f applies function f to the parser state.

putState :: forall (m :: Type -> Type) u s. Monad m => u -> ParsecT s u m ()

putState st set the user state to st.

modifyState :: forall (m :: Type -> Type) u s. Monad m => (u -> u) -> ParsecT s u m ()

modifyState f applies function f to the user state. Suppose that we want to count identifiers in a source, we could use the user state as:

 expr  = do{ x <- identifier
           ; modifyState (+1)
           ; return (Id x)
           }

optionMaybe :: forall s (m :: Type -> Type) t u a. Stream s m t => ParsecT s u m a -> ParsecT s u m (Maybe a)

optionMaybe p tries to apply parser p. If p fails without consuming input, it return Nothing, otherwise it returns Just the value returned by p.

sepEndBy1 :: forall s (m :: Type -> Type) t u a sep. Stream s m t => ParsecT s u m a -> ParsecT s u m sep -> ParsecT s u m [a]

sepEndBy1 p sep parses one or more occurrences of p, separated and optionally ended by sep. Returns a list of values returned by p.

sepEndBy :: forall s (m :: Type -> Type) t u a sep. Stream s m t => ParsecT s u m a -> ParsecT s u m sep -> ParsecT s u m [a]

sepEndBy p sep parses zero or more occurrences of p, separated and optionally ended by sep, ie. haskell style statements. Returns a list of values returned by p.

 haskellStatements  = haskellStatement `sepEndBy` semi

anyToken :: forall s (m :: Type -> Type) t u. (Stream s m t, Show t) => ParsecT s u m t

The parser anyToken accepts any kind of token. It is for example used to implement eof. Returns the accepted token.

parserTrace :: forall t s (m :: Type -> Type) u. (Show t, Stream s m t) => String -> ParsecT s u m ()

parserTrace label is an impure function, implemented with Debug.Trace that prints to the console the remaining parser state at the time it is invoked. It is intended to be used for debugging parsers by inspecting their intermediate states.

*> parseTest (oneOf "aeiou"  >> parserTrace "label") "atest"
label: "test"
...

Since: parsec-3.1.12.0

parserTraced :: forall s (m :: Type -> Type) t u b. (Stream s m t, Show t) => String -> ParsecT s u m b -> ParsecT s u m b

parserTraced label p is an impure function, implemented with Debug.Trace that prints to the console the remaining parser state at the time it is invoked. It then continues to apply parser p, and if p fails will indicate that the label has been backtracked. It is intended to be used for debugging parsers by inspecting their intermediate states.

*>  parseTest (oneOf "aeiou"  >> parserTraced "label" (oneOf "nope")) "atest"
label: "test"
label backtracked
parse error at (line 1, column 2):
...

Since: parsec-3.1.12.0

oneOf :: forall s (m :: Type -> Type) u. Stream s m Char => [Char] -> ParsecT s u m Char

oneOf cs succeeds if the current character is in the supplied list of characters cs. Returns the parsed character. See also satisfy.

  vowel  = oneOf "aeiou"

noneOf :: forall s (m :: Type -> Type) u. Stream s m Char => [Char] -> ParsecT s u m Char

As the dual of oneOf, noneOf cs succeeds if the current character not in the supplied list of characters cs. Returns the parsed character.

 consonant = noneOf "aeiou"

spaces :: forall s (m :: Type -> Type) u. Stream s m Char => ParsecT s u m ()

Skips zero or more white space characters. See also skipMany.

space :: forall s (m :: Type -> Type) u. Stream s m Char => ParsecT s u m Char

Parses a white space character (any character which satisfies isSpace) Returns the parsed character.

newline :: forall s (m :: Type -> Type) u. Stream s m Char => ParsecT s u m Char

Parses a newline character ('\n'). Returns a newline character.

crlf :: forall s (m :: Type -> Type) u. Stream s m Char => ParsecT s u m Char

Parses a carriage return character ('\r') followed by a newline character ('\n'). Returns a newline character.

endOfLine :: forall s (m :: Type -> Type) u. Stream s m Char => ParsecT s u m Char

Parses a CRLF (see crlf) or LF (see newline) end-of-line. Returns a newline character ('\n').

endOfLine = newline <|> crlf

tab :: forall s (m :: Type -> Type) u. Stream s m Char => ParsecT s u m Char

Parses a tab character ('\t'). Returns a tab character.

upper :: forall s (m :: Type -> Type) u. Stream s m Char => ParsecT s u m Char

Parses an upper case letter (according to isUpper). Returns the parsed character.

lower :: forall s (m :: Type -> Type) u. Stream s m Char => ParsecT s u m Char

Parses a lower case character (according to isLower). Returns the parsed character.

alphaNum :: forall s (m :: Type -> Type) u. Stream s m Char => ParsecT s u m Char

Parses a alphabetic or numeric Unicode characters according to isAlphaNum. Returns the parsed character.

Note that numeric digits outside the ASCII range (such as arabic-indic digits like e.g. "٤" or U+0664), as well as numeric characters which aren't digits, are parsed by this function but not by digit.

letter :: forall s (m :: Type -> Type) u. Stream s m Char => ParsecT s u m Char

Parses an alphabetic Unicode characters (lower-case, upper-case and title-case letters, plus letters of caseless scripts and modifiers letters according to isAlpha). Returns the parsed character.

digit :: forall s (m :: Type -> Type) u. Stream s m Char => ParsecT s u m Char

Parses an ASCII digit. Returns the parsed character.

hexDigit :: forall s (m :: Type -> Type) u. Stream s m Char => ParsecT s u m Char

Parses a hexadecimal digit (a digit or a letter between 'a' and 'f' or 'A' and 'F'). Returns the parsed character.

octDigit :: forall s (m :: Type -> Type) u. Stream s m Char => ParsecT s u m Char

Parses an octal digit (a character between '0' and '7'). Returns the parsed character.

anyChar :: forall s (m :: Type -> Type) u. Stream s m Char => ParsecT s u m Char

This parser succeeds for any character. Returns the parsed character.

string' :: forall s (m :: Type -> Type) u. Stream s m Char => String -> ParsecT s u m String

string' s parses a sequence of characters given by s. Doesn't consume matching prefix.

 carOrCdr    =   string' "car"
             <|> string' "cdr"

Since: parsec-3.1.16.0

module Text.Parsec.Char

data Operator s u (m :: Type -> Type) a

This data type specifies operators that work on values of type a. An operator is either binary infix or unary prefix or postfix. A binary operator has also an associated associativity.

Constructors

Infix (ParsecT s u m (a -> a -> a)) Assoc

Prefix (ParsecT s u m (a -> a))

Postfix (ParsecT s u m (a -> a))

data Assoc

This data type specifies the associativity of operators: left, right or none.

Constructors

AssocNone

AssocLeft

AssocRight

buildExpressionParser :: forall s (m :: Type -> Type) t u a. Stream s m t => OperatorTable s u m a -> ParsecT s u m a -> ParsecT s u m a

buildExpressionParser table term builds an expression parser for terms term with operators from table, taking the associativity and precedence specified in table into account. Prefix and postfix operators of the same precedence can only occur once (i.e. --2 is not allowed if - is prefix negate). Prefix and postfix operators of the same precedence associate to the left (i.e. if ++ is postfix increment, than -2++ equals -1, not -3).

The buildExpressionParser takes care of all the complexity involved in building expression parser. Here is an example of an expression parser that handles prefix signs, postfix increment and basic arithmetic.

 expr    = buildExpressionParser table term
         <?> "expression"

 term    =  parens expr
         <|> natural
         <?> "simple expression"

 table   = [ [prefix "-" negate, prefix "+" id ]
           , [postfix "++" (+1)]
           , [binary "*" (*) AssocLeft, binary "/" (div) AssocLeft ]
           , [binary "+" (+) AssocLeft, binary "-" (-)   AssocLeft ]
           ]

 binary  name fun assoc = Infix (do{ reservedOp name; return fun }) assoc
 prefix  name fun       = Prefix (do{ reservedOp name; return fun })
 postfix name fun       = Postfix (do{ reservedOp name; return fun })

type Parser = Parsec Text ()

many1 :: forall s u (m :: Type -> Type) a. ParsecT s u m a -> ParsecT s u m [a]

many1 p applies the parser p one or more times. Returns a list of the returned values of p.

 word  = many1 letter

digitToInt :: Char -> Int

Convert a single digit Char to the corresponding Int. This function fails unless its argument satisfies isHexDigit, but recognises both upper- and lower-case hexadecimal digits (that is, '0'..'9', 'a'..'f', 'A'..'F').

Examples

Expand

Characters '0' through '9' are converted properly to 0..9:

>>> map digitToInt ['0'..'9']
[0,1,2,3,4,5,6,7,8,9]

Both upper- and lower-case 'A' through 'F' are converted as well, to 10..15.

>>> map digitToInt ['a'..'f']
[10,11,12,13,14,15]
>>> map digitToInt ['A'..'F']
[10,11,12,13,14,15]

Anything else throws an exception:

>>> digitToInt 'G'
*** Exception: Char.digitToInt: not a digit 'G'
>>> digitToInt '♥'
*** Exception: Char.digitToInt: not a digit '\9829'