Cardano.Api.Protocol

Documentation

data BlockType blk where

Constructors

ByronBlockType :: BlockType (HardForkBlock '[ByronBlock])

ShelleyBlockType :: BlockType (HardForkBlock '[ShelleyBlock (TPraos StandardCrypto) StandardShelley])

CardanoBlockType :: BlockType (HardForkBlock (CardanoEras StandardCrypto))

data SomeBlockType where

Constructors

SomeBlockType :: forall blk. BlockType blk -> SomeBlockType

reflBlockType :: BlockType blk -> BlockType blk' -> Maybe (blk :~: blk')

class (RunNode blk, IOLike m) => Protocol (m :: Type -> Type) blk where

Associated Types

data ProtocolInfoArgs blk

Methods

protocolInfo :: ProtocolInfoArgs blk -> (ProtocolInfo blk, m [BlockForging m blk])

data family ProtocolInfoArgs blk

class RunNode blk => ProtocolClient blk where

Node client support for each consensus protocol.

This is like Protocol but for clients of the node, so with less onerous requirements than to run a node.

Associated Types

data ProtocolClientInfoArgs blk

Methods

protocolClientInfo :: ProtocolClientInfoArgs blk -> ProtocolClientInfo blk

data family ProtocolClientInfoArgs blk