Cardano.Api.Protocol.Version

Description

This module defines the protocol versions corresponding to the eras in the Cardano blockchain.

Documentation

data BabbageEra

Users typically interact with the latest features on the mainnet or experiment with features from the upcoming era. Hence, the protocol versions are limited to the current mainnet era and the next era (upcoming era).

data ConwayEra

pattern CurrentEra :: Era BabbageEra

How to deprecate an era

Add DEPRECATED pragma to the era type tag. {-# DEPRECATED BabbageEra "BabbageEra no longer supported, use ConwayEra" #-} data BabbageEra

Add a new era type tag. data Era version where -- | The era currently active on Cardano's mainnet. CurrentEraInternal :: Era ConwayEra -- | The era planned for the next hardfork on Cardano's mainnet. UpcomingEraInternal :: Era (UninhabitableType EraCurrentlyNonExistent)

Update pattern synonyms. @ pattern CurrentEra :: Era ConwayEra pattern CurrentEra = CurrentEraInternal

pattern UpcomingEra :: Era (UninhabitableType EraCurrentlyNonExistent) pattern UpcomingEra = UpcomingEraInternal @

Add new UseEra instance and keep the deprecated era's instance. @ instance UseEra BabbageEra where useEra = error "useEra: BabbageEra no longer supported, use ConwayEra"

instance UseEra ConwayEra where useEra = CurrentEra @

Update protocolVersionToSbe as follows: protocolVersionToSbe :: Era version -> Maybe (ShelleyBasedEra (VersionToSbe version)) protocolVersionToSbe CurrentEraInternal = Just ShelleyBasedEraBabbage protocolVersionToSbe UpcomingEraInternal = Nothing

CurrentEraInternal and UpcomingEraInternal are for internal use only. The above restriction combined with the following pattern synonyms prevents a user from pattern matching on 'Era era' and avoids the following situation:

doThing :: Era era -> ()
doThing = case
  CurrentEraInternal -> enableFeature
  UpcomingEraInternal -> disableFeature

Consumers of this library must pick one of the two eras while this library is responsibile for what happens at the boundary of the eras.

pattern UpcomingEra :: Era ConwayEra

data Era version where

Represents the eras in Cardano's blockchain.

Instead of enumerating every possible era, we use two constructors: CurrentEra and UpcomingEra. This design simplifies the handling of eras, especially for 'cardano-api' consumers who are primarily concerned with the current mainnet era and the next era for an upcoming hardfork.

Usage: - CurrentEra: Reflects the era currently active on mainnet. - UpcomingEra: Represents the era planned for the next hardfork.

After a hardfork, 'cardano-api' should be updated promptly to reflect the new mainnet era in CurrentEra.

Constructors

CurrentEraInternal :: Era BabbageEra
The era currently active on Cardano's mainnet.

UpcomingEraInternal :: Era ConwayEra
The era planned for the next hardfork on Cardano's mainnet.

class UseEra version

Type class interface for the Era type.

Minimal complete definition

useEra

type family VersionToSbe version where ...

Equations

VersionToSbe BabbageEra = BabbageEra

VersionToSbe ConwayEra = ConwayEra

useEra :: UseEra version => Era version

protocolVersionToSbe :: Era version -> Maybe (ShelleyBasedEra (VersionToSbe version))