Cardano.Api.ReexposeLedger

Documentation

data Credential (kr :: KeyRole) c

Script hash or key hash for a payment or a staking object.

Note that credentials (unlike raw key hashes) do appear to vary from era to era, since they reference the hash of a script, which can change. This parameter is a phantom, however, so in actuality the instances will remain the same.

Constructors

ScriptHashObj !(ScriptHash c)

KeyHashObj !(KeyHash kr c)

credToText :: forall (kr :: KeyRole) c. Credential kr c -> Text

newtype KeyHash (discriminator :: KeyRole) c

Discriminated hash of public Key

Constructors

KeyHash (Hash (ADDRHASH c) (VerKeyDSIGN (DSIGN c)))

data KeyRole

The role of a key.

Note that a role is not _fixed_, nor is it unique. In particular, keys may variously be used as witnesses, and so in many case we will change the role of a key to the Witness role.

It is also perfectly allowable for a key to be used in many roles; there is nothing prohibiting somebody using the same underlying key as their payment and staking key, as well as the key for their stake pool. So these roles are more intended for two purposes:

To make explicit how we are using a key in the specifications

To provide a guide to downstream implementors, for whom the profusion of keys may be confusing.

Constructors

Genesis

GenesisDelegate

Payment

Staking

StakePool

BlockIssuer

Witness

DRepRole

HotCommitteeRole

ColdCommitteeRole

newtype VKey (kd :: KeyRole) c

Discriminated verification key

We wrap the basic VerKeyDSIGN in order to add the key role.

Constructors

VKey

Fields

unVKey :: VerKeyDSIGN (DSIGN c)

data ShelleyTxCert era

A heavyweight certificate.

Constructors

ShelleyTxCertDelegCert !(ShelleyDelegCert (EraCrypto era))

ShelleyTxCertPool !(PoolCert (EraCrypto era))

ShelleyTxCertGenesisDeleg !(GenesisDelegCert (EraCrypto era))

ShelleyTxCertMir !(MIRCert (EraCrypto era))

data ShelleyDelegCert c

Constructors

ShelleyRegCert !(StakeCredential c)
A stake credential registration certificate.

ShelleyUnRegCert !(StakeCredential c)
A stake credential deregistration certificate.

ShelleyDelegCert !(StakeCredential c) !(KeyHash 'StakePool c)
A stake delegation certificate.

Bundled Patterns

pattern Delegate :: Delegation c -> ShelleyDelegCert c

pattern DeRegKey :: StakeCredential c -> ShelleyDelegCert c

pattern RegKey :: StakeCredential c -> ShelleyDelegCert c

class EraTxCert era => ShelleyEraTxCert era where

Methods

mkRegTxCert :: StakeCredential (EraCrypto era) -> TxCert era

getRegTxCert :: TxCert era -> Maybe (StakeCredential (EraCrypto era))

mkUnRegTxCert :: StakeCredential (EraCrypto era) -> TxCert era

getUnRegTxCert :: TxCert era -> Maybe (StakeCredential (EraCrypto era))

mkDelegStakeTxCert :: StakeCredential (EraCrypto era) -> KeyHash 'StakePool (EraCrypto era) -> TxCert era

getDelegStakeTxCert :: TxCert era -> Maybe (StakeCredential (EraCrypto era), KeyHash 'StakePool (EraCrypto era))

mkGenesisDelegTxCert :: GenesisDelegCert (EraCrypto era) -> TxCert era

getGenesisDelegTxCert :: TxCert era -> Maybe (GenesisDelegCert (EraCrypto era))

mkMirTxCert :: MIRCert (EraCrypto era) -> TxCert era

getMirTxCert :: TxCert era -> Maybe (MIRCert (EraCrypto era))

data PState era

The state used by the POOL rule, which tracks stake pool information.

Constructors

PState

Fields

psStakePoolParams :: !(Map (KeyHash 'StakePool (EraCrypto era)) (PoolParams (EraCrypto era)))

The stake pool parameters.

psFutureStakePoolParams :: !(Map (KeyHash 'StakePool (EraCrypto era)) (PoolParams (EraCrypto era)))

The future stake pool parameters. Changes to existing stake pool parameters are staged in order to give delegators time to react to changes. See section 11.2, "Example Illustration of the Reward Cycle", of the Shelley Ledger Specification for a sequence diagram.

psRetiring :: !(Map (KeyHash 'StakePool (EraCrypto era)) EpochNo)

A map of retiring stake pools to the epoch when they retire.

psDeposits :: !(Map (KeyHash 'StakePool (EraCrypto era)) Coin)

A map of the deposits for each pool

data GenesisDelegCert c

Genesis key delegation certificate

Constructors

GenesisDelegCert !(KeyHash 'Genesis c) !(KeyHash 'GenesisDelegate c) !(Hash c (VerKeyVRF c))

data GenDelegPair c

Constructors

GenDelegPair

Fields

genDelegKeyHash :: !(KeyHash 'GenesisDelegate c)

genDelegVrfHash :: !(Hash c (VerKeyVRF c))

data StakeReference c

Constructors

StakeRefBase !(StakeCredential c)

StakeRefPtr !Ptr

StakeRefNull

data WitVKey (kr :: KeyRole) c where

Proof/Witness that a transaction is authorized by the given key holder.

Bundled Patterns

pattern WitVKey :: (Typeable kr, Crypto c) => VKey kr c -> SignedDSIGN c (Hash c EraIndependentTxBody) -> WitVKey kr c

hashKey :: forall c (kd :: KeyRole). Crypto c => VKey kd c -> KeyHash kd c

Hash a given public key

hashVerKeyVRF :: (VRFAlgorithm v, HashAlgorithm h) => VerKeyVRF v -> Hash h (VerKeyVRF v)

hashWithSerialiser :: HashAlgorithm h => (a -> Encoding) -> a -> Hash h a

A variation on hashWith, but specially for CBOR encodings.

data PoolParams c

A stake pool.

Constructors

PoolParams

Fields

ppId :: !(KeyHash 'StakePool c)

ppVrf :: !(Hash c (VerKeyVRF c))

ppPledge :: !Coin

ppCost :: !Coin

ppMargin :: !UnitInterval

ppRewardAcnt :: !(RewardAcnt c)

ppOwners :: !(Set (KeyHash 'Staking c))

ppRelays :: !(StrictSeq StakePoolRelay)

ppMetadata :: !(StrictMaybe PoolMetadata)

class HasKeyRole (a :: KeyRole -> Type -> Type)

data MIRPot

Constructors

ReservesMIR

TreasuryMIR

data MIRTarget c

MIRTarget specifies if funds from either the reserves or the treasury are to be handed out to a collection of reward accounts or instead transfered to the opposite pot.

Constructors

StakeAddressesMIR !(Map (Credential 'Staking c) DeltaCoin)

SendToOppositePotMIR !Coin

data MIRCert c

Move instantaneous rewards certificate

Constructors

MIRCert

Fields

mirPot :: !MIRPot

mirRewards :: !(MIRTarget c)

data StakePoolRelay

Constructors

SingleHostAddr !(StrictMaybe Port) !(StrictMaybe IPv4) !(StrictMaybe IPv6)
One or both of IPv4 & IPv6

SingleHostName !(StrictMaybe Port) !DnsName
An A or AAAA DNS record

MultiHostName !DnsName
A SRV DNS record

data PoolMetadata

Constructors

PoolMetadata

Fields

pmUrl :: !Url

pmHash :: !ByteString

class (Era era, ToJSON (TxCert era), DecCBOR (TxCert era), EncCBOR (TxCert era), ToCBOR (TxCert era), FromCBOR (TxCert era), NoThunks (TxCert era), NFData (TxCert era), Show (TxCert era), Eq (TxCert era)) => EraTxCert era where

Associated Types

type TxCert era = (r :: Type) | r -> era

type TxCertUpgradeError era

type TxCertUpgradeError era = Void

Methods

upgradeTxCert :: TxCert (PreviousEra era) -> Either (TxCertUpgradeError era) (TxCert era)

Every era, except Shelley, must be able to upgrade a TxCert from a previous era. However, not all certificates can be upgraded, because some eras lose some of the certificates, thus return type is an Either. Eg. from Babbage to Conway: MIR and Genesis certificates were removed.

getVKeyWitnessTxCert :: TxCert era -> Maybe (KeyHash 'Witness (EraCrypto era))

Return a witness key whenever a certificate requires one

getScriptWitnessTxCert :: TxCert era -> Maybe (ScriptHash (EraCrypto era))

Return a ScriptHash for certificate types that require a witness

mkRegPoolTxCert :: PoolParams (EraCrypto era) -> TxCert era

getRegPoolTxCert :: TxCert era -> Maybe (PoolParams (EraCrypto era))

mkRetirePoolTxCert :: KeyHash 'StakePool (EraCrypto era) -> EpochNo -> TxCert era

getRetirePoolTxCert :: TxCert era -> Maybe (KeyHash 'StakePool (EraCrypto era), EpochNo)

lookupRegStakeTxCert :: TxCert era -> Maybe (Credential 'Staking (EraCrypto era))

Extract staking credential from any certificate that can register such credential

lookupUnRegStakeTxCert :: TxCert era -> Maybe (Credential 'Staking (EraCrypto era))

Extract staking credential from any certificate that can unregister such credential

getTotalDepositsTxCerts

Arguments

:: Foldable f

=> PParams era

-> (KeyHash 'StakePool (EraCrypto era) -> Bool)
Check whether stake pool is registered or not

-> f (TxCert era)

-> Coin

Compute the total deposits from a list of certificates.

getTotalRefundsTxCerts

Arguments

:: Foldable f

=> PParams era

-> (Credential 'Staking (EraCrypto era) -> Maybe Coin)
Lookup current deposit for Staking credential if one is registered

-> (Credential 'DRepRole (EraCrypto era) -> Maybe Coin)
Lookup current deposit for DRep credential if one is registered

-> f (TxCert era)

-> Coin

Compute the total refunds from a list of certificates.

data StrictMaybe a

Constructors

SNothing

SJust !a

pattern DelegTxCert :: ConwayEraTxCert era => StakeCredential (EraCrypto era) -> Delegatee (EraCrypto era) -> TxCert era

pattern RegPoolTxCert :: EraTxCert era => PoolParams (EraCrypto era) -> TxCert era

pattern RetirePoolTxCert :: EraTxCert era => KeyHash 'StakePool (EraCrypto era) -> EpochNo -> TxCert era

pattern RegDepositTxCert :: ConwayEraTxCert era => StakeCredential (EraCrypto era) -> Coin -> TxCert era

pattern UnRegDepositTxCert :: ConwayEraTxCert era => StakeCredential (EraCrypto era) -> Coin -> TxCert era

pattern UnRegDRepTxCert :: ConwayEraTxCert era => Credential 'DRepRole (EraCrypto era) -> Coin -> TxCert era

pattern AuthCommitteeHotKeyTxCert :: ConwayEraTxCert era => Credential 'ColdCommitteeRole (EraCrypto era) -> Credential 'HotCommitteeRole (EraCrypto era) -> TxCert era

pattern ResignCommitteeColdTxCert :: ConwayEraTxCert era => Credential 'ColdCommitteeRole (EraCrypto era) -> StrictMaybe (Anchor (EraCrypto era)) -> TxCert era

pattern RegTxCert :: ShelleyEraTxCert era => StakeCredential (EraCrypto era) -> TxCert era

pattern UnRegTxCert :: ShelleyEraTxCert era => StakeCredential (EraCrypto era) -> TxCert era

pattern DelegStakeTxCert :: ShelleyEraTxCert era => StakeCredential (EraCrypto era) -> KeyHash 'StakePool (EraCrypto era) -> TxCert era

pattern RegDepositDelegTxCert :: ConwayEraTxCert era => StakeCredential (EraCrypto era) -> Delegatee (EraCrypto era) -> Coin -> TxCert era

pattern RegDRepTxCert :: ConwayEraTxCert era => Credential 'DRepRole (EraCrypto era) -> Coin -> StrictMaybe (Anchor (EraCrypto era)) -> TxCert era

pattern MirTxCert :: (ShelleyEraTxCert era, ProtVerAtMost era 8) => MIRCert (EraCrypto era) -> TxCert era

pattern GenesisDelegTxCert :: (ShelleyEraTxCert era, ProtVerAtMost era 8) => KeyHash 'Genesis (EraCrypto era) -> KeyHash 'GenesisDelegate (EraCrypto era) -> Hash (EraCrypto era) (VerKeyVRF (EraCrypto era)) -> TxCert era

pattern UpdateDRepTxCert :: ConwayEraTxCert era => Credential 'DRepRole (EraCrypto era) -> StrictMaybe (Anchor (EraCrypto era)) -> TxCert era

newtype Coin

The amount of value held by a transaction output.

Constructors

Coin

Fields

unCoin :: Integer

class (Era era, Eq (PParamsHKD Identity era), Ord (PParamsHKD Identity era), Show (PParamsHKD Identity era), NFData (PParamsHKD Identity era), EncCBOR (PParamsHKD Identity era), DecCBOR (PParamsHKD Identity era), ToCBOR (PParamsHKD Identity era), FromCBOR (PParamsHKD Identity era), NoThunks (PParamsHKD Identity era), ToJSON (PParamsHKD Identity era), FromJSON (PParamsHKD Identity era), Eq (PParamsHKD StrictMaybe era), Ord (PParamsHKD StrictMaybe era), Show (PParamsHKD StrictMaybe era), NFData (PParamsHKD StrictMaybe era), EncCBOR (PParamsHKD StrictMaybe era), DecCBOR (PParamsHKD StrictMaybe era), ToCBOR (PParamsHKD StrictMaybe era), FromCBOR (PParamsHKD StrictMaybe era), NoThunks (PParamsHKD StrictMaybe era), ToJSON (PParamsHKD StrictMaybe era)) => EraPParams era where

Minimal complete definition

emptyPParamsIdentity, emptyPParamsStrictMaybe, upgradePParamsHKD, downgradePParamsHKD, hkdMinFeeAL, hkdMinFeeBL, hkdMaxBBSizeL, hkdMaxTxSizeL, hkdMaxBHSizeL, hkdKeyDepositL, hkdPoolDepositL, hkdEMaxL, hkdNOptL, hkdA0L, hkdRhoL, hkdTauL, hkdDL, hkdExtraEntropyL, hkdProtocolVersionL, hkdMinUTxOValueL, hkdMinPoolCostL

Associated Types

type PParamsHKD (f :: Type -> Type) era = (r :: Type) | r -> era

Protocol parameters where the fields are represented with a HKD

type UpgradePParams (f :: Type -> Type) era

type DowngradePParams (f :: Type -> Type) era

Methods

applyPPUpdates :: PParams era -> PParamsUpdate era -> PParams era

Applies a protocol parameters update

emptyPParamsIdentity :: PParamsHKD Identity era

emptyPParamsStrictMaybe :: PParamsHKD StrictMaybe era

upgradePParamsHKD :: forall (f :: Type -> Type). (HKDFunctor f, EraPParams (PreviousEra era)) => UpgradePParams f era -> PParamsHKD f (PreviousEra era) -> PParamsHKD f era

Upgrade PParams from previous era to the current one

downgradePParamsHKD :: forall (f :: Type -> Type). (HKDFunctor f, EraPParams (PreviousEra era)) => DowngradePParams f era -> PParamsHKD f era -> PParamsHKD f (PreviousEra era)

Downgrade PParams from the current era to the previous one

hkdMinFeeAL :: forall (f :: Type -> Type). HKDFunctor f => Lens' (PParamsHKD f era) (HKD f Coin)

The linear factor for the minimum fee calculation

hkdMinFeeBL :: forall (f :: Type -> Type). HKDFunctor f => Lens' (PParamsHKD f era) (HKD f Coin)

The constant factor for the minimum fee calculation

hkdMaxBBSizeL :: forall (f :: Type -> Type). HKDFunctor f => Lens' (PParamsHKD f era) (HKD f Word32)

Maximal block body size

hkdMaxTxSizeL :: forall (f :: Type -> Type). HKDFunctor f => Lens' (PParamsHKD f era) (HKD f Word32)

Maximal transaction size

hkdMaxBHSizeL :: forall (f :: Type -> Type). HKDFunctor f => Lens' (PParamsHKD f era) (HKD f Word16)

Maximal block header size

hkdKeyDepositL :: forall (f :: Type -> Type). HKDFunctor f => Lens' (PParamsHKD f era) (HKD f Coin)

The amount of a key registration deposit

hkdPoolDepositL :: forall (f :: Type -> Type). HKDFunctor f => Lens' (PParamsHKD f era) (HKD f Coin)

The amount of a pool registration deposit

hkdEMaxL :: forall (f :: Type -> Type). HKDFunctor f => Lens' (PParamsHKD f era) (HKD f EpochInterval)

epoch bound on pool retirement

hkdNOptL :: forall (f :: Type -> Type). HKDFunctor f => Lens' (PParamsHKD f era) (HKD f Natural)

Desired number of pools

hkdA0L :: forall (f :: Type -> Type). HKDFunctor f => Lens' (PParamsHKD f era) (HKD f NonNegativeInterval)

Pool influence

hkdRhoL :: forall (f :: Type -> Type). HKDFunctor f => Lens' (PParamsHKD f era) (HKD f UnitInterval)

Monetary expansion

hkdTauL :: forall (f :: Type -> Type). HKDFunctor f => Lens' (PParamsHKD f era) (HKD f UnitInterval)

Treasury expansion

hkdDL :: forall (f :: Type -> Type). (HKDFunctor f, ProtVerAtMost era 6) => Lens' (PParamsHKD f era) (HKD f UnitInterval)

Decentralization parameter

ppDG :: SimpleGetter (PParams era) UnitInterval

Decentralization parameter getter

hkdExtraEntropyL :: forall (f :: Type -> Type). (HKDFunctor f, ProtVerAtMost era 6) => Lens' (PParamsHKD f era) (HKD f Nonce)

Extra entropy

hkdProtocolVersionL :: forall (f :: Type -> Type). (HKDFunctor f, ProtVerAtMost era 8) => Lens' (PParamsHKD f era) (HKD f ProtVer)

Protocol version

ppProtocolVersionL :: Lens' (PParams era) ProtVer

ppuProtocolVersionL :: Lens' (PParamsUpdate era) (StrictMaybe ProtVer)

PParamsUpdate Protocol version

hkdMinUTxOValueL :: forall (f :: Type -> Type). (HKDFunctor f, ProtVerAtMost era 4) => Lens' (PParamsHKD f era) (HKD f Coin)

Minimum UTxO value

hkdMinPoolCostL :: forall (f :: Type -> Type). HKDFunctor f => Lens' (PParamsHKD f era) (HKD f Coin)

Minimum Stake Pool Cost

class (Crypto (EraCrypto era), Typeable era, KnownNat (ProtVerLow era), KnownNat (ProtVerHigh era), ProtVerLow era <= ProtVerHigh era, MinVersion <= ProtVerLow era, MinVersion <= ProtVerHigh era, CmpNat (ProtVerLow era) MaxVersion ~ 'LT, CmpNat (ProtVerHigh era) MaxVersion ~ 'LT, ProtVerLow era <= MaxVersion, ProtVerHigh era <= MaxVersion) => Era era where

Associated Types

type EraCrypto era

type PreviousEra era = (r :: Type) | r -> era

Map an era to its predecessor.

For example:

type instance PreviousEra (AllegraEra c) = ShelleyEra c

type ProtVerLow era :: Nat

Lowest major protocol version for this era

type ProtVerHigh era :: Nat

Highest major protocol version for this era. By default se to ProtVerLow

type ProtVerHigh era = ProtVerLow era

Methods

eraName :: String

Textual name of the current era.

Designed to be used with TypeApplications:

>>> eraName @(ByronEra StandardCrypto)
Byron

data Network

Constructors

Testnet

Mainnet

data PoolCert c

Constructors

RegPool !(PoolParams c)
A stake pool registration certificate.

RetirePool !(KeyHash 'StakePool c) !EpochNo
A stake pool retirement certificate.

newtype PParams era

Protocol parameters

Constructors

PParams (PParamsHKD Identity era)

data PParamsUpdate era

The type of updates to Protocol parameters

type family Value era

A value is something which quantifies a transaction output.

addDeltaCoin :: Coin -> DeltaCoin -> Coin

toDeltaCoin :: Coin -> DeltaCoin

toEraCBOR :: (Era era, EncCBOR t) => t -> Encoding

Convert a type that implements EncCBOR to plain Encoding using the lowest protocol version for the supplied era

fromEraCBOR :: (Era era, DecCBOR t) => Decoder s t

Convert a type that implements DecCBOR to plain Decoder using the lowest protocol version for the supplied era

ppMinUTxOValueL :: (EraPParams era, ProtVerAtMost era 4) => Lens' (PParams era) Coin

Minimum UTxO value

data Anchor c

Constructors

Anchor

Fields

anchorUrl :: !Url

anchorDataHash :: !(SafeHash c AnchorData)

data Delegatee c

First type argument is the deposit

Constructors

DelegStake !(KeyHash 'StakePool c)

DelegVote !(DRep c)

DelegStakeVote !(KeyHash 'StakePool c) !(DRep c)

data DRep c

Constructors

DRepAlwaysAbstain

DRepAlwaysNoConfidence

Bundled Patterns

pattern DRepCredential :: Credential 'DRepRole c -> DRep c

data DRepState c

Constructors

DRepState

Fields

drepExpiry :: !EpochNo

drepAnchor :: !(StrictMaybe (Anchor c))

drepDeposit :: !Coin

data ConwayTxCert era

Constructors

ConwayTxCertDeleg !(ConwayDelegCert (EraCrypto era))

ConwayTxCertPool !(PoolCert (EraCrypto era))

ConwayTxCertGov !(ConwayGovCert (EraCrypto era))

data ConwayDelegCert c

Certificates for registration and delegation of stake to Pools and DReps. Comparing to previous eras, there is now ability to:

Register and delegate with a single certificate: ConwayRegDelegCert

Ability to delegate to DReps with DelegVote and DelegStakeVote

Ability to specify the deposit amount. Deposits during registration and unregistration in Conway are optional, which will change in the future era. They are optional only for the smooth transition from Babbage to Conway. Validity of deposits is checked by the CERT rule.

Constructors

ConwayRegCert !(StakeCredential c) !(StrictMaybe Coin)
Register staking credential. Deposit, when present, must match the expected deposit amount specified by ppKeyDepositL in the protocol parameters.

ConwayUnRegCert !(StakeCredential c) !(StrictMaybe Coin)
De-Register the staking credential. Deposit, if present, must match the amount that was left as a deposit upon stake credential registration.

ConwayDelegCert !(StakeCredential c) !(Delegatee c)
Redelegate to another delegatee. Staking credential must already be registered.

ConwayRegDelegCert !(StakeCredential c) !(Delegatee c) !Coin
This is a new type of certificate, which allows to register staking credential and delegate within a single certificate. Deposit is required and must match the expected deposit amount specified by ppKeyDepositL in the protocol parameters.

class ShelleyEraTxCert era => ConwayEraTxCert era where

Methods

mkRegDepositTxCert :: StakeCredential (EraCrypto era) -> Coin -> TxCert era

getRegDepositTxCert :: TxCert era -> Maybe (StakeCredential (EraCrypto era), Coin)

mkUnRegDepositTxCert :: StakeCredential (EraCrypto era) -> Coin -> TxCert era

getUnRegDepositTxCert :: TxCert era -> Maybe (StakeCredential (EraCrypto era), Coin)

mkDelegTxCert :: StakeCredential (EraCrypto era) -> Delegatee (EraCrypto era) -> TxCert era

getDelegTxCert :: TxCert era -> Maybe (StakeCredential (EraCrypto era), Delegatee (EraCrypto era))

mkRegDepositDelegTxCert :: StakeCredential (EraCrypto era) -> Delegatee (EraCrypto era) -> Coin -> TxCert era

getRegDepositDelegTxCert :: TxCert era -> Maybe (StakeCredential (EraCrypto era), Delegatee (EraCrypto era), Coin)

mkAuthCommitteeHotKeyTxCert :: Credential 'ColdCommitteeRole (EraCrypto era) -> Credential 'HotCommitteeRole (EraCrypto era) -> TxCert era

getAuthCommitteeHotKeyTxCert :: TxCert era -> Maybe (Credential 'ColdCommitteeRole (EraCrypto era), Credential 'HotCommitteeRole (EraCrypto era))

mkResignCommitteeColdTxCert :: Credential 'ColdCommitteeRole (EraCrypto era) -> StrictMaybe (Anchor (EraCrypto era)) -> TxCert era

getResignCommitteeColdTxCert :: TxCert era -> Maybe (Credential 'ColdCommitteeRole (EraCrypto era), StrictMaybe (Anchor (EraCrypto era)))

mkRegDRepTxCert :: Credential 'DRepRole (EraCrypto era) -> Coin -> StrictMaybe (Anchor (EraCrypto era)) -> TxCert era

getRegDRepTxCert :: TxCert era -> Maybe (Credential 'DRepRole (EraCrypto era), Coin, StrictMaybe (Anchor (EraCrypto era)))

mkUnRegDRepTxCert :: Credential 'DRepRole (EraCrypto era) -> Coin -> TxCert era

getUnRegDRepTxCert :: TxCert era -> Maybe (Credential 'DRepRole (EraCrypto era), Coin)

mkUpdateDRepTxCert :: Credential 'DRepRole (EraCrypto era) -> StrictMaybe (Anchor (EraCrypto era)) -> TxCert era

getUpdateDRepTxCert :: TxCert era -> Maybe (Credential 'DRepRole (EraCrypto era), StrictMaybe (Anchor (EraCrypto era)))

data ConwayGovCert c

Constructors

ConwayRegDRep !(Credential 'DRepRole c) !Coin !(StrictMaybe (Anchor c))

ConwayUnRegDRep !(Credential 'DRepRole c) !Coin

ConwayUpdateDRep !(Credential 'DRepRole c) !(StrictMaybe (Anchor c))

ConwayAuthCommitteeHotKey !(Credential 'ColdCommitteeRole c) !(Credential 'HotCommitteeRole c)

ConwayResignCommitteeColdKey !(Credential 'ColdCommitteeRole c) !(StrictMaybe (Anchor c))

data ConwayGenesis c

Constructors

ConwayGenesis

Fields

cgUpgradePParams :: !(UpgradeConwayPParams Identity)

cgConstitution :: !(Constitution (ConwayEra c))

cgCommittee :: !(Committee (ConwayEra c))

cgDelegs :: ListMap (Credential 'Staking c) (Delegatee c)

cgInitialDReps :: ListMap (Credential 'DRepRole c) (DRepState c)

data UpgradeConwayPParams (f :: Type -> Type)

Constructors

UpgradeConwayPParams

Fields

ucppPoolVotingThresholds :: !(HKD f PoolVotingThresholds)

ucppDRepVotingThresholds :: !(HKD f DRepVotingThresholds)

ucppCommitteeMinSize :: !(HKD f Natural)

ucppCommitteeMaxTermLength :: !(HKD f EpochInterval)

ucppGovActionLifetime :: !(HKD f EpochInterval)

ucppGovActionDeposit :: !(HKD f Coin)

ucppDRepDeposit :: !(HKD f Coin)

ucppDRepActivity :: !(HKD f EpochInterval)

type family GovState era = (r :: Type) | r -> era

data GovActionId c

Constructors

GovActionId

Fields

gaidTxId :: !(TxId c)

gaidGovActionIx :: !GovActionIx

data Vote

Constructors

VoteNo

VoteYes

Abstain

data Voter c

Constructors

CommitteeVoter !(Credential 'HotCommitteeRole c)

DRepVoter !(Credential 'DRepRole c)

StakePoolVoter !(KeyHash 'StakePool c)

data VotingProcedure era

Constructors

VotingProcedure

Fields

vProcVote :: !Vote

vProcAnchor :: !(StrictMaybe (Anchor (EraCrypto era)))

data ProposalProcedure era

Constructors

ProposalProcedure

Fields

pProcDeposit :: !Coin

pProcReturnAddr :: !(RewardAcnt (EraCrypto era))

pProcGovAction :: !(GovAction era)

pProcAnchor :: !(Anchor (EraCrypto era))

newtype VotingProcedures era

Constructors

VotingProcedures

Fields

unVotingProcedures :: Map (Voter (EraCrypto era)) (Map (GovActionId (EraCrypto era)) (VotingProcedure era))

data PoolVotingThresholds

Constructors

PoolVotingThresholds

Fields

pvtMotionNoConfidence :: !UnitInterval

pvtCommitteeNormal :: !UnitInterval

pvtCommitteeNoConfidence :: !UnitInterval

pvtHardForkInitiation :: !UnitInterval

pvtPPSecurityGroup :: !UnitInterval

data DRepVotingThresholds

Constructors

DRepVotingThresholds

Fields

dvtMotionNoConfidence :: !UnitInterval

dvtCommitteeNormal :: !UnitInterval

dvtCommitteeNoConfidence :: !UnitInterval

dvtUpdateToConstitution :: !UnitInterval

dvtHardForkInitiation :: !UnitInterval

dvtPPNetworkGroup :: !UnitInterval

dvtPPEconomicGroup :: !UnitInterval

dvtPPTechnicalGroup :: !UnitInterval

dvtPPGovGroup :: !UnitInterval

dvtTreasuryWithdrawal :: !UnitInterval

dvtPPNetworkGroupL :: Lens' DRepVotingThresholds UnitInterval

dvtPPGovGroupL :: Lens' DRepVotingThresholds UnitInterval

dvtPPTechnicalGroupL :: Lens' DRepVotingThresholds UnitInterval

dvtPPEconomicGroupL :: Lens' DRepVotingThresholds UnitInterval

dvtUpdateToConstitutionL :: Lens' DRepVotingThresholds UnitInterval

drepExpiryL :: forall c f. Functor f => (EpochNo -> f EpochNo) -> DRepState c -> f (DRepState c)

drepAnchorL :: forall c f. Functor f => (StrictMaybe (Anchor c) -> f (StrictMaybe (Anchor c))) -> DRepState c -> f (DRepState c)

drepDepositL :: forall c f. Functor f => (Coin -> f Coin) -> DRepState c -> f (DRepState c)

csCommitteeCredsL :: forall era f. Functor f => (Map (Credential 'ColdCommitteeRole (EraCrypto era)) (Maybe (Credential 'HotCommitteeRole (EraCrypto era))) -> f (Map (Credential 'ColdCommitteeRole (EraCrypto era)) (Maybe (Credential 'HotCommitteeRole (EraCrypto era))))) -> CommitteeState era -> f (CommitteeState era)

data Annotated b a

Constructors

Annotated

Fields

unAnnotated :: !b

annotation :: !a

data Tx

Transaction

NB: transaction witnesses are stored separately

Constructors

UnsafeTx

Fields

txInputs :: !(NonEmpty TxIn)

Inputs of transaction.

txOutputs :: !(NonEmpty TxOut)

Outputs of transaction.

txAttributes :: !TxAttributes

Attributes of transaction

byronProtVer :: Version

serialize' :: EncCBOR a => Version -> a -> ByteString

Serialize a Haskell value to an external binary representation.

The output is represented as a strict ByteString.

toPlainDecoder :: Version -> Decoder s a -> Decoder s a

Extract the underlying Decoder by specifying the concrete version to be used.

toCBOR :: ToCBOR a => a -> Encoding

fromCBOR :: FromCBOR a => Decoder s a

data ByteSpan

A pair of offsets delimiting the beginning and end of a substring of a ByteString

Constructors

ByteSpan !ByteOffset !ByteOffset

slice :: ByteString -> ByteSpan -> ByteString

Extract a substring of a given ByteString corresponding to the offsets.

data Decoder s a

secondsToNominalDiffTimeMicro :: Micro -> NominalDiffTimeMicro

newtype CoinPerByte

Constructors

CoinPerByte

Fields

unCoinPerByte :: Coin

newtype CoinPerWord

Constructors

CoinPerWord

Fields

unCoinPerWord :: Coin

data Prices

Prices per execution unit

Constructors

Prices

Fields

prMem :: !NonNegativeInterval

prSteps :: !NonNegativeInterval

data CostModels

For a known version of Plutus, attempting to construct a cost model with too few parameters (depending on the version) will result in an error. CostModelApplyError exists to collect these errors in the CostModels type. The CostModels type itself needs to be flexible enough to accept any map of Word8 to '[Integer]', so that cost models can be placed in the protocol parameters ahead of changes to the Plutus evaluation context. In this way, serializing a cost model, updating software, and deserializing can result in errors going away.

Additionally, CostModels needs to be able to store cost models for future version of Plutus, which we cannot yet even validate. These are stored in costModelsUnknown.

data AlonzoGenesis

All configuration that is necessary to bootstrap AlonzoEra from ShelleyGenesis

ppPricesL :: AlonzoEraPParams era => Lens' (PParams era) Prices

boundRational :: BoundedRational r => Rational -> Maybe r

Returns Nothing when supplied value is not within bounds or when precision is too high to be represented by the underlying type

Example

Expand

>>> :set -XTypeApplications
>>> import Data.Ratio
>>> boundRational @UnitInterval $ 2 % 3
Just (2 % 3)
>>> boundRational @UnitInterval (-0.5)
Nothing
>>> boundRational @UnitInterval (1.5)
Nothing
>>> boundRational @UnitInterval 0
Just (0 % 1)
>>> boundRational @PositiveUnitInterval 0
Nothing

unboundRational :: BoundedRational r => r -> Rational

Promote bounded rational type into the unbounded Rational.

data DnsName

dnsToText :: DnsName -> Text

newtype EpochInterval

Denotes a positive change in the EpochNo

Constructors

EpochInterval

Fields

unEpochInterval :: Word32

textToDns :: MonadFail m => Int -> Text -> m DnsName

Turn a Text into a DnsName, fail if the Text has more than n Bytes

data Url

urlToText :: Url -> Text

textToUrl :: MonadFail m => Int -> Text -> m Url

Turn a Text into a Url, fail if the Text has more than n Bytes

portToWord16 :: Port -> Word16

data ProtVer

Constructors

ProtVer

Fields

pvMajor :: !Version

pvMinor :: !Natural

strictMaybeToMaybe :: StrictMaybe a -> Maybe a

maybeToStrictMaybe :: Maybe a -> StrictMaybe a

newtype AnchorData

Constructors

AnchorData ByteString

hashAnchorData :: Crypto c => AnchorData -> SafeHash c AnchorData

Hash AnchorData

data UnitInterval

Type to represent a value in the unit interval [0; 1]

mkVersion :: (Integral i, MonadFail m) => i -> m Version

Construct a Version and fail if the supplied value is not a supported version number.

data NonNegativeInterval

Type to represent a value in the interval [0; +∞)

hashToBytes :: Hash h a -> ByteString

The representation of the hash as bytes.

hashFromBytes

Arguments

:: HashAlgorithm h

=> ByteString
It must have an exact length, as given by sizeHash.

-> Maybe (Hash h a)

Make a hash from it bytes representation.

class (HashAlgorithm (HASH c), HashAlgorithm (ADDRHASH c), DSIGNAlgorithm (DSIGN c), KESAlgorithm (KES c), VRFAlgorithm (VRF c), ContextDSIGN (DSIGN c) ~ (), ContextKES (KES c) ~ (), ContextVRF (VRF c) ~ (), Typeable c) => Crypto c

data StandardCrypto

The same crypto used on the net

type family ADDRHASH c

newtype EpochNo

An epoch, i.e. the number of the epoch.

Constructors

EpochNo

Fields

unEpochNo :: Word64

data SafeHash c index

A SafeHash is a hash of something that is safe to hash. Such types store their own serialisation bytes. The prime example is (MemoBytes t), but other examples are things that consist of only ByteStrings (i.e. they are their own serialization) or for some other reason store their original bytes.

We do NOT export the constructor SafeHash, but instead export other functions such as 'hashWithCrypto, hashAnnotated and extractHash which have constraints that limit their application to types which preserve their original serialization bytes.

unsafeMakeSafeHash :: Hash (HASH c) index -> SafeHash c index

Don't use this except in Testing to make Arbitrary instances, etc. Defined here, only because the Constructor is in scope here.

extractHash :: SafeHash c i -> Hash (HASH c) i

Extract the hash out of a SafeHash