Cardano.Api.Serialise.Bech32

Description

Bech32 Serialisation

The types and functions of this module now live in the cardano-keys package; re-exported here for compatibility, together with the Error instance this API adds on top of them.

Documentation

module Cardano.Keys.Serialise.Bech32

Orphan instances

Error Bech32DecodeError

Instance details

Methods

prettyError :: Bech32DecodeError -> Doc ann