Cardano.Api.Serialise.Cip129

Documentation

class (SerialiseAsRawBytes a, HasTypeProxy a) => Cip129 a where

Cip-129 is a typeclass that captures the serialisation requirements of https://cips.cardano.org/cip/CIP-0129 which pertain to governance credentials and governance action ids.

Minimal complete definition

cip129Bech32PrefixFor, cip129Header, cip129SerialiseRaw, cip129DeserialiseRaw

Associated Types

data Cip129Header a

A sum type with all possible headers for CIP-129 identifier

Methods

cip129Bech32PrefixFor :: AsType a -> HumanReadablePart

The human readable part of the Bech32 encoding for the credential.

cip129Bech32PrefixesPermitted :: AsType a -> [Text]

Permitted bech32 prefixes according to CIP-129.

default cip129Bech32PrefixesPermitted :: AsType a -> [Text]

cip129Header :: Cip129Header a -> Word8

A Word8 value of Cip129 header

cip129SerialiseRaw :: a -> ByteString

Serialise a value to a binary representation used in CIP 129. It's usually distinct from CBOR serialisation. Internal conversion function. Use serialiseToBech32Cip129 instead of calling this function directly.

cip129DeserialiseRaw :: ByteString -> Either Cip129EncodingError a

Deserialise a value from the bytes representation. Internal conversion function. Use deserialiseFromBech32Cip129 instead of calling this function directly.

data Cip129EncodingError

CIP-129 decoding errors

deserialiseFromBech32Cip129

Arguments

:: Cip129 a

=> Text
A Bech32-encoded governance identifier

-> Either Cip129EncodingError a

Deserialise a governance identifier from CIP-129 format.

serialiseToBech32Cip129 :: Cip129 a => a -> Text

Serialise a accoding to the serialisation requirements of https://cips.cardano.org/cip/CIP-0129 which currently pertain to governance credentials.

serialiseGovActionIdToBech32Cip129 :: GovActionId -> Text

Deprecated: Use serialiseToBech32Cip129 instead

Governance Action ID According to Cip129 there is no header byte for GovActionId. Instead they append the txid and index to form the payload.

deserialiseGovActionIdFromBech32Cip129 :: Text -> Either Cip129EncodingError GovActionId

Deprecated: Use deserialiseFromBech32Cip129 instead

data family AsType t

A family of singleton types used in this API to indicate which type to use where it would otherwise be ambiguous or merely unclear.

Values of this type are passed to deserialisation functions for example.