Cardano.Api.Serialise.DeserialiseAnyOf

Description

Class of errors used in the Api.

Documentation

data InputFormat a where

Input format/encoding.

Constructors

InputFormatBech32 :: forall a. SerialiseAsBech32 a => InputFormat a
Bech32 encoding.

InputFormatHex :: forall a. SerialiseAsRawBytes a => InputFormat a
Hex/Base16 encoding.

InputFormatTextEnvelope :: forall a. HasTextEnvelope a => InputFormat a
Text envelope format.

data InputDecodeError

Input decoding error.

Constructors

InputTextEnvelopeError !TextEnvelopeError
The provided data seems to be a valid text envelope, but some error occurred in deserialising it.

InputBech32DecodeError !Bech32DecodeError
The provided data is valid Bech32, but some error occurred in deserialising it.

InputInvalidError
The provided data does not represent a valid value of the provided type.

deserialiseInput :: NonEmpty (InputFormat a) -> ByteString -> Either InputDecodeError a

Deserialise an input of some type that is formatted in some way.

deserialiseInputAnyOf :: [FromSomeType SerialiseAsBech32 b] -> [FromSomeType HasTextEnvelope b] -> ByteString -> Either InputDecodeError b

Deserialise an input of some type that is formatted in some way.

The provided ByteString can either be Bech32-encoded or in the text envelope format.

readFormattedFile

Arguments

:: NonEmpty (InputFormat a)
one of expected input formats

-> FilePath

-> IO (Either (FileError InputDecodeError) a)

Read formatted file

readFormattedFileTextEnvelope :: HasTextEnvelope a => File content 'In -> IO (Either (FileError InputDecodeError) a)

Read text envelope file

readFormattedFileAnyOf :: forall content b. [FromSomeType SerialiseAsBech32 b] -> [FromSomeType HasTextEnvelope b] -> File content 'In -> IO (Either (FileError InputDecodeError) b)

Read in in any of the format in the text envelope

renderInputDecodeError :: InputDecodeError -> Doc ann

Render an error message for a InputDecodeError.