Cardano.Api.Serialise.Json

Description

JSON serialisation

Documentation

serialiseToJSON :: ToJSON a => a -> ByteString

class ToJSON a where

Minimal complete definition

Nothing

Methods

toJSON :: a -> Value

toEncoding :: a -> Encoding

toJSONList :: [a] -> Value

toEncodingList :: [a] -> Encoding

omitField :: a -> Bool

class ToJSONKey a

deserialiseFromJSON :: FromJSON a => ByteString -> Either JsonDecodeError a

prettyPrintJSON :: ToJSON a => a -> ByteString

class FromJSON a where

Minimal complete definition

Nothing

Methods

parseJSON :: Value -> Parser a

parseJSONList :: Value -> Parser [a]

omittedField :: Maybe a

class FromJSONKey a

newtype JsonDecodeError

Constructors

JsonDecodeError String

readFileJSON :: FromJSON a => FilePath -> IO (Either (FileError JsonDecodeError) a)

writeFileJSON :: ToJSON a => FilePath -> a -> IO (Either (FileError ()) ())

textWithMaxLength :: String -> Int -> Value -> Parser Text

Parser for Text that validates that the number of characters is under a given maximum. The String parameter is meant to be the name of the field in order to be able to give context in case of error.

toRationalJSON :: Rational -> Value