Cardano.Api.Serialise.SerialiseUsing

Description

Raw binary serialisation

Documentation

newtype UsingRawBytes a

For use with deriving via, to provide ToCBOR and FromCBOR instances, based on the SerialiseAsRawBytes instance.

deriving (ToCBOR, FromCBOR) via (UsingRawBytes Blah)

Constructors

UsingRawBytes a

newtype UsingRawBytesHex a

For use with deriving via, to provide instances for any/all of Show, ToJSON, FromJSON, ToJSONKey, FromJSONKey' using a hex encoding, based on the SerialiseAsRawBytes instance.

deriving (Show, Pretty) via (UsingRawBytesHex Blah)
deriving (ToJSON, FromJSON) via (UsingRawBytesHex Blah)
deriving (ToJSONKey, FromJSONKey) via (UsingRawBytesHex Blah)

Constructors

UsingRawBytesHex a

newtype UsingBech32 a

For use with deriving via, to provide instances for any/all of Show, IsString, ToJSON, FromJSON, ToJSONKey, FromJSONKey' using a bech32 encoding, based on the SerialiseAsBech32 instance.

deriving (Show, Pretty) via (UsingBech32 Blah)
deriving (ToJSON, FromJSON) via (UsingBech32 Blah)
deriving (ToJSONKey, FromJSONKey) via (UsingBech32 Blah)

Constructors

UsingBech32 a