Cardano.Api.Serialise.TextEnvelope

TextEnvelope Serialisation

class SerialiseAsCBOR a => HasTextEnvelope a where

Minimal complete definition

textEnvelopeType

Methods

textEnvelopeType :: AsType a -> TextEnvelopeType

textEnvelopeDefaultDescr :: a -> TextEnvelopeDescr

textEnvelopeTypeInEra :: HasTextEnvelope (f era) => CardanoEra era -> AsType (f era) -> TextEnvelopeType

data TextEnvelope

A TextEnvelope is a structured envelope for serialised binary values with an external format with a semi-readable textual format.

It contains a "type" field, e.g. "PublicKeyByron" or "TxSignedShelley" to indicate the type of the encoded data. This is used as a sanity check and to help readers.

It also contains a "title" field which is free-form, and could be used to indicate the role or purpose to a reader.

Constructors

TextEnvelope

Fields

teType :: !TextEnvelopeType

teDescription :: !TextEnvelopeDescr

teRawCBOR :: !ByteString

newtype TextEnvelopeType

Constructors

TextEnvelopeType String

newtype TextEnvelopeDescr

Constructors

TextEnvelopeDescr String

textEnvelopeRawCBOR :: TextEnvelope -> ByteString

data TextEnvelopeError

The errors that the pure TextEnvelope parsing/decoding functions can return.

Constructors

TextEnvelopeTypeError ![TextEnvelopeType] !TextEnvelopeType
expected, actual

TextEnvelopeDecodeError !DecoderError

TextEnvelopeAesonDecodeError !String

TextEnvelopeUnknownKeyWitness !TextEnvelopeDescr

TextEnvelopeUnknownType !Text

serialiseToTextEnvelope :: HasTextEnvelope a => Maybe TextEnvelopeDescr -> a -> TextEnvelope

deserialiseFromTextEnvelope :: HasTextEnvelope a => TextEnvelope -> Either TextEnvelopeError a

readFileTextEnvelope :: HasTextEnvelope a => File content 'In -> IO (Either (FileError TextEnvelopeError) a)

writeFileTextEnvelope :: HasTextEnvelope a => File content 'Out -> Maybe TextEnvelopeDescr -> a -> IO (Either (FileError ()) ())

readTextEnvelopeFromFile :: FilePath -> IO (Either (FileError TextEnvelopeError) TextEnvelope)

readTextEnvelopeOfTypeFromFile :: TextEnvelopeType -> FilePath -> IO (Either (FileError TextEnvelopeError) TextEnvelope)

textEnvelopeToJSON :: HasTextEnvelope a => Maybe TextEnvelopeDescr -> a -> ByteString

serialiseTextEnvelope :: TextEnvelope -> ByteString

Serialise text envelope to pretty JSON

legacyComparison :: TextEnvelopeType -> TextEnvelopeType -> Bool

This is a backwards-compatibility patch to ensure that old envelopes generated by serialiseTxLedgerCddl can be deserialised after switching to the serialiseToTextEnvelope.

Reading one of several key types

data FromSomeType (c :: Type -> Constraint) b where

Constructors

FromSomeType :: forall (c :: Type -> Constraint) a b. c a => AsType a -> (a -> b) -> FromSomeType c b

deserialiseFromTextEnvelopeAnyOf :: [FromSomeType HasTextEnvelope b] -> TextEnvelope -> Either TextEnvelopeError b

readFileTextEnvelopeAnyOf :: [FromSomeType HasTextEnvelope b] -> File content 'In -> IO (Either (FileError TextEnvelopeError) b)

Data family instances

data family AsType t

A family of singleton types used in this API to indicate which type to use where it would otherwise be ambiguous or merely unclear.

Values of this type are passed to deserialisation functions for example.

CDDL Serialisation

data TextEnvelopeCddlError

Constructors

TextEnvelopeCddlErrCBORDecodingError DecoderError

TextEnvelopeCddlAesonDecodeError FilePath String

TextEnvelopeCddlUnknownKeyWitness

TextEnvelopeCddlTypeError

Fields

[Text]

Expected types

Text

Actual types

TextEnvelopeCddlErrUnknownType Text

TextEnvelopeCddlErrByronKeyWitnessUnsupported

data FromSomeTypeCDDL c b where

This GADT allows us to deserialise a tx or key witness without having to provide the era.

Constructors

FromCDDLTx

Fields

:: forall b. Text

CDDL type that we want

-> (InAnyShelleyBasedEra Tx -> b)

-> FromSomeTypeCDDL TextEnvelope b

FromCDDLWitness

Fields

:: forall b. Text

CDDL type that we want

-> (InAnyShelleyBasedEra KeyWitness -> b)

-> FromSomeTypeCDDL TextEnvelope b

cddlTypeToEra :: Text -> Either TextEnvelopeCddlError AnyShelleyBasedEra

Deprecated: Use textEnvelopeTypeToEra instead.

Reading one of several transaction or key witness types

readFileTextEnvelopeCddlAnyOf :: [FromSomeTypeCDDL TextEnvelope b] -> FilePath -> IO (Either (FileError TextEnvelopeCddlError) b)

Deprecated: Use readFileTextEnvelopeAnyOf instead.

deserialiseFromTextEnvelopeCddlAnyOf :: [FromSomeTypeCDDL TextEnvelope b] -> TextEnvelope -> Either TextEnvelopeCddlError b

Deprecated: Use deserialiseFromTextEnvelopeAnyOf instead.

writeTxFileTextEnvelope :: ShelleyBasedEra era -> File content 'Out -> Tx era -> IO (Either (FileError ()) ())

writeTxFileTextEnvelopeCddl :: ShelleyBasedEra era -> File content 'Out -> Tx era -> IO (Either (FileError ()) ())

Deprecated: Use writeTxFileTextEnvelope instead.

writeTxFileTextEnvelopeCanonical :: ShelleyBasedEra era -> File content 'Out -> Tx era -> IO (Either (FileError ()) ())

Write transaction in the text envelope format. The CBOR will be in canonical format according to RFC 7049. It is also a requirement of CIP-21, which is not fully implemented.

RFC 7049: https://datatracker.ietf.org/doc/html/rfc7049#section-3.9

CIP-21: https://github.com/cardano-foundation/CIPs/blob/master/CIP-0021/README.md#canonical-cbor-serialization-format

writeTxFileTextEnvelopeCanonicalCddl :: ShelleyBasedEra era -> File content 'Out -> Tx era -> IO (Either (FileError ()) ())

Deprecated: Use writeTxFileTextEnvelopeCanonical instead.

Write transaction in the text envelope format. The CBOR will be in canonical format according to RFC 7049. It is also a requirement of CIP-21, which is not fully implemented.

RFC 7049: https://datatracker.ietf.org/doc/html/rfc7049#section-3.9

CIP-21: https://github.com/cardano-foundation/CIPs/blob/master/CIP-0021/README.md#canonical-cbor-serialization-format

writeTxWitnessFileTextEnvelope :: ShelleyBasedEra era -> File () 'Out -> KeyWitness era -> IO (Either (FileError ()) ())

writeTxWitnessFileTextEnvelopeCddl :: ShelleyBasedEra era -> File () 'Out -> KeyWitness era -> IO (Either (FileError ()) ())

Deprecated: Use writeTxWitnessFileTextEnvelope instead.

deserialiseByronTxCddl :: TextEnvelope -> Either TextEnvelopeCddlError (ATxAux ByteString)

Deprecated: Use deserialiseByronTx instead.

serialiseWitnessLedgerCddl :: ShelleyBasedEra era -> KeyWitness era -> TextEnvelope

Deprecated: Use serialiseWitnessLedger instead.

deserialiseWitnessLedgerCddl :: ShelleyBasedEra era -> TextEnvelope -> Either TextEnvelopeCddlError (KeyWitness era)

Deprecated: Use deserialiseWitnessLedger instead.

Byron tx serialization

serialiseByronTx :: ATxAux ByteString -> TextEnvelope

deserialiseByronTx :: TextEnvelope -> Either TextEnvelopeError (ATxAux ByteString)

writeByronTxFileTextEnvelopeCddl :: File content 'Out -> ATxAux ByteString -> IO (Either (FileError ()) ())

Deprecated: Use writeByronTxFileTextEnvelope instead.