Cardano.Api.SerialiseBech32

Description

Bech32 Serialisation

Documentation

class (HasTypeProxy a, SerialiseAsRawBytes a) => SerialiseAsBech32 a where

Methods

bech32PrefixFor :: a -> Text

The human readable prefix to use when encoding this value to Bech32.

bech32PrefixesPermitted :: AsType a -> [Text]

The set of human readable prefixes that can be used for this type.

serialiseToBech32 :: SerialiseAsBech32 a => a -> Text

data Bech32DecodeError

Bech32 decoding error.

Constructors

Bech32DecodingError !DecodingError
There was an error decoding the string as Bech32.

Bech32UnexpectedPrefix !Text !(Set Text)
The human-readable prefix in the Bech32-encoded string is not one of the ones expected.

Bech32DataPartToBytesError !Text
There was an error in extracting a ByteString from the data part of the Bech32-encoded string.

Bech32DeserialiseFromBytesError !ByteString
There was an error in deserialising the bytes into a value of the expected type.

Bech32WrongPrefix !Text !Text
The human-readable prefix in the Bech32-encoded string does not correspond to the prefix that should be used for the payload value.

deserialiseFromBech32 :: SerialiseAsBech32 a => AsType a -> Text -> Either Bech32DecodeError a

deserialiseAnyOfFromBech32 :: [FromSomeType SerialiseAsBech32 b] -> Text -> Either Bech32DecodeError b