Cardano.Api.SerialiseTextEnvelope

Description

TextEnvelope Serialisation

Documentation

class SerialiseAsCBOR a => HasTextEnvelope a where

Minimal complete definition

textEnvelopeType

Methods

textEnvelopeType :: AsType a -> TextEnvelopeType

textEnvelopeDefaultDescr :: a -> TextEnvelopeDescr

textEnvelopeTypeInEra :: HasTextEnvelope (f era) => CardanoEra era -> AsType (f era) -> TextEnvelopeType

data TextEnvelope

A TextEnvelope is a structured envelope for serialised binary values with an external format with a semi-readable textual format.

It contains a "type" field, e.g. "PublicKeyByron" or "TxSignedShelley" to indicate the type of the encoded data. This is used as a sanity check and to help readers.

It also contains a "title" field which is free-form, and could be used to indicate the role or purpose to a reader.

Constructors

TextEnvelope

Fields

teType :: !TextEnvelopeType

teDescription :: !TextEnvelopeDescr

teRawCBOR :: !ByteString

newtype TextEnvelopeType

Constructors

TextEnvelopeType String

newtype TextEnvelopeDescr

Constructors

TextEnvelopeDescr String

textEnvelopeRawCBOR :: TextEnvelope -> ByteString

data TextEnvelopeError

The errors that the pure TextEnvelope parsing/decoding functions can return.

Constructors

TextEnvelopeTypeError ![TextEnvelopeType] !TextEnvelopeType
expected, actual

TextEnvelopeDecodeError !DecoderError

TextEnvelopeAesonDecodeError !String

serialiseToTextEnvelope :: HasTextEnvelope a => Maybe TextEnvelopeDescr -> a -> TextEnvelope

deserialiseFromTextEnvelope :: HasTextEnvelope a => AsType a -> TextEnvelope -> Either TextEnvelopeError a

readFileTextEnvelope :: HasTextEnvelope a => AsType a -> File content 'In -> IO (Either (FileError TextEnvelopeError) a)

writeFileTextEnvelope :: HasTextEnvelope a => File content 'Out -> Maybe TextEnvelopeDescr -> a -> IO (Either (FileError ()) ())

readTextEnvelopeFromFile :: FilePath -> IO (Either (FileError TextEnvelopeError) TextEnvelope)

readTextEnvelopeOfTypeFromFile :: TextEnvelopeType -> FilePath -> IO (Either (FileError TextEnvelopeError) TextEnvelope)

textEnvelopeToJSON :: HasTextEnvelope a => Maybe TextEnvelopeDescr -> a -> ByteString

legacyComparison :: TextEnvelopeType -> TextEnvelopeType -> Bool

This is a backwards-compatibility patch to ensure that old envelopes generated by serialiseTxLedgerCddl can be deserialised after switching to the serialiseToTextEnvelope.

Reading one of several key types

data FromSomeType (c :: Type -> Constraint) b where

Constructors

FromSomeType :: forall (c :: Type -> Constraint) a b. c a => AsType a -> (a -> b) -> FromSomeType c b

deserialiseFromTextEnvelopeAnyOf :: [FromSomeType HasTextEnvelope b] -> TextEnvelope -> Either TextEnvelopeError b

readFileTextEnvelopeAnyOf :: [FromSomeType HasTextEnvelope b] -> File content 'In -> IO (Either (FileError TextEnvelopeError) b)

Data family instances

data family AsType t

A family of singleton types used in this API to indicate which type to use where it would otherwise be ambiguous or merely unclear.

Values of this type are passed to deserialisation functions for example.