Cardano.Api.Shelley

Description

This module provides a library interface that is intended to be the complete API for Shelley covering everything, including exposing constructors for the lower level types.

Documentation

module Cardano.Api

Genesis

data ShelleyGenesis c

Shelley genesis information

Note that this is needed only for a pure Shelley network, hence it being defined here rather than in its own module. In mainnet, Shelley will transition naturally from Byron, and thus will never have its own genesis information.

Constructors

ShelleyGenesis

Fields

sgSystemStart :: !UTCTime

sgNetworkMagic :: !Word32

sgNetworkId :: !Network

sgActiveSlotsCoeff :: !PositiveUnitInterval

sgSecurityParam :: !Word64

sgEpochLength :: !EpochSize

sgSlotsPerKESPeriod :: !Word64

sgMaxKESEvolutions :: !Word64

sgSlotLength :: !NominalDiffTimeMicro

sgUpdateQuorum :: !Word64

sgMaxLovelaceSupply :: !Word64

sgProtocolParams :: !(PParams (ShelleyEra c))

sgGenDelegs :: !(Map (KeyHash 'Genesis c) (GenDelegPair c))

sgInitialFunds :: ListMap (Addr c) Coin

sgInitialFunds is intentionally kept lazy, as it can otherwise cause out-of-memory problems in testing and benchmarking.

sgStaking :: ShelleyGenesisStaking c

sgStaking is intentionally kept lazy, as it can otherwise cause out-of-memory problems in testing and benchmarking.

shelleyGenesisDefaults :: ShelleyGenesis StandardCrypto

alonzoGenesisDefaults :: CardanoEra era -> AlonzoGenesis

decodeAlonzoGenesis :: forall era t (m :: Type -> Type). MonadTransError String t m => Maybe (CardanoEra era) -> ByteString -> t m AlonzoGenesis

conwayGenesisDefaults :: ConwayGenesis StandardCrypto

Cryptographic key interface

class (Eq (VerificationKey keyrole), Show (VerificationKey keyrole), SerialiseAsRawBytes (Hash keyrole), HasTextEnvelope (VerificationKey keyrole), HasTextEnvelope (SigningKey keyrole)) => Key keyrole where

Associated Types

data VerificationKey keyrole

data SigningKey keyrole

Methods

getVerificationKey :: SigningKey keyrole -> VerificationKey keyrole

deterministicSigningKey :: AsType keyrole -> Seed -> SigningKey keyrole

deterministicSigningKeySeedSize :: AsType keyrole -> Word

verificationKeyHash :: VerificationKey keyrole -> Hash keyrole

data family VerificationKey keyrole

data family SigningKey keyrole

Hashes

data family Hash keyrole

Type Proxies

data family AsType t

Payment addresses

Constructing and inspecting Shelley payment addresses

data Address addrtype where

Constructors

ShelleyAddress :: Network -> PaymentCredential StandardCrypto -> StakeReference StandardCrypto -> Address ShelleyAddr

toShelleyAddr :: AddressInEra era -> Addr StandardCrypto

fromShelleyAddr :: ShelleyBasedEra era -> Addr StandardCrypto -> AddressInEra era

fromShelleyAddrIsSbe :: ShelleyBasedEra era -> Addr StandardCrypto -> AddressInEra era

fromShelleyAddrToAny :: Addr StandardCrypto -> AddressAny

toShelleyStakeCredential :: StakeCredential -> StakeCredential StandardCrypto

fromShelleyStakeCredential :: StakeCredential StandardCrypto -> StakeCredential

data NetworkId

Constructors

Mainnet

Testnet !NetworkMagic

Stake addresses

data PaymentCredential

Constructors

PaymentCredentialByKey (Hash PaymentKey)

PaymentCredentialByScript ScriptHash

data StakeAddress where

Constructors

StakeAddress :: Network -> StakeCredential StandardCrypto -> StakeAddress

data StakeAddressReference

Constructors

StakeAddressByValue StakeCredential

StakeAddressByPointer StakeAddressPointer

NoStakeAddress

data StakeCredential

Constructors

StakeCredentialByKey (Hash StakeKey)

StakeCredentialByScript ScriptHash

toShelleyStakeAddr :: StakeAddress -> RewardAccount StandardCrypto

fromShelleyStakeAddr :: RewardAccount StandardCrypto -> StakeAddress

fromShelleyStakeReference :: StakeReference StandardCrypto -> StakeAddressReference

fromShelleyPaymentCredential :: PaymentCredential StandardCrypto -> PaymentCredential

Building transactions

Constructing and inspecting transactions

data TxBody era where

Constructors

ShelleyTxBody :: forall era. ShelleyBasedEra era -> TxBody (ShelleyLedgerEra era) -> [Script (ShelleyLedgerEra era)] -> TxBodyScriptData era -> Maybe (TxAuxData (ShelleyLedgerEra era)) -> TxScriptValidity era -> TxBody era

newtype TxId

Constructors

TxId (Hash StandardCrypto EraIndependentTxBody)

toShelleyTxId :: TxId -> TxId StandardCrypto

fromShelleyTxId :: TxId StandardCrypto -> TxId

getTxIdShelley :: (EraCrypto (ShelleyLedgerEra era) ~ StandardCrypto, EraTxBody (ShelleyLedgerEra era)) => ShelleyBasedEra era -> TxBody (ShelleyLedgerEra era) -> TxId

data TxIn

Constructors

TxIn TxId TxIx

toShelleyTxIn :: TxIn -> TxIn StandardCrypto

fromShelleyTxIn :: TxIn StandardCrypto -> TxIn

data TxOut ctx era

Constructors

TxOut (AddressInEra era) (TxOutValue era) (TxOutDatum ctx era) (ReferenceScript era)

toShelleyTxOut :: (HasCallStack, ShelleyLedgerEra era ~ ledgerera) => ShelleyBasedEra era -> TxOut CtxUTxO era -> TxOut ledgerera

fromShelleyTxOut :: ShelleyBasedEra era -> TxOut (ShelleyLedgerEra era) -> TxOut ctx era

newtype TxIx

Constructors

TxIx Word

toMaryValue :: Value -> MaryValue StandardCrypto

fromMaryValue :: MaryValue StandardCrypto -> Value

calcMinimumDeposit :: Value -> Lovelace -> Lovelace

Arbitrary signing

signArbitraryBytesKes :: SigningKey KesKey -> Period -> ByteString -> SignedKES (KES StandardCrypto) ByteString

Signing transactions

Creating transaction witnesses one by one, or all in one go.

data Tx era where

Constructors

ShelleyTx :: forall era. ShelleyBasedEra era -> Tx (ShelleyLedgerEra era) -> Tx era

Incremental signing and separate witnesses

data KeyWitness era where

Constructors

ShelleyBootstrapWitness :: forall era. ShelleyBasedEra era -> BootstrapWitness StandardCrypto -> KeyWitness era

ShelleyKeyWitness :: forall era. ShelleyBasedEra era -> WitVKey 'Witness StandardCrypto -> KeyWitness era

data ShelleyWitnessSigningKey

Constructors

WitnessPaymentKey (SigningKey PaymentKey)

WitnessPaymentExtendedKey (SigningKey PaymentExtendedKey)

WitnessStakeKey (SigningKey StakeKey)

WitnessStakeExtendedKey (SigningKey StakeExtendedKey)

WitnessStakePoolKey (SigningKey StakePoolKey)

WitnessGenesisKey (SigningKey GenesisKey)

WitnessGenesisExtendedKey (SigningKey GenesisExtendedKey)

WitnessGenesisDelegateKey (SigningKey GenesisDelegateKey)

WitnessGenesisDelegateExtendedKey (SigningKey GenesisDelegateExtendedKey)

data ShelleySigningKey

Constructors

ShelleyNormalSigningKey (SignKeyDSIGN StandardCrypto)

ShelleyExtendedSigningKey XPrv

getShelleyKeyWitnessVerificationKey :: ShelleySigningKey -> VKey 'Witness StandardCrypto

getTxBodyAndWitnesses :: Tx era -> (TxBody era, [KeyWitness era])

makeShelleySignature :: SignableRepresentation tosign => tosign -> ShelleySigningKey -> SignedDSIGN StandardCrypto tosign

toShelleySigningKey :: ShelleyWitnessSigningKey -> ShelleySigningKey

Blocks

fromConsensusBlock :: CardanoBlock StandardCrypto ~ block => block -> BlockInMode

toConsensusBlock :: CardanoBlock StandardCrypto ~ block => BlockInMode -> block

fromConsensusTip :: CardanoBlock StandardCrypto ~ block => Tip block -> ChainTip

fromConsensusPointHF :: forall block (xs :: [Type]). HeaderHash block ~ OneEraHash xs => Point block -> ChainPoint

toConsensusPointHF :: forall block (xs :: [Type]). HeaderHash block ~ OneEraHash xs => ChainPoint -> Point block

Transaction metadata

Embedding additional structured data within transactions.

toShelleyMetadata :: Map Word64 TxMetadataValue -> Map Word64 Metadatum

fromShelleyMetadata :: Map Word64 Metadatum -> Map Word64 TxMetadataValue

toShelleyMetadatum :: TxMetadataValue -> Metadatum

fromShelleyMetadatum :: Metadatum -> TxMetadataValue

Protocol parameters

newtype LedgerProtocolParameters era

Constructors

LedgerProtocolParameters

Fields

unLedgerProtocolParameters :: PParams (ShelleyLedgerEra era)

data EraBasedProtocolParametersUpdate era where

Constructors

ShelleyEraBasedProtocolParametersUpdate :: CommonProtocolParametersUpdate -> DeprecatedAfterMaryPParams ShelleyEra -> DeprecatedAfterBabbagePParams ShelleyEra -> ShelleyToAlonzoPParams ShelleyEra -> EraBasedProtocolParametersUpdate ShelleyEra

AllegraEraBasedProtocolParametersUpdate :: CommonProtocolParametersUpdate -> DeprecatedAfterMaryPParams AllegraEra -> ShelleyToAlonzoPParams AllegraEra -> DeprecatedAfterBabbagePParams ShelleyEra -> EraBasedProtocolParametersUpdate AllegraEra

MaryEraBasedProtocolParametersUpdate :: CommonProtocolParametersUpdate -> DeprecatedAfterMaryPParams MaryEra -> ShelleyToAlonzoPParams MaryEra -> DeprecatedAfterBabbagePParams ShelleyEra -> EraBasedProtocolParametersUpdate MaryEra

AlonzoEraBasedProtocolParametersUpdate :: CommonProtocolParametersUpdate -> ShelleyToAlonzoPParams AlonzoEra -> AlonzoOnwardsPParams AlonzoEra -> DeprecatedAfterBabbagePParams ShelleyEra -> EraBasedProtocolParametersUpdate AlonzoEra

BabbageEraBasedProtocolParametersUpdate :: CommonProtocolParametersUpdate -> AlonzoOnwardsPParams BabbageEra -> DeprecatedAfterBabbagePParams ShelleyEra -> IntroducedInBabbagePParams BabbageEra -> EraBasedProtocolParametersUpdate BabbageEra

ConwayEraBasedProtocolParametersUpdate :: CommonProtocolParametersUpdate -> AlonzoOnwardsPParams ConwayEra -> IntroducedInBabbagePParams ConwayEra -> IntroducedInConwayPParams (ShelleyLedgerEra ConwayEra) -> EraBasedProtocolParametersUpdate ConwayEra

data CommonProtocolParametersUpdate

Constructors

CommonProtocolParametersUpdate

Fields

cppMinFeeA :: StrictMaybe Coin

cppMinFeeB :: StrictMaybe Coin

cppMaxBlockBodySize :: StrictMaybe Word32

cppMaxTxSize :: StrictMaybe Word32

cppMaxBlockHeaderSize :: StrictMaybe Word16

cppKeyDeposit :: StrictMaybe Coin

cppPoolDeposit :: StrictMaybe Coin

cppPoolRetireMaxEpoch :: StrictMaybe EpochInterval

cppStakePoolTargetNum :: StrictMaybe Natural

cppPoolPledgeInfluence :: StrictMaybe NonNegativeInterval

cppTreasuryExpansion :: StrictMaybe UnitInterval

cppMonetaryExpansion :: StrictMaybe UnitInterval

cppMinPoolCost :: StrictMaybe Coin

data AlonzoOnwardsPParams ledgerera

Constructors

AlonzoOnwardsPParams

Fields

alCostModels :: StrictMaybe CostModels

alPrices :: StrictMaybe Prices

alMaxTxExUnits :: StrictMaybe ExUnits

alMaxBlockExUnits :: StrictMaybe ExUnits

alMaxValSize :: StrictMaybe Natural

alCollateralPercentage :: StrictMaybe Natural

alMaxCollateralInputs :: StrictMaybe Natural

newtype DeprecatedAfterBabbagePParams ledgerera

Constructors

DeprecatedAfterBabbagePParams (StrictMaybe ProtVer)

newtype DeprecatedAfterMaryPParams ledgerera

Constructors

DeprecatedAfterMaryPParams (StrictMaybe Coin)

data ShelleyToAlonzoPParams ledgerera

Constructors

ShelleyToAlonzoPParams (StrictMaybe Nonce) (StrictMaybe UnitInterval)

newtype IntroducedInBabbagePParams era

Constructors

IntroducedInBabbagePParams (StrictMaybe CoinPerByte)

data IntroducedInConwayPParams era

Constructors

IntroducedInConwayPParams

Fields

icPoolVotingThresholds :: StrictMaybe PoolVotingThresholds

icDRepVotingThresholds :: StrictMaybe DRepVotingThresholds

icMinCommitteeSize :: StrictMaybe Natural

icCommitteeTermLength :: StrictMaybe EpochInterval

icGovActionLifetime :: StrictMaybe EpochInterval

icGovActionDeposit :: StrictMaybe Coin

icDRepDeposit :: StrictMaybe Coin

icDRepActivity :: StrictMaybe EpochInterval

icMinFeeRefScriptCostPerByte :: StrictMaybe NonNegativeInterval

createEraBasedProtocolParamUpdate :: ShelleyBasedEra era -> EraBasedProtocolParametersUpdate era -> PParamsUpdate (ShelleyLedgerEra era)

convertToLedgerProtocolParameters :: ShelleyBasedEra era -> ProtocolParameters -> Either ProtocolParametersConversionError (LedgerProtocolParameters era)

data ProtocolParameters

Constructors

ProtocolParameters

Fields

protocolParamProtocolVersion :: (Natural, Natural)

protocolParamDecentralization :: Maybe Rational

protocolParamExtraPraosEntropy :: Maybe PraosNonce

protocolParamMaxBlockHeaderSize :: Natural

protocolParamMaxBlockBodySize :: Natural

protocolParamMaxTxSize :: Natural

protocolParamTxFeeFixed :: Coin

protocolParamTxFeePerByte :: Coin

protocolParamMinUTxOValue :: Maybe Coin

protocolParamStakeAddressDeposit :: Coin

protocolParamStakePoolDeposit :: Coin

protocolParamMinPoolCost :: Coin

protocolParamPoolRetireMaxEpoch :: EpochInterval

protocolParamStakePoolTargetNum :: Natural

protocolParamPoolPledgeInfluence :: Rational

protocolParamMonetaryExpansion :: Rational

protocolParamTreasuryCut :: Rational

protocolParamCostModels :: Map AnyPlutusScriptVersion CostModel

protocolParamPrices :: Maybe ExecutionUnitPrices

protocolParamMaxTxExUnits :: Maybe ExecutionUnits

protocolParamMaxBlockExUnits :: Maybe ExecutionUnits

protocolParamMaxValueSize :: Maybe Natural

protocolParamCollateralPercent :: Maybe Natural

protocolParamMaxCollateralInputs :: Maybe Natural

protocolParamUTxOCostPerByte :: Maybe Coin

checkProtocolParameters :: ShelleyBasedEra era -> ProtocolParameters -> Either ProtocolParametersError ()

data ProtocolParametersError

Constructors

PParamsErrorMissingMinUTxoValue !AnyCardanoEra

PParamsErrorMissingAlonzoProtocolParameter

Scripts

fromShelleyBasedScript :: ShelleyBasedEra era -> Script (ShelleyLedgerEra era) -> ScriptInEra era

toShelleyScript :: ScriptInEra era -> Script (ShelleyLedgerEra era)

toShelleyMultiSig :: SimpleScript -> Either MultiSigError (MultiSig (ShelleyLedgerEra ShelleyEra))

fromShelleyMultiSig :: MultiSig (ShelleyLedgerEra ShelleyEra) -> SimpleScript

toAllegraTimelock :: (AllegraEraScript era, EraCrypto era ~ StandardCrypto, NativeScript era ~ Timelock era) => SimpleScript -> NativeScript era

fromAllegraTimelock :: (AllegraEraScript era, EraCrypto era ~ StandardCrypto) => NativeScript era -> SimpleScript

toShelleyScriptHash :: ScriptHash -> ScriptHash StandardCrypto

fromShelleyScriptHash :: ScriptHash StandardCrypto -> ScriptHash

data PlutusScript lang where

Constructors

PlutusScriptSerialised :: forall lang. ShortByteString -> PlutusScript lang

data PlutusScriptOrReferenceInput lang

Constructors

PScript (PlutusScript lang)

PReferenceScript TxIn

data SimpleScriptOrReferenceInput lang

Constructors

SScript SimpleScript

SReferenceScript TxIn

toPlutusData :: ScriptData -> Data

fromPlutusData :: Data -> ScriptData

toAlonzoData :: Era ledgerera => HashableScriptData -> Data ledgerera

fromAlonzoData :: Data ledgerera -> HashableScriptData

toAlonzoPrices :: ExecutionUnitPrices -> Either ProtocolParametersConversionError Prices

fromAlonzoPrices :: Prices -> ExecutionUnitPrices

toAlonzoExUnits :: ExecutionUnits -> ExUnits

fromAlonzoExUnits :: ExUnits -> ExecutionUnits

toScriptIndex :: AlonzoEraOnwards era -> PlutusPurpose AsIx (ShelleyLedgerEra era) -> ScriptWitnessIndex

scriptDataFromJsonDetailedSchema :: Value -> Either ScriptDataJsonSchemaError HashableScriptData

scriptDataToJsonDetailedSchema :: HashableScriptData -> Value

calculateExecutionUnitsLovelace :: Prices -> ExecutionUnits -> Maybe Coin

Reference Scripts

data ReferenceScript era where

Constructors

ReferenceScript :: forall era. BabbageEraOnwards era -> ScriptInAnyLang -> ReferenceScript era

ReferenceScriptNone :: forall era. ReferenceScript era

refScriptToShelleyScript :: ShelleyBasedEra era -> ReferenceScript era -> StrictMaybe (Script (ShelleyLedgerEra era))

Certificates

data Certificate era where

Constructors

ShelleyRelatedCertificate :: forall era. ShelleyToBabbageEra era -> ShelleyTxCert (ShelleyLedgerEra era) -> Certificate era

ConwayCertificate :: forall era. ConwayEraOnwards era -> ConwayTxCert (ShelleyLedgerEra era) -> Certificate era

toShelleyCertificate :: Certificate era -> TxCert (ShelleyLedgerEra era)

fromShelleyCertificate :: ShelleyBasedEra era -> TxCert (ShelleyLedgerEra era) -> Certificate era

toShelleyPoolParams :: StakePoolParameters -> PoolParams StandardCrypto

Operational certificates

data OperationalCertificate

Constructors

OperationalCertificate !(OCert StandardCrypto) !(VerificationKey StakePoolKey)

data OperationalCertificateIssueCounter

Constructors

OperationalCertificateIssueCounter

Fields

opCertIssueCount :: !Word64

opCertIssueColdKey :: !(VerificationKey StakePoolKey)

data OperationalCertIssueError

Constructors

OperationalCertKeyMismatch (VerificationKey StakePoolKey) (VerificationKey StakePoolKey)

Stake Pool

data StakePoolMetadata

Constructors

StakePoolMetadata !Text !Text !Text !Text

stakePoolName :: StakePoolMetadata -> Text

stakePoolDescription :: StakePoolMetadata -> Text

stakePoolTicker :: StakePoolMetadata -> Text

stakePoolHomepage :: StakePoolMetadata -> Text

data StakePoolMetadataReference

Constructors

StakePoolMetadataReference Text (Hash StakePoolMetadata)

stakePoolMetadataURL :: StakePoolMetadataReference -> Text

stakePoolMetadataHash :: StakePoolMetadataReference -> Hash StakePoolMetadata

data StakePoolParameters

Constructors

StakePoolParameters PoolId (Hash VrfKey) Coin Rational StakeAddress Coin [Hash StakeKey] [StakePoolRelay] (Maybe StakePoolMetadataReference)

stakePoolId :: StakePoolParameters -> PoolId

stakePoolVRF :: StakePoolParameters -> Hash VrfKey

stakePoolCost :: StakePoolParameters -> Coin

stakePoolMargin :: StakePoolParameters -> Rational

stakePoolRewardAccount :: StakePoolParameters -> StakeAddress

stakePoolPledge :: StakePoolParameters -> Coin

stakePoolOwners :: StakePoolParameters -> [Hash StakeKey]

stakePoolRelays :: StakePoolParameters -> [StakePoolRelay]

stakePoolMetadata :: StakePoolParameters -> Maybe StakePoolMetadataReference

data StakePoolRelay

Constructors

StakePoolRelayIp (Maybe IPv4) (Maybe IPv6) (Maybe PortNumber)

StakePoolRelayDnsARecord ByteString (Maybe PortNumber)

StakePoolRelayDnsSrvRecord ByteString

newtype EpochNo

An epoch, i.e. the number of the epoch.

Constructors

EpochNo

Fields

unEpochNo :: Word64

Governance Actions

createAnchor :: Url -> ByteString -> Anchor StandardCrypto

createPreviousGovernanceActionId :: forall era (r :: GovActionPurpose). EraCrypto (ShelleyLedgerEra era) ~ StandardCrypto => TxId -> Word16 -> GovPurposeId r (ShelleyLedgerEra era)

createGovernanceActionId :: TxId -> Word16 -> GovActionId StandardCrypto

DRep

newtype DRepMetadata

Constructors

DRepMetadata ByteString

data DRepMetadataReference

Constructors

DRepMetadataReference Text (Hash DRepMetadata)

Stake pool operator's keys

data StakePoolKey

type PoolId = Hash StakePoolKey

KES keys

data KesKey

newtype KESPeriod

Constructors

KESPeriod

Fields

unKESPeriod :: Word

VRF keys

data VrfKey

Low level protocol interaction with a Cardano node

data LocalNodeConnectInfo

Constructors

LocalNodeConnectInfo ConsensusModeParams NetworkId SocketPath

data LocalNodeClientProtocols block point tip slot tx txid txerr (query :: Type -> Type) (m :: Type -> Type)

Constructors

LocalNodeClientProtocols (LocalChainSyncClient block point tip m) (Maybe (LocalTxSubmissionClient tx txerr m ())) (Maybe (LocalStateQueryClient block point query m ())) (Maybe (LocalTxMonitorClient txid tx slot m ()))

Shelley based eras

type family ShelleyLedgerEra era = (ledgerera :: Type) | ledgerera -> era where ...

Equations

ShelleyLedgerEra ShelleyEra = StandardShelley

ShelleyLedgerEra AllegraEra = StandardAllegra

ShelleyLedgerEra MaryEra = StandardMary

ShelleyLedgerEra AlonzoEra = StandardAlonzo

ShelleyLedgerEra BabbageEra = StandardBabbage

ShelleyLedgerEra ConwayEra = StandardConway

Ledger Events

data LedgerEvent

Constructors

PoolRegistration

PoolReRegistration

IncrementalRewardsDistribution EpochNo (Map StakeCredential (Set (Reward StandardCrypto)))

RewardsDistribution EpochNo (Map StakeCredential (Set (Reward StandardCrypto)))

MIRDistribution MIRDistributionDetails

PoolReap PoolReapDetails

SuccessfulPlutusScript (NonEmpty (PlutusWithContext StandardCrypto))

FailedPlutusScript (NonEmpty (PlutusWithContext StandardCrypto))

NewGovernanceProposals (TxId StandardCrypto) AnyProposals

EpochBoundaryRatificationState AnyRatificationState

data AnyProposals

Constructors

EraPParams era => AnyProposals (Proposals era)

data AnyRatificationState

Constructors

EraPParams era => AnyRatificationState (RatifyState era)

data MIRDistributionDetails

Constructors

MIRDistributionDetails

Fields

mirddReservePayouts :: Map StakeCredential Coin

mirddTreasuryPayouts :: Map StakeCredential Coin

mirddReservesToTreasury :: Coin

mirddTreasuryToReserves :: Coin

data PoolReapDetails

Constructors

PoolReapDetails

Fields

prdEpochNo :: EpochNo

prdRefunded :: Map StakeCredential (Map (Hash StakePoolKey) Coin)

prdUnclaimed :: Map StakeCredential (Map (Hash StakePoolKey) Coin)

toLedgerEvent :: ConvertLedgerEvent blk => WrapLedgerEvent blk -> Maybe LedgerEvent

Local State Query

newtype DebugLedgerState era

Constructors

DebugLedgerState

Fields

unDebugLedgerState :: NewEpochState (ShelleyLedgerEra era)

decodeDebugLedgerState :: FromCBOR (DebugLedgerState era) => SerialisedDebugLedgerState era -> Either (ByteString, DecoderError) (DebugLedgerState era)

newtype ProtocolState era

Constructors

ProtocolState (Serialised (ChainDepState (ConsensusProtocol era)))

decodeProtocolState :: FromCBOR (ChainDepState (ConsensusProtocol era)) => ProtocolState era -> Either (ByteString, DecoderError) (ChainDepState (ConsensusProtocol era))

newtype SerialisedDebugLedgerState era

Constructors

SerialisedDebugLedgerState (Serialised (NewEpochState (ShelleyLedgerEra era)))

newtype CurrentEpochState era

Constructors

CurrentEpochState (EpochState (ShelleyLedgerEra era))

newtype SerialisedCurrentEpochState era

Constructors

SerialisedCurrentEpochState (Serialised (EpochState (ShelleyLedgerEra era)))

decodeCurrentEpochState :: ShelleyBasedEra era -> SerialisedCurrentEpochState era -> Either DecoderError (CurrentEpochState era)

newtype PoolState era

Constructors

PoolState (PState (ShelleyLedgerEra era))

newtype SerialisedPoolState era

Constructors

SerialisedPoolState (Serialised (PState (ShelleyLedgerEra era)))

decodePoolState :: (Era (ShelleyLedgerEra era), DecCBOR (PState (ShelleyLedgerEra era))) => SerialisedPoolState era -> Either DecoderError (PoolState era)

newtype PoolDistribution era

Constructors

PoolDistribution

Fields

unPoolDistr :: PoolDistr (EraCrypto (ShelleyLedgerEra era))

newtype SerialisedPoolDistribution era

Constructors

SerialisedPoolDistribution (Serialised (PoolDistr (EraCrypto (ShelleyLedgerEra era))))

decodePoolDistribution :: Crypto (EraCrypto (ShelleyLedgerEra era)) => ShelleyBasedEra era -> SerialisedPoolDistribution era -> Either DecoderError (PoolDistribution era)

newtype StakeSnapshot era

Constructors

StakeSnapshot (StakeSnapshots (EraCrypto (ShelleyLedgerEra era)))

newtype SerialisedStakeSnapshots era

Constructors

SerialisedStakeSnapshots (Serialised (StakeSnapshots (EraCrypto (ShelleyLedgerEra era))))

decodeStakeSnapshot :: FromCBOR (StakeSnapshots (EraCrypto (ShelleyLedgerEra era))) => SerialisedStakeSnapshots era -> Either DecoderError (StakeSnapshot era)

newtype UTxO era

Constructors

UTxO

Fields

unUTxO :: Map TxIn (TxOut CtxUTxO era)

data AcquiringFailure

Constructors

AFPointTooOld

AFPointNotOnChain

newtype SystemStart

System start

Slots are counted from the system start.

Constructors

SystemStart

Fields

getSystemStart :: UTCTime

Governance

data GovernanceAction era

Constructors

MotionOfNoConfidence (StrictMaybe (GovPurposeId 'CommitteePurpose (ShelleyLedgerEra era)))

ProposeNewConstitution (StrictMaybe (GovPurposeId 'ConstitutionPurpose (ShelleyLedgerEra era))) (Anchor StandardCrypto) (StrictMaybe (ScriptHash StandardCrypto))

ProposeNewCommittee (StrictMaybe (GovPurposeId 'CommitteePurpose (ShelleyLedgerEra era))) [Credential 'ColdCommitteeRole StandardCrypto] (Map (Credential 'ColdCommitteeRole StandardCrypto) EpochNo) Rational

InfoAct

TreasuryWithdrawal [(Network, StakeCredential, Coin)] !(StrictMaybe (ScriptHash StandardCrypto))

InitiateHardfork (StrictMaybe (GovPurposeId 'HardForkPurpose (ShelleyLedgerEra era))) ProtVer

UpdatePParams (StrictMaybe (GovPurposeId 'PParamUpdatePurpose (ShelleyLedgerEra era))) (PParamsUpdate (ShelleyLedgerEra era)) !(StrictMaybe (ScriptHash StandardCrypto))

newtype GovernanceActionId era

Constructors

GovernanceActionId

Fields

unGovernanceActionId :: GovActionId (EraCrypto (ShelleyLedgerEra era))

newtype Proposal era

Constructors

Proposal

Fields

unProposal :: ProposalProcedure (ShelleyLedgerEra era)

newtype VotingProcedure era

Constructors

VotingProcedure

Fields

unVotingProcedure :: VotingProcedure (ShelleyLedgerEra era)

newtype VotingProcedures era

Constructors

VotingProcedures

Fields

unVotingProcedures :: VotingProcedures (ShelleyLedgerEra era)

data GovernancePoll

Constructors

GovernancePoll

Fields

govPollQuestion :: Text

govPollAnswers :: [Text]

govPollNonce :: Maybe Word

data GovernancePollAnswer

Constructors

GovernancePollAnswer

Fields

govAnsPoll :: Hash GovernancePoll

govAnsChoice :: Word

data GovernancePollError

Constructors

ErrGovernancePollMismatch GovernancePollMismatchError

ErrGovernancePollNoAnswer

ErrGovernancePollUnauthenticated

ErrGovernancePollMalformedAnswer DecoderError

ErrGovernancePollInvalidAnswer GovernancePollInvalidAnswerError

data Vote

Constructors

No

Yes

Abstain

newtype Voter era

Constructors

Voter (Voter (EraCrypto (ShelleyLedgerEra era)))

createProposalProcedure :: ShelleyBasedEra era -> Network -> Coin -> StakeCredential -> GovernanceAction era -> Anchor StandardCrypto -> Proposal era

createVotingProcedure :: ConwayEraOnwards era -> Vote -> Maybe (Url, Text) -> VotingProcedure era

renderGovernancePollError :: GovernancePollError -> Text

fromProposalProcedure :: ShelleyBasedEra era -> Proposal era -> (Coin, StakeCredential, GovernanceAction era)

hashGovernancePoll :: GovernancePoll -> Hash GovernancePoll

verifyPollAnswer :: GovernancePoll -> InAnyShelleyBasedEra Tx -> Either GovernancePollError [Hash PaymentKey]

Various calculations

data LeadershipError

Constructors

LeaderErrDecodeLedgerStateFailure

LeaderErrDecodeProtocolStateFailure (ByteString, DecoderError)

LeaderErrDecodeProtocolEpochStateFailure DecoderError

LeaderErrGenesisSlot

LeaderErrStakePoolHasNoStake PoolId

LeaderErrStakeDistribUnstable SlotNo SlotNo SlotNo SlotNo

LeaderErrSlotRangeCalculationFailure Text

LeaderErrCandidateNonceStillEvolving

currentEpochEligibleLeadershipSlots :: ShelleyBasedEra era -> ShelleyGenesis StandardCrypto -> EpochInfo (Either Text) -> PParams (ShelleyLedgerEra era) -> ProtocolState era -> PoolId -> SigningKey VrfKey -> SerialisedPoolDistribution era -> EpochNo -> Either LeadershipError (Set SlotNo)

evaluateTransactionExecutionUnitsShelley :: ShelleyBasedEra era -> SystemStart -> LedgerEpochInfo -> LedgerProtocolParameters era -> UTxO era -> Tx (ShelleyLedgerEra era) -> Either (TransactionValidityError era) (Map ScriptWitnessIndex (Either ScriptExecutionError (EvalTxExecutionUnitsLog, ExecutionUnits)))

nextEpochEligibleLeadershipSlots :: ShelleyBasedEra era -> ShelleyGenesis StandardCrypto -> SerialisedCurrentEpochState era -> ProtocolState era -> PoolId -> SigningKey VrfKey -> PParams (ShelleyLedgerEra era) -> EpochInfo (Either Text) -> (ChainTip, EpochNo) -> Either LeadershipError (Set SlotNo)

Conversions

shelleyPayAddrToPlutusPubKHash :: Address ShelleyAddr -> Maybe PubKeyHash

toConsensusGenTx :: CardanoBlock StandardCrypto ~ block => TxInMode -> GenTx block

fromAlonzoCostModels :: CostModels -> Map AnyPlutusScriptVersion CostModel

toLedgerNonce :: Maybe PraosNonce -> Nonce

toShelleyNetwork :: NetworkId -> Network

fromShelleyPoolParams :: PoolParams StandardCrypto -> StakePoolParameters

fromLedgerPParamsUpdate :: ShelleyBasedEra era -> PParamsUpdate (ShelleyLedgerEra era) -> ProtocolParametersUpdate

emptyVotingProcedures :: VotingProcedures era

mergeVotingProcedures :: VotingProcedures era -> VotingProcedures era -> Either (VotesMergingConflict era) (VotingProcedures era)

singletonVotingProcedures :: ConwayEraOnwards era -> Voter (EraCrypto (ShelleyLedgerEra era)) -> GovActionId (EraCrypto (ShelleyLedgerEra era)) -> VotingProcedure (ShelleyLedgerEra era) -> VotingProcedures era

newtype VotesMergingConflict era

Constructors

VotesMergingConflict (Voter (EraCrypto (ShelleyLedgerEra era)), [GovActionId (EraCrypto (ShelleyLedgerEra era))])