Cardano.Api.Tx.Body

Description

Transaction bodies

Documentation

parseTxId :: Parser TxId

Transaction bodies

data TxBody era where

Constructors

ShelleyTxBody

Fields

:: forall era. ShelleyBasedEra era

-> TxBody (ShelleyLedgerEra era)

-> [Script (ShelleyLedgerEra era)]

-> TxBodyScriptData era

-> Maybe (TxAuxData (ShelleyLedgerEra era))

-> TxScriptValidity era

Mark script as expected to pass or fail validation

-> TxBody era

Bundled Patterns

pattern TxBody :: TxBodyContent ViewTx era -> TxBody era

createTransactionBody :: ShelleyBasedEra era -> TxBodyContent BuildTx era -> Either TxBodyError (TxBody era)

createAndValidateTransactionBody :: ShelleyBasedEra era -> TxBodyContent BuildTx era -> Either TxBodyError (TxBody era)

data TxBodyContent build era

Constructors

TxBodyContent

Fields

txIns :: TxIns build era

txInsCollateral :: TxInsCollateral era

txInsReference :: TxInsReference build era

txOuts :: [TxOut CtxTx era]

txTotalCollateral :: TxTotalCollateral era

txReturnCollateral :: TxReturnCollateral CtxTx era

txFee :: TxFee era

txValidityLowerBound :: TxValidityLowerBound era

txValidityUpperBound :: TxValidityUpperBound era

txMetadata :: TxMetadataInEra era

txAuxScripts :: TxAuxScripts era

txExtraKeyWits :: TxExtraKeyWitnesses era

txProtocolParams :: BuildTxWith build (Maybe (LedgerProtocolParameters era))

txWithdrawals :: TxWithdrawals build era

txCertificates :: TxCertificates build era

txUpdateProposal :: TxUpdateProposal era

txMintValue :: TxMintValue build era

txScriptValidity :: TxScriptValidity era

txProposalProcedures :: Maybe (Featured ConwayEraOnwards era (TxProposalProcedures build era))

txVotingProcedures :: Maybe (Featured ConwayEraOnwards era (TxVotingProcedures build era))

txCurrentTreasuryValue :: Maybe (Featured ConwayEraOnwards era Coin)

Current treasury value

txTreasuryDonation :: Maybe (Featured ConwayEraOnwards era Coin)

Treasury donation to perform

Byron only

makeByronTransactionBody :: TxIns BuildTx ByronEra -> [TxOut CtxTx ByronEra] -> Either TxBodyError (Annotated Tx ByteString)

Transaction body builders

defaultTxBodyContent :: ShelleyBasedEra era -> TxBodyContent BuildTx era

defaultTxFee :: ShelleyBasedEra era -> TxFee era

defaultTxValidityUpperBound :: ShelleyBasedEra era -> TxValidityUpperBound era

setTxIns :: TxIns build era -> TxBodyContent build era -> TxBodyContent build era

modTxIns :: (TxIns build era -> TxIns build era) -> TxBodyContent build era -> TxBodyContent build era

addTxIn :: (TxIn, BuildTxWith build (Witness WitCtxTxIn era)) -> TxBodyContent build era -> TxBodyContent build era

setTxInsCollateral :: TxInsCollateral era -> TxBodyContent build era -> TxBodyContent build era

setTxInsReference :: TxInsReference build era -> TxBodyContent build era -> TxBodyContent build era

setTxOuts :: [TxOut CtxTx era] -> TxBodyContent build era -> TxBodyContent build era

modTxOuts :: ([TxOut CtxTx era] -> [TxOut CtxTx era]) -> TxBodyContent build era -> TxBodyContent build era

addTxOut :: TxOut CtxTx era -> TxBodyContent build era -> TxBodyContent build era

setTxTotalCollateral :: TxTotalCollateral era -> TxBodyContent build era -> TxBodyContent build era

setTxReturnCollateral :: TxReturnCollateral CtxTx era -> TxBodyContent build era -> TxBodyContent build era

setTxFee :: TxFee era -> TxBodyContent build era -> TxBodyContent build era

setTxValidityLowerBound :: TxValidityLowerBound era -> TxBodyContent build era -> TxBodyContent build era

setTxValidityUpperBound :: TxValidityUpperBound era -> TxBodyContent build era -> TxBodyContent build era

setTxMetadata :: TxMetadataInEra era -> TxBodyContent build era -> TxBodyContent build era

setTxAuxScripts :: TxAuxScripts era -> TxBodyContent build era -> TxBodyContent build era

setTxExtraKeyWits :: TxExtraKeyWitnesses era -> TxBodyContent build era -> TxBodyContent build era

setTxProtocolParams :: BuildTxWith build (Maybe (LedgerProtocolParameters era)) -> TxBodyContent build era -> TxBodyContent build era

setTxWithdrawals :: TxWithdrawals build era -> TxBodyContent build era -> TxBodyContent build era

setTxCertificates :: TxCertificates build era -> TxBodyContent build era -> TxBodyContent build era

setTxUpdateProposal :: TxUpdateProposal era -> TxBodyContent build era -> TxBodyContent build era

setTxProposalProcedures :: Maybe (Featured ConwayEraOnwards era (TxProposalProcedures build era)) -> TxBodyContent build era -> TxBodyContent build era

setTxVotingProcedures :: Maybe (Featured ConwayEraOnwards era (TxVotingProcedures build era)) -> TxBodyContent build era -> TxBodyContent build era

setTxMintValue :: TxMintValue build era -> TxBodyContent build era -> TxBodyContent build era

setTxScriptValidity :: TxScriptValidity era -> TxBodyContent build era -> TxBodyContent build era

setTxCurrentTreasuryValue :: Maybe (Featured ConwayEraOnwards era Coin) -> TxBodyContent build era -> TxBodyContent build era

setTxTreasuryDonation :: Maybe (Featured ConwayEraOnwards era Coin) -> TxBodyContent build era -> TxBodyContent build era

data TxBodyError

Constructors

TxBodyEmptyTxIns

TxBodyEmptyTxInsCollateral

TxBodyEmptyTxOuts

TxBodyOutputNegative !Quantity !TxOutInAnyEra

TxBodyOutputOverflow !Quantity !TxOutInAnyEra

TxBodyMetadataError ![(Word64, TxMetadataRangeError)]

TxBodyMintAdaError

TxBodyInIxOverflow !TxIn

TxBodyMissingProtocolParams

TxBodyProtocolParamsConversionError !ProtocolParametersConversionError

data TxBodyScriptData era where

Constructors

TxBodyNoScriptData :: forall era. TxBodyScriptData era

TxBodyScriptData :: forall era. AlonzoEraOnwardsConstraints era => AlonzoEraOnwards era -> TxDats (ShelleyLedgerEra era) -> Redeemers (ShelleyLedgerEra era) -> TxBodyScriptData era

data TxScriptValidity era where

A representation of whether the era supports tx script validity.

The Alonzo and subsequent eras support script validity.

Constructors

TxScriptValidityNone :: forall era. TxScriptValidity era

TxScriptValidity :: forall era. AlonzoEraOnwards era -> ScriptValidity -> TxScriptValidity era
Tx script validity is supported in transactions in the Alonzo era onwards.

data ScriptValidity

Indicates whether a script is expected to fail or pass validation.

Constructors

ScriptInvalid
Script is expected to fail validation. Transactions marked as such can include scripts that fail validation. Such transactions may be submitted to the chain, in which case the collateral will be taken upon on chain script validation failure.

ScriptValid
Script is expected to pass validation. Transactions marked as such cannot include scripts that fail validation.

scriptValidityToIsValid :: ScriptValidity -> IsValid

isValidToScriptValidity :: IsValid -> ScriptValidity

txScriptValidityToIsValid :: TxScriptValidity era -> IsValid

txScriptValidityToScriptValidity :: TxScriptValidity era -> ScriptValidity

Transaction Ids

newtype TxId

Constructors

TxId (Hash StandardCrypto EraIndependentTxBody)

getTxId :: TxBody era -> TxId

Calculate the transaction identifier for a TxBody.

getTxIdByron :: ATxAux ByteString -> TxId

getTxIdShelley :: (EraCrypto (ShelleyLedgerEra era) ~ StandardCrypto, EraTxBody (ShelleyLedgerEra era)) => ShelleyBasedEra era -> TxBody (ShelleyLedgerEra era) -> TxId

Transaction inputs

data TxIn

Constructors

TxIn TxId TxIx

type TxIns build era = [(TxIn, BuildTxWith build (Witness WitCtxTxIn era))]

newtype TxIx

Constructors

TxIx Word

genesisUTxOPseudoTxIn :: NetworkId -> Hash GenesisUTxOKey -> TxIn

Compute the TxIn of the initial UTxO pseudo-transaction corresponding to the given address in the genesis initial funds.

The Shelley initial UTxO is constructed from the sgInitialFunds which is not a full UTxO but just a map from addresses to coin values.

This gets turned into a UTxO by making a pseudo-transaction for each address, with the 0th output being the coin value. So to spend from the initial UTxO we need this same TxIn to use as an input to the spending transaction.

getReferenceInputsSizeForTxIds :: ShelleyLedgerEra era ~ ledgerera => BabbageEraOnwards era -> UTxO ledgerera -> Set TxIn -> Int

Calculate the reference inputs size in bytes for provided set of transaction IDs and UTXOs.

Transaction outputs

data CtxTx

The context is a transaction body

data CtxUTxO

The context is the UTxO

data TxOut ctx era

Constructors

TxOut (AddressInEra era) (TxOutValue era) (TxOutDatum ctx era) (ReferenceScript era)

data TxOutValue era where

Constructors

TxOutValueByron :: forall era. Coin -> TxOutValue era

TxOutValueShelleyBased :: forall era. (Eq (Value (ShelleyLedgerEra era)), Show (Value (ShelleyLedgerEra era))) => ShelleyBasedEra era -> Value (ShelleyLedgerEra era) -> TxOutValue era

data TxOutDatum ctx era where

Constructors

TxOutDatumNone :: forall ctx era. TxOutDatum ctx era

TxOutDatumHash :: forall era ctx. AlonzoEraOnwards era -> Hash ScriptData -> TxOutDatum ctx era
A transaction output that only specifies the hash of the datum, but not the full datum value.

TxOutDatumInline :: forall era ctx. BabbageEraOnwards era -> HashableScriptData -> TxOutDatum ctx era
A transaction output that specifies the whole datum instead of the datum hash. Note that the datum map will not be updated with this datum, it only exists at the transaction output.

Bundled Patterns

pattern TxOutDatumInTx :: AlonzoEraOnwards era -> HashableScriptData -> TxOutDatum CtxTx era

toCtxUTxOTxOut :: TxOut CtxTx era -> TxOut CtxUTxO era

lovelaceToTxOutValue :: ShelleyBasedEra era -> Coin -> TxOutValue era

prettyRenderTxOut :: TxOutInAnyEra -> Text

txOutValueToLovelace :: TxOutValue era -> Coin

txOutValueToValue :: TxOutValue era -> Value

parseHash :: SerialiseAsRawBytes (Hash a) => AsType (Hash a) -> Parser (Hash a)

data TxOutInAnyEra where

Constructors

TxOutInAnyEra :: forall era. CardanoEra era -> TxOut CtxTx era -> TxOutInAnyEra

txOutInAnyEra :: CardanoEra era -> TxOut CtxTx era -> TxOutInAnyEra

Convenience constructor for TxOutInAnyEra

Other transaction body types

data TxInsCollateral era where

Constructors

TxInsCollateralNone :: forall era. TxInsCollateral era

TxInsCollateral :: forall era. AlonzoEraOnwards era -> [TxIn] -> TxInsCollateral era

data TxInsReference build era where

Constructors

TxInsReferenceNone :: forall build era. TxInsReference build era

TxInsReference :: forall era build. BabbageEraOnwards era -> [TxIn] -> TxInsReference build era

data TxReturnCollateral ctx era where

Constructors

TxReturnCollateralNone :: forall ctx era. TxReturnCollateral ctx era

TxReturnCollateral :: forall era ctx. BabbageEraOnwards era -> TxOut ctx era -> TxReturnCollateral ctx era

data TxTotalCollateral era where

Constructors

TxTotalCollateralNone :: forall era. TxTotalCollateral era

TxTotalCollateral :: forall era. BabbageEraOnwards era -> Coin -> TxTotalCollateral era

data TxFee era where

Constructors

TxFeeExplicit :: forall era. ShelleyBasedEra era -> Coin -> TxFee era

data TxValidityLowerBound era where

Constructors

TxValidityNoLowerBound :: forall era. TxValidityLowerBound era

TxValidityLowerBound :: forall era. AllegraEraOnwards era -> SlotNo -> TxValidityLowerBound era

data TxValidityUpperBound era where

This was formerly known as the TTL.

Constructors

TxValidityUpperBound :: forall era. ShelleyBasedEra era -> Maybe SlotNo -> TxValidityUpperBound era

data TxMetadataInEra era where

Constructors

TxMetadataNone :: forall era. TxMetadataInEra era

TxMetadataInEra :: forall era. ShelleyBasedEra era -> TxMetadata -> TxMetadataInEra era

data TxAuxScripts era where

Constructors

TxAuxScriptsNone :: forall era. TxAuxScripts era

TxAuxScripts :: forall era. AllegraEraOnwards era -> [ScriptInEra era] -> TxAuxScripts era

data TxExtraKeyWitnesses era where

Constructors

TxExtraKeyWitnessesNone :: forall era. TxExtraKeyWitnesses era

TxExtraKeyWitnesses :: forall era. AlonzoEraOnwards era -> [Hash PaymentKey] -> TxExtraKeyWitnesses era

data TxWithdrawals build era where

Constructors

TxWithdrawalsNone :: forall build era. TxWithdrawals build era

TxWithdrawals :: forall era build. ShelleyBasedEra era -> [(StakeAddress, Coin, BuildTxWith build (Witness WitCtxStake era))] -> TxWithdrawals build era

data TxCertificates build era where

Constructors

TxCertificatesNone :: forall build era. TxCertificates build era

TxCertificates

Fields

:: forall era build. ShelleyBasedEra era

-> [Certificate era]

-> BuildTxWith build [(StakeCredential, Witness WitCtxStake era)]

There can be more than one script witness per stake credential

-> TxCertificates build era

data TxUpdateProposal era where

Constructors

TxUpdateProposalNone :: forall era. TxUpdateProposal era

TxUpdateProposal :: forall era. ShelleyToBabbageEra era -> UpdateProposal -> TxUpdateProposal era

data TxMintValue build era where

Constructors

TxMintNone :: forall build era. TxMintValue build era

TxMintValue :: forall era build. MaryEraOnwards era -> Value -> BuildTxWith build (Map PolicyId (ScriptWitness WitCtxMint era)) -> TxMintValue build era

data TxVotingProcedures build era where

Constructors

TxVotingProceduresNone :: forall build era. TxVotingProcedures build era

TxVotingProcedures :: forall era build. VotingProcedures (ShelleyLedgerEra era) -> BuildTxWith build (Map (Voter (EraCrypto (ShelleyLedgerEra era))) (ScriptWitness WitCtxStake era)) -> TxVotingProcedures build era

data TxProposalProcedures build era where

Constructors

TxProposalProceduresNone :: forall build era. TxProposalProcedures build era

TxProposalProcedures :: forall era build. EraPParams (ShelleyLedgerEra era) => OSet (ProposalProcedure (ShelleyLedgerEra era)) -> BuildTxWith build (Map (ProposalProcedure (ShelleyLedgerEra era)) (ScriptWitness WitCtxStake era)) -> TxProposalProcedures build era

Building vs viewing transactions

data BuildTxWith build a where

Constructors

ViewTx :: forall a. BuildTxWith ViewTx a

BuildTxWith :: forall a. a -> BuildTxWith BuildTx a

data BuildTx

data ViewTx

Inspecting ScriptWitnesses

data AnyScriptWitness era where

A ScriptWitness in any WitCtx. This lets us handle heterogeneous collections of script witnesses from multiple contexts.

Constructors

AnyScriptWitness :: forall witctx era. ScriptWitness witctx era -> AnyScriptWitness era

data ScriptWitnessIndex

Identify the location of a ScriptWitness within the context of a TxBody. These are indexes of the objects within the transaction that need or can use script witnesses: inputs, minted assets, withdrawals and certificates. These are simple numeric indices, enumerated from zero. Thus the indices are not stable if the transaction body is modified.

Constructors

ScriptWitnessIndexTxIn !Word32
The n'th transaction input, in the order of the TxIds.

ScriptWitnessIndexMint !Word32
The n'th minting PolicyId, in the order of the PolicyIds.

ScriptWitnessIndexCertificate !Word32
The n'th certificate, in the list order of the certificates.

ScriptWitnessIndexWithdrawal !Word32
The n'th withdrawal, in the order of the StakeAddresss.

ScriptWitnessIndexVoting !Word32
The n'th vote, in the order of the votes.

ScriptWitnessIndexProposing !Word32
The n'th proposal, in the order of the proposals.

renderScriptWitnessIndex :: ScriptWitnessIndex -> String

collectTxBodyScriptWitnesses :: ShelleyBasedEra era -> TxBodyContent BuildTx era -> [(ScriptWitnessIndex, AnyScriptWitness era)]

toScriptIndex :: AlonzoEraOnwards era -> PlutusPurpose AsIx (ShelleyLedgerEra era) -> ScriptWitnessIndex

Conversion to inline data

scriptDataToInlineDatum :: Era ledgerera => HashableScriptData -> Datum ledgerera

Conversion of ScriptData to binary data which allows for the storage of data onchain within a transaction output.

Internal conversion functions & types

toByronTxId :: TxId -> TxId

toShelleyTxId :: TxId -> TxId StandardCrypto

toShelleyTxIn :: TxIn -> TxIn StandardCrypto

This function may overflow on the transaction index. Call sites must ensure that all uses of this function are appropriately guarded.

toShelleyTxOut :: ShelleyLedgerEra era ~ ledgerera => ShelleyBasedEra era -> TxOut CtxUTxO era -> TxOut ledgerera

toShelleyTxOutAny :: forall ctx era ledgerera. ShelleyLedgerEra era ~ ledgerera => ShelleyBasedEra era -> TxOut ctx era -> TxOut ledgerera

A variant of 'toShelleyTxOutAny that is used only internally to this module that works with a TxOut in any context (including CtxTx) by ignoring embedded datums (taking only their hash).

fromShelleyTxId :: TxId StandardCrypto -> TxId

fromShelleyTxIn :: TxIn StandardCrypto -> TxIn

fromShelleyTxOut :: ShelleyBasedEra era -> TxOut (ShelleyLedgerEra era) -> TxOut ctx era

fromByronTxIn :: TxIn -> TxIn

fromLedgerTxOuts :: ShelleyBasedEra era -> TxBody (ShelleyLedgerEra era) -> TxBodyScriptData era -> [TxOut CtxTx era]

renderTxIn :: TxIn -> Text

Misc helpers

calculateExecutionUnitsLovelace :: Prices -> ExecutionUnits -> Maybe Coin

orderStakeAddrs :: [(StakeAddress, x, v)] -> [(StakeAddress, x, v)]

orderTxIns :: [(TxIn, v)] -> [(TxIn, v)]

Data family instances

data family AsType t

A family of singleton types used in this API to indicate which type to use where it would otherwise be ambiguous or merely unclear.

Values of this type are passed to deserialisation functions for example.

getTxBodyContent :: TxBody era -> TxBodyContent ViewTx era

validateTxIns :: [(TxIn, BuildTxWith BuildTx (Witness WitCtxTxIn era))] -> Either TxBodyError ()

guardShelleyTxInsOverflow :: [TxIn] -> Either TxBodyError ()

validateTxOuts :: ShelleyBasedEra era -> [TxOut CtxTx era] -> Either TxBodyError ()

validateMetadata :: TxMetadataInEra era -> Either TxBodyError ()

validateMintValue :: TxMintValue build era -> Either TxBodyError ()

validateTxInsCollateral :: TxInsCollateral era -> Set Language -> Either TxBodyError ()

validateProtocolParameters :: BuildTxWith BuildTx (Maybe (LedgerProtocolParameters era)) -> Set Language -> Either TxBodyError ()