Cardano.Api.TxBody

Description

Transaction bodies

Transaction bodies

data TxBody era where

Constructors

ByronTxBody ∷ Annotated Tx ByteString → TxBody ByronEra

ShelleyTxBody ∷ ShelleyBasedEra era → TxBody (ShelleyLedgerEra era) → Maybe (AuxiliaryData (ShelleyLedgerEra era)) → TxBody era

makeTransactionBody ∷ ∀ era. IsCardanoEra era ⇒ TxBodyContent era → Either (TxBodyError era) (TxBody era)

data TxBodyContent era

Constructors

TxBodyContent

Fields

txIns ∷ [TxIn]

txOuts ∷ [TxOut era]

txFee ∷ TxFee era

txValidityRange ∷ (TxValidityLowerBound era, TxValidityUpperBound era)

txMetadata ∷ TxMetadataInEra era

txAuxScripts ∷ TxAuxScripts era

txWithdrawals ∷ TxWithdrawals era

txCertificates ∷ TxCertificates era

txUpdateProposal ∷ TxUpdateProposal era

txMintValue ∷ TxMintValue era

data TxBodyError era

Constructors

TxBodyEmptyTxIns

TxBodyEmptyTxOuts

TxBodyOutputNegative Quantity (TxOut era)

TxBodyOutputOverflow Quantity (TxOut era)

TxBodyMetadataError [(Word64, TxMetadataRangeError)]

TxBodyMintAdaError

Transitional utils

makeByronTransaction ∷ [TxIn] → [TxOut ByronEra] → Either (TxBodyError ByronEra) (TxBody ByronEra)

Transitional function to help the CLI move to the updated TxBody API.

makeShelleyTransaction ∷ [TxIn] → [TxOut ShelleyEra] → SlotNo → Lovelace → [Certificate] → [(StakeAddress, Lovelace)] → Maybe TxMetadata → Maybe UpdateProposal → Either (TxBodyError ShelleyEra) (TxBody ShelleyEra)

Transitional function to help the CLI move to the updated TxBody API.

Transaction Ids

newtype TxId

Constructors

TxId (Hash StandardCrypto EraIndependentTxBody)

getTxId ∷ TxBody era → TxId

Calculate the transaction identifier for a TxBody.

Transaction inputs

data TxIn

Constructors

TxIn TxId TxIx

newtype TxIx

Constructors

TxIx Word

genesisUTxOPseudoTxIn ∷ NetworkId → Hash GenesisUTxOKey → TxIn

Compute the TxIn of the initial UTxO pseudo-transaction corresponding to the given address in the genesis initial funds.

The Shelley initial UTxO is constructed from the sgInitialFunds which is not a full UTxO but just a map from addresses to coin values.

This gets turned into a UTxO by making a pseudo-transaction for each address, with the 0th output being the coin value. So to spend from the initial UTxO we need this same TxIn to use as an input to the spending transaction.

Transaction outputs

data TxOut era

Constructors

TxOut (AddressInEra era) (TxOutValue era)

data TxOutValue era where

Constructors

TxOutAdaOnly ∷ OnlyAdaSupportedInEra era → Lovelace → TxOutValue era

TxOutValue ∷ MultiAssetSupportedInEra era → Value → TxOutValue era

Other transaction body types

data TxFee era where

Constructors

TxFeeImplicit ∷ TxFeesImplicitInEra era → TxFee era

TxFeeExplicit ∷ TxFeesExplicitInEra era → Lovelace → TxFee era

data TxValidityLowerBound era where

Constructors

TxValidityNoLowerBound ∷ TxValidityLowerBound era

TxValidityLowerBound ∷ ValidityLowerBoundSupportedInEra era → SlotNo → TxValidityLowerBound era

data TxValidityUpperBound era where

This was formerly known as the TTL.

Constructors

TxValidityNoUpperBound ∷ ValidityNoUpperBoundSupportedInEra era → TxValidityUpperBound era

TxValidityUpperBound ∷ ValidityUpperBoundSupportedInEra era → SlotNo → TxValidityUpperBound era

data TxMetadataInEra era where

Constructors

TxMetadataNone ∷ TxMetadataInEra era

TxMetadataInEra ∷ TxMetadataSupportedInEra era → TxMetadata → TxMetadataInEra era

data TxAuxScripts era where

Constructors

TxAuxScriptsNone ∷ TxAuxScripts era

TxAuxScripts ∷ AuxScriptsSupportedInEra era → [ScriptInEra era] → TxAuxScripts era

data TxWithdrawals era where

Constructors

TxWithdrawalsNone ∷ TxWithdrawals era

TxWithdrawals ∷ WithdrawalsSupportedInEra era → [(StakeAddress, Lovelace)] → TxWithdrawals era

data TxCertificates era where

Constructors

TxCertificatesNone ∷ TxCertificates era

TxCertificates ∷ CertificatesSupportedInEra era → [Certificate] → TxCertificates era

data TxUpdateProposal era where

Constructors

TxUpdateProposalNone ∷ TxUpdateProposal era

TxUpdateProposal ∷ UpdateProposalSupportedInEra era → UpdateProposal → TxUpdateProposal era

data TxMintValue era where

Constructors

TxMintNone ∷ TxMintValue era

TxMintValue ∷ MultiAssetSupportedInEra era → Value → TxMintValue era

Era-dependent transaction body features

data MultiAssetSupportedInEra era where

A representation of whether the era supports multi-asset transactions.

The Mary and subsequent eras support multi-asset transactions.

The negation of this is OnlyAdaSupportedInEra.

Constructors

MultiAssetInMaryEra ∷ MultiAssetSupportedInEra MaryEra
Multi-asset transactions are supported in the Mary era.

data OnlyAdaSupportedInEra era where

A representation of whether the era supports only ada transactions.

Prior to the Mary era only ada transactions are supported. Multi-assets are supported from the Mary era onwards.

This is the negation of MultiAssetSupportedInEra. It exists since we need evidence to be positive.

Constructors

AdaOnlyInByronEra ∷ OnlyAdaSupportedInEra ByronEra

AdaOnlyInShelleyEra ∷ OnlyAdaSupportedInEra ShelleyEra

AdaOnlyInAllegraEra ∷ OnlyAdaSupportedInEra AllegraEra

data TxFeesExplicitInEra era where

A representation of whether the era requires explicitly specified fees in transactions.

The Byron era tx fees are implicit (as the difference bettween the sum of outputs and sum of inputs), but all later eras the fees are specified in the transaction explicitly.

Constructors

TxFeesExplicitInShelleyEra ∷ TxFeesExplicitInEra ShelleyEra

TxFeesExplicitInAllegraEra ∷ TxFeesExplicitInEra AllegraEra

TxFeesExplicitInMaryEra ∷ TxFeesExplicitInEra MaryEra

data TxFeesImplicitInEra era where

A representation of whether the era requires implicitly specified fees in transactions.

This is the negation of TxFeesExplicitInEra.

Constructors

TxFeesImplicitInByronEra ∷ TxFeesImplicitInEra ByronEra

data ValidityUpperBoundSupportedInEra era where

A representation of whether the era supports transactions with an upper bound on the range of slots in which they are valid.

The Shelley and subsequent eras support an upper bound on the validity range. In the Shelley era specifically it is actually required. It is optional in later eras.

Constructors

ValidityUpperBoundInShelleyEra ∷ ValidityUpperBoundSupportedInEra ShelleyEra

ValidityUpperBoundInAllegraEra ∷ ValidityUpperBoundSupportedInEra AllegraEra

ValidityUpperBoundInMaryEra ∷ ValidityUpperBoundSupportedInEra MaryEra

data ValidityNoUpperBoundSupportedInEra era where

A representation of whether the era supports transactions having no upper bound on the range of slots in which they are valid.

Note that the ShelleyEra does not support omitting a validity upper bound. It was introduced as a required field in Shelley and then made optional in Allegra and subsequent eras.

The Byron era supports this by virtue of the fact that it does not support validity ranges at all.

Constructors

ValidityNoUpperBoundInByronEra ∷ ValidityNoUpperBoundSupportedInEra ByronEra

ValidityNoUpperBoundInAllegraEra ∷ ValidityNoUpperBoundSupportedInEra AllegraEra

ValidityNoUpperBoundInMaryEra ∷ ValidityNoUpperBoundSupportedInEra MaryEra

data ValidityLowerBoundSupportedInEra era where

A representation of whether the era supports transactions with a lower bound on the range of slots in which they are valid.

The Allegra and subsequent eras support an optional lower bound on the validity range. No equivalent of ValidityNoUpperBoundSupportedInEra is needed since all eras support having no lower bound.

Constructors

ValidityLowerBoundInAllegraEra ∷ ValidityLowerBoundSupportedInEra AllegraEra

ValidityLowerBoundInMaryEra ∷ ValidityLowerBoundSupportedInEra MaryEra

data TxMetadataSupportedInEra era where

A representation of whether the era supports transaction metadata.

Transaction metadata is supported from the Shelley era onwards.

Constructors

TxMetadataInShelleyEra ∷ TxMetadataSupportedInEra ShelleyEra

TxMetadataInAllegraEra ∷ TxMetadataSupportedInEra AllegraEra

TxMetadataInMaryEra ∷ TxMetadataSupportedInEra MaryEra

data AuxScriptsSupportedInEra era where

A representation of whether the era supports auxiliary scripts in transactions.

Auxiliary scripts are supported from the Allegra era onwards.

Constructors

AuxScriptsInAllegraEra ∷ AuxScriptsSupportedInEra AllegraEra

AuxScriptsInMaryEra ∷ AuxScriptsSupportedInEra MaryEra

data WithdrawalsSupportedInEra era where

A representation of whether the era supports withdrawals from reward accounts.

The Shelley and subsequent eras support stake addresses, their associated reward accounts and support for withdrawals from them.

Constructors

WithdrawalsInShelleyEra ∷ WithdrawalsSupportedInEra ShelleyEra

WithdrawalsInAllegraEra ∷ WithdrawalsSupportedInEra AllegraEra

WithdrawalsInMaryEra ∷ WithdrawalsSupportedInEra MaryEra

data CertificatesSupportedInEra era where

A representation of whether the era supports Certificates embedded in transactions.

The Shelley and subsequent eras support such certificates.

Constructors

CertificatesInShelleyEra ∷ CertificatesSupportedInEra ShelleyEra

CertificatesInAllegraEra ∷ CertificatesSupportedInEra AllegraEra

CertificatesInMaryEra ∷ CertificatesSupportedInEra MaryEra

data UpdateProposalSupportedInEra era where

A representation of whether the era supports UpdateProposals embedded in transactions.

The Shelley and subsequent eras support such update proposals. They Byron era has a notion of an update proposal, but it is a standalone chain object and not embedded in a transaction.

Constructors

UpdateProposalInShelleyEra ∷ UpdateProposalSupportedInEra ShelleyEra

UpdateProposalInAllegraEra ∷ UpdateProposalSupportedInEra AllegraEra

UpdateProposalInMaryEra ∷ UpdateProposalSupportedInEra MaryEra

Feature availability functions

multiAssetSupportedInEra ∷ CardanoEra era → Either (OnlyAdaSupportedInEra era) (MultiAssetSupportedInEra era)

txFeesExplicitInEra ∷ CardanoEra era → Either (TxFeesImplicitInEra era) (TxFeesExplicitInEra era)

validityUpperBoundSupportedInEra ∷ CardanoEra era → Maybe (ValidityUpperBoundSupportedInEra era)

validityNoUpperBoundSupportedInEra ∷ CardanoEra era → Maybe (ValidityNoUpperBoundSupportedInEra era)

validityLowerBoundSupportedInEra ∷ CardanoEra era → Maybe (ValidityLowerBoundSupportedInEra era)

txMetadataSupportedInEra ∷ CardanoEra era → Maybe (TxMetadataSupportedInEra era)

auxScriptsSupportedInEra ∷ CardanoEra era → Maybe (AuxScriptsSupportedInEra era)

withdrawalsSupportedInEra ∷ CardanoEra era → Maybe (WithdrawalsSupportedInEra era)

certificatesSupportedInEra ∷ CardanoEra era → Maybe (CertificatesSupportedInEra era)

updateProposalSupportedInEra ∷ CardanoEra era → Maybe (UpdateProposalSupportedInEra era)

Internal conversion functions

toShelleyTxId ∷ TxId → TxId StandardCrypto

toShelleyTxIn ∷ TxIn → TxIn StandardCrypto

toShelleyTxOut ∷ ∀ era ledgerera. (ShelleyLedgerEra era ~ ledgerera, IsShelleyBasedEra era, ShelleyBased ledgerera) ⇒ TxOut era → TxOut ledgerera

fromShelleyTxId ∷ TxId StandardCrypto → TxId

fromShelleyTxIn ∷ TxIn StandardCrypto → TxIn

fromShelleyTxOut ∷ ∀ era ledgerera. (ShelleyLedgerEra era ~ ledgerera, IsShelleyBasedEra era, ShelleyBased ledgerera) ⇒ TxOut ledgerera → TxOut era

Data family instances

data family AsType t

A family of singleton types used in this API to indicate which type to use where it would otherwise be ambiguous or merely unclear.

Values of this type are passed to deserialisation functions for example.

Conversion functions

fromByronTxIn ∷ TxIn → TxIn