Cardano.Api.Utils

Description

Internal utils for the other Api modules

Documentation

(?!) :: Maybe a -> e -> Either e a

(?!.) :: Either e a -> (e -> e') -> Either e' a

formatParsecError :: ParseError -> String

failEither :: MonadFail m => Either String a -> m a

failEitherWith :: MonadFail m => (e -> String) -> Either e a -> m a

noInlineMaybeToStrictMaybe :: Maybe a -> StrictMaybe a

note :: MonadFail m => String -> Maybe a -> m a

parseFilePath :: String -> String -> Parser FilePath

readFileBlocking :: FilePath -> IO ByteString

runParsecParser :: Parser a -> Text -> Parser a

textShow :: Show a => a -> Text

modifyWith :: (a -> a) -> a -> a

Aids type inference. Use this function to ensure the value is a function that modifies a value.

CLI option parsing

bounded :: (Bounded a, Integral a, Show a) => String -> ReadM a

unsafeBoundedRational :: (HasCallStack, Typeable r, BoundedRational r) => Rational -> r

Convert Rational to a bounded rational. Throw an exception when the rational is out of bounds.