Cardano.Api.Value

Description

Currency values

Documentation

newtype Coin

The amount of value held by a transaction output.

Constructors

Coin

Fields

unCoin :: Integer

Multi-asset values

newtype Quantity

Constructors

Quantity Integer

parseQuantity :: Parser Quantity

Quantity (word64) parser. Only accepts positive quantities.

newtype PolicyId

Constructors

PolicyId

Fields

unPolicyId :: ScriptHash

parsePolicyId :: Parser PolicyId

Policy ID parser.

scriptPolicyId :: Script lang -> PolicyId

newtype AssetName

Constructors

UnsafeAssetName ByteString

parseAssetName :: Parser AssetName

Asset name parser.

data AssetId

Constructors

AdaAssetId

AssetId !PolicyId !AssetName

parseAssetId :: Parser AssetId

Asset ID parser.

data Value

selectAsset :: Value -> AssetId -> Quantity

valueFromList :: [(AssetId, Quantity)] -> Value

Deprecated: Use fromList instead.

valueToList :: Value -> [(AssetId, Quantity)]

Deprecated: Use toList instead.

filterValue :: (AssetId -> Bool) -> Value -> Value

allPositive :: Value -> Bool

Check if the Value consists of only positive quantities.

negateValue :: Value -> Value

This lets you write a - b as a <> negateValue b.

negateLedgerValue :: ShelleyBasedEra era -> Value (ShelleyLedgerEra era) -> Value (ShelleyLedgerEra era)

calcMinimumDeposit :: Value -> Lovelace -> Lovelace

Calculate cost of making a UTxO entry for a given Value and mininimum UTxO value derived from the ProtocolParameters

newtype PolicyAssets

Map of non-ADA assets with their quantity, for a single policy

Constructors

PolicyAssets (Map AssetName Quantity)

policyAssetsToValue :: PolicyId -> PolicyAssets -> Value

valueToPolicyAssets :: Value -> Map PolicyId PolicyAssets

Converts Value to PolicyAssets. Discards any ADA value stored in Value.

multiAssetToPolicyAssets :: MultiAsset -> Map PolicyId PolicyAssets

Convert cardano-ledger's MultiAsset to a map of PolicyAssets

parseUTxOValue :: Parser Value

Parse a Value from its string representation. The resulting amounts must be positive for the parser to succeed.

parseMintingMultiAssetValue :: MaryEraOnwards era -> Parser MultiAsset

Parse a MintValue from its string representation. The string representation cannot contain ADA.

parseTxOutMultiAssetValue :: Parser Value

Parse a Value from its string representation. The resulting amounts must be positive for the parser to succeed.

Ada / L.Coin specifically

type Lovelace = Coin

A Coin is a Lovelace.

quantityToLovelace :: Quantity -> Lovelace

lovelaceToQuantity :: Lovelace -> Quantity

selectLovelace :: Value -> Lovelace

lovelaceToValue :: Lovelace -> Value

valueToLovelace :: Value -> Maybe Lovelace

Check if the Value consists of only Lovelace and no other assets, and if so then return the Lovelace

See also selectLovelace to select the Lovelace quantity from the Value, ignoring other assets.

Alternative nested representation

newtype ValueNestedRep

An alternative nested representation for Value that groups assets that share a PolicyId.

Constructors

ValueNestedRep [ValueNestedBundle]

data ValueNestedBundle

A bundle within a ValueNestedRep for a single PolicyId, or for the special case of ada.

Constructors

ValueNestedBundleAda Quantity

ValueNestedBundle PolicyId (Map AssetName Quantity)

valueToNestedRep :: Value -> ValueNestedRep

valueFromNestedRep :: ValueNestedRep -> Value

Rendering

renderValue :: Value -> Text

Render a textual representation of a Value.

renderValuePretty :: Value -> Text

Render a "prettified" textual representation of a Value.

renderMultiAsset :: MultiAsset -> Text

renderMultiAssetPretty :: MultiAsset -> Text

Internal conversion functions

toByronLovelace :: Lovelace -> Maybe Lovelace

fromByronLovelace :: Lovelace -> Lovelace

fromShelleyDeltaLovelace :: DeltaCoin -> Lovelace

toMaryValue :: Value -> MaryValue

fromMaryPolicyID :: PolicyID -> PolicyId

fromMaryValue :: MaryValue -> Value

fromMultiAsset :: MultiAsset -> Value

fromLedgerValue :: ShelleyBasedEra era -> Value (ShelleyLedgerEra era) -> Value

toLedgerValue :: MaryEraOnwards era -> Value -> Value (ShelleyLedgerEra era)

Data family instances

data family AsType t

A family of singleton types used in this API to indicate which type to use where it would otherwise be ambiguous or merely unclear.

Values of this type are passed to deserialisation functions for example.