Cardano.Api.ValueParser

Documentation

parseValue :: ParserValueRole -> Parser Value

Parse a Value from its string representation. The role argument for which purpose the value is being parsed. This is used to enforce additional constraints on the value. Why do we parse a general value and check for additional constraints you may ask? Because we can't rule out the negation operator for transaction outputs: some users have negative values in additions, with the addition's total summing up to a positive value. So forbidding negations altogether is too restrictive.

parseTxOutMultiAssetValue :: Parser Value

parseMintingMultiAssetValue :: Parser Value

parseUTxOValue :: Parser Value

parseAssetName :: Parser AssetName

Asset name parser.

parsePolicyId :: Parser PolicyId

Policy ID parser.

data ParserValueRole

The role for which a Value is being parsed.

Constructors

RoleUTxO
The value is used as a UTxO or transaction output.

RoleMint
The value is used as a minting policy.