Cardano.Rpc.Server.Internal.UtxoRpc.Type.Rational

Documentation

utxoRpcRationalNumberToRational :: Proto RationalNumber -> Maybe Rational

Convert a protobuf RationalNumber into a Rational.

Fails with Nothing when the denominator is zero, which is also the default value of the protobuf field when it is omitted.