Cardano.Rpc.Server.Internal.UtxoRpc.Type.Script

Documentation

simpleScriptToUtxoRpcNativeScript :: SimpleScript -> Proto NativeScript

TODO: remove together with the old-API TxOutput pipeline, https://github.com/IntersectMBO/cardano-api/issues/1264

utxoRpcNativeScriptToSimpleScript :: (HasCallStack, MonadThrow m) => Proto NativeScript -> m SimpleScript

referenceScriptToUtxoRpcScript :: ReferenceScript era -> Proto Script

scriptToUtxoRpcScript :: Script lang -> Proto Script

Convert a script to the UTxO RPC Script message, dispatching on the script language to select the corresponding oneof field.

TODO: remove together with the old-API TxOutput pipeline, https://github.com/IntersectMBO/cardano-api/issues/1264

utxoRpcScriptToReferenceScript :: forall era m. (HasCallStack, MonadThrow m, IsEra era) => Proto Script -> m (ReferenceScript era)

ledgerScriptToUtxoRpcScript :: ShelleyBasedEra era -> Script (ShelleyLedgerEra era) -> Proto Script

Convert a ledger script to the UTxO RPC Script message, dispatching native scripts to the NativeScript oneof field and plutus scripts to the field of their language version.

ledgerNativeScriptToUtxoRpcNativeScript :: ShelleyBasedEra era -> NativeScript (ShelleyLedgerEra era) -> Proto NativeScript

Convert a ledger native script directly to the UTxO RPC NativeScript message.

Each group of eras has its own concrete native script type: MultiSig at Shelley, Timelock from Allegra up to Conway (all sharing the same six view patterns), and Dijkstra's own type, which additionally supports a guard script.