Cardano.Wasm.Api.Certificate.StakeCertificate

Documentation

makeStakeAddressStakeDelegationCertificateImpl :: MonadThrow m => StakeKeyHashBase16 -> PoolIdBase16 -> m CertificateCBORBase16

Make a certificate that delegates a stake address to a stake pool in the current era.

makeStakeAddressStakeDelegationCertificateUpcomingEraImpl :: MonadThrow m => StakeKeyHashBase16 -> PoolIdBase16 -> m CertificateCBORBase16

Make a certificate that delegates a stake address to a stake pool in the current upcoming era.

makeStakeAddressRegistrationCertificateImpl :: MonadThrow m => StakeKeyHashBase16 -> DepositLovelace -> m CertificateCBORBase16

Make a stake address registration certificate in the current era.

makeStakeAddressRegistrationCertificateUpcomingEraImpl :: MonadThrow m => StakeKeyHashBase16 -> DepositLovelace -> m CertificateCBORBase16

makeStakeAddressUnregistrationCertificateImpl :: MonadThrow m => StakeKeyHashBase16 -> DepositLovelace -> m CertificateCBORBase16

Make a stake address unregistration certificate in the current era.

makeStakeAddressUnregistrationCertificateUpcomingEraImpl :: MonadThrow m => StakeKeyHashBase16 -> DepositLovelace -> m CertificateCBORBase16

Make a stake address unregistration certificate in the upcoming era.