Cardano.Wasm.Api.GRPC

Documentation

newtype GrpcObject grpcClient

Internal data for the GrpcConnection virtual object. Currently, it is just a wrapper around the grpcClient, which is some object that allows us to interact with the Cardano Node via gRPC or GRPC-web.

Constructors

GrpcObject grpcClient

newGrpcConnectionImpl :: (String -> IO grpcClient) -> String -> IO (GrpcObject grpcClient)

Create a new unsigned transaction object for making a Conway era transaction.

getEraImpl :: (grpcClient -> IO Int) -> GrpcObject grpcClient -> IO Int

Get the era from the Cardano Node using gRPC or GRPC-web.

getProtocolParamsImpl :: (grpcClient -> IO ProtocolParamsJSON) -> GrpcObject grpcClient -> IO ProtocolParamsJSON

Get the protocol parameters from the Cardano Node using gRPC or GRPC-web.

getAllUtxosImpl :: (grpcClient -> IO utxos) -> GrpcObject grpcClient -> IO utxos

Get all UTXOs from the node using a gRPC or GRPC-web client.

getUtxosForAddressImpl :: (grpcClient -> String -> IO utxos) -> GrpcObject grpcClient -> String -> IO utxos

Get UTXOs for a given address using a gRPC or GRPC-web client.

submitTxImpl :: (grpcClient -> String -> IO (Either String String)) -> GrpcObject grpcClient -> String -> IO String

Submit a transaction to the Cardano Node using gRPC or GRPC-web.