Cardano.Wasm.Api.Info

Documentation

apiInfo :: ApiInfo

Provides metadata about the "virtual objects" and their methods. This is intended to help generate JavaScript wrappers.

data ApiInfo

Aggregate type for all API information.

Constructors

ApiInfo

Fields

mainObject :: VirtualObjectInfo

Information about the main virtual object of the API, which serves as the entry point.

virtualObjects :: [VirtualObjectInfo]

Info about all other (non-main) virtual objects and their methods.

initialiseFunctionDoc :: String

Documentation for the initialise function of the API (which creates the main virtual object).

initialiseFunctionReturnDoc :: String

Documentation for the return value of the initialise function (which is the main virtual object).

data VirtualObjectInfo

Information about a virtual object and its methods.

Constructors

VirtualObjectInfo

Fields

virtualObjectName :: String

Name of the virtual object in the JS API (which should match the exported class and be PascalCase).

virtualObjectDoc :: String

Documentation for the virtual object.

virtualObjectMethods :: [MethodHierarchy]

Information about the methods of the virtual object.

data MethodHierarchy

Method hierarchy. Allows grouping methods in sub-groups.

Constructors

MethodGroupEntry MethodGroup

MethodInfoEntry MethodInfo

data MethodGroup

Method group. groups several methods under a common name.

Constructors

MethodGroup

Fields

groupName :: String

Name of the method group.

groupDoc :: [String]

Documentation for the method group.

groupMethods :: [MethodHierarchy]

Methods or sub-groups belonging to this group.

data MethodInfo

Information about a single method of a virtual object.

Constructors

MethodInfo

Fields

methodName :: String

Name of the method in the virtual object of the JS API (which should match the exported function).

methodDoc :: String

General documentation for the method.

methodParams :: [ParamInfo]

Info about parameters, excluding this.

methodReturnType :: MethodReturnTypeInfo

Return type of the method.

methodReturnDoc :: String

Documentation for the return value of the method.

data ParamInfo

Information about a single parameter of a method.

Constructors

ParamInfo

Fields

paramName :: String

Name of the parameter.

paramType :: TSType

Type of the parameter (as a TypeScript type).

paramDoc :: String

Documentation for the parameter.

data MethodReturnTypeInfo

Describes the return type of a method.

Constructors

Fluent
Returns an instance of the same object type (fluent interface).

NewObject String
Returns a new instance of a specified virtual object type.

OtherType TSType
Returns a non-virtual-object type (e.g., JSString, number).

dashCaseName :: VirtualObjectInfo -> String

tsTypeAsString :: TSType -> String