Cardano.Wasm.Api.Tx

Documentation

data UnsignedTxObject

An object representing a transaction that is being built and has not been signed yet. It abstracts over the era of the transaction. It is meant to be an opaque object in the JavaScript API.

Constructors

UnsignedTxObject

Fields

unsignedTxEra :: Era era

unsignedTx :: UnsignedTx era

data SignedTxObject

An object representing a signed transaction.

Constructors

SignedTxObject (Era era) (SignedTx era)

newtype ProtocolParamsJSON

Constructors

ProtocolParamsJSON String

newTxImpl :: UnsignedTxObject

Create a new unsigned transaction object for making a transaction in the current era.

newExperimentalEraTxImpl :: (HasCallStack, MonadThrow m) => m UnsignedTxObject

Create a new unsigned transaction object for making a transaction in the current experimental era.

newConwayTxImpl :: UnsignedTxObject

Create a new unsigned transaction object for making a Conway era transaction.

addTxInputImpl :: UnsignedTxObject -> TxId -> TxIx -> UnsignedTxObject

Add a simple transaction input to an unsigned transaction object.

addSimpleTxOutImpl :: (HasCallStack, MonadThrow m) => UnsignedTxObject -> String -> Coin -> m UnsignedTxObject

Add a simple transaction output to an unsigned transaction object. It takes a destination address and an amount in lovelaces.

appendCertificateToTxImpl :: (HasCallStack, MonadThrow m) => UnsignedTxObject -> String -> m UnsignedTxObject

Append a certificate (in CBOR hex string format) to an unsigned transaction object.

estimateMinFeeImpl

Arguments

:: (HasCallStack, MonadThrow m)

=> UnsignedTxObject
The unsigned transaction object to estimate fees for.

-> ProtocolParamsJSON
The JSON for the protocol parameters of the correct era and network.

-> Int
The number of key witnesses still to be added to the transaction.

-> Int
The number of Byron key witnesses still to be added to the transaction.

-> Int
The total size in bytes of reference scripts

-> m Coin

Estimate min fees for an unsigned transaction object.

setFeeImpl :: UnsignedTxObject -> Coin -> UnsignedTxObject

Set the fee for an unsigned transaction object.

signWithPaymentKeyImpl :: UnsignedTxObject -> SigningKey PaymentKey -> SignedTxObject

Sign an unsigned transaction using a payment key.

alsoSignWithPaymentKeyImpl :: SignedTxObject -> SigningKey PaymentKey -> SignedTxObject

Add an extra signature to an already signed transaction using a payment key.

toCborImpl :: SignedTxObject -> String

Convert a signed transaction object to a base16 encoded string of its CBOR representation.