Cardano.Wasm.Api.TypeScriptDefs

Description

This module defines a basic AST for generating TypeScript declaration files, and basic pretty-printing functionality.

The reason we define a custom tool for generating TypeScript declaration files is that existing libraries like `aeson-typescript` and `servant-typescript` are not aimed at generating custom TypeScript interface declaration files for a specific API, but rather at generating TypeScript interfaces for Haskell data types and servant HTTP APIs respectively. And other libraries that align with our needs, like `language-typescript`, are not actively maintained.

Documentation

writeTypeScriptToDir :: FilePath -> TypeScriptFile -> IO ()

Output the TypeScript declaration files to the specified directory.

buildMultilineComment :: Int -> [String] -> Builder

Creates a builder for a JavaScript-style multiline comment with the specified indentation level (in spaces) and each line in the list of strings as a separate line in the comment. The first line starts with /**, subsequent lines (corresponding to each line in the list) start with ` * `, and the last line ends with ` */`. The indentation level is used to indent the entire comment block.

data TypeScriptFile

Represents the top-level structure of a TypeScript declaration file.

Constructors

TypeScriptFile

Fields

typeScriptFileName :: String

Name of the TypeScript file.

typeScriptFileContent :: [Declaration]

List of declarations in the file.

buildTypeScriptFile :: TypeScriptFile -> Builder

Creates a builder for a TypeScript declaration file. It adds a comment to the top of the file with the file name.

data Declaration

Wraps a TypeScript declaration with a comment. The TypeScript declaration can have any of the types defined by DeclarationType.

Constructors

Declaration

Fields

declarationComment :: [String]

Comments for the declaration, can be empty if no comments are needed. Each element in the list is a separate line in the comment.

declarationContent :: DeclarationType

The type and content of the declaration.

buildDeclaration :: Declaration -> Builder

Creates a builder for a TypeScript declaration.

data DeclarationType

Represents a TypeScript declaration content of some type.

Constructors

ExportDec
Export declaration.

Fields

Bool

Is it a default export?

String

Name of the symbol to export.

FunctionDec FunctionHeader
Function declaration.

InterfaceDec
Interface declaration.

Fields

String

Name of the interface.

[GroupedInterfaceContent]

Definitions of the interface.

ImportDec
Reference to import another TypeScript declaration file.

Fields

String

Name of the symbol to import.

String

Path to the TypeScript declaration file to import.

buildDeclarationType :: DeclarationType -> Builder

Creates a builder for a TypeScript declaration type and content.

data GroupedInterfaceContent

Constructors

GroupedInterfaceContent InterfaceContentGroup
A group of interface contents (potentially with subgroups).

SingleInterfaceContent InterfaceContent
A single interface content.

buildGroupedInterfaceContent :: Int -> GroupedInterfaceContent -> Builder

data InterfaceContentGroup

Constructors

InterfaceContentGroup

Fields

groupedInterfaceContentComment :: [String]

Comments for the grouped interface content.

groupedInterfaceContentName :: String

The name of the group.

groupedInterfaceContentValues :: [GroupedInterfaceContent]

The list of grouped interface contents in the group.

data FunctionParam

Represents a function parameter in TypeScript.

Constructors

FunctionParam

Fields

paramName :: String

Name of the parameter.

paramType :: String

Type of the parameter.

buildFunctionParam :: FunctionParam -> Builder

Creates a builder for a TypeScript function parameter.

data FunctionHeader

Represents a TypeScript function header.

Constructors

FunctionHeader

Fields

functionName :: String

Name of the function.

functionParams :: [FunctionParam]

List of parameters of the function.

functionReturnType :: String

Return type of the function.

buildFunctionHeader :: FunctionHeader -> Builder

Creates a builder for a TypeScript function header.

data InterfaceContent

Represents a TypeScript interface content of some type out of the ones defined by InterfaceContentType.

Constructors

InterfaceContent

Fields

interfaceContentComment :: [String]

Comments for the interface content.

interfaceContentValue :: InterfaceContentType

The type and content of the interface.

buildInterfaceContent :: Int -> InterfaceContent -> Builder

Creates a builder for a TypeScript interface content.

data InterfaceContentType

Represents a TypeScript interface type and content.

Constructors

InterfaceProperty
Defines a property in the interface.

Fields

String

Property name.

String

Property type.

InterfaceMethod
Defines a method in the interface.

Fields

String

Method name.

[FunctionParam]

Method parameters.

String

Return type of the method.

buildInterfaceContentType :: InterfaceContentType -> Builder

Creates a builder for a TypeScript interface content and type.

mconcatWith :: Builder -> [Builder] -> Builder

Concatenates a list of builders with a separator. If the list is empty, it returns an empty builder.