Cardano.Wasm.Api.Wallet

Documentation

data WalletObject

Constructors

PaymentWallet

Fields

paymentWalletNetworkId :: NetworkId

paymentWalletPaymentSigningKey :: SigningKey PaymentKey

StakeWallet

Fields

stakeWalletNetworkId :: NetworkId

stakeWalletPaymentSigningKey :: SigningKey PaymentKey

stakeWalletStakeSigningKey :: SigningKey StakeKey

generatePaymentWalletImpl :: IO WalletObject

Generate a simple payment wallet for mainnet.

generateStakeWalletImpl :: IO WalletObject

Generate a stake wallet for mainnet.

restorePaymentWalletFromSigningKeyBech32Impl :: String -> IO WalletObject

Restore a mainnet payment wallet from a Bech32 encoded signing key.

restoreStakeWalletFromSigningKeyBech32Impl :: String -> String -> IO WalletObject

Restore a mainnet stake wallet from Bech32 encoded signing keys.

generateTestnetPaymentWalletImpl :: Int -> IO WalletObject

Generate a simple payment wallet for testnet, given the testnet's network magic.

generateTestnetStakeWalletImpl :: Int -> IO WalletObject

Generate a stake wallet for testnet, given the testnet's network magic.

restoreTestnetPaymentWalletFromSigningKeyBech32Impl :: Int -> String -> IO WalletObject

Restore a testnet payment wallet from a Bech32 encoded signing key.

restoreTestnetStakeWalletFromSigningKeyBech32Impl :: Int -> String -> String -> IO WalletObject

Restore a testnet stake wallet from Bech32 encoded signing keys.

getAddressBech32Impl :: WalletObject -> String

Get the Bech32 representation of the address. (Can be shared for receiving funds.)

getBech32ForPaymentVerificationKeyImpl :: WalletObject -> String

Get the Bech32 representation of the verification key of the wallet. (Can be shared for verification.)

getBech32ForPaymentSigningKeyImpl :: WalletObject -> String

Get the Bech32 representation of the signing key of the wallet, if any. (Must be kept secret.)

getBech32ForStakeVerificationKeyImpl :: WalletObject -> String

Get the Bech32 representation of the stake verification key of the wallet. (Can be shared for verification.)

getBech32ForStakeSigningKeyImpl :: WalletObject -> String

Get the Bech32 representation of the stake signing key of the wallet, if any. (Must be kept secret.)

getBase16ForPaymentVerificationKeyHashImpl :: WalletObject -> String

Get the base16 representation of the hash of the verification key of the wallet.

getBase16ForStakeVerificationKeyHashImpl :: WalletObject -> String

Get the base16 representation of the hash of the stake verification key of the wallet.