Test.Gen.Cardano.Api.Typed

Documentation

genFeaturedInEra :: Alternative f => eon era -> f a -> f (Featured eon era a)

Generate a Featured for the given CardanoEra with the provided generator.

genMaybeFeaturedInEra :: (Eon eon, Alternative f) => (eon era -> f a) -> CardanoEra era -> f (Maybe (Featured eon era a))

Generate a Featured for the given CardanoEra with the provided generator.

Byron

genAddressInEraByron :: Gen (AddressInEra ByronEra)

genAddressByron :: Gen (Address ByronAddr)

genTxBodyByron :: HasCallStack => Gen (Annotated Tx ByteString)

Partial! It will throw if the generated transaction body is invalid.

genTxByron :: Gen (ATxAux ByteString)

genTxOutByron :: Gen (TxOut CtxTx ByronEra)

genWitnessesByron :: Gen [KeyWitness ByronEra]

genAddressInEra :: ShelleyBasedEra era -> Gen (AddressInEra era)

genAddressShelley :: Gen (Address ShelleyAddr)

genCertificate :: Typeable era => ShelleyBasedEra era -> Gen (Certificate era)

genCostModel :: MonadGen m => m CostModel

genCostModels :: MonadGen m => m CostModels

genMaybePraosNonce :: Gen (Maybe PraosNonce)

genPraosNonce :: Gen PraosNonce

genValidProtocolParameters :: ShelleyBasedEra era -> Gen (LedgerProtocolParameters era)

Generate valid protocol parameters which pass validations in Cardano.Api.ProtocolParameters

genValueNestedRep :: Gen ValueNestedRep

genValueNestedBundle :: Gen ValueNestedBundle

genByronKeyWitness :: Gen (KeyWitness ByronEra)

genCardanoKeyWitness :: Typeable era => ShelleyBasedEra era -> Gen (KeyWitness era)

genShelleyKeyWitness :: Typeable era => ShelleyBasedEra era -> Gen (KeyWitness era)

genTxId :: Gen TxId

genTxIn :: Gen TxIn

genTxOutTxContext :: ShelleyBasedEra era -> Gen (TxOut CtxTx era)

genTxOutUTxOContext :: ShelleyBasedEra era -> Gen (TxOut CtxUTxO era)

genUTxO :: ShelleyBasedEra era -> Gen (UTxO era)

Scripts

genHashableScriptData :: HasCallStack => Gen HashableScriptData

genReferenceScript :: ShelleyBasedEra era -> Gen (ReferenceScript era)

genScript :: ScriptLanguage lang -> Gen (Script lang)

genValidScript :: ScriptLanguage lang -> Gen (Script lang)

genSimpleScript :: Gen SimpleScript

genSimpleScriptMintWitness :: ShelleyBasedEra era -> Gen (Witness WitCtxMint era)

genPlutusScript :: PlutusScriptVersion lang -> Gen (PlutusScript lang)

genPlutusScript will generate the deprecated double encoded plutus scripts as well as valid plutus scripts.

genPlutusV1Script :: Gen (Script PlutusScriptV1)

genPlutusV2Script :: Gen (Script PlutusScriptV2)

genPlutusV3Script :: Gen (Script PlutusScriptV3)

genScriptInAnyLang :: Gen ScriptInAnyLang

genScriptInEra :: ShelleyBasedEra era -> Gen (ScriptInEra era)

genScriptHash :: Gen ScriptHash

genScriptData :: Gen ScriptData

Deprecated: Use genHashableScriptData

genScriptDataSchema :: Gen ScriptDataJsonSchema

genScriptValidity :: Gen ScriptValidity

genAssetName :: Gen AssetName

genAssetId :: Gen AssetId

genEpochNo :: Gen EpochNo

genExecutionUnitPrices :: Gen ExecutionUnitPrices

genExecutionUnits :: Gen ExecutionUnits

genHashScriptData :: Gen (Hash ScriptData)

genKESPeriod :: Gen KESPeriod

genNat :: Gen Natural

genNetworkId :: Gen NetworkId

genNetworkMagic :: Gen NetworkMagic

genOperationalCertificate :: Gen OperationalCertificate

genOperationalCertificateIssueCounter :: Gen OperationalCertificateIssueCounter

genOperationalCertificateWithCounter :: HasCallStack => Gen (OperationalCertificate, OperationalCertificateIssueCounter)

genPaymentCredential :: Gen PaymentCredential

genPolicyId :: Gen PolicyId

genQuantity :: Range Integer -> Gen Quantity

genRationalInt64 :: Gen Rational

genSeed :: Int -> Gen Seed

genShelleyBootstrapWitness :: Typeable era => ShelleyBasedEra era -> Gen (KeyWitness era)

genShelleyHash :: Gen (Hash Blake2b_256 EraIndependentTxBody)

genShelleyWitness :: Typeable era => ShelleyBasedEra era -> Gen (KeyWitness era)

genShelleyWitnessSigningKey :: Gen ShelleyWitnessSigningKey

genSignedQuantity :: Gen Quantity

Generate a positive or negative quantity.

genSignedNonZeroQuantity :: Gen Quantity

Generate a positive or negative, but not zero quantity.

genSigningKey :: Key keyrole => AsType keyrole -> Gen (SigningKey keyrole)

genSlotNo :: Gen SlotNo

genStakeAddress :: Gen StakeAddress

genStakeAddressReference :: Gen StakeAddressReference

genStakeCredential :: Gen StakeCredential

genTtl :: Gen SlotNo

genTx :: Typeable era => ShelleyBasedEra era -> Gen (Tx era)

genTxAuxScripts :: ShelleyBasedEra era -> Gen (TxAuxScripts era)

genTxBody :: (HasCallStack, Typeable era) => ShelleyBasedEra era -> Gen (TxBody era)

Partial! This function will throw an error when the generated transaction is invalid.

genTxBodyContent :: Typeable era => ShelleyBasedEra era -> Gen (TxBodyContent BuildTx era)

genValidTxBody

Arguments

:: Typeable era

=> ShelleyBasedEra era

-> Gen (TxBody era, TxBodyContent BuildTx era)
validated TxBody and TxBodyContent

This generator validates generated TxBodyContent and backtracks when the generated body fails the validation. That also means that it is quite slow.

genTxCertificates :: Typeable era => CardanoEra era -> Gen (TxCertificates BuildTx era)

genTxFee :: ShelleyBasedEra era -> Gen (TxFee era)

genTxIndex :: Gen TxIx

genTxInsCollateral :: CardanoEra era -> Gen (TxInsCollateral era)

genTxInsReference :: CardanoEra era -> Gen (TxInsReference era)

genTxMetadataInEra :: CardanoEra era -> Gen (TxMetadataInEra era)

genTxMintValue :: CardanoEra era -> Gen (TxMintValue BuildTx era)

genPolicyAssets :: Gen PolicyAssets

genLovelace :: Gen Coin

genPositiveLovelace :: Gen Coin

genValue :: Gen AssetId -> Gen Quantity -> Gen Value

genValueDefault :: MaryEraOnwards era -> Gen (Value (ShelleyLedgerEra era))

Generate a Value with any asset ID and a positive or negative quantity.

genVerificationKey :: (HasTypeProxy keyrole, Key keyrole) => AsType keyrole -> Gen (VerificationKey keyrole)

genVerificationKeyHash :: (HasTypeProxy keyrole, Key keyrole) => AsType keyrole -> Gen (Hash keyrole)

genUpdateProposal :: CardanoEra era -> Gen UpdateProposal

genProtocolParametersUpdate :: CardanoEra era -> Gen ProtocolParametersUpdate

genTxOutDatumHashTxContext :: ShelleyBasedEra era -> Gen (TxOutDatum CtxTx era)

genTxOutDatumHashUTxOContext :: ShelleyBasedEra era -> Gen (TxOutDatum CtxUTxO era)

genTxOutValue :: ShelleyBasedEra era -> Gen (TxOutValue era)

genTxReturnCollateral :: ShelleyBasedEra era -> Gen (TxReturnCollateral CtxTx era)

genTxScriptValidity :: CardanoEra era -> Gen (TxScriptValidity era)

genTxTotalCollateral :: CardanoEra era -> Gen (TxTotalCollateral era)

genTxUpdateProposal :: CardanoEra era -> Gen (TxUpdateProposal era)

genTxValidityLowerBound :: CardanoEra era -> Gen (TxValidityLowerBound era)

genTxValidityUpperBound :: ShelleyBasedEra era -> Gen (TxValidityUpperBound era)

genTxWithdrawals :: CardanoEra era -> Gen (TxWithdrawals build era)

genUnsignedQuantity :: Gen Quantity

genPositiveQuantity :: Gen Quantity

genValueForMinting :: Gen Value

Generate a Value suitable for minting, i.e. non-ADA asset ID and a positive or negative quantity.

genValueForTxOut :: ShelleyBasedEra era -> Gen Value

Generate a Value suitable for use in a transaction output, with non-zero ADA, any asset IDs and with positive quantities

genLedgerValueForTxOut :: ShelleyBasedEra era -> Gen (Value (ShelleyLedgerEra era))

Generate a Value suitable for usage in a transaction output, i.e. any asset ID and a positive quantity.

genLedgerMultiAssetValue :: Gen MultiAsset

genWitnesses :: Typeable era => ShelleyBasedEra era -> Gen [KeyWitness era]

genWitnessNetworkIdOrByronAddress :: Gen WitnessNetworkIdOrByronAddress

genRational :: Gen Rational

genGovernancePoll :: Gen GovernancePoll

genGovernancePollAnswer :: Gen GovernancePollAnswer

genProposals :: Applicative (BuildTxWith build) => ConwayEraOnwards era -> Gen (TxProposalProcedures build era)

genProposal :: ConwayEraOnwards era -> Gen (ProposalProcedure (ShelleyLedgerEra era))

genVotingProcedures :: Applicative (BuildTxWith build) => ConwayEraOnwards era -> Gen (TxVotingProcedures build era)

genSimpleScriptWithoutEmptyAnys :: Gen SimpleScript

genWitnessable :: AlonzoEraScript era => Gen (Witnessable 'TxInItem era)

genMintWitnessable :: AlonzoEraScript era => Gen (Witnessable 'MintItem era)

genPlutusScriptWitness :: forall (purpose :: PlutusScriptPurpose) era. Gen (PlutusScriptWitness 'PlutusV3 purpose era)

genIndexedPlutusScriptWitness :: AlonzoEraScript (ShelleyLedgerEra era) => Gen (IndexedPlutusScriptWitness 'TxInItem 'PlutusV3 'SpendingScript (ShelleyLedgerEra era))