Cardano.Api.Address

Description

Cardano addresses: payment and stake addresses.

Payment addresses

Constructing and inspecting normal payment addresses

data Address addrtype where

Addresses are used as locations where assets live. The address determines the rights needed to spend assets at the address: in particular holding some signing key or being able to satisfy the conditions of a script.

There are currently two types of address:

Byron addresses, which use the type tag ByronAddr; and

Shelley addresses, which use the type tag ShelleyAddr. Notably, Shelley addresses support scripts and stake delegation.

The address type is subtly from the ledger era in which each address type is valid: while Byron addresses are the only choice in the Byron era, the Shelley era and all subsequent eras support both Byron and Shelley addresses. The Address type param only says the type of the address (either Byron or Shelley). The AddressInEra type connects the address type with the era in which it is supported.

Constructors

ByronAddress ∷ Address → Address ByronAddr
Byron addresses were the only supported address type in the original Byron era.

ShelleyAddress ∷ Network → PaymentCredential StandardCrypto → StakeReference StandardCrypto → Address ShelleyAddr
Shelley addresses allow delegation. Shelley addresses were introduced in Shelley era and are thus supported from the Shelley era onwards

Byron addresses

data ByronAddr

A type used as a tag to distinguish Byron addresses.

makeByronAddress ∷ NetworkId → VerificationKey ByronKey → Address ByronAddr

Shelley addresses

data ShelleyAddr

A type used as a tag to distinguish Shelley addresses.

makeShelleyAddress ∷ NetworkId → PaymentCredential → StakeAddressReference → Address ShelleyAddr

data PaymentCredential

Constructors

PaymentCredentialByKey (Hash PaymentKey)

PaymentCredentialByScript ScriptHash

data StakeAddressReference

Constructors

StakeAddressByValue StakeCredential

StakeAddressByPointer StakeAddressPointer

NoStakeAddress

Addresses in any era

data AddressAny

Either a Byron address or a Shelley address.

Sometimes we need to be able to work with either of the two types of address (Byron or Shelley addresses), but without reference to an era in which the address will be used. This type serves that purpose.

Constructors

AddressByron !(Address ByronAddr)

AddressShelley !(Address ShelleyAddr)

Addresses in specific eras

data AddressInEra era where

An Address that can be used in a particular ledger era.

All current ledger eras support Byron addresses. Shelley addresses are supported in the ShelleyEra and later eras.

Constructors

AddressInEra ∷ AddressTypeInEra addrtype era → Address addrtype → AddressInEra era

data AddressTypeInEra addrtype era where

Constructors

ByronAddressInAnyEra ∷ AddressTypeInEra ByronAddr era

ShelleyAddressInEra ∷ ShelleyBasedEra era → AddressTypeInEra ShelleyAddr era

byronAddressInEra ∷ Address ByronAddr → AddressInEra era

shelleyAddressInEra ∷ IsShelleyBasedEra era ⇒ Address ShelleyAddr → AddressInEra era

anyAddressInShelleyBasedEra ∷ IsShelleyBasedEra era ⇒ AddressAny → AddressInEra era

anyAddressInEra ∷ CardanoEra era → AddressAny → Maybe (AddressInEra era)

toAddressAny ∷ Address addr → AddressAny

makeByronAddressInEra ∷ NetworkId → VerificationKey ByronKey → AddressInEra era

makeShelleyAddressInEra ∷ IsShelleyBasedEra era ⇒ NetworkId → PaymentCredential → StakeAddressReference → AddressInEra era

Stake addresses

Constructing and inspecting stake addresses

data StakeAddress where

Constructors

StakeAddress ∷ Network → StakeCredential StandardCrypto → StakeAddress

data StakeCredential

Constructors

StakeCredentialByKey (Hash StakeKey)

StakeCredentialByScript ScriptHash

makeStakeAddress ∷ NetworkId → StakeCredential → StakeAddress

data StakeKey

data StakeExtendedKey

Shelley-era stake keys using extended ed25519 cryptographic keys.

They can be used for Shelley stake addresses and witnessing transactions that use stake addresses.

These extended keys are used by HD wallets. So this type provides interoperability with HD wallets. The ITN CLI also supported this key type.

The extended verification keys can be converted (via castVerificationKey) to ordinary keys (i.e. VerificationKey StakeKey) but this is not the case for the signing keys. The signing keys can be used to witness transactions directly, with verification via their non-extended verification key (VerificationKey StakeKey).

This is a type level tag, used with other interfaces like Key.

Internal conversion functions

toShelleyAddr ∷ AddressInEra era → Addr StandardCrypto

toShelleyStakeAddr ∷ StakeAddress → RewardAcnt StandardCrypto

toShelleyStakeCredential ∷ StakeCredential → StakeCredential StandardCrypto

fromShelleyAddr ∷ IsShelleyBasedEra era ⇒ Addr StandardCrypto → AddressInEra era

fromShelleyPaymentCredential ∷ PaymentCredential StandardCrypto → PaymentCredential

fromShelleyStakeAddr ∷ RewardAcnt StandardCrypto → StakeAddress

fromShelleyStakeCredential ∷ StakeCredential StandardCrypto → StakeCredential

fromShelleyStakeReference ∷ StakeReference StandardCrypto → StakeAddressReference

Serialising addresses

class HasTypeProxy addr ⇒ SerialiseAddress addr where

Address serialisation uses different serialisation formats for different kinds of addresses, so it needs its own class.

In particular, Byron addresses are typically formatted in base 58, while Shelley addresses (payment and stake) are formatted using Bech32.

Methods

serialiseAddress ∷ addr → Text

deserialiseAddress ∷ AsType addr → Text → Maybe addr

Data family instances

data family AsType t

A family of singleton types used in this API to indicate which type to use where it would otherwise be ambiguous or merely unclear.

Values of this type are passed to deserialisation functions for example.