Cardano.Api.Anchor

Documentation

newtype AnchorUrl

The URL to build the anchor to pass to DRep registration certificate

Constructors

AnchorUrl

Fields

unAnchorUrl ∷ Url

newtype AnchorDataHash

The hash to build the anchor to pass to DRep registration certificate

Constructors

AnchorDataHash

Fields

unAnchorDataHash ∷ SafeHash StandardCrypto AnchorData