Cardano.Api.Block

Description

Blocks in the blockchain

Blocks in the context of an era

data Block era where

A blockchain block in a particular Cardano era.

Constructors

ByronBlock ∷ ByronBlock → Block ByronEra

ShelleyBlock ∷ ShelleyBasedEra era → ShelleyBlock (ConsensusProtocol era) (ShelleyLedgerEra era) → Block era

Bundled Patterns

pattern Block ∷ BlockHeader → [Tx era] → Block era
A block consists of a header and a body containing transactions.

data BlockHeader

Constructors

BlockHeader !SlotNo !(Hash BlockHeader) !BlockNo

getBlockHeader ∷ ∀ era. Block era → BlockHeader

Blocks in the context of a consensus mode

data BlockInMode mode where

A Block in one of the eras supported by a given protocol mode.

For multi-era modes such as the CardanoMode this type is a sum of the different block types for all the eras. It is used in the ChainSync protocol.

Constructors

BlockInMode ∷ IsCardanoEra era ⇒ Block era → EraInMode era mode → BlockInMode mode

fromConsensusBlock ∷ ConsensusBlockForMode mode ~ block ⇒ LedgerSupportsProtocol (ShelleyBlock (TPraos StandardCrypto) (ShelleyEra StandardCrypto)) ⇒ ConsensusMode mode → block → BlockInMode mode

toConsensusBlock ∷ ConsensusBlockForMode mode ~ block ⇒ LedgerSupportsProtocol (ShelleyBlock (TPraos StandardCrypto) (ShelleyEra StandardCrypto)) ⇒ BlockInMode mode → block

Points on the chain

data ChainPoint

Constructors

ChainPointAtGenesis

ChainPoint !SlotNo !(Hash BlockHeader)

newtype SlotNo

Constructors

SlotNo

Fields

unSlotNo ∷ Word64

newtype EpochNo

Constructors

EpochNo

Fields

unEpochNo ∷ Word64

toConsensusPoint ∷ ∀ ledgerera protocol. ShelleyCompatible protocol ledgerera ⇒ ChainPoint → Point (ShelleyBlock protocol ledgerera)

Convert a Point for single Shelley-era block type

fromConsensusPoint ∷ ∀ protocol ledgerera. ShelleyCompatible protocol ledgerera ⇒ Point (ShelleyBlock protocol ledgerera) → ChainPoint

Convert a Point for single Shelley-era block type

toConsensusPointInMode ∷ ConsensusMode mode → ChainPoint → Point (ConsensusBlockForMode mode)

fromConsensusPointInMode ∷ ConsensusMode mode → Point (ConsensusBlockForMode mode) → ChainPoint

toConsensusPointHF ∷ HeaderHash block ~ OneEraHash xs ⇒ ChainPoint → Point block

Convert a Point for multi-era block type

Tip of the chain

data ChainTip

This is like a ChainPoint but is conventionally used for the tip of the chain: that is the most recent block at the end of the chain.

It also carries the BlockNo of the chain tip.

Constructors

ChainTipAtGenesis

ChainTip !SlotNo !(Hash BlockHeader) !BlockNo

newtype BlockNo

Constructors

BlockNo

Fields

unBlockNo ∷ Word64

chainTipToChainPoint ∷ ChainTip → ChainPoint

fromConsensusTip ∷ ConsensusBlockForMode mode ~ block ⇒ ConsensusMode mode → Tip block → ChainTip

Data family instances

data family Hash keyrole ∷ Type

chainPointToHeaderHash ∷ ChainPoint → Maybe (Hash BlockHeader)

chainPointToSlotNo ∷ ChainPoint → Maybe SlotNo

makeChainTip ∷ WithOrigin BlockNo → ChainPoint → ChainTip