Cardano.Api.Block

Description

Blocks in the blockchain

Blocks in the context of an era

data Block era where

A blockchain block in a particular Cardano era.

Constructors

ByronBlock ∷ ByronBlock → Block ByronEra

ShelleyBlock ∷ ShelleyBasedEra era → ShelleyBlock (ShelleyLedgerEra era) → Block era

data BlockHeader

Blocks in the context of a consensus mode

data BlockInMode mode where

A Block in one of the eras supported by a given protocol mode.

For multi-era modes such as the CardanoMode this type is a sum of the different block types for all the eras. It is used in the ChainSync protocol.

Constructors

BlockInMode ∷ Block era → EraInMode era mode → BlockInMode mode

fromConsensusBlock ∷ ConsensusBlockForMode mode ~ block ⇒ ConsensusMode mode → block → BlockInMode mode

Points on the chain

data ChainPoint

Constructors

ChainPointAtGenesis

ChainPoint !SlotNo !(Hash BlockHeader)

data SlotNo

data EpochNo

toConsensusPoint ∷ ∀ ledgerera. ShelleyBasedEra ledgerera ⇒ ChainPoint → Point (ShelleyBlock ledgerera)

Convert a Point for single Shelley-era block type

fromConsensusPoint ∷ ∀ ledgerera. ShelleyBasedEra ledgerera ⇒ Point (ShelleyBlock ledgerera) → ChainPoint

Convert a Point for single Shelley-era block type

toConsensusPointInMode ∷ ConsensusMode mode → ChainPoint → Point (ConsensusBlockForMode mode)

fromConsensusPointInMode ∷ ConsensusMode mode → Point (ConsensusBlockForMode mode) → ChainPoint

Tip of the chain

data ChainTip

This is like a ChainPoint but is conventionally used for the tip of the chain: that is the most recent block at the end of the chain.

It also carries the BlockNo of the chain tip.

Constructors

ChainTipAtGenesis

ChainTip !SlotNo !(Hash BlockHeader) !BlockNo

data BlockNo

chainTipToChainPoint ∷ ChainTip → ChainPoint

fromConsensusTip ∷ ConsensusBlockForMode mode ~ block ⇒ ConsensusMode mode → Tip block → ChainTip

Data family instances

data family Hash keyrole ∷ Type