Cardano.Api.Byron

Description

This module provides a library interface that is intended to be the complete API for Byron covering everything, including exposing constructors for the lower level types.

Documentation

module Cardano.Api

data family AsType t

A family of singleton types used in this API to indicate which type to use where it would otherwise be ambiguous or merely unclear.

Values of this type are passed to deserialisation functions for example.

Cryptographic key interface

data family VerificationKey keyrole

The type of cryptographic verification key, for each key role.

data family SigningKey keyrole

The type of cryptographic signing key, for each key role.

data SomeByronSigningKey

Constructors

AByronSigningKeyLegacy (SigningKey ByronKeyLegacy)

AByronSigningKey (SigningKey ByronKey)

Hashes

data family Hash keyrole

Payment addresses

Constructing and inspecting Byron payment addresses

data Address addrtype where

Addresses are used as locations where assets live. The address determines the rights needed to spend assets at the address: in particular holding some signing key or being able to satisfy the conditions of a script.

There are currently two types of address:

Byron addresses, which use the type tag ByronAddr; and

Shelley addresses, which use the type tag ShelleyAddr. Notably, Shelley addresses support scripts and stake delegation.

The address type is subtly from the ledger era in which each address type is valid: while Byron addresses are the only choice in the Byron era, the Shelley era and all subsequent eras support both Byron and Shelley addresses. The Address type param only says the type of the address (either Byron or Shelley). The AddressInEra type connects the address type with the era in which it is supported.

Constructors

ByronAddress ∷ Address → Address ByronAddr
Byron addresses were the only supported address type in the original Byron era.

data NetworkId

Constructors

Mainnet

Testnet !NetworkMagic

Building transactions

Constructing and inspecting transactions

data TxBody era where

Constructors

ByronTxBody ∷ Annotated Tx ByteString → TxBody ByronEra

newtype TxId

Constructors

TxId (Hash StandardCrypto EraIndependentTxBody)

data TxIn

Constructors

TxIn TxId TxIx

data TxOut ctx era

Constructors

TxOut (AddressInEra era) (TxOutValue era) (TxOutDatum ctx era) (ReferenceScript era)

newtype TxIx

Constructors

TxIx Word

newtype Lovelace

Constructors

Lovelace Integer

Signing transactions

Creating transaction witnesses one by one, or all in one go.

data Tx era where

Constructors

ByronTx ∷ ATxAux ByteString → Tx ByronEra

Incremental signing and separate witnesses

data KeyWitness era where

Constructors

ByronKeyWitness ∷ TxInWitness → KeyWitness ByronEra

data WitnessNetworkIdOrByronAddress

Either a network ID or a Byron address to be used in constructing a Shelley bootstrap witness.

Constructors

WitnessNetworkId !NetworkId
Network ID.

If this value is used in the construction of a Shelley bootstrap witness, the result will not consist of a derivation path. If that is required, specify a WitnessByronAddress value instead.

WitnessByronAddress !(Address ByronAddr)
Byron address.

If this value is used in the construction of a Shelley bootstrap witness, both the network ID and derivation path will be extracted from the address and used in the construction of the witness.

Errors

class Show e ⇒ Error e where

Methods

displayError ∷ e → String

data FileError e

Constructors

FileError FilePath e

FileErrorTempFile

Fields

FilePath

Target path

FilePath

Temporary path

Handle

FileDoesNotExistError FilePath

FileIOError FilePath IOException

Low level protocol interaction with a Cardano node

data LocalNodeConnectInfo mode

Constructors

LocalNodeConnectInfo (ConsensusModeParams mode) NetworkId SocketPath

data ByronMode

The Byron-only consensus mode consists of only the Byron era.

This was used on the mainnet before the deployment of the multi-era CardanoMode. It is now of little practical use, though it illustrates how a single-era consensus mode works. It may be sensible to remove this at some stage.

data ConsensusMode mode where

This GADT provides a value-level representation of all the consensus modes. This enables pattern matching on the era to allow them to be treated in a non-uniform way.

Constructors

ByronMode ∷ ConsensusMode ByronMode

data LocalNodeClientProtocols block point tip slot tx txid txerr (query ∷ Type → Type) (m ∷ Type → TYPE LiftedRep)

The protocols we can use with a local node. Use in conjunction with connectToLocalNode.

These protocols use the types from the rest of this API. The conversion to/from the types used by the underlying wire formats is handled by connectToLocalNode.

Constructors

LocalNodeClientProtocols (LocalChainSyncClient block point tip m) (Maybe (LocalTxSubmissionClient tx txerr m ())) (Maybe (LocalStateQueryClient block point query m ())) (Maybe (LocalTxMonitorClient txid tx slot m ()))

Chain sync protocol

newtype ChainSyncClient header point tip (m ∷ Type → Type) a

A chain sync protocol client, on top of some effect m. The first choice of request is within that m.

Constructors

ChainSyncClient

Fields

runChainSyncClient ∷ m (ClientStIdle header point tip m a)

Local tx submission

newtype LocalTxSubmissionClient tx reject (m ∷ Type → Type) a

Constructors

LocalTxSubmissionClient (m (LocalTxClientStIdle tx reject m a))

Local state query

newtype LocalStateQueryClient block point (query ∷ Type → Type) (m ∷ Type → Type) a

Constructors

LocalStateQueryClient

Fields

runLocalStateQueryClient ∷ m (ClientStIdle block point query m a)

Address

newtype NetworkMagic

NetworkMagic is used to differentiate between different networks during the initial handshake.

Constructors

NetworkMagic

Fields

unNetworkMagic ∷ Word32

Update Proposal

newtype ByronUpdateProposal

Byron era update proposal

Constructors

ByronUpdateProposal

Fields

unByronUpdateProposal ∷ AProposal ByteString

data ByronProtocolParametersUpdate

Constructors

ByronProtocolParametersUpdate

Fields

bPpuScriptVersion ∷ !(Maybe Word16)

Redundant. This was meant to be the version of the Plutus smart contract language, however, there are no smart contracts nor scripts in the Byron era.

bPpuSlotDuration ∷ !(Maybe Natural)

Slot duration in milliseconds.

bPpuMaxBlockSize ∷ !(Maybe Natural)

Maximum block size in bytes.

bPpuMaxHeaderSize ∷ !(Maybe Natural)

Maximum block header size in bytes.

bPpuMaxTxSize ∷ !(Maybe Natural)

Maximum transaction size in bytes.

bPpuMaxProposalSize ∷ !(Maybe Natural)

Maximum update proposal size in bytes.

bPpuMpcThd ∷ !(Maybe LovelacePortion)

bPpuHeavyDelThd ∷ !(Maybe LovelacePortion)

Heavyweight delegation threshold. The delegate (i.e stakeholder) must possess no less than this threshold of stake in order to participate in heavyweight delegation.

bPpuUpdateVoteThd ∷ !(Maybe LovelacePortion)

bPpuUpdateProposalThd ∷ !(Maybe LovelacePortion)

bPpuUpdateProposalTTL ∷ !(Maybe SlotNumber)

bPpuSoftforkRule ∷ !(Maybe SoftforkRule)

Values defining the softfork resolution rule. When the stake belonging to block issuers, issuing a given block version, is greater than the current softfork resolution threshold, this block version is adopted.

bPpuTxFeePolicy ∷ !(Maybe TxFeePolicy)

Transaction fee policy represents a formula to compute the minimal allowed Fee for a transaction. Transactions with lesser fees won't be accepted.

bPpuUnlockStakeEpoch ∷ !(Maybe EpochNumber)

This has been re-purposed for unlocking the OuroborosBFT logic in the software. Relevant: CDEC-610

makeByronUpdateProposal ∷ NetworkId → ProtocolVersion → SoftwareVersion → SystemTag → InstallerHash → SomeByronSigningKey → ByronProtocolParametersUpdate → ByronUpdateProposal

toByronLedgerUpdateProposal ∷ ByronUpdateProposal → GenTx ByronBlock

makeProtocolParametersUpdate ∷ ByronProtocolParametersUpdate → ProtocolParametersUpdate

Vote

newtype ByronVote

Byron era votes

Constructors

ByronVote

Fields

unByronVote ∷ AVote ByteString

makeByronVote ∷ NetworkId → SomeByronSigningKey → ByronUpdateProposal → Bool → ByronVote

toByronLedgertoByronVote ∷ ByronVote → GenTx ByronBlock

Conversions

fromByronTxIn ∷ TxIn → TxIn

toByronLovelace ∷ Lovelace → Maybe Lovelace

toByronNetworkMagic ∷ NetworkId → NetworkMagic

toByronProtocolMagicId ∷ NetworkId → ProtocolMagicId

toByronRequiresNetworkMagic ∷ NetworkId → RequiresNetworkMagic

Hardcoded configuration parameters

applicationName ∷ ApplicationName

An application name. It has no functional impact in the Shelley eras onwards and therefore it is hardcoded.

applicationVersion ∷ NumSoftwareVersion

An application version. It has no functional impact in the Shelley eras onwards and therefore it is hardcoded.

softwareVersion ∷ SoftwareVersion

A software version composed of applicationVersion and applicationName. It has no functional impact in the Shelley eras onwards and therefore it is hardcoded.