Cardano.Api.Certificate

Description

Certificates embedded in transactions

Documentation

data Certificate era where

Constructors

ShelleyRelatedCertificate ∷ ShelleyToBabbageEra era → ShelleyTxCert (ShelleyLedgerEra era) → Certificate era

ConwayCertificate ∷ ConwayEraOnwards era → ConwayTxCert (ShelleyLedgerEra era) → Certificate era

Registering stake address and delegating

data StakeAddressRequirements era where

Constructors

StakeAddrRegistrationConway ∷ ConwayEraOnwards era → Lovelace → StakeCredential → StakeAddressRequirements era

StakeAddrRegistrationPreConway ∷ ShelleyToBabbageEra era → StakeCredential → StakeAddressRequirements era

data StakeDelegationRequirements era where

Constructors

StakeDelegationRequirementsConwayOnwards ∷ ConwayEraOnwards era → StakeCredential → Delegatee (EraCrypto (ShelleyLedgerEra era)) → StakeDelegationRequirements era

StakeDelegationRequirementsPreConway ∷ ShelleyToBabbageEra era → StakeCredential → PoolId → StakeDelegationRequirements era

makeStakeAddressDelegationCertificate ∷ StakeDelegationRequirements era → Certificate era

makeStakeAddressRegistrationCertificate ∷ StakeAddressRequirements era → Certificate era

makeStakeAddressUnregistrationCertificate ∷ StakeAddressRequirements era → Certificate era

type PoolId = Hash StakePoolKey

Registering stake pools

data StakePoolRegistrationRequirements era where

Constructors

StakePoolRegistrationRequirementsConwayOnwards ∷ ConwayEraOnwards era → PoolParams (EraCrypto (ShelleyLedgerEra era)) → StakePoolRegistrationRequirements era

StakePoolRegistrationRequirementsPreConway ∷ ShelleyToBabbageEra era → PoolParams (EraCrypto (ShelleyLedgerEra era)) → StakePoolRegistrationRequirements era

data StakePoolRetirementRequirements era where

Constructors

StakePoolRetirementRequirementsConwayOnwards ∷ ConwayEraOnwards era → PoolId → EpochNo → StakePoolRetirementRequirements era

StakePoolRetirementRequirementsPreConway ∷ ShelleyToBabbageEra era → PoolId → EpochNo → StakePoolRetirementRequirements era

makeStakePoolRegistrationCertificate ∷ StakePoolRegistrationRequirements era → Certificate era

makeStakePoolRetirementCertificate ∷ StakePoolRetirementRequirements era → Certificate era

data StakePoolParameters

Constructors

StakePoolParameters

Fields

stakePoolId ∷ PoolId

stakePoolVRF ∷ Hash VrfKey

stakePoolCost ∷ Lovelace

stakePoolMargin ∷ Rational

stakePoolRewardAccount ∷ StakeAddress

stakePoolPledge ∷ Lovelace

stakePoolOwners ∷ [Hash StakeKey]

stakePoolRelays ∷ [StakePoolRelay]

stakePoolMetadata ∷ Maybe StakePoolMetadataReference

data StakePoolRelay

Constructors

StakePoolRelayIp (Maybe IPv4) (Maybe IPv6) (Maybe PortNumber)
One or both of IPv4 & IPv6

StakePoolRelayDnsARecord ByteString (Maybe PortNumber)
An DNS name pointing to a A or AAAA record.

StakePoolRelayDnsSrvRecord ByteString
A DNS name pointing to a SRV record.

data StakePoolMetadataReference

Constructors

StakePoolMetadataReference

Fields

stakePoolMetadataURL ∷ Text

stakePoolMetadataHash ∷ Hash StakePoolMetadata

Conway specific certificates

data CommitteeColdkeyResignationRequirements era where

Constructors

CommitteeColdkeyResignationRequirements ∷ ConwayEraOnwards era → KeyHash ColdCommitteeRole (EraCrypto (ShelleyLedgerEra era)) → CommitteeColdkeyResignationRequirements era

data CommitteeHotKeyAuthorizationRequirements era where

Constructors

CommitteeHotKeyAuthorizationRequirements ∷ ConwayEraOnwards era → KeyHash ColdCommitteeRole (EraCrypto (ShelleyLedgerEra era)) → KeyHash HotCommitteeRole (EraCrypto (ShelleyLedgerEra era)) → CommitteeHotKeyAuthorizationRequirements era

data DRepRegistrationRequirements era where

Constructors

DRepRegistrationRequirements ∷ ConwayEraOnwards era → VotingCredential era → Lovelace → DRepRegistrationRequirements era

data DRepUnregistrationRequirements era where

Constructors

DRepUnregistrationRequirements ∷ ConwayEraOnwards era → VotingCredential era → Lovelace → DRepUnregistrationRequirements era

makeCommitteeColdkeyResignationCertificate ∷ CommitteeColdkeyResignationRequirements era → Certificate era

makeCommitteeHotKeyAuthorizationCertificate ∷ CommitteeHotKeyAuthorizationRequirements era → Certificate era

makeDrepRegistrationCertificate ∷ DRepRegistrationRequirements era → Maybe (Anchor (EraCrypto (ShelleyLedgerEra era))) → Certificate era

makeDrepUnregistrationCertificate ∷ DRepUnregistrationRequirements era → Certificate era

makeStakeAddressAndDRepDelegationCertificate ∷ ConwayEraOnwards era → StakeCredential → Delegatee (EraCrypto (ShelleyLedgerEra era)) → Lovelace → Certificate era

Registering DReps

data DRepMetadataReference

Constructors

DRepMetadataReference

Fields

drepMetadataURL ∷ Text

drepMetadataHash ∷ Hash DRepMetadata

Special certificates

data GenesisKeyDelegationRequirements ere where

Constructors

GenesisKeyDelegationRequirements ∷ ShelleyToBabbageEra era → Hash GenesisKey → Hash GenesisDelegateKey → Hash VrfKey → GenesisKeyDelegationRequirements era

data MirCertificateRequirements era where

Constructors

MirCertificateRequirements ∷ ShelleyToBabbageEra era → MIRPot → MIRTarget (EraCrypto (ShelleyLedgerEra era)) → MirCertificateRequirements era

makeMIRCertificate ∷ MirCertificateRequirements era → Certificate era

makeGenesisKeyDelegationCertificate ∷ GenesisKeyDelegationRequirements era → Certificate era

data MIRTarget c

MIRTarget specifies if funds from either the reserves or the treasury are to be handed out to a collection of reward accounts or instead transfered to the opposite pot.

Constructors

StakeAddressesMIR !(Map (Credential 'Staking c) DeltaCoin)

SendToOppositePotMIR !Coin

data MIRPot

Constructors

ReservesMIR

TreasuryMIR

Internal conversion functions

toShelleyCertificate ∷ Certificate era → TxCert (ShelleyLedgerEra era)

fromShelleyCertificate ∷ ShelleyBasedEra era → TxCert (ShelleyLedgerEra era) → Certificate era

toShelleyPoolParams ∷ StakePoolParameters → PoolParams StandardCrypto

fromShelleyPoolParams ∷ PoolParams StandardCrypto → StakePoolParameters

Data family instances

data family AsType t

A family of singleton types used in this API to indicate which type to use where it would otherwise be ambiguous or merely unclear.

Values of this type are passed to deserialisation functions for example.

Internal functions

filterUnRegCreds ∷ Certificate era → Maybe StakeCredential

filterUnRegDRepCreds ∷ Certificate era → Maybe (Credential DRepRole StandardCrypto)

selectStakeCredential ∷ Certificate era → Maybe StakeCredential