Cardano.Api.Certificate

Documentation

data Certificate era where

Constructors

ShelleyRelatedCertificate :: forall era. Typeable era => ShelleyToBabbageEra era -> ShelleyTxCert (ShelleyLedgerEra era) -> Certificate era

ConwayCertificate :: forall era. Typeable era => ConwayEraOnwards era -> ConwayTxCert (ShelleyLedgerEra era) -> Certificate era

Registering stake address and delegating

data StakeAddressRequirements era where

Constructors

StakeAddrRegistrationConway :: forall era. ConwayEraOnwards era -> Coin -> StakeCredential -> StakeAddressRequirements era

StakeAddrRegistrationPreConway :: forall era. ShelleyToBabbageEra era -> StakeCredential -> StakeAddressRequirements era

data StakeDelegationRequirements era where

Constructors

StakeDelegationRequirementsConwayOnwards :: forall era. ConwayEraOnwards era -> StakeCredential -> Delegatee -> StakeDelegationRequirements era

StakeDelegationRequirementsPreConway :: forall era. ShelleyToBabbageEra era -> StakeCredential -> PoolId -> StakeDelegationRequirements era

makeStakeAddressDelegationCertificate :: StakeDelegationRequirements era -> Certificate era

makeStakeAddressRegistrationCertificate :: StakeAddressRequirements era -> Certificate era

makeStakeAddressUnregistrationCertificate :: StakeAddressRequirements era -> Certificate era

type PoolId = Hash StakePoolKey

Registering stake pools

data StakePoolRegistrationRequirements era where

Constructors

StakePoolRegistrationRequirementsConwayOnwards :: forall era. ConwayEraOnwards era -> PoolParams -> StakePoolRegistrationRequirements era

StakePoolRegistrationRequirementsPreConway :: forall era. ShelleyToBabbageEra era -> PoolParams -> StakePoolRegistrationRequirements era

data StakePoolRetirementRequirements era where

Constructors

StakePoolRetirementRequirementsConwayOnwards :: forall era. ConwayEraOnwards era -> PoolId -> EpochNo -> StakePoolRetirementRequirements era

StakePoolRetirementRequirementsPreConway :: forall era. ShelleyToBabbageEra era -> PoolId -> EpochNo -> StakePoolRetirementRequirements era

makeStakePoolRegistrationCertificate :: StakePoolRegistrationRequirements era -> Certificate era

makeStakePoolRetirementCertificate :: StakePoolRetirementRequirements era -> Certificate era

data StakePoolParameters

Constructors

StakePoolParameters

Fields

stakePoolId :: PoolId

stakePoolVRF :: Hash VrfKey

stakePoolCost :: Coin

stakePoolMargin :: Rational

stakePoolRewardAccount :: StakeAddress

stakePoolPledge :: Coin

stakePoolOwners :: [Hash StakeKey]

stakePoolRelays :: [StakePoolRelay]

stakePoolMetadata :: Maybe StakePoolMetadataReference

data StakePoolRelay

Constructors

StakePoolRelayIp (Maybe IPv4) (Maybe IPv6) (Maybe PortNumber)
One or both of IPv4 & IPv6

StakePoolRelayDnsARecord ByteString (Maybe PortNumber)
An DNS name pointing to a A or AAAA record.

StakePoolRelayDnsSrvRecord ByteString
A DNS name pointing to a SRV record.

data StakePoolMetadataReference

Constructors

StakePoolMetadataReference

Fields

stakePoolMetadataURL :: Text

stakePoolMetadataHash :: Hash StakePoolMetadata

Conway specific certificates

data CommitteeColdkeyResignationRequirements era where

Constructors

CommitteeColdkeyResignationRequirements :: forall era. ConwayEraOnwards era -> Credential 'ColdCommitteeRole -> Maybe Anchor -> CommitteeColdkeyResignationRequirements era

data CommitteeHotKeyAuthorizationRequirements era where

Constructors

CommitteeHotKeyAuthorizationRequirements :: forall era. ConwayEraOnwards era -> Credential 'ColdCommitteeRole -> Credential 'HotCommitteeRole -> CommitteeHotKeyAuthorizationRequirements era

data DRepRegistrationRequirements era where

Constructors

DRepRegistrationRequirements :: forall era. ConwayEraOnwards era -> Credential 'DRepRole -> Coin -> DRepRegistrationRequirements era

data DRepUnregistrationRequirements era where

Constructors

DRepUnregistrationRequirements :: forall era. ConwayEraOnwards era -> Credential 'DRepRole -> Coin -> DRepUnregistrationRequirements era

data DRepUpdateRequirements era where

Constructors

DRepUpdateRequirements :: forall era. ConwayEraOnwards era -> Credential 'DRepRole -> DRepUpdateRequirements era

makeCommitteeColdkeyResignationCertificate :: Typeable era => CommitteeColdkeyResignationRequirements era -> Certificate era

makeCommitteeHotKeyAuthorizationCertificate :: Typeable era => CommitteeHotKeyAuthorizationRequirements era -> Certificate era

makeDrepRegistrationCertificate :: Typeable era => DRepRegistrationRequirements era -> Maybe Anchor -> Certificate era

makeDrepUnregistrationCertificate :: Typeable era => DRepUnregistrationRequirements era -> Certificate era

makeDrepUpdateCertificate :: Typeable era => DRepUpdateRequirements era -> Maybe Anchor -> Certificate era

makeStakeAddressAndDRepDelegationCertificate :: ConwayEraOnwards era -> StakeCredential -> Delegatee -> Coin -> Certificate era

Registering DReps

data DRepMetadataReference

Constructors

DRepMetadataReference

Fields

drepMetadataURL :: Text

drepMetadataHash :: Hash DRepMetadata

Special certificates

data GenesisKeyDelegationRequirements era where

Constructors

GenesisKeyDelegationRequirements :: forall era. ShelleyToBabbageEra era -> Hash GenesisKey -> Hash GenesisDelegateKey -> Hash VrfKey -> GenesisKeyDelegationRequirements era

data MirCertificateRequirements era where

Constructors

MirCertificateRequirements :: forall era. ShelleyToBabbageEra era -> MIRPot -> MIRTarget -> MirCertificateRequirements era

makeMIRCertificate :: Typeable era => MirCertificateRequirements era -> Certificate era

makeGenesisKeyDelegationCertificate :: Typeable era => GenesisKeyDelegationRequirements era -> Certificate era

data MIRTarget

MIRTarget specifies if funds from either the reserves or the treasury are to be handed out to a collection of reward accounts or instead transfered to the opposite pot.

Constructors

StakeAddressesMIR !(Map (Credential 'Staking) DeltaCoin)

SendToOppositePotMIR !Coin

data MIRPot

Constructors

ReservesMIR

TreasuryMIR

selectStakeCredentialWitness :: Certificate era -> Maybe StakeCredential

Get the stake credential witness for a certificate that requires it. Only stake address deregistration and delegation requires witnessing (witness can be script or key).

Anchor data

data AnchorDataFromCertificateError

Constructors

InvalidPoolMetadataHashError Url ByteString

getAnchorDataFromCertificate :: Certificate era -> Either AnchorDataFromCertificateError (Maybe Anchor)

Get anchor data url and hash from a certificate. A return value of Nothing means that the certificate does not contain anchor data.

Internal conversion functions

toShelleyCertificate :: Certificate era -> TxCert (ShelleyLedgerEra era)

fromShelleyCertificate :: ShelleyBasedEra era -> TxCert (ShelleyLedgerEra era) -> Certificate era

toShelleyPoolParams :: StakePoolParameters -> PoolParams

fromShelleyPoolParams :: PoolParams -> StakePoolParameters

Data family instances

data family AsType t

A family of singleton types used in this API to indicate which type to use where it would otherwise be ambiguous or merely unclear.

Values of this type are passed to deserialisation functions for example.

data family Hash keyrole

Operational Certificates

data OperationalCertificate

Constructors

OperationalCertificate !(OCert StandardCrypto) !(VerificationKey StakePoolKey)

data OperationalCertificateIssueCounter

Constructors

OperationalCertificateIssueCounter

Fields

opCertIssueCount :: !Word64

opCertIssueColdKey :: !(VerificationKey StakePoolKey)

newtype KESPeriod

Constructors

KESPeriod

Fields

unKESPeriod :: Word

data OperationalCertIssueError

Constructors

OperationalCertKeyMismatch (VerificationKey StakePoolKey) (VerificationKey StakePoolKey)
The stake pool verification key expected for the OperationalCertificateIssueCounter does not match the signing key supplied for signing.

Order: pool vkey expected, pool skey supplied

getHotKey :: OperationalCertificate -> VerificationKey KesKey

getKesPeriod :: OperationalCertificate -> Word

getOpCertCount :: OperationalCertificate -> Word64

issueOperationalCertificate :: VerificationKey KesKey -> Either AnyStakePoolSigningKey (SigningKey GenesisDelegateExtendedKey) -> KESPeriod -> OperationalCertificateIssueCounter -> Either OperationalCertIssueError (OperationalCertificate, OperationalCertificateIssueCounter)

DRep off-chain metadata

newtype DRepMetadata

A representation of the required fields for off-chain drep metadata.

Constructors

DRepMetadata

Fields

unDRepMetadata :: ByteString

hashDRepMetadata :: ByteString -> (DRepMetadata, Hash DRepMetadata)

Return the decoded metadata and the hash of the original bytes.

Stake pool off-chain metadata

data StakePoolMetadata

A representation of the required fields for off-chain stake pool metadata.

Constructors

StakePoolMetadata

Fields

stakePoolName :: !Text

A name of up to 50 characters.

stakePoolDescription :: !Text

A description of up to 255 characters.

stakePoolTicker :: !Text

A ticker of 3-5 characters, for a compact display of stake pools in a wallet.

stakePoolHomepage :: !Text

A URL to a homepage with additional information about the pool. n.b. the spec does not specify a character limit for this field.

validateAndHashStakePoolMetadata :: ByteString -> Either StakePoolMetadataValidationError (StakePoolMetadata, Hash StakePoolMetadata)

Decode and validate the provided JSON-encoded bytes as StakePoolMetadata. Return the decoded metadata and the hash of the original bytes.

data StakePoolMetadataValidationError

A stake pool metadata validation error.

Constructors

StakePoolMetadataJsonDecodeError !String

StakePoolMetadataInvalidLengthError
The length of the JSON-encoded stake pool metadata exceeds the maximum.

Fields

!Int

Maximum byte length.

!Int

Actual byte length.