Cardano.Api.Consensus

Documentation

data ShelleyGenesisStaking

Genesis Shelley staking configuration.

This allows us to configure some initial stake pools and delegation to them, in order to test Praos in a static configuration, without requiring on-chain registration and delegation.

For simplicity, pools defined in the genesis staking do not pay deposits for their registration.

Constructors

ShelleyGenesisStaking

Fields

sgsPools :: ListMap (KeyHash 'StakePool) PoolParams

Pools to register

The key in this map is the hash of the public key of the _pool_. This need not correspond to any payment or staking key, but must correspond to the cold key held by TPraosIsCoreNode.

sgsStake :: ListMap (KeyHash 'Staking) (KeyHash 'StakePool)

Stake-holding key hash credentials and the pools to delegate that stake to. We require the raw staking key hash in order to:

Avoid pointer addresses, which would be tricky when there's no slot or transaction to point to.

Avoid script credentials.

type family ChainDepState p

Protocol-specific state

NOTE: This chain is blockchain dependent, i.e., updated when new blocks come in (more precisely, new headers), and subject to rollback.

type family ChainDepState p

Protocol-specific state

NOTE: This chain is blockchain dependent, i.e., updated when new blocks come in (more precisely, new headers), and subject to rollback.

condense :: Condense a => a -> String

data family GenTx blk

Generalized transaction

The mempool (and, accordingly, blocks) consist of "generalized transactions"; this could be "proper" transactions (transferring funds) but also other kinds of things such as update proposals, delegations, etc.

txId :: HasTxId tx => tx -> TxId tx

Return the TxId of a GenTx.

NOTE: a TxId must be unique up to ledger rules, i.e., two GenTxs with the same TxId must be the same transaction according to the ledger. However, we do not assume that a TxId uniquely determines a GenTx: two GenTxs with the same TxId can differ in, e.g., witnesses.

Should be cheap as this will be called often.

unsafeExtendSafeZone :: forall (xs :: [Type]). Interpreter xs -> Interpreter xs

UNSAFE: extend the safe zone of the current era of the given Interpreter to be unbounded, ignoring any future hard forks.

This only has effect when the Interpreter was obtained in an era that was not the final one (in the final era, this is a no-op). The Interpreter will be made to believe that the current era is the final era, making its horizon unbounded, and thus never returning a PastHorizonException.

Use of this function is strongly discouraged, as it will ignore any future hard forks, and the results produced by the Interpreter can thus be incorrect.

data PastHorizonException

We tried to convert something that is past the horizon

That is, we tried to convert something that is past the point in time beyond which we lack information due to uncertainty about the next hard fork.

interpreterToEpochInfo :: forall (xs :: [Type]). Interpreter xs -> EpochInfo (Except PastHorizonException)

Construct an EpochInfo for a snapshot of the ledger state

data EraMismatch

Extra info for errors caused by applying a block, header, transaction, or query from one era to a ledger from a different era.

Constructors

EraMismatch

Fields

ledgerEraName :: !Text

Name of the era of the ledger (Byron or Shelley).

otherEraName :: !Text

Era of the block, header, transaction, or query.

data ByronBlock

Byron block

We cache two bits of information:

We cache the slot number as this is not readily available for EBBs. Having it cached allows us to e.g. give a HasHeader instance.

We cache the hash as this is expensive to compute and we need it often.

byronIdTx :: ATxAux ByteString -> TxId

class ConsensusProtocol p => PraosProtocolSupportsNode p where

The node has Praos-aware code that inspects nonces in order to support some Cardano API queries that are crucial to the user exprience

The interface being used for that has grown and needs review, but we're adding to it here under time pressure. See https://github.com/IntersectMBO/cardano-node/issues/3864

Minimal complete definition

getPraosNonces, getOpCertCounters

Associated Types

type PraosProtocolSupportsNodeCrypto p

Methods

getOpCertCounters :: proxy p -> ChainDepState p -> Map (KeyHash 'BlockIssuer) Word64

type family PraosProtocolSupportsNodeCrypto p