Cardano.Api.Consensus

Consensus modes

Consensus modes. The node supports several different modes with different combinations of consensus protocols and ledger eras.

The protocols supported in each era

type family ConsensusProtocol era where ...

Equations

ConsensusProtocol ShelleyEra = TPraos StandardCrypto

ConsensusProtocol AllegraEra = TPraos StandardCrypto

ConsensusProtocol MaryEra = TPraos StandardCrypto

ConsensusProtocol AlonzoEra = TPraos StandardCrypto

ConsensusProtocol BabbageEra = Praos StandardCrypto

ConsensusProtocol ConwayEra = Praos StandardCrypto

ConsensusProtocol DijkstraEra = Praos StandardCrypto

type family ChainDepStateProtocol era where ...

Equations

ChainDepStateProtocol ShelleyEra = TPraosState

ChainDepStateProtocol AllegraEra = TPraosState

ChainDepStateProtocol MaryEra = TPraosState

ChainDepStateProtocol AlonzoEra = TPraosState

ChainDepStateProtocol BabbageEra = PraosState

ChainDepStateProtocol ConwayEra = PraosState

Connection parameters for each mode

data ConsensusModeParams where

The consensus-mode-specific parameters needed to connect to a local node that is using each consensus mode.

It is in fact only the Byron era that requires extra parameters, but this is of course inherited by the CardanoMode that uses the Byron era. The reason this parameter is needed stems from unfortunate design decisions from the legacy Byron era. The slots per epoch are needed to be able to decode epoch boundary blocks from the Byron era.

It is possible in future that we may be able to eliminate this parameter by discovering it from the node during the initial handshake.

Constructors

CardanoModeParams :: EpochSlots -> ConsensusModeParams

newtype EpochSlots

The number of slots per epoch.

Constructors

EpochSlots

Fields

unEpochSlots :: Word64

Conversions to and from types in the consensus library

type family ConsensusCryptoForBlock block where ...

Equations

ConsensusCryptoForBlock ByronBlockHFC = StandardCrypto

ConsensusCryptoForBlock (ShelleyBlockHFC (TPraos StandardCrypto) ShelleyEra) = ShelleyEra

ConsensusCryptoForBlock (CardanoBlock StandardCrypto) = StandardCrypto

type family ConsensusBlockForEra era where ...

A closed type family that maps between the consensus mode (from this API) and the block type used by the consensus libraries.

Equations

ConsensusBlockForEra ByronEra = ByronBlock

ConsensusBlockForEra ShelleyEra = StandardShelleyBlock

ConsensusBlockForEra AllegraEra = StandardAllegraBlock

ConsensusBlockForEra MaryEra = StandardMaryBlock

ConsensusBlockForEra AlonzoEra = StandardAlonzoBlock

ConsensusBlockForEra BabbageEra = StandardBabbageBlock

ConsensusBlockForEra ConwayEra = StandardConwayBlock

ConsensusBlockForEra DijkstraEra = StandardDijkstraBlock

toConsensusEraIndex :: forall (xs :: [Type]) era. CardanoBlock StandardCrypto ~ HardForkBlock xs => CardanoEra era -> EraIndex xs

fromConsensusEraIndex :: EraIndex (CardanoEras StandardCrypto) -> AnyCardanoEra

Transactions in the consensus mode

Transactions in the context of a consensus mode, and other types used in the transaction submission protocol.

Transaction in a consensus mode

data TxInMode where

A Tx in one of the eras supported by a given protocol mode.

For multi-era modes such as the CardanoMode this type is a sum of the different transaction types for all the eras. It is used in the LocalTxSubmission protocol.

Constructors

TxInMode :: forall era. ShelleyBasedEra era -> Tx era -> TxInMode
Shelley based transactions.

TxInByronSpecial :: GenTx ByronBlock -> TxInMode
Legacy Byron transactions and things we can post to the chain which are not actually transactions. This covers: update proposals, votes and delegation certs.

fromConsensusGenTx :: CardanoBlock StandardCrypto ~ block => GenTx block -> TxInMode

toConsensusGenTx :: CardanoBlock StandardCrypto ~ block => TxInMode -> GenTx block

Transaction id in a consensus mode

data TxIdInMode where

A TxId in one of the eras supported by a given protocol mode.

For multi-era modes such as the CardanoMode this type is a sum of the different transaction types for all the eras. It is used in the LocalTxMonitoring protocol.

TODO Rename to TxIdInEra

Constructors

TxIdInMode :: forall era. CardanoEra era -> TxId -> TxIdInMode

toConsensusTxId :: CardanoBlock StandardCrypto ~ block => TxIdInMode -> TxId (GenTx block)

Transaction validation errors

data TxValidationError era where

The transaction validations errors that can occur from trying to submit a transaction to a local node. The errors are specific to an era.

Constructors

ByronTxValidationError :: forall era. ApplyTxErr ByronBlock -> TxValidationError era

ShelleyTxValidationError :: forall era. ShelleyBasedEra era -> ApplyTxErr (ShelleyBlock (ConsensusProtocol era) (ShelleyLedgerEra era)) -> TxValidationError era

data TxValidationErrorInCardanoMode where

A TxValidationError in one of the eras supported by a given protocol mode.

This is used in the LocalStateQuery protocol.

Constructors

TxValidationErrorInCardanoMode :: forall era. TxValidationError era -> TxValidationErrorInCardanoMode

TxValidationEraMismatch :: EraMismatch -> TxValidationErrorInCardanoMode

fromConsensusApplyTxErr :: CardanoBlock StandardCrypto ~ block => ApplyTxErr block -> TxValidationErrorInCardanoMode

Consensus protocol

data BlockType blk where

Constructors

ByronBlockType :: BlockType (HardForkBlock '[ByronBlock])

ShelleyBlockType :: BlockType (HardForkBlock '[ShelleyBlock (TPraos StandardCrypto) ShelleyEra])

CardanoBlockType :: BlockType (HardForkBlock (CardanoEras StandardCrypto))

data SomeBlockType where

Constructors

SomeBlockType :: forall blk. BlockType blk -> SomeBlockType

reflBlockType :: BlockType blk -> BlockType blk' -> Maybe (blk :~: blk')

class (RunNode blk, IOLike m) => Protocol (m :: Type -> Type) blk where

Associated Types

data ProtocolInfoArgs blk

Methods

protocolInfo :: ProtocolInfoArgs blk -> (ProtocolInfo blk, Tracer m KESAgentClientTrace -> m [MkBlockForging m blk])

data family ProtocolInfoArgs blk

class RunNode blk => ProtocolClient blk where

Node client support for each consensus protocol.

This is like Protocol but for clients of the node, so with less onerous requirements than to run a node.

Associated Types

data ProtocolClientInfoArgs blk

Methods

protocolClientInfo :: ProtocolClientInfoArgs blk -> ProtocolClientInfo blk

data family ProtocolClientInfoArgs blk

Reexports from ouroboros-consensus

data ByronBlock

Byron block

We cache two bits of information:

We cache the slot number as this is not readily available for EBBs. Having it cached allows us to e.g. give a HasHeader instance.

We cache the hash as this is expensive to compute and we need it often.

type family ChainDepState p

Protocol-specific state

NOTE: This chain is blockchain dependent, i.e., updated when new blocks come in (more precisely, new headers), and subject to rollback.

data family GenTx blk

Generalized transaction

The mempool (and, accordingly, blocks) consist of "generalized transactions"; this could be "proper" transactions (transferring funds) but also other kinds of things such as update proposals, delegations, etc.

data EraMismatch

Extra info for errors caused by applying a block, header, transaction, or query from one era to a ledger from a different era.

Constructors

EraMismatch

Fields

ledgerEraName :: !Text

Name of the era of the ledger (Byron or Shelley).

otherEraName :: !Text

Era of the block, header, transaction, or query.

data PastHorizonException

We tried to convert something that is past the horizon

That is, we tried to convert something that is past the point in time beyond which we lack information due to uncertainty about the next hard fork.

class ConsensusProtocol p => PraosProtocolSupportsNode p

The node has Praos-aware code that inspects nonces in order to support some Cardano API queries that are crucial to the user exprience

The interface being used for that has grown and needs review, but we're adding to it here under time pressure. See https://github.com/IntersectMBO/cardano-node/issues/3864

Minimal complete definition

getPraosNonces, getOpCertCounters

type family PraosProtocolSupportsNodeCrypto p

data ShelleyGenesisStaking

Genesis Shelley staking configuration.

This allows us to configure some initial stake pools and delegation to them, in order to test Praos in a static configuration, without requiring on-chain registration and delegation.

For simplicity, pools defined in the genesis staking do not pay deposits for their registration.

Constructors

ShelleyGenesisStaking

Fields

sgsPools :: ListMap (KeyHash 'StakePool) PoolParams

Pools to register

The key in this map is the hash of the public key of the _pool_. This need not correspond to any payment or staking key, but must correspond to the cold key held by TPraosIsCoreNode.

sgsStake :: ListMap (KeyHash 'Staking) (KeyHash 'StakePool)

Stake-holding key hash credentials and the pools to delegate that stake to. We require the raw staking key hash in order to:

Avoid pointer addresses, which would be tricky when there's no slot or transaction to point to.

Avoid script credentials.

byronIdTx :: ATxAux ByteString -> TxId

condense :: Condense a => a -> String

getOpCertCounters :: PraosProtocolSupportsNode p => proxy p -> ChainDepState p -> Map (KeyHash 'BlockIssuer) Word64

interpreterToEpochInfo :: forall (xs :: [Type]). Interpreter xs -> EpochInfo (Except PastHorizonException)

Construct an EpochInfo for a snapshot of the ledger state

unsafeExtendSafeZone :: forall (xs :: [Type]). Interpreter xs -> Interpreter xs

UNSAFE: extend the safe zone of the current era of the given Interpreter to be unbounded, ignoring any future hard forks.

This only has effect when the Interpreter was obtained in an era that was not the final one (in the final era, this is a no-op). The Interpreter will be made to believe that the current era is the final era, making its horizon unbounded, and thus never returning a PastHorizonException.

Use of this function is strongly discouraged, as it will ignore any future hard forks, and the results produced by the Interpreter can thus be incorrect.

txId :: HasTxId tx => tx -> TxId tx

Return the TxId of a GenTx.

NOTE: a TxId must be unique up to ledger rules, i.e., two GenTxs with the same TxId must be the same transaction according to the ledger. However, we do not assume that a TxId uniquely determines a GenTx: two GenTxs with the same TxId can differ in, e.g., witnesses.

Should be cheap as this will be called often.