Cardano.Api.Convenience.Query

Description

Convenience query functions

Documentation

data QueryConvenienceError

Constructors

AcqFailure AcquiringFailure

QueryEraMismatch EraMismatch

ByronEraNotSupported

EraConsensusModeMismatch !AnyConsensusMode !AnyCardanoEra

QceUnsupportedNtcVersion !UnsupportedNtcVersionError

determineEra ∷ ConsensusModeParams mode → LocalNodeConnectInfo mode → IO (Either AcquiringFailure AnyCardanoEra)

Query the node to determine which era it is in.

Simplest query related

executeQueryCardanoMode ∷ SocketPath → CardanoEra era → NetworkId → QueryInMode CardanoMode (Either EraMismatch result) → IO (Either QueryConvenienceError result)

Execute a query against the local node. The local node must be in CardanoMode.

executeQueryAnyMode ∷ ∀ result era mode. CardanoEra era → LocalNodeConnectInfo mode → QueryInMode mode (Either EraMismatch result) → IO (Either QueryConvenienceError result)

Execute a query against the local node in any mode.

queryStateForBalancedTx ∷ CardanoEra era → [TxIn] → [Certificate era] → LocalStateQueryExpr block point (QueryInMode CardanoMode) r IO (Either QueryConvenienceError (UTxO era, LedgerProtocolParameters era, EraHistory CardanoMode, SystemStart, Set PoolId, Map StakeCredential Lovelace, Map (Credential DRepRole StandardCrypto) Lovelace))

A convenience function to query the relevant information, from the local node, for Cardano.Api.Convenience.Construction.constructBalancedTx

renderQueryConvenienceError ∷ QueryConvenienceError → Text