Cardano.Api.Convenience.Query

Description

Convenience query functions

Documentation

data QueryConvenienceError

Constructors

AcqFailure AcquiringFailure

SockErr EnvSocketError

QueryEraMismatch EraMismatch

ByronEraNotSupported

EraConsensusModeMismatch !AnyConsensusMode !AnyCardanoEra

determineEra ∷ ConsensusModeParams mode → LocalNodeConnectInfo mode → IO (Either AcquiringFailure AnyCardanoEra)

Query the node to determine which era it is in.

Simplest query related

executeQueryCardanoMode ∷ CardanoEra era → NetworkId → QueryInMode CardanoMode (Either EraMismatch result) → IO (Either QueryConvenienceError result)

Execute a query against the local node. The local node must be in CardanoMode.

queryStateForBalancedTx ∷ CardanoEra era → NetworkId → [TxIn] → IO (Either QueryConvenienceError (UTxO era, ProtocolParameters, EraHistory CardanoMode, SystemStart, Set PoolId))

A convenience function to query the relevant information, from the local node, for Cardano.Api.Convenience.Construction.constructBalancedTx

renderQueryConvenienceError ∷ QueryConvenienceError → Text