Cardano.Api.Eon.ShelleyBasedEra

Shelley-based eras

data ShelleyBasedEra era where

While the Byron and Shelley eras are quite different, there are several eras that are based on Shelley with only minor differences. It is useful to be able to treat the Shelley-based eras in a mostly-uniform way.

Values of this type witness the fact that the era is Shelley-based. This can be used to constrain the era to being a Shelley-based on. It allows non-uniform handling making case distinctions on the constructor.

Constructors

ShelleyBasedEraShelley ∷ ShelleyBasedEra ShelleyEra

ShelleyBasedEraAllegra ∷ ShelleyBasedEra AllegraEra

ShelleyBasedEraMary ∷ ShelleyBasedEra MaryEra

ShelleyBasedEraAlonzo ∷ ShelleyBasedEra AlonzoEra

ShelleyBasedEraBabbage ∷ ShelleyBasedEra BabbageEra

ShelleyBasedEraConway ∷ ShelleyBasedEra ConwayEra

class IsCardanoEra era ⇒ IsShelleyBasedEra era where

The class of eras that are based on Shelley. This allows uniform handling of Shelley-based eras, but also non-uniform by making case distinctions on the ShelleyBasedEra constructors.

Methods

shelleyBasedEra ∷ ShelleyBasedEra era

data AnyShelleyBasedEra where

Constructors

AnyShelleyBasedEra ∷ IsShelleyBasedEra era ⇒ ShelleyBasedEra era → AnyShelleyBasedEra

data InAnyShelleyBasedEra thing where

This pairs up some era-dependent type with a ShelleyBasedEra value that tells us what era it is, but hides the era type. This is useful when the era is not statically known, for example when deserialising from a file.

Constructors

InAnyShelleyBasedEra ∷ IsShelleyBasedEra era ⇒ ShelleyBasedEra era → thing era → InAnyShelleyBasedEra thing

shelleyBasedToCardanoEra ∷ ShelleyBasedEra era → CardanoEra era

Converts a ShelleyBasedEra to the broader CardanoEra.

inEonForShelleyBasedEra ∷ Eon eon ⇒ a → (eon era → a) → ShelleyBasedEra era → a

Determine the value to use for a feature in a given ShelleyBasedEra.

inEonForShelleyBasedEraMaybe ∷ Eon eon ⇒ (eon era → a) → ShelleyBasedEra era → Maybe a

forShelleyBasedEraInEon ∷ Eon eon ⇒ ShelleyBasedEra era → a → (eon era → a) → a

forShelleyBasedEraInEonMaybe ∷ Eon eon ⇒ ShelleyBasedEra era → (eon era → a) → Maybe a

forShelleyBasedEraMaybeEon ∷ Eon eon ⇒ ShelleyBasedEra era → Maybe (eon era)

Cardano eras, as Byron vs Shelley-based

data CardanoEraStyle era where

This is the same essential information as CardanoEra but instead of a flat set of alternative eras, it is factored into the legcy Byron era and the current Shelley-based eras.

This way of factoring the eras is useful because in many cases the major differences are between the Byron and Shelley-based eras, and the Shelley-based eras can often be treated uniformly.

Constructors

LegacyByronEra ∷ CardanoEraStyle ByronEra

ShelleyBasedEra ∷ IsShelleyBasedEra era ⇒ ShelleyBasedEra era → CardanoEraStyle era

cardanoEraStyle ∷ CardanoEra era → CardanoEraStyle era

The CardanoEraStyle for a CardanoEra.

Assertions on era

requireShelleyBasedEra ∷ Applicative m ⇒ CardanoEra era → m (Maybe (ShelleyBasedEra era))

Mapping to era types from the Shelley ledger library

type family ShelleyLedgerEra era = ledgerera | ledgerera → era where ...

A type family that connects our era type tags to equivalent type tags used in the Shelley ledger library.

This type mapping connect types from this API with types in the Shelley ledger library which allows writing conversion functions in a more generic way.

Equations

ShelleyLedgerEra ShelleyEra = StandardShelley

ShelleyLedgerEra AllegraEra = StandardAllegra

ShelleyLedgerEra MaryEra = StandardMary

ShelleyLedgerEra AlonzoEra = StandardAlonzo

ShelleyLedgerEra BabbageEra = StandardBabbage

ShelleyLedgerEra ConwayEra = StandardConway

eraProtVerLow ∷ ShelleyBasedEra era → Version

Lookup the lower major protocol version for the shelley based era. In other words this is the major protocol version that the era has started in.