Cardano.Api.Eras

Description

Cardano eras, sometimes we have to distinguish them.

Eras

data ByronEra

A type used as a tag to distinguish the Byron era.

data ShelleyEra

A type used as a tag to distinguish the Shelley era.

data AllegraEra

A type used as a tag to distinguish the Allegra era.

data MaryEra

A type used as a tag to distinguish the Mary era.

data AlonzoEra

A type used as a tag to distinguish the Alonzo era.

data BabbageEra

A type used as a tag to distinguish the Babbage era.

data ConwayEra

A type used as a tag to distinguish the Conway era.

data CardanoEra era where

This GADT provides a value-level representation of all the Cardano eras. This enables pattern matching on the era to allow them to be treated in a non-uniform way.

This can be used in combination with the IsCardanoEra class to get access to this value.

In combination this can often enable code that handles all eras, and does so uniformly where possible, and non-uniformly where necessary.

Constructors

ByronEra ∷ CardanoEra ByronEra

ShelleyEra ∷ CardanoEra ShelleyEra

AllegraEra ∷ CardanoEra AllegraEra

MaryEra ∷ CardanoEra MaryEra

AlonzoEra ∷ CardanoEra AlonzoEra

BabbageEra ∷ CardanoEra BabbageEra

ConwayEra ∷ CardanoEra ConwayEra

class HasTypeProxy era ⇒ IsCardanoEra era where

The class of Cardano eras. This allows uniform handling of all Cardano eras, but also non-uniform by making case distinctions on the CardanoEra constructors, or the CardanoEraStyle constructors via cardanoEraStyle.

Methods

cardanoEra ∷ CardanoEra era

data AnyCardanoEra where

Constructors

AnyCardanoEra ∷ IsCardanoEra era ⇒ CardanoEra era → AnyCardanoEra

anyCardanoEra ∷ CardanoEra era → AnyCardanoEra

Like the AnyCardanoEra constructor but does not demand a IsCardanoEra class constraint.

data InAnyCardanoEra thing where

This pairs up some era-dependent type with a CardanoEra value that tells us what era it is, but hides the era type. This is useful when the era is not statically known, for example when deserialising from a file.

Constructors

InAnyCardanoEra ∷ IsCardanoEra era ⇒ CardanoEra era → thing era → InAnyCardanoEra thing

Deprecated aliases

type Byron = ByronEra

Deprecated: Use ByronEra or ByronAddr as appropriate

type Shelley = ShelleyEra

Deprecated: Use ShelleyEra or ShelleyAddr as appropriate

type Allegra = AllegraEra

Deprecated: Use AllegraEra instead

type Mary = MaryEra

Deprecated: Use MaryEra instead

Shelley-based eras

data ShelleyBasedEra era where

While the Byron and Shelley eras are quite different, there are several eras that are based on Shelley with only minor differences. It is useful to be able to treat the Shelley-based eras in a mostly-uniform way.

Values of this type witness the fact that the era is Shelley-based. This can be used to constrain the era to being a Shelley-based on. It allows non-uniform handling making case distinctions on the constructor.

Constructors

ShelleyBasedEraShelley ∷ ShelleyBasedEra ShelleyEra

ShelleyBasedEraAllegra ∷ ShelleyBasedEra AllegraEra

ShelleyBasedEraMary ∷ ShelleyBasedEra MaryEra

ShelleyBasedEraAlonzo ∷ ShelleyBasedEra AlonzoEra

ShelleyBasedEraBabbage ∷ ShelleyBasedEra BabbageEra

ShelleyBasedEraConway ∷ ShelleyBasedEra ConwayEra

class IsCardanoEra era ⇒ IsShelleyBasedEra era where

The class of eras that are based on Shelley. This allows uniform handling of Shelley-based eras, but also non-uniform by making case distinctions on the ShelleyBasedEra constructors.

Methods

shelleyBasedEra ∷ ShelleyBasedEra era

data InAnyShelleyBasedEra thing where

This pairs up some era-dependent type with a ShelleyBasedEra value that tells us what era it is, but hides the era type. This is useful when the era is not statically known, for example when deserialising from a file.

Constructors

InAnyShelleyBasedEra ∷ IsShelleyBasedEra era ⇒ ShelleyBasedEra era → thing era → InAnyShelleyBasedEra thing

shelleyBasedToCardanoEra ∷ ShelleyBasedEra era → CardanoEra era

Converts a ShelleyBasedEra to the broader CardanoEra.

Mapping to era types from the Shelley ledger library

type family ShelleyLedgerEra era where ...

A type family that connects our era type tags to equivalent type tags used in the Shelley ledger library.

This type mapping connect types from this API with types in the Shelley ledger library which allows writing conversion functions in a more generic way.

Equations

ShelleyLedgerEra ShelleyEra = StandardShelley

ShelleyLedgerEra AllegraEra = StandardAllegra

ShelleyLedgerEra MaryEra = StandardMary

ShelleyLedgerEra AlonzoEra = StandardAlonzo

ShelleyLedgerEra BabbageEra = StandardBabbage

ShelleyLedgerEra ConwayEra = StandardConway

Cardano eras, as Byron vs Shelley-based

data CardanoEraStyle era where

This is the same essential information as CardanoEra but instead of a flat set of alternative eras, it is factored into the legcy Byron era and the current Shelley-based eras.

This way of factoring the eras is useful because in many cases the major differences are between the Byron and Shelley-based eras, and the Shelley-based eras can often be treated uniformly.

Constructors

LegacyByronEra ∷ CardanoEraStyle ByronEra

ShelleyBasedEra ∷ IsShelleyBasedEra era ⇒ ShelleyBasedEra era → CardanoEraStyle era

cardanoEraStyle ∷ CardanoEra era → CardanoEraStyle era

The CardanoEraStyle for a CardanoEra.

Data family instances

data family AsType t

A family of singleton types used in this API to indicate which type to use where it would otherwise be ambiguous or merely unclear.

Values of this type are passed to deserialisation functions for example.