Cardano.Api.Eras.Case

Documentation

caseByronOrShelleyBasedEra ∷ (ByronEraOnly era → a) → (ShelleyBasedEraConstraints era ⇒ ShelleyBasedEra era → a) → CardanoEra era → a

caseByronOrShelleyBasedEra f g era applies f to byron and g to other eras.

caseByronToAllegraOrMaryEraOnwards ∷ (ByronToAllegraEraConstraints era ⇒ ByronToAllegraEra era → a) → (MaryEraOnwardsConstraints era ⇒ MaryEraOnwards era → a) → CardanoEra era → a

caseByronToAllegraOrMaryEraOnwards f g era applies f to byron, shelley, and allegra; and g to mary and later eras.

caseByronToMaryOrAlonzoEraOnwards ∷ (ByronToMaryEraConstraints era ⇒ ByronToMaryEra era → a) → (AlonzoEraOnwardsConstraints era ⇒ AlonzoEraOnwards era → a) → CardanoEra era → a

caseByronToMaryOrAlonzoEraOnwards f g era applies f to byron, shelley, allegra, and mary; and g to alonzo and later eras.

caseByronToAlonzoOrBabbageEraOnwards ∷ (ByronToAlonzoEraConstraints era ⇒ ByronToAlonzoEra era → a) → (BabbageEraOnwardsConstraints era ⇒ BabbageEraOnwards era → a) → CardanoEra era → a

caseByronToAlonzoOrBabbageEraOnwards f g era applies f to byron, shelley, allegra, mary, and alonzo; and g to babbage and later eras.

caseByronAndAllegraEraOnwardsOrShelleyEraOnly ∷ (ByronAndAllegraEraOnwardsConstraints era ⇒ ByronAndAllegraEraOnwards era → a) → (ShelleyEraOnlyConstraints era ⇒ ShelleyEraOnly era → a) → CardanoEra era → a

caseByronAndAllegraEraOnwardsOrShelleyEraOnly f g era applies f to all eras except shelley; and applies g to shelley.

caseShelleyEraOnlyOrAllegraEraOnwards ∷ (ShelleyEraOnlyConstraints era ⇒ ShelleyEraOnly era → a) → (AllegraEraOnwardsConstraints era ⇒ AllegraEraOnwards era → a) → ShelleyBasedEra era → a

caseShelleyEraOnlyOrAllegraEraOnwards f g era applies f to shelley; and applies g to allegra and later eras.

caseShelleyToAllegraOrMaryEraOnwards ∷ (ShelleyToAllegraEraConstraints era ⇒ ShelleyToAllegraEra era → a) → (MaryEraOnwardsConstraints era ⇒ MaryEraOnwards era → a) → ShelleyBasedEra era → a

caseShelleyToAllegraOrMaryEraOnwards f g era applies f to shelley and allegra; and applies g to mary and later eras.

caseShelleyToMaryOrAlonzoEraOnwards ∷ (ShelleyToMaryEraConstraints era ⇒ ShelleyToMaryEra era → a) → (AlonzoEraOnwardsConstraints era ⇒ AlonzoEraOnwards era → a) → ShelleyBasedEra era → a

caseShelleyToMaryOrAlonzoEraOnwards f g era applies f to shelley, allegra, and mary; and applies g to alonzo and later eras.

caseShelleyToAlonzoOrBabbageEraOnwards ∷ (ShelleyToAlonzoEraConstraints era ⇒ ShelleyToAlonzoEra era → a) → (BabbageEraOnwardsConstraints era ⇒ BabbageEraOnwards era → a) → ShelleyBasedEra era → a

caseShelleyToAlonzoOrBabbageEraOnwards f g era applies f to shelley, allegra, mary, and alonzo; and applies g to babbage and later eras.

caseShelleyToBabbageOrConwayEraOnwards ∷ (ShelleyToBabbageEraConstraints era ⇒ ShelleyToBabbageEra era → a) → (ConwayEraOnwardsConstraints era ⇒ ConwayEraOnwards era → a) → ShelleyBasedEra era → a

caseShelleyToBabbageOrConwayEraOnwards f g era applies f to eras before conway; and applies g to conway and later eras.

caseMaryEraOnlyOrAlonzoEraOnwards ∷ (MaryEraOnly era → a) → (AlonzoEraOnwards era → a) → MaryEraOnwards era → a

caseMaryEraOnlyOrAlonzoEraOnwards f g era applies f to mary; and applies g to alonzo and later eras.

caseAlonzoOnlyOrBabbageEraOnwards ∷ (AlonzoEraOnly era → a) → (BabbageEraOnwards era → a) → AlonzoEraOnwards era → a

caseAlonzoOnlyOrBabbageEraOnwards f g era applies f to alonzo; and applies g to babbage and later eras.

noByronEraInShelleyBasedEra ∷ ShelleyBasedEra era → ByronEraOnly era → a

Deprecated: Use disjointByronEraOnlyAndShelleyBasedEra instead

disjointAlonzoEraOnlyAndBabbageEraOnwards ∷ AlonzoEraOnly era → BabbageEraOnwards era → a

disjointByronEraOnlyAndShelleyBasedEra ∷ ByronEraOnly era → ShelleyBasedEra era → a

shelleyToAllegraEraToByronToAllegraEra ∷ ShelleyToAllegraEra era → ByronToAllegraEra era

alonzoEraOnlyToAlonzoEraOnwards ∷ AlonzoEraOnly era → AlonzoEraOnwards era

alonzoEraOnwardsToMaryEraOnwards ∷ AlonzoEraOnwards era → MaryEraOnwards era

babbageEraOnwardsToMaryEraOnwards ∷ BabbageEraOnwards era → MaryEraOnwards era

babbageEraOnwardsToAlonzoEraOnwards ∷ BabbageEraOnwards era → AlonzoEraOnwards era