Cardano.Api.Experimental

Description

This module provides an experimental library interface intended to replace the existing API. It is subject to significant changes. Please, use it with caution.

Creating transactions

For details and an example of creating a transaction using the experimental API, see the Cardano.Api.Internal.Experimental.Tx documentation.

Contents

Transaction-related

newtype UnsignedTx era

A transaction that can contain everything except key witnesses.

Constructors

UnsignedTx (Tx (LedgerEra era))

newtype UnsignedTxError

Constructors

UnsignedTxError TxBodyError

makeUnsignedTx :: Era era -> TxBodyContent BuildTx era -> Either TxBodyError (UnsignedTx era)

makeKeyWitness :: Era era -> UnsignedTx era -> ShelleyWitnessSigningKey -> WitVKey 'Witness

signTx :: Era era -> [BootstrapWitness] -> [WitVKey 'Witness] -> UnsignedTx era -> Tx (LedgerEra era)

convertTxBodyToUnsignedTx :: HasCallStack => ShelleyBasedEra era -> TxBody era -> UnsignedTx era

type EraCommonConstraints era = (AlonzoEraTx (LedgerEra era), BabbageEraTxBody (LedgerEra era), EraTx (LedgerEra era), EraUTxO (LedgerEra era), ShelleyLedgerEra era ~ LedgerEra era, HashAnnotated (TxBody (LedgerEra era)) EraIndependentTxBody, IsEra era)

obtainCommonConstraints :: Era era -> (EraCommonConstraints era => a) -> a

hashTxBody :: HashAnnotated (TxBody era) EraIndependentTxBody => TxBody era -> Hash HASH EraIndependentTxBody

evaluateTransactionExecutionUnitsShelley :: ShelleyBasedEra era -> SystemStart -> LedgerEpochInfo -> LedgerProtocolParameters era -> UTxO era -> Tx (ShelleyLedgerEra era) -> Either (TransactionValidityError era) (Map ScriptWitnessIndex (Either ScriptExecutionError (EvalTxExecutionUnitsLog, ExecutionUnits)))

Era-related

data BabbageEra

A type used as a tag to distinguish the Babbage era.

data ConwayEra

A type used as a tag to distinguish the Conway era.

data Era era where

Represents the latest Cardano blockchain eras, including the one currently on mainnet and the upcoming one.

After a hard fork takes place, the era on mainnet before the hard fork is deprecated and, after a deprecation period, removed from cardano-api. During the deprecation period, cardano-api users should update their codebase to the new mainnet era.

Constructors

BabbageEra :: Era BabbageEra
The currently active era on the Cardano mainnet.

ConwayEra :: Era ConwayEra
The upcoming era in development.

class IsEra era where

Type class interface for the Era type.

Methods

useEra :: Era era

data Some (f :: k -> Type) where

An existential wrapper for types of kind k -> Type. It can hold any era, for example, Some Era. The era witness can be brought back into scope, for example, using this pattern:

anyEra = Some ConwayEra
-- then later in the code
Some era <- pure anyEra
obtainCommonConstraints era foo

Constructors

Some :: forall {k} (f :: k -> Type) (a :: k). (Typeable a, Typeable (f a)) => f a -> Some f

type family LedgerEra era = (r :: Type) | r -> era where ...

Users typically interact with the latest features on the mainnet or experiment with features from the upcoming era. Therefore, protocol versions are limited to the current mainnet era and the next (upcoming) era.

Equations

LedgerEra BabbageEra = BabbageEra

LedgerEra ConwayEra = ConwayEra

newtype DeprecatedEra era

Constructors

DeprecatedEra (ShelleyBasedEra era)

eraToSbe :: Era era -> ShelleyBasedEra era

Deprecated: Use convert instead.

How to deprecate an era:

Add the DEPRECATED pragma to the era type tag and constructor at the same time:

{-# DEPRECATED BabbageEra "BabbageEra no longer supported, use ConwayEra" #-}
data BabbageEra

Update the Haddock documentation for the constructor of the deprecated era, mentioning the deprecation.

data Era era where
  {-# DEPRECATED BabbageEra "BabbageEra no longer supported, use ConwayEra" #-}
  BabbageEra :: Era BabbageEra
  -- | The era currently active on Cardano's mainnet.
  ConwayEra :: Era ConwayEra

Add a new IsEra instance and update the deprecated era instance to produce a compile-time error:

instance TypeError ('Text "IsEra BabbageEra: Deprecated. Update to ConwayEra") => IsEra BabbageEra where
  useEra = error "unreachable"

instance IsEra ConwayEra where
  useEra = ConwayEra

babbageEraOnwardsToEra :: BabbageEraOnwards era -> Era era

Deprecated: Use convert instead.

eraToBabbageEraOnwards :: Era era -> BabbageEraOnwards era

Deprecated: Use convert instead.

sbeToEra :: MonadError (DeprecatedEra era) m => ShelleyBasedEra era -> m (Era era)