Cardano.Api.Experimental

Description

This module provides an experimental library interface intended to replace the existing API. It is subject to significant changes. Please, use it with caution.

Creating transactions

For details and an example of creating a transaction using the experimental API, see the Cardano.Api.Experimental.Tx documentation.

Contents

Transaction-related

data UnsignedTx era

A transaction that can contain everything except key witnesses.

Constructors

EraTx (LedgerEra era) => UnsignedTx (Tx (LedgerEra era))

data SignedTx era

A transaction that has been witnesssed

Constructors

EraTx (LedgerEra era) => SignedTx (Tx (LedgerEra era))

makeUnsignedTx :: Era era -> TxBodyContent (LedgerEra era) -> UnsignedTx era

makeKeyWitness :: Era era -> UnsignedTx era -> ShelleyWitnessSigningKey -> WitVKey 'Witness

signTx :: Era era -> [BootstrapWitness] -> [WitVKey 'Witness] -> UnsignedTx era -> SignedTx era

convertTxBodyToUnsignedTx :: HasCallStack => ShelleyBasedEra era -> TxBody era -> UnsignedTx era

type EraCommonConstraints era = (AllegraEraScript (LedgerEra era), AlonzoEraScript (LedgerEra era), AlonzoEraTx (LedgerEra era), BabbageEraPParams (LedgerEra era), BabbageEraTxBody (LedgerEra era), ConwayEraTxBody (LedgerEra era), ConwayEraTxCert (LedgerEra era), Era (LedgerEra era), EraCertState (LedgerEra era), EraPlutusContext (LedgerEra era), EraTx (LedgerEra era), EraTxCert (LedgerEra era), EraTxOut (LedgerEra era), EraUTxO (LedgerEra era), Ord (PlutusPurpose AsIx (LedgerEra era)), ScriptsNeeded (LedgerEra era) ~ AlonzoScriptsNeeded (LedgerEra era), Val (Value (LedgerEra era)), Value (LedgerEra era) ~ MaryValue, FromCBOR (ChainDepState (ConsensusProtocol era)), NativeScript (LedgerEra era) ~ Timelock (LedgerEra era), PraosProtocolSupportsNode (ConsensusProtocol era), ShelleyLedgerEra era ~ LedgerEra era, LedgerEra era ~ ShelleyLedgerEra era, ToJSON (ChainDepState (ConsensusProtocol era)), HashAnnotated (TxBody (LedgerEra era)) EraIndependentTxBody, IsCardanoEra era, IsShelleyBasedEra era, IsEra era)

obtainCommonConstraints :: Era era -> (EraCommonConstraints era => a) -> a

hashTxBody :: HashAnnotated (TxBody era) EraIndependentTxBody => TxBody era -> Hash HASH EraIndependentTxBody

evaluateTransactionExecutionUnitsShelley :: ShelleyBasedEra era -> SystemStart -> LedgerEpochInfo -> LedgerProtocolParameters era -> UTxO era -> Tx (ShelleyLedgerEra era) -> Map ScriptWitnessIndex (Either ScriptExecutionError (EvalTxExecutionUnitsLog, ExecutionUnits))

data AnchorDataFromCertificateError

Constructors

InvalidPoolMetadataHashError Url ByteString

getAnchorDataFromCertificate :: Era era -> Certificate (LedgerEra era) -> Either AnchorDataFromCertificateError (Maybe Anchor)

mkTxCertificates :: Era era -> [(Certificate (LedgerEra era), AnyWitness (LedgerEra era))] -> TxCertificates (LedgerEra era)

Create TxCertificates. Note that 'Certificate era' will be deduplicated. Only Certificates with a stake credential will be in the result.

Note that, when building a transaction in Conway era, a witness is not required for staking credential registration, but this is only the case during the transitional period of Conway era and only for staking credential registration certificates without a deposit. Future eras will require a witness for registration certificates, because the one without a deposit will be removed.

Transaction fee related

estimateBalancedTxBody

Arguments

:: HasCallStack

=> Era era

-> TxBodyContent (LedgerEra era)

-> PParams (LedgerEra era)

-> Set PoolId
The set of registered stake pools, being unregistered in this transaction.

-> Map StakeCredential Coin
A map of all deposits for stake credentials that are being unregistered in this transaction.

-> Map (Credential 'DRepRole) Coin
A map of all deposits for DRep credentials that are being unregistered in this transaction.

-> Map (PlutusPurpose AsIx (LedgerEra era)) ExecutionUnits
Plutus script execution units.

-> Coin
Total potential collateral amount.

-> Int
The number of key witnesses to be added to the transaction.

-> Int
The number of Byron key witnesses to be added to the transaction.

-> Int
The size of all reference scripts in bytes.

-> AddressInEra era
Change address.

-> Value (LedgerEra era)
Total value of UTXOs being spent.

-> Either (TxFeeEstimationError era) (TxBodyContent (LedgerEra era))

Use when you do not have access to the UTxOs you intend to spend

evaluateTransactionFee

Arguments

:: PParams (LedgerEra era)

-> UnsignedTx era

-> Word
The number of Shelley key witnesses

-> Word
The number of Byron key witnesses

-> Int
Reference script size in bytes

-> Coin

Transaction fees can be computed for a proposed transaction based on the expected number of key witnesses (i.e. signatures).

When possible, use calculateMinTxFee, as it provides a more accurate estimate:

collectTxBodyScriptWitnesses :: IsEra era => TxBodyContent (LedgerEra era) -> [(ScriptWitnessIndex, AnyScriptWitness (LedgerEra era))]

Era-related

data BabbageEra

A type used as a tag to distinguish the Babbage era.

data ConwayEra

A type used as a tag to distinguish the Conway era.

data Era era where

Represents the latest Cardano blockchain eras, including the one currently on mainnet and the upcoming one.

After a hard fork takes place, the era on mainnet before the hard fork is deprecated and, after a deprecation period, removed from cardano-api. During the deprecation period, cardano-api users should update their codebase to the new mainnet era.

Constructors

ConwayEra :: Era ConwayEra
The currently active era on the Cardano mainnet.

DijkstraEra :: Era DijkstraEra

class IsEra era where

Type class interface for the Era type.

Methods

useEra :: Era era

data Some (f :: k -> Type) where

An existential wrapper for types of kind k -> Type. It can hold any era, for example, Some Era. The era witness can be brought back into scope, for example, using this pattern:

anyEra = Some ConwayEra
-- then later in the code
Some era <- pure anyEra
obtainCommonConstraints era foo

Constructors

Some :: forall {k} (f :: k -> Type) (a :: k). (Typeable a, Typeable (f a)) => f a -> Some f

type family LedgerEra era = (r :: Type) | r -> era where ...

Users typically interact with the latest features on the mainnet or experiment with features from the upcoming era. Therefore, protocol versions are limited to the current mainnet era and the next (upcoming) era.

Equations

LedgerEra ConwayEra = ConwayEra

LedgerEra DijkstraEra = DijkstraEra

newtype DeprecatedEra era

Constructors

DeprecatedEra (ShelleyBasedEra era)

eraToSbe :: Era era -> ShelleyBasedEra era

Deprecated: Use convert instead.

How to deprecate an era:

Add the DEPRECATED pragma to the era type tag and constructor at the same time:

{-# DEPRECATED BabbageEra "BabbageEra no longer supported, use ConwayEra" #-}
data BabbageEra

Update the Haddock documentation for the constructor of the deprecated era, mentioning the deprecation.

data Era era where
  {-# DEPRECATED BabbageEra "BabbageEra no longer supported, use ConwayEra" #-}
  BabbageEra :: Era BabbageEra
  -- | The era currently active on Cardano's mainnet.
  ConwayEra :: Era ConwayEra

Add a new IsEra instance and update the deprecated era instance to produce a compile-time error:

instance TypeError ('Text "IsEra BabbageEra: Deprecated. Update to ConwayEra") => IsEra BabbageEra where
  useEra = error "unreachable"

instance IsEra ConwayEra where
  useEra = ConwayEra

eraToBabbageEraOnwards :: Era era -> BabbageEraOnwards era

Deprecated: Use convert instead.

sbeToEra :: MonadError (DeprecatedEra era) m => ShelleyBasedEra era -> m (Era era)

Witness related

data AnyWitness era where

Here we consider three types of witnesses in Cardano: * key witnesses * simple script witnesses * Plutus script witnesses

Note that AnyKeyWitnessPlaceholder does not contain the actual key witness. This is because key witnesses are provided in the signing stage of the transaction. However we need this constuctor to index the witnessable things correctly when plutus scripts are being used within the transaction. AnyWitness is solely used to contruct the transaction body.

Constructors

AnyKeyWitnessPlaceholder :: forall era. AnyWitness era

AnySimpleScriptWitness :: forall era. SimpleScriptOrReferenceInput era -> AnyWitness era

AnyPlutusScriptWitness :: forall (lang :: Language) (purpose :: PlutusScriptPurpose) era. AnyPlutusScriptWitness lang purpose era -> AnyWitness era

data PlutusScriptWitness (lang :: Language) (purpose :: PlutusScriptPurpose) era where

This is a Plutus script witness. It possesses: 1. The plutus script or reference input 2. The script redeemer 3. The execution units 4. Potentially a script datum. See the PlutusScriptDatum type family for more details.

Note that Plutus script witnesses do not exist on their own. They must witness something and a redeemer pointer must be constucted to point to the thing being witnessed. See IndexedPlutusScriptWitness for more details.

Constructors

PlutusScriptWitness :: forall (lang :: Language) era (purpose :: PlutusScriptPurpose). SLanguage lang -> PlutusScriptOrReferenceInput lang era -> PlutusScriptDatum lang purpose -> ScriptRedeemer -> ExecutionUnits -> PlutusScriptWitness lang purpose era

data TxScriptWitnessRequirements era

This type collects all the requirements for script witnesses in a transaction.

Constructors

TxScriptWitnessRequirements (Set Language) [Script era] (TxDats era) (Redeemers era)

data Witnessable (thing :: WitnessableItem) era where

These are all of the "things" a plutus script can witness. We include the relevant type class constraint to avoid boilerplate when creating the PlutusPurpose in the toPlutusScriptPurpose.

Constructors

WitTxIn :: forall era. AlonzoEraScript era => TxIn -> Witnessable 'TxInItem era

WitTxCert :: forall era. (EraTxCert era, AlonzoEraScript era) => TxCert era -> StakeCredential -> Witnessable 'CertItem era

WitMint :: forall era. AlonzoEraScript era => PolicyId -> PolicyAssets -> Witnessable 'MintItem era

WitWithdrawal :: forall era. AlonzoEraScript era => StakeAddress -> Coin -> Witnessable 'WithdrawalItem era

WitVote :: forall era. ConwayEraScript era => Voter -> Witnessable 'VoterItem era

WitProposal :: forall era. (ConwayEraScript era, EraPParams era) => ProposalProcedure era -> Witnessable 'ProposalItem era

data WitnessableItem

Constructors

TxInItem

CertItem

MintItem

WithdrawalItem

VoterItem

ProposalItem

Simple script related

data SimpleScript era where

A simple script in a particular era. We leverage ledger's Cardano.Api.Experimental.ErasraScript type class methods to work with the script.

Constructors

SimpleScript :: forall era. EraScript era => NativeScript era -> SimpleScript era

data SimpleScriptOrReferenceInput era

Constructors

SScript (SimpleScript era)

SReferenceScript TxIn

deserialiseSimpleScript :: EraScript era => ByteString -> Either DecoderError (SimpleScript era)

hashSimpleScript :: IsEra era => SimpleScript (LedgerEra era) -> ScriptHash

Plutus related

data PlutusScriptInEra (lang :: Language) era where

A Plutus script in a particular era. Why PlutusRunnable? Mainly for deserialization benefits. The deserialization of this type looks at the major protocol version and the script language to determine if indeed the script is runnable. This is a dramatic improvement over the old api which essentially read a ByteString and hoped for the best. Any failures due to malformed/invalid scripts were caught upon transaction submission or running the script when attempting to predict the necessary execution units.

Where do we get the major protocol version from? In order to access the major protocol version we pass in an era type parameter which can be translated to the major protocol version.

Where do we get the script language from? The serialized version of PlutusRunnable encodes the script language. See `DecCBOR (PlutusRunnable l)` in cardano-ledger for more details.

Constructors

PlutusScriptInEra :: forall (lang :: Language) era. PlutusLanguage lang => PlutusRunnable lang -> PlutusScriptInEra lang era

data PlutusScriptOrReferenceInput (lang :: Language) era

You can provide the plutus script directly in the transaction or a reference input that points to the script in the UTxO. Using a reference script saves space in your transaction.

Constructors

PScript (PlutusScriptInEra lang era)

PReferenceScript TxIn

data IndexedPlutusScriptWitness (witnessable :: WitnessableItem) (lang :: Language) (purpose :: PlutusScriptPurpose) era where

A Plutus script witness along the thing it is witnessing and the index of that thing. E.g transaction input, certificate, withdrawal, minting policy, etc. A Plutus script witness only makes sense in the context of what it is witnessing and the index of the thing it is witnessing.

Constructors

IndexedPlutusScriptWitness :: forall era (witnessable :: WitnessableItem) (lang :: Language) (purpose :: PlutusScriptPurpose). AlonzoEraScript era => Witnessable witnessable era -> PlutusPurpose AsIx era -> AnyPlutusScriptWitness lang purpose era -> IndexedPlutusScriptWitness witnessable lang purpose era

data PlutusScriptPurpose

Every Plutus script has a purpose that indicates what that script is witnessing.

Constructors

SpendingScript
Witnesses a transaction input

MintingScript
Witnesses a minting policy

WithdrawingScript
Witnesses a withdrawal

CertifyingScript
Witnesses a certificate

ProposingScript
Witnesses a proposal

VotingScript
Witnesses a vote

data PlutusScriptDatum (lang :: Language) (purpose :: PlutusScriptPurpose) where

Constructors

SpendingScriptDatum :: forall (lang :: Language). PlutusScriptDatumF lang 'SpendingScript -> PlutusScriptDatum lang 'SpendingScript

InlineDatum :: forall (lang :: Language) (purpose :: PlutusScriptPurpose). PlutusScriptDatum lang purpose

NoScriptDatum :: forall (lang :: Language) (purpose :: PlutusScriptPurpose). PlutusScriptDatum lang purpose

data NoScriptDatum

Constructors

NoScriptDatumAllowed

Certificate related

data Certificate era where

Constructors

Certificate :: forall era. EraTxCert era => TxCert era -> Certificate era

Registering stake address and delegating

makeStakeAddressDelegationCertificate :: IsEra era => StakeCredential -> Delegatee era -> Certificate (LedgerEra era)

makeStakeAddressRegistrationCertificate :: IsEra era => StakeCredential -> Coin -> Certificate (LedgerEra era)

makeStakeAddressUnregistrationCertificate :: IsEra era => StakeCredential -> Coin -> Certificate (LedgerEra era)

Registering stake pools

makeStakePoolRegistrationCertificate :: IsEra era => PoolParams -> Certificate (LedgerEra era)

makeStakePoolRetirementCertificate :: IsShelleyBasedEra era => Hash StakePoolKey -> EpochNo -> Certificate (ShelleyLedgerEra era)

Governance related certificates

makeCommitteeColdkeyResignationCertificate :: IsEra era => Credential 'ColdCommitteeRole -> Maybe Anchor -> Certificate (LedgerEra era)

makeCommitteeHotKeyAuthorizationCertificate :: IsEra era => Credential 'ColdCommitteeRole -> Credential 'HotCommitteeRole -> Certificate (LedgerEra era)

makeDrepRegistrationCertificate :: IsEra era => Credential 'DRepRole -> Coin -> Maybe Anchor -> Certificate (LedgerEra era)

makeDrepUnregistrationCertificate :: IsEra era => Credential 'DRepRole -> Coin -> Certificate (LedgerEra era)

makeDrepUpdateCertificate :: IsEra era => Credential 'DRepRole -> Maybe Anchor -> Certificate (LedgerEra era)

makeStakeAddressAndDRepDelegationCertificate :: IsEra era => StakeCredential -> Delegatee -> Coin -> Certificate (LedgerEra era)

Data family instances

data family AsType t

A family of singleton types used in this API to indicate which type to use where it would otherwise be ambiguous or merely unclear.

Values of this type are passed to deserialisation functions for example.

Internal

getAnyWitnessRedeemerPointerMap :: forall era (witnessable :: WitnessableItem). AlonzoEraScript era => [(Witnessable witnessable era, AnyWitness era)] -> Redeemers era

The transaction's redeemer pointer map allows the ledger to connect a redeemer and execution unit pairing to the relevant script. The ledger basically reconstructs the indicies (redeemer pointers) of this map can then look up the relevant execution units/redeemer pairing. NB: the redeemer pointer has been renamed to 'PlutusPurpose AsIndex' in the ledger.

toPlutusScriptPurpose :: forall (thing :: WitnessableItem) era. Word32 -> Witnessable thing era -> PlutusPurpose AsIx era

To reduce boilerplate, we reuse the PlutusPurpose type from `cardano-ledger`. This type is utilized in constructing the redeemer pointers map, which links the redeemer and execution units with the entity being witnessed. The map is indexed by the redeemer pointer.

A natural question arises: How do Plutus scripts determine which execution units and redeemer are paired with them? The ledger constructs a redeemer pointer for every Plutus script, and this pointer corresponds to the one in the transaction's redeemer pointers map. For more details, refer to collectPlutusScriptsWithContext in `cardano-ledger`.

Legacy

legacyWitnessConversion :: forall era (witnessable :: WitnessableItem) ctx. AlonzoEraOnwards era -> [(Witnessable witnessable (ShelleyLedgerEra era), BuildTxWith BuildTx (Witness ctx era))] -> Either DecoderError [(Witnessable witnessable (ShelleyLedgerEra era), AnyWitness (ShelleyLedgerEra era))]

When it comes to using plutus scripts we need to provide the following to the tx:

The redeemer pointer map

The set of plutus languages in use

The set of plutus scripts in use (present in the t)

The datum map

toPlutusSLanguage :: PlutusScriptVersion lang -> SLanguage (ToLedgerPlutusLanguage lang)