Cardano.Api.Experimental.Tx

Creating transactions using the new API

Both the old and new APIs can be used to create transactions, and it is possible to transform a transaction from one format to the other as they share the same representation. However, the focus will shift towards using the new API, while the old API will be deprecated to ensure simplicity, closer alignment with the ledger, and easier maintenance.

In both the new and old APIs, constructing a transaction requires creating a TxBodyContent, along with at least one witness (for example, a ShelleyWitnessSigningKey) to sign the transaction. This process remains unchanged.

To learn how to create a transaction using the old API, see the Cardano.Api.Tx.Internal.Body documentation.

In the examples below, the following qualified modules are used:

import qualified Cardano.Api as Api                -- the general `cardano-api` exports (including the old API)
import qualified Cardano.Api.Script as Script      -- types related to scripts (Plutus and native)
import qualified Cardano.Api.Ledger as Ledger      -- cardano-ledger re-exports
import qualified Cardano.Api.Experimental as Exp   -- the experimental API

For instructions on how to do this, refer to the Test.Cardano.Api.Experimental documentation.

Creating a TxBodyContent

Regardless of whether the experimental or the traditional API is used, creating a TxBodyContent is necessary.

You can see how to do this in the documentation of the Cardano.Api.Tx.Internal.Body module.

Balancing a transaction

If a UTXO has exactly 12 ada, the transaction could be constructed as described in Cardano.Api.Tx.Internal.Body, and it would be valid. However:

Ada may be wasted

The UTXO that we intend to spend may not contain exactly 12 ada

The transaction may not be this simple.

For these reasons, it is recommended to balance the transaction before proceeding with signing and submitting.

For instructions on how to balance a transaction, refer to the Cardano.Api.Tx.Internal.Fee documentation.

Creating a ShelleyWitnessSigningKey

Signing a transaction requires a witness, such as a ShelleyWitnessSigningKey.

For instructions on creating a ShelleyWitnessSigningKey refer to the Cardano.Api.Tx.Internal.Sign documentation.

Creating a transaction using the new API

This section outlines how to create a transaction using the new API. First, create an UnsignedTx using the makeUnsignedTx function and the Era and TxBodyContent that we defined earlier:

case Exp.makeUnsignedTx era txBodyContent of
  Left err -> error (show err)
  Right unsignedTx -> ...

Next, use the key witness to sign the unsigned transaction with the makeKeyWitness function:

let transactionWitness = Exp.makeKeyWitness era unsignedTx (Api.WitnessPaymentKey signingKey)

Finally, sign the transaction using the signTx function:

let newApiSignedTx :: Ledger.Tx (Exp.LedgerEra Exp.ConwayEra) = Exp.signTx era [] [transactionWitness] unsignedTx

The empty list represents the bootstrap witnesses, which are not needed in this case.

The transaction is now signed.

Converting a transaction from the new API to the old API

A transaction created with the new API can be easily converted to the old API by wrapping it with the ShelleyTx constructor:

let oldStyleTx :: Api.Tx Api.ConwayEra = ShelleyTx sbe newApiSignedTx

Inspecting transactions

When using a Tx created with the experimental API, the TxBody and TxWits can be extracted using the txBody and txWits lenses from Cardano.Api.Ledger respectively.

Contents

data UnsignedTx era

A transaction that can contain everything except key witnesses.

Constructors

EraTx era => UnsignedTx (Tx 'TopTx era)

data SignedTx era

A transaction that has been witnesssed

Constructors

EraTx (ShelleyLedgerEra era) => SignedTx (Tx 'TopTx (ShelleyLedgerEra era))

data MakeUnsignedTxError

Error that can occur when constructing an unsigned transaction.

Constructors

MakeUnsignedTxMissingProtocolParams
Plutus scripts are present in the transaction but no protocol parameters were provided. Protocol parameters are required to compute the script integrity hash (script_data_hash).

makeUnsignedTx :: Era era -> TxBodyContent (LedgerEra era) -> Either MakeUnsignedTxError (UnsignedTx (LedgerEra era))

makeKeyWitness :: HasCallStack => Era era -> UnsignedTx (LedgerEra era) -> ShelleyWitnessSigningKey -> WitVKey 'Witness

signTx :: Era era -> [BootstrapWitness] -> [WitVKey 'Witness] -> UnsignedTx (LedgerEra era) -> SignedTx era

convertTxBodyToUnsignedTx :: HasCallStack => ShelleyBasedEra era -> TxBody era -> UnsignedTx (LedgerEra era)

hashTxBody :: HashAnnotated (TxBody 'TopTx era) EraIndependentTxBody => TxBody 'TopTx era -> Hash HASH EraIndependentTxBody

getUnsignedTxFee :: UnsignedTx era -> Coin

TxBodyContent

data TxBodyContent era

Constructors

TxBodyContent

Fields

txIns :: [(TxIn, AnyWitness era)]

txInsCollateral :: [TxIn]

txInsReference :: TxInsReference era

txOuts :: [TxOut era]

txTotalCollateral :: Maybe TxTotalCollateral

txReturnCollateral :: Maybe (TxReturnCollateral era)

txFee :: Coin

txValidityLowerBound :: Maybe SlotNo

txValidityUpperBound :: Maybe SlotNo

txMetadata :: TxMetadata

txAuxScripts :: [SimpleScript era]

txExtraKeyWits :: TxExtraKeyWitnesses

txProtocolParams :: Maybe (PParams era)

txWithdrawals :: TxWithdrawals era

txCertificates :: TxCertificates era

txMintValue :: TxMintValue era

txScriptValidity :: ScriptValidity

txProposalProcedures :: Maybe (TxProposalProcedures era)

txVotingProcedures :: Maybe (TxVotingProcedures era)

txCurrentTreasuryValue :: Maybe Coin

Current treasury value

txTreasuryDonation :: Maybe Coin

Treasury donation to perform

txSupplementalDatums :: Map DataHash (Data era)

Supplemental datums are datums whose hashes correspond to output datum hashes. They are included in the transaction witness set for communication purposes only.

defaultTxBodyContent :: TxBodyContent era

mkTxCertificates :: Era era -> [(Certificate (LedgerEra era), AnyWitness (LedgerEra era))] -> TxCertificates (LedgerEra era)

Create TxCertificates. Note that 'Certificate era' will be deduplicated. Only Certificates with a stake credential will be in the result.

Note that, when building a transaction in Conway era, a witness is not required for staking credential registration, but this is only the case during the transitional period of Conway era and only for staking credential registration certificates without a deposit. Future eras will require a witness for registration certificates, because the one without a deposit will be removed.

mkTxVotingProcedures :: [(VotingProcedures era, AnyWitness era)] -> Either (VotingError era) (TxVotingProcedures era)

Create voting procedures from map of voting procedures and optional witnesses. Validates the function argument, to make sure the list of votes is legal. See mergeVotingProcedures for validation rules.

mkTxProposalProcedures :: IsEra era => [(ProposalProcedure (LedgerEra era), AnyWitness (LedgerEra era))] -> TxProposalProcedures (LedgerEra era)

A smart constructor for TxProposalProcedures. It makes sure that the value produced is consistent - the witnessed proposals are also present in the first constructor parameter.

modTxOuts :: ([TxOut era] -> [TxOut era]) -> TxBodyContent era -> TxBodyContent era

setTxAuxScripts :: [SimpleScript era] -> TxBodyContent era -> TxBodyContent era

setTxCertificates :: TxCertificates era -> TxBodyContent era -> TxBodyContent era

setTxReturnCollateral :: TxReturnCollateral era -> TxBodyContent era -> TxBodyContent era

setTxTotalCollateral :: TxTotalCollateral -> TxBodyContent era -> TxBodyContent era

setTxCurrentTreasuryValue :: Coin -> TxBodyContent era -> TxBodyContent era

setTxExtraKeyWits :: TxExtraKeyWitnesses -> TxBodyContent era -> TxBodyContent era

setTxFee :: Coin -> TxBodyContent era -> TxBodyContent era

setTxIns :: [(TxIn, AnyWitness era)] -> TxBodyContent era -> TxBodyContent era

setTxInsCollateral :: [TxIn] -> TxBodyContent era -> TxBodyContent era

setTxInsReference :: TxInsReference era -> TxBodyContent era -> TxBodyContent era

setTxMetadata :: TxMetadata -> TxBodyContent era -> TxBodyContent era

setTxMintValue :: TxMintValue era -> TxBodyContent era -> TxBodyContent era

setTxOuts :: [TxOut era] -> TxBodyContent era -> TxBodyContent era

setTxProposalProcedures :: TxProposalProcedures era -> TxBodyContent era -> TxBodyContent era

setTxProtocolParams :: PParams era -> TxBodyContent era -> TxBodyContent era

setTxScriptValidity :: ScriptValidity -> TxBodyContent era -> TxBodyContent era

setTxSupplementalDatums :: Map DataHash (Data era) -> TxBodyContent era -> TxBodyContent era

setTxTreasuryDonation :: Coin -> TxBodyContent era -> TxBodyContent era

setTxValidityLowerBound :: SlotNo -> TxBodyContent era -> TxBodyContent era

setTxValidityUpperBound :: SlotNo -> TxBodyContent era -> TxBodyContent era

setTxVotingProcedures :: TxVotingProcedures era -> TxBodyContent era -> TxBodyContent era

setTxWithdrawals :: TxWithdrawals era -> TxBodyContent era -> TxBodyContent era

TxBodyContent sub type

newtype TxCertificates era

Constructors

TxCertificates

Fields

unTxCertificates :: OMap (Certificate era) (Maybe (StakeCredential, AnyWitness era))

newtype TxMintValue era

Constructors

TxMintValue

Fields

unTxMintValue :: Map PolicyId (PolicyAssets, AnyScriptWitness era)

data TxOut era where

Constructors

TxOut :: forall era. EraTxOut era => TxOut era -> TxOut era

newtype TxProposalProcedures era

Constructors

TxProposalProcedures (OMap (ProposalProcedure era) (AnyWitness era))

data TxVotingProcedures era

Constructors

TxVotingProcedures (VotingProcedures era) (Map Voter (AnyWitness era))

newtype TxWithdrawals era

Constructors

TxWithdrawals

Fields

unTxWithdrawals :: [(StakeAddress, Coin, AnyWitness era)]

newtype TxReturnCollateral era

Constructors

TxReturnCollateral

Fields

unTxReturnCollateral :: TxOut era

newtype TxTotalCollateral

Constructors

TxTotalCollateral

Fields

unTxTotalCollateral :: Coin

newtype TxExtraKeyWitnesses

Constructors

TxExtraKeyWitnesses [Hash PaymentKey]

data TxInsReference era

Constructors

TxInsReference [TxIn] (Set (Datum CtxTx era))

Witness

Any witness (key, simple script, plutus script).

data AnyWitness era where

Here we consider three types of witnesses in Cardano: * key witnesses * simple script witnesses * Plutus script witnesses

Note that AnyKeyWitnessPlaceholder does not contain the actual key witness. This is because key witnesses are provided in the signing stage of the transaction. However we need this constuctor to index the witnessable things correctly when plutus scripts are being used within the transaction. AnyWitness is solely used to contruct the transaction body.

Constructors

AnyKeyWitnessPlaceholder :: forall era. AnyWitness era

AnySimpleScriptWitness :: forall era. SimpleScriptOrReferenceInput era -> AnyWitness era

AnyPlutusScriptWitness :: forall (lang :: Language) (purpose :: PlutusScriptPurpose) era. AnyPlutusScriptWitness lang purpose era -> AnyWitness era

getAnyWitnessScript :: AlonzoEraScript era => AnyWitness era -> Maybe (Script era)

getAnyWitnessReferenceInput :: AnyWitness era -> Maybe TxIn

getAnyWitnessPlutusLanguage :: AnyWitness era -> Maybe Language

getAnyWitnessScriptData :: Era era => AnyWitness era -> TxDats era

NB this does not include datums from inline datums existing at tx outputs!

All the parts that constitute a plutus script witness but also including simple scripts

data TxScriptWitnessRequirements era

This type collects all the requirements for script witnesses in a transaction.

Constructors

TxScriptWitnessRequirements (Set Language) [Script era] (TxDats era) (Redeemers era)

Plutus related

data Datum ctx era where

Constructors

TxOutDatumHash :: forall ctx era. DataHash -> Datum ctx era

TxOutSupplementalDatum :: forall era. DataHash -> Data era -> Datum CtxTx era

TxOutDatumInline :: forall era ctx. DataHash -> Data era -> Datum ctx era

getDatums

Arguments

:: IsEra era

=> TxInsReference (LedgerEra era)
reference inputs

-> Map DataHash (Data (LedgerEra era))
supplemental datums

-> TxDats (LedgerEra era)

Collect datums for the transaction witness set (TxDats): 1. supplemental datums provided explicitly 2. datums from reference inputs

Supplemental datums are datums whose hashes correspond to datum hashes in transaction outputs. They are included for communication purposes only (the Alonzo ledger spec uses subset equality for these).

Note that this function does not check whose datum hashes are present in the reference inputs. This means if there are redundant datums in TxInsReference, a submission of such transaction will fail.

extractDatumsAndHashes :: Datum ctx era -> Maybe (DataHash, Data era)

Collecting plutus script witness related transaction requirements.

collectPlutusScriptHashes :: IsEra era => UnsignedTx (LedgerEra era) -> UTxO (LedgerEra era) -> Map ScriptWitnessIndex ScriptHash

Collect all plutus script hashes that are needed to validate the given transaction and return them in a map with their corresponding ScriptWitnessIndex as key.

extractAllIndexedPlutusScriptWitnesses :: Era era -> TxBodyContent (LedgerEra era) -> Either DecoderError [AnyIndexedPlutusScriptWitness (LedgerEra era)]

getTxScriptWitnessesRequirements :: forall era (witnessable :: WitnessableItem). (AlonzoEraScript era, Monoid (TxScriptWitnessRequirements era)) => [(Witnessable witnessable era, AnyWitness era)] -> TxScriptWitnessRequirements era

obtainMonoidConstraint :: Era era -> (Monoid (TxScriptWitnessRequirements (LedgerEra era)) => a) -> a

Transaction evaluation

evaluateTransaction

Arguments

:: IsEra era

=> SystemStart
Start time of the blockchain

-> LedgerEpochInfo
Epoch info for slot/time conversions

-> PParams (LedgerEra era)
Protocol parameters

-> Set PoolId
Registered stake pools

-> Map StakeCredential Coin
Stake delegation deposits

-> Map (Credential 'DRepRole) Coin
DRep delegation deposits

-> UTxO (LedgerEra era)
UTxO set for the transaction inputs

-> Tx 'TopTx (LedgerEra era)
Signed transaction to evaluate

-> TxEvaluationResult (LedgerEra era)

Run all scripts, compute the minimum fee, and check the balance.

evaluateSignedTx

Arguments

:: IsEra era

=> SystemStart
Start time of the blockchain

-> LedgerEpochInfo
Epoch info for slot/time conversions

-> PParams (LedgerEra era)
Protocol parameters

-> Set PoolId
Registered stake pools

-> Map StakeCredential Coin
Stake delegation deposits

-> Map (Credential 'DRepRole) Coin
DRep delegation deposits

-> UTxO (LedgerEra era)
UTxO set for the transaction inputs

-> SignedTx era
Signed transaction to evaluate

-> TxEvaluationResult (LedgerEra era)

Like evaluateTransaction but accepts a SignedTx directly.

data TxEvaluationResult era

Result of evaluating a signed transaction against the current ledger state.

Constructors

Show (Value era) => TxEvaluationResult

Fields

txEvalFee :: Coin

Computed minimum fee for the transaction

txEvalExecutionUnits :: Map ScriptWitnessIndex (Either ScriptExecutionError (EvalTxExecutionUnitsLog, ExecutionUnits))

Per-redeemer execution units or script errors

txEvalBalance :: Value era

Remaining balance (consumed - produced); mempty when balanced

evaluateTransactionExecutionUnits :: IsEra era => SystemStart -> LedgerEpochInfo -> PParams (LedgerEra era) -> UTxO (LedgerEra era) -> Tx 'TopTx (LedgerEra era) -> Map ScriptWitnessIndex (Either ScriptExecutionError (EvalTxExecutionUnitsLog, ExecutionUnits))

Compute the ExecutionUnits required for each script in the transaction.

This process involves executing all scripts and counting the actual execution units consumed.

Balancing transactions

calculateMinimumUTxO :: HasCallStack => PParams (LedgerEra era) -> TxOut (LedgerEra era) -> Coin

makeTransactionBodyAutoBalance

Arguments

:: (HasCallStack, IsEra era)

=> SystemStart

-> LedgerEpochInfo

-> PParams (LedgerEra era)

-> Set PoolId
The set of registered stake pools, being unregistered in this transaction.

-> Map StakeCredential Coin
The map of all deposits for stake credentials that are being unregistered in this transaction

-> Map (Credential 'DRepRole) Coin
The map of all deposits for DRep credentials that are being unregistered in this transaction

-> UTxO (LedgerEra era)
The transaction inputs (including reference and collateral ones), not the entire UTxO.

-> TxBodyContent (LedgerEra era)

-> AddressInEra era
Change address

-> Maybe Word
Override key witnesses

-> Either (TxBodyErrorAutoBalance (LedgerEra era)) (UnsignedTx (LedgerEra era), TxBodyContent (LedgerEra era))

This is similar to makeTransactionBody but with greater automation to calculate suitable values for several things.

In particular:

It calculates the correct script ExecutionUnits (ignoring the provided values, which can thus be zero).

It calculates the transaction fees based on the script ExecutionUnits, the current ProtocolParameters, and an estimate of the number of key witnesses (i.e. signatures). There is an override for the number of key witnesses.

It accepts a change address, calculates the balance of the transaction and puts the excess change into the change output.

It also checks that the balance is positive and the change is above the minimum threshold.

To do this, it requires more information than makeTransactionBody, all of which can be queried from a local node.

data TxBodyErrorAutoBalance era

Constructors

TxBodyErrorBalanceNegative Coin MultiAsset
There is not enough ada and non-ada to cover both the outputs and the fees. The transaction should be changed to provide more input assets, or otherwise adjusted to need less (e.g. outputs, script etc).

TxBodyErrorAdaBalanceTooSmall
There is enough ada to cover both the outputs and the fees, but the resulting change is too small: it is under the minimum value for new UTXO entries. The transaction should be changed to provide more input ada.

Fields

(TxOut era)

Offending TxOut

Coin

Minimum UTxO

Coin

Tx balance

TxBodyErrorMinUTxONotMet
The minimum spendable UTxO threshold has not been met.

Fields

(TxOut era)

Offending TxOut

Coin

Minimum UTXO

TxBodyErrorScriptWitnessIndexMissingFromExecUnitsMap ScriptWitnessIndex (Map ScriptWitnessIndex ExecutionUnits)

TxBodyScriptExecutionError [(ScriptWitnessIndex, ScriptExecutionError)]
One or more scripts failed to execute correctly.

TxBodyScriptBadScriptValidity
One or more scripts were expected to fail validation, but none did.

BalanceIsNegative

Fields

Coin

Negative balance

(UnsignedTx (LedgerEra ConwayEra))

The transaction body

NotEnoughAdaInUTxO

Fields

MaryValue

Total UTxO value

Coin

Total deposits

MaryValue

Balance

TxBodyErrorMakeUnsignedTx MakeUnsignedTxError

data TxFeeEstimationError era

Constructors

TxFeeEstimationScriptExecutionError (TxBodyErrorAutoBalance (LedgerEra era))

TxFeeEstimationBalanceError (TxBodyErrorAutoBalance (LedgerEra era))

TxFeeEstimationMakeUnsignedTxError MakeUnsignedTxError

Internal functions

extractExecutionUnits :: TxScriptWitnessRequirements era -> [ExecutionUnits]

getTxScriptWitnessRequirements :: forall era (witnessable :: WitnessableItem). (AlonzoEraScript era, Monoid (TxScriptWitnessRequirements era)) => [(Witnessable witnessable era, AnyWitness era)] -> TxScriptWitnessRequirements era

extractWitnessableTxIns :: IsEra era => [(TxIn, AnyWitness (LedgerEra era))] -> [(Witnessable 'TxInItem (LedgerEra era), AnyWitness (LedgerEra era))]

extractWitnessableMints :: IsEra era => TxMintValue (LedgerEra era) -> [(Witnessable 'MintItem (LedgerEra era), AnyScriptWitness (LedgerEra era))]

extractWitnessableCertificates :: IsEra era => TxCertificates (LedgerEra era) -> [(Witnessable 'CertItem (LedgerEra era), AnyWitness (LedgerEra era))]

extractWitnessableWithdrawals :: IsEra era => TxWithdrawals (LedgerEra era) -> [(Witnessable 'WithdrawalItem (LedgerEra era), AnyWitness (LedgerEra era))]

extractWitnessableVotes :: IsEra era => Maybe (TxVotingProcedures (LedgerEra era)) -> [(Witnessable 'VoterItem (LedgerEra era), AnyWitness (LedgerEra era))]

extractWitnessableProposals :: IsEra era => Maybe (TxProposalProcedures (LedgerEra era)) -> [(Witnessable 'ProposalItem (LedgerEra era), AnyWitness (LedgerEra era))]