Cardano.Api.Fees

Description

Fee calculation

Documentation

transactionFee

Arguments

∷ ∀ era. IsShelleyBasedEra era

⇒ Natural
The fixed tx fee

→ Natural
The tx fee per byte

→ Tx era

→ Lovelace

For a concrete fully-constructed transaction, determine the minimum fee that it needs to pay.

This function is simple, but if you are doing input selection then you probably want to consider estimateTransactionFee.

estimateTransactionFee

Arguments

∷ ∀ era. IsShelleyBasedEra era

⇒ NetworkId

→ Natural
The fixed tx fee

→ Natural
The tx fee per byte

→ Tx era

→ Int
The number of extra UTxO transaction inputs

→ Int
The number of extra transaction outputs

→ Int
The number of extra Shelley key witnesses

→ Int
The number of extra Byron key witnesses

→ Lovelace

This can estimate what the transaction fee will be, based on a starting base transaction, plus the numbers of the additional components of the transaction that may be added.

So for example with wallet coin selection, the base transaction should contain all the things not subject to coin selection (such as script inputs, metadata, withdrawals, certs etc)