Cardano.Api.Genesis

Documentation

data ShelleyGenesis era

Constructors

ShelleyGenesis

Fields

sgSystemStart ∷ !UTCTime

sgNetworkMagic ∷ !Word32

sgNetworkId ∷ !Network

sgActiveSlotsCoeff ∷ !PositiveUnitInterval

sgSecurityParam ∷ !Word64

sgEpochLength ∷ !EpochSize

sgSlotsPerKESPeriod ∷ !Word64

sgMaxKESEvolutions ∷ !Word64

sgSlotLength ∷ !NominalDiffTime

sgUpdateQuorum ∷ !Word64

sgMaxLovelaceSupply ∷ !Word64

sgProtocolParams ∷ !(ShelleyPParams era)

sgGenDelegs ∷ !(Map (KeyHash 'Genesis (Crypto era)) (GenDelegPair (Crypto era)))

sgInitialFunds ∷ ListMap (Addr (Crypto era)) Coin

sgStaking ∷ ShelleyGenesisStaking (Crypto era)

shelleyGenesisDefaults ∷ ShelleyGenesis crypto

Some reasonable starting defaults for constructing a ShelleyGenesis.

You must override at least the following fields for this to be useful:

sgSystemStart the time of the first block

sgNetworkMagic to a suitable testnet or mainnet network magic number.

sgGenDelegs to have some initial nodes

sgInitialFunds to have any money in the system

sgMaxLovelaceSupply must be at least the sum of the sgInitialFunds but more if you want to allow for rewards.