Cardano.Api.Genesis

Documentation

data ShelleyGenesis

Shelley genesis information

Note that this is needed only for a pure Shelley network, hence it being defined here rather than in its own module. In mainnet, Shelley will transition naturally from Byron, and thus will never have its own genesis information.

Constructors

ShelleyGenesis

Fields

sgSystemStart :: !UTCTime

sgNetworkMagic :: !Word32

sgNetworkId :: !Network

sgActiveSlotsCoeff :: !PositiveUnitInterval

sgSecurityParam :: !(NonZero Word64)

sgEpochLength :: !EpochSize

sgSlotsPerKESPeriod :: !Word64

sgMaxKESEvolutions :: !Word64

sgSlotLength :: !NominalDiffTimeMicro

sgUpdateQuorum :: !Word64

sgMaxLovelaceSupply :: !Word64

sgProtocolParams :: !(PParams ShelleyEra)

sgGenDelegs :: !(Map (KeyHash 'Genesis) GenDelegPair)

sgInitialFunds :: ListMap Addr Coin

sgInitialFunds is intentionally kept lazy, as it can otherwise cause out-of-memory problems in testing and benchmarking.

sgStaking :: ShelleyGenesisStaking

sgStaking is intentionally kept lazy, as it can otherwise cause out-of-memory problems in testing and benchmarking.

shelleyGenesisDefaults :: ShelleyGenesis

Some reasonable starting defaults for constructing a ShelleyGenesis.

You must override at least the following fields for this to be useful:

sgSystemStart the time of the first block

sgNetworkMagic to a suitable testnet or mainnet network magic number.

sgGenDelegs to have some initial nodes

sgInitialFunds to have any money in the system

sgMaxLovelaceSupply must be at least the sum of the sgInitialFunds but more if you want to allow for rewards.

alonzoGenesisDefaults :: CardanoEra era -> AlonzoGenesis

Some reasonable starting defaults for constructing a AlonzoGenesis. Based on https://github.com/IntersectMBO/cardano-node/blob/master/cardano-testnet/src/Testnet/Defaults.hs The era determines Plutus V2 cost model parameters: * Conway: 185 * <= Babbage: 175

decodeAlonzoGenesis

Arguments

:: forall era t (m :: Type -> Type). MonadTransError String t m

=> Maybe (CardanoEra era)
An optional era witness in which we're reading the genesis

-> ByteString
Genesis JSON

-> t m AlonzoGenesis

Decode Alonzo genesis in an optionally era sensitive way.

Because the Plutus V2 cost model has changed between Babbage and Conway era, we need to know the era if we want to decde Alonzo Genesis with a cost model baked in. If the V2 cost model is present in genesis, you need to provide an era witness.

When an era witness is provided, for Plutus V2 model the function additionally: 1. Does extra cost model parameters name validation: Checks for mandatory 175 parameters if provided in a map form. 2. If >= Conway: adds defaults for new 10 parameters, if they were not provided (maxBound) 3. Removes extra parameters above the max count: Babbage - 175, Conway - 185.

conwayGenesisDefaults :: ConwayGenesis

Some reasonable starting defaults for constructing a ConwayGenesis. Based on https://github.com/IntersectMBO/cardano-node/blob/master/cardano-testnet/src/Testnet/Defaults.hs

Configuration

type ByronGenesisConfig = Config

type ShelleyGenesisConfig = ShelleyConfig

type AlonzoGenesisConfig = AlonzoGenesis

type ConwayGenesisConfig = ConwayGenesis

data ShelleyConfig

Constructors

ShelleyConfig

Fields

scConfig :: !ShelleyGenesis

scGenesisHash :: !GenesisHashShelley

newtype GenesisHashByron

Constructors

GenesisHashByron

Fields

unGenesisHashByron :: Text

newtype GenesisHashShelley

Constructors

GenesisHashShelley

Fields

unGenesisHashShelley :: Hash Blake2b_256 ByteString

newtype GenesisHashAlonzo

Constructors

GenesisHashAlonzo

Fields

unGenesisHashAlonzo :: Hash Blake2b_256 ByteString

newtype GenesisHashConway

Constructors

GenesisHashConway

Fields

unGenesisHashConway :: Hash Blake2b_256 ByteString

Files

type ByronGenesisFile = File ByronGenesisConfig

type ShelleyGenesisFile = File ShelleyGenesisConfig

type AlonzoGenesisFile = File AlonzoGenesisConfig

type ConwayGenesisFile = File ConwayGenesisConfig

Protocol parameters fixed in the genesis file

data GenesisParameters era

Constructors

GenesisParameters

Fields

protocolParamSystemStart :: UTCTime

The reference time the system started. The time of slot zero. The time epoch against which all Ouroboros time slots are measured.

protocolParamNetworkId :: NetworkId

The network identifier for this blockchain instance. This distinguishes the mainnet from testnets, and different testnets from each other.

protocolParamActiveSlotsCoefficient :: Rational

The Ouroboros Praos active slot coefficient, aka f.

protocolParamSecurity :: NonZero Word64

The Ouroboros security parameters, aka k. This is the maximum number of blocks the node would ever be prepared to roll back by.

Clients of the node following the chain should be prepared to handle the node switching forks up to this long.

protocolParamEpochLength :: EpochSize

The number of Ouroboros time slots in an Ouroboros epoch.

protocolParamSlotLength :: NominalDiffTime

The time duration of a slot.

protocolParamSlotsPerKESPeriod :: Int

For Ouroboros Praos, the length of a KES period as a number of time slots. The KES keys get evolved once per KES period.

protocolParamMaxKESEvolutions :: Int

The maximum number of times a KES key can be evolved before it is no longer considered valid. This can be less than the maximum number of times given the KES key size. For example the mainnet KES key size would allow 64 evolutions, but the max KES evolutions param is 62.

protocolParamUpdateQuorum :: Int

In the Shelley era, prior to decentralised governance, this is the number of genesis key delegates that need to agree for an update proposal to be enacted.

protocolParamMaxLovelaceSupply :: Coin

The maximum supply for Lovelace. This determines the initial value of the reserves.

protocolInitialUpdateableProtocolParameters :: PParams (ShelleyLedgerEra era)

The initial values of the updateable ProtocolParameters.

newtype EpochSize

Constructors

EpochSize

Fields

unEpochSize :: Word64

Utilities

unsafeBoundedRational :: (HasCallStack, Typeable r, BoundedRational r) => Rational -> r

Convert Rational to a bounded rational. Throw an exception when the rational is out of bounds.