Cardano.Api.GenesisParameters

Description

Parameters fixed in the genesis file: GenesisParameters

Protocol parameters fixed in the genesis file

data GenesisParameters era

Constructors

GenesisParameters

Fields

protocolParamSystemStart ∷ UTCTime

The reference time the system started. The time of slot zero. The time epoch against which all Ouroboros time slots are measured.

protocolParamNetworkId ∷ NetworkId

The network identifier for this blockchain instance. This distinguishes the mainnet from testnets, and different testnets from each other.

protocolParamActiveSlotsCoefficient ∷ Rational

The Ouroboros Praos active slot coefficient, aka f.

protocolParamSecurity ∷ Int

The Ouroboros security parameters, aka k. This is the maximum number of blocks the node would ever be prepared to roll back by.

Clients of the node following the chain should be prepared to handle the node switching forks up to this long.

protocolParamEpochLength ∷ EpochSize

The number of Ouroboros time slots in an Ouroboros epoch.

protocolParamSlotLength ∷ NominalDiffTime

The time duration of a slot.

protocolParamSlotsPerKESPeriod ∷ Int

For Ouroboros Praos, the length of a KES period as a number of time slots. The KES keys get evolved once per KES period.

protocolParamMaxKESEvolutions ∷ Int

The maximum number of times a KES key can be evolved before it is no longer considered valid. This can be less than the maximum number of times given the KES key size. For example the mainnet KES key size would allow 64 evolutions, but the max KES evolutions param is 62.

protocolParamUpdateQuorum ∷ Int

In the Shelley era, prior to decentralised governance, this is the number of genesis key delegates that need to agree for an update proposal to be enacted.

protocolParamMaxLovelaceSupply ∷ Lovelace

The maximum supply for Lovelace. This determines the initial value of the reserves.

protocolInitialUpdateableProtocolParameters ∷ PParams (ShelleyLedgerEra era)

The initial values of the updateable ProtocolParameters.

newtype EpochSize

Constructors

EpochSize

Fields

unEpochSize ∷ Word64

Internal conversion functions

fromShelleyGenesis ∷ ShelleyGenesis StandardCrypto → GenesisParameters ShelleyEra