Cardano.Api.Governance.Actions.ProposalProcedure

Documentation

data AnyGovernanceAction

Constructors

∀ era. AnyGovernanceAction (GovAction era)

data GovernanceAction

Constructors

MotionOfNoConfidence (StrictMaybe (PrevGovActionId CommitteePurpose StandardCrypto))

ProposeNewConstitution (StrictMaybe (PrevGovActionId ConstitutionPurpose StandardCrypto)) (Anchor StandardCrypto)

ProposeNewCommittee

Fields

(StrictMaybe (PrevGovActionId CommitteePurpose StandardCrypto))

[Hash CommitteeColdKey]

Old constitutional committee

(Map (Hash CommitteeColdKey) EpochNo)

New committee members with epoch number when each of them expires

Rational

Quorum of the committee that is necessary for a successful vote

InfoAct

TreasuryWithdrawal [(Network, StakeCredential, Lovelace)]

InitiateHardfork (StrictMaybe (PrevGovActionId HardForkPurpose StandardCrypto)) ProtVer

UpdatePParams (StrictMaybe (PrevGovActionId PParamUpdatePurpose StandardCrypto)) ProtocolParametersUpdate

toGovernanceAction ∷ ShelleyBasedEra era → GovernanceAction → GovAction (ShelleyLedgerEra era)

fromGovernanceAction ∷ EraCrypto (ShelleyLedgerEra era) ~ StandardCrypto ⇒ ShelleyBasedEra era → GovAction (ShelleyLedgerEra era) → GovernanceAction

newtype Proposal era

Constructors

Proposal

Fields

unProposal ∷ ProposalProcedure (ShelleyLedgerEra era)

createProposalProcedure

Arguments

∷ ShelleyBasedEra era

→ Network

→ Lovelace
Deposit

→ Hash StakeKey
Return address

→ GovernanceAction

→ Anchor StandardCrypto

→ Proposal era

fromProposalProcedure ∷ ShelleyBasedEra era → Proposal era → (Lovelace, Hash StakeKey, GovernanceAction)

createPreviousGovernanceActionId

Arguments

∷ TxId

→ Word32
Governance action transation index

→ PrevGovActionId (r ∷ GovActionPurpose) StandardCrypto

createGovernanceActionId ∷ TxId → Word32 → GovActionId StandardCrypto

createAnchor ∷ Url → ByteString → Anchor StandardCrypto

toCommitteeMember ∷ Hash CommitteeColdKey → Credential ColdCommitteeRole StandardCrypto

fromCommitteeMember ∷ Credential ColdCommitteeRole StandardCrypto → Hash CommitteeColdKey