Cardano.Api.Governance.Actions.VotingProcedure

Documentation

newtype GovernanceActionId era

Constructors

GovernanceActionId

Fields

unGovernanceActionId ∷ GovActionId (EraCrypto (ShelleyLedgerEra era))

data Voter era

Constructors

VoterCommittee (VotingCredential era)
Constitutional committee

VoterDRep (VotingCredential era)
Delegated representative

VoterSpo (Hash StakePoolKey)
Stake pool operator

data Vote

Constructors

No

Yes

Abstain

toVoterRole ∷ ConwayEraOnwards era → Voter era → Voter (EraCrypto (ShelleyLedgerEra era))

fromVoterRole ∷ ConwayEraOnwards era → Voter (EraCrypto (ShelleyLedgerEra era)) → Voter era

toVote ∷ Vote → Vote

toVotingCredential ∷ ConwayEraOnwards era → StakeCredential → Either DecoderError (VotingCredential era)

eraDecodeVotingCredential ∷ ConwayEraOnwards era → ByteString → Either DecoderError (VotingCredential era)

newtype VotingCredential era

Constructors

VotingCredential

Fields

unVotingCredential ∷ Credential 'DRepRole (EraCrypto (ShelleyLedgerEra era))

createVotingProcedure

Arguments

∷ ConwayEraOnwards era

→ Vote

→ Maybe (Url, Text)
Anchor

→ VotingProcedure era

newtype VotingProcedure era

Constructors

VotingProcedure

Fields

unVotingProcedure ∷ VotingProcedure (ShelleyLedgerEra era)

newtype VotingProcedures era

Constructors

VotingProcedures

Fields

unVotingProcedures ∷ VotingProcedures (ShelleyLedgerEra era)

emptyVotingProcedures ∷ VotingProcedures era

singletonVotingProcedures ∷ ConwayEraOnwards era → Voter (EraCrypto (ShelleyLedgerEra era)) → GovActionId (EraCrypto (ShelleyLedgerEra era)) → VotingProcedure (ShelleyLedgerEra era) → VotingProcedures era

unsafeMergeVotingProcedures ∷ VotingProcedures era → VotingProcedures era → VotingProcedures era

Right biased merge of Voting procedures. TODO Conway we need an alternative version of this function that can report conflicts as it is not safe to just throw away votes.