Cardano.Api.Hash

Documentation

data family Hash keyrole ∷ Type

class CastHash roleA roleB where

Methods

castHash ∷ Hash roleA → Hash roleB

data family AsType t

A family of singleton types used in this API to indicate which type to use where it would otherwise be ambiguous or merely unclear.

Values of this type are passed to deserialisation functions for example.