Cardano.Api.IO.Compat.Posix

Documentation

data VRFPrivateKeyFilePermissionError

checkVrfFilePermissionsImpl ∷ File content direction → ExceptT VRFPrivateKeyFilePermissionError IO ()

Make sure the VRF private key file is readable only by the current process owner the node is running under.

handleFileForWritingWithOwnerPermissionImpl ∷ FilePath → (Handle → IO ()) → IO (Either (FileError e) ())

writeSecretsImpl ∷ FilePath → [Char] → [Char] → (a → ByteString) → [a] → IO ()