Cardano.Api.IPC

Description

Node IPC protocols

Node interaction

Operations that involve talking to a local Cardano node.

connectToLocalNode ∷ LocalNodeConnectInfo mode → LocalNodeClientProtocolsInMode mode → IO ()

Establish a connection to a local node and execute the given set of protocol handlers.

connectToLocalNodeWithVersion ∷ LocalNodeConnectInfo mode → (NodeToClientVersion → LocalNodeClientProtocolsInMode mode) → IO ()

Establish a connection to a local node and execute the given set of protocol handlers parameterized on the negotiated node-to-client protocol version.

data LocalNodeConnectInfo mode

Constructors

LocalNodeConnectInfo

Fields

localConsensusModeParams ∷ ConsensusModeParams mode

localNodeNetworkId ∷ NetworkId

localNodeSocketPath ∷ SocketPath

localConsensusMode ∷ LocalNodeConnectInfo mode → ConsensusMode mode

data LocalNodeClientParams where

This type defines the boundary between the mode-parametrised style used in this API and the block-parametrised style used by the underlying network and consensus libraries.

This interface itself is in the block-parametrised style, with the block type itself being an hidden/existential type.

It bundles together all the necessary class instances, the consensus protocol client identifier, and the set of client side mini-protocol handlers for the node-to-client protocol.

Constructors

LocalNodeClientParamsSingleBlock ∷ (ProtocolClient block, LedgerSupportsProtocol (ShelleyBlock (TPraos StandardCrypto) (ShelleyEra StandardCrypto))) ⇒ ProtocolClientInfoArgs block → (NodeToClientVersion → LocalNodeClientProtocolsForBlock block) → LocalNodeClientParams

LocalNodeClientParamsCardano ∷ (ProtocolClient block, CardanoHardForkConstraints (ConsensusCryptoForBlock block)) ⇒ ProtocolClientInfoArgs block → (NodeToClientVersion → LocalNodeClientProtocolsForBlock block) → LocalNodeClientParams

mkLocalNodeClientParams ∷ ∀ mode block. ConsensusBlockForMode mode ~ block ⇒ ConsensusModeParams mode → (NodeToClientVersion → LocalNodeClientProtocolsInMode mode) → LocalNodeClientParams

Convert from the mode-parametrised style to the block-parametrised style.

data LocalNodeClientProtocols block point tip slot tx txid txerr query m

The protocols we can use with a local node. Use in conjunction with connectToLocalNode.

These protocols use the types from the rest of this API. The conversion to/from the types used by the underlying wire formats is handled by connectToLocalNode.

Constructors

LocalNodeClientProtocols

Fields

localChainSyncClient ∷ LocalChainSyncClient block point tip m

localTxSubmissionClient ∷ Maybe (LocalTxSubmissionClient tx txerr m ())

localStateQueryClient ∷ Maybe (LocalStateQueryClient block point query m ())

localTxMonitoringClient ∷ Maybe (LocalTxMonitorClient txid tx slot m ())

data LocalChainSyncClient block point tip m

Constructors

NoLocalChainSyncClient

LocalChainSyncClientPipelined (ChainSyncClientPipelined block point tip m ())

LocalChainSyncClient (ChainSyncClient block point tip m ())

type LocalNodeClientProtocolsInMode mode = LocalNodeClientProtocols (BlockInMode mode) ChainPoint ChainTip SlotNo (TxInMode mode) (TxIdInMode mode) (TxValidationErrorInMode mode) (QueryInMode mode) IO

Modes

TODO move to Cardano.Api

data ByronMode

The Byron-only consensus mode consists of only the Byron era.

This was used on the mainnet before the deployment of the multi-era CardanoMode. It is now of little practical use, though it illustrates how a single-era consensus mode works. It may be sensible to remove this at some stage.

data ShelleyMode

The Shelley-only consensus mode consists of only the Shelley era.

This was used for the early Shelley testnets prior to the use of the multi-era CardanoMode. It is useful for setting up Shelley test networks (e.g. for benchmarking) without having to go through the complication of the hard fork from Byron to Shelley eras. It also shows how a single-era consensus mode works. It may be replaced by other single-era modes in future.

data CardanoMode

The Cardano consensus mode consists of all the eras currently in use on the Cardano mainnet. This is currently: the ByronEra; ShelleyEra, AllegraEra and MaryEra, in that order.

This mode will be extended with new eras as the Cardano mainnet develops.

data ConsensusModeParams mode where

The consensus-mode-specific parameters needed to connect to a local node that is using each consensus mode.

It is in fact only the Byron era that requires extra parameters, but this is of course inherited by the CardanoMode that uses the Byron era. The reason this parameter is needed stems from unfortunate design decisions from the legacy Byron era. The slots per epoch are needed to be able to decode epoch boundary blocks from the Byron era.

It is possible in future that we may be able to eliminate this parameter by discovering it from the node during the initial handshake.

Constructors

ByronModeParams ∷ EpochSlots → ConsensusModeParams ByronMode

ShelleyModeParams ∷ ConsensusModeParams ShelleyMode

CardanoModeParams ∷ EpochSlots → ConsensusModeParams CardanoMode

newtype EpochSlots

The number of slots per epoch.

Constructors

EpochSlots

Fields

unEpochSlots ∷ Word64

Chain sync protocol

newtype ChainSyncClient header point tip (m ∷ Type → Type) a

A chain sync protocol client, on top of some effect m. The first choice of request is within that m.

Constructors

ChainSyncClient

Fields

runChainSyncClient ∷ m (ClientStIdle header point tip m a)

newtype ChainSyncClientPipelined header point tip (m ∷ Type → TYPE LiftedRep) a

Pipelined chain sync client. It can only pipeline MsgRequestNext messages, while the MsgFindIntersect are non pipelined. This has a penalty cost of an RTT, but they are sent relatively seldom and their response might impact how many messages one would like to pipeline. It also simplifies the receiver callback.

Constructors

ChainSyncClientPipelined

Fields

runChainSyncClientPipelined ∷ m (ClientPipelinedStIdle 'Z header point tip m a)

data BlockInMode mode where

A Block in one of the eras supported by a given protocol mode.

For multi-era modes such as the CardanoMode this type is a sum of the different block types for all the eras. It is used in the ChainSync protocol.

Constructors

BlockInMode ∷ IsCardanoEra era ⇒ Block era → EraInMode era mode → BlockInMode mode

Local tx submission

newtype LocalTxSubmissionClient tx reject (m ∷ Type → Type) a

Constructors

LocalTxSubmissionClient

Fields

runLocalTxSubmissionClient ∷ m (LocalTxClientStIdle tx reject m a)

data TxInMode mode where

A Tx in one of the eras supported by a given protocol mode.

For multi-era modes such as the CardanoMode this type is a sum of the different transaction types for all the eras. It is used in the LocalTxSubmission protocol.

Constructors

TxInMode ∷ Tx era → EraInMode era mode → TxInMode mode
Everything we consider a normal transaction.

TxInByronSpecial ∷ GenTx ByronBlock → EraInMode ByronEra mode → TxInMode mode
Byron has various things we can post to the chain which are not actually transactions. This covers: update proposals, votes and delegation certs.

data TxValidationErrorInMode mode where

A TxValidationError in one of the eras supported by a given protocol mode.

This is used in the LocalStateQuery protocol.

Constructors

TxValidationErrorInMode ∷ TxValidationError era → EraInMode era mode → TxValidationErrorInMode mode

TxValidationEraMismatch ∷ EraMismatch → TxValidationErrorInMode mode

data TxValidationError era

The transaction validations errors that can occur from trying to submit a transaction to a local node. The errors are specific to an era.

submitTxToNodeLocal ∷ ∀ mode. LocalNodeConnectInfo mode → TxInMode mode → IO (SubmitResult (TxValidationErrorInMode mode))

data SubmitResult reason

Isomorphic with Maybe but with a name that better describes its purpose and usage.

Constructors

SubmitSuccess

SubmitFail reason

Local state query

newtype LocalStateQueryClient block point (query ∷ Type → Type) (m ∷ Type → Type) a

Constructors

LocalStateQueryClient

Fields

runLocalStateQueryClient ∷ m (ClientStIdle block point query m a)

data AcquiringFailure

Establish a connection to a node and execute a single query using the local state query protocol.

Constructors

AFPointTooOld

AFPointNotOnChain

data QueryInMode mode result where

Constructors

QueryCurrentEra ∷ ConsensusModeIsMultiEra mode → QueryInMode mode AnyCardanoEra

QueryInEra ∷ EraInMode era mode → QueryInEra era result → QueryInMode mode (Either EraMismatch result)

QueryEraHistory ∷ ConsensusModeIsMultiEra mode → QueryInMode mode (EraHistory mode)

QuerySystemStart ∷ QueryInMode mode SystemStart

QueryChainBlockNo ∷ QueryInMode mode (WithOrigin BlockNo)

QueryChainPoint ∷ ConsensusMode mode → QueryInMode mode ChainPoint

data QueryInEra era result where

Constructors

QueryByronUpdateState ∷ QueryInEra ByronEra ByronUpdateState

QueryInShelleyBasedEra ∷ ShelleyBasedEra era → QueryInShelleyBasedEra era result → QueryInEra era result

data QueryInShelleyBasedEra era result where

Constructors

QueryEpoch ∷ QueryInShelleyBasedEra era EpochNo

QueryGenesisParameters ∷ QueryInShelleyBasedEra era (GenesisParameters ShelleyEra)

QueryProtocolParameters ∷ QueryInShelleyBasedEra era (PParams (ShelleyLedgerEra era))

QueryProtocolParametersUpdate ∷ QueryInShelleyBasedEra era (Map (Hash GenesisKey) ProtocolParametersUpdate)

QueryStakeDistribution ∷ QueryInShelleyBasedEra era (Map (Hash StakePoolKey) Rational)

QueryUTxO ∷ QueryUTxOFilter → QueryInShelleyBasedEra era (UTxO era)

QueryStakeAddresses ∷ Set StakeCredential → NetworkId → QueryInShelleyBasedEra era (Map StakeAddress Lovelace, Map StakeAddress PoolId)

QueryStakePools ∷ QueryInShelleyBasedEra era (Set PoolId)

QueryStakePoolParameters ∷ Set PoolId → QueryInShelleyBasedEra era (Map PoolId StakePoolParameters)

QueryDebugLedgerState ∷ QueryInShelleyBasedEra era (SerialisedDebugLedgerState era)

QueryProtocolState ∷ QueryInShelleyBasedEra era (ProtocolState era)

QueryCurrentEpochState ∷ QueryInShelleyBasedEra era (SerialisedCurrentEpochState era)

QueryPoolState ∷ Maybe (Set PoolId) → QueryInShelleyBasedEra era (SerialisedPoolState era)

QueryPoolDistribution ∷ Maybe (Set PoolId) → QueryInShelleyBasedEra era (SerialisedPoolDistribution era)

QueryStakeSnapshot ∷ Maybe (Set PoolId) → QueryInShelleyBasedEra era (SerialisedStakeSnapshots era)

QueryStakeDelegDeposits ∷ Set StakeCredential → QueryInShelleyBasedEra era (Map StakeCredential Lovelace)

QueryConstitution ∷ QueryInShelleyBasedEra era (Maybe (Constitution (ShelleyLedgerEra era)))

QueryGovState ∷ QueryInShelleyBasedEra era (GovState (ShelleyLedgerEra era))

QueryDRepState ∷ Set (Credential DRepRole StandardCrypto) → QueryInShelleyBasedEra era (Map (Credential DRepRole StandardCrypto) (DRepState StandardCrypto))

QueryDRepStakeDistr ∷ Set (DRep StandardCrypto) → QueryInShelleyBasedEra era (Map (DRep StandardCrypto) Lovelace)

QueryCommitteeState ∷ QueryInShelleyBasedEra era (CommitteeState (ShelleyLedgerEra era))

queryNodeLocalState ∷ ∀ mode result. LocalNodeConnectInfo mode → Maybe ChainPoint → QueryInMode mode result → IO (Either AcquiringFailure result)

Local tx monitoring

newtype LocalTxMonitorClient txid tx slot (m ∷ Type → Type) a

A tx monitor client, on top of some effect m.

Constructors

LocalTxMonitorClient

Fields

runLocalTxMonitorClient ∷ m (ClientStIdle txid tx slot m a)

data LocalTxMonitoringQuery mode

Constructors

LocalTxMonitoringQueryTx (TxIdInMode mode)
Query if a particular tx exists in the mempool. Note that, the absence of a transaction does not imply anything about how the transaction was processed: it may have been dropped, or inserted in a block.

LocalTxMonitoringSendNextTx
The mempool is modeled as an ordered list of transactions and thus, can be traversed linearly. LocalTxMonitoringSendNextTx requests the next transaction from the current list. This must be a transaction that was not previously sent to the client for this particular snapshot.

LocalTxMonitoringMempoolInformation
Ask the server about the current mempool's capacity and sizes. This is fixed in a given snapshot.

data LocalTxMonitoringResult mode

Constructors

LocalTxMonitoringTxExists TxId SlotNo
Slot number at which the mempool snapshot was taken

LocalTxMonitoringTxDoesNotExist TxId SlotNo
Slot number at which the mempool snapshot was taken

LocalTxMonitoringNextTx (Maybe (TxInMode mode)) SlotNo
Slot number at which the mempool snapshot was taken

LocalTxMonitoringMempoolSizeAndCapacity MempoolSizeAndCapacity SlotNo
Slot number at which the mempool snapshot was taken

data MempoolSizeAndCapacity

Describes the MemPool sizes and capacity for a given snapshot.

Constructors

MempoolSizeAndCapacity

Fields

capacityInBytes ∷ !Word32

The maximum capacity of the mempool. Note that this may dynamically change when the ledger state is updated.

sizeInBytes ∷ !Word32

The summed byte size of all the transactions in the mempool.

numberOfTxs ∷ !Word32

The number of transactions in the mempool

queryTxMonitoringLocal ∷ ∀ mode. LocalNodeConnectInfo mode → LocalTxMonitoringQuery mode → IO (LocalTxMonitoringResult mode)

data EraHistory mode where

Constructors

EraHistory ∷ ConsensusBlockForMode mode ~ HardForkBlock xs ⇒ ConsensusMode mode → Interpreter xs → EraHistory mode

getProgress ∷ SlotNo → EraHistory mode → Either PastHorizonException (RelativeTime, SlotLength)

Common queries

getLocalChainTip ∷ LocalNodeConnectInfo mode → IO ChainTip

Helpers

data ConsensusMode mode where

This GADT provides a value-level representation of all the consensus modes. This enables pattern matching on the era to allow them to be treated in a non-uniform way.

Constructors

ByronMode ∷ ConsensusMode ByronMode

ShelleyMode ∷ ConsensusMode ShelleyMode

CardanoMode ∷ ConsensusMode CardanoMode

consensusModeOnly ∷ ConsensusModeParams mode → ConsensusMode mode

toAcquiringFailure ∷ AcquireFailure → AcquiringFailure

data NodeToClientVersion

Enumeration of node to client protocol versions.

Constructors

NodeToClientV_9
enabled CardanoNodeToClientVersion7, i.e., Alonzo

NodeToClientV_10
added GetChainBlockNo and GetChainPoint queries

NodeToClientV_11
added GetRewardInfoPools Block query

NodeToClientV_12
added LocalTxMonitor mini-protocol

NodeToClientV_13
enabled CardanoNodeToClientVersion9, i.e., Babbage

NodeToClientV_14
added GetPoolDistr, GetPoolState, GetSnapshots

NodeToClientV_15
added query to NodeToClientVersionData

NodeToClientV_16
enabled CardanoNodeToClientVersion11, i.e., Conway and GetStakeDelegDeposits.

data UnsupportedNtcVersionError

Constructors

UnsupportedNtcVersionError !MinNodeToClientVersion !NodeToClientVersion