Cardano.Api.IPC

Description

Node IPC protocols

Node interaction

Operations that involve talking to a local Cardano node.

connectToLocalNode ∷ LocalNodeConnectInfo mode → LocalNodeClientProtocolsInMode mode → IO ()

Establish a connection to a local node and execute the given set of protocol handlers.

data LocalNodeConnectInfo mode

Constructors

LocalNodeConnectInfo

Fields

localConsensusModeParams ∷ ConsensusModeParams mode

localNodeNetworkId ∷ NetworkId

localNodeSocketPath ∷ FilePath

localConsensusMode ∷ LocalNodeConnectInfo mode → ConsensusMode mode

data LocalNodeClientProtocols block point tip tx txerr query m

Constructors

LocalNodeClientProtocols

Fields

localChainSyncClient ∷ Maybe (ChainSyncClient block point tip m ())

localTxSubmissionClient ∷ Maybe (LocalTxSubmissionClient tx txerr m ())

localStateQueryClient ∷ Maybe (LocalStateQueryClient block point query m ())

type LocalNodeClientProtocolsInMode mode = LocalNodeClientProtocols (BlockInMode mode) ChainPoint ChainTip (TxInMode mode) (TxValidationErrorInMode mode) (QueryInMode mode) IO

Chain sync protocol

newtype ChainSyncClient header point tip (m ∷ Type → Type) a

Constructors

ChainSyncClient

Fields

runChainSyncClient ∷ m (ClientStIdle header point tip m a)

data BlockInMode mode where

A Block in one of the eras supported by a given protocol mode.

For multi-era modes such as the CardanoMode this type is a sum of the different block types for all the eras. It is used in the ChainSync protocol.

Constructors

BlockInMode ∷ Block era → EraInMode era mode → BlockInMode mode

Local tx submission

newtype LocalTxSubmissionClient tx reject (m ∷ Type → Type) a

Constructors

LocalTxSubmissionClient

Fields

runLocalTxSubmissionClient ∷ m (LocalTxClientStIdle tx reject m a)

data TxInMode mode where

A Tx in one of the eras supported by a given protocol mode.

For multi-era modes such as the CardanoMode this type is a sum of the different transaction types for all the eras. It is used in the LocalTxSubmission protocol.

Constructors

TxInMode ∷ Tx era → EraInMode era mode → TxInMode mode
Everything we consider a normal transaction.

TxInByronSpecial ∷ GenTx ByronBlock → EraInMode ByronEra mode → TxInMode mode
Byron has various things we can post to the chain which are not actually transactions. This covers: update proposals, votes and delegation certs.

data TxValidationErrorInMode mode where

A TxValidationError in one of the eras supported by a given protocol mode.

This is used in the LocalStateQuery protocol.

Constructors

TxValidationErrorInMode ∷ TxValidationError era → EraInMode era mode → TxValidationErrorInMode mode

TxValidationEraMismatch ∷ EraMismatch → TxValidationErrorInMode mode

data TxValidationError era

The transaction validations errors that can occur from trying to submit a transaction to a local node. The errors are specific to an era.

submitTxToNodeLocal ∷ ∀ mode. LocalNodeConnectInfo mode → TxInMode mode → IO (SubmitResult (TxValidationErrorInMode mode))

data SubmitResult reason

Constructors

SubmitSuccess

SubmitFail reason

Local state query

newtype LocalStateQueryClient block point (query ∷ Type → Type) (m ∷ Type → Type) a

Constructors

LocalStateQueryClient

Fields

runLocalStateQueryClient ∷ m (ClientStIdle block point query m a)

data QueryInMode mode result where

Constructors

QueryCurrentEra ∷ ConsensusModeIsMultiEra mode → QueryInMode mode AnyCardanoEra

QueryInEra ∷ EraInMode era mode → QueryInEra era result → QueryInMode mode (Either EraMismatch result)

data QueryInEra era result where

Constructors

QueryByronUpdateState ∷ QueryInEra ByronEra ByronUpdateState

QueryInShelleyBasedEra ∷ ShelleyBasedEra era → QueryInShelleyBasedEra era result → QueryInEra era result

data QueryInShelleyBasedEra era result where

Constructors

QueryChainPoint ∷ QueryInShelleyBasedEra era ChainPoint

QueryEpoch ∷ QueryInShelleyBasedEra era EpochNo

QueryGenesisParameters ∷ QueryInShelleyBasedEra era GenesisParameters

QueryProtocolParameters ∷ QueryInShelleyBasedEra era ProtocolParameters

QueryProtocolParametersUpdate ∷ QueryInShelleyBasedEra era (Map (Hash GenesisKey) ProtocolParametersUpdate)

QueryStakeDistribution ∷ QueryInShelleyBasedEra era (Map (Hash StakePoolKey) Rational)

QueryUTxO ∷ Maybe (Set AddressAny) → QueryInShelleyBasedEra era (UTxO era)

QueryStakeAddresses ∷ Set StakeCredential → QueryInShelleyBasedEra era (Map StakeCredential Lovelace, Map StakeCredential PoolId)

queryNodeLocalState ∷ ∀ mode result. LocalNodeConnectInfo mode → ChainPoint → QueryInMode mode result → IO (Either AcquireFailure result)

Establish a connection to a node and execute a single query using the local state query protocol.