Cardano.Api.IPC.Monad

Documentation

data LocalStateQueryExpr block point query r m a

Monadic type for constructing local state query expressions.

Use queryExpr in a do block to construct queries of this type and convert the expression to a LocalStateQueryClient with setupLocalStateQueryExpr.

Some consideration was made to use Applicative instead of Monad as the abstraction in order to support pipelining, but we actually have a fair amount of code where the next query depends on the result of the former and therefore actually need Monad.

In order to make pipelining still possible we can explore the use of Selective Functors which would allow us to straddle both worlds.

executeLocalStateQueryExpr ∷ LocalNodeConnectInfo mode → Maybe ChainPoint → LocalStateQueryExpr (BlockInMode mode) ChainPoint (QueryInMode mode) () IO a → IO (Either AcquiringFailure a)

Execute a local state query expression.

queryExpr ∷ QueryInMode mode a → LocalStateQueryExpr block point (QueryInMode mode) r IO (Either UnsupportedNtcVersionError a)

Use queryExpr in a do block to construct monadic local state queries.

determineEraExpr ∷ ConsensusModeParams mode → LocalStateQueryExpr block point (QueryInMode mode) r IO (Either UnsupportedNtcVersionError AnyCardanoEra)

A monad expression that determines what era the node is in.