Cardano.Api.IPC.Monad

Documentation

data LocalStateQueryScript block point query r m a

Monadic type for constructing local state queries.

Use sendMsgQuery in a do block to construct queries of this type and convert the expression to a LocalStateQueryClient with setupLocalStateQueryScript.

sendMsgQuery ∷ Monad m ⇒ query a → LocalStateQueryScript block point query r m a

Use sendMsgQuery in a do block to construct monadic local state queries.

setupLocalStateQueryScript ∷ STM x → Maybe ChainPoint → NodeToClientVersion → TMVar (Maybe (Either AcquireFailure a)) → LocalStateQueryScript (BlockInMode CardanoMode) ChainPoint (QueryInMode CardanoMode) () IO a → LocalStateQueryClient (BlockInMode CardanoMode) ChainPoint (QueryInMode CardanoMode) IO ()

Use sendMsgQuery in a do block to construct monadic local state queries.