Cardano.Api.LedgerEvent

Documentation

data LedgerEvent

Constructors

PoolRegistration
The given pool is being registered for the first time on chain.

PoolReRegistration
The given pool already exists and is being re-registered.

IncrementalRewardsDistribution EpochNo (Map StakeCredential (Set (Reward StandardCrypto)))
Incremental rewards are being computed.

RewardsDistribution EpochNo (Map StakeCredential (Set (Reward StandardCrypto)))
Reward distribution has completed.

MIRDistribution MIRDistributionDetails
MIR are being distributed.

PoolReap PoolReapDetails
Pools have been reaped and deposits refunded.

SuccessfulPlutusScript (NonEmpty PlutusDebug)
A number of succeeded Plutus script evaluations.

FailedPlutusScript (NonEmpty PlutusDebug)
A number of failed Plutus script evaluations.

NewGovernanceProposals (TxId StandardCrypto) AnyProposals
Newly submittted governance proposals in a single transaction.

EpochBoundaryRatificationState AnyRatificationState
The current state of governance matters at the epoch boundary. I.E the current constitution, committee, protocol parameters, etc.

data MIRDistributionDetails

Details of fund transfers due to MIR certificates.

Note that the transfers from reserves to treasury and treasury to reserves are inverse; a transfer of 100 ADA in either direction will result in a net movement of 0, but we include both directions for assistance in debugging.

Constructors

MIRDistributionDetails

Fields

mirddReservePayouts :: Map StakeCredential Lovelace

mirddTreasuryPayouts :: Map StakeCredential Lovelace

mirddReservesToTreasury :: Lovelace

mirddTreasuryToReserves :: Lovelace

data PoolReapDetails

Constructors

PoolReapDetails

Fields

prdEpochNo :: EpochNo

prdRefunded :: Map StakeCredential (Map (Hash StakePoolKey) Lovelace)

Refunded deposits. The pools referenced are now retired, and the StakeCredential accounts are credited with the deposits.

prdUnclaimed :: Map StakeCredential (Map (Hash StakePoolKey) Lovelace)

Unclaimed deposits. The StakeCredential referenced in this map is not actively registered at the time of the pool reaping, and as such the funds are returned to the treasury.

toLedgerEvent :: ConvertLedgerEvent blk => WrapLedgerEvent blk -> Maybe LedgerEvent