Cardano.Api.LedgerState

Initialization / Accumulation

envSecurityParam :: Env -> Word64

data LedgerState

Constructors

LedgerState

Fields

clsState :: CardanoLedgerState StandardCrypto EmptyMK

clsTables :: LedgerTables (LedgerState (HardForkBlock (CardanoEras StandardCrypto))) ValuesMK

Bundled Patterns

pattern LedgerStateByron :: LedgerState ByronBlock EmptyMK -> LedgerState

Deprecated: Use 'LedgerState $ Consensus.LedgerStateByron' instead

pattern LedgerStateShelley :: LedgerState StandardShelleyBlock EmptyMK -> LedgerState

Deprecated: Use 'LedgerState $ Consensus.LedgerStateShelley' instead

pattern LedgerStateAllegra :: LedgerState StandardAllegraBlock EmptyMK -> LedgerState

Deprecated: Use 'LedgerState $ Consensus.LedgerStateAllegra' instead

pattern LedgerStateMary :: LedgerState StandardMaryBlock EmptyMK -> LedgerState

Deprecated: Use 'LedgerState $ Consensus.LedgerStateMary' instead

pattern LedgerStateAlonzo :: LedgerState StandardAlonzoBlock EmptyMK -> LedgerState

Deprecated: Use 'LedgerState $ Consensus.LedgerAlonzo' instead

pattern LedgerStateBabbage :: LedgerState StandardBabbageBlock EmptyMK -> LedgerState

Deprecated: Use 'LedgerState $ Consensus.LedgerBabbage' instead

pattern LedgerStateConway :: LedgerState StandardConwayBlock EmptyMK -> LedgerState

Deprecated: Use 'LedgerState $ Consensus.LedgerConway' instead

encodeLedgerState :: LedgerState -> Encoding

decodeLedgerState :: Decoder s LedgerState

initialLedgerState

Arguments

:: forall t (m :: Type -> Type). MonadIOTransError InitialLedgerStateError t m

=> NodeConfigFile 'In
Path to the cardano-node config file (e.g. to cardano-node projectconfigurationcardano/mainnet-config.json)

-> t m (Env, LedgerState)
The environment and initial ledger state

Get the environment and initial ledger state.

applyBlock

Arguments

:: Env
The environment returned by initialLedgerState

-> LedgerState
The current ledger state

-> ValidationMode

-> BlockInMode
Some block to apply

-> Either LedgerStateError (LedgerState, [LedgerEvent])
The new ledger state (or an error).

Apply a single block to the current ledger state.

data ValidationMode

How to do validation when applying a block to a ledger state.

Constructors

FullValidation
Do all validation implied by the ledger layer's applyBlock.

QuickValidation
Only check that the previous hash from the block matches the head hash of the ledger state.

applyBlockWithEvents

Arguments

:: Env

-> LedgerState

-> Bool
True to validate

-> CardanoBlock StandardCrypto

-> Either LedgerStateError LedgerStateEvents

data AnyNewEpochState where

Constructors

AnyNewEpochState :: forall era. ShelleyBasedEra era -> NewEpochState (ShelleyLedgerEra era) -> LedgerTables (LedgerState (CardanoBlock StandardCrypto)) ValuesMK -> AnyNewEpochState

getAnyNewEpochState :: ShelleyBasedEra era -> LedgerState -> Either LedgerStateError AnyNewEpochState

Retrieve new epoch state from the ledger state, or an error on failure

getLedgerTablesUTxOValues :: ShelleyBasedEra era -> LedgerTables (LedgerState (CardanoBlock StandardCrypto)) ValuesMK -> Map TxIn (TxOut CtxUTxO era)

Traversing the block chain

foldBlocks

Arguments

:: forall a t (m :: Type -> Type). (Show a, MonadIOTransError FoldBlocksError t m)

=> NodeConfigFile 'In
Path to the cardano-node config file (e.g. to cardano-node projectconfigurationcardano/mainnet-config.json)

-> SocketPath
Path to local cardano-node socket. This is the path specified by the --socket-path command line option when running the node.

-> ValidationMode

-> a
The initial accumulator state.

-> (Env -> LedgerState -> [LedgerEvent] -> BlockInMode -> a -> IO (a, FoldStatus))
Accumulator function Takes:

Environment (this is a constant over the whole fold).

The Ledger state (with block i applied) at block i.

The Ledger events resulting from applying block i.

Block i.

The accumulator state at block i - 1.

And returns:

The accumulator state at block i

A type indicating whether to stop or continue folding.

Note: This function can safely assume no rollback will occur even though internally this is implemented with a client protocol that may require rollback. This is achieved by only calling the accumulator on states/blocks that are older than the security parameter, k. This has the side effect of truncating the last k blocks before the node's tip.

-> t m a
The final state

Monadic fold over all blocks and ledger states. Stopping k blocks before the node's tip where k is the security parameter.

data FoldStatus

Type that lets us decide whether to continue or stop the fold from within our accumulation function.

Constructors

ContinueFold

StopFold

DebugFold

chainSyncClientWithLedgerState

Arguments

:: forall (m :: Type -> Type) a. Monad m

=> Env

-> LedgerState
Initial ledger state

-> ValidationMode

-> ChainSyncClient (BlockInMode, Either LedgerStateError (LedgerState, [LedgerEvent])) ChainPoint ChainTip m a
A client to wrap. The block is annotated with a 'Either LedgerStateError LedgerState'. This is either an error from validating a block or the current LedgerState from applying the current block. If we trust the node, then we generally expect blocks to validate. Also note that after a block fails to validate we may still roll back to a validated block, in which case the valid LedgerState will be passed here again.

-> ChainSyncClient BlockInMode ChainPoint ChainTip m a
A client that acts just like the wrapped client but doesn't require the LedgerState annotation on the block type.

Wrap a ChainSyncClient with logic that tracks the ledger state.

chainSyncClientPipelinedWithLedgerState :: forall (m :: Type -> Type) a. Monad m => Env -> LedgerState -> ValidationMode -> ChainSyncClientPipelined (BlockInMode, Either LedgerStateError (LedgerState, [LedgerEvent])) ChainPoint ChainTip m a -> ChainSyncClientPipelined BlockInMode ChainPoint ChainTip m a

See chainSyncClientWithLedgerState.

Ledger state conditions

data ConditionResult

Type isomorphic to bool, representing condition check result

Constructors

ConditionNotMet

ConditionMet

fromConditionResult :: ConditionResult -> Bool

toConditionResult :: Bool -> ConditionResult

foldEpochState

Arguments

:: forall t (m :: Type -> Type) s. MonadIOTransError FoldBlocksError t m

=> NodeConfigFile 'In
Path to the cardano-node config file (e.g. to cardano-node projectconfigurationcardano/mainnet-config.json)

-> SocketPath
Path to local cardano-node socket. This is the path specified by the --socket-path command line option when running the node.

-> ValidationMode

-> EpochNo
Termination epoch

-> s
an initial value for the condition state

-> (AnyNewEpochState -> SlotNo -> BlockNo -> StateT s IO ConditionResult)
Condition you want to check against the new epoch state.

SlotNo - Current (not necessarily the tip) slot number

BlockNo - Current (not necessarily the tip) block number

Note: This function can safely assume no rollback will occur even though internally this is implemented with a client protocol that may require rollback. This is achieved by only calling the accumulator on states/blocks that are older than the security parameter, k. This has the side effect of truncating the last k blocks before the node's tip.

-> t m (ConditionResult, s)
The final state

Reconstructs the ledger's new epoch state and applies a supplied condition to it for every block. This function only terminates if the condition is met or we have reached the termination epoch. We need to provide a termination epoch otherwise blocks would be applied indefinitely.

Ledger events

data LedgerEvent

Constructors

PoolRegistration
The given pool is being registered for the first time on chain.

PoolReRegistration
The given pool already exists and is being re-registered.

IncrementalRewardsDistribution EpochNo (Map StakeCredential (Set Reward))
Incremental rewards are being computed.

RewardsDistribution EpochNo (Map StakeCredential (Set Reward))
Reward distribution has completed.

MIRDistribution MIRDistributionDetails
MIR are being distributed.

PoolReap PoolReapDetails
Pools have been reaped and deposits refunded.

SuccessfulPlutusScript (NonEmpty PlutusWithContext)
A number of succeeded Plutus script evaluations.

FailedPlutusScript (NonEmpty PlutusWithContext)
A number of failed Plutus script evaluations.

NewGovernanceProposals TxId AnyProposals
Newly submittted governance proposals in a single transaction.

RemovedGovernanceVotes
Governance votes that were invalidated.

Fields

TxId

(Set (Voter, GovActionId))

Votes that were replaced in this tx.

(Set (Credential 'DRepRole))

Any votes from these DReps in this or in previous txs are removed

EpochBoundaryRatificationState AnyRatificationState
The current state of governance matters at the epoch boundary. I.E the current constitution, committee, protocol parameters, etc.

data AnyProposals

Constructors

EraPParams era => AnyProposals (Proposals era)

data AnyRatificationState

Constructors

EraPParams era => AnyRatificationState (RatifyState era)

data MIRDistributionDetails

Details of fund transfers due to MIR certificates.

Note that the transfers from reserves to treasury and treasury to reserves are inverse; a transfer of 100 ADA in either direction will result in a net movement of 0, but we include both directions for assistance in debugging.

Constructors

MIRDistributionDetails

Fields

mirddReservePayouts :: Map StakeCredential Coin

mirddTreasuryPayouts :: Map StakeCredential Coin

mirddReservesToTreasury :: Coin

mirddTreasuryToReserves :: Coin

data PoolReapDetails

Constructors

PoolReapDetails

Fields

prdEpochNo :: EpochNo

prdRefunded :: Map StakeCredential (Map (Hash StakePoolKey) Coin)

Refunded deposits. The pools referenced are now retired, and the StakeCredential accounts are credited with the deposits.

prdUnclaimed :: Map StakeCredential (Map (Hash StakePoolKey) Coin)

Unclaimed deposits. The StakeCredential referenced in this map is not actively registered at the time of the pool reaping, and as such the funds are returned to the treasury.

convertRetiredPoolsMap :: Map StakeCredential (Map (KeyHash 'StakePool) Coin) -> Map StakeCredential (Map (Hash StakePoolKey) Coin)

Errors

data LedgerStateError

Constructors

ApplyBlockHashMismatch Text
When using QuickValidation, the block hash did not match the expected block hash after applying a new block to the current ledger state.

ApplyBlockError (CardanoLedgerError StandardCrypto)
When using FullValidation, an error occurred when applying a new block to the current ledger state.

InvalidRollback
Encountered a rollback larger than the security parameter.

Fields

SlotNo

Oldest known slot number that we can roll back to.

ChainPoint

Rollback was attempted to this point.

TerminationEpochReached EpochNo
The ledger state condition you were interested in was not met prior to the termination epoch.

UnexpectedLedgerState

Fields

AnyShelleyBasedEra

Expected era

(NS (Current (Flip LedgerState EmptyMK)) (CardanoEras StandardCrypto))

Ledgerstate from an unexpected era

ByronEraUnsupported

DebugError !String

data FoldBlocksError

Constructors

FoldBlocksInitialLedgerStateError !InitialLedgerStateError

FoldBlocksApplyBlockError !LedgerStateError

FoldBlocksIOException !IOException

FoldBlocksMuxError !Error

data GenesisConfigError

Constructors

NEError !Text

NEByronConfig !FilePath !ConfigurationError

NEShelleyConfig !FilePath !Text

NEAlonzoConfig !FilePath !Text

NEConwayConfig !FilePath !Text

NECardanoConfig !Text

data InitialLedgerStateError

Constructors

ILSEConfigFile Text
Failed to read or parse the network config file.

ILSEGenesisFile GenesisConfigError
Failed to read or parse a genesis file linked from the network config file.

ILSELedgerConsensusConfig GenesisConfigError
Failed to derive the Ledger or Consensus config.

Leadership schedule

data LeadershipError

Constructors

LeaderErrDecodeLedgerStateFailure

LeaderErrDecodeProtocolStateFailure (ByteString, DecoderError)

LeaderErrDecodeProtocolEpochStateFailure DecoderError

LeaderErrGenesisSlot

LeaderErrStakePoolHasNoStake PoolId

LeaderErrStakeDistribUnstable

Fields

SlotNo

Current slot

SlotNo

Stable after

SlotNo

Stability window size

SlotNo

Predicted last slot of the epoch

LeaderErrSlotRangeCalculationFailure Text

LeaderErrCandidateNonceStillEvolving

constructGlobals :: ShelleyGenesis -> EpochInfo (Either Text) -> Globals

currentEpochEligibleLeadershipSlots

Arguments

:: ShelleyBasedEra era

-> ShelleyGenesis

-> EpochInfo (Either Text)

-> PParams (ShelleyLedgerEra era)

-> ProtocolState era

-> PoolId

-> SigningKey VrfKey

-> SerialisedPoolDistribution era

-> EpochNo
Current EpochInfo

-> Either LeadershipError (Set SlotNo)

Return the slots at which a particular stake pool operator is expected to mint a block.

nextEpochEligibleLeadershipSlots

Arguments

:: ShelleyBasedEra era

-> ShelleyGenesis

-> SerialisedCurrentEpochState era
We need the mark stake distribution in order to predict the following epoch's leadership schedule

-> ProtocolState era

-> PoolId
Potential slot leading stake pool

-> SigningKey VrfKey
VRF signing key of the stake pool

-> PParams (ShelleyLedgerEra era)

-> EpochInfo (Either Text)

-> (ChainTip, EpochNo)

-> Either LeadershipError (Set SlotNo)

Node Config

data NodeConfig

Constructors

NodeConfig

Fields

ncPBftSignatureThreshold :: !(Maybe Double)

ncByronGenesisFile :: !(File ByronGenesisConfig 'In)

ncByronGenesisHash :: !GenesisHashByron

ncShelleyGenesisFile :: !(File ShelleyGenesisConfig 'In)

ncShelleyGenesisHash :: !GenesisHashShelley

ncAlonzoGenesisFile :: !(File AlonzoGenesis 'In)

ncAlonzoGenesisHash :: !GenesisHashAlonzo

ncConwayGenesisFile :: !(Maybe (File ConwayGenesisConfig 'In))

ncConwayGenesisHash :: !(Maybe GenesisHashConway)

ncRequiresNetworkMagic :: !RequiresNetworkMagic

ncByronProtocolVersion :: !ProtocolVersion

ncHardForkTriggers :: !CardanoHardForkTriggers

Network Config

type NodeConfigFile = File NodeConfig

readNodeConfig :: (MonadError Text m, MonadIO m) => NodeConfigFile 'In -> m NodeConfig

Genesis Config

data GenesisConfig

Constructors

GenesisCardano !NodeConfig !Config !GenesisHashShelley !(TransitionConfig LatestKnownEra)

readCardanoGenesisConfig

Arguments

:: forall t (m :: Type -> Type) era. MonadIOTransError GenesisConfigError t m

=> Maybe (CardanoEra era)
Provide era witness to read Alonzo Genesis in an era-sensitive manner (see decodeAlonzGenesis for more details)

-> NodeConfig

-> t m GenesisConfig

mkProtocolInfoCardano :: GenesisConfig -> (ProtocolInfo (CardanoBlock StandardCrypto), IO [BlockForging IO (CardanoBlock StandardCrypto)])

Byron Genesis Config

readByronGenesisConfig :: forall t (m :: Type -> Type). MonadIOTransError GenesisConfigError t m => NodeConfig -> t m Config

Shelley Genesis Config

data ShelleyConfig

Constructors

ShelleyConfig

Fields

scConfig :: !ShelleyGenesis

scGenesisHash :: !GenesisHashShelley

newtype GenesisHashShelley

Constructors

GenesisHashShelley

Fields

unGenesisHashShelley :: Hash Blake2b_256 ByteString

readShelleyGenesisConfig :: forall t (m :: Type -> Type). MonadIOTransError GenesisConfigError t m => NodeConfig -> t m ShelleyConfig

shelleyPraosNonce :: GenesisHashShelley -> Nonce

Compute the Nonce from the hash of the Genesis file.

Alonzo Genesis Config

newtype GenesisHashAlonzo

Constructors

GenesisHashAlonzo

Fields

unGenesisHashAlonzo :: Hash Blake2b_256 ByteString

readAlonzoGenesisConfig

Arguments

:: forall t (m :: Type -> Type) era. MonadIOTransError GenesisConfigError t m

=> Maybe (CardanoEra era)
Provide era witness to read Alonzo Genesis in an era-sensitive manner (see decodeAlonzGenesis for more details)

-> NodeConfig

-> t m AlonzoGenesis

Conway Genesis Config

newtype GenesisHashConway

Constructors

GenesisHashConway

Fields

unGenesisHashConway :: Hash Blake2b_256 ByteString

readConwayGenesisConfig :: forall t (m :: Type -> Type). MonadIOTransError GenesisConfigError t m => NodeConfig -> t m ConwayGenesis

If the conway genesis file does not exist we simply put in a default.

Environment

data Env

Constructors

Env

Fields

envLedgerConfig :: CardanoLedgerConfig StandardCrypto

envConsensusConfig :: CardanoConsensusConfig StandardCrypto

genesisConfigToEnv :: GenesisConfig -> Either GenesisConfigError Env