Cardano.Api.LedgerState

Initialization / Accumulation

data Env

Constructors

Env

Fields

envLedgerConfig ∷ HardForkLedgerConfig (CardanoEras StandardCrypto)

envProtocolConfig ∷ ConsensusConfig (HardForkProtocol (CardanoEras StandardCrypto))

envSecurityParam ∷ Env → Word64

newtype LedgerState

Constructors

LedgerState

Fields

clsState ∷ LedgerState (HardForkBlock (CardanoEras StandardCrypto))

Bundled Patterns

pattern LedgerStateByron ∷ LedgerState ByronBlock → LedgerState

pattern LedgerStateShelley ∷ LedgerState (ShelleyBlock (ShelleyEra StandardCrypto)) → LedgerState

pattern LedgerStateAllegra ∷ LedgerState (ShelleyBlock (AllegraEra StandardCrypto)) → LedgerState

pattern LedgerStateMary ∷ LedgerState (ShelleyBlock (MaryEra StandardCrypto)) → LedgerState

initialLedgerState

Arguments

∷ FilePath
Path to the cardano-node config file (e.g. to cardano-node projectconfigurationcardano/mainnet-config.json)

→ ExceptT InitialLedgerStateError IO (Env, LedgerState)
The environment and initial ledger state

Get the environment and initial ledger state.

applyBlock

Arguments

∷ Env
The environment returned by initialLedgerState

→ LedgerState
The current ledger state

→ Bool
True to perform validation. If True, tickThenApply will be used instead of tickThenReapply.

→ Block era
Some block to apply

→ Either Text LedgerState
The new ledger state (or an error).

Apply a single block to the current ledger state.

Traversing the block chain

foldBlocks

Arguments

∷ ∀ a. FilePath
Path to the cardano-node config file (e.g. to cardano-node projectconfigurationcardano/mainnet-config.json)

→ FilePath
Path to local cardano-node socket. This is the path specified by the --socket-path command line option when running the node.

→ NetworkId
The network ID.

→ Bool
True to enable validation. Under the hood this will use applyBlock instead of reapplyBlock from the ApplyBlock type class.

→ a
The initial accumulator state.

→ (Env → LedgerState → BlockInMode CardanoMode → a → IO a)
Accumulator function Takes: * Environment (this is a constant over the whole fold) * The current Ledger state (with the current block applied) * The current Block * The previous state

And this should return the new state.

Note: This function can safely assume no rollback will occur even though internally this is implemented with a client protocol that may require rollback. This is achieved by only calling the accumulator on states/blocks that are older than the security parameter, k. This has the side effect of truncating the last k blocks before the node's tip.

→ ExceptT FoldBlocksError IO a
The final state

Monadic fold over all blocks and ledger states. Stopping k blocks before the node's tip where k is the security parameter.

Errors

data FoldBlocksError

Constructors

FoldBlocksInitialLedgerStateError InitialLedgerStateError

FoldBlocksApplyBlockError Text

data GenesisConfigError

Constructors

NEError !Text

NEByronConfig !FilePath !ConfigurationError

NEShelleyConfig !FilePath !Text

NECardanoConfig !Text

data InitialLedgerStateError

Constructors

ILSEConfigFile Text
Failed to read or parse the network config file.

ILSEGenesisFile GenesisConfigError
Failed to read or parse a genesis file linked from the network config file.

ILSELedgerConsensusConfig GenesisConfigError
Failed to derive the Ledger or Consensus config.

renderFoldBlocksError ∷ FoldBlocksError → Text

renderGenesisConfigError ∷ GenesisConfigError → Text

renderInitialLedgerStateError ∷ InitialLedgerStateError → Text