Cardano.Api.Monad.Error

Description

This module serves purpose as a single source of abstractions used in handling MonadError and ExceptT through 'cardano-api'.

Documentation

type MonadTransError e (t :: (Type -> Type) -> Type -> Type) (m :: Type -> Type) = (Monad m, MonadTrans t, MonadError e (t m))

Convenience alias

type MonadIOTransError e (t :: (Type -> Type) -> Type -> Type) (m :: Type -> Type) = (MonadIO m, MonadIO (t m), MonadCatch m, MonadTrans t, MonadError e (t m))

Same as MonadTransError, but with also MonadIO constraint

liftExceptT :: forall e t (m :: Type -> Type) a. MonadTransError e t m => ExceptT e m a -> t m a

Lift ExceptT into MonadTransError

modifyError

Arguments

:: forall e' t (m :: Type -> Type) e a. MonadTransError e' t m

=> (e -> e')
mapping function

-> ExceptT e m a
value

-> t m a
result with modified error

Modify an ExceptT error and lift it to MonadError transformer stack.

This implementation avoids nesting problem of modifyError from mtl. The issue with modifyError (from mtl) is that when you use it on a function, you actually end up with ExceptT e1 (ExceptT e2 m a):

modifyError (f :: e2 -> e1) (foo :: ExceptT e2 (ExceptT e1 m) a) :: ExceptT e1 m a

and if you use modifyError (mtl) again, the more nested you get e.g. ExceptT e1 (ExceptT e2 (ExceptT e3 m a)). With a deeper monad stack you pay the overhead with every use of >>=.

This function avoids that, but at the cost of limiting its application to transformers.

handleIOExceptionsWith

Arguments

:: (MonadError e' m, MonadCatch m, Exception e)

=> (e -> e')
mapping function

-> m a
action that can throw

-> m a
result with caught exception

Wrap an exception and lift it into MonadError.

handleIOExceptionsLiftWith

Arguments

:: (MonadIOTransError e' t m, Exception e)

=> (e -> e')
mapping function

-> m a
action that can throw

-> t m a
action with caught error lifted into MonadError stack

Wrap an exception and lift it into MonadError stack.

hoistIOEither :: forall e t (m :: Type -> Type) a. MonadIOTransError e t m => IO (Either e a) -> t m a

Lift an IO action that returns Either into MonadIOTransError

liftMaybe

Arguments

:: MonadError e m

=> e
Error to throw, if Nothing

-> Maybe a

-> m a

Lift Maybe into MonadError

failEither :: MonadFail m => Either String a -> m a

Lift Either to MonadFail

failEitherWith

Arguments

:: MonadFail m

=> (e -> String)
convert Left to String

-> Either e a

-> m a

Lift Either to MonadFail

(?!) :: MonadError e m => Maybe a -> e -> m a infixl 8

Infix liftMaybe, with arguments flipped

(?!&) :: Either e a -> (e -> e') -> Either e' a infixl 8

Map over the Left value of 'Either e a'. Infix first with its arguments flipped.

newtype ExceptT e (m :: Type -> Type) a

A monad transformer that adds exceptions to other monads.

ExceptT constructs a monad parameterized over two things:

e - The exception type.

m - The inner monad.

The return function yields a computation that produces the given value, while >>= sequences two subcomputations, exiting on the first exception.

Constructors

ExceptT (m (Either e a))

class Monad m => MonadError e (m :: Type -> Type) | m -> e where

The strategy of combining computations that can throw exceptions by bypassing bound functions from the point an exception is thrown to the point that it is handled.

Is parameterized over the type of error information and the monad type constructor. It is common to use Either String as the monad type constructor for an error monad in which error descriptions take the form of strings. In that case and many other common cases the resulting monad is already defined as an instance of the MonadError class. You can also define your own error type and/or use a monad type constructor other than Either String or Either IOError. In these cases you will have to explicitly define instances of the MonadError class. (If you are using the deprecated Control.Monad.Error or Control.Monad.Trans.Error, you may also have to define an Error instance.)

Methods

throwError :: e -> m a

Is used within a monadic computation to begin exception processing.

catchError :: m a -> (e -> m a) -> m a

A handler function to handle previous errors and return to normal execution. A common idiom is:

do { action1; action2; action3 } `catchError` handler

where the action functions can call throwError. Note that handler and the do-block must have the same return type.

runExcept :: Except e a -> Either e a

Extractor for computations in the exception monad. (The inverse of except).

mapExcept :: (Either e a -> Either e' b) -> Except e a -> Except e' b

Map the unwrapped computation using the given function.

runExcept (mapExcept f m) = f (runExcept m)

withExcept :: (e -> e') -> Except e a -> Except e' a

Transform any exceptions thrown by the computation using the given function (a specialization of withExceptT).

runExceptT :: ExceptT e m a -> m (Either e a)

The inverse of ExceptT.

mapExceptT :: (m (Either e a) -> n (Either e' b)) -> ExceptT e m a -> ExceptT e' n b

Map the unwrapped computation using the given function.

runExceptT (mapExceptT f m) = f (runExceptT m)

liftEither :: MonadError e m => Either e a -> m a

Lifts an Either e into any MonadError e.

do { val <- liftEither =<< action1; action2 }

where action1 returns an Either to represent errors.

Since: mtl-2.2.2

module Control.Monad.IO.Class

module Control.Monad.Trans.Class

module Control.Monad.Trans.Except

left :: forall (m :: Type -> Type) x a. Monad m => x -> ExceptT x m a

right :: forall (m :: Type -> Type) a x. Monad m => a -> ExceptT x m a

runExceptT :: ExceptT e m a -> m (Either e a)

The inverse of ExceptT.

mapExceptT :: (m (Either e a) -> n (Either e' b)) -> ExceptT e m a -> ExceptT e' n b

Map the unwrapped computation using the given function.

runExceptT (mapExceptT f m) = f (runExceptT m)

hoistMaybe :: forall (m :: Type -> Type) x a. Monad m => x -> Maybe a -> ExceptT x m a

handleIOExceptT :: forall (m :: Type -> Type) x a. MonadIO m => (IOException -> x) -> IO a -> ExceptT x m a

firstExceptT :: forall (m :: Type -> Type) x y a. Functor m => (x -> y) -> ExceptT x m a -> ExceptT y m a

hoistEither :: forall (m :: Type -> Type) x a. Monad m => Either x a -> ExceptT x m a

newExceptT :: m (Either x a) -> ExceptT x m a

bimapExceptT :: forall (m :: Type -> Type) x y a b. Functor m => (x -> y) -> (a -> b) -> ExceptT x m a -> ExceptT y m b

bracketExceptT :: forall (m :: Type -> Type) e a b c. Monad m => ExceptT e m a -> (a -> ExceptT e m b) -> (a -> ExceptT e m c) -> ExceptT e m c

bracketExceptionT :: forall (m :: Type -> Type) e a c b. MonadMask m => ExceptT e m a -> (a -> ExceptT e m c) -> (a -> ExceptT e m b) -> ExceptT e m b

catchExceptT :: (MonadCatch m, Exception e) => m a -> (e -> x) -> ExceptT x m a

catchIOExceptT :: forall (m :: Type -> Type) a x. MonadIO m => IO a -> (IOException -> x) -> ExceptT x m a

catchLeftT :: forall (m :: Type -> Type) e a. Monad m => ExceptT e m a -> (e -> ExceptT e m a) -> ExceptT e m a

catchesExceptT :: (Foldable f, MonadCatch m) => m a -> f (Handler m x) -> ExceptT x m a

exceptT :: Monad m => (x -> m b) -> (a -> m b) -> ExceptT x m a -> m b

handleExceptT :: (MonadCatch m, Exception e) => (e -> x) -> m a -> ExceptT x m a

handleLeftT :: forall (m :: Type -> Type) e a. Monad m => (e -> ExceptT e m a) -> ExceptT e m a -> ExceptT e m a

handlesExceptT :: (Foldable f, MonadCatch m) => f (Handler m x) -> m a -> ExceptT x m a

hoistExceptT :: (forall b. m b -> n b) -> ExceptT x m a -> ExceptT x n a

hushM :: Monad m => Either e a -> (e -> m ()) -> m (Maybe a)

onLeft :: forall e x (m :: Type -> Type) a. Monad m => (e -> ExceptT x m a) -> ExceptT x m (Either e a) -> ExceptT x m a

onNothing :: forall x (m :: Type -> Type) a. Monad m => ExceptT x m a -> ExceptT x m (Maybe a) -> ExceptT x m a

secondExceptT :: forall (m :: Type -> Type) a b x. Functor m => (a -> b) -> ExceptT x m a -> ExceptT x m b