Cardano.Api.Plutus

Languages

data SimpleScript'

data PlutusScriptV1

The original simple script language which supports

require a signature from a given key (by verification key hash)

n-way and combinator

n-way or combinator

m-of-n combinator

This version of the language was introduced in the ShelleyEra.

The second version of the simple script language. It has all the features of the original simple script language plus new atomic predicates:

require the time be before a given slot number

require the time be after a given slot number

This version of the language was introduced in the AllegraEra.

However we opt for a single type level tag SimpleScript' as the second version of of the language introduced in the Allegra era is a superset of the language introduced in the Shelley era.

Place holder type to show what the pattern is to extend to multiple languages, not just multiple versions of a single language.

data PlutusScriptV2

data PlutusScriptV3

data ScriptLanguage lang where

Constructors

SimpleScriptLanguage :: ScriptLanguage SimpleScript'

PlutusScriptLanguage :: forall lang. IsPlutusScriptLanguage lang => PlutusScriptVersion lang -> ScriptLanguage lang

data PlutusScriptVersion lang where

Constructors

PlutusScriptV1 :: PlutusScriptVersion PlutusScriptV1

PlutusScriptV2 :: PlutusScriptVersion PlutusScriptV2

PlutusScriptV3 :: PlutusScriptVersion PlutusScriptV3

data AnyScriptLanguage where

Constructors

AnyScriptLanguage :: forall lang. ScriptLanguage lang -> AnyScriptLanguage

data AnyPlutusScriptVersion where

Constructors

AnyPlutusScriptVersion :: forall lang. IsPlutusScriptLanguage lang => PlutusScriptVersion lang -> AnyPlutusScriptVersion

class IsScriptLanguage lang => IsPlutusScriptLanguage lang where

Methods

plutusScriptVersion :: PlutusScriptVersion lang

class HasTypeProxy lang => IsScriptLanguage lang where

Methods

scriptLanguage :: ScriptLanguage lang

type family ToLedgerPlutusLanguage lang :: Language where ...

Equations

ToLedgerPlutusLanguage PlutusScriptV1 = 'PlutusV1

ToLedgerPlutusLanguage PlutusScriptV2 = 'PlutusV2

ToLedgerPlutusLanguage PlutusScriptV3 = 'PlutusV3

Scripts in a specific language

data Script lang where

A script in a particular language.

See also ScriptInAnyLang for a script in any of the known languages.

See also ScriptInEra for a script in a language that is available within a particular era.

Note that some but not all scripts have an external JSON syntax, hence this type has no JSON serialisation instances. The SimpleScript family of languages do have a JSON syntax and thus have ToJSON/FromJSON instances.

Constructors

SimpleScript :: !SimpleScript -> Script SimpleScript'

PlutusScript :: forall lang. IsPlutusScriptLanguage lang => !(PlutusScriptVersion lang) -> !(PlutusScript lang) -> Script lang

data PlutusScriptInEra era lang where

Constructors

PlutusScriptInEra :: forall lang era. PlutusScript lang -> PlutusScriptInEra era lang

Scripts in any language

data ScriptInAnyLang where

Sometimes it is necessary to handle all languages without making static type distinctions between languages. For example, when reading external input, or before the era context is known.

Use toScriptInEra to convert to a script in the context of an era.

Constructors

ScriptInAnyLang :: forall lang. ScriptLanguage lang -> Script lang -> ScriptInAnyLang

toScriptInAnyLang :: Script lang -> ScriptInAnyLang

Convert a script in a specific statically-known language to a ScriptInAnyLang.

No inverse to this is provided, just do case analysis on the ScriptLanguage field within the ScriptInAnyLang constructor.

removePlutusScriptDoubleEncoding :: ByteString -> ByteString

Previously we were double encoding the plutus script bytes. This function removes a layer of encoding to return the original plutus script bytes if it exists.

Scripts in an era

data ScriptInEra era where

Constructors

ScriptInEra :: forall lang era. ScriptLanguageInEra lang era -> Script lang -> ScriptInEra era

toScriptInEra :: ShelleyBasedEra era -> ScriptInAnyLang -> Maybe (ScriptInEra era)

Given a target era and a script in some language, check if the language is supported in that era, and if so return a ScriptInEra.

eraOfScriptInEra :: ScriptInEra era -> ShelleyBasedEra era

class HasScriptLanguageInEra lang era where

Smart-constructor for ScriptLanguageInEra to write functions manipulating scripts that do not commit to a particular era.

Methods

scriptLanguageInEra :: ScriptLanguageInEra lang era

class ToAlonzoScript lang era where

Methods

toLedgerScript :: PlutusScript lang -> AlonzoScript (ShelleyLedgerEra era)

Reference scripts

data ReferenceScript era where

A reference scripts is a script that can exist at a transaction output. This greatly reduces the size of transactions that use scripts as the script no longer has to be added to the transaction, they can now be referenced via a transaction output.

Constructors

ReferenceScript :: forall era. BabbageEraOnwards era -> ScriptInAnyLang -> ReferenceScript era

ReferenceScriptNone :: forall era. ReferenceScript era

refScriptToShelleyScript :: ShelleyBasedEra era -> ReferenceScript era -> StrictMaybe (Script (ShelleyLedgerEra era))

Use of a script in an era as a witness

data WitCtxTxIn

A tag type for the context in which a script is used in a transaction.

This type tags the context as being to witness a transaction input.

data WitCtxMint

A tag type for the context in which a script is used in a transaction.

This type tags the context as being to witness minting.

data WitCtxStake

A tag type for the context in which a script is used in a transaction.

This type tags the context as being to witness the use of stake addresses in certificates, withdrawals, voting and proposals.

data WitCtx witctx where

This GADT provides a value-level representation of all the witness contexts. This enables pattern matching on the context to allow them to be treated in a non-uniform way.

Constructors

WitCtxTxIn :: WitCtx WitCtxTxIn

WitCtxMint :: WitCtx WitCtxMint

WitCtxStake :: WitCtx WitCtxStake

data ScriptWitness witctx era where

A use of a script within a transaction body to witness that something is being used in an authorised manner. That can be

spending a transaction input

minting tokens

using a certificate (stake address certs specifically)

withdrawing from a reward account

For simple script languages, the use of the script is the same in all contexts. For Plutus scripts, using a script involves supplying a redeemer. In addition, Plutus scripts used for spending inputs must also supply the datum value used when originally creating the TxOut that is now being spent.

Constructors

SimpleScriptWitness :: forall era witctx. ScriptLanguageInEra SimpleScript' era -> SimpleScriptOrReferenceInput SimpleScript' -> ScriptWitness witctx era

PlutusScriptWitness :: forall lang era witctx. IsPlutusScriptLanguage lang => ScriptLanguageInEra lang era -> PlutusScriptVersion lang -> PlutusScriptOrReferenceInput lang -> ScriptDatum witctx -> ScriptRedeemer -> ExecutionUnits -> ScriptWitness witctx era

getScriptWitnessReferenceInput :: ScriptWitness witctx era -> Maybe TxIn

getScriptWitnessScript :: ScriptWitness witctx era -> Maybe (ScriptInEra era)

getScriptWitnessReferenceInputOrScript :: ScriptWitness witctx era -> Either (ScriptInEra era) TxIn

We cannot always extract a script from a script witness due to reference scripts. Reference scripts exist in the UTxO, so without access to the UTxO we cannot retrieve the script. So in the cases for script reference, the result contains Right TxIn.

data Witness witctx era where

Constructors

KeyWitness :: forall witctx era. KeyWitnessInCtx witctx -> Witness witctx era

ScriptWitness :: forall witctx era. ScriptWitnessInCtx witctx -> ScriptWitness witctx era -> Witness witctx era

data KeyWitnessInCtx witctx where

Constructors

KeyWitnessForSpending :: KeyWitnessInCtx WitCtxTxIn

KeyWitnessForStakeAddr :: KeyWitnessInCtx WitCtxStake

data ScriptWitnessInCtx witctx where

Constructors

ScriptWitnessForSpending :: ScriptWitnessInCtx WitCtxTxIn

ScriptWitnessForMinting :: ScriptWitnessInCtx WitCtxMint

ScriptWitnessForStakeAddr :: ScriptWitnessInCtx WitCtxStake

class IsScriptWitnessInCtx ctx where

Methods

scriptWitnessInCtx :: ScriptWitnessInCtx ctx

data ScriptDatum witctx where

Constructors

ScriptDatumForTxIn :: Maybe HashableScriptData -> ScriptDatum WitCtxTxIn

InlineScriptDatum :: ScriptDatum WitCtxTxIn

NoScriptDatumForMint :: ScriptDatum WitCtxMint

NoScriptDatumForStake :: ScriptDatum WitCtxStake

type ScriptRedeemer = HashableScriptData

Languages supported in each era

data ScriptLanguageInEra lang era where

Constructors

SimpleScriptInShelley :: ScriptLanguageInEra SimpleScript' ShelleyEra

SimpleScriptInAllegra :: ScriptLanguageInEra SimpleScript' AllegraEra

SimpleScriptInMary :: ScriptLanguageInEra SimpleScript' MaryEra

SimpleScriptInAlonzo :: ScriptLanguageInEra SimpleScript' AlonzoEra

SimpleScriptInBabbage :: ScriptLanguageInEra SimpleScript' BabbageEra

SimpleScriptInConway :: ScriptLanguageInEra SimpleScript' ConwayEra

PlutusScriptV1InAlonzo :: ScriptLanguageInEra PlutusScriptV1 AlonzoEra

PlutusScriptV1InBabbage :: ScriptLanguageInEra PlutusScriptV1 BabbageEra

PlutusScriptV1InConway :: ScriptLanguageInEra PlutusScriptV1 ConwayEra

PlutusScriptV2InBabbage :: ScriptLanguageInEra PlutusScriptV2 BabbageEra

PlutusScriptV2InConway :: ScriptLanguageInEra PlutusScriptV2 ConwayEra

PlutusScriptV3InConway :: ScriptLanguageInEra PlutusScriptV3 ConwayEra

scriptLanguageSupportedInEra :: ShelleyBasedEra era -> ScriptLanguage lang -> Maybe (ScriptLanguageInEra lang era)

Check if a given script language is supported in a given era, and if so return the evidence.

sbeToSimpleScriptLanguageInEra :: ShelleyBasedEra era -> ScriptLanguageInEra SimpleScript' era

languageOfScriptLanguageInEra :: ScriptLanguageInEra lang era -> ScriptLanguage lang

eraOfScriptLanguageInEra :: ScriptLanguageInEra lang era -> ShelleyBasedEra era

The simple script language

data SimpleScript

Constructors

RequireSignature !(Hash PaymentKey)

RequireTimeBefore !SlotNo

RequireTimeAfter !SlotNo

RequireAllOf ![SimpleScript]

RequireAnyOf ![SimpleScript]

RequireMOf !Int ![SimpleScript]

data SimpleScriptOrReferenceInput lang

Constructors

SScript SimpleScript

SReferenceScript TxIn

The Plutus script language

data PlutusScript lang where

Plutus scripts.

Note that Plutus scripts have a binary serialisation but no JSON serialisation.

Constructors

PlutusScriptSerialised :: forall lang. ShortByteString -> PlutusScript lang

data PlutusScriptOrReferenceInput lang

Scripts can now exist in the UTxO at a transaction output. We can reference these scripts via specification of a reference transaction input in order to witness spending inputs, withdrawals, certificates or to mint tokens. This datatype encapsulates this concept.

Constructors

PScript (PlutusScript lang)

PReferenceScript TxIn

examplePlutusScriptAlwaysSucceeds :: WitCtx witctx -> PlutusScript PlutusScriptV1

An example Plutus script that always succeeds, irrespective of inputs.

For example, if one were to use this for a payment address then it would allow anyone to spend from it.

The exact script depends on the context in which it is to be used.

examplePlutusScriptAlwaysFails :: WitCtx witctx -> PlutusScript PlutusScriptV1

An example Plutus script that always fails, irrespective of inputs.

For example, if one were to use this for a payment address then it would be impossible for anyone to ever spend from it.

The exact script depends on the context in which it is to be used.

Script execution units

data ExecutionUnits

The units for how long a script executes for and how much memory it uses. This is used to declare the resources used by a particular use of a script.

This type is also used to describe the limits for the maximum overall execution units per transaction or per block.

Constructors

ExecutionUnits

Fields

executionSteps :: Natural

This corresponds roughly to the time to execute a script.

executionMemory :: Natural

This corresponds roughly to the peak memory used during script execution.

Script hashes

newtype ScriptHash

We have this type separate from the Hash type to avoid the script hash type being parametrised by the era. The representation is era independent, and there are many places where we want to use a script hash where we don't want things to be era-parametrised.

Constructors

ScriptHash ScriptHash

parseScriptHash :: Parser ScriptHash

Parser for hex rerepsentation of ScriptHash

hashScript :: Script lang -> ScriptHash

Internal conversion functions

toShelleyScript :: ScriptInEra era -> Script (ShelleyLedgerEra era)

fromShelleyBasedScript :: ShelleyBasedEra era -> Script (ShelleyLedgerEra era) -> ScriptInEra era

toShelleyMultiSig :: SimpleScript -> Either MultiSigError (MultiSig (ShelleyLedgerEra ShelleyEra))

Conversion for the MultiSig language used by the Shelley era.

fromShelleyMultiSig :: MultiSig (ShelleyLedgerEra ShelleyEra) -> SimpleScript

Conversion for the MultiSig language used by the Shelley era.

toAllegraTimelock :: (AllegraEraScript era, NativeScript era ~ Timelock era) => SimpleScript -> NativeScript era

Conversion for the Timelock language that is shared between the Allegra and Mary eras.

fromAllegraTimelock :: AllegraEraScript era => NativeScript era -> SimpleScript

Conversion for the Timelock language that is shared between the Allegra and Mary eras.

toAlonzoExUnits :: ExecutionUnits -> ExUnits

fromAlonzoExUnits :: ExUnits -> ExecutionUnits

toShelleyScriptHash :: ScriptHash -> ScriptHash

fromShelleyScriptHash :: ScriptHash -> ScriptHash

toPlutusData :: ScriptData -> Data

fromPlutusData :: Data -> ScriptData

toAlonzoData :: Era ledgerera => HashableScriptData -> Data ledgerera

fromAlonzoData :: Data ledgerera -> HashableScriptData

toAlonzoLanguage :: AnyPlutusScriptVersion -> Language

fromAlonzoLanguage :: Language -> AnyPlutusScriptVersion

fromShelleyScriptToReferenceScript :: ShelleyBasedEra era -> Script (ShelleyLedgerEra era) -> ReferenceScript era

scriptInEraToRefScript :: ScriptInEra era -> ReferenceScript era

data DebugPlutusFailure

A structured representation of Plutus script validation failures, providing detailed information about the failed execution for debugging purposes. This type contains the same information as the data constructor ValidationFailure but with named fields and fixed crypto parameters for easier debugging and error reporting.

Constructors

DebugPlutusFailure

Fields

dpfEvaluationError :: EvaluationError

dpfScriptWithContext :: PlutusWithContext

dpfExecutionUnits :: ExUnits

dpfExecutionLogs :: [Text]

renderDebugPlutusFailure :: DebugPlutusFailure -> Text

collectPlutusScriptHashes :: AlonzoEraOnwards era -> Tx era -> UTxO era -> Map ScriptWitnessIndex ScriptHash

Collect all plutus script hashes that are needed to validate the given transaction and return them in a map with their corresponding ScriptWitnessIndex as key.

Script data

data HashableScriptData

Constructors

HashableScriptData

Fields

!ByteString

Original ScriptData bytes

!ScriptData

hashScriptDataBytes :: HashableScriptData -> Hash ScriptData

getOriginalScriptDataBytes :: HashableScriptData -> ByteString

getScriptData :: HashableScriptData -> ScriptData

unsafeHashableScriptData :: ScriptData -> HashableScriptData

Warning: Creating HashableScriptData from a ScriptData value pretty much guarantees the original bytes used to create the ScriptData value will be different if we serialize HashableScriptData again. Do not use this.

data ScriptData

Constructors

ScriptDataConstructor

Fields

Integer

Tag for the constructor

[ScriptData]

Constructor arguments

ScriptDataMap [(ScriptData, ScriptData)]
Key value pairs

ScriptDataList [ScriptData]
Elements

ScriptDataNumber Integer

ScriptDataBytes ByteString

parseScriptDataHash :: Parser (Hash ScriptData)

Parser for hex representation

Validating metadata

validateScriptData :: ScriptData -> Either ScriptDataRangeError ()

Validate script data. This is for use with existing constructed script data values, e.g. constructed manually or decoded from CBOR directly.

newtype ScriptDataRangeError

An error in script data due to an out-of-range value.

Constructors

ScriptDataConstructorOutOfRange Integer
The constructor number is outside the maximum range of -2^64-1 .. 2^64-1.

Conversion to/from JSON

data ScriptDataJsonSchema

Script data is similar to JSON but not exactly the same. It has some deliberate limitations such as no support for floating point numbers or special forms for null or boolean values. It also has limitations on the length of strings. On the other hand, unlike JSON, it distinguishes between byte strings and text strings. It also supports any value as map keys rather than just string. It also supports alternatives / tagged unions, used for representing constructors for Plutus data values.

We provide two different mappings between script data and JSON, useful for different purposes:

A mapping that allows almost any JSON value to be converted into script data. This does not require a specific JSON schema for the input. It does not expose the full representation capability of script data.

A mapping that exposes the full representation capability of script data, but relies on a specific JSON schema for the input JSON.

In the "no schema" mapping, the idea is that (almost) any JSON can be turned into script data and then converted back, without loss. That is, we can round-trip the JSON.

The subset of JSON supported is all JSON except:

No null or bool values

No floating point, only integers in the range of a 64bit signed integer

A limitation on string lengths

The approach for this mapping is to use whichever representation as script data is most compact. In particular:

JSON lists and maps represented as CBOR lists and maps

JSON strings represented as CBOR strings

JSON hex strings with "0x" prefix represented as CBOR byte strings

JSON integer numbers represented as CBOR signed or unsigned numbers

JSON maps with string keys that parse as numbers or hex byte strings, represented as CBOR map keys that are actually numbers or byte strings.

The string length limit depends on whether the hex string representation is used or not. For text strings the limit is 64 bytes for the UTF8 representation of the text string. For byte strings the limit is 64 bytes for the raw byte form (ie not the input hex, but after hex decoding).

In the "detailed schema" mapping, the idea is that we expose the full representation capability of the script data in the form of a JSON schema. This means the full representation is available and can be controlled precisely. It also means any script data can be converted into the JSON and back without loss. That is we can round-trip the script data via the JSON and also round-trip schema-compliant JSON via script data.

Warning*: While the JSON representation does round-trip, the CBOR through JSON does not. When serialising and deserialising HashableScriptData through JSON e.g: CBOR -> HashableScriptData -> JSON -> HashableScriptData -> CBOR the original CBOR representation is lost and the resulting CBOR *will* be different resulting in a different script data hash, which is calculated from CBOR. This is because cardano-ledger does not canonicalise CBOR representation, so you can have few slightly different serialised representations of a data structure, which represent the same value.

See: https://github.com/IntersectMBO/cardano-api/issues/612#issuecomment-2701256007

Constructors

ScriptDataJsonNoSchema
Use the "no schema" mapping between JSON and script data as described above.

ScriptDataJsonDetailedSchema
Use the "detailed schema" mapping between JSON and script data as described above.

scriptDataFromJson :: ScriptDataJsonSchema -> Value -> Either ScriptDataJsonError HashableScriptData

Convert a value from JSON into script data, using the given choice of mapping between JSON and script data.

This may fail with a conversion error if the JSON is outside the supported subset for the chosen mapping. See ScriptDataJsonSchema for the details.

scriptDataToJson :: ScriptDataJsonSchema -> HashableScriptData -> Value

Convert a script data value into JSON , using the given choice of mapping between JSON and script data.

This conversion is total but is not necessarily invertible. See ScriptDataJsonSchema for the details.

data ScriptDataJsonError

Constructors

ScriptDataJsonSchemaError !Value !ScriptDataJsonSchemaError

ScriptDataRangeError !Value !ScriptDataRangeError

data ScriptDataJsonSchemaError

Constructors

ScriptDataJsonNullNotAllowed

ScriptDataJsonBoolNotAllowed

ScriptDataJsonNumberNotInteger !Double

ScriptDataJsonNotObject !Value

ScriptDataJsonBadObject ![(Text, Value)]

ScriptDataJsonBadMapPair !Value

ScriptDataJsonTypeMismatch !Text !Value

scriptDataFromJsonDetailedSchema :: Value -> Either ScriptDataJsonSchemaError HashableScriptData

Convert a script data value from JSON using detailed schema.

This conversion is total but is not necessarily invertible. See ScriptDataJsonSchema for the details.

scriptDataToJsonDetailedSchema :: HashableScriptData -> Value

Convert a script data value into JSON using detailed schema.

This conversion is total but is not necessarily invertible. See ScriptDataJsonSchema for the details.

newtype ScriptBytesError

Constructors

ScriptBytesError String

data ScriptDataJsonBytesError

Constructors

ScriptDataJsonBytesErrorValue ScriptDataJsonError

ScriptDataJsonBytesErrorInvalid ScriptDataRangeError

scriptDataJsonToHashable

Arguments

:: ScriptDataJsonSchema

-> Value
ScriptData Value

-> Either ScriptDataJsonBytesError HashableScriptData

This allows us to take JSON formatted ScriptData and encode it in the CDDL format whilst preserving the original bytes.

Data family instances

data family AsType t

A family of singleton types used in this API to indicate which type to use where it would otherwise be ambiguous or merely unclear.

Values of this type are passed to deserialisation functions for example.

data family Hash keyrole