Cardano.Api.Protocol

Documentation

data BlockType blk where

Constructors

ByronBlockType ∷ BlockType ByronBlockHFC

ShelleyBlockType ∷ BlockType (ShelleyBlockHFC (TPraos StandardCrypto) StandardShelley)

CardanoBlockType ∷ BlockType (CardanoBlock StandardCrypto)

class (RunNode blk, IOLike m) ⇒ Protocol m blk where

Associated Types

data ProtocolInfoArgs m blk

Methods

protocolInfo ∷ ProtocolInfoArgs m blk → ProtocolInfo m blk

class RunNode blk ⇒ ProtocolClient blk where

Node client support for each consensus protocol.

This is like Protocol but for clients of the node, so with less onerous requirements than to run a node.

Associated Types

data ProtocolClientInfoArgs blk

Methods

protocolClientInfo ∷ ProtocolClientInfoArgs blk → ProtocolClientInfo blk