Cardano.Api.Protocol

The enumeration of supported protocols

data Protocol

Constructors

ByronProtocol !EpochSlots

ShelleyProtocol

CardanoProtocol !EpochSlots

Node client support

Support for the context needed to run a client of a node that is using a protocol.

localNodeConnectInfo ∷ Protocol → NetworkId → FilePath → LocalNodeConnectInfoForSomeMode

withlocalNodeConnectInfo ∷ Protocol → NetworkId → FilePath → (∀ mode block. RunNode block ⇒ LocalNodeConnectInfo mode block → a) → a

data LocalNodeConnectInfoForSomeMode where

Constructors

LocalNodeConnectInfoForSomeMode ∷ RunNode block ⇒ LocalNodeConnectInfo mode block → LocalNodeConnectInfoForSomeMode