Cardano.Api.ProtocolParameters

Description

The various Cardano protocol parameters, including:

the current values of updatable protocol parameters: ProtocolParameters

updates to protocol parameters: ProtocolParametersUpdate

update proposals that can be embedded in transactions: UpdateProposal

parameters fixed in the genesis file: GenesisParameters

The updatable protocol parameters

data EpochNo

An epoch, i.e. the number of the epoch.

The updatable protocol parameters

newtype LedgerProtocolParameters era

Constructors

LedgerProtocolParameters

Fields

unLedgerProtocolParameters :: PParams (ShelleyLedgerEra era)

data EraBasedProtocolParametersUpdate era where

Each constructor corresponds to the set of protocol parameters available in a given era.

Constructors

ShelleyEraBasedProtocolParametersUpdate :: CommonProtocolParametersUpdate -> DeprecatedAfterMaryPParams ShelleyEra -> DeprecatedAfterBabbagePParams ShelleyEra -> ShelleyToAlonzoPParams ShelleyEra -> EraBasedProtocolParametersUpdate ShelleyEra

AllegraEraBasedProtocolParametersUpdate :: CommonProtocolParametersUpdate -> DeprecatedAfterMaryPParams AllegraEra -> ShelleyToAlonzoPParams AllegraEra -> DeprecatedAfterBabbagePParams ShelleyEra -> EraBasedProtocolParametersUpdate AllegraEra

MaryEraBasedProtocolParametersUpdate :: CommonProtocolParametersUpdate -> DeprecatedAfterMaryPParams MaryEra -> ShelleyToAlonzoPParams MaryEra -> DeprecatedAfterBabbagePParams ShelleyEra -> EraBasedProtocolParametersUpdate MaryEra

AlonzoEraBasedProtocolParametersUpdate :: CommonProtocolParametersUpdate -> ShelleyToAlonzoPParams AlonzoEra -> AlonzoOnwardsPParams AlonzoEra -> DeprecatedAfterBabbagePParams ShelleyEra -> EraBasedProtocolParametersUpdate AlonzoEra

BabbageEraBasedProtocolParametersUpdate :: CommonProtocolParametersUpdate -> AlonzoOnwardsPParams BabbageEra -> DeprecatedAfterBabbagePParams ShelleyEra -> IntroducedInBabbagePParams BabbageEra -> EraBasedProtocolParametersUpdate BabbageEra

ConwayEraBasedProtocolParametersUpdate :: CommonProtocolParametersUpdate -> AlonzoOnwardsPParams ConwayEra -> IntroducedInBabbagePParams ConwayEra -> IntroducedInConwayPParams (ShelleyLedgerEra ConwayEra) -> EraBasedProtocolParametersUpdate ConwayEra

data AlonzoOnwardsPParams ledgerera

Constructors

AlonzoOnwardsPParams

Fields

alCostModels :: StrictMaybe CostModels

alPrices :: StrictMaybe Prices

alMaxTxExUnits :: StrictMaybe ExUnits

alMaxBlockExUnits :: StrictMaybe ExUnits

alMaxValSize :: StrictMaybe Natural

alCollateralPercentage :: StrictMaybe Natural

alMaxCollateralInputs :: StrictMaybe Natural

data CommonProtocolParametersUpdate

Protocol parameters common to each era. This can only ever be reduced if parameters are deprecated.

Constructors

CommonProtocolParametersUpdate

Fields

cppMinFeeA :: StrictMaybe Coin

cppMinFeeB :: StrictMaybe Coin

cppMaxBlockBodySize :: StrictMaybe Word32

cppMaxTxSize :: StrictMaybe Word32

cppMaxBlockHeaderSize :: StrictMaybe Word16

cppKeyDeposit :: StrictMaybe Coin

cppPoolDeposit :: StrictMaybe Coin

cppPoolRetireMaxEpoch :: StrictMaybe EpochInterval

cppStakePoolTargetNum :: StrictMaybe Word16

cppPoolPledgeInfluence :: StrictMaybe NonNegativeInterval

cppTreasuryExpansion :: StrictMaybe UnitInterval

cppMonetaryExpansion :: StrictMaybe UnitInterval

cppMinPoolCost :: StrictMaybe Coin

newtype DeprecatedAfterBabbagePParams ledgerera

Constructors

DeprecatedAfterBabbagePParams (StrictMaybe ProtVer)

newtype DeprecatedAfterMaryPParams ledgerera

Constructors

DeprecatedAfterMaryPParams (StrictMaybe Coin)

data ShelleyToAlonzoPParams ledgerera

Constructors

ShelleyToAlonzoPParams

Fields

(StrictMaybe Nonce)

Extra entropy

(StrictMaybe UnitInterval)

Decentralization parameter

newtype IntroducedInBabbagePParams era

Constructors

IntroducedInBabbagePParams (StrictMaybe CoinPerByte)
Coins per UTxO byte

data IntroducedInConwayPParams era

Constructors

IntroducedInConwayPParams

Fields

icPoolVotingThresholds :: StrictMaybe PoolVotingThresholds

icDRepVotingThresholds :: StrictMaybe DRepVotingThresholds

icMinCommitteeSize :: StrictMaybe Natural

icCommitteeTermLength :: StrictMaybe EpochInterval

icGovActionLifetime :: StrictMaybe EpochInterval

icGovActionDeposit :: StrictMaybe Coin

icDRepDeposit :: StrictMaybe Coin

icDRepActivity :: StrictMaybe EpochInterval

icMinFeeRefScriptCostPerByte :: StrictMaybe NonNegativeInterval

createEraBasedProtocolParamUpdate :: ShelleyBasedEra era -> EraBasedProtocolParametersUpdate era -> PParamsUpdate (ShelleyLedgerEra era)

createPParams :: ShelleyBasedEra era -> EraBasedProtocolParametersUpdate era -> PParams (ShelleyLedgerEra era)

Deprecated

data ProtocolParametersUpdate

The representation of a change in the ProtocolParameters.

Constructors

ProtocolParametersUpdate

Fields

protocolUpdateProtocolVersion :: Maybe (Natural, Natural)

Protocol version, major and minor. Updating the major version is used to trigger hard forks.

protocolUpdateDecentralization :: Maybe Rational

The decentralization parameter. This is fraction of slots that belong to the BFT overlay schedule, rather than the Praos schedule. So 1 means fully centralised, while 0 means fully decentralised.

This is the "d" parameter from the design document.

protocolUpdateExtraPraosEntropy :: Maybe (Maybe PraosNonce)

Extra entropy for the Praos per-epoch nonce.

This can be used to add extra entropy during the decentralisation process. If the extra entropy can be demonstrated to be generated randomly then this method can be used to show that the initial federated operators did not subtly bias the initial schedule so that they retain undue influence after decentralisation.

protocolUpdateMaxBlockHeaderSize :: Maybe Word16

The maximum permitted size of a block header.

This must be at least as big as the largest legitimate block headers but should not be too much larger, to help prevent DoS attacks.

Caution: setting this to be smaller than legitimate block headers is a sure way to brick the system!

protocolUpdateMaxBlockBodySize :: Maybe Word32

The maximum permitted size of the block body (that is, the block payload, without the block header).

This should be picked with the Praos network delta security parameter in mind. Making this too large can severely weaken the Praos consensus properties.

Caution: setting this to be smaller than a transaction that can change the protocol parameters is a sure way to brick the system!

protocolUpdateMaxTxSize :: Maybe Word32

The maximum permitted size of a transaction.

Typically this should not be too high a fraction of the block size, otherwise wastage from block fragmentation becomes a problem, and the current implementation does not use any sophisticated box packing algorithm.

protocolUpdateTxFeeFixed :: Maybe Coin

The constant factor for the minimum fee calculation.

protocolUpdateTxFeePerByte :: Maybe Coin

The linear factor for the minimum fee calculation.

protocolUpdateMinUTxOValue :: Maybe Coin

The minimum permitted value for new UTxO entries, ie for transaction outputs.

protocolUpdateStakeAddressDeposit :: Maybe Coin

The deposit required to register a stake address.

protocolUpdateStakePoolDeposit :: Maybe Coin

The deposit required to register a stake pool.

protocolUpdateMinPoolCost :: Maybe Coin

The minimum value that stake pools are permitted to declare for their cost parameter.

protocolUpdatePoolRetireMaxEpoch :: Maybe EpochInterval

The maximum number of epochs into the future that stake pools are permitted to schedule a retirement.

protocolUpdateStakePoolTargetNum :: Maybe Word16

The equilibrium target number of stake pools.

This is the "k" incentives parameter from the design document.

protocolUpdatePoolPledgeInfluence :: Maybe Rational

The influence of the pledge in stake pool rewards.

This is the "a_0" incentives parameter from the design document.

protocolUpdateMonetaryExpansion :: Maybe Rational

The monetary expansion rate. This determines the fraction of the reserves that are added to the fee pot each epoch.

This is the "rho" incentives parameter from the design document.

protocolUpdateTreasuryCut :: Maybe Rational

The fraction of the fee pot each epoch that goes to the treasury.

This is the "tau" incentives parameter from the design document.

protocolUpdateCostModels :: Map AnyPlutusScriptVersion CostModel

Cost models for script languages that use them.

Introduced in Alonzo

protocolUpdatePrices :: Maybe ExecutionUnitPrices

Price of execution units for script languages that use them.

Introduced in Alonzo

protocolUpdateMaxTxExUnits :: Maybe ExecutionUnits

Max total script execution resources units allowed per tx

Introduced in Alonzo

protocolUpdateMaxBlockExUnits :: Maybe ExecutionUnits

Max total script execution resources units allowed per block

Introduced in Alonzo

protocolUpdateMaxValueSize :: Maybe Natural

Max size of a Value in a tx output.

Introduced in Alonzo

protocolUpdateCollateralPercent :: Maybe Natural

The percentage of the script contribution to the txfee that must be provided as collateral inputs when including Plutus scripts.

Introduced in Alonzo

protocolUpdateMaxCollateralInputs :: Maybe Natural

The maximum number of collateral inputs allowed in a transaction.

Introduced in Alonzo

protocolUpdateUTxOCostPerByte :: Maybe Coin

Cost in ada per byte of UTxO storage.

Introduced in Babbage

Errors

data ProtocolParametersError

Constructors

PParamsErrorMissingMinUTxoValue !AnyCardanoEra

PParamsErrorMissingAlonzoProtocolParameter

data ProtocolParametersConversionError

Constructors

PpceOutOfBounds !ProtocolParameterName !Rational

PpceVersionInvalid !ProtocolParameterVersion

PpceInvalidCostModel !CostModel !CostModelApplyError

PpceMissingParameter !ProtocolParameterName

PraosNonce

data PraosNonce

makePraosNonce :: ByteString -> PraosNonce

Execution units, prices and cost models,

data ExecutionUnits

The units for how long a script executes for and how much memory it uses. This is used to declare the resources used by a particular use of a script.

This type is also used to describe the limits for the maximum overall execution units per transaction or per block.

Constructors

ExecutionUnits

Fields

executionSteps :: Natural

This corresponds roughly to the time to execute a script.

executionMemory :: Natural

This corresponds roughly to the peak memory used during script execution.

data ExecutionUnitPrices

The prices for ExecutionUnits as a fraction of a Coin.

These are used to determine the fee for the use of a script within a transaction, based on the ExecutionUnits needed by the use of the script.

Constructors

ExecutionUnitPrices

Fields

priceExecutionSteps :: Rational

priceExecutionMemory :: Rational

newtype CostModels

Constructors

CostModels

Fields

unCostModels :: Map AnyPlutusScriptVersion CostModel

newtype CostModel

Constructors

CostModel [Int64]

fromAlonzoCostModels :: CostModels -> Map AnyPlutusScriptVersion CostModel

Update proposals to change the protocol parameters

data UpdateProposal

Constructors

UpdateProposal !(Map (Hash GenesisKey) ProtocolParametersUpdate) !EpochNo

makeShelleyUpdateProposal :: ProtocolParametersUpdate -> [Hash GenesisKey] -> EpochNo -> UpdateProposal

Internal conversion functions

toLedgerNonce :: Maybe PraosNonce -> Nonce

toLedgerUpdate :: ShelleyBasedEra era -> UpdateProposal -> Either ProtocolParametersConversionError (Update (ShelleyLedgerEra era))

fromLedgerUpdate :: ShelleyLedgerEra era ~ ledgerera => ShelleyBasedEra era -> Update ledgerera -> UpdateProposal

toLedgerProposedPPUpdates :: ShelleyBasedEra era -> Map (Hash GenesisKey) ProtocolParametersUpdate -> Either ProtocolParametersConversionError (ProposedPPUpdates (ShelleyLedgerEra era))

fromLedgerProposedPPUpdates :: ShelleyLedgerEra era ~ ledgerera => ShelleyBasedEra era -> ProposedPPUpdates ledgerera -> Map (Hash GenesisKey) ProtocolParametersUpdate

toLedgerPParamsUpdate :: ShelleyBasedEra era -> ProtocolParametersUpdate -> Either ProtocolParametersConversionError (PParamsUpdate (ShelleyLedgerEra era))

fromLedgerPParamsUpdate :: ShelleyBasedEra era -> PParamsUpdate (ShelleyLedgerEra era) -> ProtocolParametersUpdate

toAlonzoPrices :: ExecutionUnitPrices -> Either ProtocolParametersConversionError Prices

fromAlonzoPrices :: Prices -> ExecutionUnitPrices

toAlonzoScriptLanguage :: AnyPlutusScriptVersion -> Language

fromAlonzoScriptLanguage :: Language -> AnyPlutusScriptVersion

toAlonzoCostModel :: CostModel -> Language -> Either ProtocolParametersConversionError CostModel

fromAlonzoCostModel :: CostModel -> CostModel

toAlonzoCostModels :: Map AnyPlutusScriptVersion CostModel -> Either ProtocolParametersConversionError CostModels

Data family instances

data family AsType t

A family of singleton types used in this API to indicate which type to use where it would otherwise be ambiguous or merely unclear.

Values of this type are passed to deserialisation functions for example.

Era-dependent protocol features