Cardano.Api.ProtocolParameters

Description

The various Cardano protocol parameters, including:

the current values of updateable protocol parameters: ProtocolParameters

updates to protocol parameters: ProtocolParametersUpdate

update proposals that can be embedded in transactions: UpdateProposal

parameters fixed in the genesis file: GenesisParameters

The updateable protocol paramaters

data ProtocolParameters

The values of the set of updateable protocol paramaters. At any particular point on the chain there is a current set of paramaters in use.

These paramaters can be updated (at epoch boundaries) via an UpdateProposal, which contains a ProtocolParametersUpdate.

The ProtocolParametersUpdate is essentially a diff for the ProtocolParameters.

There are also paramaters fixed in the Genesis file. See GenesisParameters.

Constructors

ProtocolParameters

Fields

protocolParamProtocolVersion ∷ (Natural, Natural)

Protocol version, major and minor. Updating the major version is used to trigger hard forks.

protocolParamDecentralization ∷ Rational

The decentralization parameter. This is fraction of slots that belong to the BFT overlay schedule, rather than the Praos schedule. So 1 means fully centralised, while 0 means fully decentralised.

This is the "d" parameter from the design document.

protocolParamExtraPraosEntropy ∷ Maybe PraosNonce

Extra entropy for the Praos per-epoch nonce.

This can be used to add extra entropy during the decentralisation process. If the extra entropy can be demonstrated to be generated randomly then this method can be used to show that the initial federated operators did not subtly bias the initial schedule so that they retain undue influence after decentralisation.

protocolParamMaxBlockHeaderSize ∷ Natural

The maximum permitted size of a block header.

This must be at least as big as the largest legitimate block headers but should not be too much larger, to help prevent DoS attacks.

Caution: setting this to be smaller than legitimate block headers is a sure way to brick the system!

protocolParamMaxBlockBodySize ∷ Natural

The maximum permitted size of the block body (that is, the block payload, without the block header).

This should be picked with the Praos network delta security parameter in mind. Making this too large can severely weaken the Praos consensus properties.

Caution: setting this to be smaller than a transaction that can change the protocol parameters is a sure way to brick the system!

protocolParamMaxTxSize ∷ Natural

The maximum permitted size of a transaction.

Typically this should not be too high a fraction of the block size, otherwise wastage from block fragmentation becomes a problem, and the current implementation does not use any sophisticated box packing algorithm.

protocolParamTxFeeFixed ∷ Natural

The constant factor for the minimum fee calculation.

protocolParamTxFeePerByte ∷ Natural

The linear factor for the minimum fee calculation.

protocolParamMinUTxOValue ∷ Lovelace

The minimum permitted value for new UTxO entries, ie for transaction outputs.

protocolParamStakeAddressDeposit ∷ Lovelace

The deposit required to register a stake address.

protocolParamStakePoolDeposit ∷ Lovelace

The deposit required to register a stake pool.

protocolParamMinPoolCost ∷ Lovelace

The minimum value that stake pools are permitted to declare for their cost parameter.

protocolParamPoolRetireMaxEpoch ∷ EpochNo

The maximum number of epochs into the future that stake pools are permitted to schedule a retirement.

protocolParamStakePoolTargetNum ∷ Natural

The equilibrium target number of stake pools.

This is the "k" incentives parameter from the design document.

protocolParamPoolPledgeInfluence ∷ Rational

The influence of the pledge in stake pool rewards.

This is the "a_0" incentives parameter from the design document.

protocolParamMonetaryExpansion ∷ Rational

The monetary expansion rate. This determines the fraction of the reserves that are added to the fee pot each epoch.

This is the "rho" incentives parameter from the design document.

protocolParamTreasuryCut ∷ Rational

The fraction of the fee pot each epoch that goes to the treasury.

This is the "tau" incentives parameter from the design document.

data EpochNo

Updates to the protocol paramaters

data ProtocolParametersUpdate

The representation of a change in the ProtocolParameters.

Constructors

ProtocolParametersUpdate

Fields

protocolUpdateProtocolVersion ∷ Maybe (Natural, Natural)

Protocol version, major and minor. Updating the major version is used to trigger hard forks.

protocolUpdateDecentralization ∷ Maybe Rational

The decentralization parameter. This is fraction of slots that belong to the BFT overlay schedule, rather than the Praos schedule. So 1 means fully centralised, while 0 means fully decentralised.

This is the "d" parameter from the design document.

protocolUpdateExtraPraosEntropy ∷ Maybe (Maybe PraosNonce)

Extra entropy for the Praos per-epoch nonce.

This can be used to add extra entropy during the decentralisation process. If the extra entropy can be demonstrated to be generated randomly then this method can be used to show that the initial federated operators did not subtly bias the initial schedule so that they retain undue influence after decentralisation.

protocolUpdateMaxBlockHeaderSize ∷ Maybe Natural

The maximum permitted size of a block header.

This must be at least as big as the largest legitimate block headers but should not be too much larger, to help prevent DoS attacks.

Caution: setting this to be smaller than legitimate block headers is a sure way to brick the system!

protocolUpdateMaxBlockBodySize ∷ Maybe Natural

The maximum permitted size of the block body (that is, the block payload, without the block header).

This should be picked with the Praos network delta security parameter in mind. Making this too large can severely weaken the Praos consensus properties.

Caution: setting this to be smaller than a transaction that can change the protocol parameters is a sure way to brick the system!

protocolUpdateMaxTxSize ∷ Maybe Natural

The maximum permitted size of a transaction.

Typically this should not be too high a fraction of the block size, otherwise wastage from block fragmentation becomes a problem, and the current implementation does not use any sophisticated box packing algorithm.

protocolUpdateTxFeeFixed ∷ Maybe Natural

The constant factor for the minimum fee calculation.

protocolUpdateTxFeePerByte ∷ Maybe Natural

The linear factor for the minimum fee calculation.

protocolUpdateMinUTxOValue ∷ Maybe Lovelace

The minimum permitted value for new UTxO entries, ie for transaction outputs.

protocolUpdateStakeAddressDeposit ∷ Maybe Lovelace

The deposit required to register a stake address.

protocolUpdateStakePoolDeposit ∷ Maybe Lovelace

The deposit required to register a stake pool.

protocolUpdateMinPoolCost ∷ Maybe Lovelace

The minimum value that stake pools are permitted to declare for their cost parameter.

protocolUpdatePoolRetireMaxEpoch ∷ Maybe EpochNo

The maximum number of epochs into the future that stake pools are permitted to schedule a retirement.

protocolUpdateStakePoolTargetNum ∷ Maybe Natural

The equilibrium target number of stake pools.

This is the "k" incentives parameter from the design document.

protocolUpdatePoolPledgeInfluence ∷ Maybe Rational

The influence of the pledge in stake pool rewards.

This is the "a_0" incentives parameter from the design document.

protocolUpdateMonetaryExpansion ∷ Maybe Rational

The monetary expansion rate. This determines the fraction of the reserves that are added to the fee pot each epoch.

This is the "rho" incentives parameter from the design document.

protocolUpdateTreasuryCut ∷ Maybe Rational

The fraction of the fee pot each epoch that goes to the treasury.

This is the "tau" incentives parameter from the design document.

PraosNonce

data PraosNonce

makePraosNonce ∷ ByteString → PraosNonce

Update proposals to change the protocol paramaters

data UpdateProposal

Constructors

UpdateProposal !(Map (Hash GenesisKey) ProtocolParametersUpdate) !EpochNo

makeShelleyUpdateProposal ∷ ProtocolParametersUpdate → [Hash GenesisKey] → EpochNo → UpdateProposal

Protocol paramaters fixed in the genesis file

data GenesisParameters

Constructors

GenesisParameters

Fields

protocolParamSystemStart ∷ UTCTime

The reference time the system started. The time of slot zero. The time epoch against which all Ouroboros time slots are measured.

protocolParamNetworkId ∷ NetworkId

The network identifier for this blockchain instance. This distinguishes the mainnet from testnets, and different testnets from each other.

protocolParamActiveSlotsCoefficient ∷ Rational

The Ouroboros Praos active slot coefficient, aka f.

protocolParamSecurity ∷ Int

The Ouroboros security paramaters, aka k. This is the maximum number of blocks the node would ever be prepared to roll back by.

Clients of the node following the chain should be prepared to handle the node switching forks up to this long.

protocolParamEpochLength ∷ EpochSize

The number of Ouroboros time slots in an Ouroboros epoch.

protocolParamSlotLength ∷ NominalDiffTime

The time duration of a slot.

protocolParamSlotsPerKESPeriod ∷ Int

For Ouroboros Praos, the length of a KES period as a number of time slots. The KES keys get evolved once per KES period.

protocolParamMaxKESEvolutions ∷ Int

The maximum number of times a KES key can be evolved before it is no longer considered valid. This can be less than the maximum number of times given the KES key size. For example the mainnet KES key size would allow 64 evolutions, but the max KES evolutions param is 62.

protocolParamUpdateQuorum ∷ Int

In the Shelley era, prior to decentralised governance, this is the number of genesis key delegates that need to agree for an update proposal to be enacted.

protocolParamMaxLovelaceSupply ∷ Lovelace

The maximum supply for Lovelace. This determines the initial value of the reserves.

protocolInitialUpdateableProtocolParameters ∷ ProtocolParameters

The initial values of the updateable ProtocolParameters.

newtype EpochSize

Constructors

EpochSize

Fields

unEpochSize ∷ Word64

Internal conversion functions

toShelleyPParamsUpdate ∷ ProtocolParametersUpdate → PParamsUpdate ledgerera

toShelleyProposedPPUpdates ∷ Crypto ledgerera ~ StandardCrypto ⇒ Map (Hash GenesisKey) ProtocolParametersUpdate → ProposedPPUpdates ledgerera

toShelleyUpdate ∷ Crypto ledgerera ~ StandardCrypto ⇒ UpdateProposal → Update ledgerera

fromShelleyPParams ∷ PParams ledgerera → ProtocolParameters

fromShelleyPParamsUpdate ∷ PParamsUpdate ledgerera → ProtocolParametersUpdate

fromShelleyProposedPPUpdates ∷ Crypto ledgerera ~ StandardCrypto ⇒ ProposedPPUpdates ledgerera → Map (Hash GenesisKey) ProtocolParametersUpdate

fromShelleyUpdate ∷ Crypto ledgerera ~ StandardCrypto ⇒ Update ledgerera → UpdateProposal

fromShelleyGenesis ∷ ShelleyGenesis era → GenesisParameters

Data family instances

data family AsType t

A family of singleton types used in this API to indicate which type to use where it would otherwise be ambiguous or merely unclear.

Values of this type are passed to deserialisation functions for example.