Cardano.Api.Query

Description

Queries from local clients to the node.

Queries

data QueryInMode mode result where

Constructors

QueryCurrentEra ∷ ConsensusModeIsMultiEra mode → QueryInMode mode AnyCardanoEra

QueryInEra ∷ EraInMode era mode → QueryInEra era result → QueryInMode mode (Either EraMismatch result)

QueryEraHistory ∷ ConsensusModeIsMultiEra mode → QueryInMode mode (EraHistory mode)

QuerySystemStart ∷ QueryInMode mode SystemStart

QueryChainBlockNo ∷ QueryInMode mode (WithOrigin BlockNo)

QueryChainPoint ∷ ConsensusMode mode → QueryInMode mode ChainPoint

data QueryInEra era result where

Constructors

QueryByronUpdateState ∷ QueryInEra ByronEra ByronUpdateState

QueryInShelleyBasedEra ∷ ShelleyBasedEra era → QueryInShelleyBasedEra era result → QueryInEra era result

data QueryInShelleyBasedEra era result where

Constructors

QueryEpoch ∷ QueryInShelleyBasedEra era EpochNo

QueryGenesisParameters ∷ QueryInShelleyBasedEra era (GenesisParameters ShelleyEra)

QueryProtocolParameters ∷ QueryInShelleyBasedEra era (PParams (ShelleyLedgerEra era))

QueryProtocolParametersUpdate ∷ QueryInShelleyBasedEra era (Map (Hash GenesisKey) ProtocolParametersUpdate)

QueryStakeDistribution ∷ QueryInShelleyBasedEra era (Map (Hash StakePoolKey) Rational)

QueryUTxO ∷ QueryUTxOFilter → QueryInShelleyBasedEra era (UTxO era)

QueryStakeAddresses ∷ Set StakeCredential → NetworkId → QueryInShelleyBasedEra era (Map StakeAddress Lovelace, Map StakeAddress PoolId)

QueryStakePools ∷ QueryInShelleyBasedEra era (Set PoolId)

QueryStakePoolParameters ∷ Set PoolId → QueryInShelleyBasedEra era (Map PoolId StakePoolParameters)

QueryDebugLedgerState ∷ QueryInShelleyBasedEra era (SerialisedDebugLedgerState era)

QueryProtocolState ∷ QueryInShelleyBasedEra era (ProtocolState era)

QueryCurrentEpochState ∷ QueryInShelleyBasedEra era (SerialisedCurrentEpochState era)

QueryPoolState ∷ Maybe (Set PoolId) → QueryInShelleyBasedEra era (SerialisedPoolState era)

QueryPoolDistribution ∷ Maybe (Set PoolId) → QueryInShelleyBasedEra era (SerialisedPoolDistribution era)

QueryStakeSnapshot ∷ Maybe (Set PoolId) → QueryInShelleyBasedEra era (SerialisedStakeSnapshots era)

QueryStakeDelegDeposits ∷ Set StakeCredential → QueryInShelleyBasedEra era (Map StakeCredential Lovelace)

QueryConstitution ∷ QueryInShelleyBasedEra era (Maybe (Constitution (ShelleyLedgerEra era)))

QueryGovState ∷ QueryInShelleyBasedEra era (GovState (ShelleyLedgerEra era))

QueryDRepState ∷ Set (Credential DRepRole StandardCrypto) → QueryInShelleyBasedEra era (Map (Credential DRepRole StandardCrypto) (DRepState StandardCrypto))

QueryDRepStakeDistr ∷ Set (DRep StandardCrypto) → QueryInShelleyBasedEra era (Map (DRep StandardCrypto) Lovelace)

QueryCommitteeState ∷ QueryInShelleyBasedEra era (CommitteeState (ShelleyLedgerEra era))

data QueryUTxOFilter

Getting the whole UTxO is obviously not efficient since the result can be huge. Filtering by address is also not efficient because it requires a linear search.

The QueryUTxOFilterByTxIn is efficient since it fits with the structure of the UTxO (which is indexed by TxIn).

Constructors

QueryUTxOWhole
O(n) time and space for utxo size n

QueryUTxOByAddress (Set AddressAny)
O(n) time, O(m) space for utxo size n, and address set size m

QueryUTxOByTxIn (Set TxIn)
O(m log n) time, O(m) space for utxo size n, and address set size m

newtype UTxO era

Constructors

UTxO

Fields

unUTxO ∷ Map TxIn (TxOut CtxUTxO era)

data UTxOInAnyEra where

Constructors

UTxOInAnyEra ∷ CardanoEra era → UTxO era → UTxOInAnyEra

Internal conversion functions

toConsensusQuery ∷ ∀ mode block result. ConsensusBlockForMode mode ~ block ⇒ QueryInMode mode result → Some (Query block)

fromConsensusQueryResult ∷ ∀ mode block result result'. HasCallStack ⇒ ConsensusBlockForMode mode ~ block ⇒ QueryInMode mode result → Query block result' → result' → result

Wrapper types used in queries

newtype SerialisedDebugLedgerState era

Constructors

SerialisedDebugLedgerState (Serialised (NewEpochState (ShelleyLedgerEra era)))

newtype ProtocolState era

Constructors

ProtocolState (Serialised (ChainDepState (ConsensusProtocol era)))

decodeProtocolState ∷ FromCBOR (ChainDepState (ConsensusProtocol era)) ⇒ ProtocolState era → Either (ByteString, DecoderError) (ChainDepState (ConsensusProtocol era))

newtype DebugLedgerState era

Constructors

DebugLedgerState

Fields

unDebugLedgerState ∷ NewEpochState (ShelleyLedgerEra era)

decodeDebugLedgerState ∷ ∀ era. FromCBOR (DebugLedgerState era) ⇒ SerialisedDebugLedgerState era → Either ByteString (DebugLedgerState era)

newtype SerialisedCurrentEpochState era

Constructors

SerialisedCurrentEpochState (Serialised (EpochState (ShelleyLedgerEra era)))

newtype CurrentEpochState era

Constructors

CurrentEpochState (EpochState (ShelleyLedgerEra era))

decodeCurrentEpochState ∷ ShelleyBasedEra era → SerialisedCurrentEpochState era → Either DecoderError (CurrentEpochState era)

newtype SerialisedPoolState era

Constructors

SerialisedPoolState (Serialised (PState (ShelleyLedgerEra era)))

newtype PoolState era

Constructors

PoolState (PState (ShelleyLedgerEra era))

decodePoolState ∷ ∀ era. Era (ShelleyLedgerEra era) ⇒ DecCBOR (PState (ShelleyLedgerEra era)) ⇒ SerialisedPoolState era → Either DecoderError (PoolState era)

newtype SerialisedPoolDistribution era

Constructors

SerialisedPoolDistribution (Serialised (PoolDistr (EraCrypto (ShelleyLedgerEra era))))

newtype PoolDistribution era

Constructors

PoolDistribution

Fields

unPoolDistr ∷ PoolDistr (EraCrypto (ShelleyLedgerEra era))

decodePoolDistribution ∷ ∀ era. Crypto (EraCrypto (ShelleyLedgerEra era)) ⇒ ShelleyBasedEra era → SerialisedPoolDistribution era → Either DecoderError (PoolDistribution era)

newtype SerialisedStakeSnapshots era

Constructors

SerialisedStakeSnapshots (Serialised (StakeSnapshots (EraCrypto (ShelleyLedgerEra era))))

newtype StakeSnapshot era

Constructors

StakeSnapshot (StakeSnapshots (EraCrypto (ShelleyLedgerEra era)))

decodeStakeSnapshot ∷ ∀ era. FromCBOR (StakeSnapshots (EraCrypto (ShelleyLedgerEra era))) ⇒ SerialisedStakeSnapshots era → Either DecoderError (StakeSnapshot era)

data EraHistory mode where

Constructors

EraHistory ∷ ConsensusBlockForMode mode ~ HardForkBlock xs ⇒ ConsensusMode mode → Interpreter xs → EraHistory mode

newtype SystemStart

System start

Slots are counted from the system start.

Constructors

SystemStart

Fields

getSystemStart ∷ UTCTime

newtype LedgerEpochInfo

Constructors

LedgerEpochInfo

Fields

unLedgerEpochInfo ∷ EpochInfo (Either Text)

toLedgerEpochInfo ∷ EraHistory mode → LedgerEpochInfo

newtype SlotsInEpoch

Constructors

SlotsInEpoch Word64

newtype SlotsToEpochEnd

Constructors

SlotsToEpochEnd Word64

slotToEpoch ∷ SlotNo → EraHistory mode → Either PastHorizonException (EpochNo, SlotsInEpoch, SlotsToEpochEnd)

data family LedgerState blk

Ledger state associated with a block

getProgress ∷ SlotNo → EraHistory mode → Either PastHorizonException (RelativeTime, SlotLength)

getSlotForRelativeTime ∷ RelativeTime → EraHistory mode → Either PastHorizonException SlotNo

Returns the slot number for provided relative time from SystemStart

Internal conversion functions

toLedgerUTxO ∷ ShelleyBasedEra era → UTxO era → UTxO (ShelleyLedgerEra era)

fromLedgerUTxO ∷ ShelleyBasedEra era → UTxO (ShelleyLedgerEra era) → UTxO era