Cardano.Api.Query

Description

Queries from local clients to the node.

Queries

data QueryInMode mode result where

Constructors

QueryCurrentEra ∷ ConsensusModeIsMultiEra mode → QueryInMode mode AnyCardanoEra

QueryInEra ∷ EraInMode era mode → QueryInEra era result → QueryInMode mode (Either EraMismatch result)

QueryEraHistory ∷ ConsensusModeIsMultiEra mode → QueryInMode mode (EraHistory mode)

QuerySystemStart ∷ QueryInMode mode SystemStart

QueryChainBlockNo ∷ QueryInMode mode (WithOrigin BlockNo)

QueryChainPoint ∷ ConsensusMode mode → QueryInMode mode ChainPoint

data QueryInEra era result where

Constructors

QueryByronUpdateState ∷ QueryInEra ByronEra ByronUpdateState

QueryInShelleyBasedEra ∷ ShelleyBasedEra era → QueryInShelleyBasedEra era result → QueryInEra era result

data QueryInShelleyBasedEra era result where

Constructors

QueryEpoch ∷ QueryInShelleyBasedEra era EpochNo

QueryGenesisParameters ∷ QueryInShelleyBasedEra era GenesisParameters

QueryProtocolParameters ∷ QueryInShelleyBasedEra era ProtocolParameters

QueryProtocolParametersUpdate ∷ QueryInShelleyBasedEra era (Map (Hash GenesisKey) ProtocolParametersUpdate)

QueryStakeDistribution ∷ QueryInShelleyBasedEra era (Map (Hash StakePoolKey) Rational)

QueryUTxO ∷ QueryUTxOFilter → QueryInShelleyBasedEra era (UTxO era)

QueryStakeAddresses ∷ Set StakeCredential → NetworkId → QueryInShelleyBasedEra era (Map StakeAddress Lovelace, Map StakeAddress PoolId)

QueryStakePools ∷ QueryInShelleyBasedEra era (Set PoolId)

QueryStakePoolParameters ∷ Set PoolId → QueryInShelleyBasedEra era (Map PoolId StakePoolParameters)

QueryDebugLedgerState ∷ QueryInShelleyBasedEra era (SerialisedDebugLedgerState era)

QueryProtocolState ∷ QueryInShelleyBasedEra era (ProtocolState era)

QueryCurrentEpochState ∷ QueryInShelleyBasedEra era (SerialisedCurrentEpochState era)

QueryPoolState ∷ Maybe (Set PoolId) → QueryInShelleyBasedEra era (SerialisedPoolState era)

QueryPoolDistribution ∷ Maybe (Set PoolId) → QueryInShelleyBasedEra era (SerialisedPoolDistribution era)

QueryStakeSnapshot ∷ Maybe (Set PoolId) → QueryInShelleyBasedEra era (SerialisedStakeSnapshots era)

data QueryUTxOFilter

Getting the whole UTxO is obviously not efficient since the result can be huge. Filtering by address is also not efficient because it requires a linear search.

The QueryUTxOFilterByTxIn is efficient since it fits with the structure of the UTxO (which is indexed by TxIn).

Constructors

QueryUTxOWhole
O(n) time and space for utxo size n

QueryUTxOByAddress (Set AddressAny)
O(n) time, O(m) space for utxo size n, and address set size m

QueryUTxOByTxIn (Set TxIn)
O(m log n) time, O(m) space for utxo size n, and address set size m

newtype UTxO era

Constructors

UTxO

Fields

unUTxO ∷ Map TxIn (TxOut CtxUTxO era)

data UTxOInAnyEra where

Constructors

UTxOInAnyEra ∷ CardanoEra era → UTxO era → UTxOInAnyEra

Internal conversion functions

toConsensusQuery ∷ ∀ mode block result. ConsensusBlockForMode mode ~ block ⇒ QueryInMode mode result → Some (Query block)

fromConsensusQueryResult ∷ ∀ mode block result result'. ConsensusBlockForMode mode ~ block ⇒ QueryInMode mode result → Query block result' → result' → result

Wrapper types used in queries

newtype SerialisedDebugLedgerState era

Constructors

SerialisedDebugLedgerState (Serialised (NewEpochState (ShelleyLedgerEra era)))

newtype ProtocolState era

Constructors

ProtocolState (Serialised (ChainDepState (ConsensusProtocol era)))

decodeProtocolState ∷ FromCBOR (ChainDepState (ConsensusProtocol era)) ⇒ ProtocolState era → Either (ByteString, DecoderError) (ChainDepState (ConsensusProtocol era))

data DebugLedgerState era where

Constructors

DebugLedgerState ∷ ShelleyLedgerEra era ~ ledgerera ⇒ NewEpochState ledgerera → DebugLedgerState era

decodeDebugLedgerState ∷ ∀ era. FromCBOR (DebugLedgerState era) ⇒ SerialisedDebugLedgerState era → Either ByteString (DebugLedgerState era)

newtype SerialisedCurrentEpochState era

Constructors

SerialisedCurrentEpochState (Serialised (EpochState (ShelleyLedgerEra era)))

newtype CurrentEpochState era

Constructors

CurrentEpochState (EpochState (ShelleyLedgerEra era))

decodeCurrentEpochState ∷ ∀ era. (Era (ShelleyLedgerEra era), HashAnnotated (TxBody (ShelleyLedgerEra era)) EraIndependentTxBody (Crypto (ShelleyLedgerEra era))) ⇒ FromSharedCBOR (TxOut (ShelleyLedgerEra era)) ⇒ Share (TxOut (ShelleyLedgerEra era)) ~ Interns (Credential 'Staking (Crypto (ShelleyLedgerEra era))) ⇒ FromCBOR (PParams (ShelleyLedgerEra era)) ⇒ FromCBOR (Value (ShelleyLedgerEra era)) ⇒ FromCBOR (State (EraRule "PPUP" (ShelleyLedgerEra era))) ⇒ SerialisedCurrentEpochState era → Either DecoderError (CurrentEpochState era)

newtype SerialisedPoolState era

Constructors

SerialisedPoolState (Serialised (PState (Crypto (ShelleyLedgerEra era))))

newtype PoolState era

Constructors

PoolState (PState (Crypto (ShelleyLedgerEra era)))

decodePoolState ∷ ∀ era. FromCBOR (PState (Crypto (ShelleyLedgerEra era))) ⇒ SerialisedPoolState era → Either DecoderError (PoolState era)

newtype SerialisedPoolDistribution era

Constructors

SerialisedPoolDistribution (Serialised (PoolDistr (Crypto (ShelleyLedgerEra era))))

newtype PoolDistribution era

Constructors

PoolDistribution

Fields

unPoolDistr ∷ PoolDistr (Crypto (ShelleyLedgerEra era))

decodePoolDistribution ∷ ∀ era. FromCBOR (PoolDistr (Crypto (ShelleyLedgerEra era))) ⇒ SerialisedPoolDistribution era → Either DecoderError (PoolDistribution era)

newtype SerialisedStakeSnapshots era

Constructors

SerialisedStakeSnapshots (Serialised (StakeSnapshots (Crypto (ShelleyLedgerEra era))))

newtype StakeSnapshot era

Constructors

StakeSnapshot (StakeSnapshots (Crypto (ShelleyLedgerEra era)))

decodeStakeSnapshot ∷ ∀ era. FromCBOR (StakeSnapshots (Crypto (ShelleyLedgerEra era))) ⇒ SerialisedStakeSnapshots era → Either DecoderError (StakeSnapshot era)

data EraHistory mode where

Constructors

EraHistory ∷ ConsensusBlockForMode mode ~ HardForkBlock xs ⇒ ConsensusMode mode → Interpreter xs → EraHistory mode

newtype SystemStart

Constructors

SystemStart

Fields

getSystemStart ∷ UTCTime

newtype LedgerEpochInfo

Constructors

LedgerEpochInfo

Fields

unLedgerEpochInfo ∷ EpochInfo (Either Text)

toLedgerEpochInfo ∷ EraHistory mode → LedgerEpochInfo

newtype SlotsInEpoch

Constructors

SlotsInEpoch Word64

newtype SlotsToEpochEnd

Constructors

SlotsToEpochEnd Word64

slotToEpoch ∷ SlotNo → EraHistory mode → Either PastHorizonException (EpochNo, SlotsInEpoch, SlotsToEpochEnd)

data family LedgerState blk

getProgress ∷ SlotNo → EraHistory mode → Either PastHorizonException (RelativeTime, SlotLength)

Internal conversion functions

toLedgerUTxO ∷ ShelleyLedgerEra era ~ ledgerera ⇒ Crypto ledgerera ~ StandardCrypto ⇒ ShelleyBasedEra era → UTxO era → UTxO ledgerera

fromLedgerUTxO ∷ ShelleyLedgerEra era ~ ledgerera ⇒ Crypto ledgerera ~ StandardCrypto ⇒ ShelleyBasedEra era → UTxO ledgerera → UTxO era