Cardano.Api.Query.Expr

Documentation

queryChainBlockNo ∷ LocalStateQueryExpr block point (QueryInMode mode) r IO (Either UnsupportedNtcVersionError (WithOrigin BlockNo))

queryChainPoint ∷ LocalStateQueryExpr block point (QueryInMode CardanoMode) r IO (Either UnsupportedNtcVersionError ChainPoint)

queryConstitution ∷ EraInMode era mode → ShelleyBasedEra era → LocalStateQueryExpr block point (QueryInMode mode) r IO (Either UnsupportedNtcVersionError (Either EraMismatch (Maybe (Constitution (ShelleyLedgerEra era)))))

queryCurrentEpochState ∷ EraInMode era mode → ShelleyBasedEra era → LocalStateQueryExpr block point (QueryInMode mode) r IO (Either UnsupportedNtcVersionError (Either EraMismatch (SerialisedCurrentEpochState era)))

queryCurrentEra ∷ LocalStateQueryExpr block point (QueryInMode CardanoMode) r IO (Either UnsupportedNtcVersionError AnyCardanoEra)

queryDebugLedgerState ∷ EraInMode era mode → ShelleyBasedEra era → LocalStateQueryExpr block point (QueryInMode mode) r IO (Either UnsupportedNtcVersionError (Either EraMismatch (SerialisedDebugLedgerState era)))

queryEpoch ∷ EraInMode era mode → ShelleyBasedEra era → LocalStateQueryExpr block point (QueryInMode mode) r IO (Either UnsupportedNtcVersionError (Either EraMismatch EpochNo))

queryConstitutionHash ∷ EraInMode era mode → ShelleyBasedEra era → LocalStateQueryExpr block point (QueryInMode mode) r IO (Either UnsupportedNtcVersionError (Either EraMismatch (Maybe (SafeHash (EraCrypto (ShelleyLedgerEra era)) AnchorData))))

queryEraHistory ∷ LocalStateQueryExpr block point (QueryInMode CardanoMode) r IO (Either UnsupportedNtcVersionError (EraHistory CardanoMode))

queryGenesisParameters ∷ EraInMode era mode → ShelleyBasedEra era → LocalStateQueryExpr block point (QueryInMode mode) r IO (Either UnsupportedNtcVersionError (Either EraMismatch (GenesisParameters ShelleyEra)))

queryPoolDistribution ∷ EraInMode era mode → ShelleyBasedEra era → Maybe (Set PoolId) → LocalStateQueryExpr block point (QueryInMode mode) r IO (Either UnsupportedNtcVersionError (Either EraMismatch (SerialisedPoolDistribution era)))

queryPoolState ∷ EraInMode era mode → ShelleyBasedEra era → Maybe (Set PoolId) → LocalStateQueryExpr block point (QueryInMode mode) r IO (Either UnsupportedNtcVersionError (Either EraMismatch (SerialisedPoolState era)))

queryProtocolParameters ∷ EraInMode era mode → ShelleyBasedEra era → LocalStateQueryExpr block point (QueryInMode mode) r IO (Either UnsupportedNtcVersionError (Either EraMismatch (PParams (ShelleyLedgerEra era))))

queryProtocolParametersUpdate ∷ EraInMode era mode → ShelleyBasedEra era → LocalStateQueryExpr block point (QueryInMode mode) r IO (Either UnsupportedNtcVersionError (Either EraMismatch (Map (Hash GenesisKey) ProtocolParametersUpdate)))

queryProtocolState ∷ EraInMode era mode → ShelleyBasedEra era → LocalStateQueryExpr block point (QueryInMode mode) r IO (Either UnsupportedNtcVersionError (Either EraMismatch (ProtocolState era)))

queryStakeAddresses ∷ EraInMode era mode → ShelleyBasedEra era → Set StakeCredential → NetworkId → LocalStateQueryExpr block point (QueryInMode mode) r IO (Either UnsupportedNtcVersionError (Either EraMismatch (Map StakeAddress Lovelace, Map StakeAddress PoolId)))

queryStakeDelegDeposits ∷ EraInMode era mode → ShelleyBasedEra era → Set StakeCredential → LocalStateQueryExpr block point (QueryInMode mode) r IO (Either UnsupportedNtcVersionError (Either EraMismatch (Map StakeCredential Lovelace)))

queryStakeDistribution ∷ EraInMode era mode → ShelleyBasedEra era → LocalStateQueryExpr block point (QueryInMode mode) r IO (Either UnsupportedNtcVersionError (Either EraMismatch (Map (Hash StakePoolKey) Rational)))

queryStakePoolParameters ∷ EraInMode era mode → ShelleyBasedEra era → Set PoolId → LocalStateQueryExpr block point (QueryInMode mode) r IO (Either UnsupportedNtcVersionError (Either EraMismatch (Map PoolId StakePoolParameters)))

queryStakePools ∷ EraInMode era mode → ShelleyBasedEra era → LocalStateQueryExpr block point (QueryInMode mode) r IO (Either UnsupportedNtcVersionError (Either EraMismatch (Set PoolId)))

queryStakeSnapshot ∷ EraInMode era mode → ShelleyBasedEra era → Maybe (Set PoolId) → LocalStateQueryExpr block point (QueryInMode mode) r IO (Either UnsupportedNtcVersionError (Either EraMismatch (SerialisedStakeSnapshots era)))

querySystemStart ∷ LocalStateQueryExpr block point (QueryInMode mode) r IO (Either UnsupportedNtcVersionError SystemStart)

queryUtxo ∷ EraInMode era mode → ShelleyBasedEra era → QueryUTxOFilter → LocalStateQueryExpr block point (QueryInMode mode) r IO (Either UnsupportedNtcVersionError (Either EraMismatch (UTxO era)))

determineEraExpr ∷ ConsensusModeParams mode → LocalStateQueryExpr block point (QueryInMode mode) r IO (Either UnsupportedNtcVersionError AnyCardanoEra)

A monad expression that determines what era the node is in.

queryCommitteeState ∷ EraInMode era mode → ShelleyBasedEra era → LocalStateQueryExpr block point (QueryInMode mode) r IO (Either UnsupportedNtcVersionError (Either EraMismatch (CommitteeState (ShelleyLedgerEra era))))

queryDRepStakeDistribution

Arguments

∷ EraInMode era mode

→ ShelleyBasedEra era

→ Set (DRep StandardCrypto)
An empty DRep set means that distributions for all DReps will be returned

→ LocalStateQueryExpr block point (QueryInMode mode) r IO (Either UnsupportedNtcVersionError (Either EraMismatch (Map (DRep StandardCrypto) Lovelace)))

queryDRepState

Arguments

∷ EraInMode era mode

→ ShelleyBasedEra era

→ Set (Credential DRepRole StandardCrypto)
An empty credentials set means that states for all DReps will be returned

→ LocalStateQueryExpr block point (QueryInMode mode) r IO (Either UnsupportedNtcVersionError (Either EraMismatch (Map (Credential DRepRole StandardCrypto) (DRepState StandardCrypto))))

queryGovState ∷ EraInMode era mode → ShelleyBasedEra era → LocalStateQueryExpr block point (QueryInMode mode) r IO (Either UnsupportedNtcVersionError (Either EraMismatch (GovState (ShelleyLedgerEra era))))