Cardano.Api.ReexposeLedger

Documentation

data Credential (kr :: KeyRole) c

Script hash or key hash for a payment or a staking object.

Note that credentials (unlike raw key hashes) do appear to vary from era to era, since they reference the hash of a script, which can change. This parameter is a phantom, however, so in actuality the instances will remain the same.

Constructors

ScriptHashObj !(ScriptHash c)

KeyHashObj !(KeyHash kr c)

newtype KeyHash (discriminator :: KeyRole) c

Discriminated hash of public Key

Constructors

KeyHash (Hash (ADDRHASH c) (VerKeyDSIGN (DSIGN c)))

data KeyRole

The role of a key.

Note that a role is not _fixed_, nor is it unique. In particular, keys may variously be used as witnesses, and so in many case we will change the role of a key to the Witness role.

It is also perfectly allowable for a key to be used in many roles; there is nothing prohibiting somebody using the same underlying key as their payment and staking key, as well as the key for their stake pool. So these roles are more intended for two purposes:

To make explicit how we are using a key in the specifications

To provide a guide to downstream implementors, for whom the profusion of keys may be confusing.

Constructors

Genesis

GenesisDelegate

Payment

Staking

StakePool

BlockIssuer

Witness

DRepRole

HotCommitteeRole

ColdCommitteeRole

data ShelleyTxCert era

A heavyweight certificate.

Constructors

ShelleyTxCertDelegCert !(ShelleyDelegCert (EraCrypto era))

ShelleyTxCertPool !(PoolCert (EraCrypto era))

ShelleyTxCertGenesisDeleg !(GenesisDelegCert (EraCrypto era))

ShelleyTxCertMir !(MIRCert (EraCrypto era))

data ShelleyDelegCert c

Constructors

ShelleyRegCert !(StakeCredential c)
A stake credential registration certificate.

ShelleyUnRegCert !(StakeCredential c)
A stake credential deregistration certificate.

ShelleyDelegCert !(StakeCredential c) !(KeyHash 'StakePool c)
A stake delegation certificate.

Bundled Patterns

pattern Delegate :: Delegation c -> ShelleyDelegCert c

pattern DeRegKey :: StakeCredential c -> ShelleyDelegCert c

pattern RegKey :: StakeCredential c -> ShelleyDelegCert c

class EraTxCert era => ShelleyEraTxCert era where

Methods

mkRegTxCert :: StakeCredential (EraCrypto era) -> TxCert era

getRegTxCert :: TxCert era -> Maybe (StakeCredential (EraCrypto era))

mkUnRegTxCert :: StakeCredential (EraCrypto era) -> TxCert era

getUnRegTxCert :: TxCert era -> Maybe (StakeCredential (EraCrypto era))

mkDelegStakeTxCert :: StakeCredential (EraCrypto era) -> KeyHash 'StakePool (EraCrypto era) -> TxCert era

getDelegStakeTxCert :: TxCert era -> Maybe (StakeCredential (EraCrypto era), KeyHash 'StakePool (EraCrypto era))

mkGenesisDelegTxCert :: GenesisDelegCert (EraCrypto era) -> TxCert era

getGenesisDelegTxCert :: TxCert era -> Maybe (GenesisDelegCert (EraCrypto era))

mkMirTxCert :: MIRCert (EraCrypto era) -> TxCert era

getMirTxCert :: TxCert era -> Maybe (MIRCert (EraCrypto era))

data GenesisDelegCert c

Genesis key delegation certificate

Constructors

GenesisDelegCert !(KeyHash 'Genesis c) !(KeyHash 'GenesisDelegate c) !(Hash c (VerKeyVRF c))

data PoolParams c

A stake pool.

Constructors

PoolParams

Fields

ppId :: !(KeyHash 'StakePool c)

ppVrf :: !(Hash c (VerKeyVRF c))

ppPledge :: !Coin

ppCost :: !Coin

ppMargin :: !UnitInterval

ppRewardAcnt :: !(RewardAcnt c)

ppOwners :: !(Set (KeyHash 'Staking c))

ppRelays :: !(StrictSeq StakePoolRelay)

ppMetadata :: !(StrictMaybe PoolMetadata)

class HasKeyRole (a :: KeyRole -> Type -> Type)

data MIRPot

Constructors

ReservesMIR

TreasuryMIR

data MIRTarget c

MIRTarget specifies if funds from either the reserves or the treasury are to be handed out to a collection of reward accounts or instead transfered to the opposite pot.

Constructors

StakeAddressesMIR !(Map (Credential 'Staking c) DeltaCoin)

SendToOppositePotMIR !Coin

data MIRCert c

Move instantaneous rewards certificate

Constructors

MIRCert

Fields

mirPot :: !MIRPot

mirRewards :: !(MIRTarget c)

data StakePoolRelay

Constructors

SingleHostAddr !(StrictMaybe Port) !(StrictMaybe IPv4) !(StrictMaybe IPv6)
One or both of IPv4 & IPv6

SingleHostName !(StrictMaybe Port) !DnsName
An A or AAAA DNS record

MultiHostName !DnsName
A SRV DNS record

data PoolMetadata

Constructors

PoolMetadata

Fields

pmUrl :: !Url

pmHash :: !ByteString

class (Era era, DecCBOR (TxCert era), EncCBOR (TxCert era), ToCBOR (TxCert era), FromCBOR (TxCert era), NoThunks (TxCert era), NFData (TxCert era), Show (TxCert era), Eq (TxCert era), ToExpr (TxCert era)) => EraTxCert era where

Associated Types

type TxCert era = (r :: Type) | r -> era

type TxCertUpgradeError era

type TxCertUpgradeError era = Void

Methods

upgradeTxCert :: TxCert (PreviousEra era) -> Either (TxCertUpgradeError era) (TxCert era)

Every era, except Shelley, must be able to upgrade a TxCert from a previous era. However, not all certificates can be upgraded, because some eras lose some of the certificates, thus return type is an Either. Eg. from Babbage to Conway: MIR and Genesis certificates were removed.

getVKeyWitnessTxCert :: TxCert era -> Maybe (KeyHash 'Witness (EraCrypto era))

Return a witness key whenever a certificate requires one

getScriptWitnessTxCert :: TxCert era -> Maybe (ScriptHash (EraCrypto era))

Return a ScriptHash for certificate types that require a witness

mkRegPoolTxCert :: PoolParams (EraCrypto era) -> TxCert era

getRegPoolTxCert :: TxCert era -> Maybe (PoolParams (EraCrypto era))

mkRetirePoolTxCert :: KeyHash 'StakePool (EraCrypto era) -> EpochNo -> TxCert era

getRetirePoolTxCert :: TxCert era -> Maybe (KeyHash 'StakePool (EraCrypto era), EpochNo)

lookupRegStakeTxCert :: TxCert era -> Maybe (Credential 'Staking (EraCrypto era))

Extract staking credential from any certificate that can register such credential

lookupUnRegStakeTxCert :: TxCert era -> Maybe (Credential 'Staking (EraCrypto era))

Extract staking credential from any certificate that can unregister such credential

getTotalDepositsTxCerts

Arguments

:: Foldable f

=> PParams era

-> (KeyHash 'StakePool (EraCrypto era) -> Bool)
Check whether stake pool is registered or not

-> f (TxCert era)

-> Coin

Compute the total deposits from a list of certificates.

getTotalRefundsTxCerts

Arguments

:: Foldable f

=> PParams era

-> (Credential 'Staking (EraCrypto era) -> Maybe Coin)
Lookup current deposit for Staking credential if one is registered

-> (Credential 'DRepRole (EraCrypto era) -> Maybe Coin)
Lookup current deposit for DRep credential if one is registered

-> f (TxCert era)

-> Coin

Compute the total refunds from a list of certificates.

data StrictMaybe a

Constructors

SNothing

SJust !a

pattern DelegTxCert :: ConwayEraTxCert era => StakeCredential (EraCrypto era) -> Delegatee (EraCrypto era) -> TxCert era

pattern RegPoolTxCert :: EraTxCert era => PoolParams (EraCrypto era) -> TxCert era

pattern RetirePoolTxCert :: EraTxCert era => KeyHash 'StakePool (EraCrypto era) -> EpochNo -> TxCert era

pattern RegDepositTxCert :: ConwayEraTxCert era => StakeCredential (EraCrypto era) -> Coin -> TxCert era

pattern UnRegDepositTxCert :: ConwayEraTxCert era => StakeCredential (EraCrypto era) -> Coin -> TxCert era

pattern UnRegDRepTxCert :: ConwayEraTxCert era => Credential 'DRepRole (EraCrypto era) -> Coin -> TxCert era

pattern AuthCommitteeHotKeyTxCert :: ConwayEraTxCert era => Credential 'ColdCommitteeRole (EraCrypto era) -> Credential 'HotCommitteeRole (EraCrypto era) -> TxCert era

pattern ResignCommitteeColdTxCert :: ConwayEraTxCert era => Credential 'ColdCommitteeRole (EraCrypto era) -> StrictMaybe (Anchor (EraCrypto era)) -> TxCert era

pattern RegTxCert :: ShelleyEraTxCert era => StakeCredential (EraCrypto era) -> TxCert era

pattern UnRegTxCert :: ShelleyEraTxCert era => StakeCredential (EraCrypto era) -> TxCert era

pattern DelegStakeTxCert :: ShelleyEraTxCert era => StakeCredential (EraCrypto era) -> KeyHash 'StakePool (EraCrypto era) -> TxCert era

pattern RegDepositDelegTxCert :: ConwayEraTxCert era => StakeCredential (EraCrypto era) -> Delegatee (EraCrypto era) -> Coin -> TxCert era

pattern RegDRepTxCert :: ConwayEraTxCert era => Credential 'DRepRole (EraCrypto era) -> Coin -> StrictMaybe (Anchor (EraCrypto era)) -> TxCert era

pattern MirTxCert :: (ShelleyEraTxCert era, ProtVerAtMost era 8) => MIRCert (EraCrypto era) -> TxCert era

pattern GenesisDelegTxCert :: (ShelleyEraTxCert era, ProtVerAtMost era 8) => KeyHash 'Genesis (EraCrypto era) -> KeyHash 'GenesisDelegate (EraCrypto era) -> Hash (EraCrypto era) (VerKeyVRF (EraCrypto era)) -> TxCert era

pattern UpdateDRepTxCert :: ConwayEraTxCert era => Credential 'DRepRole (EraCrypto era) -> StrictMaybe (Anchor (EraCrypto era)) -> TxCert era

newtype Coin

The amount of value held by a transaction output.

Constructors

Coin

Fields

unCoin :: Integer

type family EraCrypto era

data Network

Constructors

Testnet

Mainnet

data PoolCert c

Constructors

RegPool !(PoolParams c)
A stake pool registration certificate.

RetirePool !(KeyHash 'StakePool c) !EpochNo
A stake pool retirement certificate.

newtype PParams era

Protocol parameters

Constructors

PParams (PParamsHKD Identity era)

data PParamsUpdate era

The type of updates to Protocol parameters

type family Value era

A value is something which quantifies a transaction output.

addDeltaCoin :: Coin -> DeltaCoin -> Coin

toDeltaCoin :: Coin -> DeltaCoin

toEraCBOR :: (Era era, EncCBOR t) => t -> Encoding

Convert a type that implements EncCBOR to plain Encoding using the lowest protocol version for the supplied era

fromEraCBOR :: (Era era, DecCBOR t) => Decoder s t

Convert a type that implements DecCBOR to plain Decoder using the lowest protocol version for the supplied era

data Anchor c

Constructors

Anchor

Fields

anchorUrl :: !Url

anchorDataHash :: !(SafeHash c AnchorData)

data Delegatee c

First type argument is the deposit

Constructors

DelegStake !(KeyHash 'StakePool c)

DelegVote !(DRep c)

DelegStakeVote !(KeyHash 'StakePool c) !(DRep c)

data DRep c

Constructors

DRepAlwaysAbstain

DRepAlwaysNoConfidence

Bundled Patterns

pattern DRepCredential :: Credential 'DRepRole c -> DRep c

data DRepState c

data ConwayTxCert era

Constructors

ConwayTxCertDeleg !(ConwayDelegCert (EraCrypto era))

ConwayTxCertPool !(PoolCert (EraCrypto era))

ConwayTxCertGov !(ConwayGovCert (EraCrypto era))

data ConwayDelegCert c

Certificates for registration and delegation of stake to Pools and DReps. Comparing to previous eras, there is now ability to:

Register and delegate with a single certificate: ConwayRegDelegCert

Ability to delegate to DReps with DelegVote and DelegStakeVote

Ability to specify the deposit amount. Deposits during registration and unregistration in Conway are optional, which will change in the future era. They are optional only for the smooth transition from Babbage to Conway. Validity of deposits is checked by the CERT rule.

Constructors

ConwayRegCert !(StakeCredential c) !(StrictMaybe Coin)
Register staking credential. Deposit, when present, must match the expected deposit amount specified by ppKeyDepositL in the protocol parameters.

ConwayUnRegCert !(StakeCredential c) !(StrictMaybe Coin)
De-Register the staking credential. Deposit, if present, must match the amount that was left as a deposit upon stake credential registration.

ConwayDelegCert !(StakeCredential c) !(Delegatee c)
Redelegate to another delegatee. Staking credential must already be registered.

ConwayRegDelegCert !(StakeCredential c) !(Delegatee c) !Coin
This is a new type of certificate, which allows to register staking credential and delegate within a single certificate. Deposit is required and must match the expected deposit amount specified by ppKeyDepositL in the protocol parameters.

class ShelleyEraTxCert era => ConwayEraTxCert era where

Methods

mkRegDepositTxCert :: StakeCredential (EraCrypto era) -> Coin -> TxCert era

getRegDepositTxCert :: TxCert era -> Maybe (StakeCredential (EraCrypto era), Coin)

mkUnRegDepositTxCert :: StakeCredential (EraCrypto era) -> Coin -> TxCert era

getUnRegDepositTxCert :: TxCert era -> Maybe (StakeCredential (EraCrypto era), Coin)

mkDelegTxCert :: StakeCredential (EraCrypto era) -> Delegatee (EraCrypto era) -> TxCert era

getDelegTxCert :: TxCert era -> Maybe (StakeCredential (EraCrypto era), Delegatee (EraCrypto era))

mkRegDepositDelegTxCert :: StakeCredential (EraCrypto era) -> Delegatee (EraCrypto era) -> Coin -> TxCert era

getRegDepositDelegTxCert :: TxCert era -> Maybe (StakeCredential (EraCrypto era), Delegatee (EraCrypto era), Coin)

mkAuthCommitteeHotKeyTxCert :: Credential 'ColdCommitteeRole (EraCrypto era) -> Credential 'HotCommitteeRole (EraCrypto era) -> TxCert era

getAuthCommitteeHotKeyTxCert :: TxCert era -> Maybe (Credential 'ColdCommitteeRole (EraCrypto era), Credential 'HotCommitteeRole (EraCrypto era))

mkResignCommitteeColdTxCert :: Credential 'ColdCommitteeRole (EraCrypto era) -> StrictMaybe (Anchor (EraCrypto era)) -> TxCert era

getResignCommitteeColdTxCert :: TxCert era -> Maybe (Credential 'ColdCommitteeRole (EraCrypto era), StrictMaybe (Anchor (EraCrypto era)))

mkRegDRepTxCert :: Credential 'DRepRole (EraCrypto era) -> Coin -> StrictMaybe (Anchor (EraCrypto era)) -> TxCert era

getRegDRepTxCert :: TxCert era -> Maybe (Credential 'DRepRole (EraCrypto era), Coin, StrictMaybe (Anchor (EraCrypto era)))

mkUnRegDRepTxCert :: Credential 'DRepRole (EraCrypto era) -> Coin -> TxCert era

getUnRegDRepTxCert :: TxCert era -> Maybe (Credential 'DRepRole (EraCrypto era), Coin)

mkUpdateDRepTxCert :: Credential 'DRepRole (EraCrypto era) -> StrictMaybe (Anchor (EraCrypto era)) -> TxCert era

getUpdateDRepTxCert :: TxCert era -> Maybe (Credential 'DRepRole (EraCrypto era), StrictMaybe (Anchor (EraCrypto era)))

data ConwayGovCert c

Constructors

ConwayRegDRep !(Credential 'DRepRole c) !Coin !(StrictMaybe (Anchor c))

ConwayUnRegDRep !(Credential 'DRepRole c) !Coin

ConwayUpdateDRep !(Credential 'DRepRole c) !(StrictMaybe (Anchor c))

ConwayAuthCommitteeHotKey !(Credential 'ColdCommitteeRole c) !(Credential 'HotCommitteeRole c)

ConwayResignCommitteeColdKey !(Credential 'ColdCommitteeRole c) !(StrictMaybe (Anchor c))

type family GovState era = (r :: Type) | r -> era

data GovActionId c

Constructors

GovActionId

Fields

gaidTxId :: !(TxId c)

gaidGovActionIx :: !GovActionIx

data Vote

Constructors

VoteNo

VoteYes

Abstain

data Voter c

Constructors

CommitteeVoter !(Credential 'HotCommitteeRole c)

DRepVoter !(Credential 'DRepRole c)

StakePoolVoter !(KeyHash 'StakePool c)

data VotingProcedure era

Constructors

VotingProcedure

Fields

vProcVote :: !Vote

vProcAnchor :: !(StrictMaybe (Anchor (EraCrypto era)))

data PoolVotingThresholds

Constructors

PoolVotingThresholds

Fields

pvtMotionNoConfidence :: !UnitInterval

pvtCommitteeNormal :: !UnitInterval

pvtCommitteeNoConfidence :: !UnitInterval

pvtHardForkInitiation :: !UnitInterval

data DRepVotingThresholds

Constructors

DRepVotingThresholds

Fields

dvtMotionNoConfidence :: !UnitInterval

dvtCommitteeNormal :: !UnitInterval

dvtCommitteeNoConfidence :: !UnitInterval

dvtUpdateToConstitution :: !UnitInterval

dvtHardForkInitiation :: !UnitInterval

dvtPPNetworkGroup :: !UnitInterval

dvtPPEconomicGroup :: !UnitInterval

dvtPPTechnicalGroup :: !UnitInterval

dvtPPGovGroup :: !UnitInterval

dvtTreasuryWithdrawal :: !UnitInterval

dvtPPNetworkGroupL :: Lens' DRepVotingThresholds UnitInterval

dvtPPGovGroupL :: Lens' DRepVotingThresholds UnitInterval

dvtPPTechnicalGroupL :: Lens' DRepVotingThresholds UnitInterval

dvtPPEconomicGroupL :: Lens' DRepVotingThresholds UnitInterval

dvtUpdateToConstitutionL :: Lens' DRepVotingThresholds UnitInterval

drepExpiryL :: forall c f. Functor f => (EpochNo -> f EpochNo) -> DRepState c -> f (DRepState c)

drepAnchorL :: forall c f. Functor f => (StrictMaybe (Anchor c) -> f (StrictMaybe (Anchor c))) -> DRepState c -> f (DRepState c)

drepDepositL :: forall c f. Functor f => (Coin -> f Coin) -> DRepState c -> f (DRepState c)

csCommitteeCredsL :: forall era f. Functor f => (Map (Credential 'ColdCommitteeRole (EraCrypto era)) (Maybe (Credential 'HotCommitteeRole (EraCrypto era))) -> f (Map (Credential 'ColdCommitteeRole (EraCrypto era)) (Maybe (Credential 'HotCommitteeRole (EraCrypto era))))) -> CommitteeState era -> f (CommitteeState era)

data Annotated b a

data Tx

Transaction

NB: transaction witnesses are stored separately

Constructors

UnsafeTx

Fields

txInputs :: !(NonEmpty TxIn)

Inputs of transaction.

txOutputs :: !(NonEmpty TxOut)

Outputs of transaction.

txAttributes :: !TxAttributes

Attributes of transaction

newtype CoinPerByte

Constructors

CoinPerByte

Fields

unCoinPerByte :: Coin

newtype CoinPerWord

Constructors

CoinPerWord

Fields

unCoinPerWord :: Coin

data Prices

Prices per execution unit

Constructors

Prices

Fields

prMem :: !NonNegativeInterval

prSteps :: !NonNegativeInterval

boundRational :: BoundedRational r => Rational -> Maybe r

Returns Nothing when supplied value is not within bounds or when precision is too high to be represented by the underlying type

Example

Expand

>>> :set -XTypeApplications
>>> import Data.Ratio
>>> boundRational @UnitInterval $ 2 % 3
Just (2 % 3)
>>> boundRational @UnitInterval (-0.5)
Nothing
>>> boundRational @UnitInterval (1.5)
Nothing
>>> boundRational @UnitInterval 0
Just (0 % 1)
>>> boundRational @PositiveUnitInterval 0
Nothing

unboundRational :: BoundedRational r => r -> Rational

Promote bounded rational type into the unbounded Rational.

data DnsName

dnsToText :: DnsName -> Text

textToDns :: Text -> Maybe DnsName

data Url

urlToText :: Url -> Text

textToUrl :: Text -> Maybe Url

portToWord16 :: Port -> Word16

strictMaybeToMaybe :: StrictMaybe a -> Maybe a

maybeToStrictMaybe :: Maybe a -> StrictMaybe a

hashToBytes :: Hash h a -> ByteString

hashFromBytes :: HashAlgorithm h => ByteString -> Maybe (Hash h a)

class (HashAlgorithm (HASH c), HashAlgorithm (ADDRHASH c), DSIGNAlgorithm (DSIGN c), KESAlgorithm (KES c), VRFAlgorithm (VRF c), ContextDSIGN (DSIGN c) ~ (), ContextKES (KES c) ~ (), ContextVRF (VRF c) ~ (), Typeable c) => Crypto c

data StandardCrypto

The same crypto used on the net

newtype EpochNo

Constructors

EpochNo

Fields

unEpochNo :: Word64