Cardano.Api.Rewards

Documentation

newtype DelegationsAndRewards

A mapping of Shelley reward accounts to both the stake pool that they delegate to and their reward account balance. TODO: Move to cardano-api

Constructors

DelegationsAndRewards (Map StakeAddress Lovelace, Map StakeAddress PoolId)

mergeDelegsAndRewards ∷ DelegationsAndRewards → [(StakeAddress, Maybe Lovelace, Maybe PoolId)]