Cardano.Api.Script

Languages

data SimpleScript'

data PlutusScriptV1

The original simple script language which supports

require a signature from a given key (by verification key hash)

n-way and combinator

n-way or combinator

m-of-n combinator

This version of the language was introduced in the ShelleyEra.

The second version of the simple script language. It has all the features of the original simple script language plus new atomic predicates:

require the time be before a given slot number

require the time be after a given slot number

This version of the language was introduced in the AllegraEra.

However we opt for a single type level tag SimpleScript' as the second version of of the language introduced in the Allegra era is a superset of the language introduced in the Shelley era.

Place holder type to show what the pattern is to extend to multiple languages, not just multiple versions of a single language.

data PlutusScriptV2

data ScriptLanguage lang where

Constructors

SimpleScriptLanguage ∷ ScriptLanguage SimpleScript'

PlutusScriptLanguage ∷ PlutusScriptVersion lang → ScriptLanguage lang

data PlutusScriptVersion lang where

Constructors

PlutusScriptV1 ∷ PlutusScriptVersion PlutusScriptV1

PlutusScriptV2 ∷ PlutusScriptVersion PlutusScriptV2

data AnyScriptLanguage where

Constructors

AnyScriptLanguage ∷ ScriptLanguage lang → AnyScriptLanguage

data AnyPlutusScriptVersion where

Constructors

AnyPlutusScriptVersion ∷ PlutusScriptVersion lang → AnyPlutusScriptVersion

class IsScriptLanguage lang ⇒ IsPlutusScriptLanguage lang where

Methods

plutusScriptVersion ∷ PlutusScriptVersion lang

class HasTypeProxy lang ⇒ IsScriptLanguage lang where

Methods

scriptLanguage ∷ ScriptLanguage lang

Scripts in a specific language

data Script lang where

A script in a particular language.

See also ScriptInAnyLang for a script in any of the known languages.

See also ScriptInEra for a script in a language that is available within a particular era.

Note that some but not all scripts have an external JSON syntax, hence this type has no JSON serialisation instances. The Script family of languages do have a JSON syntax and thus have ToJSON/FromJSON instances.

Constructors

SimpleScript ∷ !SimpleScript → Script SimpleScript'

PlutusScript ∷ !(PlutusScriptVersion lang) → !(PlutusScript lang) → Script lang

Scripts in any language

data ScriptInAnyLang where

Sometimes it is necessary to handle all languages without making static type distinctions between languages. For example, when reading external input, or before the era context is known.

Use toScriptInEra to convert to a script in the context of an era.

Constructors

ScriptInAnyLang ∷ ScriptLanguage lang → Script lang → ScriptInAnyLang

toScriptInAnyLang ∷ Script lang → ScriptInAnyLang

Convert a script in a specific statically-known language to a ScriptInAnyLang.

No inverse to this is provided, just do case analysis on the ScriptLanguage field within the ScriptInAnyLang constructor.

Scripts in an era

data ScriptInEra era where

Constructors

ScriptInEra ∷ ScriptLanguageInEra lang era → Script lang → ScriptInEra era

toScriptInEra ∷ CardanoEra era → ScriptInAnyLang → Maybe (ScriptInEra era)

Given a target era and a script in some language, check if the language is supported in that era, and if so return a ScriptInEra.

eraOfScriptInEra ∷ ScriptInEra era → ShelleyBasedEra era

Reference scripts

data ReferenceScript era where

A reference scripts is a script that can exist at a transaction output. This greatly reduces the size of transactions that use scripts as the script no longer has to be added to the transaction, they can now be referenced via a transaction output.

Constructors

ReferenceScript ∷ ReferenceTxInsScriptsInlineDatumsSupportedInEra era → ScriptInAnyLang → ReferenceScript era

ReferenceScriptNone ∷ ReferenceScript era

data ReferenceTxInsScriptsInlineDatumsSupportedInEra era where

Constructors

ReferenceTxInsScriptsInlineDatumsInBabbageEra ∷ ReferenceTxInsScriptsInlineDatumsSupportedInEra BabbageEra

ReferenceTxInsScriptsInlineDatumsInConwayEra ∷ ReferenceTxInsScriptsInlineDatumsSupportedInEra ConwayEra

refInsScriptsAndInlineDatsSupportedInEra ∷ CardanoEra era → Maybe (ReferenceTxInsScriptsInlineDatumsSupportedInEra era)

refScriptToShelleyScript ∷ CardanoEra era → ReferenceScript era → StrictMaybe (Script (ShelleyLedgerEra era))

Use of a script in an era as a witness

data WitCtxTxIn

A tag type for the context in which a script is used in a transaction.

This type tags the context as being to witness a transaction input.

data WitCtxMint

A tag type for the context in which a script is used in a transaction.

This type tags the context as being to witness minting.

data WitCtxStake

A tag type for the context in which a script is used in a transaction.

This type tags the context as being to witness the use of stake addresses in both certificates and withdrawals.

data WitCtx witctx where

This GADT provides a value-level representation of all the witness contexts. This enables pattern matching on the context to allow them to be treated in a non-uniform way.

Constructors

WitCtxTxIn ∷ WitCtx WitCtxTxIn

WitCtxMint ∷ WitCtx WitCtxMint

WitCtxStake ∷ WitCtx WitCtxStake

data ScriptWitness witctx era where

A use of a script within a transaction body to witness that something is being used in an authorised manner. That can be

spending a transaction input

minting tokens

using a certificate (stake address certs specifically)

withdrawing from a reward account

For simple script languages, the use of the script is the same in all contexts. For Plutus scripts, using a script involves supplying a redeemer. In addition, Plutus scripts used for spending inputs must also supply the datum value used when originally creating the TxOut that is now being spent.

Constructors

SimpleScriptWitness ∷ ScriptLanguageInEra SimpleScript' era → SimpleScriptOrReferenceInput SimpleScript' → ScriptWitness witctx era

PlutusScriptWitness ∷ ScriptLanguageInEra lang era → PlutusScriptVersion lang → PlutusScriptOrReferenceInput lang → ScriptDatum witctx → ScriptRedeemer → ExecutionUnits → ScriptWitness witctx era

data Witness witctx era where

Constructors

KeyWitness ∷ KeyWitnessInCtx witctx → Witness witctx era

ScriptWitness ∷ ScriptWitnessInCtx witctx → ScriptWitness witctx era → Witness witctx era

data KeyWitnessInCtx witctx where

Constructors

KeyWitnessForSpending ∷ KeyWitnessInCtx WitCtxTxIn

KeyWitnessForStakeAddr ∷ KeyWitnessInCtx WitCtxStake

data ScriptWitnessInCtx witctx where

Constructors

ScriptWitnessForSpending ∷ ScriptWitnessInCtx WitCtxTxIn

ScriptWitnessForMinting ∷ ScriptWitnessInCtx WitCtxMint

ScriptWitnessForStakeAddr ∷ ScriptWitnessInCtx WitCtxStake

class IsScriptWitnessInCtx ctx where

Methods

scriptWitnessInCtx ∷ ScriptWitnessInCtx ctx

data ScriptDatum witctx where

Constructors

ScriptDatumForTxIn ∷ HashableScriptData → ScriptDatum WitCtxTxIn

InlineScriptDatum ∷ ScriptDatum WitCtxTxIn

NoScriptDatumForMint ∷ ScriptDatum WitCtxMint

NoScriptDatumForStake ∷ ScriptDatum WitCtxStake

type ScriptRedeemer = HashableScriptData

scriptWitnessScript ∷ ScriptWitness witctx era → Maybe (ScriptInEra era)

Languages supported in each era

data ScriptLanguageInEra lang era where

Constructors

SimpleScriptInShelley ∷ ScriptLanguageInEra SimpleScript' ShelleyEra

SimpleScriptInAllegra ∷ ScriptLanguageInEra SimpleScript' AllegraEra

SimpleScriptInMary ∷ ScriptLanguageInEra SimpleScript' MaryEra

SimpleScriptInAlonzo ∷ ScriptLanguageInEra SimpleScript' AlonzoEra

SimpleScriptInBabbage ∷ ScriptLanguageInEra SimpleScript' BabbageEra

SimpleScriptInConway ∷ ScriptLanguageInEra SimpleScript' ConwayEra

PlutusScriptV1InAlonzo ∷ ScriptLanguageInEra PlutusScriptV1 AlonzoEra

PlutusScriptV1InBabbage ∷ ScriptLanguageInEra PlutusScriptV1 BabbageEra

PlutusScriptV1InConway ∷ ScriptLanguageInEra PlutusScriptV1 ConwayEra

PlutusScriptV2InBabbage ∷ ScriptLanguageInEra PlutusScriptV2 BabbageEra

PlutusScriptV2InConway ∷ ScriptLanguageInEra PlutusScriptV2 ConwayEra

scriptLanguageSupportedInEra ∷ CardanoEra era → ScriptLanguage lang → Maybe (ScriptLanguageInEra lang era)

Check if a given script language is supported in a given era, and if so return the evidence.

languageOfScriptLanguageInEra ∷ ScriptLanguageInEra lang era → ScriptLanguage lang

eraOfScriptLanguageInEra ∷ ScriptLanguageInEra lang era → ShelleyBasedEra era

The simple script language

data SimpleScript

Constructors

RequireSignature !(Hash PaymentKey)

RequireTimeBefore !SlotNo

RequireTimeAfter !SlotNo

RequireAllOf ![SimpleScript]

RequireAnyOf ![SimpleScript]

RequireMOf !Int ![SimpleScript]

data SimpleScriptOrReferenceInput lang

Constructors

SScript SimpleScript

SReferenceScript TxIn (Maybe ScriptHash)

The Plutus script language

data PlutusScript lang where

Plutus scripts.

Note that Plutus scripts have a binary serialisation but no JSON serialisation.

Constructors

PlutusScriptSerialised ∷ ShortByteString → PlutusScript lang

data PlutusScriptOrReferenceInput lang

Scripts can now exist in the UTxO at a transaction output. We can reference these scripts via specification of a reference transaction input in order to witness spending inputs, withdrawals, certificates or to mint tokens. This datatype encapsulates this concept.

Constructors

PScript (PlutusScript lang)

PReferenceScript TxIn (Maybe ScriptHash)

examplePlutusScriptAlwaysSucceeds ∷ WitCtx witctx → PlutusScript PlutusScriptV1

An example Plutus script that always succeeds, irrespective of inputs.

For example, if one were to use this for a payment address then it would allow anyone to spend from it.

The exact script depends on the context in which it is to be used.

examplePlutusScriptAlwaysFails ∷ WitCtx witctx → PlutusScript PlutusScriptV1

An example Plutus script that always fails, irrespective of inputs.

For example, if one were to use this for a payment address then it would be impossible for anyone to ever spend from it.

The exact script depends on the context in which it is to be used.

Script data

data ScriptData

Constructors

ScriptDataConstructor

Fields

Integer

Tag for the constructor

[ScriptData]

Constructor arguments

ScriptDataMap [(ScriptData, ScriptData)]
Key value pairs

ScriptDataList [ScriptData]
Elements

ScriptDataNumber Integer

ScriptDataBytes ByteString

Script execution units

data ExecutionUnits

The units for how long a script executes for and how much memory it uses. This is used to declare the resources used by a particular use of a script.

This type is also used to describe the limits for the maximum overall execution units per transaction or per block.

Constructors

ExecutionUnits

Fields

executionSteps ∷ Natural

This corresponds roughly to the time to execute a script.

executionMemory ∷ Natural

This corresponds roughly to the peak memory used during script execution.

Script hashes

newtype ScriptHash

We have this type separate from the Hash type to avoid the script hash type being parametrised by the era. The representation is era independent, and there are many places where we want to use a script hash where we don't want things to be era-parametrised.

Constructors

ScriptHash (ScriptHash StandardCrypto)

hashScript ∷ Script lang → ScriptHash

Internal conversion functions

toShelleyScript ∷ ScriptInEra era → Script (ShelleyLedgerEra era)

fromShelleyBasedScript ∷ ShelleyBasedEra era → Script (ShelleyLedgerEra era) → ScriptInEra era

toShelleyMultiSig ∷ SimpleScript → Either MultiSigError (MultiSig StandardCrypto)

Conversion for the MultiSig language used by the Shelley era.

fromShelleyMultiSig ∷ MultiSig StandardCrypto → SimpleScript

Conversion for the MultiSig language used by the Shelley era.

toAllegraTimelock ∷ SimpleScript → Timelock StandardCrypto

Conversion for the Timelock language that is shared between the Allegra and Mary eras.

fromAllegraTimelock ∷ Timelock StandardCrypto → SimpleScript

Conversion for the Timelock language that is shared between the Allegra and Mary eras.

toAlonzoExUnits ∷ ExecutionUnits → ExUnits

fromAlonzoExUnits ∷ ExUnits → ExecutionUnits

toShelleyScriptHash ∷ ScriptHash → ScriptHash StandardCrypto

fromShelleyScriptHash ∷ ScriptHash StandardCrypto → ScriptHash

toPlutusData ∷ ScriptData → Data

fromPlutusData ∷ Data → ScriptData

toAlonzoData ∷ HashableScriptData → Data ledgerera

fromAlonzoData ∷ Data ledgerera → HashableScriptData

toAlonzoLanguage ∷ AnyPlutusScriptVersion → Language

fromAlonzoLanguage ∷ Language → AnyPlutusScriptVersion

fromShelleyScriptToReferenceScript ∷ ShelleyBasedEra era → Script (ShelleyLedgerEra era) → ReferenceScript era

Data family instances

data family AsType t

A family of singleton types used in this API to indicate which type to use where it would otherwise be ambiguous or merely unclear.

Values of this type are passed to deserialisation functions for example.

data family Hash keyrole ∷ Type