Cardano.Api.Script

Languages

data SimpleScriptV1

The original simple script language which supports

require a signature from a given key (by verification key hash)

n-way and combinator

n-way or combinator

m-of-n combinator

This version of the language was introduced in the ShelleyEra.

data SimpleScriptV2

The second version of the simple script language. It has all the features of SimpleScriptV1 plus new atomic predicates:

require the time be before a given slot number

require the time be after a given slot number

This version of the language was introduced in the AllegraEra.

data ScriptLanguage lang where

Constructors

SimpleScriptLanguage ∷ SimpleScriptVersion lang → ScriptLanguage lang

PlutusScriptLanguage ∷ PlutusScriptVersion lang → ScriptLanguage lang

data SimpleScriptVersion lang where

Constructors

SimpleScriptV1 ∷ SimpleScriptVersion SimpleScriptV1

SimpleScriptV2 ∷ SimpleScriptVersion SimpleScriptV2

data PlutusScriptVersion lang

data AnyScriptLanguage where

Constructors

AnyScriptLanguage ∷ ScriptLanguage lang → AnyScriptLanguage

class HasTypeProxy lang ⇒ IsScriptLanguage lang where

Methods

scriptLanguage ∷ ScriptLanguage lang

class IsScriptLanguage lang ⇒ IsSimpleScriptLanguage lang where

Methods

simpleScriptVersion ∷ SimpleScriptVersion lang

Scripts in a specific language

data Script lang where

A script in a particular language.

See also ScriptInAnyLang for a script in any of the languages that is available within a particular era.

See also ScriptInEra for a script in a language that is available within a particular era.

Constructors

SimpleScript ∷ !(SimpleScriptVersion lang) → !(SimpleScript lang) → Script lang

PlutusScript ∷ !(PlutusScriptVersion lang) → () → Script lang

Scripts in any language

data ScriptInAnyLang where

Sometimes it is necessary to handle all languages without making static type distinctions between languages. For example, when reading external input, or before the era context is known.

Use toScriptInEra to convert to a script in the context of an era.

Constructors

ScriptInAnyLang ∷ ScriptLanguage lang → Script lang → ScriptInAnyLang

toScriptInAnyLang ∷ Script lang → ScriptInAnyLang

Convert a script in a specific statically-known language to a ScriptInAnyLang.

No inverse to this is provided, just do case analysis on the ScriptLanguage field within the ScriptInAnyLang constructor.

Scripts in an era

data ScriptInEra era where

Constructors

ScriptInEra ∷ ScriptLanguageInEra lang era → Script lang → ScriptInEra era

toScriptInEra ∷ CardanoEra era → ScriptInAnyLang → Maybe (ScriptInEra era)

Given a target era and a script in some language, check if the language is supported in that era, and if so return a ScriptInEra.

eraOfScriptInEra ∷ ScriptInEra era → ShelleyBasedEra era

Languages supported in each era

data ScriptLanguageInEra lang era where

Constructors

SimpleScriptV1InShelley ∷ ScriptLanguageInEra SimpleScriptV1 ShelleyEra

SimpleScriptV1InAllegra ∷ ScriptLanguageInEra SimpleScriptV1 AllegraEra

SimpleScriptV1InMary ∷ ScriptLanguageInEra SimpleScriptV1 MaryEra

SimpleScriptV2InAllegra ∷ ScriptLanguageInEra SimpleScriptV2 AllegraEra

SimpleScriptV2InMary ∷ ScriptLanguageInEra SimpleScriptV2 MaryEra

scriptLanguageSupportedInEra ∷ CardanoEra era → ScriptLanguage lang → Maybe (ScriptLanguageInEra lang era)

Check if a given script language is supported in a given era, and if so return the evidence.

languageOfScriptLanguageInEra ∷ ScriptLanguageInEra lang era → ScriptLanguage lang

eraOfScriptLanguageInEra ∷ ScriptLanguageInEra lang era → ShelleyBasedEra era

The simple script language

data SimpleScript lang where

Constructors

RequireSignature ∷ !(Hash PaymentKey) → SimpleScript lang

RequireTimeBefore ∷ !(TimeLocksSupported lang) → !SlotNo → SimpleScript lang

RequireTimeAfter ∷ !(TimeLocksSupported lang) → !SlotNo → SimpleScript lang

RequireAllOf ∷ [SimpleScript lang] → SimpleScript lang

RequireAnyOf ∷ [SimpleScript lang] → SimpleScript lang

RequireMOf ∷ Int → [SimpleScript lang] → SimpleScript lang

data TimeLocksSupported lang where

Time lock feature in the SimpleScript language.

The constructors of this type serve as evidence that the timelocks feature is supported in particular versions of the language.

Constructors

TimeLocksInSimpleScriptV2 ∷ TimeLocksSupported SimpleScriptV2

timeLocksSupported ∷ SimpleScriptVersion lang → Maybe (TimeLocksSupported lang)

adjustSimpleScriptVersion ∷ SimpleScriptVersion lang' → SimpleScript lang → Maybe (SimpleScript lang')

Try converting the SimpleScript into a different version of the language.

This will work when the script only uses the features of the target language version. For example converting from SimpleScriptV2 to SimpleScriptV1 will work if the script happens not to use time locks feature. On the other hand converting SimpleScriptV1 to SimpleScriptV2 will always work because it is backwards compatible.

Script hashes

newtype ScriptHash

We have this type separate from the Hash type to avoid the script hash type being parametrised by the era. The representation is era independent, and there are many places where we want to use a script hash where we don't want things to be era-parametrised.

Constructors

ScriptHash (ScriptHash StandardCrypto)

hashScript ∷ Script lang → ScriptHash

Internal conversion functions

toShelleyScript ∷ ScriptInEra era → Script (ShelleyLedgerEra era)

toShelleyMultiSig ∷ SimpleScript SimpleScriptV1 → MultiSig StandardCrypto

Conversion for the MultiSig language used by the Shelley era.

fromShelleyMultiSig ∷ MultiSig StandardCrypto → SimpleScript lang

Conversion for the MultiSig language used by the Shelley era.

toAllegraTimelock ∷ ∀ lang. SimpleScript lang → Timelock StandardCrypto

Conversion for the Timelock language that is shared between the Allegra and Mary eras.

fromAllegraTimelock ∷ TimeLocksSupported lang → Timelock StandardCrypto → SimpleScript lang

Conversion for the Timelock language that is shared between the Allegra and Mary eras.

toShelleyScriptHash ∷ ScriptHash → ScriptHash StandardCrypto

fromShelleyScriptHash ∷ ScriptHash StandardCrypto → ScriptHash

Data family instances

data family AsType t

A family of singleton types used in this API to indicate which type to use where it would otherwise be ambiguous or merely unclear.

Values of this type are passed to deserialisation functions for example.