Cardano.Api.Serialise.TextEnvelope

TextEnvelope Serialisation

class SerialiseAsCBOR a => HasTextEnvelope a where

Minimal complete definition

textEnvelopeType

Methods

textEnvelopeType :: AsType a -> TextEnvelopeType

textEnvelopeDefaultDescr :: a -> TextEnvelopeDescr

textEnvelopeTypeInEra :: HasTextEnvelope (f era) => CardanoEra era -> AsType (f era) -> TextEnvelopeType

data TextEnvelope

A TextEnvelope is a structured envelope for serialised binary values with an external format with a semi-readable textual format.

It contains a "type" field, e.g. "PublicKeyByron" or "TxSignedShelley" to indicate the type of the encoded data. This is used as a sanity check and to help readers.

It also contains a "title" field which is free-form, and could be used to indicate the role or purpose to a reader.

Constructors

TextEnvelope

Fields

teType :: !TextEnvelopeType

teDescription :: !TextEnvelopeDescr

teRawCBOR :: !ByteString

newtype TextEnvelopeType

Constructors

TextEnvelopeType String

newtype TextEnvelopeDescr

Constructors

TextEnvelopeDescr String

textEnvelopeRawCBOR :: TextEnvelope -> ByteString

data TextEnvelopeError

The errors that the pure TextEnvelope parsing/decoding functions can return.

Constructors

TextEnvelopeTypeError ![TextEnvelopeType] !TextEnvelopeType
expected, actual

TextEnvelopeDecodeError !DecoderError

TextEnvelopeAesonDecodeError !String

serialiseToTextEnvelope :: HasTextEnvelope a => Maybe TextEnvelopeDescr -> a -> TextEnvelope

deserialiseFromTextEnvelope :: HasTextEnvelope a => TextEnvelope -> Either TextEnvelopeError a

readFileTextEnvelope :: HasTextEnvelope a => File content 'In -> IO (Either (FileError TextEnvelopeError) a)

writeFileTextEnvelope :: HasTextEnvelope a => File content 'Out -> Maybe TextEnvelopeDescr -> a -> IO (Either (FileError ()) ())

readTextEnvelopeFromFile :: FilePath -> IO (Either (FileError TextEnvelopeError) TextEnvelope)

readTextEnvelopeOfTypeFromFile :: TextEnvelopeType -> FilePath -> IO (Either (FileError TextEnvelopeError) TextEnvelope)

textEnvelopeToJSON :: HasTextEnvelope a => Maybe TextEnvelopeDescr -> a -> ByteString

serialiseTextEnvelope :: TextEnvelope -> ByteString

Serialise text envelope to pretty JSON

legacyComparison :: TextEnvelopeType -> TextEnvelopeType -> Bool

This is a backwards-compatibility patch to ensure that old envelopes generated by serialiseTxLedgerCddl can be deserialised after switching to the serialiseToTextEnvelope.

Reading one of several key types

data FromSomeType (c :: Type -> Constraint) b where

Constructors

FromSomeType :: forall (c :: Type -> Constraint) a b. c a => AsType a -> (a -> b) -> FromSomeType c b

deserialiseFromTextEnvelopeAnyOf :: [FromSomeType HasTextEnvelope b] -> TextEnvelope -> Either TextEnvelopeError b

readFileTextEnvelopeAnyOf :: [FromSomeType HasTextEnvelope b] -> File content 'In -> IO (Either (FileError TextEnvelopeError) b)

Data family instances

data family AsType t

A family of singleton types used in this API to indicate which type to use where it would otherwise be ambiguous or merely unclear.

Values of this type are passed to deserialisation functions for example.

CDDL Serialisation

data TextEnvelopeCddlError

Constructors

TextEnvelopeCddlErrCBORDecodingError DecoderError

TextEnvelopeCddlAesonDecodeError FilePath String

TextEnvelopeCddlUnknownKeyWitness

TextEnvelopeCddlTypeError

Fields

[Text]

Expected types

Text

Actual types

TextEnvelopeCddlErrUnknownType Text

TextEnvelopeCddlErrByronKeyWitnessUnsupported

data FromSomeTypeCDDL c b where

This GADT allows us to deserialise a tx or key witness without having to provide the era.

Constructors

FromCDDLTx

Fields

:: forall b. Text

CDDL type that we want

-> (InAnyShelleyBasedEra Tx -> b)

-> FromSomeTypeCDDL TextEnvelope b

FromCDDLWitness

Fields

:: forall b. Text

CDDL type that we want

-> (InAnyShelleyBasedEra KeyWitness -> b)

-> FromSomeTypeCDDL TextEnvelope b

cddlTypeToEra :: Text -> Either TextEnvelopeCddlError AnyShelleyBasedEra

Reading one of several transaction or key witness types

readFileTextEnvelopeCddlAnyOf :: [FromSomeTypeCDDL TextEnvelope b] -> FilePath -> IO (Either (FileError TextEnvelopeCddlError) b)

deserialiseFromTextEnvelopeCddlAnyOf :: [FromSomeTypeCDDL TextEnvelope b] -> TextEnvelope -> Either TextEnvelopeCddlError b

writeTxFileTextEnvelopeCddl :: ShelleyBasedEra era -> File content 'Out -> Tx era -> IO (Either (FileError ()) ())

writeTxFileTextEnvelopeCanonicalCddl :: ShelleyBasedEra era -> File content 'Out -> Tx era -> IO (Either (FileError ()) ())

Write transaction in the text envelope format. The CBOR will be in canonical format according to RFC 7049. It is also a requirement of CIP-21, which is not fully implemented.

RFC 7049: https://datatracker.ietf.org/doc/html/rfc7049#section-3.9

CIP-21: https://github.com/cardano-foundation/CIPs/blob/master/CIP-0021/README.md#canonical-cbor-serialization-format

writeTxWitnessFileTextEnvelopeCddl :: ShelleyBasedEra era -> File () 'Out -> KeyWitness era -> IO (Either (FileError ()) ())

deserialiseByronTxCddl :: TextEnvelope -> Either TextEnvelopeCddlError (ATxAux ByteString)

serialiseWitnessLedgerCddl :: ShelleyBasedEra era -> KeyWitness era -> TextEnvelope

deserialiseWitnessLedgerCddl :: ShelleyBasedEra era -> TextEnvelope -> Either TextEnvelopeCddlError (KeyWitness era)

Byron tx serialization

serializeByronTx :: ATxAux ByteString -> TextEnvelope

writeByronTxFileTextEnvelopeCddl :: File content 'Out -> ATxAux ByteString -> IO (Either (FileError ()) ())