Cardano.Api.Shelley

Description

This module provides a library interface that is intended to be the complete API for Shelley covering everything, including exposing constructors for the lower level types.

Documentation

module Cardano.Api

Genesis

data ShelleyGenesis era

Shelley genesis information

Note that this is needed only for a pure Shelley network, hence it being defined here rather than in its own module. In mainnet, Shelley will transition naturally from Byron, and thus will never have its own genesis information.

Constructors

ShelleyGenesis

Fields

sgSystemStart :: !UTCTime

sgNetworkMagic :: !Word32

sgNetworkId :: !Network

sgActiveSlotsCoeff :: !PositiveUnitInterval

sgSecurityParam :: !Word64

sgEpochLength :: !EpochSize

sgSlotsPerKESPeriod :: !Word64

sgMaxKESEvolutions :: !Word64

sgSlotLength :: !NominalDiffTime

sgUpdateQuorum :: !Word64

sgMaxLovelaceSupply :: !Word64

sgProtocolParams :: !(PParams era)

sgGenDelegs :: !(Map (KeyHash 'Genesis (Crypto era)) (GenDelegPair (Crypto era)))

sgInitialFunds :: !(Map (Addr (Crypto era)) Coin)

sgStaking :: !(ShelleyGenesisStaking (Crypto era))

shelleyGenesisDefaults :: ShelleyGenesis crypto

Some reasonable starting defaults for constructing a ShelleyGenesis.

You must override at least the following fields for this to be useful:

sgSystemStart the time of the first block

sgNetworkMagic to a suitable testnet or mainnet network magic number.

sgGenDelegs to have some initial nodes

sgInitialFunds to have any money in the system

sgMaxLovelaceSupply must be at least the sum of the sgInitialFunds but more if you want to allow for rewards.

Cryptographic key interface

class (Eq (VerificationKey keyrole), Show (VerificationKey keyrole), SerialiseAsRawBytes (Hash keyrole), HasTextEnvelope (VerificationKey keyrole), HasTextEnvelope (SigningKey keyrole)) => Key keyrole where

An interface for cryptographic keys used for signatures with a SigningKey and a VerificationKey key.

This interface does not provide actual signing or verifying functions since this API is concerned with the management of keys: generating and serialising.

Associated Types

data VerificationKey keyrole :: Type

The type of cryptographic verification key, for each key role.

data SigningKey keyrole :: Type

The type of cryptographic signing key, for each key role.

Methods

getVerificationKey :: SigningKey keyrole -> VerificationKey keyrole

Get the corresponding verification key from a signing key.

deterministicSigningKey :: AsType keyrole -> Seed -> SigningKey keyrole

Generate a SigningKey deterministically, given a Seed. The required size of the seed is given by deterministicSigningKeySeedSize.

deterministicSigningKeySeedSize :: AsType keyrole -> Word

verificationKeyHash :: VerificationKey keyrole -> Hash keyrole

Hashes

data family Hash keyrole :: Type

Payment addresses

Constructing and inspecting Shelley payment addresses

data Address addrtype where

Addresses are used as locations where assets live. The address determines the rights needed to spend assets at the address: in particular holding some signing key or being able to satisfy the conditions of a script.

There are currently two types of address:

Byron addresses, which use the type tag ByronAddr; and

Shelley addresses, which use the type tag ShelleyAddr. Notably, Shelley addresses support scripts and stake delegation.

The address type is subtly from the ledger era in which each address type is valid: while Byron addresses are the only choice in the Byron era, the Shelley era and all subsequent eras support both Byron and Shelley addresses. The Address type param only says the type of the address (either Byron or Shelley). The AddressInEra type connects the address type with the era in which it is supported.

Constructors

ShelleyAddress :: Network -> PaymentCredential StandardCrypto -> StakeReference StandardCrypto -> Address ShelleyAddr
Shelley addresses allow delegation. Shelley addresses were introduced in Shelley era and are thus supported from the Shelley era onwards

toShelleyAddr :: AddressInEra era -> Addr StandardCrypto

fromShelleyAddr :: ShelleyBasedEra era -> Addr StandardCrypto -> AddressInEra era

fromShelleyAddrIsSbe :: IsShelleyBasedEra era => Addr StandardCrypto -> AddressInEra era

fromShelleyAddrToAny :: Addr StandardCrypto -> AddressAny

toShelleyStakeCredential :: StakeCredential -> StakeCredential StandardCrypto

fromShelleyStakeCredential :: StakeCredential StandardCrypto -> StakeCredential

data NetworkId

Constructors

Mainnet

Testnet !NetworkMagic

Stake addresses

data PaymentCredential

Constructors

PaymentCredentialByKey (Hash PaymentKey)

PaymentCredentialByScript ScriptHash

data StakeAddress where

Constructors

StakeAddress :: Network -> StakeCredential StandardCrypto -> StakeAddress

data StakeAddressReference

Constructors

StakeAddressByValue StakeCredential

StakeAddressByPointer StakeAddressPointer

NoStakeAddress

data StakeCredential

Constructors

StakeCredentialByKey (Hash StakeKey)

StakeCredentialByScript ScriptHash

toShelleyStakeAddr :: StakeAddress -> RewardAcnt StandardCrypto

fromShelleyStakeAddr :: RewardAcnt StandardCrypto -> StakeAddress

fromShelleyStakeReference :: StakeReference StandardCrypto -> StakeAddressReference

fromShelleyPaymentCredential :: PaymentCredential StandardCrypto -> PaymentCredential

Building transactions

Constructing and inspecting transactions

data TxBody era where

Constructors

ShelleyTxBody

Fields

:: ShelleyBasedEra era

-> TxBody (ShelleyLedgerEra era)

-> [Script (ShelleyLedgerEra era)]

-> TxBodyScriptData era

-> Maybe (AuxiliaryData (ShelleyLedgerEra era))

-> TxScriptValidity era

Mark script as expected to pass or fail validation

-> TxBody era

newtype TxId

Constructors

TxId (Hash StandardCrypto EraIndependentTxBody)

toShelleyTxId :: TxId -> TxId StandardCrypto

fromShelleyTxId :: TxId StandardCrypto -> TxId

getTxIdShelley :: Crypto (ShelleyLedgerEra era) ~ StandardCrypto => UsesTxBody (ShelleyLedgerEra era) => ShelleyBasedEra era -> TxBody (ShelleyLedgerEra era) -> TxId

data TxIn

Constructors

TxIn TxId TxIx

toShelleyTxIn :: TxIn -> TxIn StandardCrypto

This function may overflow on the transaction index. Call sites must ensure that all uses of this function are appropriately guarded.

fromShelleyTxIn :: TxIn StandardCrypto -> TxIn

data TxOut ctx era

Constructors

TxOut (AddressInEra era) (TxOutValue era) (TxOutDatum ctx era) (ReferenceScript era)

toShelleyTxOut :: forall era ledgerera. ShelleyLedgerEra era ~ ledgerera => ShelleyBasedEra era -> TxOut CtxUTxO era -> TxOut ledgerera

fromShelleyTxOut :: ShelleyLedgerEra era ~ ledgerera => ShelleyBasedEra era -> TxOut ledgerera -> TxOut ctx era

newtype TxIx

Constructors

TxIx Word

newtype Lovelace

Constructors

Lovelace Integer

toShelleyLovelace :: Lovelace -> Coin

fromShelleyLovelace :: Coin -> Lovelace

toMaryValue :: Value -> Value StandardCrypto

fromMaryValue :: Value StandardCrypto -> Value

calcMinimumDeposit :: Value -> Lovelace -> Lovelace

Calculate cost of making a UTxO entry for a given Value and mininimum UTxO value derived from the ProtocolParameters

Signing transactions

Creating transaction witnesses one by one, or all in one go.

data Tx era where

Constructors

ShelleyTx :: ShelleyBasedEra era -> Tx (ShelleyLedgerEra era) -> Tx era

Incremental signing and separate witnesses

data KeyWitness era where

Constructors

ShelleyBootstrapWitness :: ShelleyBasedEra era -> BootstrapWitness StandardCrypto -> KeyWitness era

ShelleyKeyWitness :: ShelleyBasedEra era -> WitVKey Witness StandardCrypto -> KeyWitness era

data ShelleyWitnessSigningKey

Constructors

WitnessPaymentKey (SigningKey PaymentKey)

WitnessPaymentExtendedKey (SigningKey PaymentExtendedKey)

WitnessStakeKey (SigningKey StakeKey)

WitnessStakeExtendedKey (SigningKey StakeExtendedKey)

WitnessStakePoolKey (SigningKey StakePoolKey)

WitnessGenesisKey (SigningKey GenesisKey)

WitnessGenesisExtendedKey (SigningKey GenesisExtendedKey)

WitnessGenesisDelegateKey (SigningKey GenesisDelegateKey)

WitnessGenesisDelegateExtendedKey (SigningKey GenesisDelegateExtendedKey)

data ShelleySigningKey

We support making key witnesses with both normal and extended signing keys.

Constructors

ShelleyNormalSigningKey (SignKeyDSIGN StandardCrypto)
A normal ed25519 signing key

ShelleyExtendedSigningKey XPrv
An extended ed25519 signing key

getShelleyKeyWitnessVerificationKey :: ShelleySigningKey -> VKey Witness StandardCrypto

getTxBodyAndWitnesses :: Tx era -> (TxBody era, [KeyWitness era])

makeShelleySignature :: SignableRepresentation tosign => tosign -> ShelleySigningKey -> SignedDSIGN StandardCrypto tosign

toShelleySigningKey :: ShelleyWitnessSigningKey -> ShelleySigningKey

Blocks

fromConsensusBlock :: ConsensusBlockForMode mode ~ block => LedgerSupportsProtocol (ShelleyBlock (TPraos StandardCrypto) (ShelleyEra StandardCrypto)) => ConsensusMode mode -> block -> BlockInMode mode

toConsensusBlock :: ConsensusBlockForMode mode ~ block => LedgerSupportsProtocol (ShelleyBlock (TPraos StandardCrypto) (ShelleyEra StandardCrypto)) => BlockInMode mode -> block

fromConsensusTip :: ConsensusBlockForMode mode ~ block => ConsensusMode mode -> Tip block -> ChainTip

fromConsensusPointInMode :: ConsensusMode mode -> Point (ConsensusBlockForMode mode) -> ChainPoint

toConsensusPointInMode :: ConsensusMode mode -> ChainPoint -> Point (ConsensusBlockForMode mode)

toConsensusPointHF :: HeaderHash block ~ OneEraHash xs => ChainPoint -> Point block

Convert a Point for multi-era block type

Transaction metadata

Embedding additional structured data within transactions.

toShelleyMetadata :: Map Word64 TxMetadataValue -> Map Word64 Metadatum

fromShelleyMetadata :: Map Word64 Metadatum -> Map Word64 TxMetadataValue

toShelleyMetadatum :: TxMetadataValue -> Metadatum

fromShelleyMetadatum :: Metadatum -> TxMetadataValue

Protocol parameters

data ProtocolParameters

The values of the set of updatable protocol parameters. At any particular point on the chain there is a current set of parameters in use.

These parameters can be updated (at epoch boundaries) via an UpdateProposal, which contains a ProtocolParametersUpdate.

The ProtocolParametersUpdate is essentially a diff for the ProtocolParameters.

There are also parameters fixed in the Genesis file. See GenesisParameters.

Constructors

ProtocolParameters

Fields

protocolParamProtocolVersion :: (Natural, Natural)

Protocol version, major and minor. Updating the major version is used to trigger hard forks. (Major , Minor )

protocolParamDecentralization :: Maybe Rational

The decentralization parameter. This is fraction of slots that belong to the BFT overlay schedule, rather than the Praos schedule. So 1 means fully centralised, while 0 means fully decentralised.

This is the "d" parameter from the design document.

Deprecated in Babbage

protocolParamExtraPraosEntropy :: Maybe PraosNonce

Extra entropy for the Praos per-epoch nonce.

This can be used to add extra entropy during the decentralisation process. If the extra entropy can be demonstrated to be generated randomly then this method can be used to show that the initial federated operators did not subtly bias the initial schedule so that they retain undue influence after decentralisation.

protocolParamMaxBlockHeaderSize :: Natural

The maximum permitted size of a block header.

This must be at least as big as the largest legitimate block headers but should not be too much larger, to help prevent DoS attacks.

Caution: setting this to be smaller than legitimate block headers is a sure way to brick the system!

protocolParamMaxBlockBodySize :: Natural

The maximum permitted size of the block body (that is, the block payload, without the block header).

This should be picked with the Praos network delta security parameter in mind. Making this too large can severely weaken the Praos consensus properties.

Caution: setting this to be smaller than a transaction that can change the protocol parameters is a sure way to brick the system!

protocolParamMaxTxSize :: Natural

The maximum permitted size of a transaction.

Typically this should not be too high a fraction of the block size, otherwise wastage from block fragmentation becomes a problem, and the current implementation does not use any sophisticated box packing algorithm.

protocolParamTxFeeFixed :: Natural

The constant factor for the minimum fee calculation.

protocolParamTxFeePerByte :: Natural

The linear factor for the minimum fee calculation.

protocolParamMinUTxOValue :: Maybe Lovelace

The minimum permitted value for new UTxO entries, ie for transaction outputs.

protocolParamStakeAddressDeposit :: Lovelace

The deposit required to register a stake address.

protocolParamStakePoolDeposit :: Lovelace

The deposit required to register a stake pool.

protocolParamMinPoolCost :: Lovelace

The minimum value that stake pools are permitted to declare for their cost parameter.

protocolParamPoolRetireMaxEpoch :: EpochNo

The maximum number of epochs into the future that stake pools are permitted to schedule a retirement.

protocolParamStakePoolTargetNum :: Natural

The equilibrium target number of stake pools.

This is the "k" incentives parameter from the design document.

protocolParamPoolPledgeInfluence :: Rational

The influence of the pledge in stake pool rewards.

This is the "a_0" incentives parameter from the design document.

protocolParamMonetaryExpansion :: Rational

The monetary expansion rate. This determines the fraction of the reserves that are added to the fee pot each epoch.

This is the "rho" incentives parameter from the design document.

protocolParamTreasuryCut :: Rational

The fraction of the fee pot each epoch that goes to the treasury.

This is the "tau" incentives parameter from the design document.

protocolParamUTxOCostPerWord :: Maybe Lovelace

Cost in ada per word of UTxO storage.

Introduced in Alonzo

protocolParamCostModels :: Map AnyPlutusScriptVersion CostModel

Cost models for script languages that use them.

Introduced in Alonzo

protocolParamPrices :: Maybe ExecutionUnitPrices

Price of execution units for script languages that use them.

Introduced in Alonzo

protocolParamMaxTxExUnits :: Maybe ExecutionUnits

Max total script execution resources units allowed per tx

Introduced in Alonzo

protocolParamMaxBlockExUnits :: Maybe ExecutionUnits

Max total script execution resources units allowed per block

Introduced in Alonzo

protocolParamMaxValueSize :: Maybe Natural

Max size of a Value in a tx output.

Introduced in Alonzo

protocolParamCollateralPercent :: Maybe Natural

The percentage of the script contribution to the txfee that must be provided as collateral inputs when including Plutus scripts.

Introduced in Alonzo

protocolParamMaxCollateralInputs :: Maybe Natural

The maximum number of collateral inputs allowed in a transaction.

Introduced in Alonzo

protocolParamUTxOCostPerByte :: Maybe Lovelace

Cost in ada per byte of UTxO storage.

Introduced in Babbage

checkProtocolParameters :: forall era. IsCardanoEra era => ShelleyBasedEra era -> ProtocolParameters -> Either ProtocolParametersError ()

data ProtocolParametersError

Constructors

PParamsErrorMissingMinUTxoValue AnyCardanoEra

PParamsErrorMissingAlonzoProtocolParameter

Scripts

toShelleyScript :: ScriptInEra era -> Script (ShelleyLedgerEra era)

toShelleyMultiSig :: SimpleScript SimpleScriptV1 -> MultiSig StandardCrypto

Conversion for the MultiSig language used by the Shelley era.

fromShelleyMultiSig :: MultiSig StandardCrypto -> SimpleScript lang

Conversion for the MultiSig language used by the Shelley era.

toAllegraTimelock :: forall lang. SimpleScript lang -> Timelock StandardCrypto

Conversion for the Timelock language that is shared between the Allegra and Mary eras.

fromAllegraTimelock :: TimeLocksSupported lang -> Timelock StandardCrypto -> SimpleScript lang

Conversion for the Timelock language that is shared between the Allegra and Mary eras.

toShelleyScriptHash :: ScriptHash -> ScriptHash StandardCrypto

fromShelleyScriptHash :: ScriptHash StandardCrypto -> ScriptHash

data PlutusScript lang where

Plutus scripts.

Note that Plutus scripts have a binary serialisation but no JSON serialisation.

Constructors

PlutusScriptSerialised :: ShortByteString -> PlutusScript lang

data PlutusScriptOrReferenceInput lang

Scripts can now exist in the UTxO at a transaction output. We can reference these scripts via specification of a reference transaction input in order to witness spending inputs, withdrawals, certificates or to mint tokens. This datatype encapsulates this concept.

Constructors

PScript (PlutusScript lang)

PReferenceScript TxIn (Maybe ScriptHash)

data SimpleScriptOrReferenceInput lang

Constructors

SScript (SimpleScript lang)

SReferenceScript TxIn (Maybe ScriptHash)

toPlutusData :: ScriptData -> Data

fromPlutusData :: Data -> ScriptData

toAlonzoData :: ScriptData -> Data ledgerera

fromAlonzoData :: Data ledgerera -> ScriptData

toAlonzoPrices :: ExecutionUnitPrices -> Maybe Prices

fromAlonzoPrices :: Prices -> ExecutionUnitPrices

toAlonzoExUnits :: ExecutionUnits -> ExUnits

fromAlonzoExUnits :: ExUnits -> ExecutionUnits

toAlonzoRdmrPtr :: ScriptWitnessIndex -> RdmrPtr

fromAlonzoRdmrPtr :: RdmrPtr -> ScriptWitnessIndex

scriptDataFromJsonDetailedSchema :: Value -> Either ScriptDataJsonSchemaError ScriptData

scriptDataToJsonDetailedSchema :: ScriptData -> Value

calculateExecutionUnitsLovelace :: ExecutionUnitPrices -> ExecutionUnits -> Maybe Lovelace

Reference Scripts

data ReferenceScript era where

A reference scripts is a script that can exist at a transaction output. This greatly reduces the size of transactions that use scripts as the script no longer has to be added to the transaction, they can now be referenced via a transaction output.

Constructors

ReferenceScript :: ReferenceTxInsScriptsInlineDatumsSupportedInEra era -> ScriptInAnyLang -> ReferenceScript era

ReferenceScriptNone :: ReferenceScript era

data ReferenceTxInsScriptsInlineDatumsSupportedInEra era where

Constructors

ReferenceTxInsScriptsInlineDatumsInBabbageEra :: ReferenceTxInsScriptsInlineDatumsSupportedInEra BabbageEra

refInsScriptsAndInlineDatsSupportedInEra :: CardanoEra era -> Maybe (ReferenceTxInsScriptsInlineDatumsSupportedInEra era)

refScriptToShelleyScript :: CardanoEra era -> ReferenceScript era -> StrictMaybe (Script (ShelleyLedgerEra era))

Certificates

data Certificate

Constructors

StakeAddressRegistrationCertificate StakeCredential

StakeAddressDeregistrationCertificate StakeCredential

StakeAddressDelegationCertificate StakeCredential PoolId

StakePoolRegistrationCertificate StakePoolParameters

StakePoolRetirementCertificate PoolId EpochNo

GenesisKeyDelegationCertificate (Hash GenesisKey) (Hash GenesisDelegateKey) (Hash VrfKey)

MIRCertificate MIRPot MIRTarget

toShelleyCertificate :: Certificate -> DCert StandardCrypto

fromShelleyCertificate :: DCert StandardCrypto -> Certificate

Operational certificates

data OperationalCertificate

Constructors

OperationalCertificate !(OCert StandardCrypto) !(VerificationKey StakePoolKey)

data OperationalCertificateIssueCounter

Constructors

OperationalCertificateIssueCounter

Fields

opCertIssueCount :: !Word64

opCertIssueColdKey :: !(VerificationKey StakePoolKey)

data OperationalCertIssueError

Constructors

OperationalCertKeyMismatch (VerificationKey StakePoolKey) (VerificationKey StakePoolKey)
The stake pool verification key expected for the OperationalCertificateIssueCounter does not match the signing key supplied for signing.

Order: pool vkey expected, pool skey supplied

Stake Pool

data StakePoolMetadata

A representation of the required fields for off-chain stake pool metadata.

Constructors

StakePoolMetadata !Text !Text !Text !Text

stakePoolName :: StakePoolMetadata -> Text

A name of up to 50 characters.

stakePoolDescription :: StakePoolMetadata -> Text

A description of up to 255 characters.

stakePoolTicker :: StakePoolMetadata -> Text

A ticker of 3-5 characters, for a compact display of stake pools in a wallet.

stakePoolHomepage :: StakePoolMetadata -> Text

A URL to a homepage with additional information about the pool. n.b. the spec does not specify a character limit for this field.

data StakePoolMetadataReference

Constructors

StakePoolMetadataReference Text (Hash StakePoolMetadata)

stakePoolMetadataURL :: StakePoolMetadataReference -> Text

stakePoolMetadataHash :: StakePoolMetadataReference -> Hash StakePoolMetadata

data StakePoolParameters

Constructors

StakePoolParameters PoolId (Hash VrfKey) Lovelace Rational StakeAddress Lovelace [Hash StakeKey] [StakePoolRelay] (Maybe StakePoolMetadataReference)

stakePoolId :: StakePoolParameters -> PoolId

stakePoolVRF :: StakePoolParameters -> Hash VrfKey

stakePoolCost :: StakePoolParameters -> Lovelace

stakePoolMargin :: StakePoolParameters -> Rational

stakePoolRewardAccount :: StakePoolParameters -> StakeAddress

stakePoolPledge :: StakePoolParameters -> Lovelace

stakePoolOwners :: StakePoolParameters -> [Hash StakeKey]

stakePoolRelays :: StakePoolParameters -> [StakePoolRelay]

stakePoolMetadata :: StakePoolParameters -> Maybe StakePoolMetadataReference

data StakePoolRelay

Constructors

StakePoolRelayIp (Maybe IPv4) (Maybe IPv6) (Maybe PortNumber)
One or both of IPv4 & IPv6

StakePoolRelayDnsARecord ByteString (Maybe PortNumber)
An DNS name pointing to a A or AAAA record.

StakePoolRelayDnsSrvRecord ByteString
A DNS name pointing to a SRV record.

newtype EpochNo

An epoch, i.e. the number of the epoch.

Constructors

EpochNo

Fields

unEpochNo :: Word64

Stake pool operator's keys

data StakePoolKey

type PoolId = Hash StakePoolKey

KES keys

data KesKey

newtype KESPeriod

Constructors

KESPeriod

Fields

unKESPeriod :: Word

VRF keys

data VrfKey

Low level protocol interaction with a Cardano node

data LocalNodeConnectInfo mode

Constructors

LocalNodeConnectInfo (ConsensusModeParams mode) NetworkId FilePath

data ShelleyMode

The Shelley-only consensus mode consists of only the Shelley era.

This was used for the early Shelley testnets prior to the use of the multi-era CardanoMode. It is useful for setting up Shelley test networks (e.g. for benchmarking) without having to go through the complication of the hard fork from Byron to Shelley eras. It also shows how a single-era consensus mode works. It may be replaced by other single-era modes in future.

data ConsensusMode mode where

This GADT provides a value-level representation of all the consensus modes. This enables pattern matching on the era to allow them to be treated in a non-uniform way.

Constructors

ByronMode :: ConsensusMode ByronMode

ShelleyMode :: ConsensusMode ShelleyMode

data LocalNodeClientProtocols block point tip slot tx txid txerr query m

The protocols we can use with a local node. Use in conjunction with connectToLocalNode.

These protocols use the types from the rest of this API. The conversion to/from the types used by the underlying wire formats is handled by connectToLocalNode.

Constructors

LocalNodeClientProtocols (LocalChainSyncClient block point tip m) (Maybe (LocalTxSubmissionClient tx txerr m ())) (Maybe (LocalStateQueryClient block point query m ())) (Maybe (LocalTxMonitorClient txid tx slot m ()))

Shelley based eras

type family ShelleyLedgerEra era where ...

A type family that connects our era type tags to equivalent type tags used in the Shelley ledger library.

This type mapping connect types from this API with types in the Shelley ledger library which allows writing conversion functions in a more generic way.

Equations

ShelleyLedgerEra ShelleyEra = StandardShelley

ShelleyLedgerEra AllegraEra = StandardAllegra

ShelleyLedgerEra MaryEra = StandardMary

ShelleyLedgerEra AlonzoEra = StandardAlonzo

ShelleyLedgerEra BabbageEra = StandardBabbage

Local State Query

data DebugLedgerState era where

Constructors

DebugLedgerState :: ShelleyLedgerEra era ~ ledgerera => NewEpochState ledgerera -> DebugLedgerState era

decodeDebugLedgerState :: forall era. FromCBOR (DebugLedgerState era) => SerialisedDebugLedgerState era -> Either ByteString (DebugLedgerState era)

newtype ProtocolState era

Constructors

ProtocolState (Serialised (ChainDepState (ConsensusProtocol era)))

decodeProtocolState :: FromCBOR (ChainDepState (ConsensusProtocol era)) => ProtocolState era -> Either (ByteString, DecoderError) (ChainDepState (ConsensusProtocol era))

newtype SerialisedDebugLedgerState era

Constructors

SerialisedDebugLedgerState (Serialised (NewEpochState (ShelleyLedgerEra era)))

newtype SerialisedCurrentEpochState era

Constructors

SerialisedCurrentEpochState (Serialised (EpochState (ShelleyLedgerEra era)))

decodeCurrentEpochState :: forall era. Era (ShelleyLedgerEra era) => Share (TxOut (ShelleyLedgerEra era)) ~ Interns (Credential 'Staking (Crypto (ShelleyLedgerEra era))) => FromSharedCBOR (TxOut (ShelleyLedgerEra era)) => Share (TxOut (ShelleyLedgerEra era)) ~ Interns (Credential 'Staking (Crypto (ShelleyLedgerEra era))) => FromCBOR (PParams (ShelleyLedgerEra era)) => FromCBOR (Value (ShelleyLedgerEra era)) => FromCBOR (State (EraRule "PPUP" (ShelleyLedgerEra era))) => SerialisedCurrentEpochState era -> Either DecoderError (CurrentEpochState era)

newtype UTxO era

Constructors

UTxO

Fields

unUTxO :: Map TxIn (TxOut CtxUTxO era)

data AcquiringFailure

Establish a connection to a node and execute a single query using the local state query protocol.

Constructors

AFPointTooOld

AFPointNotOnChain

newtype SystemStart

System start

Slots are counted from the system start.

Constructors

SystemStart

Fields

getSystemStart :: UTCTime

Various calculations

data LeadershipError

Constructors

LeaderErrDecodeLedgerStateFailure

LeaderErrDecodeProtocolStateFailure (ByteString, DecoderError)

LeaderErrDecodeProtocolEpochStateFailure DecoderError

LeaderErrGenesisSlot

LeaderErrStakePoolHasNoStake PoolId

LeaderErrStakeDistribUnstable

Fields

SlotNo

Current slot

SlotNo

Stable after

SlotNo

Stability window size

SlotNo

Predicted last slot of the epoch

LeaderErrSlotRangeCalculationFailure Text

LeaderErrCandidateNonceStillEvolving

currentEpochEligibleLeadershipSlots

Arguments

:: forall era ledgerera. ShelleyLedgerEra era ~ ledgerera

=> Era ledgerera

=> PraosProtocolSupportsNode (ConsensusProtocol era)

=> HasField "_d" (PParams ledgerera) UnitInterval

=> Share (TxOut (ShelleyLedgerEra era)) ~ Interns (Credential 'Staking (Crypto (ShelleyLedgerEra era)))

=> FromCBOR (ChainDepState (ConsensusProtocol era))

=> ShelleyBasedEra era

-> ShelleyGenesis StandardShelley

-> EpochInfo (Either Text)

-> ProtocolParameters

-> ProtocolState era

-> PoolId

-> SigningKey VrfKey

-> SerialisedCurrentEpochState era

-> EpochNo
Current EpochInfo

-> Either LeadershipError (Set SlotNo)

Return the slots at which a particular stake pool operator is expected to mint a block.

nextEpochEligibleLeadershipSlots

Arguments

:: forall era. HasField "_d" (PParams (ShelleyLedgerEra era)) UnitInterval

=> Era (ShelleyLedgerEra era)

=> Share (TxOut (ShelleyLedgerEra era)) ~ Interns (Credential 'Staking (Crypto (ShelleyLedgerEra era)))

=> FromCBOR (ChainDepState (ConsensusProtocol era))

=> PraosProtocolSupportsNode (ConsensusProtocol era)

=> ShelleyBasedEra era

-> ShelleyGenesis StandardShelley

-> SerialisedCurrentEpochState era
We need the mark stake distribution in order to predict the following epoch's leadership schedule

-> ProtocolState era

-> PoolId
Potential slot leading stake pool

-> SigningKey VrfKey
VRF signing key of the stake pool

-> ProtocolParameters

-> EpochInfo (Either Text)

-> (ChainTip, EpochNo)

-> Either LeadershipError (Set SlotNo)

Conversions

shelleyPayAddrToPlutusPubKHash :: Address ShelleyAddr -> Maybe PubKeyHash

Converts a Shelley payment address to a Plutus public key hash.

toConsensusGenTx :: ConsensusBlockForMode mode ~ block => TxInMode mode -> GenTx block

fromAlonzoCostModels :: CostModels -> Map AnyPlutusScriptVersion CostModel

toShelleyNetwork :: NetworkId -> Network

fromShelleyPParams :: PParams ledgerera -> ProtocolParameters