Cardano.Api.Shelley

Description

This module provides a library interface that is intended to be the complete API for Shelley covering everything, including exposing constructors for the lower level types.

Documentation

module Cardano.Api

Genesis

data ShelleyGenesis era

Constructors

ShelleyGenesis

Fields

sgSystemStart ∷ !UTCTime

sgNetworkMagic ∷ !Word32

sgNetworkId ∷ !Network

sgActiveSlotsCoeff ∷ !Rational

sgSecurityParam ∷ !Word64

sgEpochLength ∷ !EpochSize

sgSlotsPerKESPeriod ∷ !Word64

sgMaxKESEvolutions ∷ !Word64

sgSlotLength ∷ !NominalDiffTime

sgUpdateQuorum ∷ !Word64

sgMaxLovelaceSupply ∷ !Word64

sgProtocolParams ∷ !(PParams era)

sgGenDelegs ∷ !(Map (KeyHash 'Genesis (Crypto era)) (GenDelegPair (Crypto era)))

sgInitialFunds ∷ !(Map (Addr (Crypto era)) Coin)

sgStaking ∷ !(ShelleyGenesisStaking (Crypto era))

shelleyGenesisDefaults ∷ ShelleyGenesis crypto

Some reasonable starting defaults for constructing a ShelleyGenesis.

You must override at least the following fields for this to be useful:

sgSystemStart the time of the first block

sgNetworkMagic to a suitable testnet or mainnet network magic number.

sgGenDelegs to have some initial nodes

sgInitialFunds to have any money in the system

sgMaxLovelaceSupply must be at least the sum of the sgInitialFunds but more if you want to allow for rewards.

Cryptographic key interface

Payment addresses

Constructing and inspecting Shelley payment addresses

data Address addrtype where

Addresses are used as locations where assets live. The address determines the rights needed to spend assets at the address: in particular holding some signing key or being able to satisfy the conditions of a script.

There are currently two types of address:

Byron addresses, which use the type tag ByronAddr; and

Shelley addresses, which use the type tag ShelleyAddr. Notably, Shelley addresses support scripts and stake delegation.

The address type is subtly from the ledger era in which each address type is valid: while Byron addresses are the only choice in the Byron era, the Shelley era and all subsequent eras support both Byron and Shelley addresses. The Address type param only says the type of the address (either Byron or Shelley). The AddressInEra type connects the address type with the era in which it is supported.

Constructors

ShelleyAddress ∷ Network → PaymentCredential StandardCrypto → StakeReference StandardCrypto → Address ShelleyAddr
Shelley addresses allow delegation. Shelley addresses were introduced in Shelley era and are thus supported from the Shelley era onwards

toShelleyAddr ∷ AddressInEra era → Addr StandardCrypto

fromShelleyAddr ∷ IsShelleyBasedEra era ⇒ Addr StandardCrypto → AddressInEra era

toShelleyStakeCredential ∷ StakeCredential → StakeCredential StandardCrypto

fromShelleyStakeCredential ∷ StakeCredential StandardCrypto → StakeCredential

data NetworkId

Constructors

Mainnet

Testnet !NetworkMagic

Stake addresses

data PaymentCredential

Constructors

PaymentCredentialByKey (Hash PaymentKey)

PaymentCredentialByScript ScriptHash

data StakeAddress where

Constructors

StakeAddress ∷ Network → StakeCredential StandardCrypto → StakeAddress

data StakeAddressReference

Constructors

StakeAddressByValue StakeCredential

StakeAddressByPointer StakeAddressPointer

NoStakeAddress

data StakeCredential

Constructors

StakeCredentialByKey (Hash StakeKey)

StakeCredentialByScript ScriptHash

toShelleyStakeAddr ∷ StakeAddress → RewardAcnt StandardCrypto

fromShelleyStakeAddr ∷ RewardAcnt StandardCrypto → StakeAddress

fromShelleyStakeReference ∷ StakeReference StandardCrypto → StakeAddressReference

fromShelleyPaymentCredential ∷ PaymentCredential StandardCrypto → PaymentCredential

Building transactions

Constructing and inspecting transactions

data TxBody era where

Constructors

ShelleyTxBody ∷ ShelleyBasedEra era → TxBody (ShelleyLedgerEra era) → Maybe (AuxiliaryData (ShelleyLedgerEra era)) → TxBody era

newtype TxId

Constructors

TxId (Hash StandardCrypto EraIndependentTxBody)

toShelleyTxId ∷ TxId → TxId StandardCrypto

fromShelleyTxId ∷ TxId StandardCrypto → TxId

data TxIn

Constructors

TxIn TxId TxIx

toShelleyTxIn ∷ TxIn → TxIn StandardCrypto

fromShelleyTxIn ∷ TxIn StandardCrypto → TxIn

data TxOut era

Constructors

TxOut (AddressInEra era) (TxOutValue era)

toShelleyTxOut ∷ ∀ era ledgerera. (ShelleyLedgerEra era ~ ledgerera, IsShelleyBasedEra era, ShelleyBased ledgerera) ⇒ TxOut era → TxOut ledgerera

fromShelleyTxOut ∷ TxOut StandardShelley → TxOut ShelleyEra

newtype TxIx

Constructors

TxIx Word

newtype Lovelace

Constructors

Lovelace Integer

toShelleyLovelace ∷ Lovelace → Coin

fromShelleyLovelace ∷ Coin → Lovelace

toMaryValue ∷ Value → Value StandardCrypto

fromMaryValue ∷ Value StandardCrypto → Value

Signing transactions

Creating transaction witnesses one by one, or all in one go.

data Tx era where

Constructors

ShelleyTx ∷ ShelleyBasedEra era → Tx (ShelleyLedgerEra era) → Tx era

Incremental signing and separate witnesses

data Witness era where

Constructors

ShelleyBootstrapWitness ∷ ShelleyBasedEra era → BootstrapWitness StandardCrypto → Witness era

ShelleyKeyWitness ∷ ShelleyBasedEra era → WitVKey Witness StandardCrypto → Witness era

ShelleyScriptWitness ∷ ShelleyBasedEra era → Script (ShelleyLedgerEra era) → Witness era

data ShelleyWitnessSigningKey

Constructors

WitnessPaymentKey (SigningKey PaymentKey)

WitnessPaymentExtendedKey (SigningKey PaymentExtendedKey)

WitnessStakeKey (SigningKey StakeKey)

WitnessStakeExtendedKey (SigningKey StakeExtendedKey)

WitnessStakePoolKey (SigningKey StakePoolKey)

WitnessGenesisKey (SigningKey GenesisKey)

WitnessGenesisExtendedKey (SigningKey GenesisExtendedKey)

WitnessGenesisDelegateKey (SigningKey GenesisDelegateKey)

WitnessGenesisDelegateExtendedKey (SigningKey GenesisDelegateExtendedKey)

data ShelleySigningKey

We support making key witnesses with both normal and extended signing keys.

getShelleyKeyWitnessVerificationKey ∷ ShelleySigningKey → VKey Witness StandardCrypto

makeShelleySignature ∷ SignableRepresentation tosign ⇒ tosign → ShelleySigningKey → SignedDSIGN StandardCrypto tosign

toShelleySigningKey ∷ ShelleyWitnessSigningKey → ShelleySigningKey

Transaction metadata

Embedding additional structured data within transactions.

toShelleyMetadata ∷ Map Word64 TxMetadataValue → Map Word64 Metadatum

fromShelleyMetadata ∷ Map Word64 Metadatum → Map Word64 TxMetadataValue

Protocol parameters

data ProtocolParameters

The values of the set of updateable protocol paramaters. At any particular point on the chain there is a current set of paramaters in use.

These paramaters can be updated (at epoch boundaries) via an UpdateProposal, which contains a ProtocolParametersUpdate.

The ProtocolParametersUpdate is essentially a diff for the ProtocolParameters.

There are also paramaters fixed in the Genesis file. See GenesisParameters.

Constructors

ProtocolParameters

Fields

protocolParamProtocolVersion ∷ (Natural, Natural)

Protocol version, major and minor. Updating the major version is used to trigger hard forks.

protocolParamDecentralization ∷ Rational

The decentralization parameter. This is fraction of slots that belong to the BFT overlay schedule, rather than the Praos schedule. So 1 means fully centralised, while 0 means fully decentralised.

This is the "d" parameter from the design document.

protocolParamExtraPraosEntropy ∷ Maybe PraosNonce

Extra entropy for the Praos per-epoch nonce.

This can be used to add extra entropy during the decentralisation process. If the extra entropy can be demonstrated to be generated randomly then this method can be used to show that the initial federated operators did not subtly bias the initial schedule so that they retain undue influence after decentralisation.

protocolParamMaxBlockHeaderSize ∷ Natural

The maximum permitted size of a block header.

This must be at least as big as the largest legitimate block headers but should not be too much larger, to help prevent DoS attacks.

Caution: setting this to be smaller than legitimate block headers is a sure way to brick the system!

protocolParamMaxBlockBodySize ∷ Natural

The maximum permitted size of the block body (that is, the block payload, without the block header).

This should be picked with the Praos network delta security parameter in mind. Making this too large can severely weaken the Praos consensus properties.

Caution: setting this to be smaller than a transaction that can change the protocol parameters is a sure way to brick the system!

protocolParamMaxTxSize ∷ Natural

The maximum permitted size of a transaction.

Typically this should not be too high a fraction of the block size, otherwise wastage from block fragmentation becomes a problem, and the current implementation does not use any sophisticated box packing algorithm.

protocolParamTxFeeFixed ∷ Natural

The constant factor for the minimum fee calculation.

protocolParamTxFeePerByte ∷ Natural

The linear factor for the minimum fee calculation.

protocolParamMinUTxOValue ∷ Lovelace

The minimum permitted value for new UTxO entries, ie for transaction outputs.

protocolParamStakeAddressDeposit ∷ Lovelace

The deposit required to register a stake address.

protocolParamStakePoolDeposit ∷ Lovelace

The deposit required to register a stake pool.

protocolParamMinPoolCost ∷ Lovelace

The minimum value that stake pools are permitted to declare for their cost parameter.

protocolParamPoolRetireMaxEpoch ∷ EpochNo

The maximum number of epochs into the future that stake pools are permitted to schedule a retirement.

protocolParamStakePoolTargetNum ∷ Natural

The equilibrium target number of stake pools.

This is the "k" incentives parameter from the design document.

protocolParamPoolPledgeInfluence ∷ Rational

The influence of the pledge in stake pool rewards.

This is the "a_0" incentives parameter from the design document.

protocolParamMonetaryExpansion ∷ Rational

The monetary expansion rate. This determines the fraction of the reserves that are added to the fee pot each epoch.

This is the "rho" incentives parameter from the design document.

protocolParamTreasuryCut ∷ Rational

The fraction of the fee pot each epoch that goes to the treasury.

This is the "tau" incentives parameter from the design document.

Scripts

toShelleyScript ∷ ScriptInEra era → Script (ShelleyLedgerEra era)

toShelleyMultiSig ∷ SimpleScript SimpleScriptV1 → MultiSig StandardCrypto

Conversion for the MultiSig language used by the Shelley era.

fromShelleyMultiSig ∷ MultiSig StandardCrypto → SimpleScript lang

Conversion for the MultiSig language used by the Shelley era.

toAllegraTimelock ∷ ∀ lang. SimpleScript lang → Timelock StandardCrypto

Conversion for the Timelock language that is shared between the Allegra and Mary eras.

fromAllegraTimelock ∷ TimeLocksSupported lang → Timelock StandardCrypto → SimpleScript lang

Conversion for the Timelock language that is shared between the Allegra and Mary eras.

toShelleyScriptHash ∷ ScriptHash → ScriptHash StandardCrypto

fromShelleyScriptHash ∷ ScriptHash StandardCrypto → ScriptHash

Certificates

data Certificate

Constructors

StakeAddressRegistrationCertificate StakeCredential

StakeAddressDeregistrationCertificate StakeCredential

StakeAddressDelegationCertificate StakeCredential PoolId

StakePoolRegistrationCertificate StakePoolParameters

StakePoolRetirementCertificate PoolId EpochNo

GenesisKeyDelegationCertificate (Hash GenesisKey) (Hash GenesisDelegateKey) (Hash VrfKey)

MIRCertificate MIRPot [(StakeCredential, Lovelace)]

toShelleyCertificate ∷ Certificate → DCert StandardCrypto

fromShelleyCertificate ∷ DCert StandardCrypto → Certificate

Operational certificates

data OperationalCertificate

Constructors

OperationalCertificate !(OCert StandardCrypto) !(VerificationKey StakePoolKey)

data OperationalCertificateIssueCounter

Constructors

OperationalCertificateIssueCounter !Word64 !(VerificationKey StakePoolKey)

data OperationalCertIssueError

Constructors

OperationalCertKeyMismatch (VerificationKey StakePoolKey) (VerificationKey StakePoolKey)
The stake pool verification key expected for the OperationalCertificateIssueCounter does not match the signing key supplied for signing.

Order: pool vkey expected, pool skey supplied

Stake Pool

data StakePoolMetadata

A representation of the required fields for off-chain stake pool metadata.

Constructors

StakePoolMetadata !Text !Text !Text !Text

stakePoolName ∷ StakePoolMetadata → Text

A name of up to 50 characters.

stakePoolDescription ∷ StakePoolMetadata → Text

A description of up to 255 characters.

stakePoolTicker ∷ StakePoolMetadata → Text

A ticker of 3-5 characters, for a compact display of stake pools in a wallet.

stakePoolHomepage ∷ StakePoolMetadata → Text

A URL to a homepage with additional information about the pool. n.b. the spec does not specify a character limit for this field.

data StakePoolMetadataReference

Constructors

StakePoolMetadataReference Text (Hash StakePoolMetadata)

stakePoolMetadataURL ∷ StakePoolMetadataReference → Text

stakePoolMetadataHash ∷ StakePoolMetadataReference → Hash StakePoolMetadata

data StakePoolParameters

Constructors

StakePoolParameters PoolId (Hash VrfKey) Lovelace Rational StakeAddress Lovelace [Hash StakeKey] [StakePoolRelay] (Maybe StakePoolMetadataReference)

stakePoolId ∷ StakePoolParameters → PoolId

stakePoolVRF ∷ StakePoolParameters → Hash VrfKey

stakePoolCost ∷ StakePoolParameters → Lovelace

stakePoolMargin ∷ StakePoolParameters → Rational

stakePoolRewardAccount ∷ StakePoolParameters → StakeAddress

stakePoolPledge ∷ StakePoolParameters → Lovelace

stakePoolOwners ∷ StakePoolParameters → [Hash StakeKey]

stakePoolRelays ∷ StakePoolParameters → [StakePoolRelay]

stakePoolMetadata ∷ StakePoolParameters → Maybe StakePoolMetadataReference

data StakePoolRelay

Constructors

StakePoolRelayIp (Maybe IPv4) (Maybe IPv6) (Maybe PortNumber)
One or both of IPv4 & IPv6

StakePoolRelayDnsARecord ByteString (Maybe PortNumber)
An DNS name pointing to a A or AAAA record.

StakePoolRelayDnsSrvRecord ByteString
A DNS name pointing to a SRV record.

newtype EpochNo

Constructors

EpochNo

Fields

unEpochNo ∷ Word64

Stake pool operator's keys

data StakePoolKey

type PoolId = Hash StakePoolKey

KES keys

data KesKey

newtype KESPeriod

Constructors

KESPeriod

Fields

unKESPeriod ∷ Word

VRF keys

data VrfKey

Low level protocol interaction with a Cardano node

data LocalNodeConnectInfo mode

Constructors

LocalNodeConnectInfo (ConsensusModeParams mode) NetworkId FilePath

data ShelleyMode

The Shelley-only consensus mode consists of only the Shelley era.

This was used for the early Shelley testnets prior to the use of the multi-era CardanoMode. It is useful for setting up Shelley test networks (e.g. for benchmarking) without having to go through the complication of the hard fork from Byron to Shelley eras. It also shows how a single-era consensus mode works. It may be replaced by other single-era modes in future.

data CardanoMode

The Cardano consensus mode consists of all the eras currently in use on the Cardano mainnet. This is currently: the ByronEra; ShelleyEra, AllegraEra and MaryEra, in that order.

This mode will be extended with new eras as the Cardano mainnet develops.

data ConsensusMode mode where

This GADT provides a value-level representation of all the consensus modes. This enables pattern matching on the era to allow them to be treated in a non-uniform way.

Constructors

ShelleyMode ∷ ConsensusMode ShelleyMode

CardanoMode ∷ ConsensusMode CardanoMode

data LocalNodeClientProtocols block point tip tx txerr query m

Constructors

LocalNodeClientProtocols (Maybe (ChainSyncClient block point tip m ())) (Maybe (LocalTxSubmissionClient tx txerr m ())) (Maybe (LocalStateQueryClient block point query m ()))

Shelley based eras

type family ShelleyLedgerEra era where ...

A type family that connects our era type tags to equivalent type tags used in the Shelley ledger library.

This type mapping connect types from this API with types in the Shelley ledger library which allows writing conversion functions in a more generic way.

Equations

ShelleyLedgerEra ShelleyEra = StandardShelley

ShelleyLedgerEra AllegraEra = StandardAllegra

ShelleyLedgerEra MaryEra = StandardMary

Local State Query

data QueryInShelleyBasedEra era result where

Constructors

QueryChainPoint ∷ QueryInShelleyBasedEra era ChainPoint

QueryEpoch ∷ QueryInShelleyBasedEra era EpochNo

QueryGenesisParameters ∷ QueryInShelleyBasedEra era GenesisParameters

QueryProtocolParameters ∷ QueryInShelleyBasedEra era ProtocolParameters

QueryProtocolParametersUpdate ∷ QueryInShelleyBasedEra era (Map (Hash GenesisKey) ProtocolParametersUpdate)

QueryStakeDistribution ∷ QueryInShelleyBasedEra era (Map (Hash StakePoolKey) Rational)

QueryUTxO ∷ Maybe (Set AddressAny) → QueryInShelleyBasedEra era (UTxO era)

QueryStakeAddresses ∷ Set StakeCredential → NetworkId → QueryInShelleyBasedEra era (Map StakeAddress Lovelace, Map StakeAddress PoolId)

QueryLedgerState ∷ QueryInShelleyBasedEra era (SerialisedLedgerState era)

QueryProtocolState ∷ QueryInShelleyBasedEra era (ProtocolState era)

data LedgerState era where

Constructors

LedgerState ∷ ShelleyLedgerEra era ~ ledgerera ⇒ NewEpochState ledgerera → LedgerState era

newtype ProtocolState era

Constructors

ProtocolState (Serialised (ChainDepState (Crypto (ShelleyLedgerEra era))))

newtype SerialisedLedgerState era

Constructors

SerialisedLedgerState (Serialised (NewEpochState (ShelleyLedgerEra era)))

newtype UTxO era

Constructors

UTxO (Map TxIn (TxOut era))

Conversions

toShelleyNetwork ∷ NetworkId → Network

fromShelleyPParams ∷ PParams ledgerera → ProtocolParameters