Cardano.Api.Shelley

Description

This module provides a library interface that is intended to be the complete API for Shelley covering everything, including exposing constructors for the lower level types.

Documentation

module Cardano.Api

Cryptographic key interface

Payment addresses

Constructing and inspecting Shelley payment addresses

data Address addrtype where

Addresses are used as locations where assets live. The address determines the rights needed to spend assets at the address: in particular holding some signing key or being able to satisfy the conditions of a script.

There are currently two types of address:

Byron addresses, which use the type tag ByronAddr; and

Shelley addresses, which use the type tag ShelleyAddr. Notably, Shelley addresses support scripts and stake delegation.

The address type is subtly from the ledger era in which each address type is valid: while Byron addresses are the only choice in the Byron era, the Shelley era and all subsequent eras support both Byron and Shelley addresses. The Address type param only says the type of the address (either Byron or Shelley). The AddressInEra type connects the address type with the era in which it is supported.

Constructors

ShelleyAddress ∷ Network → PaymentCredential StandardCrypto → StakeReference StandardCrypto → Address ShelleyAddr
Shelley addresses allow delegation. Shelley addresses were introduced in Shelley era and are thus supported from the Shelley era onwards

toShelleyAddr ∷ AddressInEra era → Addr StandardCrypto

fromShelleyAddr ∷ IsShelleyBasedEra era ⇒ Addr StandardCrypto → AddressInEra era

toShelleyStakeCredential ∷ StakeCredential → StakeCredential StandardCrypto

fromShelleyStakeCredential ∷ StakeCredential StandardCrypto → StakeCredential

data NetworkId

Constructors

Mainnet

Testnet !NetworkMagic

Stake addresses

data PaymentCredential

Constructors

PaymentCredentialByKey (Hash PaymentKey)

PaymentCredentialByScript ScriptHash

data StakeAddress where

Constructors

StakeAddress ∷ Network → StakeCredential StandardCrypto → StakeAddress

data StakeAddressReference

Constructors

StakeAddressByValue StakeCredential

StakeAddressByPointer StakeAddressPointer

NoStakeAddress

data StakeCredential

Constructors

StakeCredentialByKey (Hash StakeKey)

StakeCredentialByScript ScriptHash

toShelleyStakeAddr ∷ StakeAddress → RewardAcnt StandardCrypto

fromShelleyStakeAddr ∷ RewardAcnt StandardCrypto → StakeAddress

fromShelleyStakeReference ∷ StakeReference StandardCrypto → StakeAddressReference

fromShelleyPaymentCredential ∷ PaymentCredential StandardCrypto → PaymentCredential

Building transactions

Constructing and inspecting transactions

data TxBody era where

Constructors

ShelleyTxBody ∷ ShelleyBasedEra era → TxBody (ShelleyLedgerEra era) → Maybe (AuxiliaryData (ShelleyLedgerEra era)) → TxBody era

newtype TxId

Constructors

TxId (Hash StandardCrypto EraIndependentTxBody)

toShelleyTxId ∷ TxId → TxId StandardCrypto

fromShelleyTxId ∷ TxId StandardCrypto → TxId

data TxIn

Constructors

TxIn TxId TxIx

toShelleyTxIn ∷ TxIn → TxIn StandardCrypto

fromShelleyTxIn ∷ TxIn StandardCrypto → TxIn

data TxOut era

Constructors

TxOut (AddressInEra era) (TxOutValue era)

toShelleyTxOut ∷ ∀ era ledgerera. (ShelleyLedgerEra era ~ ledgerera, IsShelleyBasedEra era, ShelleyBased ledgerera) ⇒ TxOut era → TxOut ledgerera

fromShelleyTxOut ∷ ∀ era ledgerera. (ShelleyLedgerEra era ~ ledgerera, IsShelleyBasedEra era, ShelleyBased ledgerera) ⇒ TxOut ledgerera → TxOut era

newtype TxIx

Constructors

TxIx Word

newtype Lovelace

Constructors

Lovelace Integer

toShelleyLovelace ∷ Lovelace → Coin

fromShelleyLovelace ∷ Coin → Lovelace

toMaryValue ∷ Value → Value StandardCrypto

fromMaryValue ∷ Value StandardCrypto → Value

newtype SlotNo

Constructors

SlotNo Word64

Signing transactions

Creating transaction witnesses one by one, or all in one go.

data Tx era where

Constructors

ShelleyTx ∷ ShelleyBasedEra era → Tx (ShelleyLedgerEra era) → Tx era

Incremental signing and separate witnesses

data Witness era where

Constructors

ShelleyBootstrapWitness ∷ ShelleyBasedEra era → BootstrapWitness StandardCrypto → Witness era

ShelleyKeyWitness ∷ ShelleyBasedEra era → WitVKey Witness StandardCrypto → Witness era

ShelleyScriptWitness ∷ ShelleyBasedEra era → Script (ShelleyLedgerEra era) → Witness era

data ShelleyWitnessSigningKey

Constructors

WitnessPaymentKey (SigningKey PaymentKey)

WitnessPaymentExtendedKey (SigningKey PaymentExtendedKey)

WitnessStakeKey (SigningKey StakeKey)

WitnessStakeExtendedKey (SigningKey StakeExtendedKey)

WitnessStakePoolKey (SigningKey StakePoolKey)

WitnessGenesisKey (SigningKey GenesisKey)

WitnessGenesisExtendedKey (SigningKey GenesisExtendedKey)

WitnessGenesisDelegateKey (SigningKey GenesisDelegateKey)

WitnessGenesisDelegateExtendedKey (SigningKey GenesisDelegateExtendedKey)

data ShelleySigningKey

We support making key witnesses with both normal and extended signing keys.

getShelleyKeyWitnessVerificationKey ∷ ShelleySigningKey → VKey Witness StandardCrypto

makeShelleySignature ∷ SignableRepresentation tosign ⇒ tosign → ShelleySigningKey → SignedDSIGN StandardCrypto tosign

toShelleySigningKey ∷ ShelleyWitnessSigningKey → ShelleySigningKey

Reading one of several key types

data FromSomeType (c ∷ Type → Constraint) b where

Constructors

FromSomeType ∷ c a ⇒ AsType a → (a → b) → FromSomeType c b

Transaction metadata

Embedding additional structured data within transactions.

toShelleyMetadata ∷ Map Word64 TxMetadataValue → Map Word64 Metadatum

fromShelleyMetadata ∷ Map Word64 Metadatum → Map Word64 TxMetadataValue

Protocol parameter updates

newtype EpochNo

Constructors

EpochNo

Fields

unEpochNo ∷ Word64

newtype NetworkMagic

Constructors

NetworkMagic

Fields

unNetworkMagic ∷ Word32

Scripts

toShelleyScript ∷ ScriptInEra era → Script (ShelleyLedgerEra era)

toShelleyMultiSig ∷ SimpleScript SimpleScriptV1 → MultiSig StandardCrypto

Conversion for the MultiSig language used by the Shelley era.

fromShelleyMultiSig ∷ MultiSig StandardCrypto → SimpleScript lang

Conversion for the MultiSig language used by the Shelley era.

toAllegraTimelock ∷ ∀ lang. SimpleScript lang → Timelock StandardCrypto

Conversion for the Timelock language that is shared between the Allegra and Mary eras.

fromAllegraTimelock ∷ TimeLocksSupported lang → Timelock StandardCrypto → SimpleScript lang

Conversion for the Timelock language that is shared between the Allegra and Mary eras.

toShelleyScriptHash ∷ ScriptHash → ScriptHash StandardCrypto

fromShelleyScriptHash ∷ ScriptHash StandardCrypto → ScriptHash

Certificates

data Certificate

Constructors

StakeAddressRegistrationCertificate StakeCredential

StakeAddressDeregistrationCertificate StakeCredential

StakeAddressDelegationCertificate StakeCredential PoolId

StakePoolRegistrationCertificate StakePoolParameters

StakePoolRetirementCertificate PoolId EpochNo

GenesisKeyDelegationCertificate (Hash GenesisKey) (Hash GenesisDelegateKey) (Hash VrfKey)

MIRCertificate MIRPot [(StakeCredential, Lovelace)]

toShelleyCertificate ∷ Certificate → DCert StandardCrypto

fromShelleyCertificate ∷ DCert StandardCrypto → Certificate

Operational certificates

data OperationalCertificate

Constructors

OperationalCertificate !(OCert StandardCrypto) !(VerificationKey StakePoolKey)

data OperationalCertificateIssueCounter

Constructors

OperationalCertificateIssueCounter !Word64 !(VerificationKey StakePoolKey)

data OperationalCertIssueError

Constructors

OperationalCertKeyMismatch (VerificationKey StakePoolKey) (VerificationKey StakePoolKey)
The stake pool verification key expected for the OperationalCertificateIssueCounter does not match the signing key supplied for signing.

Order: pool vkey expected, pool skey supplied

Stake Pool

data StakePoolMetadata

A representation of the required fields for off-chain stake pool metadata.

Constructors

StakePoolMetadata !Text !Text !Text !Text

stakePoolName ∷ StakePoolMetadata → Text

A name of up to 50 characters.

stakePoolDescription ∷ StakePoolMetadata → Text

A description of up to 255 characters.

stakePoolTicker ∷ StakePoolMetadata → Text

A ticker of 3-5 characters, for a compact display of stake pools in a wallet.

stakePoolHomepage ∷ StakePoolMetadata → Text

A URL to a homepage with additional information about the pool. n.b. the spec does not specify a character limit for this field.

data StakePoolMetadataReference

Constructors

StakePoolMetadataReference Text (Hash StakePoolMetadata)

stakePoolMetadataURL ∷ StakePoolMetadataReference → Text

stakePoolMetadataHash ∷ StakePoolMetadataReference → Hash StakePoolMetadata

data StakePoolParameters

Constructors

StakePoolParameters PoolId (Hash VrfKey) Lovelace Rational StakeAddress Lovelace [Hash StakeKey] [StakePoolRelay] (Maybe StakePoolMetadataReference)

stakePoolId ∷ StakePoolParameters → PoolId

stakePoolVRF ∷ StakePoolParameters → Hash VrfKey

stakePoolCost ∷ StakePoolParameters → Lovelace

stakePoolMargin ∷ StakePoolParameters → Rational

stakePoolRewardAccount ∷ StakePoolParameters → StakeAddress

stakePoolPledge ∷ StakePoolParameters → Lovelace

stakePoolOwners ∷ StakePoolParameters → [Hash StakeKey]

stakePoolRelays ∷ StakePoolParameters → [StakePoolRelay]

stakePoolMetadata ∷ StakePoolParameters → Maybe StakePoolMetadataReference

data StakePoolRelay

Constructors

StakePoolRelayIp (Maybe IPv4) (Maybe IPv6) (Maybe PortNumber)
One or both of IPv4 & IPv6

StakePoolRelayDnsARecord ByteString (Maybe PortNumber)
An DNS name pointing to a A or AAAA record.

StakePoolRelayDnsSrvRecord ByteString
A DNS name pointing to a SRV record.

Stake pool operator's keys

data StakePoolKey

type PoolId = Hash StakePoolKey

KES keys

data KesKey

VRF keys

data VrfKey

Low level protocol interaction with a Cardano node

data LocalNodeConnectInfo mode block

Constructors

LocalNodeConnectInfo FilePath NetworkId (NodeConsensusMode mode block)

data ShelleyMode

data CardanoMode

data NodeConsensusMode mode block where

Constructors

ShelleyMode ∷ NodeConsensusMode ShelleyMode (ShelleyBlockHFC StandardShelley)

CardanoMode ∷ EpochSlots → NodeConsensusMode CardanoMode (CardanoBlock StandardCrypto)

data LocalNodeClientProtocols block

Constructors

LocalNodeClientProtocols (Maybe (ChainSyncClient block (Point block) (Tip block) IO ())) (Maybe (LocalTxSubmissionClient (GenTx block) (ApplyTxErr block) IO ())) (Maybe (LocalStateQueryClient block (Point block) (Query block) IO ()))

withNodeProtocolClient ∷ NodeConsensusMode mode block → ((SerialiseNodeToClientConstraints block, SupportedNetworkProtocolVersion block) ⇒ ProtocolClient block (BlockProtocol block) → a) → a

Shelley based eras

type family ShelleyLedgerEra era where ...

A type family that connects our era type tags to equivalent type tags used in the Shelley ledger library.

This type mapping connect types from this API with types in the Shelley ledger library which allows writing conversion functions in a more generic way.

Equations

ShelleyLedgerEra ShelleyEra = StandardShelley

ShelleyLedgerEra AllegraEra = StandardAllegra

ShelleyLedgerEra MaryEra = StandardMary

Conversions

toShelleyNetwork ∷ NetworkId → Network