Cardano.Api.SpecialByron

Description

Special Byron values that we can submit to a node to propose an update proposal or to vote on an update proposal. These are not transactions.

Documentation

newtype ByronUpdateProposal

Byron era update proposal

Constructors

ByronUpdateProposal

Fields

unByronUpdateProposal ∷ AProposal ByteString

data ByronProtocolParametersUpdate

Constructors

ByronProtocolParametersUpdate

Fields

bPpuScriptVersion ∷ !(Maybe Word16)

Redundant. This was meant to be the version of the Plutus smart contract language, however, there are no smart contracts nor scripts in the Byron era.

bPpuSlotDuration ∷ !(Maybe Natural)

Slot duration in milliseconds.

bPpuMaxBlockSize ∷ !(Maybe Natural)

Maximum block size in bytes.

bPpuMaxHeaderSize ∷ !(Maybe Natural)

Maximum block header size in bytes.

bPpuMaxTxSize ∷ !(Maybe Natural)

Maximum transaction size in bytes.

bPpuMaxProposalSize ∷ !(Maybe Natural)

Maximum update proposal size in bytes.

bPpuMpcThd ∷ !(Maybe LovelacePortion)

bPpuHeavyDelThd ∷ !(Maybe LovelacePortion)

Heavyweight delegation threshold. The delegate (i.e stakeholder) must possess no less than this threshold of stake in order to participate in heavyweight delegation.

bPpuUpdateVoteThd ∷ !(Maybe LovelacePortion)

bPpuUpdateProposalThd ∷ !(Maybe LovelacePortion)

bPpuUpdateProposalTTL ∷ !(Maybe SlotNumber)

bPpuSoftforkRule ∷ !(Maybe SoftforkRule)

Values defining the softfork resolution rule. When the stake belonging to block issuers, issuing a given block version, is greater than the current softfork resolution threshold, this block version is adopted.

bPpuTxFeePolicy ∷ !(Maybe TxFeePolicy)

Transaction fee policy represents a formula to compute the minimal allowed Fee for a transaction. Transactions with lesser fees won't be accepted.

bPpuUnlockStakeEpoch ∷ !(Maybe EpochNumber)

This has been re-purposed for unlocking the OuroborosBFT logic in the software. Relevant: CDEC-610

data family AsType t

A family of singleton types used in this API to indicate which type to use where it would otherwise be ambiguous or merely unclear.

Values of this type are passed to deserialisation functions for example.

makeProtocolParametersUpdate ∷ ByronProtocolParametersUpdate → ProtocolParametersUpdate

toByronLedgerUpdateProposal ∷ ByronUpdateProposal → GenTx ByronBlock

newtype ByronVote

Byron era votes

Constructors

ByronVote

Fields

unByronVote ∷ AVote ByteString

makeByronUpdateProposal ∷ NetworkId → ProtocolVersion → SoftwareVersion → SystemTag → InstallerHash → SomeByronSigningKey → ByronProtocolParametersUpdate → ByronUpdateProposal

makeByronVote ∷ NetworkId → SomeByronSigningKey → ByronUpdateProposal → Bool → ByronVote

toByronLedgertoByronVote ∷ ByronVote → GenTx ByronBlock