Cardano.Api.StakePoolMetadata

Description

Stake pool off-chain metadata

Stake pool off-chain metadata

data StakePoolMetadata

A representation of the required fields for off-chain stake pool metadata.

Constructors

StakePoolMetadata

Fields

stakePoolName ∷ !Text

A name of up to 50 characters.

stakePoolDescription ∷ !Text

A description of up to 255 characters.

stakePoolTicker ∷ !Text

A ticker of 3-5 characters, for a compact display of stake pools in a wallet.

stakePoolHomepage ∷ !Text

A URL to a homepage with additional information about the pool. n.b. the spec does not specify a character limit for this field.

validateAndHashStakePoolMetadata ∷ ByteString → Either StakePoolMetadataValidationError (StakePoolMetadata, Hash StakePoolMetadata)

Decode and validate the provided JSON-encoded bytes as StakePoolMetadata. Return the decoded metadata and the hash of the original bytes.

data StakePoolMetadataValidationError

A stake pool metadata validation error.

Constructors

StakePoolMetadataJsonDecodeError !String

StakePoolMetadataInvalidLengthError
The length of the JSON-encoded stake pool metadata exceeds the maximum.

Fields

!Int

Maximum byte length.

!Int

Actual byte length.

Data family instances

data family AsType t

A family of singleton types used in this API to indicate which type to use where it would otherwise be ambiguous or merely unclear.

Values of this type are passed to deserialisation functions for example.

data family Hash keyrole ∷ Type