Cardano.Api.Tx

Description

Complete, signed transactions

Signing transactions

Creating transaction witnesses one by one, or all in one go.

data Tx era where

Constructors

ByronTx ∷ ATxAux ByteString → Tx ByronEra

ShelleyTx ∷ ShelleyBasedEra era → Tx (ShelleyLedgerEra era) → Tx era

Bundled Patterns

pattern Tx ∷ TxBody era → [KeyWitness era] → Tx era

getTxBody ∷ ∀ era. Tx era → TxBody era

getTxWitnesses ∷ ∀ era. Tx era → [KeyWitness era]

data ScriptValidity

Indicates whether a script is expected to fail or pass validation.

Constructors

ScriptInvalid
Script is expected to fail validation. Transactions marked as such can include scripts that fail validation. Such transactions may be submitted to the chain, in which case the collateral will be taken upon on chain script validation failure.

ScriptValid
Script is expected to pass validation. Transactions marked as such cannot include scripts that fail validation.

Signing in one go

data ShelleySigningKey

We support making key witnesses with both normal and extended signing keys.

Constructors

ShelleyNormalSigningKey (SignKeyDSIGN StandardCrypto)
A normal ed25519 signing key

ShelleyExtendedSigningKey XPrv
An extended ed25519 signing key

toShelleySigningKey ∷ ShelleyWitnessSigningKey → ShelleySigningKey

signByronTransaction ∷ NetworkId → TxBody ByronEra → [SigningKey ByronKey] → Tx ByronEra

signShelleyTransaction ∷ IsShelleyBasedEra era ⇒ TxBody era → [ShelleyWitnessSigningKey] → Tx era

Incremental signing and separate witnesses

makeSignedTransaction ∷ ∀ era. [KeyWitness era] → TxBody era → Tx era

data KeyWitness era where

Constructors

ByronKeyWitness ∷ TxInWitness → KeyWitness ByronEra

ShelleyBootstrapWitness ∷ ShelleyBasedEra era → BootstrapWitness StandardCrypto → KeyWitness era

ShelleyKeyWitness ∷ ShelleyBasedEra era → WitVKey Witness StandardCrypto → KeyWitness era

makeByronKeyWitness ∷ ∀ key. IsByronKey key ⇒ NetworkId → TxBody ByronEra → SigningKey key → KeyWitness ByronEra

data ShelleyWitnessSigningKey

Constructors

WitnessPaymentKey (SigningKey PaymentKey)

WitnessPaymentExtendedKey (SigningKey PaymentExtendedKey)

WitnessStakeKey (SigningKey StakeKey)

WitnessStakeExtendedKey (SigningKey StakeExtendedKey)

WitnessStakePoolKey (SigningKey StakePoolKey)

WitnessGenesisKey (SigningKey GenesisKey)

WitnessGenesisExtendedKey (SigningKey GenesisExtendedKey)

WitnessGenesisDelegateKey (SigningKey GenesisDelegateKey)

WitnessGenesisDelegateExtendedKey (SigningKey GenesisDelegateExtendedKey)

WitnessGenesisUTxOKey (SigningKey GenesisUTxOKey)

makeShelleyKeyWitness ∷ ∀ era. IsShelleyBasedEra era ⇒ TxBody era → ShelleyWitnessSigningKey → KeyWitness era

data WitnessNetworkIdOrByronAddress

Either a network ID or a Byron address to be used in constructing a Shelley bootstrap witness.

Constructors

WitnessNetworkId !NetworkId
Network ID.

If this value is used in the construction of a Shelley bootstrap witness, the result will not consist of a derivation path. If that is required, specify a WitnessByronAddress value instead.

WitnessByronAddress !(Address ByronAddr)
Byron address.

If this value is used in the construction of a Shelley bootstrap witness, both the network ID and derivation path will be extracted from the address and used in the construction of the witness.

makeShelleyBootstrapWitness ∷ ∀ era. IsShelleyBasedEra era ⇒ WitnessNetworkIdOrByronAddress → TxBody era → SigningKey ByronKey → KeyWitness era

makeShelleySignature ∷ SignableRepresentation tosign ⇒ tosign → ShelleySigningKey → SignedDSIGN StandardCrypto tosign

getShelleyKeyWitnessVerificationKey ∷ ShelleySigningKey → VKey Witness StandardCrypto

getTxBodyAndWitnesses ∷ Tx era → (TxBody era, [KeyWitness era])

Data family instances

data family AsType t

A family of singleton types used in this API to indicate which type to use where it would otherwise be ambiguous or merely unclear.

Values of this type are passed to deserialisation functions for example.