Cardano.Api.TxBody

Description

Transaction bodies

Documentation

parseTxId ∷ Parser TxId

Transaction bodies

data TxBody era where

Constructors

ByronTxBody ∷ Annotated Tx ByteString → TxBody ByronEra

ShelleyTxBody

Fields

∷ ShelleyBasedEra era

→ TxBody (ShelleyLedgerEra era)

→ [Script (ShelleyLedgerEra era)]

→ TxBodyScriptData era

→ Maybe (AuxiliaryData (ShelleyLedgerEra era))

→ TxScriptValidity era

Mark script as expected to pass or fail validation

→ TxBody era

Bundled Patterns

pattern TxBody ∷ TxBodyContent ViewTx era → TxBody era

makeTransactionBody ∷ ∀ era. IsCardanoEra era ⇒ TxBodyContent BuildTx era → Either TxBodyError (TxBody era)

Deprecated: Use createAndValidateTransactionBody.

createTransactionBody ∷ ShelleyBasedEra era → TxBodyContent BuildTx era → TxBody era

createAndValidateTransactionBody ∷ ∀ era. IsCardanoEra era ⇒ TxBodyContent BuildTx era → Either TxBodyError (TxBody era)

data TxBodyContent build era

Constructors

TxBodyContent

Fields

txIns ∷ TxIns build era

txInsCollateral ∷ TxInsCollateral era

txInsReference ∷ TxInsReference build era

txOuts ∷ [TxOut CtxTx era]

txTotalCollateral ∷ TxTotalCollateral era

txReturnCollateral ∷ TxReturnCollateral CtxTx era

txFee ∷ TxFee era

txValidityRange ∷ (TxValidityLowerBound era, TxValidityUpperBound era)

txMetadata ∷ TxMetadataInEra era

txAuxScripts ∷ TxAuxScripts era

txExtraKeyWits ∷ TxExtraKeyWitnesses era

txProtocolParams ∷ BuildTxWith build (Maybe ProtocolParameters)

txWithdrawals ∷ TxWithdrawals build era

txCertificates ∷ TxCertificates build era

txUpdateProposal ∷ TxUpdateProposal era

txMintValue ∷ TxMintValue build era

txScriptValidity ∷ TxScriptValidity era

data TxBodyError

Constructors

TxBodyEmptyTxIns

TxBodyEmptyTxInsCollateral

TxBodyEmptyTxOuts

TxBodyOutputNegative Quantity TxOutInAnyEra

TxBodyOutputOverflow Quantity TxOutInAnyEra

TxBodyMetadataError [(Word64, TxMetadataRangeError)]

TxBodyMintAdaError

TxBodyMissingProtocolParams

TxBodyInIxOverflow TxIn

data TxBodyScriptData era where

Constructors

TxBodyNoScriptData ∷ TxBodyScriptData era

TxBodyScriptData ∷ ScriptDataSupportedInEra era → TxDats (ShelleyLedgerEra era) → Redeemers (ShelleyLedgerEra era) → TxBodyScriptData era

data TxScriptValidity era where

A representation of whether the era supports tx script validity.

The Alonzo and subsequent eras support script validity.

Constructors

TxScriptValidityNone ∷ TxScriptValidity era

TxScriptValidity ∷ TxScriptValiditySupportedInEra era → ScriptValidity → TxScriptValidity era
Tx script validity is supported in transactions in the Alonzo era onwards.

data TxScriptValiditySupportedInEra era where

Constructors

TxScriptValiditySupportedInAlonzoEra ∷ TxScriptValiditySupportedInEra AlonzoEra

TxScriptValiditySupportedInBabbageEra ∷ TxScriptValiditySupportedInEra BabbageEra

TxScriptValiditySupportedInConwayEra ∷ TxScriptValiditySupportedInEra ConwayEra

data ScriptValidity

Indicates whether a script is expected to fail or pass validation.

Constructors

ScriptInvalid
Script is expected to fail validation. Transactions marked as such can include scripts that fail validation. Such transactions may be submitted to the chain, in which case the collateral will be taken upon on chain script validation failure.

ScriptValid
Script is expected to pass validation. Transactions marked as such cannot include scripts that fail validation.

scriptValidityToIsValid ∷ ScriptValidity → IsValid

isValidToScriptValidity ∷ IsValid → ScriptValidity

scriptValidityToTxScriptValidity ∷ ShelleyBasedEra era → ScriptValidity → TxScriptValidity era

txScriptValidityToIsValid ∷ TxScriptValidity era → IsValid

txScriptValidityToScriptValidity ∷ TxScriptValidity era → ScriptValidity

Transaction Ids

newtype TxId

Constructors

TxId (Hash StandardCrypto EraIndependentTxBody)

getTxId ∷ ∀ era. TxBody era → TxId

Calculate the transaction identifier for a TxBody.

getTxIdShelley ∷ Crypto (ShelleyLedgerEra era) ~ StandardCrypto ⇒ EraTxBody (ShelleyLedgerEra era) ⇒ ShelleyBasedEra era → TxBody (ShelleyLedgerEra era) → TxId

Transaction inputs

data TxIn

Constructors

TxIn TxId TxIx

type TxIns build era = [(TxIn, BuildTxWith build (Witness WitCtxTxIn era))]

newtype TxIx

Constructors

TxIx Word

genesisUTxOPseudoTxIn ∷ NetworkId → Hash GenesisUTxOKey → TxIn

Compute the TxIn of the initial UTxO pseudo-transaction corresponding to the given address in the genesis initial funds.

The Shelley initial UTxO is constructed from the sgInitialFunds which is not a full UTxO but just a map from addresses to coin values.

This gets turned into a UTxO by making a pseudo-transaction for each address, with the 0th output being the coin value. So to spend from the initial UTxO we need this same TxIn to use as an input to the spending transaction.

Transaction outputs

data CtxTx

The context is a transaction body

data CtxUTxO

The context is the UTxO

data TxOut ctx era

Constructors

TxOut (AddressInEra era) (TxOutValue era) (TxOutDatum ctx era) (ReferenceScript era)

data TxOutValue era where

Constructors

TxOutAdaOnly ∷ OnlyAdaSupportedInEra era → Lovelace → TxOutValue era

TxOutValue ∷ MultiAssetSupportedInEra era → Value → TxOutValue era

data TxOutDatum ctx era where

Constructors

TxOutDatumNone ∷ TxOutDatum ctx era

TxOutDatumHash ∷ ScriptDataSupportedInEra era → Hash ScriptData → TxOutDatum ctx era
A transaction output that only specifies the hash of the datum, but not the full datum value.

TxOutDatumInline ∷ ReferenceTxInsScriptsInlineDatumsSupportedInEra era → HashableScriptData → TxOutDatum ctx era
A transaction output that specifies the whole datum instead of the datum hash. Note that the datum map will not be updated with this datum, it only exists at the transaction output.

Bundled Patterns

pattern TxOutDatumInTx ∷ ScriptDataSupportedInEra era → HashableScriptData → TxOutDatum CtxTx era

toCtxUTxOTxOut ∷ TxOut CtxTx era → TxOut CtxUTxO era

lovelaceToTxOutValue ∷ IsCardanoEra era ⇒ Lovelace → TxOutValue era

prettyRenderTxOut ∷ TxOutInAnyEra → Text

txOutValueToLovelace ∷ TxOutValue era → Lovelace

txOutValueToValue ∷ TxOutValue era → Value

parseHash ∷ SerialiseAsRawBytes (Hash a) ⇒ AsType (Hash a) → Parser (Hash a)

data TxOutInAnyEra where

Constructors

TxOutInAnyEra ∷ CardanoEra era → TxOut CtxTx era → TxOutInAnyEra

txOutInAnyEra ∷ IsCardanoEra era ⇒ TxOut CtxTx era → TxOutInAnyEra

Convenience constructor for TxOutInAnyEra

Other transaction body types

data TxInsCollateral era where

Constructors

TxInsCollateralNone ∷ TxInsCollateral era

TxInsCollateral ∷ CollateralSupportedInEra era → [TxIn] → TxInsCollateral era

data TxInsReference build era where

Constructors

TxInsReferenceNone ∷ TxInsReference build era

TxInsReference ∷ ReferenceTxInsScriptsInlineDatumsSupportedInEra era → [TxIn] → TxInsReference build era

data TxReturnCollateral ctx era where

Constructors

TxReturnCollateralNone ∷ TxReturnCollateral ctx era

TxReturnCollateral ∷ TxTotalAndReturnCollateralSupportedInEra era → TxOut ctx era → TxReturnCollateral ctx era

data TxTotalCollateral era where

Constructors

TxTotalCollateralNone ∷ TxTotalCollateral era

TxTotalCollateral ∷ TxTotalAndReturnCollateralSupportedInEra era → Lovelace → TxTotalCollateral era

data TxFee era where

Constructors

TxFeeImplicit ∷ TxFeesImplicitInEra era → TxFee era

TxFeeExplicit ∷ TxFeesExplicitInEra era → Lovelace → TxFee era

data TxValidityLowerBound era where

Constructors

TxValidityNoLowerBound ∷ TxValidityLowerBound era

TxValidityLowerBound ∷ ValidityLowerBoundSupportedInEra era → SlotNo → TxValidityLowerBound era

data TxValidityUpperBound era where

This was formerly known as the TTL.

Constructors

TxValidityNoUpperBound ∷ ValidityNoUpperBoundSupportedInEra era → TxValidityUpperBound era

TxValidityUpperBound ∷ ValidityUpperBoundSupportedInEra era → SlotNo → TxValidityUpperBound era

data TxMetadataInEra era where

Constructors

TxMetadataNone ∷ TxMetadataInEra era

TxMetadataInEra ∷ TxMetadataSupportedInEra era → TxMetadata → TxMetadataInEra era

data TxAuxScripts era where

Constructors

TxAuxScriptsNone ∷ TxAuxScripts era

TxAuxScripts ∷ AuxScriptsSupportedInEra era → [ScriptInEra era] → TxAuxScripts era

data TxExtraKeyWitnesses era where

Constructors

TxExtraKeyWitnessesNone ∷ TxExtraKeyWitnesses era

TxExtraKeyWitnesses ∷ TxExtraKeyWitnessesSupportedInEra era → [Hash PaymentKey] → TxExtraKeyWitnesses era

data TxWithdrawals build era where

Constructors

TxWithdrawalsNone ∷ TxWithdrawals build era

TxWithdrawals ∷ WithdrawalsSupportedInEra era → [(StakeAddress, Lovelace, BuildTxWith build (Witness WitCtxStake era))] → TxWithdrawals build era

data TxCertificates build era where

Constructors

TxCertificatesNone ∷ TxCertificates build era

TxCertificates ∷ CertificatesSupportedInEra era → [Certificate] → BuildTxWith build (Map StakeCredential (Witness WitCtxStake era)) → TxCertificates build era

data TxUpdateProposal era where

Constructors

TxUpdateProposalNone ∷ TxUpdateProposal era

TxUpdateProposal ∷ UpdateProposalSupportedInEra era → UpdateProposal → TxUpdateProposal era

data TxMintValue build era where

Constructors

TxMintNone ∷ TxMintValue build era

TxMintValue ∷ MultiAssetSupportedInEra era → Value → BuildTxWith build (Map PolicyId (ScriptWitness WitCtxMint era)) → TxMintValue build era

Building vs viewing transactions

data BuildTxWith build a where

Constructors

ViewTx ∷ BuildTxWith ViewTx a

BuildTxWith ∷ a → BuildTxWith BuildTx a

data BuildTx

data ViewTx

Era-dependent transaction body features

data CollateralSupportedInEra era where

A representation of whether the era supports transactions with inputs used only for collateral for script fees.

The Alonzo and subsequent eras support collateral inputs.

Constructors

CollateralInAlonzoEra ∷ CollateralSupportedInEra AlonzoEra

CollateralInBabbageEra ∷ CollateralSupportedInEra BabbageEra

CollateralInConwayEra ∷ CollateralSupportedInEra ConwayEra

data MultiAssetSupportedInEra era where

A representation of whether the era supports multi-asset transactions.

The Mary and subsequent eras support multi-asset transactions.

The negation of this is OnlyAdaSupportedInEra.

Constructors

MultiAssetInMaryEra ∷ MultiAssetSupportedInEra MaryEra
Multi-asset transactions are supported in the Mary era.

MultiAssetInAlonzoEra ∷ MultiAssetSupportedInEra AlonzoEra
Multi-asset transactions are supported in the Alonzo era.

MultiAssetInBabbageEra ∷ MultiAssetSupportedInEra BabbageEra
Multi-asset transactions are supported in the Babbage era.

MultiAssetInConwayEra ∷ MultiAssetSupportedInEra ConwayEra
Multi-asset transactions are supported in the Conway era.

data OnlyAdaSupportedInEra era where

A representation of whether the era supports only ada transactions.

Prior to the Mary era only ada transactions are supported. Multi-assets are supported from the Mary era onwards.

This is the negation of MultiAssetSupportedInEra. It exists since we need evidence to be positive.

Constructors

AdaOnlyInByronEra ∷ OnlyAdaSupportedInEra ByronEra

AdaOnlyInShelleyEra ∷ OnlyAdaSupportedInEra ShelleyEra

AdaOnlyInAllegraEra ∷ OnlyAdaSupportedInEra AllegraEra

data TxFeesExplicitInEra era where

A representation of whether the era requires explicitly specified fees in transactions.

The Byron era tx fees are implicit (as the difference bettween the sum of outputs and sum of inputs), but all later eras the fees are specified in the transaction explicitly.

Constructors

TxFeesExplicitInShelleyEra ∷ TxFeesExplicitInEra ShelleyEra

TxFeesExplicitInAllegraEra ∷ TxFeesExplicitInEra AllegraEra

TxFeesExplicitInMaryEra ∷ TxFeesExplicitInEra MaryEra

TxFeesExplicitInAlonzoEra ∷ TxFeesExplicitInEra AlonzoEra

TxFeesExplicitInBabbageEra ∷ TxFeesExplicitInEra BabbageEra

TxFeesExplicitInConwayEra ∷ TxFeesExplicitInEra ConwayEra

data TxFeesImplicitInEra era where

A representation of whether the era requires implicitly specified fees in transactions.

This is the negation of TxFeesExplicitInEra.

Constructors

TxFeesImplicitInByronEra ∷ TxFeesImplicitInEra ByronEra

data ValidityUpperBoundSupportedInEra era where

A representation of whether the era supports transactions with an upper bound on the range of slots in which they are valid.

The Shelley and subsequent eras support an upper bound on the validity range. In the Shelley era specifically it is actually required. It is optional in later eras.

Constructors

ValidityUpperBoundInShelleyEra ∷ ValidityUpperBoundSupportedInEra ShelleyEra

ValidityUpperBoundInAllegraEra ∷ ValidityUpperBoundSupportedInEra AllegraEra

ValidityUpperBoundInMaryEra ∷ ValidityUpperBoundSupportedInEra MaryEra

ValidityUpperBoundInAlonzoEra ∷ ValidityUpperBoundSupportedInEra AlonzoEra

ValidityUpperBoundInBabbageEra ∷ ValidityUpperBoundSupportedInEra BabbageEra

ValidityUpperBoundInConwayEra ∷ ValidityUpperBoundSupportedInEra ConwayEra

data ValidityNoUpperBoundSupportedInEra era where

A representation of whether the era supports transactions having no upper bound on the range of slots in which they are valid.

Note that the ShelleyEra does not support omitting a validity upper bound. It was introduced as a required field in Shelley and then made optional in Allegra and subsequent eras.

The Byron era supports this by virtue of the fact that it does not support validity ranges at all.

Constructors

ValidityNoUpperBoundInByronEra ∷ ValidityNoUpperBoundSupportedInEra ByronEra

ValidityNoUpperBoundInAllegraEra ∷ ValidityNoUpperBoundSupportedInEra AllegraEra

ValidityNoUpperBoundInMaryEra ∷ ValidityNoUpperBoundSupportedInEra MaryEra

ValidityNoUpperBoundInAlonzoEra ∷ ValidityNoUpperBoundSupportedInEra AlonzoEra

ValidityNoUpperBoundInBabbageEra ∷ ValidityNoUpperBoundSupportedInEra BabbageEra

ValidityNoUpperBoundInConwayEra ∷ ValidityNoUpperBoundSupportedInEra ConwayEra

data ValidityLowerBoundSupportedInEra era where

A representation of whether the era supports transactions with a lower bound on the range of slots in which they are valid.

The Allegra and subsequent eras support an optional lower bound on the validity range. No equivalent of ValidityNoUpperBoundSupportedInEra is needed since all eras support having no lower bound.

Constructors

ValidityLowerBoundInAllegraEra ∷ ValidityLowerBoundSupportedInEra AllegraEra

ValidityLowerBoundInMaryEra ∷ ValidityLowerBoundSupportedInEra MaryEra

ValidityLowerBoundInAlonzoEra ∷ ValidityLowerBoundSupportedInEra AlonzoEra

ValidityLowerBoundInBabbageEra ∷ ValidityLowerBoundSupportedInEra BabbageEra

ValidityLowerBoundInConwayEra ∷ ValidityLowerBoundSupportedInEra ConwayEra

data TxMetadataSupportedInEra era where

A representation of whether the era supports transaction metadata.

Transaction metadata is supported from the Shelley era onwards.

Constructors

TxMetadataInShelleyEra ∷ TxMetadataSupportedInEra ShelleyEra

TxMetadataInAllegraEra ∷ TxMetadataSupportedInEra AllegraEra

TxMetadataInMaryEra ∷ TxMetadataSupportedInEra MaryEra

TxMetadataInAlonzoEra ∷ TxMetadataSupportedInEra AlonzoEra

TxMetadataInBabbageEra ∷ TxMetadataSupportedInEra BabbageEra

TxMetadataInConwayEra ∷ TxMetadataSupportedInEra ConwayEra

data AuxScriptsSupportedInEra era where

A representation of whether the era supports auxiliary scripts in transactions.

Auxiliary scripts are supported from the Allegra era onwards.

Constructors

AuxScriptsInAllegraEra ∷ AuxScriptsSupportedInEra AllegraEra

AuxScriptsInMaryEra ∷ AuxScriptsSupportedInEra MaryEra

AuxScriptsInAlonzoEra ∷ AuxScriptsSupportedInEra AlonzoEra

AuxScriptsInBabbageEra ∷ AuxScriptsSupportedInEra BabbageEra

AuxScriptsInConwayEra ∷ AuxScriptsSupportedInEra ConwayEra

data TxExtraKeyWitnessesSupportedInEra era where

A representation of whether the era supports transactions that specify in the body that they need extra key witnesses, and where this fact is visible to scripts.

Extra key witnesses visible to scripts are supported from the Alonzo era onwards.

Constructors

ExtraKeyWitnessesInAlonzoEra ∷ TxExtraKeyWitnessesSupportedInEra AlonzoEra

ExtraKeyWitnessesInBabbageEra ∷ TxExtraKeyWitnessesSupportedInEra BabbageEra

ExtraKeyWitnessesInConwayEra ∷ TxExtraKeyWitnessesSupportedInEra ConwayEra

data ScriptDataSupportedInEra era where

A representation of whether the era supports script data in transactions.

Constructors

ScriptDataInAlonzoEra ∷ ScriptDataSupportedInEra AlonzoEra
Script data is supported in transactions in the Alonzo era.

ScriptDataInBabbageEra ∷ ScriptDataSupportedInEra BabbageEra

ScriptDataInConwayEra ∷ ScriptDataSupportedInEra ConwayEra

data WithdrawalsSupportedInEra era where

A representation of whether the era supports withdrawals from reward accounts.

The Shelley and subsequent eras support stake addresses, their associated reward accounts and support for withdrawals from them.

Constructors

WithdrawalsInShelleyEra ∷ WithdrawalsSupportedInEra ShelleyEra

WithdrawalsInAllegraEra ∷ WithdrawalsSupportedInEra AllegraEra

WithdrawalsInMaryEra ∷ WithdrawalsSupportedInEra MaryEra

WithdrawalsInAlonzoEra ∷ WithdrawalsSupportedInEra AlonzoEra

WithdrawalsInBabbageEra ∷ WithdrawalsSupportedInEra BabbageEra

WithdrawalsInConwayEra ∷ WithdrawalsSupportedInEra ConwayEra

data CertificatesSupportedInEra era where

A representation of whether the era supports Certificates embedded in transactions.

The Shelley and subsequent eras support such certificates.

Constructors

CertificatesInShelleyEra ∷ CertificatesSupportedInEra ShelleyEra

CertificatesInAllegraEra ∷ CertificatesSupportedInEra AllegraEra

CertificatesInMaryEra ∷ CertificatesSupportedInEra MaryEra

CertificatesInAlonzoEra ∷ CertificatesSupportedInEra AlonzoEra

CertificatesInBabbageEra ∷ CertificatesSupportedInEra BabbageEra

CertificatesInConwayEra ∷ CertificatesSupportedInEra ConwayEra

data UpdateProposalSupportedInEra era where

A representation of whether the era supports UpdateProposals embedded in transactions.

The Shelley and subsequent eras support such update proposals. They Byron era has a notion of an update proposal, but it is a standalone chain object and not embedded in a transaction.

Constructors

UpdateProposalInShelleyEra ∷ UpdateProposalSupportedInEra ShelleyEra

UpdateProposalInAllegraEra ∷ UpdateProposalSupportedInEra AllegraEra

UpdateProposalInMaryEra ∷ UpdateProposalSupportedInEra MaryEra

UpdateProposalInAlonzoEra ∷ UpdateProposalSupportedInEra AlonzoEra

UpdateProposalInBabbageEra ∷ UpdateProposalSupportedInEra BabbageEra

UpdateProposalInConwayEra ∷ UpdateProposalSupportedInEra ConwayEra

data TxTotalAndReturnCollateralSupportedInEra era where

Constructors

TxTotalAndReturnCollateralInBabbageEra ∷ TxTotalAndReturnCollateralSupportedInEra BabbageEra

TxTotalAndReturnCollateralInConwayEra ∷ TxTotalAndReturnCollateralSupportedInEra ConwayEra

Feature availability functions

collateralSupportedInEra ∷ CardanoEra era → Maybe (CollateralSupportedInEra era)

multiAssetSupportedInEra ∷ CardanoEra era → Either (OnlyAdaSupportedInEra era) (MultiAssetSupportedInEra era)

txFeesExplicitInEra ∷ CardanoEra era → Either (TxFeesImplicitInEra era) (TxFeesExplicitInEra era)

validityUpperBoundSupportedInEra ∷ CardanoEra era → Maybe (ValidityUpperBoundSupportedInEra era)

validityNoUpperBoundSupportedInEra ∷ CardanoEra era → Maybe (ValidityNoUpperBoundSupportedInEra era)

validityLowerBoundSupportedInEra ∷ CardanoEra era → Maybe (ValidityLowerBoundSupportedInEra era)

txMetadataSupportedInEra ∷ CardanoEra era → Maybe (TxMetadataSupportedInEra era)

auxScriptsSupportedInEra ∷ CardanoEra era → Maybe (AuxScriptsSupportedInEra era)

extraKeyWitnessesSupportedInEra ∷ CardanoEra era → Maybe (TxExtraKeyWitnessesSupportedInEra era)

scriptDataSupportedInEra ∷ CardanoEra era → Maybe (ScriptDataSupportedInEra era)

withdrawalsSupportedInEra ∷ CardanoEra era → Maybe (WithdrawalsSupportedInEra era)

certificatesSupportedInEra ∷ CardanoEra era → Maybe (CertificatesSupportedInEra era)

updateProposalSupportedInEra ∷ CardanoEra era → Maybe (UpdateProposalSupportedInEra era)

txScriptValiditySupportedInShelleyBasedEra ∷ ShelleyBasedEra era → Maybe (TxScriptValiditySupportedInEra era)

txScriptValiditySupportedInCardanoEra ∷ CardanoEra era → Maybe (TxScriptValiditySupportedInEra era)

totalAndReturnCollateralSupportedInEra ∷ CardanoEra era → Maybe (TxTotalAndReturnCollateralSupportedInEra era)

Inspecting ScriptWitnesses

data AnyScriptWitness era where

A ScriptWitness in any WitCtx. This lets us handle heterogeneous collections of script witnesses from multiple contexts.

Constructors

AnyScriptWitness ∷ ScriptWitness witctx era → AnyScriptWitness era

data ScriptWitnessIndex

Identify the location of a ScriptWitness within the context of a TxBody. These are indexes of the objects within the transaction that need or can use script witnesses: inputs, minted assets, withdrawals and certificates. These are simple numeric indices, enumerated from zero. Thus the indices are not stable if the transaction body is modified.

Constructors

ScriptWitnessIndexTxIn !Word
The n'th transaction input, in the order of the TxIds.

ScriptWitnessIndexMint !Word
The n'th minting PolicyId, in the order of the PolicyIds.

ScriptWitnessIndexCertificate !Word
The n'th certificate, in the list order of the certificates.

ScriptWitnessIndexWithdrawal !Word
The n'th withdrawal, in the order of the StakeAddresss.

renderScriptWitnessIndex ∷ ScriptWitnessIndex → String

collectTxBodyScriptWitnesses ∷ ∀ era. TxBodyContent BuildTx era → [(ScriptWitnessIndex, AnyScriptWitness era)]

Conversion to inline data

scriptDataToInlineDatum ∷ HashableScriptData → Datum ledgerera

Conversion of ScriptData to binary data which allows for the storage of data onchain within a transaction output.

Internal conversion functions & types

toByronTxId ∷ TxId → TxId

toShelleyTxId ∷ TxId → TxId StandardCrypto

toShelleyTxIn ∷ TxIn → TxIn StandardCrypto

This function may overflow on the transaction index. Call sites must ensure that all uses of this function are appropriately guarded.

toShelleyTxOut ∷ ∀ era ledgerera. ShelleyLedgerEra era ~ ledgerera ⇒ ShelleyBasedEra era → TxOut CtxUTxO era → TxOut ledgerera

toShelleyTxOutAny ∷ ∀ ctx era ledgerera. ShelleyLedgerEra era ~ ledgerera ⇒ ShelleyBasedEra era → TxOut ctx era → TxOut ledgerera

A variant of 'toShelleyTxOutAny that is used only internally to this module that works with a TxOut in any context (including CtxTx) by ignoring embedded datums (taking only their hash).

fromShelleyTxId ∷ TxId StandardCrypto → TxId

fromShelleyTxIn ∷ TxIn StandardCrypto → TxIn

fromShelleyTxOut ∷ ShelleyLedgerEra era ~ ledgerera ⇒ ShelleyBasedEra era → TxOut ledgerera → TxOut ctx era

toAlonzoRdmrPtr ∷ ScriptWitnessIndex → RdmrPtr

fromAlonzoRdmrPtr ∷ RdmrPtr → ScriptWitnessIndex

fromByronTxIn ∷ TxIn → TxIn

fromLedgerTxOuts ∷ ∀ era. ShelleyBasedEra era → TxBody (ShelleyLedgerEra era) → TxBodyScriptData era → [TxOut CtxTx era]

renderTxIn ∷ TxIn → Text

Misc helpers

calculateExecutionUnitsLovelace ∷ ExecutionUnitPrices → ExecutionUnits → Maybe Lovelace

orderStakeAddrs ∷ [(StakeAddress, x, v)] → [(StakeAddress, x, v)]

orderTxIns ∷ [(TxIn, v)] → [(TxIn, v)]

Data family instances

data family AsType t

A family of singleton types used in this API to indicate which type to use where it would otherwise be ambiguous or merely unclear.

Values of this type are passed to deserialisation functions for example.