Cardano.Api.TxBody

Description

Transaction bodies

Documentation

parseTxId :: Parser TxId

Transaction bodies

data TxBody era where

Constructors

ShelleyTxBody

Fields

:: ShelleyBasedEra era

-> TxBody (ShelleyLedgerEra era)

-> [Script (ShelleyLedgerEra era)]

-> TxBodyScriptData era

-> Maybe (TxAuxData (ShelleyLedgerEra era))

-> TxScriptValidity era

Mark script as expected to pass or fail validation

-> TxBody era

Bundled Patterns

pattern TxBody :: TxBodyContent ViewTx era -> TxBody era

createTransactionBody :: ShelleyBasedEra era -> TxBodyContent BuildTx era -> Either TxBodyError (TxBody era)

createAndValidateTransactionBody :: ShelleyBasedEra era -> TxBodyContent BuildTx era -> Either TxBodyError (TxBody era)

data TxBodyContent build era

Constructors

TxBodyContent

Fields

txIns :: TxIns build era

txInsCollateral :: TxInsCollateral era

txInsReference :: TxInsReference build era

txOuts :: [TxOut CtxTx era]

txTotalCollateral :: TxTotalCollateral era

txReturnCollateral :: TxReturnCollateral CtxTx era

txFee :: TxFee era

txValidityLowerBound :: TxValidityLowerBound era

txValidityUpperBound :: TxValidityUpperBound era

txMetadata :: TxMetadataInEra era

txAuxScripts :: TxAuxScripts era

txExtraKeyWits :: TxExtraKeyWitnesses era

txProtocolParams :: BuildTxWith build (Maybe (LedgerProtocolParameters era))

txWithdrawals :: TxWithdrawals build era

txCertificates :: TxCertificates build era

txUpdateProposal :: TxUpdateProposal era

txMintValue :: TxMintValue build era

txScriptValidity :: TxScriptValidity era

txProposalProcedures :: Maybe (Featured ConwayEraOnwards era [Proposal era])

txVotingProcedures :: Maybe (Featured ConwayEraOnwards era (VotingProcedures era))

Byron only

makeByronTransactionBody :: TxIns BuildTx ByronEra -> [TxOut CtxTx ByronEra] -> Either TxBodyError (Annotated Tx ByteString)

Transaction body builders

defaultTxBodyContent :: ShelleyBasedEra era -> TxBodyContent BuildTx era

defaultTxFee :: ShelleyBasedEra era -> TxFee era

defaultTxValidityUpperBound :: ShelleyBasedEra era -> TxValidityUpperBound era

setTxIns :: TxIns build era -> TxBodyContent build era -> TxBodyContent build era

modTxIns :: (TxIns build era -> TxIns build era) -> TxBodyContent build era -> TxBodyContent build era

addTxIn :: (TxIn, BuildTxWith build (Witness WitCtxTxIn era)) -> TxBodyContent build era -> TxBodyContent build era

setTxInsCollateral :: TxInsCollateral era -> TxBodyContent build era -> TxBodyContent build era

setTxInsReference :: TxInsReference build era -> TxBodyContent build era -> TxBodyContent build era

setTxOuts :: [TxOut CtxTx era] -> TxBodyContent build era -> TxBodyContent build era

modTxOuts :: ([TxOut CtxTx era] -> [TxOut CtxTx era]) -> TxBodyContent build era -> TxBodyContent build era

addTxOut :: TxOut CtxTx era -> TxBodyContent build era -> TxBodyContent build era

setTxTotalCollateral :: TxTotalCollateral era -> TxBodyContent build era -> TxBodyContent build era

setTxReturnCollateral :: TxReturnCollateral CtxTx era -> TxBodyContent build era -> TxBodyContent build era

setTxFee :: TxFee era -> TxBodyContent build era -> TxBodyContent build era

setTxValidityLowerBound :: TxValidityLowerBound era -> TxBodyContent build era -> TxBodyContent build era

setTxValidityUpperBound :: TxValidityUpperBound era -> TxBodyContent build era -> TxBodyContent build era

setTxMetadata :: TxMetadataInEra era -> TxBodyContent build era -> TxBodyContent build era

setTxAuxScripts :: TxAuxScripts era -> TxBodyContent build era -> TxBodyContent build era

setTxExtraKeyWits :: TxExtraKeyWitnesses era -> TxBodyContent build era -> TxBodyContent build era

setTxProtocolParams :: BuildTxWith build (Maybe (LedgerProtocolParameters era)) -> TxBodyContent build era -> TxBodyContent build era

setTxWithdrawals :: TxWithdrawals build era -> TxBodyContent build era -> TxBodyContent build era

setTxCertificates :: TxCertificates build era -> TxBodyContent build era -> TxBodyContent build era

setTxUpdateProposal :: TxUpdateProposal era -> TxBodyContent build era -> TxBodyContent build era

setTxMintValue :: TxMintValue build era -> TxBodyContent build era -> TxBodyContent build era

setTxScriptValidity :: TxScriptValidity era -> TxBodyContent build era -> TxBodyContent build era

data TxBodyError

Constructors

TxBodyEmptyTxIns

TxBodyEmptyTxInsCollateral

TxBodyEmptyTxOuts

TxBodyOutputNegative !Quantity !TxOutInAnyEra

TxBodyOutputOverflow !Quantity !TxOutInAnyEra

TxBodyMetadataError ![(Word64, TxMetadataRangeError)]

TxBodyMintAdaError

TxBodyInIxOverflow !TxIn

TxBodyMissingProtocolParams

TxBodyProtocolParamsConversionError !ProtocolParametersConversionError

data TxBodyScriptData era where

Constructors

TxBodyNoScriptData :: TxBodyScriptData era

TxBodyScriptData :: AlonzoEraOnwards era -> TxDats (ShelleyLedgerEra era) -> Redeemers (ShelleyLedgerEra era) -> TxBodyScriptData era

data TxScriptValidity era where

A representation of whether the era supports tx script validity.

The Alonzo and subsequent eras support script validity.

Constructors

TxScriptValidityNone :: TxScriptValidity era

TxScriptValidity :: AlonzoEraOnwards era -> ScriptValidity -> TxScriptValidity era
Tx script validity is supported in transactions in the Alonzo era onwards.

data ScriptValidity

Indicates whether a script is expected to fail or pass validation.

Constructors

ScriptInvalid
Script is expected to fail validation. Transactions marked as such can include scripts that fail validation. Such transactions may be submitted to the chain, in which case the collateral will be taken upon on chain script validation failure.

ScriptValid
Script is expected to pass validation. Transactions marked as such cannot include scripts that fail validation.

scriptValidityToIsValid :: ScriptValidity -> IsValid

isValidToScriptValidity :: IsValid -> ScriptValidity

scriptValidityToTxScriptValidity :: ShelleyBasedEra era -> ScriptValidity -> TxScriptValidity era

txScriptValidityToIsValid :: TxScriptValidity era -> IsValid

txScriptValidityToScriptValidity :: TxScriptValidity era -> ScriptValidity

Transaction Ids

newtype TxId

Constructors

TxId (Hash StandardCrypto EraIndependentTxBody)

getTxId :: TxBody era -> TxId

Calculate the transaction identifier for a TxBody.

getTxIdByron :: ATxAux ByteString -> TxId

getTxIdShelley :: EraCrypto (ShelleyLedgerEra era) ~ StandardCrypto => EraTxBody (ShelleyLedgerEra era) => ShelleyBasedEra era -> TxBody (ShelleyLedgerEra era) -> TxId

Transaction inputs

data TxIn

Constructors

TxIn TxId TxIx

type TxIns build era = [(TxIn, BuildTxWith build (Witness WitCtxTxIn era))]

newtype TxIx

Constructors

TxIx Word

genesisUTxOPseudoTxIn :: NetworkId -> Hash GenesisUTxOKey -> TxIn

Compute the TxIn of the initial UTxO pseudo-transaction corresponding to the given address in the genesis initial funds.

The Shelley initial UTxO is constructed from the sgInitialFunds which is not a full UTxO but just a map from addresses to coin values.

This gets turned into a UTxO by making a pseudo-transaction for each address, with the 0th output being the coin value. So to spend from the initial UTxO we need this same TxIn to use as an input to the spending transaction.

Transaction outputs

data CtxTx

The context is a transaction body

data CtxUTxO

The context is the UTxO

data TxOut ctx era

Constructors

TxOut (AddressInEra era) (TxOutValue era) (TxOutDatum ctx era) (ReferenceScript era)

data TxOutValue era where

Constructors

TxOutValueByron :: Lovelace -> TxOutValue era

TxOutValueShelleyBased :: (Eq (Value (ShelleyLedgerEra era)), Show (Value (ShelleyLedgerEra era))) => ShelleyBasedEra era -> Value (ShelleyLedgerEra era) -> TxOutValue era

data TxOutDatum ctx era where

Constructors

TxOutDatumNone :: TxOutDatum ctx era

TxOutDatumHash :: AlonzoEraOnwards era -> Hash ScriptData -> TxOutDatum ctx era
A transaction output that only specifies the hash of the datum, but not the full datum value.

TxOutDatumInline :: BabbageEraOnwards era -> HashableScriptData -> TxOutDatum ctx era
A transaction output that specifies the whole datum instead of the datum hash. Note that the datum map will not be updated with this datum, it only exists at the transaction output.

Bundled Patterns

pattern TxOutDatumInTx :: AlonzoEraOnwards era -> HashableScriptData -> TxOutDatum CtxTx era

toCtxUTxOTxOut :: TxOut CtxTx era -> TxOut CtxUTxO era

lovelaceToTxOutValue :: ShelleyBasedEra era -> Lovelace -> TxOutValue era

prettyRenderTxOut :: TxOutInAnyEra -> Text

txOutValueToLovelace :: TxOutValue era -> Lovelace

txOutValueToValue :: TxOutValue era -> Value

parseHash :: SerialiseAsRawBytes (Hash a) => AsType (Hash a) -> Parser (Hash a)

data TxOutInAnyEra where

Constructors

TxOutInAnyEra :: CardanoEra era -> TxOut CtxTx era -> TxOutInAnyEra

txOutInAnyEra :: CardanoEra era -> TxOut CtxTx era -> TxOutInAnyEra

Convenience constructor for TxOutInAnyEra

Other transaction body types

data TxInsCollateral era where

Constructors

TxInsCollateralNone :: TxInsCollateral era

TxInsCollateral :: AlonzoEraOnwards era -> [TxIn] -> TxInsCollateral era

data TxInsReference build era where

Constructors

TxInsReferenceNone :: TxInsReference build era

TxInsReference :: BabbageEraOnwards era -> [TxIn] -> TxInsReference build era

data TxReturnCollateral ctx era where

Constructors

TxReturnCollateralNone :: TxReturnCollateral ctx era

TxReturnCollateral :: BabbageEraOnwards era -> TxOut ctx era -> TxReturnCollateral ctx era

data TxTotalCollateral era where

Constructors

TxTotalCollateralNone :: TxTotalCollateral era

TxTotalCollateral :: BabbageEraOnwards era -> Lovelace -> TxTotalCollateral era

data TxFee era where

Constructors

TxFeeExplicit :: ShelleyBasedEra era -> Lovelace -> TxFee era

data TxValidityLowerBound era where

Constructors

TxValidityNoLowerBound :: TxValidityLowerBound era

TxValidityLowerBound :: AllegraEraOnwards era -> SlotNo -> TxValidityLowerBound era

data TxValidityUpperBound era where

This was formerly known as the TTL.

Constructors

TxValidityUpperBound :: ShelleyBasedEra era -> Maybe SlotNo -> TxValidityUpperBound era

data TxMetadataInEra era where

Constructors

TxMetadataNone :: TxMetadataInEra era

TxMetadataInEra :: ShelleyBasedEra era -> TxMetadata -> TxMetadataInEra era

data TxAuxScripts era where

Constructors

TxAuxScriptsNone :: TxAuxScripts era

TxAuxScripts :: AllegraEraOnwards era -> [ScriptInEra era] -> TxAuxScripts era

data TxExtraKeyWitnesses era where

Constructors

TxExtraKeyWitnessesNone :: TxExtraKeyWitnesses era

TxExtraKeyWitnesses :: AlonzoEraOnwards era -> [Hash PaymentKey] -> TxExtraKeyWitnesses era

data TxWithdrawals build era where

Constructors

TxWithdrawalsNone :: TxWithdrawals build era

TxWithdrawals :: ShelleyBasedEra era -> [(StakeAddress, Lovelace, BuildTxWith build (Witness WitCtxStake era))] -> TxWithdrawals build era

data TxCertificates build era where

Constructors

TxCertificatesNone :: TxCertificates build era

TxCertificates :: ShelleyBasedEra era -> [Certificate era] -> BuildTxWith build (Map StakeCredential (Witness WitCtxStake era)) -> TxCertificates build era

data TxUpdateProposal era where

Constructors

TxUpdateProposalNone :: TxUpdateProposal era

TxUpdateProposal :: ShelleyToBabbageEra era -> UpdateProposal -> TxUpdateProposal era

data TxMintValue build era where

Constructors

TxMintNone :: TxMintValue build era

TxMintValue :: MaryEraOnwards era -> Value -> BuildTxWith build (Map PolicyId (ScriptWitness WitCtxMint era)) -> TxMintValue build era

Building vs viewing transactions

data BuildTxWith build a where

Constructors

ViewTx :: BuildTxWith ViewTx a

BuildTxWith :: a -> BuildTxWith BuildTx a

data BuildTx

data ViewTx

Inspecting ScriptWitnesses

data AnyScriptWitness era where

A ScriptWitness in any WitCtx. This lets us handle heterogeneous collections of script witnesses from multiple contexts.

Constructors

AnyScriptWitness :: ScriptWitness witctx era -> AnyScriptWitness era

data ScriptWitnessIndex

Identify the location of a ScriptWitness within the context of a TxBody. These are indexes of the objects within the transaction that need or can use script witnesses: inputs, minted assets, withdrawals and certificates. These are simple numeric indices, enumerated from zero. Thus the indices are not stable if the transaction body is modified.

Constructors

ScriptWitnessIndexTxIn !Word
The n'th transaction input, in the order of the TxIds.

ScriptWitnessIndexMint !Word
The n'th minting PolicyId, in the order of the PolicyIds.

ScriptWitnessIndexCertificate !Word
The n'th certificate, in the list order of the certificates.

ScriptWitnessIndexWithdrawal !Word
The n'th withdrawal, in the order of the StakeAddresss.

renderScriptWitnessIndex :: ScriptWitnessIndex -> String

collectTxBodyScriptWitnesses :: forall era. ShelleyBasedEra era -> TxBodyContent BuildTx era -> [(ScriptWitnessIndex, AnyScriptWitness era)]

Conversion to inline data

scriptDataToInlineDatum :: Era ledgerera => HashableScriptData -> Datum ledgerera

Conversion of ScriptData to binary data which allows for the storage of data onchain within a transaction output.

Internal conversion functions & types

toByronTxId :: TxId -> TxId

toShelleyTxId :: TxId -> TxId StandardCrypto

toShelleyTxIn :: TxIn -> TxIn StandardCrypto

This function may overflow on the transaction index. Call sites must ensure that all uses of this function are appropriately guarded.

toShelleyTxOut :: forall era ledgerera. ShelleyLedgerEra era ~ ledgerera => ShelleyBasedEra era -> TxOut CtxUTxO era -> TxOut ledgerera

toShelleyTxOutAny :: forall ctx era ledgerera. ShelleyLedgerEra era ~ ledgerera => ShelleyBasedEra era -> TxOut ctx era -> TxOut ledgerera

A variant of 'toShelleyTxOutAny that is used only internally to this module that works with a TxOut in any context (including CtxTx) by ignoring embedded datums (taking only their hash).

fromShelleyTxId :: TxId StandardCrypto -> TxId

fromShelleyTxIn :: TxIn StandardCrypto -> TxIn

fromShelleyTxOut :: forall era ctx. ShelleyBasedEra era -> TxOut (ShelleyLedgerEra era) -> TxOut ctx era

toAlonzoRdmrPtr :: ScriptWitnessIndex -> RdmrPtr

fromAlonzoRdmrPtr :: RdmrPtr -> ScriptWitnessIndex

fromByronTxIn :: TxIn -> TxIn

fromLedgerTxOuts :: forall era. ShelleyBasedEra era -> TxBody (ShelleyLedgerEra era) -> TxBodyScriptData era -> [TxOut CtxTx era]

renderTxIn :: TxIn -> Text

Misc helpers

calculateExecutionUnitsLovelace :: Prices -> ExecutionUnits -> Maybe Lovelace

orderStakeAddrs :: [(StakeAddress, x, v)] -> [(StakeAddress, x, v)]

orderTxIns :: [(TxIn, v)] -> [(TxIn, v)]

Data family instances

data family AsType t

A family of singleton types used in this API to indicate which type to use where it would otherwise be ambiguous or merely unclear.

Values of this type are passed to deserialisation functions for example.

getTxBodyContent :: TxBody era -> TxBodyContent ViewTx era