Cardano.Api.TxIn

Description

Transaction bodies

Transaction inputs

data TxIn

Constructors

TxIn TxId TxIx

newtype TxIx

Constructors

TxIx Word

Transaction Ids

newtype TxId

Constructors

TxId (Hash StandardCrypto EraIndependentTxBody)

parseTxId ∷ Parser TxId

Data family instances

data family AsType t

A family of singleton types used in this API to indicate which type to use where it would otherwise be ambiguous or merely unclear.

Values of this type are passed to deserialisation functions for example.

Internal conversion functions

toByronTxId ∷ TxId → TxId

toShelleyTxId ∷ TxId → TxId StandardCrypto

fromShelleyTxId ∷ TxId StandardCrypto → TxId

toByronTxIn ∷ TxIn → TxIn

fromByronTxIn ∷ TxIn → TxIn

toShelleyTxIn ∷ TxIn → TxIn StandardCrypto

This function may overflow on the transaction index. Call sites must ensure that all uses of this function are appropriately guarded.

fromShelleyTxIn ∷ TxIn StandardCrypto → TxIn

renderTxIn ∷ TxIn → Text