Cardano.Api.Value

Description

Currency values

Documentation

newtype Lovelace

Constructors

Lovelace Integer

Multi-asset values

newtype Quantity

Constructors

Quantity Integer

newtype PolicyId

Constructors

PolicyId

Fields

unPolicyId ∷ ScriptHash

scriptPolicyId ∷ Script lang → PolicyId

newtype AssetName

Constructors

AssetName ByteString

data AssetId

Constructors

AdaAssetId

AssetId !PolicyId !AssetName

data Value

selectAsset ∷ Value → AssetId → Quantity

valueFromList ∷ [(AssetId, Quantity)] → Value

valueToList ∷ Value → [(AssetId, Quantity)]

filterValue ∷ (AssetId → Bool) → Value → Value

negateValue ∷ Value → Value

This lets you write a - b as a <> negateValue b.

calcMinimumDeposit ∷ Value → Lovelace → Lovelace

Calculate cost of making a UTxO entry for a given Value and mininimum UTxO value derived from the ProtocolParameters

Ada / Lovelace specifically

quantityToLovelace ∷ Quantity → Lovelace

lovelaceToQuantity ∷ Lovelace → Quantity

selectLovelace ∷ Value → Lovelace

lovelaceToValue ∷ Lovelace → Value

valueToLovelace ∷ Value → Maybe Lovelace

Check if the Value consists of only Lovelace and no other assets, and if so then return the Lovelace.

See also selectLovelace to select the Lovelace quantity from the Value, ignoring other assets.

Alternative nested representation

newtype ValueNestedRep

An alternative nested representation for Value that groups assets that share a PolicyId.

Constructors

ValueNestedRep [ValueNestedBundle]

data ValueNestedBundle

A bundle within a ValueNestedRep for a single PolicyId, or for the special case of ada.

Constructors

ValueNestedBundleAda Quantity

ValueNestedBundle PolicyId (Map AssetName Quantity)

valueToNestedRep ∷ Value → ValueNestedRep

valueFromNestedRep ∷ ValueNestedRep → Value

Rendering

renderValue ∷ Value → Text

Render a textual representation of a Value.

renderValuePretty ∷ Value → Text

Render a "prettified" textual representation of a Value.

Internal conversion functions

toByronLovelace ∷ Lovelace → Maybe Lovelace

fromByronLovelace ∷ Lovelace → Lovelace

toShelleyLovelace ∷ Lovelace → Coin

fromShelleyLovelace ∷ Coin → Lovelace

fromShelleyDeltaLovelace ∷ DeltaCoin → Lovelace

toMaryValue ∷ Value → MaryValue StandardCrypto

fromMaryValue ∷ MaryValue StandardCrypto → Value

Data family instances

data family AsType t

A family of singleton types used in this API to indicate which type to use where it would otherwise be ambiguous or merely unclear.

Values of this type are passed to deserialisation functions for example.