Cardano.Rpc.Server.Config

Documentation

type RpcConfig = RpcConfigF Identity

type PartialRpcConfig = RpcConfigF Last

data RpcConfigF (m :: Type -> Type)

RPC server configuration, which is a part of cardano-node configuration.

Constructors

RpcConfig

Fields

isEnabled :: !(m Bool)

whether the RPC server is enabled

rpcSocketPath :: !(m SocketPath)

path to the socket file where the RPC server listens

nodeSocketPath :: !(m SocketPath)

cardano-node socket path. Only valid if RPC endpoint is enabled.

makeRpcConfig :: MonadError String m => PartialRpcConfig -> m RpcConfig

Build RPC Config

Uses the following defaults if the values are not provided * RPC is disabled * rpc.sock is placed in the same path as the node socket

Validates if the node socket is enabled if RPC is enabled.

nodeSocketPathToRpcSocketPath :: SocketPath -> SocketPath

Convert node socket path to a default rpc socket path. By default it's rpc.sock in the same directory as node socket path.