Test.Gen.Cardano.Api.Typed

Documentation

genFeaturedInEra ∷ Alternative f ⇒ eon era → f a → f (Featured eon era a)

Generate a Featured for the given CardanoEra with the provided generator.

genMaybeFeaturedInEra ∷ Eon eon ⇒ Alternative f ⇒ (eon era → f a) → CardanoEra era → f (Maybe (Featured eon era a))

Generate a Featured for the given CardanoEra with the provided generator.

genAddressByron ∷ Gen (Address ByronAddr)

genAddressInEra ∷ CardanoEra era → Gen (AddressInEra era)

genAddressShelley ∷ Gen (Address ShelleyAddr)

genCertificate ∷ ShelleyBasedEra era → Gen (Certificate era)

genCostModel ∷ Gen CostModel

genMaybePraosNonce ∷ Gen (Maybe PraosNonce)

genPraosNonce ∷ Gen PraosNonce

genValidProtocolParameters ∷ ShelleyBasedEra era → Gen (LedgerProtocolParameters era)

Generate valid protocol parameters which pass validations in Cardano.Api.ProtocolParameters

genProtocolParameters ∷ CardanoEra era → Gen ProtocolParameters

genValueNestedRep ∷ Gen ValueNestedRep

genValueNestedBundle ∷ Gen ValueNestedBundle

genByronKeyWitness ∷ Gen (KeyWitness ByronEra)

genCardanoKeyWitness ∷ CardanoEra era → Gen (KeyWitness era)

genShelleyKeyWitness ∷ IsShelleyBasedEra era ⇒ CardanoEra era → Gen (KeyWitness era)

genTxId ∷ Gen TxId

genTxIn ∷ Gen TxIn

genTxOutTxContext ∷ CardanoEra era → Gen (TxOut CtxTx era)

genTxOutUTxOContext ∷ CardanoEra era → Gen (TxOut CtxUTxO era)

genUTxO ∷ CardanoEra era → Gen (UTxO era)

Scripts

genHashableScriptData ∷ Gen HashableScriptData

genReferenceScript ∷ CardanoEra era → Gen (ReferenceScript era)

genScript ∷ ScriptLanguage lang → Gen (Script lang)

genSimpleScript ∷ Gen SimpleScript

genPlutusScript ∷ PlutusScriptVersion lang → Gen (PlutusScript lang)

genScriptInAnyLang ∷ Gen ScriptInAnyLang

genScriptInEra ∷ CardanoEra era → Gen (ScriptInEra era)

genScriptHash ∷ Gen ScriptHash

genScriptData ∷ Gen ScriptData

Deprecated: Use genHashableScriptData

genScriptDataSchema ∷ Gen ScriptDataJsonSchema

genScriptValidity ∷ Gen ScriptValidity

genAssetName ∷ Gen AssetName

genAssetId ∷ Gen AssetId

genEpochNo ∷ Gen EpochNo

genExecutionUnitPrices ∷ Gen ExecutionUnitPrices

genExecutionUnits ∷ Gen ExecutionUnits

genHashScriptData ∷ Gen (Hash ScriptData)

genKESPeriod ∷ Gen KESPeriod

genNat ∷ Gen Natural

genNetworkId ∷ Gen NetworkId

genNetworkMagic ∷ Gen NetworkMagic

genOperationalCertificate ∷ Gen OperationalCertificate

genOperationalCertificateIssueCounter ∷ Gen OperationalCertificateIssueCounter

genOperationalCertificateWithCounter ∷ Gen (OperationalCertificate, OperationalCertificateIssueCounter)

genPaymentCredential ∷ Gen PaymentCredential

genPolicyId ∷ Gen PolicyId

genQuantity ∷ Range Integer → Gen Quantity

genRationalInt64 ∷ Gen Rational

genSeed ∷ Int → Gen Seed

genShelleyBootstrapWitness ∷ IsShelleyBasedEra era ⇒ CardanoEra era → Gen (KeyWitness era)

genShelleyHash ∷ Gen (Hash Blake2b_256 EraIndependentTxBody)

genShelleyWitness ∷ IsShelleyBasedEra era ⇒ CardanoEra era → Gen (KeyWitness era)

genShelleyWitnessSigningKey ∷ Gen ShelleyWitnessSigningKey

genSignedQuantity ∷ Gen Quantity

Generate a positive or negative quantity.

genSignedNonZeroQuantity ∷ Gen Quantity

Generate a positive or negative, but not zero quantity.

genSigningKey ∷ Key keyrole ⇒ AsType keyrole → Gen (SigningKey keyrole)

genSlotNo ∷ Gen SlotNo

genStakeAddress ∷ Gen StakeAddress

genStakeAddressReference ∷ Gen StakeAddressReference

genStakeCredential ∷ Gen StakeCredential

genTtl ∷ Gen SlotNo

genTx ∷ ∀ era. IsCardanoEra era ⇒ CardanoEra era → Gen (Tx era)

genTxAuxScripts ∷ CardanoEra era → Gen (TxAuxScripts era)

genTxBody ∷ IsCardanoEra era ⇒ CardanoEra era → Gen (TxBody era)

genTxBodyContent ∷ CardanoEra era → Gen (TxBodyContent BuildTx era)

genTxCertificates ∷ CardanoEra era → Gen (TxCertificates BuildTx era)

genTxFee ∷ CardanoEra era → Gen (TxFee era)

genTxIndex ∷ Gen TxIx

genTxInsCollateral ∷ CardanoEra era → Gen (TxInsCollateral era)

genTxInsReference ∷ CardanoEra era → Gen (TxInsReference BuildTx era)

genTxMetadataInEra ∷ CardanoEra era → Gen (TxMetadataInEra era)

genTxMintValue ∷ CardanoEra era → Gen (TxMintValue BuildTx era)

genLovelace ∷ Gen Lovelace

genPositiveLovelace ∷ Gen Lovelace

genValue ∷ Gen AssetId → Gen Quantity → Gen Value

genValueDefault ∷ Gen Value

Generate a Value with any asset ID and a positive or negative quantity.

genVerificationKey ∷ Key keyrole ⇒ AsType keyrole → Gen (VerificationKey keyrole)

genVerificationKeyHash ∷ Key keyrole ⇒ AsType keyrole → Gen (Hash keyrole)

genUpdateProposal ∷ CardanoEra era → Gen UpdateProposal

genProtocolParametersUpdate ∷ CardanoEra era → Gen ProtocolParametersUpdate

genTxOutDatumHashTxContext ∷ CardanoEra era → Gen (TxOutDatum CtxTx era)

genTxOutDatumHashUTxOContext ∷ CardanoEra era → Gen (TxOutDatum CtxUTxO era)

genTxOutValue ∷ CardanoEra era → Gen (TxOutValue era)

genTxReturnCollateral ∷ CardanoEra era → Gen (TxReturnCollateral CtxTx era)

genTxScriptValidity ∷ CardanoEra era → Gen (TxScriptValidity era)

genTxTotalCollateral ∷ CardanoEra era → Gen (TxTotalCollateral era)

genTxUpdateProposal ∷ CardanoEra era → Gen (TxUpdateProposal era)

genTxValidityLowerBound ∷ CardanoEra era → Gen (TxValidityLowerBound era)

genTxValidityRange ∷ CardanoEra era → Gen (TxValidityLowerBound era, TxValidityUpperBound era)

genTxValidityUpperBound ∷ CardanoEra era → Gen (TxValidityUpperBound era)

genTxWithdrawals ∷ CardanoEra era → Gen (TxWithdrawals BuildTx era)

genUnsignedQuantity ∷ Gen Quantity

genPositiveQuantity ∷ Gen Quantity

genValueForMinting ∷ Gen Value

Generate a Value suitable for minting, i.e. non-ADA asset ID and a positive or negative quantity.

genValueForTxOut ∷ Gen Value

Generate a Value suitable for usage in a transaction output, i.e. any asset ID and a positive quantity.

genWitnesses ∷ CardanoEra era → Gen [KeyWitness era]

genWitnessNetworkIdOrByronAddress ∷ Gen WitnessNetworkIdOrByronAddress

genRational ∷ Gen Rational

genGovernancePoll ∷ Gen GovernancePoll

genGovernancePollAnswer ∷ Gen GovernancePollAnswer

genProposals ∷ ∀ era. ConwayEraOnwards era → Gen [Proposal era]

genProposal ∷ ConwayEraOnwards era → Gen (Proposal era)

genVotingProcedures ∷ ConwayEraOnwards era → Gen (VotingProcedures era)